Description of Attitude data

Table of counts extracted from Andrich (1988). Data the response patterns observed for an eight item survey.

Details

The items in this survey were:

Capital punishment is one of the most hideous practices of our time.

The state cannot teach the sacredness of human life by destroying it.

Capital punishment is not an effective deterrent to crime.

I don't believe in capital punishment but I am not sure it isn't necessary.

I think capital punishment is necessary but I wish it were not.

Until we find a more civilized way to prevent crime we must have capital punishment.

Capital punishment is justified because it does act as a deterrent to crime.

Capital punishment gives the criminal what he deserves.

References

Andrich, D. (1988). The Application of an Unfolding Model of the PIRT Type to the Measurement of Attitude. Applied Psychological Measurement, 12, 33-51.

Author

Phil Chalmers rphilip.chalmers@gmail.com

Examples

head(Attitude)
#>   hideous state_teaching deterrent believe_not.necessary necessary_wish.not
#> 1       0              1         1                     0                  0
#> 2       1              1         1                     0                  0
#> 3       0              1         1                     1                  0
#> 4       1              1         1                     1                  0
#> 5       0              1         1                     1                  1
#> 6       0              1         1                     1                  0
#>   must.have justified deserved freq
#> 1         0         0        0    4
#> 2         0         0        0   10
#> 3         0         0        0    3
#> 4         0         0        0    8
#> 5         0         0        0    1
#> 6         0         1        0    1
df <- expand.table(Attitude)
itemstats(df)
#> $overall
#>   N mean_total.score sd_total.score  ave.r  sd.r  alpha SEM.alpha
#>  54            3.852          1.053 -0.064 0.524 -0.849     1.433
#> 
#> $itemstats
#>                        N  mean    sd total.r total.r_if_rm alpha_if_rm
#> hideous               54 0.444 0.502   0.056        -0.388      -0.340
#> state_teaching        54 0.648 0.482  -0.290        -0.616      -0.046
#> deterrent             54 0.667 0.476  -0.251        -0.587      -0.088
#> believe_not.necessary 54 0.463 0.503   0.523         0.053      -1.260
#> necessary_wish.not    54 0.481 0.504   0.669         0.249      -1.836
#> must.have             54 0.444 0.502   0.591         0.141      -1.499
#> justified             54 0.352 0.482   0.402        -0.061      -0.959
#> deserved              54 0.352 0.482   0.402        -0.061      -0.959
#> 
#> $proportions
#>                           0     1
#> hideous               0.556 0.444
#> state_teaching        0.352 0.648
#> deterrent             0.333 0.667
#> believe_not.necessary 0.537 0.463
#> necessary_wish.not    0.519 0.481
#> must.have             0.556 0.444
#> justified             0.648 0.352
#> deserved              0.648 0.352
#> 

if (FALSE) { # \dontrun{

# estimate SSLM with estimated " latitude of acceptance" (rho)
mod.rho <- mirt(df, 1, itemtype = 'sslm')
coef(mod.rho)
coef(mod.rho, simplify=TRUE)  # slope-intercept-log_rho
coef(mod.rho, simplify=TRUE, IRTpars=TRUE)  # discrimination-difficulty-rho
plot(mod.rho)
plot(mod.rho, type = 'trace')

# without estimating rho, and fixing to rho^2 = 1  (hence,
#   log_rho = -exp(1) = -2.718282 in order to obtain (exp(exp(log_rho))) = 1)
syntax <- "Theta = 1-8
           FIXED = (1-8, log_rho1)
           START = (1-8, log_rho1, -2.71828)"
mod <- mirt(df, syntax, itemtype = 'sslm')  # model found in Andrich (1988)
coef(mod)
coef(mod, simplify=TRUE)  # slope-intercept-log_rho
coef(mod, simplify=TRUE, IRTpars=TRUE)  # discrimination-difficulty-rho
plot(mod)
plot(mod, type = 'trace') # notice that all curves have a fixed height of .5

# goodness of fit (less constrained model fits better)
anova(mod, mod.rho) # original model fits much worse
M2(mod)
M2(mod.rho)
itemfit(mod, p.adjust='fdr')
itemfit(mod.rho, p.adjust='fdr')

} # }