MDISC {mirt}	R Documentation

Compute multidimensional discrimination index

Description

Returns a vector containing the MDISC values for each item in the model input object (Reckase, 2009).

Usage

MDISC(x, group = NULL)

Arguments

x
an object of class 'SingleGroupClass', or an object of class 'MultipleGroupClass' if a suitable group input were supplied

group
group argument to pass to extract.group function. Required when the input object is a multiple-group model

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi:10.18637/jss.v048.i06

Reckase, M. D. (2009). Multidimensional Item Response Theory. Springer.

See Also

extract.group

Examples

## No test: 

mod <- mirt(Science, 2)
MDISC(mod)

##  Comfort     Work   Future  Benefit 
## 1.338530 2.050473 1.875270 1.722043

## End(No test)

[Package mirt version 1.40 Index]