MultipleGroupClass-class {mirt}	R Documentation

Class "MultipleGroupClass"

Description

Defines the object returned from multipleGroup.

Slots

Call:

function call

Data:

list of data, sometimes in different forms

Options:

list of estimation options

Fit:

a list of fit information

Model:

a list of model-based information

ParObjects:

a list of the S4 objects used during estimation

OptimInfo:

a list of arguments from the optimization process

Internals:

a list of internal arguments for secondary computations (inspecting this object is generally not required)

vcov:

a matrix represented the asymtotic covariance matrix of the parameter estimates

time:

a data.frame indicating the breakdown of computation times in seconds

Methods

coef

signature(object = "MultipleGroupClass")

print

signature(x = "MultipleGroupClass")

show

signature(object = "MultipleGroupClass")

anova

signature(object = "MultipleGroupClass")

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

[Package mirt version 1.16 Index]