SIBTEST {mirt}	R Documentation

Simultaneous Item Bias Test (SIBTEST)

Description

Classical test theory approach to detecting DIF for unidimensional tests by applying a regression-corrected matched-total score approach. SIBTEST is similar to the Mantel-Haenszel approach for detecting DIF but uses a regression correction based on the KR-20/coefficient alpha reliability index to correct the observed differences when the latent trait distributions are not equal. Function supports the standard SIBTEST for dichotomous and poltomous data (compensatory) and also supports crossed DIF testing (i.e., non-compensatory) for dichotomous data.

Usage

SIBTEST(dat, group, focal_set, match_set, focal_name = "focal",
  guess_correction = 0, Jmin = 2, cross = FALSE, permute = 1000,
  pk_focal = FALSE, correction = TRUE)

Arguments

dat
integer dataset to be tested containing dichotomous or polytomous responses

group
a vector indicating group membership

focal_set
an integer vector indicating which items to inspect with SIBTEST. Including only one value will perform a DIF test, while including more than one will perform a simultaneous bundle test (DBF); including all non-matched items will perform DTF. If missing, a simultaneous test using all the items not listed in match_set will be used (i.e., DTF)

match_set
an integer vector indicating which items to use as the items which are matched (i.e., contain no DIF). These are analogous to 'achor' items in the likelihood method to locate DIF. If missing, all items other than the items found in the focal_set will be used

focal_name
name of the focal group. E.g., 'focal'

guess_correction
a vector of numbers from 0 to 1 indicating how much to correct the items for guessing. It's length should be the same as ncol(dat)

Jmin
the minimum number of observations required when splitting the data into focal and reference groups conditioned on the matched set

cross
logical; perform the crossing test for DIF? Can only be used when the data consist of dichotomous responses. Default is FALSE

permute
number of permutations to perform when cross = TRUE. Default is 1000

pk_focal
logical; using the group weights from the focal group instead of the total sample? Default is FALSE as per Shealy and Stout's recommendation

correction
logical; apply the composite correction for the difference between focal composite scores using the true-score regression technique? Default is TRUE, reflecting Shealy and Stout's method

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

References

Chang, H. H., Mazzeo, J. & Roussos, L. (1996). DIF for Polytomously Scored Items: An Adaptation of the SIBTEST Procedure. Journal of Educational Measurement, 33, 333-353.

Li, H.-H. & Stout, W. (1996). A new procedure for detetion of crossing DIF. Psychometrika, 61, 647-677.

Shealy, R. & Stout, W. (1993). A model-based standardization approach that separates true bias/DIF from group ability differences and ddetect test bias/DTF as well as item bias/DIF. Psychometrika, 58, 159-194.

Examples

## No test: 

library(mirt)

set.seed(1234)
n <- 30
N <- 500
a <- matrix(1, n)
d <- matrix(rnorm(n), n)
group <- c(rep('reference', N), rep('focal', N*2))

## -------------
# groups completely equal
dat1 <- simdata(a, d, N, itemtype = 'dich')
dat2 <- simdata(a, d, N*2, itemtype = 'dich')
dat <- rbind(dat1, dat2)

#DIF (all other items as anchors)
SIBTEST(dat, group, focal_set = 6)

##         n_matched_set n_focal_set           B         z         p
## SIBTEST            29           1 -0.03695718 -1.358991 0.1741496

#DIF (specific anchors)
SIBTEST(dat, group, match_set = 1:5, focal_set = 6)

##         n_matched_set n_focal_set           B         z         p
## SIBTEST             5           1 -0.03277425 -1.176036 0.2395803

# DBF (all and specific anchors, respectively)
SIBTEST(dat, group, focal_set = 11:30)

##         n_matched_set n_focal_set        B         z         p
## SIBTEST            10          20 0.101876 0.6088667 0.5426128

SIBTEST(dat, group, match_set = 1:5, focal_set = 11:30)

##         n_matched_set n_focal_set          B          z         p
## SIBTEST             5          20 0.01004398 0.04987587 0.9602213

#DTF
SIBTEST(dat, group, focal_set = 11:30)

##         n_matched_set n_focal_set        B         z         p
## SIBTEST            10          20 0.101876 0.6088667 0.5426128

SIBTEST(dat, group, match_set = 1:10) #equivalent

##         n_matched_set n_focal_set        B         z         p
## SIBTEST            10          20 0.101876 0.6088667 0.5426128

# different hyper pars
dat1 <- simdata(a, d, N, itemtype = 'dich')
dat2 <- simdata(a, d, N*2, itemtype = 'dich', mu = .5, sigma = matrix(1.5))
dat <- rbind(dat1, dat2)
SIBTEST(dat, group, 6:30)

##         n_matched_set n_focal_set          B        z        p
## SIBTEST             5          25 -0.3302013 -1.29682 0.194693

SIBTEST(dat, group, 11:30)

##         n_matched_set n_focal_set          B          z         p
## SIBTEST            10          20 -0.1351675 -0.7223809 0.4700604

#DIF testing with anchors 1 through 5
SIBTEST(dat, group, 6, match_set = 1:5)

##         n_matched_set n_focal_set           B         z          p
## SIBTEST             5           1 -0.04912936 -1.804087 0.07121769

SIBTEST(dat, group, 7, match_set = 1:5)

##         n_matched_set n_focal_set           B          z         p
## SIBTEST             5           1 -0.00737568 -0.2617505 0.7935138

SIBTEST(dat, group, 8, match_set = 1:5)

##         n_matched_set n_focal_set           B         z         p
## SIBTEST             5           1 0.004200776 0.1501167 0.8806726

#DIF testing with all other items as anchors
SIBTEST(dat, group, 6)

##         n_matched_set n_focal_set           B         z          p
## SIBTEST            29           1 -0.04631279 -1.661794 0.09655412

SIBTEST(dat, group, 7)

##         n_matched_set n_focal_set          B         z         p
## SIBTEST            29           1 0.01081096 0.3667616 0.7137968

SIBTEST(dat, group, 8)

##         n_matched_set n_focal_set           B          z        p
## SIBTEST            29           1 0.001678146 0.05794552 0.953792

#crossed SIBTEST
SIBTEST(dat, group, 6, match_set = 1:5, cross=TRUE)

##                 n_matched_set n_focal_set          B  z    p
## Crossed_SIBTEST             5           1 0.04912936 NA 0.54

SIBTEST(dat, group, 7, match_set = 1:5, cross=TRUE)

##                 n_matched_set n_focal_set          B  z     p
## Crossed_SIBTEST             5           1 0.04750447 NA 0.272

SIBTEST(dat, group, 8, match_set = 1:5, cross=TRUE)

##                 n_matched_set n_focal_set            B  z p
## Crossed_SIBTEST             5           1 -0.003186207 NA 1

## -------------
## systematic differing slopes and intercepts (clear DTF)
dat1 <- simdata(a, d, N, itemtype = 'dich')
dat2 <- simdata(a + c(numeric(15), rnorm(n-15, 1, .25)), d + c(numeric(15), rnorm(n-15, 1, 1)),
  N*2, itemtype = 'dich')
dat <- rbind(dat1, dat2)
SIBTEST(dat, group, 6:30)

##         n_matched_set n_focal_set         B         z p
## SIBTEST             5          25 -2.329058 -8.536827 0

SIBTEST(dat, group, 11:30)

##         n_matched_set n_focal_set         B        z p
## SIBTEST            10          20 -2.755746 -15.4522 0

#DIF testing using valid anchors
SIBTEST(dat, group, focal_set = 6, match_set = 1:5)

##         n_matched_set n_focal_set          B        z           p
## SIBTEST             5           1 0.07151388 2.590642 0.009579702

SIBTEST(dat, group, focal_set = 7, match_set = 1:5)

##         n_matched_set n_focal_set          B        z         p
## SIBTEST             5           1 0.04192564 1.473055 0.1407363

SIBTEST(dat, group, focal_set = 30, match_set = 1:5)

##         n_matched_set n_focal_set          B         z            p
## SIBTEST             5           1 -0.1061777 -3.981502 6.848102e-05

SIBTEST(dat, group, focal_set = 11, match_set = 1:10, cross=TRUE)

##                 n_matched_set n_focal_set           B  z p
## Crossed_SIBTEST            10           1 1.88186e-05 NA 1

SIBTEST(dat, group, focal_set = 30, match_set = 1:15, cross=TRUE)

##                 n_matched_set n_focal_set         B  z p
## Crossed_SIBTEST            15           1 0.1502871 NA 0

## End(No test)

[Package mirt version 1.16 Index]