createItem {mirt}	R Documentation

Create a user defined item with correct generic functions

Description

Initializes the proper S4 class and methods necessary for mirt functions to use in estimation. To use the defined objects pass to the mirt(..., customItems = list()) command, and ensure that the classes are properly labeled and unique in the list.

Usage

createItem(name, par, est, P, gr = NULL, hss = NULL, lbound = NULL,
  ubound = NULL)

Arguments

name
a character indicating the item class name to be defined

par
a named vector of the starting values for the parameters

est
a logical vector indicating which parameters should be freely estimated by default

P
the probability trace function for all categories (first column is category 1, second category two, etc). First input contains a vector of all the item parameters, the second input must be a matrix called Theta, and the third input must be the number of categories called ncat. Function also must return a matrix object of category probabilities

gr
gradient function (vector of first derivatives) of the log-likelihood used in estimation. The function must be of the form gr(x, Theta), where x is the object defined by createItem() and Theta is a matrix of latent trait parameters. Tabulated (EM) or raw (MHRM) data are located in the x@dat slot, and are used to form the complete data log-likelihood. If not specified a numeric approximation will be used

hss
Hessian function (matrix of second derivatives) of the log-likelihood used in estimation. If not specified a numeric approximation will be used (required for the MH-RM algorithm only). The input is idential to the gr argument

lbound
optional vector indicating the lower bounds of the parameters. If not specified then the bounds will be set to -Inf

ubound
optional vector indicating the lower bounds of the parameters. If not specified then the bounds will be set to Inf

Details

The summary() function will not return proper standardized loadings since the function is not sure how to handle them (no slopes could be defined at all!). Instead loadings of .001 are filled in as place-holders.

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

Examples

## No test: 

name <- 'old2PL'
par <- c(a = .5, b = -2)
est <- c(TRUE, TRUE)
P.old2PL <- function(par,Theta, ncat){
     a <- par[1]
     b <- par[2]
     P1 <- 1 / (1 + exp(-1*a*(Theta - b)))
     cbind(1-P1, P1)
}

x <- createItem(name, par=par, est=est, P=P.old2PL)

#So, let's estimate it!
dat <- expand.table(LSAT7)
sv <- mirt(dat, 1, c(rep('2PL',4), 'old2PL'), customItems=list(old2PL=x), pars = 'values')
tail(sv) #looks good

##    group   item     class   name parnum value lbound ubound   est
## 15   all Item.4      dich      g     15   0.0  0e+00      1 FALSE
## 16   all Item.4      dich      u     16   1.0  0e+00      1 FALSE
## 17   all Item.5    custom      a     17   0.5   -Inf    Inf  TRUE
## 18   all Item.5    custom      b     18  -2.0   -Inf    Inf  TRUE
## 19   all  GROUP GroupPars MEAN_1     19   0.0   -Inf    Inf FALSE
## 20   all  GROUP GroupPars COV_11     20   1.0  1e-04    Inf FALSE
##    prior.type prior_1 prior_2
## 15       none     NaN     NaN
## 16       none     NaN     NaN
## 17       none     NaN     NaN
## 18       none     NaN     NaN
## 19       none     NaN     NaN
## 20       none     NaN     NaN

mod <- mirt(dat, 1, c(rep('2PL',4), 'old2PL'), customItems=list(old2PL=x))
coef(mod)

## $Item.1
##        a1     d g u
## par 0.989 1.856 0 1
## 
## $Item.2
##        a1     d g u
## par 1.081 0.808 0 1
## 
## $Item.3
##        a1     d g u
## par 1.703 1.803 0 1
## 
## $Item.4
##        a1     d g u
## par 0.766 0.486 0 1
## 
## $Item.5
##         a      b
## par 0.737 -2.518
## 
## $GroupPars
##     MEAN_1 COV_11
## par      0      1

mod2 <- mirt(dat, 1, c(rep('2PL',4), 'old2PL'), customItems=list(old2PL=x), method = 'MHRM')
coef(mod2)

## $Item.1
##        a1     d g u
## par 0.985 1.851 0 1
## 
## $Item.2
##        a1    d g u
## par 1.104 0.81 0 1
## 
## $Item.3
##        a1     d g u
## par 1.625 1.755 0 1
## 
## $Item.4
##        a1     d g u
## par 0.767 0.484 0 1
## 
## $Item.5
##         a     b
## par 0.768 -2.43
## 
## $GroupPars
##     MEAN_1 COV_11
## par      0      1

###non-linear
name <- 'nonlin'
par <- c(a1 = .5, a2 = .1, d = 0)
est <- c(TRUE, TRUE, TRUE)
P.nonlin <- function(par,Theta, ncat=2){
     a1 <- par[1]
     a2 <- par[2]
     d <- par[3]
     P1 <- 1 / (1 + exp(-1*(a1*Theta + a2*Theta^2 + d)))
     cbind(1-P1, P1)
}

x2 <- createItem(name, par=par, est=est, P=P.nonlin)

mod <- mirt(dat, 1, c(rep('2PL',4), 'nonlin'), customItems=list(nonlin=x2))
coef(mod)

## $Item.1
##        a1     d g u
## par 0.984 1.854 0 1
## 
## $Item.2
##        a1     d g u
## par 1.087 0.809 0 1
## 
## $Item.3
##        a1     d g u
## par 1.705 1.803 0 1
## 
## $Item.4
##        a1     d g u
## par 0.762 0.486 0 1
## 
## $Item.5
##        a1    a2     d
## par 0.806 0.065 1.818
## 
## $GroupPars
##     MEAN_1 COV_11
## par      0      1

###nominal response model (Bock 1972 version)
Tnom.dev <- function(ncat) {
   T <- matrix(1/ncat, ncat, ncat - 1)
   diag(T[-1, ]) <-  diag(T[-1, ]) - 1
   return(T)
}

name <- 'nom'
par <- c(alp=c(3,0,-3),gam=rep(.4,3))
est <- rep(TRUE, length(par))
P.nom <- function(par, Theta, ncat){
   alp <- par[1:(ncat-1)]
   gam <- par[ncat:length(par)]
   a <- Tnom.dev(ncat) %*% alp
   c <- Tnom.dev(ncat) %*% gam
   z <- matrix(0, nrow(Theta), ncat)
   for(i in 1:ncat)
       z[,i] <- a[i] * Theta + c[i]
   P <- exp(z) / rowSums(exp(z))
   P
}

nom1 <- createItem(name, par=par, est=est, P=P.nom)
nommod <- mirt(Science, 1, 'nom1', customItems=list(nom1=nom1))
coef(nommod)

## $Comfort
##       alp1   alp2   alp3   gam1   gam2   gam3
## par -1.553 -2.016 -3.026 -3.641 -5.907 -4.534
## 
## $Work
##       alp1   alp2   alp3   gam1   gam2   gam3
## par -0.579 -1.261 -2.522 -1.463 -2.326 -0.325
## 
## $Future
##       alp1  alp2   alp3   gam1   gam2   gam3
## par -1.547 -3.78 -6.096 -3.659 -5.852 -3.939
## 
## $Benefit
##       alp1   alp2  alp3   gam1   gam2   gam3
## par -0.809 -1.359 -2.34 -2.146 -2.913 -1.623
## 
## $GroupPars
##     MEAN_1 COV_11
## par      0      1

Tnom.dev(4) %*% coef(nommod)[[1]][1:3] #a

##          [,1]
## [1,] -1.64875
## [2,] -0.09575
## [3,]  0.36725
## [4,]  1.37725

Tnom.dev(4) %*% coef(nommod)[[1]][4:6] #d

##         [,1]
## [1,] -3.5205
## [2,]  0.1205
## [3,]  2.3865
## [4,]  1.0135

## End(No test)

[Package mirt version 1.14 Index]