Draw plausible parameter instantiations from a given model

Draws plausible parameters from a model using parametric sampling (if the information matrix was computed) or via boostrap sampling. Primarily for use with the DRF function.

draw_parameters(
  mod,
  draws,
  method = c("parametric", "boostrap"),
  redraws = 20,
  ...
)

Arguments

mod
estimated single or multiple-group model

draws
number of draws to obtain

method
type of plausible values to obtain. Can be 'parametric', for the parametric sampling scheme which uses the estimated information matrix, or 'boostrap' to obtain values from the boot function. Default is 'parametric'

redraws
number of redraws to perform when the given parameteric sample does not satisfy the upper and lower parameter bounds. If a valid set cannot be found within this number of draws then an error will be thrown

...
additional arguments to be passed

Value

returns a draws x p matrix of plausible parameters, where each row correspeonds to a single set

Examples

# \donttest{
set.seed(1234)
n <- 40
N <- 500

# only first 5 items as anchors
model <- 'F = 1-40
          CONSTRAINB = (1-5, a1), (1-5, d)'

a <- matrix(1, n)
d <- matrix(rnorm(n), n)
group <- c(rep('Group_1', N), rep('Group_2', N))

## -------------
# groups completely equal
dat1 <- simdata(a, d, N, itemtype = 'dich')
dat2 <- simdata(a, d, N, itemtype = 'dich')
dat <- rbind(dat1, dat2)
mod <- multipleGroup(dat, model, group=group, SE=TRUE,
                     invariance=c('free_means', 'free_var'))

param_set <- draw_parameters(mod, 100)
head(param_set)

#>           [,1]      [,2] [,3] [,4]      [,5]      [,6] [,7] [,8]      [,9]
#> [1,] 0.9706048 -1.098878    0    1 0.9727960 0.5352826    0    1 0.9975924
#> [2,] 1.2034241 -1.195367    0    1 1.1838115 0.3778588    0    1 1.2080660
#> [3,] 1.0471443 -1.054472    0    1 1.0009382 0.4542811    0    1 0.9138505
#> [4,] 0.9432668 -1.063708    0    1 1.0044127 0.3206982    0    1 1.1236906
#> [5,] 1.1482297 -1.373231    0    1 0.8825741 0.3002710    0    1 1.1290716
#> [6,] 1.2537973 -1.203807    0    1 1.1366696 0.2797301    0    1 1.3226339
#>          [,10] [,11] [,12]    [,13]     [,14] [,15] [,16]     [,17]     [,18]
#> [1,] 1.2162798     0     1 1.220515 -2.349255     0     1 0.7764887 0.5415735
#> [2,] 0.9403822     0     1 1.297873 -2.477683     0     1 0.8743893 0.5011681
#> [3,] 1.0843719     0     1 1.426406 -2.361161     0     1 0.8359967 0.5721702
#> [4,] 1.1503707     0     1 1.150072 -2.257685     0     1 0.8397915 0.4518688
#> [5,] 1.1047093     0     1 1.264975 -2.388279     0     1 0.9085445 0.3569120
#> [6,] 1.2730987     0     1 1.455955 -2.407915     0     1 0.8075839 0.4227599
#>      [,19] [,20]     [,21]     [,22] [,23] [,24]     [,25]      [,26] [,27]
#> [1,]     0     1 1.2924688 0.4386223     0     1 1.2447620 -0.5611534     0
#> [2,]     0     1 1.4037454 0.4075320     0     1 1.2207934 -0.5351847     0
#> [3,]     0     1 1.3427371 0.4001179     0     1 1.1119137 -0.4987708     0
#> [4,]     0     1 1.1078633 0.3475802     0     1 1.0244886 -0.4207340     0
#> [5,]     0     1 0.8790791 0.4401321     0     1 0.9610584 -0.5051074     0
#> [6,]     0     1 1.0709512 0.5038575     0     1 1.0859846 -0.3060513     0
#>      [,28]     [,29]      [,30] [,31] [,32]     [,33]      [,34] [,35] [,36]
#> [1,]     1 1.0790518 -0.4291185     0     1 0.8648857 -0.4493249     0     1
#> [2,]     1 1.0562855 -0.5598180     0     1 0.6718351 -0.4956853     0     1
#> [3,]     1 0.9770038 -0.5621551     0     1 0.8960956 -0.2969312     0     1
#> [4,]     1 1.1258879 -0.7201494     0     1 0.7213739 -0.5868639     0     1
#> [5,]     1 1.0577001 -0.7344510     0     1 0.9888427 -0.4142642     0     1
#> [6,]     1 1.0870495 -0.6148905     0     1 0.9704698 -0.4526229     0     1
#>          [,37]      [,38] [,39] [,40]     [,41]      [,42] [,43] [,44]
#> [1,] 0.7600400 -0.7431602     0     1 0.9557773 -0.4969073     0     1
#> [2,] 1.0559383 -0.9283458     0     1 1.0378306 -0.3021953     0     1
#> [3,] 1.2221618 -0.9504793     0     1 1.1070464 -0.3176478     0     1
#> [4,] 1.0013162 -0.6090048     0     1 0.7911614 -0.1717180     0     1
#> [5,] 0.8962249 -0.9318161     0     1 1.0505351 -0.4829546     0     1
#> [6,] 0.9281949 -0.9230024     0     1 0.9395066 -0.6024924     0     1
#>          [,45]      [,46] [,47] [,48]     [,49]      [,50] [,51] [,52]
#> [1,] 1.2129670 -1.0511074     0     1 0.9957647 -0.6266565     0     1
#> [2,] 0.9274543 -0.9824084     0     1 1.2532438 -0.6915689     0     1
#> [3,] 1.2548161 -1.0174359     0     1 0.9948778 -0.7517325     0     1
#> [4,] 1.0258378 -1.1818701     0     1 0.9974017 -0.5050230     0     1
#> [5,] 0.9173356 -1.0198218     0     1 1.1713840 -1.0601061     0     1
#> [6,] 1.0961087 -0.9907984     0     1 1.0441425 -0.7965599     0     1
#>          [,53]     [,54] [,55] [,56]    [,57]    [,58] [,59] [,60]     [,61]
#> [1,] 0.6450821 0.1659600     0     1 1.209958 1.063678     0     1 0.9436980
#> [2,] 0.9454362 0.2219314     0     1 1.410315 1.109297     0     1 0.9244916
#> [3,] 0.7499241 0.1453393     0     1 1.289951 1.001839     0     1 0.6907016
#> [4,] 0.8125422 0.2678060     0     1 1.120443 1.138378     0     1 0.8493085
#> [5,] 0.9677509 0.1213460     0     1 1.507925 1.005989     0     1 1.0663289
#> [6,] 0.9895537 0.1259535     0     1 1.495404 1.043689     0     1 1.0219372
#>            [,62] [,63] [,64]     [,65]      [,66] [,67] [,68]    [,69]
#> [1,]  0.08044846     0     1 1.0765375 -0.4416947     0     1 1.000702
#> [2,] -0.08672655     0     1 0.7957369 -0.4288258     0     1 1.180028
#> [3,] -0.12560582     0     1 1.1883050 -0.3745693     0     1 1.141610
#> [4,] -0.11639886     0     1 1.0650982 -0.3271925     0     1 1.238449
#> [5,]  0.03649571     0     1 0.7665531 -0.4072869     0     1 0.980612
#> [6,] -0.19089492     0     1 1.0204720 -0.6024851     0     1 1.365000
#>           [,70] [,71] [,72]     [,73]      [,74] [,75] [,76]    [,77]    [,78]
#> [1,] -0.9347871     0     1 0.9187332 -0.5312207     0     1 1.101202 2.489329
#> [2,] -1.2240485     0     1 1.1399053 -0.6957816     0     1 1.385267 2.782407
#> [3,] -0.8800437     0     1 1.0555542 -0.4449645     0     1 1.941471 3.120022
#> [4,] -1.0487994     0     1 0.8592031 -0.7141686     0     1 1.232000 3.205202
#> [5,] -1.0533224     0     1 1.2325744 -0.9093296     0     1 1.404092 2.959562
#> [6,] -0.9361257     0     1 0.8204183 -0.6744054     0     1 1.213615 2.454721
#>      [,79] [,80]     [,81]       [,82] [,83] [,84]    [,85]      [,86] [,87]
#> [1,]     0     1 0.8046958  0.12832353     0     1 1.300801 -0.4042582     0
#> [2,]     0     1 1.1982318  0.03307734     0     1 1.425367 -0.7236752     0
#> [3,]     0     1 0.9683640  0.18068410     0     1 1.301984 -0.6592910     0
#> [4,]     0     1 0.9669823  0.15875756     0     1 1.316646 -0.7100784     0
#> [5,]     0     1 0.8649566 -0.11421439     0     1 1.134909 -0.5920451     0
#> [6,]     0     1 0.9230991 -0.02066030     0     1 1.213411 -0.5062486     0
#>      [,88]     [,89]      [,90] [,91] [,92]     [,93]     [,94] [,95] [,96]
#> [1,]     1 0.8826901 -0.4524264     0     1 0.9275022 0.6287660     0     1
#> [2,]     1 0.7377546 -0.6080442     0     1 1.0673266 0.5808530     0     1
#> [3,]     1 0.9466481 -0.4511876     0     1 1.2137328 0.6268076     0     1
#> [4,]     1 0.6398234 -0.3143388     0     1 0.9251352 0.4975599     0     1
#> [5,]     1 0.6819819 -0.4377236     0     1 0.9448951 0.4707067     0     1
#> [6,]     1 0.9563707 -0.4195523     0     1 0.9462990 0.4669390     0     1
#>          [,97]      [,98] [,99] [,100]   [,101]    [,102] [,103] [,104]
#> [1,] 1.0422506 -0.7140121     0      1 1.423986 -1.405418      0      1
#> [2,] 1.1395638 -0.4931220     0      1 1.523760 -1.631331      0      1
#> [3,] 1.1985876 -0.3970168     0      1 1.372366 -1.322367      0      1
#> [4,] 0.8198764 -0.6825928     0      1 1.396905 -1.500892      0      1
#> [5,] 0.7564137 -0.5884395     0      1 1.177499 -1.642744      0      1
#> [6,] 1.0927339 -0.7074842     0      1 1.634473 -1.327893      0      1
#>        [,105]    [,106] [,107] [,108]    [,109]     [,110] [,111] [,112]
#> [1,] 1.214749 0.8786176      0      1 0.8184427 -0.8921982      0      1
#> [2,] 1.423700 0.6305353      0      1 1.2914217 -0.7786995      0      1
#> [3,] 1.651226 0.8892790      0      1 1.2126947 -1.1035592      0      1
#> [4,] 1.299977 0.7738131      0      1 0.9150064 -0.9336558      0      1
#> [5,] 1.182235 0.5468914      0      1 1.0862362 -0.7795187      0      1
#> [6,] 1.245990 0.5440097      0      1 1.2414273 -0.9310122      0      1
#>         [,113]      [,114] [,115] [,116]    [,117]     [,118] [,119] [,120]
#> [1,] 0.9719441 -0.06413765      0      1 1.0657304 -0.8227668      0      1
#> [2,] 1.3237706 -0.16181588      0      1 0.9502281 -0.8350891      0      1
#> [3,] 1.2197267 -0.03988397      0      1 1.0942635 -0.8057415      0      1
#> [4,] 1.1109770  0.09783249      0      1 1.1343204 -0.9935824      0      1
#> [5,] 0.8890796  0.02626983      0      1 0.8242011 -0.8462716      0      1
#> [6,] 1.1361609 -0.19957941      0      1 1.0864674 -0.9150995      0      1
#>         [,121]   [,122] [,123] [,124]    [,125]     [,126] [,127] [,128]
#> [1,] 1.0972335 1.393080      0      1 1.0361196 -0.3948651      0      1
#> [2,] 1.0693135 1.135582      0      1 0.9589949 -0.5644483      0      1
#> [3,] 0.9702945 1.165392      0      1 0.9983275 -0.4308975      0      1
#> [4,] 0.8462496 1.307773      0      1 0.8108319 -0.2390250      0      1
#> [5,] 1.0807468 1.403967      0      1 0.8284389 -0.2524584      0      1
#> [6,] 0.9497868 1.198771      0      1 0.9381207 -0.3553973      0      1
#>         [,129]     [,130] [,131] [,132]    [,133]     [,134] [,135] [,136]
#> [1,] 1.1308819 -0.5775562      0      1 1.0045512 -0.4125562      0      1
#> [2,] 0.9516824 -0.5972809      0      1 0.8707829 -0.5708825      0      1
#> [3,] 1.0780658 -0.5535222      0      1 1.0764039 -0.5821323      0      1
#> [4,] 0.9581735 -0.5939089      0      1 1.0607243 -0.5254642      0      1
#> [5,] 0.9244505 -0.4994486      0      1 0.7160516 -0.5743700      0      1
#> [6,] 0.9770499 -0.6483429      0      1 1.0220090 -0.8272470      0      1
#>         [,137]    [,138] [,139] [,140]    [,141]     [,142] [,143] [,144]
#> [1,] 1.0677212 -1.517162      0      1 0.9116918 -1.0025017      0      1
#> [2,] 0.9054174 -1.279623      0      1 1.0041872 -0.9310365      0      1
#> [3,] 1.2718682 -1.593387      0      1 1.0790987 -1.0509264      0      1
#> [4,] 0.9612130 -1.818600      0      1 0.5605145 -0.8319758      0      1
#> [5,] 0.9356005 -1.470430      0      1 0.9135123 -0.9788414      0      1
#> [6,] 1.5335504 -1.850800      0      1 0.9565746 -0.9303182      0      1
#>         [,145]    [,146] [,147] [,148]    [,149]    [,150] [,151] [,152]
#> [1,] 1.1257480 -2.213538      0      1 1.1227891 -1.446339      0      1
#> [2,] 1.0675163 -2.393822      0      1 1.2075418 -1.667363      0      1
#> [3,] 0.9572845 -2.102742      0      1 0.9011092 -1.052149      0      1
#> [4,] 1.0060204 -2.508684      0      1 1.2785127 -1.548192      0      1
#> [5,] 0.7382971 -2.229296      0      1 0.9666307 -1.337177      0      1
#> [6,] 1.0320585 -2.348507      0      1 1.1075151 -1.493520      0      1
#>         [,153]     [,154] [,155] [,156]    [,157]     [,158] [,159] [,160]
#> [1,] 0.8400795 -0.1194495      0      1 1.2650736 -0.2332116      0      1
#> [2,] 1.1229336 -0.2637633      0      1 1.0471831 -0.2192196      0      1
#> [3,] 0.9479736 -0.2225754      0      1 0.9788813 -0.5875727      0      1
#> [4,] 0.8927255 -0.2097865      0      1 1.1063023 -0.4153062      0      1
#> [5,] 0.7816276 -0.1653238      0      1 0.8622267 -0.6058538      0      1
#> [6,] 0.9366711 -0.2552546      0      1 1.1834648 -0.6288336      0      1
#>      [,161] [,162]    [,163]    [,164] [,165] [,166]    [,167]    [,168] [,169]
#> [1,]      0      1 0.9706048 -1.098878      0      1 0.9727960 0.5352826      0
#> [2,]      0      1 1.2034241 -1.195367      0      1 1.1838115 0.3778588      0
#> [3,]      0      1 1.0471443 -1.054472      0      1 1.0009382 0.4542811      0
#> [4,]      0      1 0.9432668 -1.063708      0      1 1.0044127 0.3206982      0
#> [5,]      0      1 1.1482297 -1.373231      0      1 0.8825741 0.3002710      0
#> [6,]      0      1 1.2537973 -1.203807      0      1 1.1366696 0.2797301      0
#>      [,170]    [,171]    [,172] [,173] [,174]   [,175]    [,176] [,177] [,178]
#> [1,]      1 0.9975924 1.2162798      0      1 1.220515 -2.349255      0      1
#> [2,]      1 1.2080660 0.9403822      0      1 1.297873 -2.477683      0      1
#> [3,]      1 0.9138505 1.0843719      0      1 1.426406 -2.361161      0      1
#> [4,]      1 1.1236906 1.1503707      0      1 1.150072 -2.257685      0      1
#> [5,]      1 1.1290716 1.1047093      0      1 1.264975 -2.388279      0      1
#> [6,]      1 1.3226339 1.2730987      0      1 1.455955 -2.407915      0      1
#>         [,179]    [,180] [,181] [,182]    [,183]    [,184] [,185] [,186]
#> [1,] 0.7764887 0.5415735      0      1 0.8892622 0.6465678      0      1
#> [2,] 0.8743893 0.5011681      0      1 0.7066536 0.2760210      0      1
#> [3,] 0.8359967 0.5721702      0      1 0.6160197 0.4073347      0      1
#> [4,] 0.8397915 0.4518688      0      1 0.6456903 0.3409698      0      1
#> [5,] 0.9085445 0.3569120      0      1 0.8527239 0.2573574      0      1
#> [6,] 0.8075839 0.4227599      0      1 0.9304753 0.3858307      0      1
#>         [,187]     [,188] [,189] [,190]    [,191]     [,192] [,193] [,194]
#> [1,] 0.9727143 -0.4337066      0      1 0.7812762 -0.2888699      0      1
#> [2,] 1.2229728 -0.7865907      0      1 1.0345929 -0.5646957      0      1
#> [3,] 1.1516654 -0.4008829      0      1 0.8710344 -0.3345320      0      1
#> [4,] 0.8348929 -0.7144063      0      1 0.8116943 -0.3838335      0      1
#> [5,] 0.8321565 -0.4626858      0      1 1.0246498 -0.4709226      0      1
#> [6,] 1.0948861 -0.6877331      0      1 0.7846392 -0.6877435      0      1
#>         [,195]     [,196] [,197] [,198]    [,199]     [,200] [,201] [,202]
#> [1,] 0.6572298 -0.4538809      0      1 0.8185024 -0.4998381      0      1
#> [2,] 0.8924683 -0.5651475      0      1 0.8628747 -0.7602620      0      1
#> [3,] 0.8640987 -0.4810945      0      1 0.9031096 -0.6908476      0      1
#> [4,] 0.7341494 -0.4938451      0      1 1.0702813 -0.7012946      0      1
#> [5,] 0.6771189 -0.6815690      0      1 0.7710429 -0.9668419      0      1
#> [6,] 0.9844191 -0.5934458      0      1 1.1247363 -0.8968057      0      1
#>         [,203]     [,204] [,205] [,206]    [,207]     [,208] [,209] [,210]
#> [1,] 0.9658046 -0.6057154      0      1 1.1733796 -0.5062916      0      1
#> [2,] 1.1252629 -0.5846615      0      1 1.3686317 -1.2317292      0      1
#> [3,] 1.1674640 -0.3292052      0      1 1.3168926 -0.9205226      0      1
#> [4,] 0.9781098 -0.5227718      0      1 1.2839629 -1.0471953      0      1
#> [5,] 0.9717642 -0.4432126      0      1 0.9012885 -0.9755325      0      1
#> [6,] 1.0422950 -0.6493477      0      1 1.4595473 -1.1484454      0      1
#>         [,211]     [,212] [,213] [,214]    [,215]     [,216] [,217] [,218]
#> [1,] 0.9212279 -0.5807230      0      1 0.9589077  0.6557833      0      1
#> [2,] 1.0403151 -1.0026626      0      1 0.9717297  0.2243922      0      1
#> [3,] 0.9254918 -0.6320219      0      1 0.9866448  0.4251984      0      1
#> [4,] 0.9948208 -0.9398925      0      1 1.1084900  0.2034137      0      1
#> [5,] 0.8969267 -0.6949632      0      1 0.6904504 -0.1525740      0      1
#> [6,] 1.1055952 -0.9358742      0      1 0.9582110  0.2517182      0      1
#>         [,219]    [,220] [,221] [,222]    [,223]      [,224] [,225] [,226]
#> [1,] 1.0320964 1.1164190      0      1 1.0023245 -0.08506484      0      1
#> [2,] 1.1084796 0.9410142      0      1 0.9961261 -0.39739515      0      1
#> [3,] 1.4542887 1.3553555      0      1 0.9412990  0.16100780      0      1
#> [4,] 0.8681597 1.1885255      0      1 1.0064765 -0.10890743      0      1
#> [5,] 0.9987949 1.0329216      0      1 0.8110843 -0.27269986      0      1
#> [6,] 1.2817554 1.1402921      0      1 1.0009243 -0.43636168      0      1
#>         [,227]     [,228] [,229] [,230]    [,231]     [,232] [,233] [,234]
#> [1,] 0.9998687 -0.1713356      0      1 0.8536589 -0.5320088      0      1
#> [2,] 1.1091674 -0.5411797      0      1 0.7517521 -0.8973061      0      1
#> [3,] 1.1221906 -0.1655266      0      1 0.9071202 -0.7165930      0      1
#> [4,] 1.0564329 -0.3298487      0      1 0.9015334 -0.9963294      0      1
#> [5,] 0.9582945 -0.4822184      0      1 0.8327169 -0.9452970      0      1
#> [6,] 1.4446876 -0.3465541      0      1 1.1828422 -0.7659421      0      1
#>         [,235]     [,236] [,237] [,238]    [,239]   [,240] [,241] [,242]
#> [1,] 0.6794971 -0.4987268      0      1 0.7568743 2.527114      0      1
#> [2,] 0.6771065 -0.6090472      0      1 0.8539503 2.551203      0      1
#> [3,] 0.9404983 -0.6519686      0      1 1.1916635 3.194427      0      1
#> [4,] 0.7902267 -0.7963981      0      1 1.4224879 2.809448      0      1
#> [5,] 0.8200226 -0.6996367      0      1 0.8573490 2.304936      0      1
#> [6,] 0.8670305 -0.6746071      0      1 1.3149767 2.361556      0      1
#>         [,243]      [,244] [,245] [,246]   [,247]        [,248] [,249] [,250]
#> [1,] 0.8561581  0.34027019      0      1 1.311794 -0.0052237950      0      1
#> [2,] 0.9361211 -0.01142372      0      1 1.173043 -0.4756803967      0      1
#> [3,] 0.7765916  0.23591911      0      1 1.237412  0.0002723278      0      1
#> [4,] 0.6785039  0.12843874      0      1 1.438880 -0.3376065037      0      1
#> [5,] 0.8558469  0.01432018      0      1 1.133198 -0.3429461996      0      1
#> [6,] 0.9324091 -0.17887487      0      1 1.313674 -0.3942895132      0      1
#>         [,251]     [,252] [,253] [,254]    [,255]    [,256] [,257] [,258]
#> [1,] 1.0261471 -0.2914226      0      1 0.9906072 1.0998463      0      1
#> [2,] 0.8498339 -0.6124563      0      1 1.0337755 0.6105145      0      1
#> [3,] 1.0001788 -0.3545236      0      1 0.7574189 0.7682795      0      1
#> [4,] 0.6735881 -0.4542847      0      1 1.1434754 0.9263200      0      1
#> [5,] 0.8170677 -0.4312472      0      1 0.7785616 0.6770115      0      1
#> [6,] 0.7770773 -0.4879678      0      1 1.0136366 0.6633859      0      1
#>         [,259]     [,260] [,261] [,262]    [,263]    [,264] [,265] [,266]
#> [1,] 1.0918225 -0.1812104      0      1 1.1183459 -1.117679      0      1
#> [2,] 0.9297291 -0.4264246      0      1 0.7440416 -1.460978      0      1
#> [3,] 0.9717042 -0.3140387      0      1 0.9313255 -1.275485      0      1
#> [4,] 1.0072748 -0.3911584      0      1 0.7952016 -1.192197      0      1
#> [5,] 0.9568830 -0.5953112      0      1 0.8720127 -1.639160      0      1
#> [6,] 1.2954108 -0.6106062      0      1 1.1755908 -1.359659      0      1
#>        [,267]    [,268] [,269] [,270]    [,271]     [,272] [,273] [,274]
#> [1,] 1.240181 0.8975621      0      1 0.7313241 -0.8019469      0      1
#> [2,] 1.188990 0.6104321      0      1 0.6508736 -0.9735487      0      1
#> [3,] 1.006342 0.5738313      0      1 0.9767276 -0.8190859      0      1
#> [4,] 1.111344 0.6693267      0      1 0.7370846 -0.8824282      0      1
#> [5,] 1.064623 0.3135511      0      1 0.9831481 -0.8929524      0      1
#> [6,] 1.437696 0.5088852      0      1 1.1537616 -1.0524459      0      1
#>         [,275]    [,276] [,277] [,278]    [,279]     [,280] [,281] [,282]
#> [1,] 1.1984849 0.4046287      0      1 0.8155366 -0.7747891      0      1
#> [2,] 1.1317054 0.2291912      0      1 0.9491661 -1.1324928      0      1
#> [3,] 1.1224339 0.2376811      0      1 1.1172256 -0.8126866      0      1
#> [4,] 1.1273508 0.2587197      0      1 1.1437877 -1.0920410      0      1
#> [5,] 0.9569644 0.1044591      0      1 0.9348731 -1.2465764      0      1
#> [6,] 1.1703140 0.2755453      0      1 1.0248876 -1.0449714      0      1
#>         [,283]    [,284] [,285] [,286]    [,287]     [,288] [,289] [,290]
#> [1,] 0.9925055 1.2184443      0      1 1.0288838 -0.3406296      0      1
#> [2,] 1.1914722 0.9809516      0      1 1.1411978 -0.5130387      0      1
#> [3,] 0.9767975 1.1022258      0      1 1.1375561 -0.3778279      0      1
#> [4,] 0.9623837 0.9706666      0      1 0.7950108 -0.4993759      0      1
#> [5,] 0.9638797 0.8767520      0      1 0.7634671 -0.6291256      0      1
#> [6,] 1.1964328 0.8926733      0      1 1.2589790 -0.3867625      0      1
#>         [,291]     [,292] [,293] [,294]    [,295]      [,296] [,297] [,298]
#> [1,] 1.0745801 -0.2511955      0      1 1.2258999 -0.04476311      0      1
#> [2,] 0.8554924 -0.9426309      0      1 0.9119326 -0.46512799      0      1
#> [3,] 0.9716621 -0.5119468      0      1 0.9188902 -0.32915226      0      1
#> [4,] 0.8554429 -0.6178045      0      1 1.1681639 -0.39203035      0      1
#> [5,] 0.7125117 -0.5518046      0      1 0.8453645 -0.51168269      0      1
#> [6,] 1.1615025 -0.7085532      0      1 1.0674229 -0.42715893      0      1
#>         [,299]    [,300] [,301] [,302]    [,303]    [,304] [,305] [,306]
#> [1,] 0.7059637 -1.603078      0      1 0.8383790 -0.906908      0      1
#> [2,] 0.9650198 -1.695621      0      1 0.9093199 -1.349227      0      1
#> [3,] 1.1385185 -1.498731      0      1 0.9150513 -1.050642      0      1
#> [4,] 0.8906113 -1.911941      0      1 0.8914043 -1.052137      0      1
#> [5,] 1.0320490 -2.002488      0      1 0.6229432 -1.222561      0      1
#> [6,] 0.9652706 -1.876189      0      1 1.0030988 -1.233448      0      1
#>         [,307]    [,308] [,309] [,310]    [,311]     [,312] [,313] [,314]
#> [1,] 1.0556258 -2.281384      0      1 1.0811296 -0.9902174      0      1
#> [2,] 0.9445830 -2.228572      0      1 0.7258335 -1.2619422      0      1
#> [3,] 1.1376002 -2.128935      0      1 1.1139359 -0.9778998      0      1
#> [4,] 0.9226204 -2.015896      0      1 0.8837094 -1.2196036      0      1
#> [5,] 0.7387255 -2.204863      0      1 0.8786617 -1.3794804      0      1
#> [6,] 0.8632878 -2.373486      0      1 1.1281751 -1.1560399      0      1
#>         [,315]      [,316] [,317] [,318]    [,319]      [,320] [,321] [,322]
#> [1,] 0.6812042 -0.09213167      0      1 0.6210672 -0.19476202      0      1
#> [2,] 1.0468361 -0.25202156      0      1 1.0676203 -0.46469666      0      1
#> [3,] 0.9235185 -0.09710572      0      1 0.9649329 -0.07104352      0      1
#> [4,] 0.8356593 -0.11258088      0      1 0.8674260 -0.42169272      0      1
#> [5,] 0.7901139 -0.33854975      0      1 0.7725759 -0.34355760      0      1
#> [6,] 1.1085552 -0.24548305      0      1 0.7029192 -0.50570950      0      1
#>           [,323]    [,324]
#> [1,] -0.38627615 1.0792132
#> [2,] -0.07201150 0.9552173
#> [3,] -0.28567396 0.9005742
#> [4,] -0.14555720 0.9988952
#> [5,] -0.03501515 1.1799284
#> [6,] -0.10210340 0.8243333

# }