empirical_rxx {mirt}	R Documentation

Function to calculate the empirical (marginal) reliability

Description

Given secondary latent trait estimates and their associated standard errors returned from fscores, compute the empirical reliability.

Usage

empirical_rxx(Theta_SE)

Arguments

Theta_SE
a matrix of latent trait estimates returned from fscores with the options full.scores = TRUE and full.scores.SE = TRUE

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

See Also

fscores, marginal_rxx

Examples

## No test: 

dat <- expand.table(deAyala)
mod <- mirt(dat, 1)

theta_se <- fscores(mod, full.scores.SE = TRUE)
empirical_rxx(theta_se)

##        F1 
## 0.6200703

theta_se <- fscores(mod, full.scores.SE = TRUE, method = 'ML')
empirical_rxx(theta_se)

##        F1 
## 0.5636644

## End(No test)

[Package mirt version 1.24 Index]