expected.item {mirt}	R Documentation

Function to calculate expected value of item

Description

Given an internal mirt object extracted from an estimated model compute the expected value for an item given the ability parameter(s).

Usage

expected.item(x, Theta, min = 0)

Arguments

x
an extracted internal mirt object containing item information (see extract.item)

Theta
a vector (unidimensional) or matrix (multidimensional) of latent trait values

min
a constant value added to the expected values indicating the lowest theoretical category. Default is 0

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi:10.18637/jss.v048.i06

See Also

extract.item, expected.test

Examples

mod <- mirt(Science, 1)
extr.2 <- extract.item(mod, 2)
Theta <- matrix(seq(-6,6, length.out=200))
expected <- expected.item(extr.2, Theta, min(Science[,1])) #min() of first item
head(data.frame(expected, Theta=Theta))

##   expected     Theta
## 1 1.013391 -6.000000
## 2 1.014407 -5.939698
## 3 1.015498 -5.879397
## 4 1.016672 -5.819095
## 5 1.017933 -5.758794
## 6 1.019289 -5.698492

[Package mirt version 1.38.3 Index]