expected.test {mirt}	R Documentation

Function to calculate expected test score

Description

Given an estimated model compute the expected test score. Returns the expected values in the same form as the data used to estimate the model.

Usage

expected.test(x, Theta, group = NULL, mins = TRUE, individual = FALSE,
  which.items = NULL)

Arguments

x
an estimated mirt object

Theta
a matrix of latent trait values

group
a number signifying which group the item should be extracted from (applies to 'MultipleGroupClass' objects only)

mins
logical; include the minimum value constants in the dataset. If FALSE, the expected values for each item are determined from the scoring 0:(ncat-1)

individual
logical; return tracelines for individual items?

which.items
an integer vector indicating which items to include in the expected test score. Default uses all possible items

See Also

expected.item

Examples

## No test: 
dat <- expand.table(deAyala)
model <- 'F = 1-5
          CONSTRAIN = (1-5, a1)'
mod <- mirt(dat, model)

Theta <- matrix(seq(-6,6,.01))
tscore <- expected.test(mod, Theta)
tail(cbind(Theta, tscore))

##                tscore
## [1196,] 5.95 4.999138
## [1197,] 5.96 4.999150
## [1198,] 5.97 4.999162
## [1199,] 5.98 4.999174
## [1200,] 5.99 4.999186
## [1201,] 6.00 4.999198

# use only first two items (i.e., a bundle)
bscore <- expected.test(mod, Theta, which.items = 1:2)
tail(cbind(Theta, bscore))

##                bscore
## [1196,] 5.95 1.999905
## [1197,] 5.96 1.999906
## [1198,] 5.97 1.999908
## [1199,] 5.98 1.999909
## [1200,] 5.99 1.999910
## [1201,] 6.00 1.999912

## End(No test)

[Package mirt version 1.16 Index]