fixef {mirt}	R Documentation

Compute latent regression fixed effect expected values

Description

Create expected values for fixed effects parameters in latent regression models.

Usage

fixef(x)

Arguments

x
an estimated model object from the mixedmirt or mirt function

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

See Also

mirt, mixedmirt

Examples

## No test: 

#simulate data
set.seed(1234)
N <- 1000

# covariates
X1 <- rnorm(N); X2 <- rnorm(N)
covdata <- data.frame(X1, X2)
Theta <- matrix(0.5 * X1 + -1 * X2 + rnorm(N, sd = 0.5))

#items and response data
a <- matrix(1, 20); d <- matrix(rnorm(20))
dat <- simdata(a, d, 1000, itemtype = 'dich', Theta=Theta)

#conditional model using X1 and X2 as predictors of Theta
mod1 <- mirt(dat, 1, 'Rasch', covdata=covdata, formula = ~ X1 + X2)

#latent regression fixed effects (i.e., expected values)
fixef(mod1)

##                 F1
## 1     0.5920773981
## 2    -0.1609157082
## 3     2.1328439652
## 4    -1.8756553438
## 5    -0.4894065537
## 6     2.2022933369
## 7    -1.2555824305
## 8    -0.0584655889
## 9     0.3888400112
## 10   -0.9197307901
## 11   -1.5517009394
## 12   -2.1141311134
## 13    0.8131009325
## 14    0.4786279068
## 15    0.3543268860
## 16   -0.1249216255
## 17    0.1843514771
## 18    1.9025220967
## 19   -0.1560365162
## 20    2.4677712834
## 21    0.0366416447
## 22    0.5940900682
## 23   -0.1114655702
## 24    0.9945884329
## 25    0.0512630402
## 26   -0.3754129620
## 27   -1.0486574472
## 28   -1.2648519066
## 29    0.2281383146
## 30    0.7554671940
## 31   -0.4866432313
## 32    1.2104533990
## 33    0.8297618660
## 34    0.5260318208
## 35    0.1339757910
## 36    0.3061669026
## 37   -2.4512543454
## 38   -0.5952086453
## 39    0.5732326141
## 40   -0.9269577601
## 41    0.5591636855
## 42    0.6796611546
## 43    0.1216135628
## 44   -0.6895731826
## 45   -2.5034738601
## 46   -2.1155553378
## 47   -1.4320483277
## 48   -0.6431750508
## 49    1.4792683845
## 50    0.4822525843
## 51   -0.7280367366
## 52   -0.8033994584
## 53   -0.6417892665
## 54   -0.5914409008
## 55   -1.9534352781
## 56   -0.0987412933
## 57    1.1491295059
## 58    0.0300083022
## 59    0.7451528708
## 60    0.2195123905
## 61   -1.0462223651
## 62    1.6287439190
## 63    0.5189425435
## 64    0.5361331410
## 65   -0.9551368693
## 66    2.2298729511
## 67   -0.1206288832
## 68    0.3995864764
## 69    0.8072230702
## 70   -0.0784547943
## 71    1.3412633384
## 72    0.1982689012
## 73    1.5397084110
## 74    1.1957835578
## 75   -0.0814569439
## 76   -0.3373849256
## 77   -0.8158681663
## 78    0.5694475976
## 79    0.7013117640
## 80    0.8666939618
## 81   -0.1476189642
## 82   -0.8667398966
## 83   -1.0236155518
## 84    0.5365613686
## 85   -0.1159935217
## 86   -0.0284525430
## 87   -0.5685745926
## 88    0.2366446887
## 89    1.0981317095
## 90   -0.5403829672
## 91   -0.3355863302
## 92   -1.4507014116
## 93    0.9652916934
## 94    0.0911258932
## 95   -0.1752303478
## 96   -2.0051232259
## 97   -1.9546914847
## 98    0.6965626551
## 99    0.6396135113
## 100   1.1342110288
## 101   0.1022481668
## 102   0.3253966021
## 103  -0.0474133370
## 104   0.4436476731
## 105   0.6809732833
## 106  -0.7299960446
## 107  -1.1372896416
## 108   0.6654755575
## 109  -1.1074123929
## 110   0.1683622092
## 111   0.1855543340
## 112   0.6288890800
## 113  -0.5571402093
## 114   0.5845216898
## 115  -0.0568936208
## 116  -0.0166306085
## 117  -1.5738006797
## 118  -0.2441216577
## 119  -0.6557783300
## 120  -1.0965456757
## 121   2.8778997585
## 122  -0.0412921508
## 123   0.4999394251
## 124   1.5313999318
## 125   1.9328032870
## 126  -0.2362208328
## 127   0.1175502489
## 128   0.9662462956
## 129  -1.2541498104
## 130   0.9732039151
## 131   0.1388656376
## 132  -0.2866280169
## 133  -2.1605588682
## 134   0.1508606161
## 135  -1.5416634415
## 136  -1.3510128088
## 137   0.2472543756
## 138   0.6610192412
## 139   0.3371567093
## 140  -1.6741208303
## 141   1.2540554547
## 142  -0.2971084206
## 143   0.3949405922
## 144  -0.1579574500
## 145   0.6242974853
## 146  -0.8444039788
## 147  -0.5754813849
## 148  -1.4517957605
## 149   1.2486262684
## 150  -0.5894302299
## 151   0.1845238506
## 152   1.0506254936
## 153   0.4629972871
## 154  -0.4408588694
## 155  -0.7407880178
## 156  -1.1074061475
## 157  -1.1766364512
## 158   0.1547026880
## 159  -0.5979551327
## 160  -1.6408399338
## 161  -1.6378763543
## 162   1.4739538605
## 163  -1.1791520683
## 164   0.2500853320
## 165   0.8489900074
## 166   1.4466692590
## 167   1.1968171537
## 168  -0.7359009365
## 169  -1.6268958429
## 170   0.6301256577
## 171   0.9678797320
## 172   0.2283085826
## 173  -1.1437708810
## 174   0.1433598886
## 175   0.4858773709
## 176  -1.4774481320
## 177  -1.6683873249
## 178   1.1720784708
## 179   1.3045438264
## 180   0.8170309701
## 181  -0.7003601811
## 182   1.4144758093
## 183   0.4710931220
## 184  -1.0258772185
## 185   2.9649392619
## 186   0.2693859758
## 187   0.8189677333
## 188   0.5149241008
## 189  -0.4133359888
## 190  -0.7629909872
## 191  -0.3799142362
## 192  -0.5627342466
## 193   0.3023139519
## 194   1.8993746572
## 195  -0.7747836778
## 196   0.4206615773
## 197   1.0920101837
## 198   1.6897238805
## 199   0.5217477724
## 200  -0.5316094828
## 201   1.6795961326
## 202  -2.0574949065
## 203  -0.7899034642
## 204   1.9281172444
## 205  -0.9890359284
## 206  -0.6631368574
## 207   0.3666190942
## 208   0.1776390834
## 209  -0.1301852625
## 210  -1.2135970080
## 211   1.2686538435
## 212   0.4500849317
## 213   1.8014294027
## 214  -0.6740608413
## 215  -0.3670394647
## 216  -0.3905978894
## 217  -0.6172403772
## 218  -1.7870878673
## 219  -0.7668216409
## 220  -0.9467099723
## 221  -0.6459634987
## 222   0.9442245467
## 223   0.4285143451
## 224   0.1543964774
## 225   0.0297347562
## 226   0.5721748308
## 227   0.9330346934
## 228   0.2904238216
## 229   0.0018083008
## 230  -0.8397497292
## 231  -1.1785918429
## 232  -1.0318407619
## 233   2.1641403402
## 234   0.9732102506
## 235  -0.7851672604
## 236   0.1924952527
## 237  -2.1104484071
## 238  -0.3195336485
## 239  -1.1099773247
## 240   1.0335491963
## 241   0.0897511612
## 242  -0.1337873868
## 243  -1.8689989001
## 244   0.5531651912
## 245   0.0817972066
## 246   1.4966875276
## 247   0.0682944159
## 248   0.0018850907
## 249  -0.5111139858
## 250   0.5579093900
## 251  -1.4837330921
## 252   0.6979636271
## 253   0.1139849577
## 254   0.4551389484
## 255  -1.4559392325
## 256   0.5867116726
## 257   0.5696776678
## 258  -1.2256692610
## 259   1.4008528700
## 260  -0.0190745269
## 261   1.1151279490
## 262  -1.6350300636
## 263  -0.6332163126
## 264   0.8287228404
## 265  -0.7567332652
## 266  -0.5707925853
## 267   0.4453447951
## 268  -0.2614026001
## 269   1.4947142394
## 270  -0.0429826312
## 271   1.5055141897
## 272   1.0994631510
## 273  -0.1109990679
## 274   0.4126569261
## 275   0.2013909828
## 276  -0.3927725927
## 277  -1.9225972027
## 278   1.3907706027
## 279  -1.5674414762
## 280   1.8080739868
## 281   0.5643464117
## 282   1.1625784448
## 283   1.1783666102
## 284  -1.0822554990
## 285  -1.4760142360
## 286  -1.2261493606
## 287   0.1749877832
## 288  -0.2158542406
## 289  -1.9509562698
## 290  -1.0104756849
## 291  -0.0525671024
## 292  -0.0646428075
## 293   0.2570582626
## 294  -0.6401009356
## 295   1.0497218616
## 296   1.6141352928
## 297  -2.2179616303
## 298   0.5205629094
## 299  -0.0614180518
## 300  -0.3571036608
## 301  -1.3723807650
## 302  -0.8731276470
## 303  -2.0994618602
## 304   0.0770828722
## 305   0.5088312996
## 306  -0.4444896618
## 307  -0.9444971422
## 308   3.0726962378
## 309  -0.1372339891
## 310   0.3579562784
## 311   1.0141613731
## 312   0.1804223235
## 313   1.8031528220
## 314  -0.3728546030
## 315  -0.2611974570
## 316   1.1441886230
## 317   2.8326968695
## 318   1.2306238314
## 319  -0.4146076017
## 320  -0.2669842714
## 321   0.6936689947
## 322  -0.3142056466
## 323   0.6608311069
## 324  -1.5398591619
## 325  -0.8234281492
## 326  -1.3488776684
## 327  -0.2442197181
## 328  -0.1416572364
## 329   0.6160057920
## 330  -1.6056846186
## 331  -0.4675493977
## 332   0.1186819634
## 333  -1.5720259766
## 334   0.9277748501
## 335  -0.0718026150
## 336  -1.9179552559
## 337   0.1122499102
## 338   0.7121106045
## 339   0.0367881264
## 340  -1.6828358918
## 341  -0.1659596287
## 342   0.1047678036
## 343   1.1932353986
## 344   1.0251006184
## 345   1.6475018346
## 346   0.1943620902
## 347  -1.2623504402
## 348   0.1298793325
## 349  -0.6552166546
## 350  -0.6047781746
## 351  -1.4128796560
## 352   0.2222950659
## 353  -2.0725442962
## 354  -0.4569631707
## 355  -2.2568349268
## 356  -0.1712841140
## 357   0.4963654617
## 358   0.0881309512
## 359  -2.7937913585
## 360   1.0825661213
## 361   0.9952163297
## 362   0.6144294844
## 363  -0.4185947801
## 364  -0.8153789963
## 365   0.6626223311
## 366   0.8916160234
## 367  -2.2741643023
## 368   0.6519249858
## 369   0.3377105737
## 370   0.5605456657
## 371   0.2971020997
## 372  -1.9443760394
## 373  -0.3073785547
## 374  -0.1117986251
## 375   1.1776800338
## 376   0.4221909338
## 377  -2.2119246374
## 378  -0.0653298802
## 379  -0.4024411684
## 380   0.7659207396
## 381   0.6544926629
## 382  -2.1113586382
## 383  -1.7773713666
## 384   2.3320343337
## 385  -0.4204023039
## 386   0.3007657989
## 387   0.8873169021
## 388  -1.6122352229
## 389   0.2313272634
## 390   0.2271407889
## 391   0.1580765922
## 392  -0.6805489970
## 393  -1.5336855782
## 394   0.6940730566
## 395  -0.9719465787
## 396   0.4375822815
## 397   0.9867611649
## 398   1.6448951993
## 399   1.4031266693
## 400  -0.3290587485
## 401   1.3957211901
## 402   1.6039478431
## 403  -2.2156467604
## 404  -0.2903415584
## 405  -1.1608734538
## 406  -1.5954167624
## 407   0.6458155465
## 408  -0.5917206321
## 409  -0.4711392330
## 410  -0.7301174617
## 411  -1.5523353847
## 412   1.8987704391
## 413  -0.4194862701
## 414  -1.7491693748
## 415   0.0239647739
## 416  -0.5094714890
## 417  -0.3128767193
## 418  -0.1667139926
## 419  -1.2699187004
## 420  -0.3852124119
## 421   0.9731794332
## 422   0.3863924559
## 423  -0.6325121836
## 424  -0.0782592018
## 425   1.7393777807
## 426   1.9744181251
## 427   0.9003078489
## 428  -0.6379224964
## 429   0.0620288040
## 430   0.0051503961
## 431  -1.2314328243
## 432  -1.9966537689
## 433  -1.0217053501
## 434  -0.6397361625
## 435   1.2729962378
## 436   1.3588887371
## 437  -0.0146709106
## 438  -0.0785191120
## 439   0.2446273924
## 440  -0.3120749669
## 441   0.3509100999
## 442  -1.3567663950
## 443   0.1246156243
## 444  -0.1008900876
## 445  -0.4104414657
## 446   1.6298762449
## 447  -1.1736327157
## 448  -0.3820170906
## 449   0.7835906979
## 450   0.6763255471
## 451  -0.6827959251
## 452   1.1688365497
## 453   1.0275176755
## 454  -0.1533431104
## 455   2.4343314618
## 456  -0.4994164654
## 457  -0.8159427039
## 458  -0.3637624682
## 459   0.9598610577
## 460  -0.1404644239
## 461   0.5200049015
## 462  -0.5584641612
## 463  -2.0264357572
## 464   0.9674529829
## 465   0.0005219902
## 466  -0.7093910354
## 467  -0.7280053751
## 468  -0.1649446870
## 469  -0.8817650233
## 470  -0.3861564093
## 471   1.8315026686
## 472   1.8692032377
## 473  -0.4648557842
## 474   0.5560442257
## 475   0.9499179976
## 476  -0.9692388791
## 477  -0.3415042056
## 478   1.1036743349
## 479  -0.2107165351
## 480  -1.9371995602
## 481  -1.3927990479
## 482   1.6609471747
## 483  -0.2687572515
## 484  -0.1018213021
## 485   1.5093686428
## 486   2.0639099720
## 487  -1.4167362987
## 488  -0.3743726147
## 489   0.1263409286
## 490   0.6917366855
## 491  -0.4649044542
## 492  -0.6950040039
## 493   2.0175715353
## 494  -0.7983909827
## 495  -0.3705155245
## 496  -0.0628548663
## 497  -0.3381833083
## 498  -0.3104188158
## 499  -1.4357821310
## 500  -0.2261199379
## 501  -1.3069676289
## 502   0.7446222315
## 503   1.0911052698
## 504   0.5080848933
## 505   1.2498163552
## 506   0.2546654588
## 507  -1.2667135397
## 508   1.5395927415
## 509   0.7219416252
## 510   0.5566489960
## 511  -0.1734418095
## 512   0.1743817368
## 513  -1.2293598132
## 514  -0.2370913304
## 515  -0.4333337594
## 516   0.9811651952
## 517  -4.3995024569
## 518  -0.6911163909
## 519  -1.0299298510
## 520   2.6086691827
## 521   0.9953322001
## 522  -0.8556581306
## 523   0.2865760114
## 524  -0.6183740185
## 525  -0.9548091509
## 526  -0.0025034785
## 527   0.0729450621
## 528   1.0606438996
## 529  -1.1195299992
## 530   2.0809753752
## 531   0.3895071934
## 532  -0.4506273929
## 533   0.0678512742
## 534  -0.6169397333
## 535  -0.6257132617
## 536   0.2540400906
## 537  -0.0842027784
## 538   0.2587754841
## 539  -0.4890283040
## 540   0.7176626223
## 541   0.8549011836
## 542   0.9501061816
## 543  -1.2689255138
## 544  -1.1507782135
## 545   0.3156979299
## 546  -0.6472184284
## 547   0.0474940479
## 548   0.8730605839
## 549   0.9368236969
## 550   1.6733185286
## 551   1.3113193165
## 552  -1.3728492456
## 553  -1.0107208246
## 554   0.4841806834
## 555  -0.8844533864
## 556  -1.5285291188
## 557   2.1642346220
## 558  -2.6708779999
## 559  -0.7477887252
## 560   0.1680315757
## 561   0.1641210316
## 562   0.5565945945
## 563   0.7367757755
## 564   0.7037632260
## 565   2.5194592412
## 566   1.3856662904
## 567  -1.4479772943
## 568   0.9196417179
## 569  -0.8911671088
## 570  -0.4217384835
## 571  -0.2296408833
## 572  -1.4120613936
## 573   1.1050160151
## 574   0.4504655206
## 575  -0.0545124500
## 576  -2.2107480024
## 577  -0.2004361561
## 578  -1.0378441251
## 579   0.9116880147
## 580  -0.4426653632
## 581  -0.9033957666
## 582   0.6153675104
## 583   1.0868961956
## 584   0.9057301463
## 585   0.5053345670
## 586  -0.6635023741
## 587  -2.0466173858
## 588   1.9625243068
## 589   0.6929749372
## 590   1.8887338355
## 591  -0.0292584089
## 592  -0.3261701208
## 593  -0.0348515820
## 594  -0.7625967012
## 595  -0.5687244115
## 596   0.3322820262
## 597  -1.6277836343
## 598  -0.0191398113
## 599  -1.0326680579
## 600   0.4368527338
## 601   0.4177583791
## 602   1.5629573997
## 603  -0.8129689381
## 604  -0.9757075082
## 605   0.4712421621
## 606  -0.1613012398
## 607  -0.2163473706
## 608   2.3892857923
## 609   0.7900672520
## 610   1.4730836553
## 611  -0.6371605640
## 612  -0.0123792685
## 613  -1.0149050511
## 614   1.3244018054
## 615  -0.3208016663
## 616   1.8703296182
## 617  -0.8378476155
## 618  -1.0114031363
## 619   0.2129966275
## 620  -0.0869840527
## 621   0.6915441913
## 622  -1.6431597507
## 623  -0.5094473294
## 624  -0.4164456825
## 625  -0.3981219726
## 626   0.2021555984
## 627   1.2359695512
## 628  -0.9513304128
## 629   1.4614829857
## 630  -0.2085271860
## 631   2.7248845269
## 632   0.3292121377
## 633   0.1515321470
## 634   2.4367686312
## 635   0.9556997893
## 636   0.6758950163
## 637  -0.5225820236
## 638  -0.2857649133
## 639  -0.7133915628
## 640   1.2034860000
## 641  -0.0300768756
## 642  -0.7712275406
## 643   0.1491497426
## 644   0.8550319001
## 645  -1.1832532181
## 646  -2.2971473815
## 647  -2.1124436454
## 648  -0.8531248537
## 649  -0.1776910267
## 650   0.1513976480
## 651  -0.0298633365
## 652  -0.0705227662
## 653  -0.0548864080
## 654   2.2621847315
## 655  -0.1190762865
## 656  -0.1391955165
## 657   1.4744781065
## 658   1.6362200999
## 659   0.3114974921
## 660   3.4405435808
## 661  -2.3595083233
## 662  -0.4166607367
## 663   0.7120922475
## 664  -0.6120962003
## 665  -0.7588717664
## 666   1.3123903673
## 667   0.2791781945
## 668  -1.5994594828
## 669   2.0279056932
## 670   0.1280757751
## 671   1.2802578451
## 672  -0.9611647710
## 673  -0.2699550621
## 674  -0.8838353378
## 675  -0.9990897938
## 676   0.4861003766
## 677   0.9643100353
## 678   0.2606689377
## 679   0.2109216150
## 680  -2.6018233238
## 681   0.4779730768
## 682  -0.6553364651
## 683  -0.2021415291
## 684   0.2823588339
## 685  -0.5665149443
## 686  -0.8148325995
## 687   0.7177208236
## 688   0.2254366245
## 689  -0.1847573076
## 690  -0.2253546322
## 691   2.0483177254
## 692   0.0820708419
## 693   0.1647288743
## 694   2.4941902980
## 695   0.6021358391
## 696  -0.1896741010
## 697   1.6949648253
## 698  -0.7198463313
## 699   2.0251842552
## 700   1.4579988200
## 701   0.6957171034
## 702  -0.7200487030
## 703   0.0482039379
## 704   0.3927089994
## 705  -1.7526253919
## 706  -0.0519776340
## 707  -0.3889417993
## 708  -1.6294409432
## 709  -0.0592498380
## 710  -1.9317197991
## 711   1.7242548597
## 712  -0.8645861953
## 713   1.6172980189
## 714   1.0647048320
## 715   0.4001035182
## 716  -0.8694830275
## 717  -0.9142655525
## 718  -0.6057743843
## 719  -0.4448995570
## 720   0.1712160781
## 721  -2.5529142768
## 722   1.1013200549
## 723  -0.5243537969
## 724   0.4493183245
## 725   0.8800368875
## 726   1.4805166321
## 727   0.9682396603
## 728   0.1806671282
## 729   2.0314694844
## 730   0.0628743625
## 731   0.1047823973
## 732   1.8748889404
## 733   0.7058331338
## 734   1.1742562275
## 735  -0.1070471287
## 736  -1.0931003380
## 737   2.9359271365
## 738  -2.5011676630
## 739   1.2379796735
## 740  -0.7558905040
## 741  -1.7573845099
## 742   0.8188596962
## 743   1.9165326827
## 744  -1.5246209045
## 745   0.0828385947
## 746   0.9949038371
## 747  -0.6801644512
## 748  -1.7675795511
## 749   0.2882691565
## 750  -1.8656897029
## 751  -1.9866759062
## 752   0.1241117058
## 753   1.4607749293
## 754   0.7333844725
## 755  -1.8192871370
## 756  -0.3417477567
## 757  -1.2250645428
## 758  -0.7517040205
## 759  -0.2502028719
## 760  -0.0807756737
## 761  -0.8491454347
## 762   0.2543166641
## 763  -2.4261473714
## 764   0.2546009663
## 765  -1.0620899202
## 766  -0.7542316196
## 767  -0.0117516358
## 768   1.6236914073
## 769   0.8153898544
## 770  -0.0629232922
## 771   0.1957235712
## 772  -0.7543081874
## 773  -0.9489968731
## 774  -0.3933030535
## 775  -0.8657296539
## 776  -0.8579806665
## 777  -0.4885846919
## 778  -1.1875038314
## 779   0.4155688926
## 780  -0.3971093521
## 781   1.3630292853
## 782  -0.7278240616
## 783  -1.3355276604
## 784   0.3675979706
## 785  -0.0864549955
## 786  -0.5479315833
## 787   1.4827519129
## 788  -2.5613911892
## 789  -0.5639652430
## 790  -0.6379506427
## 791   0.4895568790
## 792   1.7722070390
## 793  -1.2875331206
## 794   0.4596889208
## 795  -0.6181428998
## 796  -2.0233672816
## 797   1.7808914248
## 798  -1.6593004407
## 799  -0.2814846020
## 800  -0.6118920968
## 801   0.0040120381
## 802  -0.2209645358
## 803   1.1430802029
## 804   0.8212480101
## 805  -0.5327486759
## 806   0.5970970676
## 807   1.0675552316
## 808   0.2931088657
## 809  -0.1227192770
## 810   0.5463511724
## 811   1.0114233757
## 812   1.5526052576
## 813  -0.9328595114
## 814   0.2747751566
## 815  -1.9832975396
## 816  -1.1305133920
## 817   0.1721638827
## 818  -0.0310561559
## 819   0.4310652799
## 820  -0.7719684972
## 821  -0.6029042657
## 822   2.2371175674
## 823   1.3564993424
## 824  -0.2610187212
## 825   3.0932479233
## 826   0.9935736429
## 827  -1.7548993200
## 828   1.4071775682
## 829  -0.8652338552
## 830  -1.6052257829
## 831  -0.6422604770
## 832  -1.2249145220
## 833   0.6665842340
## 834  -0.3128741906
## 835   0.1883997103
## 836  -0.8080021014
## 837  -1.3482530920
## 838  -0.0127702325
## 839  -1.4153805139
## 840   0.3687327108
## 841  -0.4886493920
## 842   0.6299224774
## 843   1.1535635276
## 844   0.1383510514
## 845  -0.1013220074
## 846  -1.4577334193
## 847   0.2376042885
## 848   0.2209592165
## 849   0.6008623985
## 850  -2.8294025363
## 851  -0.5322663403
## 852  -0.0391381487
## 853  -1.8703827519
## 854  -1.0876961278
## 855  -2.2029207033
## 856   0.1729760488
## 857   0.1662096346
## 858   0.1923704052
## 859  -0.3389168832
## 860  -0.2440089926
## 861   0.2128979489
## 862   1.4500998391
## 863  -0.0218038609
## 864  -1.9753871215
## 865  -1.9326982537
## 866   0.7503273405
## 867  -0.7133916957
## 868   0.1003388824
## 869  -0.8977851552
## 870   1.4463049719
## 871   1.5832227866
## 872  -2.2844754468
## 873  -0.9199290317
## 874   1.9584661242
## 875  -0.4430399021
## 876  -1.2035447296
## 877   0.0211938193
## 878   0.4350036094
## 879  -0.3460584480
## 880  -1.2887215069
## 881  -0.8948842023
## 882  -0.9439662625
## 883  -1.6886354965
## 884  -1.2421822873
## 885   0.7076314556
## 886   0.6983053500
## 887  -0.9718408245
## 888   0.0755143448
## 889   1.2320413187
## 890  -0.3289172778
## 891   0.8621893114
## 892  -1.7041970685
## 893   2.9063939291
## 894   0.1718011281
## 895  -0.1401200916
## 896  -0.1128473946
## 897   0.4756762281
## 898  -2.4771019354
## 899   0.9263068487
## 900   0.3732595194
## 901  -2.6494026082
## 902  -0.9327995620
## 903  -0.9637530291
## 904   0.4810906031
## 905   0.0062815654
## 906   1.8496019976
## 907  -0.3949273157
## 908  -0.4430048131
## 909   1.0164363446
## 910   0.0009142350
## 911  -1.2036676361
## 912  -1.8759568159
## 913   1.5938362714
## 914  -1.0217428526
## 915  -0.5189851249
## 916  -0.5658403180
## 917   1.8693653025
## 918  -0.7113145210
## 919   1.9090483479
## 920   0.8070603365
## 921   0.0297081456
## 922  -1.5403072313
## 923   1.2763954250
## 924  -0.1899671241
## 925  -0.4447117343
## 926   1.6831413745
## 927  -0.1930201678
## 928  -1.2465953040
## 929  -1.6512926265
## 930  -1.5525182175
## 931   0.1161391206
## 932  -0.4705906388
## 933   2.1320684473
## 934   0.7321115417
## 935   0.5823500022
## 936   0.1337214652
## 937  -1.1680070176
## 938   0.3397427897
## 939   1.2532463343
## 940  -0.2388096291
## 941   1.4840557750
## 942  -0.3357097700
## 943   0.1802397903
## 944  -1.4902682012
## 945  -0.6685785009
## 946   0.2636495738
## 947   0.8730040347
## 948  -0.4833730704
## 949  -1.2342221037
## 950  -1.2894217914
## 951   1.9159274893
## 952   1.0998410098
## 953  -1.0762183965
## 954  -0.2579516369
## 955   0.6769062437
## 956   0.5285161440
## 957  -2.3478839598
## 958   1.0058814213
## 959  -1.6263401055
## 960  -0.3689751772
## 961  -0.1994192310
## 962  -0.4943583087
## 963   1.5349352833
## 964  -0.3971915019
## 965   0.0676944510
## 966   0.3802756773
## 967  -0.5452569224
## 968  -1.1008080625
## 969   1.2440747597
## 970  -1.9712929349
## 971   0.2601790150
## 972   1.9425631366
## 973  -0.9301550901
## 974  -0.2862862277
## 975   0.5730567960
## 976   1.3793865418
## 977   0.5778235559
## 978  -0.7553127537
## 979  -0.7900875757
## 980   0.9074737924
## 981   0.7540109167
## 982   1.1603987567
## 983  -0.2573644088
## 984  -0.5353037513
## 985   0.1830381171
## 986   1.1115570323
## 987   0.5912813612
## 988   1.1106280542
## 989  -0.3470183850
## 990  -1.4630581256
## 991   0.0828179628
## 992  -0.8129469401
## 993   1.0429339640
## 994   0.5082483976
## 995  -2.8542426440
## 996  -0.3794118274
## 997  -1.0681441274
## 998   0.0013293725
## 999   1.7971503356
## 1000  0.9762739569

# with mixedmirt()
mod1b <- mixedmirt(dat, covdata, 1, lr.fixed = ~ X1 + X2, fixed = ~ 0 + items)
fixef(mod1b)

##                F1
## 1     0.593462526
## 2    -0.161272290
## 3     2.135701400
## 4    -1.877621871
## 5    -0.490328314
## 6     2.205484680
## 7    -1.257289557
## 8    -0.058341041
## 9     0.389660085
## 10   -0.920797196
## 11   -1.553902162
## 12   -2.116993518
## 13    0.814657193
## 14    0.479339353
## 15    0.354495713
## 16   -0.125070631
## 17    0.184835809
## 18    1.905808251
## 19   -0.155948126
## 20    2.470621719
## 21    0.036645420
## 22    0.595197128
## 23   -0.111464241
## 24    0.995939009
## 25    0.051614221
## 26   -0.375422436
## 27   -1.050497213
## 28   -1.266397110
## 29    0.228495426
## 30    0.756997417
## 31   -0.487824946
## 32    1.212503295
## 33    0.831317550
## 34    0.527038269
## 35    0.134821408
## 36    0.307096450
## 37   -2.454171902
## 38   -0.595598526
## 39    0.574230488
## 40   -0.928201757
## 41    0.559455498
## 42    0.681126771
## 43    0.122136476
## 44   -0.690524494
## 45   -2.506937163
## 46   -2.118431659
## 47   -1.433818058
## 48   -0.643673700
## 49    1.481750995
## 50    0.483189914
## 51   -0.728448559
## 52   -0.804407666
## 53   -0.642341945
## 54   -0.591952944
## 55   -1.956378471
## 56   -0.099114274
## 57    1.150251607
## 58    0.030358021
## 59    0.745669584
## 60    0.220304702
## 61   -1.048090500
## 62    1.630250590
## 63    0.519754988
## 64    0.537221223
## 65   -0.956604114
## 66    2.232607891
## 67   -0.120367683
## 68    0.399664707
## 69    0.807943788
## 70   -0.078707618
## 71    1.343325222
## 72    0.198752857
## 73    1.542222317
## 74    1.197369771
## 75   -0.082394873
## 76   -0.337844052
## 77   -0.816578196
## 78    0.570608133
## 79    0.702289921
## 80    0.868152492
## 81   -0.147776413
## 82   -0.868007338
## 83   -1.024652586
## 84    0.537455499
## 85   -0.116505770
## 86   -0.028770138
## 87   -0.569665656
## 88    0.237167018
## 89    1.099897246
## 90   -0.540745941
## 91   -0.336082030
## 92   -1.453034617
## 93    0.966107994
## 94    0.090873086
## 95   -0.175305570
## 96   -2.008349232
## 97   -1.957255002
## 98    0.697290186
## 99    0.640216208
## 100   1.135124405
## 101   0.102242863
## 102   0.326083799
## 103  -0.047512221
## 104   0.444527787
## 105   0.682345686
## 106  -0.731185258
## 107  -1.138688489
## 108   0.666433905
## 109  -1.109256336
## 110   0.168641817
## 111   0.185916856
## 112   0.630111869
## 113  -0.557562244
## 114   0.585088111
## 115  -0.056989973
## 116  -0.016982415
## 117  -1.575734849
## 118  -0.244563771
## 119  -0.657051966
## 120  -1.098223261
## 121   2.882404789
## 122  -0.041048827
## 123   0.499901184
## 124   1.533462637
## 125   1.935062456
## 126  -0.236615866
## 127   0.117979009
## 128   0.968327564
## 129  -1.255830649
## 130   0.974613136
## 131   0.138681537
## 132  -0.287168419
## 133  -2.163424619
## 134   0.150830376
## 135  -1.543388869
## 136  -1.352948814
## 137   0.247759042
## 138   0.662801390
## 139   0.337314584
## 140  -1.676453316
## 141   1.256116909
## 142  -0.298113326
## 143   0.395298636
## 144  -0.158158034
## 145   0.625056815
## 146  -0.845860465
## 147  -0.577019857
## 148  -1.454138773
## 149   1.250349559
## 150  -0.590473940
## 151   0.184955945
## 152   1.052204393
## 153   0.463066795
## 154  -0.441193823
## 155  -0.741976091
## 156  -1.109645374
## 157  -1.178355547
## 158   0.155354253
## 159  -0.598534183
## 160  -1.643212539
## 161  -1.640064946
## 162   1.476324994
## 163  -1.180804469
## 164   0.250542129
## 165   0.850137084
## 166   1.448650940
## 167   1.198000733
## 168  -0.737006738
## 169  -1.629274178
## 170   0.630518259
## 171   0.968700658
## 172   0.228643042
## 173  -1.145400910
## 174   0.144295755
## 175   0.486071381
## 176  -1.479393612
## 177  -1.670514395
## 178   1.172688013
## 179   1.306459652
## 180   0.818301668
## 181  -0.700365890
## 182   1.416692259
## 183   0.471435195
## 184  -1.027618773
## 185   2.969145083
## 186   0.269022059
## 187   0.819769505
## 188   0.515916387
## 189  -0.414248609
## 190  -0.764087574
## 191  -0.380287850
## 192  -0.562479622
## 193   0.303089222
## 194   1.902106881
## 195  -0.776827205
## 196   0.421112582
## 197   1.093447679
## 198   1.692703949
## 199   0.522487053
## 200  -0.531612112
## 201   1.681991070
## 202  -2.060935450
## 203  -0.791192163
## 204   1.930810144
## 205  -0.990680667
## 206  -0.664456083
## 207   0.366460293
## 208   0.177475939
## 209  -0.130398475
## 210  -1.215027681
## 211   1.270442585
## 212   0.450934826
## 213   1.803616150
## 214  -0.674467664
## 215  -0.367441269
## 216  -0.391364762
## 217  -0.618130933
## 218  -1.789600804
## 219  -0.767882402
## 220  -0.948414312
## 221  -0.647309160
## 222   0.945418857
## 223   0.428281650
## 224   0.154173561
## 225   0.029667606
## 226   0.573314520
## 227   0.933325192
## 228   0.290604994
## 229   0.002079668
## 230  -0.841115539
## 231  -1.180278715
## 232  -1.033321244
## 233   2.167837955
## 234   0.974662440
## 235  -0.786501122
## 236   0.193214474
## 237  -2.112428039
## 238  -0.319925469
## 239  -1.111697483
## 240   1.034906883
## 241   0.089866104
## 242  -0.134155575
## 243  -1.871444969
## 244   0.553737548
## 245   0.081640965
## 246   1.498892590
## 247   0.067866835
## 248   0.001729246
## 249  -0.512004495
## 250   0.558662970
## 251  -1.486188470
## 252   0.698621914
## 253   0.113982937
## 254   0.455579244
## 255  -1.457734334
## 256   0.587760204
## 257   0.569974881
## 258  -1.227075557
## 259   1.402907897
## 260  -0.019262844
## 261   1.116461465
## 262  -1.637409916
## 263  -0.634253272
## 264   0.830690528
## 265  -0.758030987
## 266  -0.571409584
## 267   0.446123973
## 268  -0.261338797
## 269   1.496863953
## 270  -0.043310414
## 271   1.507951163
## 272   1.100439256
## 273  -0.111433110
## 274   0.412919802
## 275   0.200896628
## 276  -0.393121632
## 277  -1.925873999
## 278   1.392524182
## 279  -1.569851796
## 280   1.810731908
## 281   0.565612224
## 282   1.164183454
## 283   1.180455572
## 284  -1.084240756
## 285  -1.477967773
## 286  -1.228162113
## 287   0.175589384
## 288  -0.216090044
## 289  -1.953349882
## 290  -1.012081505
## 291  -0.053034787
## 292  -0.064814233
## 293   0.258147162
## 294  -0.640842686
## 295   1.050687718
## 296   1.615909594
## 297  -2.220914528
## 298   0.521508129
## 299  -0.061651383
## 300  -0.357778609
## 301  -1.374265630
## 302  -0.874097496
## 303  -2.102623111
## 304   0.076805193
## 305   0.509605264
## 306  -0.444919127
## 307  -0.945753036
## 308   3.076792378
## 309  -0.137269808
## 310   0.358207743
## 311   1.015144835
## 312   0.180525915
## 313   1.806093909
## 314  -0.373412833
## 315  -0.261406304
## 316   1.145468000
## 317   2.837115899
## 318   1.231909774
## 319  -0.415025379
## 320  -0.267141255
## 321   0.694680573
## 322  -0.314698793
## 323   0.662046303
## 324  -1.541608777
## 325  -0.824707442
## 326  -1.350672734
## 327  -0.245020547
## 328  -0.141501376
## 329   0.616511877
## 330  -1.607901490
## 331  -0.468383879
## 332   0.118511924
## 333  -1.574242178
## 334   0.928985354
## 335  -0.071878612
## 336  -1.920461988
## 337   0.112528696
## 338   0.712571009
## 339   0.036928796
## 340  -1.685105294
## 341  -0.166193885
## 342   0.104799497
## 343   1.194697182
## 344   1.026229559
## 345   1.649998359
## 346   0.194204195
## 347  -1.263993601
## 348   0.130999433
## 349  -0.656040094
## 350  -0.605506638
## 351  -1.414145481
## 352   0.222416482
## 353  -2.075426918
## 354  -0.457577846
## 355  -2.259685933
## 356  -0.171762663
## 357   0.496387238
## 358   0.088611239
## 359  -2.798047627
## 360   1.083467666
## 361   0.996381747
## 362   0.614624426
## 363  -0.419237063
## 364  -0.816574339
## 365   0.663842318
## 366   0.892425432
## 367  -2.277160225
## 368   0.652807870
## 369   0.338248586
## 370   0.560524699
## 371   0.297798067
## 372  -1.946776623
## 373  -0.307943002
## 374  -0.111955154
## 375   1.179822162
## 376   0.422788885
## 377  -2.215221285
## 378  -0.065164041
## 379  -0.403257210
## 380   0.767229882
## 381   0.656220311
## 382  -2.113567864
## 383  -1.779879382
## 384   2.335053261
## 385  -0.421378374
## 386   0.300751877
## 387   0.888297122
## 388  -1.614840846
## 389   0.232274837
## 390   0.227409741
## 391   0.157693034
## 392  -0.680263664
## 393  -1.535739693
## 394   0.694708737
## 395  -0.973650206
## 396   0.437946023
## 397   0.988100670
## 398   1.647215682
## 399   1.405280736
## 400  -0.329205559
## 401   1.398351108
## 402   1.606402589
## 403  -2.218891886
## 404  -0.290435493
## 405  -1.162824209
## 406  -1.597932790
## 407   0.646235330
## 408  -0.592191290
## 409  -0.471661229
## 410  -0.731211906
## 411  -1.554172342
## 412   1.901804933
## 413  -0.419854035
## 414  -1.751018938
## 415   0.024113064
## 416  -0.510046098
## 417  -0.313803012
## 418  -0.166733527
## 419  -1.272092282
## 420  -0.385861344
## 421   0.975162880
## 422   0.386840380
## 423  -0.633444164
## 424  -0.078444841
## 425   1.741796915
## 426   1.977412592
## 427   0.901071534
## 428  -0.638949795
## 429   0.062616456
## 430   0.004867156
## 431  -1.232848386
## 432  -1.998974060
## 433  -1.022960162
## 434  -0.640428627
## 435   1.275194179
## 436   1.360349832
## 437  -0.015335436
## 438  -0.078924953
## 439   0.245503901
## 440  -0.312564424
## 441   0.351259065
## 442  -1.358801066
## 443   0.124740036
## 444  -0.100674572
## 445  -0.410567885
## 446   1.631832569
## 447  -1.175544212
## 448  -0.382444836
## 449   0.784669205
## 450   0.676557120
## 451  -0.683989284
## 452   1.170081773
## 453   1.028562583
## 454  -0.153419340
## 455   2.437779312
## 456  -0.499929626
## 457  -0.816620290
## 458  -0.363850771
## 459   0.961396929
## 460  -0.141385953
## 461   0.520764203
## 462  -0.559008698
## 463  -2.029645440
## 464   0.968668228
## 465   0.001031137
## 466  -0.710069049
## 467  -0.728363938
## 468  -0.164700305
## 469  -0.882774009
## 470  -0.386636243
## 471   1.833768440
## 472   1.871731701
## 473  -0.465678468
## 474   0.557122534
## 475   0.951395974
## 476  -0.970945743
## 477  -0.341832054
## 478   1.104957540
## 479  -0.211471619
## 480  -1.939723633
## 481  -1.394649816
## 482   1.663030225
## 483  -0.268508743
## 484  -0.102142763
## 485   1.511600350
## 486   2.065832588
## 487  -1.417845712
## 488  -0.374618884
## 489   0.126270859
## 490   0.692144324
## 491  -0.465949637
## 492  -0.695513206
## 493   2.019798237
## 494  -0.798930240
## 495  -0.370649015
## 496  -0.062934686
## 497  -0.339577372
## 498  -0.311095359
## 499  -1.438278878
## 500  -0.227146240
## 501  -1.309365935
## 502   0.746249106
## 503   1.092506239
## 504   0.508909848
## 505   1.251040539
## 506   0.255153010
## 507  -1.268064013
## 508   1.541769580
## 509   0.722913764
## 510   0.557512756
## 511  -0.173294109
## 512   0.174979419
## 513  -1.231247195
## 514  -0.237863721
## 515  -0.433857741
## 516   0.983017518
## 517  -4.405213201
## 518  -0.692374201
## 519  -1.031682931
## 520   2.612170838
## 521   0.996288640
## 522  -0.856919087
## 523   0.286507491
## 524  -0.619326389
## 525  -0.956013397
## 526  -0.002174430
## 527   0.073193514
## 528   1.062187253
## 529  -1.121242522
## 530   2.084099752
## 531   0.390028675
## 532  -0.451601505
## 533   0.068062936
## 534  -0.618531224
## 535  -0.626848201
## 536   0.254899572
## 537  -0.084223811
## 538   0.258971339
## 539  -0.489619727
## 540   0.718631906
## 541   0.856046629
## 542   0.951057076
## 543  -1.270622975
## 544  -1.152542965
## 545   0.316406688
## 546  -0.647772304
## 547   0.048292842
## 548   0.874327733
## 549   0.938029634
## 550   1.675643387
## 551   1.313883360
## 552  -1.374519076
## 553  -1.011828731
## 554   0.485196267
## 555  -0.885881613
## 556  -1.531115287
## 557   2.167885885
## 558  -2.673865042
## 559  -0.749314145
## 560   0.168048736
## 561   0.164648740
## 562   0.557872746
## 563   0.738027626
## 564   0.704804089
## 565   2.523053568
## 566   1.387253641
## 567  -1.449872975
## 568   0.920943207
## 569  -0.893062510
## 570  -0.422033899
## 571  -0.229906936
## 572  -1.413539692
## 573   1.106604143
## 574   0.451410572
## 575  -0.054233339
## 576  -2.213844850
## 577  -0.200919144
## 578  -1.038592935
## 579   0.913258369
## 580  -0.442948886
## 581  -0.904408737
## 582   0.616128452
## 583   1.088431183
## 584   0.907166328
## 585   0.505901456
## 586  -0.664287379
## 587  -2.048973609
## 588   1.965092105
## 589   0.694132167
## 590   1.891784429
## 591  -0.029365269
## 592  -0.326972804
## 593  -0.035072460
## 594  -0.764134491
## 595  -0.569617948
## 596   0.332991426
## 597  -1.630171339
## 598  -0.019492252
## 599  -1.033924307
## 600   0.437899450
## 601   0.418278394
## 602   1.565122854
## 603  -0.813556611
## 604  -0.977517382
## 605   0.471962858
## 606  -0.161478389
## 607  -0.217114851
## 608   2.392382952
## 609   0.791733725
## 610   1.474954097
## 611  -0.637893261
## 612  -0.012450273
## 613  -1.016656035
## 614   1.326388169
## 615  -0.320721324
## 616   1.873407345
## 617  -0.839741593
## 618  -1.013027794
## 619   0.213652536
## 620  -0.086860506
## 621   0.692193601
## 622  -1.645824876
## 623  -0.509879385
## 624  -0.417089171
## 625  -0.398968979
## 626   0.201790703
## 627   1.238218133
## 628  -0.952645661
## 629   1.463341672
## 630  -0.208667972
## 631   2.728505777
## 632   0.329309434
## 633   0.152416542
## 634   2.440353121
## 635   0.957727881
## 636   0.677510278
## 637  -0.523414066
## 638  -0.286227040
## 639  -0.714476839
## 640   1.204842302
## 641  -0.030238950
## 642  -0.772150506
## 643   0.149332391
## 644   0.856549930
## 645  -1.185047851
## 646  -2.300232776
## 647  -2.115650070
## 648  -0.854478458
## 649  -0.177689123
## 650   0.151614510
## 651  -0.029654675
## 652  -0.070378560
## 653  -0.054890655
## 654   2.265474236
## 655  -0.119178988
## 656  -0.139484621
## 657   1.476367034
## 658   1.638850671
## 659   0.312017645
## 660   3.445638905
## 661  -2.362646477
## 662  -0.417515539
## 663   0.713398325
## 664  -0.612591729
## 665  -0.759345978
## 666   1.314691128
## 667   0.280067213
## 668  -1.601458990
## 669   2.030388539
## 670   0.128748666
## 671   1.281837024
## 672  -0.962858278
## 673  -0.270023337
## 674  -0.884955917
## 675  -1.000466683
## 676   0.486773892
## 677   0.965931096
## 678   0.260745659
## 679   0.211361158
## 680  -2.604978416
## 681   0.478746578
## 682  -0.655618356
## 683  -0.202820757
## 684   0.283086918
## 685  -0.566949261
## 686  -0.815854705
## 687   0.718595909
## 688   0.226536635
## 689  -0.185141320
## 690  -0.225319930
## 691   2.051474015
## 692   0.081930661
## 693   0.164758869
## 694   2.497439686
## 695   0.603012035
## 696  -0.189690862
## 697   1.696728571
## 698  -0.720835363
## 699   2.027984634
## 700   1.459877701
## 701   0.696609471
## 702  -0.721716516
## 703   0.048360163
## 704   0.393796794
## 705  -1.755181654
## 706  -0.051842871
## 707  -0.389669225
## 708  -1.631582816
## 709  -0.059608246
## 710  -1.934236467
## 711   1.726220000
## 712  -0.866485998
## 713   1.619963648
## 714   1.066288744
## 715   0.400866043
## 716  -0.870347908
## 717  -0.914628339
## 718  -0.606696995
## 719  -0.445570894
## 720   0.170835033
## 721  -2.556060761
## 722   1.103275192
## 723  -0.525506327
## 724   0.449875573
## 725   0.881343730
## 726   1.483039209
## 727   0.969651191
## 728   0.181037815
## 729   2.034551888
## 730   0.063197971
## 731   0.104477197
## 732   1.877358926
## 733   0.706394960
## 734   1.176207851
## 735  -0.107092458
## 736  -1.094645454
## 737   2.939907043
## 738  -2.504900042
## 739   1.239846959
## 740  -0.757203645
## 741  -1.760355814
## 742   0.820225213
## 743   1.918562027
## 744  -1.526603708
## 745   0.083526005
## 746   0.996418676
## 747  -0.681905109
## 748  -1.770321066
## 749   0.288669820
## 750  -1.868900325
## 751  -1.989695124
## 752   0.124262050
## 753   1.462549172
## 754   0.734476568
## 755  -1.821792318
## 756  -0.342428805
## 757  -1.226521866
## 758  -0.752288755
## 759  -0.250539843
## 760  -0.080467916
## 761  -0.850680515
## 762   0.254765805
## 763  -2.429581135
## 764   0.254355912
## 765  -1.063681631
## 766  -0.755950591
## 767  -0.011725209
## 768   1.626319970
## 769   0.816498396
## 770  -0.063424362
## 771   0.194962878
## 772  -0.755063086
## 773  -0.950385658
## 774  -0.394331465
## 775  -0.867196340
## 776  -0.859227816
## 777  -0.488825415
## 778  -1.189222372
## 779   0.416281085
## 780  -0.397688521
## 781   1.364494602
## 782  -0.728763088
## 783  -1.336964849
## 784   0.367968968
## 785  -0.086544135
## 786  -0.548694805
## 787   1.484784844
## 788  -2.564193087
## 789  -0.564616046
## 790  -0.638872404
## 791   0.490072882
## 792   1.775037414
## 793  -1.289120504
## 794   0.460250288
## 795  -0.619102031
## 796  -2.026525204
## 797   1.783178794
## 798  -1.661641768
## 799  -0.281427599
## 800  -0.612331592
## 801   0.004420089
## 802  -0.220760036
## 803   1.144859049
## 804   0.821801377
## 805  -0.533529673
## 806   0.597711120
## 807   1.069631254
## 808   0.293647036
## 809  -0.123130359
## 810   0.547331478
## 811   1.012672045
## 812   1.554722761
## 813  -0.934047114
## 814   0.275501094
## 815  -1.987275803
## 816  -1.132178004
## 817   0.172426294
## 818  -0.031082552
## 819   0.431951824
## 820  -0.772792957
## 821  -0.603386553
## 822   2.240202419
## 823   1.358974395
## 824  -0.261682480
## 825   3.097457255
## 826   0.994648630
## 827  -1.757546155
## 828   1.409555292
## 829  -0.866142252
## 830  -1.607689400
## 831  -0.643157802
## 832  -1.226861359
## 833   0.667847361
## 834  -0.313602678
## 835   0.188409413
## 836  -0.809668114
## 837  -1.351211428
## 838  -0.012867558
## 839  -1.417333299
## 840   0.369330643
## 841  -0.488762634
## 842   0.630841116
## 843   1.155506261
## 844   0.138404661
## 845  -0.101372468
## 846  -1.459435277
## 847   0.237955671
## 848   0.221830460
## 849   0.601435239
## 850  -2.833039807
## 851  -0.533029967
## 852  -0.039378143
## 853  -1.872665385
## 854  -1.089919508
## 855  -2.205964150
## 856   0.173441034
## 857   0.166401374
## 858   0.192798841
## 859  -0.339031981
## 860  -0.244545250
## 861   0.213431408
## 862   1.452133166
## 863  -0.021639329
## 864  -1.978052105
## 865  -1.935636088
## 866   0.751227416
## 867  -0.714584917
## 868   0.100516463
## 869  -0.898402630
## 870   1.447961584
## 871   1.585675496
## 872  -2.287652806
## 873  -0.921265906
## 874   1.961378227
## 875  -0.443125942
## 876  -1.205678860
## 877   0.021129565
## 878   0.435201140
## 879  -0.346978134
## 880  -1.290494087
## 881  -0.896069258
## 882  -0.945028575
## 883  -1.691092829
## 884  -1.244465004
## 885   0.708017294
## 886   0.699015858
## 887  -0.972618549
## 888   0.075308011
## 889   1.233531067
## 890  -0.329665902
## 891   0.863852012
## 892  -1.707238117
## 893   2.910581394
## 894   0.171805499
## 895  -0.140186081
## 896  -0.113353096
## 897   0.476364632
## 898  -2.480888306
## 899   0.927149041
## 900   0.373536876
## 901  -2.652352788
## 902  -0.934067998
## 903  -0.964652741
## 904   0.481600049
## 905   0.006464947
## 906   1.852095319
## 907  -0.395718836
## 908  -0.443695375
## 909   1.018034335
## 910   0.001057433
## 911  -1.205598728
## 912  -1.878409481
## 913   1.596025583
## 914  -1.023160925
## 915  -0.519491998
## 916  -0.566583315
## 917   1.871507018
## 918  -0.711719455
## 919   1.912123842
## 920   0.808260799
## 921   0.029964896
## 922  -1.542516790
## 923   1.278210098
## 924  -0.190346871
## 925  -0.445131837
## 926   1.685509504
## 927  -0.193387725
## 928  -1.248783548
## 929  -1.653313512
## 930  -1.554606595
## 931   0.116393524
## 932  -0.471691423
## 933   2.135557947
## 934   0.733643860
## 935   0.582943510
## 936   0.133744657
## 937  -1.169323861
## 938   0.339462468
## 939   1.254587237
## 940  -0.238632088
## 941   1.485495001
## 942  -0.335547877
## 943   0.180477181
## 944  -1.492114220
## 945  -0.669459666
## 946   0.264192297
## 947   0.874221112
## 948  -0.483675837
## 949  -1.235039833
## 950  -1.291368977
## 951   1.919169717
## 952   1.101451252
## 953  -1.077541980
## 954  -0.258659648
## 955   0.677936741
## 956   0.529092863
## 957  -2.351107530
## 958   1.007512705
## 959  -1.628978531
## 960  -0.369508056
## 961  -0.200375374
## 962  -0.494348862
## 963   1.537001223
## 964  -0.397808552
## 965   0.067427477
## 966   0.380899811
## 967  -0.545020925
## 968  -1.102667484
## 969   1.246075457
## 970  -1.973939016
## 971   0.260837157
## 972   1.945531451
## 973  -0.932549256
## 974  -0.286800759
## 975   0.574323408
## 976   1.380762680
## 977   0.578885103
## 978  -0.756149744
## 979  -0.790814231
## 980   0.909135112
## 981   0.754412038
## 982   1.161829134
## 983  -0.258049640
## 984  -0.536185285
## 985   0.183208745
## 986   1.113036757
## 987   0.592106796
## 988   1.112276773
## 989  -0.347926905
## 990  -1.465245097
## 991   0.082845359
## 992  -0.813846057
## 993   1.043917659
## 994   0.508484105
## 995  -2.857924863
## 996  -0.379910949
## 997  -1.070004116
## 998   0.001141954
## 999   1.799618046
## 1000  0.977951635

## End(No test)

[Package mirt version 1.9 Index]