imputeMissing {mirt}	R Documentation

Imputing plausible data for missing values

Description

Given an estimated model from any of mirt's model fitting functions and an estimate of the latent trait, impute plausible missing data values. Returns the original data in a data.frame without any NA values. If a list of Theta values is supplied then a list of complete datasets is returned instead.

Usage

imputeMissing(x, Theta, ...)

Arguments

x
an estimated model x from the mirt package

Theta
a matrix containing the estimates of the latent trait scores (e.g., via fscores). Can also be a list input containing different estimates for Theta (e.g., plausible value draws)

...
additional arguments to pass

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

Examples

## No test: 
dat <- expand.table(LSAT7)
(original <- mirt(dat, 1))

## 
## Call:
## mirt(data = dat, model = 1)
## 
## Full-information item factor analysis with 1 factor(s).
## Converged within 1e-04 tolerance after 28 EM iterations.
## mirt version: 1.10 
## M-step optimizer: BFGS 
## EM acceleration: Ramsay
## Number of rectangular quadrature: 61
## 
## Log-likelihood = -2658.805
## AIC = 5337.61; AICc = 5337.833
## BIC = 5386.688; SABIC = 5354.927
## G2 (21) = 31.7, p = 0.0628
## RMSEA = 0.023, CFI = 0.939, TLI = 0.924

NAperson <- sample(1:nrow(dat), 20, replace = TRUE)
NAitem <- sample(1:ncol(dat), 20, replace = TRUE)
for(i in 1:20)
    dat[NAperson[i], NAitem[i]] <- NA
(mod <- mirt(dat, 1))

## 
## Call:
## mirt(data = dat, model = 1)
## 
## Full-information item factor analysis with 1 factor(s).
## Converged within 1e-04 tolerance after 34 EM iterations.
## mirt version: 1.10 
## M-step optimizer: BFGS 
## EM acceleration: Ramsay
## Number of rectangular quadrature: 61
## 
## Log-likelihood = -2650.772
## AIC = 5321.544; AICc = 5321.767
## BIC = 5370.622; SABIC = 5338.861

scores <- fscores(mod, method = 'MAP', full.scores = TRUE)

#re-estimate imputed dataset (good to do this multiple times and average over)
fulldata <- imputeMissing(mod, scores)
(fullmod <- mirt(fulldata, 1))

## 
## Call:
## mirt(data = fulldata, model = 1)
## 
## Full-information item factor analysis with 1 factor(s).
## Converged within 1e-04 tolerance after 30 EM iterations.
## mirt version: 1.10 
## M-step optimizer: BFGS 
## EM acceleration: Ramsay
## Number of rectangular quadrature: 61
## 
## Log-likelihood = -2661.103
## AIC = 5342.207; AICc = 5342.429
## BIC = 5391.284; SABIC = 5359.524
## G2 (21) = 31.5, p = 0.0657
## RMSEA = 0.022, CFI = 0.939, TLI = 0.924

#with multipleGroup
group <- rep(c('group1', 'group2'), each=500)
mod2 <- multipleGroup(dat, 1, group, TOL=1e-2)
fs <- fscores(mod2, full.scores=TRUE)
fulldata2 <- imputeMissing(mod2, fs)

#supply list of plausible value estimates (the best approach when Theta's are imprecise)
pv <- fscores(mod, plausible.draws = 5)
fulldata_list <- imputeMissing(mod, pv)
str(fulldata_list)

## List of 5
##  $ : int [1:1000, 1:5] 0 0 0 0 0 0 0 0 0 0 ...
##   ..- attr(*, "dimnames")=List of 2
##   .. ..$ : chr [1:1000] "1" "2" "3" "4" ...
##   .. ..$ : chr [1:5] "Item.1" "Item.2" "Item.3" "Item.4" ...
##  $ : int [1:1000, 1:5] 0 0 0 0 0 0 0 0 0 0 ...
##   ..- attr(*, "dimnames")=List of 2
##   .. ..$ : chr [1:1000] "1" "2" "3" "4" ...
##   .. ..$ : chr [1:5] "Item.1" "Item.2" "Item.3" "Item.4" ...
##  $ : int [1:1000, 1:5] 0 0 0 0 0 0 0 0 0 0 ...
##   ..- attr(*, "dimnames")=List of 2
##   .. ..$ : chr [1:1000] "1" "2" "3" "4" ...
##   .. ..$ : chr [1:5] "Item.1" "Item.2" "Item.3" "Item.4" ...
##  $ : int [1:1000, 1:5] 0 0 0 0 0 0 0 0 0 0 ...
##   ..- attr(*, "dimnames")=List of 2
##   .. ..$ : chr [1:1000] "1" "2" "3" "4" ...
##   .. ..$ : chr [1:5] "Item.1" "Item.2" "Item.3" "Item.4" ...
##  $ : int [1:1000, 1:5] 0 0 0 0 0 0 0 0 0 0 ...
##   ..- attr(*, "dimnames")=List of 2
##   .. ..$ : chr [1:1000] "1" "2" "3" "4" ...
##   .. ..$ : chr [1:5] "Item.1" "Item.2" "Item.3" "Item.4" ...

## End(No test)

[Package mirt version 1.10 Index]