Function to calculate the marginal reliability

Given an estimated model and a prior density function, compute the marginal reliability (Thissen and Wainer, 2001). This is only available for unidimensional tests.

Usage

marginal_rxx(mod, density = dnorm, ...)

Arguments

mod

an object of class 'SingleGroupClass'

density

a density function to use for integration. Default assumes the latent traits are from a normal (Gaussian) distribution

...

additional arguments passed to the density function

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi:10.18637/jss.v048.i06

Thissen, D. and Wainer, H. (2001). Test Scoring. Lawrence Erlbaum Associates.

See also

Author

Phil Chalmers rphilip.chalmers@gmail.com

Examples

dat <- expand.table(deAyala)
mod <- mirt(dat)
#> Iteration: 1, Log-Lik: -56256.669, Max-Change: 0.33820Iteration: 2, Log-Lik: -55598.944, Max-Change: 0.25622Iteration: 3, Log-Lik: -55344.183, Max-Change: 0.17901Iteration: 4, Log-Lik: -55252.392, Max-Change: 0.12170Iteration: 5, Log-Lik: -55219.175, Max-Change: 0.08221Iteration: 6, Log-Lik: -55206.845, Max-Change: 0.05582Iteration: 7, Log-Lik: -55202.096, Max-Change: 0.03866Iteration: 8, Log-Lik: -55200.153, Max-Change: 0.02623Iteration: 9, Log-Lik: -55199.330, Max-Change: 0.01879Iteration: 10, Log-Lik: -55198.648, Max-Change: 0.00688Iteration: 11, Log-Lik: -55198.588, Max-Change: 0.00523Iteration: 12, Log-Lik: -55198.554, Max-Change: 0.00418Iteration: 13, Log-Lik: -55198.501, Max-Change: 0.00094Iteration: 14, Log-Lik: -55198.498, Max-Change: 0.00025Iteration: 15, Log-Lik: -55198.497, Max-Change: 0.00024Iteration: 16, Log-Lik: -55198.497, Max-Change: 0.00021Iteration: 17, Log-Lik: -55198.497, Max-Change: 0.00096Iteration: 18, Log-Lik: -55198.496, Max-Change: 0.00057Iteration: 19, Log-Lik: -55198.496, Max-Change: 0.00015Iteration: 20, Log-Lik: -55198.496, Max-Change: 0.00007

# marginal estimate treating item parameters as known
marginal_rxx(mod)
#> [1] 0.6092894

# compare to alpha
itemstats(dat)$overall$alpha
#> [1] 0.6077281

# \donttest{

# empirical estimate (assuming the same prior)
fscores(mod, returnER = TRUE)
#>        F1 
#> 0.6200703 

# empirical rxx the alternative way, given theta scores and SEs
fs <- fscores(mod, full.scores.SE=TRUE)
head(fs)
#>             F1     SE_F1
#> [1,] -1.580133 0.6699424
#> [2,] -1.580133 0.6699424
#> [3,] -1.580133 0.6699424
#> [4,] -1.580133 0.6699424
#> [5,] -1.580133 0.6699424
#> [6,] -1.580133 0.6699424
empirical_rxx(fs)
#>        F1 
#> 0.6200703 

#############
# example demonstrating correlation attenuation

theta <- rnorm(1000)
X <- theta + rnorm(1000, sd=2)
cor(X, theta)    # correlation without measurement error (what you want)
#> [1] 0.4644682

# measured with a 10 item GRM test
nitems <- 10
a <- matrix(rlnorm(nitems,.2,.3))
diffs <- t(apply(matrix(runif(nitems*4, .3, 1), nitems), 1, cumsum))
diffs <- -(diffs - rowMeans(diffs))
d <- diffs + rnorm(nitems)
dat <- simdata(a, d, itemtype = 'graded', Theta=matrix(theta))

# correlation with total score (attenuated)
cor(rowSums(dat), X)
#> [1] 0.4279283

# fit single group model
mod <- mirt(dat)
#> Iteration: 1, Log-Lik: -12638.902, Max-Change: 0.49082Iteration: 2, Log-Lik: -12565.042, Max-Change: 0.24076Iteration: 3, Log-Lik: -12523.393, Max-Change: 0.14515Iteration: 4, Log-Lik: -12516.473, Max-Change: 0.10364Iteration: 5, Log-Lik: -12511.817, Max-Change: 0.06581Iteration: 6, Log-Lik: -12510.110, Max-Change: 0.04295Iteration: 7, Log-Lik: -12509.178, Max-Change: 0.01531Iteration: 8, Log-Lik: -12509.038, Max-Change: 0.00953Iteration: 9, Log-Lik: -12508.990, Max-Change: 0.00623Iteration: 10, Log-Lik: -12508.952, Max-Change: 0.00168Iteration: 11, Log-Lik: -12508.948, Max-Change: 0.00123Iteration: 12, Log-Lik: -12508.946, Max-Change: 0.00093Iteration: 13, Log-Lik: -12508.944, Max-Change: 0.00034Iteration: 14, Log-Lik: -12508.943, Max-Change: 0.00011Iteration: 15, Log-Lik: -12508.943, Max-Change: 0.00027Iteration: 16, Log-Lik: -12508.943, Max-Change: 0.00011Iteration: 17, Log-Lik: -12508.943, Max-Change: 0.00022Iteration: 18, Log-Lik: -12508.943, Max-Change: 0.00011Iteration: 19, Log-Lik: -12508.943, Max-Change: 0.00009

# EAP correlation (also attenuated)
fs <- fscores(mod)
cor(fs, X)
#>         [,1]
#> F1 0.4302499

# correction for attenuation, r_x.theta = r_x.theta.hat / sqrt(rxx_theta.hat)
(rxx <- marginal_rxx(mod))  # alternatively, could use empirical_rxx()
#> [1] 0.8095505
cor(fs, X) / sqrt(rxx)  # correction estimate
#>         [,1]
#> F1 0.4781881
cor(X, theta)           # compare to true correlation
#> [1] 0.4644682

# }