mixedmirt {mirt}	R Documentation

Mixed effects modeling for MIRT models

Description

mixedmirt fits MIRT models using FIML estimation to dichotomous and polytomous IRT models conditional on fixed and random effect of person and item level covariates. This can also be understood as 'explanatory IRT' if only fixed effects are modeled, or multilevel/mixed IRT if random and fixed effects are included. The method uses the MH-RM algorithm exclusively. Additionally, computation of the log-likelihood can be sped up by using parallel estimation via mirtCluster.

Usage

mixedmirt(data, covdata = NULL, model, fixed = ~1, random = NULL,
  itemtype = "Rasch", lr.fixed = ~1, lr.random = NULL,
  itemdesign = NULL, constrain = NULL, pars = NULL,
  return.design = FALSE, SE = TRUE, internal_constraints = TRUE,
  technical = list(SEtol = 1e-04), ...)

Arguments

data
a matrix or data.frame that consists of numerically ordered data, with missing data coded as NA

covdata
a data.frame that consists of the nrow(data) by K 'person level' fixed and random predictors

model
an object returned from mirt.model() declaring how the factor model is to be estimated. See mirt.model for more details

fixed
a right sided R formula for specifying the fixed effect (aka 'explanatory') predictors from covdata and itemdesign. To estimate the intercepts for each item the keyword items is reserved and automatically added to the itemdesign input. If any polytomous items are being model the items are argument is not valid since all intercept parameters are freely estimated and identified with the parameterizations found in mirt, and the first column in the fixed design matrix (commonly the intercept or a reference group) is omitted

random
a right sided formula or list of formulas containing crossed random effects of the form v1 + ... v_n | G, where G is the grouping variable and v_n are random numeric predictors within each group. If no intercept value is specified then by default the correlations between the v's and G are estimated, but can be suppressed by including the ~ -1 + ... or 0 constant. G may contain interaction terms, such as group:items to include cross or person-level interactions effects

itemtype
same as itemtype in mirt, except does not support the following item types: c('PC2PL', 'PC3PL', '2PLNRM', '3PLNRM', '3PLuNRM', '4PLNRM')

lr.fixed
an R forumala (or list of formulas) to specificy regression effects in the latent variables from the variables in covdata. This is used to construct models such as the so-called 'latent regression model' to expalin person-level ability/trait differences. If a named list of formulas is supplied (where the names correspond to the latent trait names in model) then specific regression effects can be estimated for each factor. Supplying a single formula will estimate the regression parameters for all latent traits by default

lr.random
(CURRENTLY DISABLED) a list of random effect terms for modeling variability in the latent trait scores, where the syntax uses the same style as in the random argument. Useful for building so-called 'multilevel IRT' models which are non-Rasch (multilevel Rasch models do not require these, and can be built using the fixed and random inputs alone)

itemdesign
a data.frame object used to create a design matrix for the items, where each nrow(itemdesign) == nitems and the number of columns is equal to the number of fixed effect predictors (i.e., item intercepts). By default an items variable is reserved for modeling the item intercept parameters

constrain
a list indicating parameter equality constrains. See mirt for more detail

pars
used for parameter starting values. See mirt for more detail

return.design
logical; return the design matrices before they have (potentially) been reassigned?

SE
logical; compute the standard errors by approximating the information matrix using the MHRM algorithm? Default is TRUE

internal_constraints
logical; use the internally defined constraints for constraining effects across persons and items? Default is TRUE. Setting this to FALSE runs the risk of underidentification

technical
the technical list passed to the MH-RM estimation engine, with the SEtol default increased to .0001. See mirt for further details

...
additional arguments to be passed to the MH-RM estimation engine. See mirt for more details and examples

Details

For dichotomous response models, mixedmirt follows the general form

P(x = 1|θ, ψ) = g + \frac{(u - g)}{1 + exp(-1 * [θ a + X β + Z δ])}

where X is a design matrix with associated β fixed effect intercept coefficients, and Z is a design matrix with associated δ random effects for the intercepts. For simplicity and easier interpretation, the unique item intercept values typically found in X β are extracted and reassigned within mirt's 'intercept' parameters (e.g., 'd'). To observe how the design matrices are structured prior to reassignment and estimation pass the argument return.design = TRUE.

Polytomous IRT models follow a similar format except the item intercepts are automatically estimated internally, rendering the items argument in the fixed formula redundant and therefore must be omitted from the specification. If there are a mixture of dichotomous and polytomous items the intercepts for the dichotomous models are also estimated for consistency.

The decomposition of the θ parameters is also possible to form latent regression and multilevel IRT models by using the lr.fixed and lr.random inputs. These effects decompose θ such that

θ = V Γ + W ζ + ε

where V and W are fixed and random effects design matricies for the associated coefficients.

To simulate maximum a posteriori estimates for the random effect terms use the randef function.

Value

function returns an object of class MixedClass (MixedClass-class).

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

References

Chalmers, R. P. (2015, in press). Extended Mixed-Effects Item Response Models with the MH-RM Algorithm. Journal of Educational Measurement.

See Also

mirt, randef, fixef, boot.mirt

Examples

## No test: 

#make some data
set.seed(1234)
N <- 750
a <- matrix(rlnorm(10,.3,1),10,1)
d <- matrix(rnorm(10), 10)
Theta <- matrix(sort(rnorm(N)))
pseudoIQ <- Theta * 5 + 100  + rnorm(N, 0 , 5)
pseudoIQ <- (pseudoIQ - mean(pseudoIQ))/10  #rescale variable for numerical stability
group <- factor(rep(c('G1','G2','G3'), each = N/3))
data <- simdata(a,d,N, itemtype = rep('dich',10), Theta=Theta)
covdata <- data.frame(group, pseudoIQ)
#use parallel computing
mirtCluster()

## mirtCluster() has already been defined

#specify IRT model
model <- mirt.model('Theta = 1-10')

#model with no person predictors
mod0 <- mirt(data, model, itemtype = 'Rasch')

#group as a fixed effect predictor (aka, uniform dif)
mod1 <- mixedmirt(data, covdata, model, fixed = ~ 0 + group + items)
anova(mod0, mod1)

## 
## Model 1: mirt(data = data, model = model, itemtype = "Rasch")
## Model 2: mixedmirt(data = data, covdata = covdata, model = model, fixed = ~0 + 
##     group + items)

##        AIC     AICc    SABIC      BIC    logLik      X2  df   p
## 1 8826.167 8826.525 8842.059 8876.988 -4402.084     NaN NaN NaN
## 2 8117.443 8117.938 8136.224 8177.504 -4045.722 712.724   2   0

summary(mod1)

## 
## Call:
## mixedmirt(data = data, covdata = covdata, model = model, fixed = ~0 + 
##     group + items)
## 
## --------------
## FIXED EFFECTS:
##         Estimate Std.Error z.value
## groupG1   -1.829     0.099 -18.411
## groupG2   -0.659     0.092  -7.149
## groupG3    0.493     0.092   5.361
## 
## --------------
## RANDOM EFFECT COVARIANCE(S):
## Correlations on upper diagonal
## 
## $Theta
##        Theta
## Theta 0.0767

coef(mod1)

## $Item_1
##         groupG1 groupG2 groupG3 a1  d  g  u
## par      -1.829  -0.659   0.493  1  0  0  1
## CI_2.5   -2.024  -0.839   0.313 NA NA NA NA
## CI_97.5  -1.635  -0.478   0.674 NA NA NA NA
## 
## $Item_2
##         groupG1 groupG2 groupG3 a1      d  g  u
## par      -1.829  -0.659   0.493  1 -0.170  0  1
## CI_2.5   -2.024  -0.839   0.313 NA -0.403 NA NA
## CI_97.5  -1.635  -0.478   0.674 NA  0.064 NA NA
## 
## $Item_3
##         groupG1 groupG2 groupG3 a1     d  g  u
## par      -1.829  -0.659   0.493  1 0.347  0  1
## CI_2.5   -2.024  -0.839   0.313 NA 0.118 NA NA
## CI_97.5  -1.635  -0.478   0.674 NA 0.576 NA NA
## 
## $Item_4
##         groupG1 groupG2 groupG3 a1     d  g  u
## par      -1.829  -0.659   0.493  1 0.684  0  1
## CI_2.5   -2.024  -0.839   0.313 NA 0.455 NA NA
## CI_97.5  -1.635  -0.478   0.674 NA 0.912 NA NA
## 
## $Item_5
##         groupG1 groupG2 groupG3 a1     d  g  u
## par      -1.829  -0.659   0.493  1 1.315  0  1
## CI_2.5   -2.024  -0.839   0.313 NA 1.081 NA NA
## CI_97.5  -1.635  -0.478   0.674 NA 1.548 NA NA
## 
## $Item_6
##         groupG1 groupG2 groupG3 a1     d  g  u
## par      -1.829  -0.659   0.493  1 0.578  0  1
## CI_2.5   -2.024  -0.839   0.313 NA 0.350 NA NA
## CI_97.5  -1.635  -0.478   0.674 NA 0.807 NA NA
## 
## $Item_7
##         groupG1 groupG2 groupG3 a1      d  g  u
## par      -1.829  -0.659   0.493  1  0.097  0  1
## CI_2.5   -2.024  -0.839   0.313 NA -0.134 NA NA
## CI_97.5  -1.635  -0.478   0.674 NA  0.327 NA NA
## 
## $Item_8
##         groupG1 groupG2 groupG3 a1      d  g  u
## par      -1.829  -0.659   0.493  1 -0.407  0  1
## CI_2.5   -2.024  -0.839   0.313 NA -0.644 NA NA
## CI_97.5  -1.635  -0.478   0.674 NA -0.170 NA NA
## 
## $Item_9
##         groupG1 groupG2 groupG3 a1      d  g  u
## par      -1.829  -0.659   0.493  1 -0.228  0  1
## CI_2.5   -2.024  -0.839   0.313 NA -0.462 NA NA
## CI_97.5  -1.635  -0.478   0.674 NA  0.006 NA NA
## 
## $Item_10
##         groupG1 groupG2 groupG3 a1     d  g  u
## par      -1.829  -0.659   0.493  1 3.588  0  1
## CI_2.5   -2.024  -0.839   0.313 NA 3.256 NA NA
## CI_97.5  -1.635  -0.478   0.674 NA 3.920 NA NA
## 
## $GroupPars
##         MEAN_1 COV_11
## par          0  0.077
## CI_2.5      NA  0.044
## CI_97.5     NA  0.109

#same model as above in lme4
wide <- data.frame(id=1:nrow(data),data,covdata)
long <- reshape2::melt(wide, id.vars = c('id', 'group', 'pseudoIQ'))
library(lme4)

## Loading required package: Matrix
## Loading required package: Rcpp
## 
## Attaching package: 'lme4'
## 
## The following object is masked from 'package:mirt':
## 
##     fixef

lmod0 <- glmer(value ~ 0 + variable + (1|id), long, family = binomial)

## Warning in checkConv(attr(opt, "derivs"), opt$par, ctrl =
## control$checkConv, : Model failed to converge with max|grad| = 0.00153862
## (tol = 0.001, component 6)

lmod1 <- glmer(value ~ 0 + group + variable + (1|id), long, family = binomial)
anova(lmod0, lmod1)

## Data: long
## Models:
## lmod0: value ~ 0 + variable + (1 | id)
## lmod1: value ~ 0 + group + variable + (1 | id)
##       Df    AIC    BIC  logLik deviance  Chisq Chi Df Pr(>Chisq)    
## lmod0 11 8839.4 8915.5 -4408.7   8817.4                             
## lmod1 13 8116.0 8206.0 -4045.0   8090.0 727.38      2  < 2.2e-16 ***
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1

#model using 2PL items instead of Rasch
mod1b <- mixedmirt(data, covdata, model, fixed = ~ 0 + group + items, itemtype = '2PL')
anova(mod1, mod1b) #better with 2PL models using all criteria (as expected, given simdata pars)

## 
## Model 1: mixedmirt(data = data, covdata = covdata, model = model, fixed = ~0 + 
##     group + items)
## Model 2: mixedmirt(data = data, covdata = covdata, model = model, fixed = ~0 + 
##     group + items, itemtype = "2PL")

##        AIC     AICc    SABIC      BIC    logLik      X2  df   p
## 1 8117.443 8117.938 8136.224 8177.504 -4045.722     NaN NaN NaN
## 2 7968.583 7969.975 8000.366 8070.224 -3962.291 166.861   9   0

#continuous predictor with group
mod2 <- mixedmirt(data, covdata, model, fixed = ~ 0 + group + items + pseudoIQ)
summary(mod2)

## 
## Call:
## mixedmirt(data = data, covdata = covdata, model = model, fixed = ~0 + 
##     group + items + pseudoIQ)
## 
## --------------
## FIXED EFFECTS:
##          Estimate Std.Error z.value
## groupG1    -1.715     0.104 -16.530
## groupG2    -0.656     0.092  -7.127
## groupG3     0.375     0.097   3.856
## pseudoIQ    0.215     0.056   3.862
## 
## --------------
## RANDOM EFFECT COVARIANCE(S):
## Correlations on upper diagonal
## 
## $Theta
##        Theta
## Theta 0.0693

anova(mod1b, mod2)

## 
## Model 1: mixedmirt(data = data, covdata = covdata, model = model, fixed = ~0 + 
##     group + items + pseudoIQ)
## Model 2: mixedmirt(data = data, covdata = covdata, model = model, fixed = ~0 + 
##     group + items, itemtype = "2PL")

##        AIC     AICc    SABIC      BIC    logLik      X2  df   p
## 1 8104.990 8105.561 8125.216 8169.671 -4038.495     NaN NaN NaN
## 2 7968.583 7969.975 8000.366 8070.224 -3962.291 152.408   8   0

#view fixed design matrix with and without unique item level intercepts
withint <- mixedmirt(data, covdata, model, fixed = ~ 0 + items + group, return.design = TRUE)
withoutint <- mixedmirt(data, covdata, model, fixed = ~ 0 + group, return.design = TRUE)

#notice that in result above, the intercepts 'items1 to items 10' were reassigned to 'd'
head(withint$X)

##     items1 items2 items3 items4 items5 items6 items7 items8 items9 items10
## 1.1      1      0      0      0      0      0      0      0      0       0
## 2.1      1      0      0      0      0      0      0      0      0       0
## 3.1      1      0      0      0      0      0      0      0      0       0
## 4.1      1      0      0      0      0      0      0      0      0       0
## 5.1      1      0      0      0      0      0      0      0      0       0
## 6.1      1      0      0      0      0      0      0      0      0       0
##     groupG2 groupG3
## 1.1       0       0
## 2.1       0       0
## 3.1       0       0
## 4.1       0       0
## 5.1       0       0
## 6.1       0       0

tail(withint$X)

##        items1 items2 items3 items4 items5 items6 items7 items8 items9
## 745.10      0      0      0      0      0      0      0      0      0
## 746.10      0      0      0      0      0      0      0      0      0
## 747.10      0      0      0      0      0      0      0      0      0
## 748.10      0      0      0      0      0      0      0      0      0
## 749.10      0      0      0      0      0      0      0      0      0
## 750.10      0      0      0      0      0      0      0      0      0
##        items10 groupG2 groupG3
## 745.10       1       0       1
## 746.10       1       0       1
## 747.10       1       0       1
## 748.10       1       0       1
## 749.10       1       0       1
## 750.10       1       0       1

head(withoutint$X) #no intercepts design here to be reassigned into item intercepts

##     groupG1 groupG2 groupG3
## 1.1       1       0       0
## 2.1       1       0       0
## 3.1       1       0       0
## 4.1       1       0       0
## 5.1       1       0       0
## 6.1       1       0       0

tail(withoutint$X)

##        groupG1 groupG2 groupG3
## 745.10       0       0       1
## 746.10       0       0       1
## 747.10       0       0       1
## 748.10       0       0       1
## 749.10       0       0       1
## 750.10       0       0       1

###################################################
### random effects
#make the number of groups much larger
covdata$group <- factor(rep(paste0('G',1:50), each = N/50))

#random groups
rmod1 <- mixedmirt(data, covdata, 1, fixed = ~ 0 + items, random = ~ 1|group)
summary(rmod1)

## 
## Call:
## mixedmirt(data = data, covdata = covdata, model = 1, fixed = ~0 + 
##     items, random = ~1 | group)
## 
## 
## --------------
## RANDOM EFFECT COVARIANCE(S):
## Correlations on upper diagonal
## 
## $Theta
##        F1
## F1 0.0523
## 
## $group
##           COV_group
## COV_group      1.05

coef(rmod1)

## $Item_1
##         a1      d  g  u
## par      1 -0.667  0  1
## CI_2.5  NA -0.867 NA NA
## CI_97.5 NA -0.467 NA NA
## 
## $Item_2
##         a1      d  g  u
## par      1 -0.839  0  1
## CI_2.5  NA -1.043 NA NA
## CI_97.5 NA -0.635 NA NA
## 
## $Item_3
##         a1      d  g  u
## par      1 -0.316  0  1
## CI_2.5  NA -0.512 NA NA
## CI_97.5 NA -0.120 NA NA
## 
## $Item_4
##         a1      d  g  u
## par      1  0.026  0  1
## CI_2.5  NA -0.169 NA NA
## CI_97.5 NA  0.220 NA NA
## 
## $Item_5
##         a1     d  g  u
## par      1 0.666  0  1
## CI_2.5  NA 0.466 NA NA
## CI_97.5 NA 0.866 NA NA
## 
## $Item_6
##         a1      d  g  u
## par      1 -0.081  0  1
## CI_2.5  NA -0.276 NA NA
## CI_97.5 NA  0.114 NA NA
## 
## $Item_7
##         a1      d  g  u
## par      1 -0.569  0  1
## CI_2.5  NA -0.768 NA NA
## CI_97.5 NA -0.370 NA NA
## 
## $Item_8
##         a1      d  g  u
## par      1 -1.080  0  1
## CI_2.5  NA -1.289 NA NA
## CI_97.5 NA -0.870 NA NA
## 
## $Item_9
##         a1      d  g  u
## par      1 -0.898  0  1
## CI_2.5  NA -1.103 NA NA
## CI_97.5 NA -0.693 NA NA
## 
## $Item_10
##         a1     d  g  u
## par      1 2.990  0  1
## CI_2.5  NA 2.673 NA NA
## CI_97.5 NA 3.307 NA NA
## 
## $GroupPars
##         MEAN_1 COV_11
## par          0  0.052
## CI_2.5      NA    NaN
## CI_97.5     NA    NaN
## 
## $group
##         COV_group_group
## par               1.050
## CI_2.5            0.623
## CI_97.5           1.477

#random groups and random items
rmod2 <- mixedmirt(data, covdata, 1, random = list(~ 1|group, ~ 1|items))
summary(rmod2)

## 
## Call:
## mixedmirt(data = data, covdata = covdata, model = 1, random = list(~1 | 
##     group, ~1 | items))
## 
## --------------
## FIXED EFFECTS:
##             Estimate Std.Error z.value
## (Intercept)   -0.574       NaN     NaN
## 
## --------------
## RANDOM EFFECT COVARIANCE(S):
## Correlations on upper diagonal
## 
## $Theta
##        F1
## F1 0.0564
## 
## $group
##           COV_group
## COV_group     0.887
## 
## $items
##           COV_items
## COV_items      1.07

eff <- randef(rmod2) #estimate random effects

#random slopes with fixed intercepts (suppressed correlation)
rmod3 <- mixedmirt(data, covdata, 1, fixed = ~ 0 + items, random = ~ -1 + pseudoIQ|group)
summary(rmod3)

## 
## Call:
## mixedmirt(data = data, covdata = covdata, model = 1, fixed = ~0 + 
##     items, random = ~-1 + pseudoIQ | group)
## 
## 
## --------------
## RANDOM EFFECT COVARIANCE(S):
## Correlations on upper diagonal
## 
## $Theta
##        F1
## F1 0.0638
## 
## $group
##              COV_group COV_pseudoIQ
## COV_group         1.04       0.0000
## COV_pseudoIQ      0.00       0.0427

eff <- randef(rmod3)
str(eff)

## List of 2
##  $ Theta: num [1:750, 1] -0.1168 0.0209 -0.0544 0.0366 -0.0774 ...
##   ..- attr(*, "dimnames")=List of 2
##   .. ..$ : NULL
##   .. ..$ : chr "F1"
##  $ group: num [1:50, 1:2] -1.99 -1.13 -1.4 -1.17 -1.42 ...
##   ..- attr(*, "dimnames")=List of 2
##   .. ..$ : chr [1:50] "G1" "G2" "G3" "G4" ...
##   .. ..$ : chr [1:2] "group" "pseudoIQ"

###################################################
##LLTM, and 2PL version of LLTM
data(SAT12)
data <- key2binary(SAT12,
                   key = c(1,4,5,2,3,1,2,1,3,1,2,4,2,1,5,3,4,4,1,4,3,3,4,1,3,5,1,3,1,5,4,5))
model <- mirt.model('Theta = 1-32')

# Suppose that the first 16 items were suspected to be easier than the last 16 items,
#   and we wish to test this item structure hypothesis (more intercept designs are possible
#   by including more columns).
itemdesign <- data.frame(itemorder = factor(c(rep('easier', 16), rep('harder', 16))))

#notice that the 'fixed = ~ ... + items' argument is omitted
LLTM <- mixedmirt(data, model = model, fixed = ~ 0 + itemorder, itemdesign = itemdesign,
   SE = TRUE) # SE argument ensures that the information matrix is computed accurately
summary(LLTM)

## 
## Call:
## mixedmirt(data = data, model = model, fixed = ~0 + itemorder, 
##     itemdesign = itemdesign, SE = TRUE)
## 
## --------------
## FIXED EFFECTS:
##                 Estimate Std.Error z.value
## itemordereasier    0.153     0.033   4.565
## itemorderharder    0.444     0.034  13.169
## 
## --------------
## RANDOM EFFECT COVARIANCE(S):
## Correlations on upper diagonal
## 
## $Theta
##       Theta
## Theta 0.357

coef(LLTM)

## $Item.1
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.2
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.3
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.4
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.5
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.6
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.7
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.8
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.9
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.10
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.11
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.12
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.13
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.14
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.15
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.16
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.17
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.18
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.19
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.20
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.21
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.22
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.23
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.24
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.25
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.26
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.27
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.28
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.29
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.30
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.31
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $Item.32
##         itemordereasier itemorderharder a1  d  g  u
## par               0.153           0.444  1  0  0  1
## CI_2.5            0.087           0.378 NA NA NA NA
## CI_97.5           0.218           0.510 NA NA NA NA
## 
## $GroupPars
##         MEAN_1 COV_11
## par          0  0.357
## CI_2.5      NA  0.300
## CI_97.5     NA  0.414

wald(LLTM)

##                                                                                                                                     1
## infoname itemordereasier.1.7.13.19.25.31.37.43.49.55.61.67.73.79.85.91.97.103.109.115.121.127.133.139.145.151.157.163.169.175.181.187
## par                                                                                                                             0.153
##                                                                                                                                     2
## infoname itemorderharder.2.8.14.20.26.32.38.44.50.56.62.68.74.80.86.92.98.104.110.116.122.128.134.140.146.152.158.164.170.176.182.188
## par                                                                                                                             0.444
##                   3
## infoname COV_11.194
## par           0.357

L <- matrix(c(-1, 1, 0), 1)
wald(LLTM, L) #first half different from second

##          W df p
## 1 92.15335  1 0

#compare to items with estimated slopes (2PL)
twoPL <- mixedmirt(data, model = model, fixed = ~ 0 + itemorder, itemtype = '2PL',
                   itemdesign = itemdesign)
#twoPL not mixing too well (AR should be between .2 and .5), decrease MHcand
twoPL <- mixedmirt(data, model = model, fixed = ~ 0 + itemorder, itemtype = '2PL',
                  itemdesign = itemdesign, technical = list(MHcand = 0.8))
anova(twoPL, LLTM) #much better fit

## 
## Model 1: mixedmirt(data = data, model = model, fixed = ~0 + itemorder, 
##     itemdesign = itemdesign, SE = TRUE)
## Model 2: mixedmirt(data = data, model = model, fixed = ~0 + itemorder, 
##     itemtype = "2PL", itemdesign = itemdesign, technical = list(MHcand = 0.8))

##        AIC     AICc    SABIC      BIC    logLik       X2  df   p
## 1 25463.38 25463.42 25467.04 25476.57 -12728.69      NaN NaN NaN
## 2 20480.44 20484.65 20522.00 20629.94 -10206.22 5044.935  31   0

summary(twoPL)

## 
## Call:
## mixedmirt(data = data, model = model, fixed = ~0 + itemorder, 
##     itemtype = "2PL", itemdesign = itemdesign, technical = list(MHcand = 0.8))
## 
## --------------
## FIXED EFFECTS:
##                 Estimate Std.Error z.value
## itemordereasier   -1.657     0.063 -26.331
## itemorderharder   -1.632     0.070 -23.199
## 
## --------------
## RANDOM EFFECT COVARIANCE(S):
## Correlations on upper diagonal
## 
## $Theta
##       Theta
## Theta     1

coef(twoPL)

## $Item.1
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 0.929  0  0  1
## CI_2.5           -1.781          -1.770 0.711 NA NA NA
## CI_97.5          -1.534          -1.494 1.148 NA NA NA
## 
## $Item.2
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 2.536  0  0  1
## CI_2.5           -1.781          -1.770 2.289 NA NA NA
## CI_97.5          -1.534          -1.494 2.783 NA NA NA
## 
## $Item.3
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 0.997  0  0  1
## CI_2.5           -1.781          -1.770 0.781 NA NA NA
## CI_97.5          -1.534          -1.494 1.213 NA NA NA
## 
## $Item.4
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 1.287  0  0  1
## CI_2.5           -1.781          -1.770 1.055 NA NA NA
## CI_97.5          -1.534          -1.494 1.519 NA NA NA
## 
## $Item.5
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 2.714  0  0  1
## CI_2.5           -1.781          -1.770 2.457 NA NA NA
## CI_97.5          -1.534          -1.494 2.971 NA NA NA
## 
## $Item.6
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 0.418  0  0  1
## CI_2.5           -1.781          -1.770 0.195 NA NA NA
## CI_97.5          -1.534          -1.494 0.640 NA NA NA
## 
## $Item.7
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 3.937  0  0  1
## CI_2.5           -1.781          -1.770 3.698 NA NA NA
## CI_97.5          -1.534          -1.494 4.176 NA NA NA
## 
## $Item.8
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 0.510  0  0  1
## CI_2.5           -1.781          -1.770 0.293 NA NA NA
## CI_97.5          -1.534          -1.494 0.727 NA NA NA
## 
## $Item.9
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 5.495  0  0  1
## CI_2.5           -1.781          -1.770 5.282 NA NA NA
## CI_97.5          -1.534          -1.494 5.708 NA NA NA
## 
## $Item.10
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 1.632  0  0  1
## CI_2.5           -1.781          -1.770 1.390 NA NA NA
## CI_97.5          -1.534          -1.494 1.875 NA NA NA
## 
## $Item.11
##         itemordereasier itemorderharder     a1  d  g  u
## par              -1.657          -1.632 13.706  0  0  1
## CI_2.5           -1.781          -1.770 12.068 NA NA NA
## CI_97.5          -1.534          -1.494 15.345 NA NA NA
## 
## $Item.12
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 1.240  0  0  1
## CI_2.5           -1.781          -1.770 1.005 NA NA NA
## CI_97.5          -1.534          -1.494 1.475 NA NA NA
## 
## $Item.13
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 3.053  0  0  1
## CI_2.5           -1.781          -1.770 2.822 NA NA NA
## CI_97.5          -1.534          -1.494 3.285 NA NA NA
## 
## $Item.14
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 3.508  0  0  1
## CI_2.5           -1.781          -1.770 3.246 NA NA NA
## CI_97.5          -1.534          -1.494 3.769 NA NA NA
## 
## $Item.15
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 4.789  0  0  1
## CI_2.5           -1.781          -1.770   NaN NA NA NA
## CI_97.5          -1.534          -1.494   NaN NA NA NA
## 
## $Item.16
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 1.469  0  0  1
## CI_2.5           -1.781          -1.770 1.231 NA NA NA
## CI_97.5          -1.534          -1.494 1.707 NA NA NA
## 
## $Item.17
##         itemordereasier itemorderharder     a1  d  g  u
## par              -1.657          -1.632 10.048  0  0  1
## CI_2.5           -1.781          -1.770  9.330 NA NA NA
## CI_97.5          -1.534          -1.494 10.766 NA NA NA
## 
## $Item.18
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 1.412  0  0  1
## CI_2.5           -1.781          -1.770 1.191 NA NA NA
## CI_97.5          -1.534          -1.494 1.633 NA NA NA
## 
## $Item.19
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 2.161  0  0  1
## CI_2.5           -1.781          -1.770 1.906 NA NA NA
## CI_97.5          -1.534          -1.494 2.415 NA NA NA
## 
## $Item.20
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 5.832  0  0  1
## CI_2.5           -1.781          -1.770   NaN NA NA NA
## CI_97.5          -1.534          -1.494   NaN NA NA NA
## 
## $Item.21
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 6.415  0  0  1
## CI_2.5           -1.781          -1.770   NaN NA NA NA
## CI_97.5          -1.534          -1.494   NaN NA NA NA
## 
## $Item.22
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 7.925  0  0  1
## CI_2.5           -1.781          -1.770 7.430 NA NA NA
## CI_97.5          -1.534          -1.494 8.420 NA NA NA
## 
## $Item.23
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 0.979  0  0  1
## CI_2.5           -1.781          -1.770 0.754 NA NA NA
## CI_97.5          -1.534          -1.494 1.204 NA NA NA
## 
## $Item.24
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 3.646  0  0  1
## CI_2.5           -1.781          -1.770 3.416 NA NA NA
## CI_97.5          -1.534          -1.494 3.877 NA NA NA
## 
## $Item.25
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 1.294  0  0  1
## CI_2.5           -1.781          -1.770 1.063 NA NA NA
## CI_97.5          -1.534          -1.494 1.525 NA NA NA
## 
## $Item.26
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 1.907  0  0  1
## CI_2.5           -1.781          -1.770 1.683 NA NA NA
## CI_97.5          -1.534          -1.494 2.131 NA NA NA
## 
## $Item.27
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 5.728  0  0  1
## CI_2.5           -1.781          -1.770   NaN NA NA NA
## CI_97.5          -1.534          -1.494   NaN NA NA NA
## 
## $Item.28
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 2.125  0  0  1
## CI_2.5           -1.781          -1.770 1.877 NA NA NA
## CI_97.5          -1.534          -1.494 2.373 NA NA NA
## 
## $Item.29
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 1.163  0  0  1
## CI_2.5           -1.781          -1.770 0.941 NA NA NA
## CI_97.5          -1.534          -1.494 1.385 NA NA NA
## 
## $Item.30
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 1.438  0  0  1
## CI_2.5           -1.781          -1.770 1.195 NA NA NA
## CI_97.5          -1.534          -1.494 1.682 NA NA NA
## 
## $Item.31
##         itemordereasier itemorderharder    a1  d  g  u
## par              -1.657          -1.632 5.282  0  0  1
## CI_2.5           -1.781          -1.770 5.111 NA NA NA
## CI_97.5          -1.534          -1.494 5.454 NA NA NA
## 
## $Item.32
##         itemordereasier itemorderharder     a1  d  g  u
## par              -1.657          -1.632  0.088  0  0  1
## CI_2.5           -1.781          -1.770 -0.158 NA NA NA
## CI_97.5          -1.534          -1.494  0.335 NA NA NA
## 
## $GroupPars
##         MEAN_1 COV_11
## par          0      1
## CI_2.5      NA     NA
## CI_97.5     NA     NA

wald(twoPL)

##                                                                                                                                     1
## infoname itemordereasier.1.7.13.19.25.31.37.43.49.55.61.67.73.79.85.91.97.103.109.115.121.127.133.139.145.151.157.163.169.175.181.187
## par                                                                                                                            -1.657
##                                                                                                                                     2
## infoname itemorderharder.2.8.14.20.26.32.38.44.50.56.62.68.74.80.86.92.98.104.110.116.122.128.134.140.146.152.158.164.170.176.182.188
## par                                                                                                                            -1.632
##               3      4      5      6      7      8      9     10     11
## infoname   a1.3   a1.9  a1.15  a1.21  a1.27  a1.33  a1.39  a1.45  a1.51
## par       0.929  2.536  0.997  1.287  2.714  0.418  3.937  0.510  5.495
##              12     13     14     15     16     17     18     19     20
## infoname  a1.57  a1.63  a1.69  a1.75  a1.81  a1.87  a1.93  a1.99 a1.105
## par       1.632 13.706  1.240  3.053  3.508  4.789  1.469 10.048  1.412
##              21     22     23     24     25     26     27     28     29
## infoname a1.111 a1.117 a1.123 a1.129 a1.135 a1.141 a1.147 a1.153 a1.159
## par       2.161  5.832  6.415  7.925  0.979  3.646  1.294  1.907  5.728
##              30     31     32     33     34
## infoname a1.165 a1.171 a1.177 a1.183 a1.189
## par       2.125  1.163  1.438  5.282  0.088

L <- matrix(0, 1, 34)
L[1, 1] <- 1
L[1, 2] <- -1
wald(twoPL, L) #n.s., which is the correct conclusion. Rasch approach gave wrong inference

##            W df         p
## 1 0.07593067  1 0.7828897

##LLTM with item error term
LLTMwithError <- mixedmirt(data, model = model, fixed = ~ 0 + itemorder, random = ~ 1|items,
    itemdesign = itemdesign)
summary(LLTMwithError)

## 
## Call:
## mixedmirt(data = data, model = model, fixed = ~0 + itemorder, 
##     random = ~1 | items, itemdesign = itemdesign)
## 
## --------------
## FIXED EFFECTS:
##                 Estimate Std.Error z.value
## itemordereasier    0.206     0.066   3.115
## itemorderharder    0.361     0.087   4.124
## 
## --------------
## RANDOM EFFECT COVARIANCE(S):
## Correlations on upper diagonal
## 
## $Theta
##       Theta
## Theta 0.795
## 
## $items
##           COV_items
## COV_items      2.47

#large item level variance after itemorder is regressed; not a great predictor of item difficulty
coef(LLTMwithError)

## $Item.1
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.2
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.3
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.4
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.5
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.6
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.7
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.8
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.9
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.10
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.11
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.12
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.13
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.14
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.15
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.16
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.17
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.18
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.19
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.20
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.21
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.22
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.23
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.24
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.25
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.26
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.27
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.28
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.29
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.30
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.31
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $Item.32
##         itemordereasier itemorderharder a1  d  g  u
## par               0.206           0.361  1  0  0  1
## CI_2.5            0.076           0.189 NA NA NA NA
## CI_97.5           0.335           0.532 NA NA NA NA
## 
## $GroupPars
##         MEAN_1 COV_11
## par          0  0.795
## CI_2.5      NA  0.662
## CI_97.5     NA  0.928
## 
## $items
##         COV_items_items
## par               2.468
## CI_2.5            1.263
## CI_97.5           3.674

###################################################
### Polytomous example

#make an arbitrary group difference
covdat <- data.frame(group = rep(c('m', 'f'), nrow(Science)/2))

#partial credit model
mod <- mixedmirt(Science, covdat, model=1, fixed = ~ 0 + group)
coef(mod)

## $Comfort
##         groupm a1 ak0 ak1 ak2 ak3 d0    d1    d2    d3
## par     -0.032  1   0   1   2   3  0 3.039 5.620 4.260
## CI_2.5  -0.283 NA  NA  NA  NA  NA NA 2.012 4.548 3.121
## CI_97.5  0.218 NA  NA  NA  NA  NA NA 4.066 6.692 5.399
## 
## $Work
##         groupm a1 ak0 ak1 ak2 ak3 d0    d1    d2    d3
## par     -0.032  1   0   1   2   3  0 1.870 2.783 0.965
## CI_2.5  -0.283 NA  NA  NA  NA  NA NA 1.393 2.209 0.249
## CI_97.5  0.218 NA  NA  NA  NA  NA NA 2.348 3.357 1.682
## 
## $Future
##         groupm a1 ak0 ak1 ak2 ak3 d0    d1    d2    d3
## par     -0.032  1   0   1   2   3  0 2.612 4.027 2.922
## CI_2.5  -0.283 NA  NA  NA  NA  NA NA 1.958 3.289 2.089
## CI_97.5  0.218 NA  NA  NA  NA  NA NA 3.266 4.764 3.755
## 
## $Benefit
##         groupm a1 ak0 ak1 ak2 ak3 d0    d1    d2    d3
## par     -0.032  1   0   1   2   3  0 2.419 3.317 1.995
## CI_2.5  -0.283 NA  NA  NA  NA  NA NA 1.869 2.672 1.236
## CI_97.5  0.218 NA  NA  NA  NA  NA NA 2.970 3.963 2.754
## 
## $GroupPars
##         MEAN_1 COV_11
## par          0  0.934
## CI_2.5      NA  0.633
## CI_97.5     NA  1.235

#gpcm to estimate slopes
mod2 <- mixedmirt(Science, covdat, model=1, fixed = ~ 0 + group,
                 itemtype = 'gpcm')
summary(mod2)

## 
## Call:
## mixedmirt(data = Science, covdata = covdat, model = 1, fixed = ~0 + 
##     group, itemtype = "gpcm")
## 
## --------------
## FIXED EFFECTS:
##        Estimate Std.Error z.value
## groupm   -0.164     0.126  -1.305
## 
## --------------
## RANDOM EFFECT COVARIANCE(S):
## Correlations on upper diagonal
## 
## $Theta
##    F1
## F1  1

anova(mod, mod2)

## 
## Model 1: mixedmirt(data = Science, covdata = covdat, model = 1, fixed = ~0 + 
##     group)
## Model 2: mixedmirt(data = Science, covdata = covdat, model = 1, fixed = ~0 + 
##     group, itemtype = "gpcm")

##        AIC     AICc    SABIC      BIC    logLik    X2  df     p
## 1 3268.627 3269.741 3279.803 3324.224 -1620.313   NaN NaN   NaN
## 2 3257.107 3258.744 3270.678 3324.619 -1611.554 17.52   3 6e-04

#graded model
mod3 <- mixedmirt(Science, covdat, model=1, fixed = ~ 0 + group,
                 itemtype = 'graded')
coef(mod3)

## $Comfort
##         groupm    a1    d1    d2     d3
## par     -0.207 1.010 4.937 2.722 -1.354
## CI_2.5  -0.524 0.662 3.976 2.274 -1.693
## CI_97.5  0.110 1.359 5.898 3.171 -1.014
## 
## $Work
##         groupm    a1    d1    d2     d3
## par     -0.207 1.207 3.013 0.997 -2.154
## CI_2.5  -0.524 0.865 2.533 0.684 -2.580
## CI_97.5  0.110 1.550 3.492 1.311 -1.729
## 
## $Future
##         groupm    a1    d1    d2     d3
## par     -0.207 2.494 5.615 2.441 -1.976
## CI_2.5  -0.524 1.357 3.875 1.574 -2.727
## CI_97.5  0.110 3.630 7.356 3.309 -1.225
## 
## $Benefit
##         groupm    a1    d1    d2     d3
## par     -0.207 1.068 3.429 1.085 -1.575
## CI_2.5  -0.524 0.683 2.869 0.776 -1.948
## CI_97.5  0.110 1.452 3.990 1.395 -1.202
## 
## $GroupPars
##         MEAN_1 COV_11
## par          0      1
## CI_2.5      NA     NA
## CI_97.5     NA     NA

###################################################
# latent regression with Rasch and 2PL models

set.seed(1)
n <- 300
a <- matrix(1, 10)
d <- matrix(rnorm(10))
Theta <- matrix(c(rnorm(n, 0), rnorm(n, 1), rnorm(n, 2)))
covdata <- data.frame(group=rep(c('g1','g2','g3'), each=n))
dat <- simdata(a, d, N=n*3, Theta=Theta, itemtype = 'dich')

#had we known the latent abilities, we could have computed the regression coefs
summary(lm(Theta ~ covdata$group))

## 
## Call:
## lm(formula = Theta ~ covdata$group)
## 
## Residuals:
##     Min      1Q  Median      3Q     Max 
## -2.9871 -0.6851 -0.0427  0.7170  3.8313 
## 
## Coefficients:
##                 Estimate Std. Error t value Pr(>|t|)    
## (Intercept)      0.04434    0.05970   0.743    0.458    
## covdata$groupg2  0.93468    0.08443  11.071   <2e-16 ***
## covdata$groupg3  1.88134    0.08443  22.284   <2e-16 ***
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Residual standard error: 1.034 on 897 degrees of freedom
## Multiple R-squared:  0.3563, Adjusted R-squared:  0.3549 
## F-statistic: 248.3 on 2 and 897 DF,  p-value: < 2.2e-16

#but all we have is observed test data. Latent regression helps to recover these coefs
#Rasch model approach (and mirt equivalent)
rmod0 <- mirt(dat, 1, 'Rasch') # unconditional

# these two models are equivalent
rmod1a <- mirt(dat, 1, 'Rasch', covdata = covdata, formula = ~ group)
rmod1b <- mixedmirt(dat, covdata, 1, fixed = ~ 0 + items + group)
anova(rmod0, rmod1b)

## 
## Model 1: mirt(data = dat, model = 1, itemtype = "Rasch")
## Model 2: mixedmirt(data = dat, covdata = covdata, model = 1, fixed = ~0 + 
##     items + group)

##        AIC     AICc    SABIC      BIC    logLik     X2  df   p
## 1 9743.151 9743.448 9761.043 9795.977 -4860.576    NaN NaN NaN
## 2 9494.221 9494.632 9515.366 9556.652 -4734.111 252.93   2   0

coef(rmod1a, simplify=TRUE)

## $items
##         a1      d g u
## Item_1   1 -0.437 0 1
## Item_2   1  0.257 0 1
## Item_3   1 -0.555 0 1
## Item_4   1  1.758 0 1
## Item_5   1  0.416 0 1
## Item_6   1 -0.736 0 1
## Item_7   1  0.690 0 1
## Item_8   1  0.793 0 1
## Item_9   1  0.674 0 1
## Item_10  1 -0.251 0 1
## 
## $means
## F1 
##  0 
## 
## $cov
##       F1
## F1 0.981
## 
## $lr.betas
##                F1
## (Intercept) 0.000
## groupg2     0.810
## groupg3     1.776

summary(rmod1b)

## 
## Call:
## mixedmirt(data = dat, covdata = covdata, model = 1, fixed = ~0 + 
##     items + group)
## 
## --------------
## FIXED EFFECTS:
##         Estimate Std.Error z.value
## groupg2    0.810     0.099   8.142
## groupg3    1.782     0.113  15.753
## 
## --------------
## RANDOM EFFECT COVARIANCE(S):
## Correlations on upper diagonal
## 
## $Theta
##       F1
## F1 0.954

# 2PL, requires different input to allow Theta variance to remain fixed
mod0 <- mirt(dat, 1) # unconditional
mod1a <- mirt(dat, 1, covdata = covdata, formula = ~ group, itemtype = '2PL')
mod1b <- mixedmirt(dat, covdata, 1, fixed = ~ 0 + items, lr.fixed = ~group, itemtype = '2PL')
anova(mod0, mod1b)

## 
## Model 1: mirt(data = dat, model = 1)
## Model 2: mixedmirt(data = dat, covdata = covdata, model = 1, fixed = ~0 + 
##     items, itemtype = "2PL", lr.fixed = ~group)

##        AIC     AICc    SABIC      BIC    logLik      X2  df   p
## 1 9743.773 9744.729 9776.304 9839.821 -4851.887     NaN NaN NaN
## 2 9492.185 9493.339 9527.969 9597.838 -4724.092 255.588   2   0

coef(mod1a)$lr.betas

##                F1
## (Intercept) 0.000
## groupg2     0.815
## groupg3     1.800

summary(mod1b)

## 
## Call:
## mixedmirt(data = dat, covdata = covdata, model = 1, fixed = ~0 + 
##     items, itemtype = "2PL", lr.fixed = ~group)
## 
## 
## --------------
## RANDOM EFFECT COVARIANCE(S):
## Correlations on upper diagonal
## 
## $Theta
##    F1
## F1  1
## 
## --------------
## LATENT REGRESSION FIXED EFFECTS:
## 
##                F1
## (Intercept) 0.000
## groupg2     0.846
## groupg3     1.820
## 
##             Std.Error_F1   z_F1
## (Intercept)           NA     NA
## groupg2            0.095  8.955
## groupg3            0.100 18.267

# specifying specific regression effects is accomplished by passing a list of formula
model <- mirt.model('F1 = 1-5
                     F2 = 6-10')
covdata$contvar <- rnorm(nrow(covdata))
mod2 <- mirt(dat, model, itemtype = 'Rasch', covdata=covdata,
        formula = list(F1 = ~ group + contvar, F2 = ~ group))
coef(mod2)[11:12]

## $GroupPars
##     MEAN_1 MEAN_2 COV_11 COV_21 COV_22
## par      0      0  0.891      0  0.703
## 
## $lr.betas
##                 F1    F2
## (Intercept)  0.000 0.000
## groupg2      0.820 0.721
## groupg3      1.619 1.740
## contvar     -0.063 0.000

mod2b <- mixedmirt(dat, covdata, model, fixed = ~ 0 + items,
        lr.fixed = list(F1 = ~ group + contvar, F2 = ~ group))
summary(mod2b)

## 
## Call:
## mixedmirt(data = dat, covdata = covdata, model = model, fixed = ~0 + 
##     items, lr.fixed = list(F1 = ~group + contvar, F2 = ~group))
## 
## 
## --------------
## RANDOM EFFECT COVARIANCE(S):
## Correlations on upper diagonal
## 
## $Theta
##      F1    F2
## F1 1.07 0.000
## F2 0.00 0.787
## 
## --------------
## LATENT REGRESSION FIXED EFFECTS:
## 
##                 F1    F2
## (Intercept)  0.000 0.000
## groupg2      0.855 0.678
## groupg3      1.678 1.745
## contvar     -0.049 0.000
## 
##             Std.Error_F1 Std.Error_F2   z_F1  z_F2
## (Intercept)           NA           NA     NA    NA
## groupg2            0.089         0.13  9.613 5.205
## groupg3              NaN          NaN    NaN   NaN
## contvar            0.046           NA -1.070    NA

####################################################
## Simulated Multilevel Rasch Model

set.seed(1)
N <- 2000
a <- matrix(rep(1,10),10,1)
d <- matrix(rnorm(10))
cluster = 100
random_intercept = rnorm(cluster,0,1)
Theta = numeric()
for (i in 1:cluster)
    Theta <- c(Theta, rnorm(N/cluster,0,1) + random_intercept[i])

group = factor(rep(paste0('G',1:cluster), each = N/cluster))
covdata <- data.frame(group)
dat <- simdata(a,d,N, itemtype = rep('dich',10), Theta=matrix(Theta))

# null model
mod1 <- mixedmirt(dat, covdata, 1, fixed = ~ 0 + items, random = ~ 1|group)
summary(mod1)

## 
## Call:
## mixedmirt(data = dat, covdata = covdata, model = 1, fixed = ~0 + 
##     items, random = ~1 | group)
## 
## 
## --------------
## RANDOM EFFECT COVARIANCE(S):
## Correlations on upper diagonal
## 
## $Theta
##       F1
## F1 0.993
## 
## $group
##           COV_group
## COV_group     0.665

# include level 2 predictor for 'group' variance
covdata$group_pred <- rep(random_intercept, each = N/cluster)
mod2 <- mixedmirt(dat, covdata, 1, fixed = ~ 0 + items + group_pred, random = ~ 1|group)

# including group means predicts nearly all variability in 'group'
summary(mod2)

## 
## Call:
## mixedmirt(data = dat, covdata = covdata, model = 1, fixed = ~0 + 
##     items + group_pred, random = ~1 | group)
## 
## --------------
## FIXED EFFECTS:
##            Estimate Std.Error z.value
## group_pred     0.97     0.035  27.497
## 
## --------------
## RANDOM EFFECT COVARIANCE(S):
## Correlations on upper diagonal
## 
## $Theta
##       F1
## F1 0.984
## 
## $group
##           COV_group
## COV_group     0.025

anova(mod1, mod2)

## 
## Model 1: mixedmirt(data = dat, covdata = covdata, model = 1, fixed = ~0 + 
##     items, random = ~1 | group)
## Model 2: mixedmirt(data = dat, covdata = covdata, model = 1, fixed = ~0 + 
##     items + group_pred, random = ~1 | group)

##        AIC     AICc    SABIC      BIC    logLik      X2  df   p
## 1 23685.86 23686.02 23714.95 23753.07 -11830.93     NaN NaN NaN
## 2 22909.54 22909.72 22941.05 22982.35 -11441.77 778.325   1   0

## End(No test)

[Package mirt version 1.9 Index]