M2 {mirt}	R Documentation

Compute the M2 model fit statistic

Description

Computes the M2 (Maydeu-Olivares & Joe, 2006) statistic for dichotomous data and the M2* statistic for polytomous data (collapsing over response categories for better stability; see Cai and Hansen, 2013), as well as associated fit indices that are based on fitting the null model.

Usage

M2(obj, calcNull = TRUE, quadpts = NULL, theta_lim = c(-6, 6),
  impute = 0, CI = 0.9, residmat = FALSE, QMC = FALSE, suppress = 1,
  ...)

Arguments

obj
an estimated model object from the mirt package

calcNull
logical; calculate statistics for the null model as well? Allows for statistics such as the limited information TLI and CFI

quadpts
number of quadrature points to use during estimation. If NULL, a suitable value will be chosen based on the rubric found in fscores

theta_lim
lower and upper range to evaluate latent trait integral for each dimension

impute
a number indicating how many imputations to perform (passed to imputeMissing) when there are missing data present. This requires a precomputed Theta input. Will return a data.frame object with the mean estimates of the stats and their imputed standard deviations

CI
numeric value from 0 to 1 indicating the range of the confidence interval for RMSEA. Default returns the 90% interval

residmat
logical; return the residual matrix used to compute the SRMSR statistic? Only the lower triangle of the residual correlation matrix will be returned (the upper triangle is filled with NA's)

QMC
logical; use quasi-Monte Carlo integration? Useful for higher dimensional models. If quadpts not specified, 15000 nodes are used by default

suppress
a numeric value indicating which parameter residual dependency combinations to flag as being too high. Absolute values for the standardized residuals greater than this value will be returned, while all values less than this value will be set to NA. Must be used in conjunction with the argument residmat = TRUE

...
additional arguments to pass

Value

Returns a data.frame object with the M2 statistic, along with the degrees of freedom, p-value, RMSEA (with 90% confidence interval), SRMSR if all items were ordinal, and optionally the TLI and CFI model fit statistics

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

References

Cai, L. & Hansen, M. (2013). Limited-information goodness-of-fit testing of hierarchical item factor models. British Journal of Mathematical and Statistical Psychology, 66, 245-276.

Maydeu-Olivares, A. & Joe, H. (2006). Limited information goodness-of-fit testing in multidimensional contingency tables Psychometrika, 71, 713-732.

Examples

## No test: 
dat <- expand.table(LSAT7)
(mod1 <- mirt(dat, 1))

## 
## Call:
## mirt(data = dat, model = 1)
## 
## Full-information item factor analysis with 1 factor(s).
## Converged within 1e-04 tolerance after 28 EM iterations.
## mirt version: 1.10 
## M-step optimizer: BFGS 
## EM acceleration: Ramsay
## Number of rectangular quadrature: 61
## 
## Log-likelihood = -2658.805
## AIC = 5337.61; AICc = 5337.833
## BIC = 5386.688; SABIC = 5354.927
## G2 (21) = 31.7, p = 0.0628
## RMSEA = 0.023, CFI = 0.939, TLI = 0.924

M2(mod1)

##             M2 df          p     RMSEA     RMSEA_5   RMSEA_95       TLI
## stats 11.93769  5 0.03565165 0.0372683 0.008950922 0.06496573 0.9369332
##             CFI      SRMSR
## stats 0.9684666 0.03195919

M2(mod1, residmat=TRUE) #lower triangle of residual correlation matrix

##             Item.1      Item.2       Item.3       Item.4 Item.5
## Item.1          NA          NA           NA           NA     NA
## Item.2 -0.02212010          NA           NA           NA     NA
## Item.3 -0.03265942  0.03335601           NA           NA     NA
## Item.4  0.05155184 -0.01642572 -0.012476892           NA     NA
## Item.5  0.05443241 -0.03867705 -0.001860395 7.235426e-05     NA

#M2 imputed with missing data present (run in parallel)
dat[sample(1:prod(dim(dat)), 250)] <- NA
mod2 <- mirt(dat, 1)
mirtCluster()

## mirtCluster() has already been defined

M2(mod2, impute = 10)

##                M2 df          p       RMSEA     RMSEA_5   RMSEA_95
## stats    9.574611  5 0.11290279 0.029394130 0.002356477 0.05860189
## SD_stats 2.516901  0 0.05612749 0.007198836 0.007069431 0.00603010
##                 TLI        CFI       SRMSR
## stats    0.95459681 0.97729840 0.028917316
## SD_stats 0.02452819 0.01226409 0.003213708

## End(No test)

[Package mirt version 1.10 Index]