PLCI.mirt {mirt}	R Documentation

Compute profiled-likelihood confidence intervals

Description

Computes profiled-likelihood based confidence intervals. Supports the inclusion of prior parameter distributions as well as equality constraints.

Usage

PLCI.mirt(mod, alpha = 0.05, parnum = NULL, plot = FALSE, npts = 24,
  ...)

Arguments

mod
a converged mirt model

alpha
two-tailed alpha critical level

parnum
a numeric vector indicating which parameters to estimate. Use mod2values to determine parameter numbers. If NULL, all possible parameters are used

plot
logical; plot the parameter relationship in the likelihood space for two parameters?

npts
number of points to evaluate and plot if plot = TRUE

...
additional arguments to pass to the estimation functions

See Also

boot.mirt

Examples

## No test: 
mirtCluster() #use all available cores to estimate CI's in parallel

## mirtCluster() has already been defined

dat <- expand.table(LSAT7)
mod <- mirt(dat, 1)

result <- PLCI.mirt(mod)
result

##      Item class parnam parnum     value lower_2.5 upper_97.5
## 1  Item.1  dich     a1      1 0.9879266 0.6705149  1.3806045
## 2  Item.1  dich      d      2 1.8560607 1.6198236  2.1473734
## 3  Item.2  dich     a1      5 1.0808853 0.7829707  1.4613673
## 4  Item.2  dich      d      6 0.8079786 0.6356877  0.9987544
## 5  Item.3  dich     a1      9 1.7058011 1.1949315  2.6062722
## 6  Item.3  dich      d     10 1.8042187 1.4770176  2.3755982
## 7  Item.4  dich     a1     13 0.7651859 0.5211663  1.0539014
## 8  Item.4  dich      d     14 0.4859966 0.3396014  0.6381355
## 9  Item.5  dich     a1     17 0.7357992 0.4551309  1.0566129
## 10 Item.5  dich      d     18 1.8545129 1.6454021  2.0986698

mod2 <- mirt(Science, 1)
result2 <- PLCI.mirt(mod2)
result2

##       Item  class parnam parnum      value  lower_2.5 upper_97.5
## 1  Comfort graded     a1      1  1.0414729  0.6993727   1.454282
## 2  Comfort graded     d1      2  4.8637387  4.0112138   5.968288
## 3  Comfort graded     d2      3  2.6397565  2.2337787   3.115183
## 4  Comfort graded     d3      4 -1.4658946 -1.7996820  -1.171171
## 5     Work graded     a1      5  1.2258714  0.8958606   1.621321
## 6     Work graded     d1      6  2.9240567  2.4851038   3.431261
## 7     Work graded     d2      7  0.9010791  0.6315233   1.196785
## 8     Work graded     d3      8 -2.2665744 -2.6992118  -1.893484
## 9   Future graded     a1      9  2.2950953  1.5688526   3.986465
## 10  Future graded     d1     10  5.2365326  4.1279527   7.822383
## 11  Future graded     d2     11  2.2148449  1.6580793   3.414952
## 12  Future graded     d3     12 -1.9645374 -3.0252765  -1.454014
## 13 Benefit graded     a1     13  1.0946895  0.7658947   1.500362
## 14 Benefit graded     d1     14  3.3476577  2.8453393   3.940363
## 15 Benefit graded     d2     15  0.9916032  0.7262126   1.282877
## 16 Benefit graded     d3     16 -1.6880711 -2.0454654  -1.377499

#only estimate CI's slopes
sv <- mod2values(mod2)
parnum <- sv$parnum[sv$name == 'a1']
result3 <- PLCI.mirt(mod2, parnum=parnum)
result3

##      Item  class parnam parnum    value lower_2.5 upper_97.5
## 1 Comfort graded     a1      1 1.041473 0.6993727   1.454282
## 2    Work graded     a1      5 1.225871 0.8958606   1.621321
## 3  Future graded     a1      9 2.295095 1.5688526   3.986465
## 4 Benefit graded     a1     13 1.094690 0.7658947   1.500362

## End(No test)

[Package mirt version 1.9 Index]