PLCI.mirt {mirt}	R Documentation

Compute profiled-likelihood (or posterior) confidence intervals

Description

Computes profiled-likelihood based confidence intervals. Supports the inclusion of equality constraints. Object returns the confidence intervals and whether the respective interval could be found.

Usage

PLCI.mirt(mod, parnum = NULL, alpha = 0.05, search_bound = TRUE,
  step = 0.5, lower = TRUE, upper = TRUE, inf2val = 30,
  NealeMiller = FALSE, ...)

Arguments

mod
a converged mirt model

parnum
a numeric vector indicating which parameters to estimate. Use mod2values to determine parameter numbers. If NULL, all possible parameters are used

alpha
two-tailed alpha critical level

search_bound
logical; use a fixed grid of values around the ML estimate to determine more suitable optimization bounds? Using this has much better behaviour than setting fixed upper/lower bound values and searching from more extreme ends

step
magnitude of steps used when search_bound is TRUE. Smaller values create more points to search a suitable bound for (up to the lower bound value visible with mod2values). When upper/lower bounds are detected this value will be adjusted accordingly

lower
logical; search for the lower CI?

upper
logical; search for the upper CI?

inf2val
a numeric used to change parameter bounds which are infinity to a finite number. Decreasing this too much may not allow a suitable bound to be located. Default is 30

NealeMiller
logical; use the Neale and Miller 1997 approximation? Default is FALSE

...
additional arguments to pass to the estimation functions

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

References

Neale, M. C. & Miller, M. B. (1997). The use of likelihood-based confidence intervals in genetic models. Behavior Genetircs, 27, 113-120

See Also

boot.mirt

Examples

## No test: 
mirtCluster() #use all available cores to estimate CI's in parallel

## mirtCluster() has already been defined

dat <- expand.table(LSAT7)
mod <- mirt(dat, 1)

result <- PLCI.mirt(mod)
result

##      Item class parnam parnum     value lower_2.5 upper_97.5 lower_conv
## 1  Item.1  dich     a1      1 0.9879254 0.6705382  1.3819761       TRUE
## 2  Item.1  dich      d      2 1.8560605 1.6203324  2.1474212       TRUE
## 3  Item.2  dich     a1      5 1.0808847 0.7816287  1.4614715       TRUE
## 4  Item.2  dich      d      6 0.8079786 0.6373153  0.9994391       TRUE
## 5  Item.3  dich     a1      9 1.7058006 1.1965700  2.6063139       TRUE
## 6  Item.3  dich      d     10 1.8042187 1.4754051  2.3765518       TRUE
## 7  Item.4  dich     a1     13 0.7651853 0.5211915  1.0554405       TRUE
## 8  Item.4  dich      d     14 0.4859966 0.3417040  0.6365806       TRUE
## 9  Item.5  dich     a1     17 0.7357980 0.4551545  1.0555079       TRUE
## 10 Item.5  dich      d     18 1.8545127 1.6438380  2.0976038       TRUE
##    upper_conv
## 1        TRUE
## 2        TRUE
## 3        TRUE
## 4        TRUE
## 5        TRUE
## 6        TRUE
## 7        TRUE
## 8        TRUE
## 9        TRUE
## 10       TRUE

# model with constraints
mod <- mirt(dat, 'F = 1-5
                  CONSTRAIN = (1-5, a1)')

result <- PLCI.mirt(mod)
result

##     Item class parnam parnum     value lower_2.5 upper_97.5 lower_conv
## 1 Item.1  dich     a1      1 1.0111403 0.8863986  1.1416332       TRUE
## 2 Item.1  dich      d      2 1.8681305 1.6751813  2.0690618       TRUE
## 3 Item.2  dich      d      6 0.7909392 0.6334651  0.9518291       TRUE
## 4 Item.3  dich      d     10 1.4608781 1.2848561  1.6429949       TRUE
## 5 Item.4  dich      d     14 0.5214695 0.3681970  0.6770357       TRUE
## 6 Item.5  dich      d     18 1.9928267 1.7937790  2.2005130       TRUE
##   upper_conv
## 1       TRUE
## 2       TRUE
## 3       TRUE
## 4       TRUE
## 5       TRUE
## 6       TRUE

mod2 <- mirt(Science, 1)
result2 <- PLCI.mirt(mod2)
result2

##       Item  class parnam parnum      value  lower_2.5 upper_97.5
## 1  Comfort graded     a1      1  1.0417547  0.7008507   1.453043
## 2  Comfort graded     d1      2  4.8641542  4.0112192   5.966783
## 3  Comfort graded     d2      3  2.6399417  2.2332765   3.115799
## 4  Comfort graded     d3      4 -1.4660135 -1.7996419  -1.171326
## 5     Work graded     a1      5  1.2259618  0.8942260   1.620994
## 6     Work graded     d1      6  2.9240027  2.4851417   3.430756
## 7     Work graded     d2      7  0.9011651  0.6307976   1.195238
## 8     Work graded     d3      8 -2.2665647 -2.6975968  -1.893557
## 9   Future graded     a1      9  2.2933717  1.5687253   3.986968
## 10  Future graded     d1     10  5.2339928  4.1279873   7.822344
## 11  Future graded     d2     11  2.2137728  1.6589418   3.416497
## 12  Future graded     d3     12 -1.9637062 -3.0256940  -1.453916
## 13 Benefit graded     a1     13  1.0949151  0.7659051   1.501997
## 14 Benefit graded     d1     14  3.3479196  2.8453839   3.940600
## 15 Benefit graded     d2     15  0.9916289  0.7267310   1.282509
## 16 Benefit graded     d3     16 -1.6882599 -2.0443919  -1.375841
##    lower_conv upper_conv
## 1        TRUE       TRUE
## 2        TRUE       TRUE
## 3        TRUE       TRUE
## 4        TRUE       TRUE
## 5        TRUE       TRUE
## 6        TRUE       TRUE
## 7        TRUE       TRUE
## 8        TRUE       TRUE
## 9        TRUE       TRUE
## 10       TRUE       TRUE
## 11       TRUE       TRUE
## 12       TRUE       TRUE
## 13       TRUE       TRUE
## 14       TRUE       TRUE
## 15       TRUE       TRUE
## 16       TRUE       TRUE

#only estimate CI's slopes
sv <- mod2values(mod2)
parnum <- sv$parnum[sv$name == 'a1']
result3 <- PLCI.mirt(mod2, parnum)
result3

##      Item  class parnam parnum    value lower_2.5 upper_97.5 lower_conv
## 1 Comfort graded     a1      1 1.041755 0.7008507   1.453043       TRUE
## 2    Work graded     a1      5 1.225962 0.8942260   1.620994       TRUE
## 3  Future graded     a1      9 2.293372 1.5687253   3.986968       TRUE
## 4 Benefit graded     a1     13 1.094915 0.7659051   1.501997       TRUE
##   upper_conv
## 1       TRUE
## 2       TRUE
## 3       TRUE
## 4       TRUE

## End(No test)

[Package mirt version 1.21 Index]