Model-based Reliable Change Index

Computes an IRT version of the "reliable change index" (RCI) proposed by Jacobson and Traux (1991) but modified to use IRT information about scores and measurement error (see Jabrayilov, Emons, and Sijtsma (2016)). Main benefit of the IRT approach is the inclusion of response pattern information in the pre/post data score estimates, as well as conditional standard error of measurement information. Models can be specified as separate unidimensional IRT models fitted via mirt (or extracted from multipleGroup via extract.group), or a two-dimensional model where the latent traits correspond to the two test administrations.

Usage

RCI(
  mod_pre,
  predat,
  postdat,
  mod_post = mod_pre,
  cutoffs = NULL,
  SEM.pre = NULL,
  SEM.post = NULL,
  Fisher = FALSE,
  zero_cor = TRUE,
  shiny = FALSE,
  main = "Test Scores",
  ...
)

Arguments

mod_pre

single-group model fitted by mirt. If not supplied the information will be extracted from the data input objects to compute the classical test theory version of the RCI statistics

predat

a vector (if one individual) or matrix/data.frame of response data to be scored, where each individuals' responses are included in exactly one row.

If total score information only is to be used instead of the complete response matrix then a matrix object with exactly ncol = 1 columns should be provided

postdat

same as predat, but with respect to the post/follow-up measurement. Ignored when a two-dimensional IRT model is included.

For the original RCI approach, a matrix containing the complete responses, or a matrix with one column containing only the total scores, can be provided

mod_post

(optional) IRT model for post-test if different from pre-test; otherwise, the pre-test model will be used. Ignored when a two-dimensional model IRT is included

cutoffs

optional vector of length 2 indicating the type of cut-offs to report (e.g., c(-1.96, 1.96) reflects the 95 percent z-score type cut-off)

SEM.pre

standard error of measurement for the pretest. This can be used instead of rxx.pre and SD.pre

SEM.post

(optional) standard error of measurement for the post-test. Using this will create a pooled version of the SEM; otherwise, SEM.post = SEM.pre

Fisher

logical; use the Fisher/expected information function to compute the SE terms? If FALSE the SE information will be extracted from the select fscores method (default). Only applicable for unidimensional models

zero_cor

logical; when the supplied mod_pre is a two-factor model should the covariance/correlation between the latent traits be forced to be 0?

shiny

logical; launch an interactive shiny applications for real-time scoring of supplied total-scores or response vectors? Only requires mod_pre and (optional) mod_post inputs

main

main label to use when shiny=TRUE

...

additional arguments passed to fscores

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi:10.18637/jss.v048.i06

Jacobson, N. S., & Truax, P. (1991). Clinical significance: A statistical approach to defining meaningful change in psychotherapy research. Journal of Consulting and Clinical Psychology, 59, 12-19.

Jabrayilov, R. , Emons, W. H. M., & Sijtsma, K. (2016). Comparison of Classical Test Theory and Item Response Theory in Individual Change Assessment. Applied Psychological Measurement, 40 (8), 559-572.

Author

Phil Chalmers rphilip.chalmers@gmail.com

Examples

# \donttest{

# simulate some data
N <- 1000
J <- 20     # number of items
a <- matrix(rlnorm(J,.2,.3))
d <- rnorm(J)

theta <- matrix(rnorm(N))
dat_pre <- simdata(a, d, itemtype = '2PL', Theta = theta)

# first 3 cases decrease by 1/2
theta2 <- theta - c(1/2, 1/2, 1/2, numeric(N-3))
dat_post <- simdata(a, d, itemtype = '2PL', Theta = theta2)

mod <- mirt(dat_pre)
#> 

# all changes using fitted model from pre data
RCI(mod, predat=dat_pre, postdat=dat_post)
#> Error in pnorm(abs(z), lower.tail = FALSE): Non-numeric argument to mathematical function

# single response pattern change using EAP information
RCI(mod, predat=dat_pre[1,], postdat=dat_post[1,])
#> Error in pnorm(abs(z), lower.tail = FALSE): Non-numeric argument to mathematical function

# WLE estimator with Fisher information for SE (see Jabrayilov et al. 2016)
RCI(mod, predat = dat_pre[1,], postdat = dat_post[1,],
    method = 'WLE', Fisher = TRUE)
#> Error in pnorm(abs(z), lower.tail = FALSE): Non-numeric argument to mathematical function

# multiple respondents
RCI(mod, predat = dat_pre[1:6,], postdat = dat_post[1:6,])
#> Error in pnorm(abs(z), lower.tail = FALSE): Non-numeric argument to mathematical function

# include large-sample z-type cutoffs
RCI(mod, predat = dat_pre[1:6,], postdat = dat_post[1:6,],
    cutoffs = c(-1.96, 1.96))
#> Error in pnorm(abs(z), lower.tail = FALSE): Non-numeric argument to mathematical function

######
# CTT version by omitting IRT model
    # Requires either sample or population SEM's as input
(istats <- itemstats(dat_pre)$overall)
#>     N mean_total.score sd_total.score ave.r  sd.r alpha SEM.alpha
#>  1000            9.785          4.846 0.214 0.065 0.846     1.901
SEM.alpha <- istats$SEM.alpha    # SEM estimate of dat_pre

# assumes SEM.post = SEM.pre
RCI(predat = dat_pre, postdat = dat_post, SEM.pre=SEM.alpha)
#> # A tibble: 1,000 × 6
#>    pre.score post.score  diff    SE      z        p
#>        <dbl>      <dbl> <dbl> <dbl>  <dbl>    <dbl>
#>  1        18         19     1  2.69  0.372 0.710   
#>  2        18          9    -9  2.69 -3.35  0.000812
#>  3        15          9    -6  2.69 -2.23  0.0256  
#>  4        11         10    -1  2.69 -0.372 0.710   
#>  5         4          5     1  2.69  0.372 0.710   
#>  6         4          8     4  2.69  1.49  0.137   
#>  7        11          8    -3  2.69 -1.12  0.264   
#>  8        11         15     4  2.69  1.49  0.137   
#>  9         3          3     0  2.69  0     1       
#> 10         4          8     4  2.69  1.49  0.137   
#> # ℹ 990 more rows

# include cutoffs
RCI(predat = dat_pre, postdat = dat_post, SEM.pre=SEM.alpha,
    cutoffs=c(-1.96, 1.96))
#> # A tibble: 1,000 × 7
#>    pre.score post.score  diff    SE      z        p cut_decision
#>        <dbl>      <dbl> <dbl> <dbl>  <dbl>    <dbl> <chr>       
#>  1        18         19     1  2.69  0.372 0.710    unchanged   
#>  2        18          9    -9  2.69 -3.35  0.000812 decreased   
#>  3        15          9    -6  2.69 -2.23  0.0256   decreased   
#>  4        11         10    -1  2.69 -0.372 0.710    unchanged   
#>  5         4          5     1  2.69  0.372 0.710    unchanged   
#>  6         4          8     4  2.69  1.49  0.137    unchanged   
#>  7        11          8    -3  2.69 -1.12  0.264    unchanged   
#>  8        11         15     4  2.69  1.49  0.137    unchanged   
#>  9         3          3     0  2.69  0     1        unchanged   
#> 10         4          8     4  2.69  1.49  0.137    unchanged   
#> # ℹ 990 more rows

# allows SEM.post != SEM.pre
(istats.post <- itemstats(dat_post)$overall)
#>     N mean_total.score sd_total.score ave.r  sd.r alpha SEM.alpha
#>  1000            9.745          4.774 0.206 0.063  0.84      1.91
SEM.alpha.post <- istats.post$SEM.alpha

RCI(predat = dat_pre, postdat = dat_post,
   SEM.pre=SEM.alpha, SEM.post=SEM.alpha.post)
#> # A tibble: 1,000 × 6
#>    pre.score post.score  diff    SE      z        p
#>        <dbl>      <dbl> <dbl> <dbl>  <dbl>    <dbl>
#>  1        18         19     1  2.69  0.371 0.711   
#>  2        18          9    -9  2.69 -3.34  0.000836
#>  3        15          9    -6  2.69 -2.23  0.0260  
#>  4        11         10    -1  2.69 -0.371 0.711   
#>  5         4          5     1  2.69  0.371 0.711   
#>  6         4          8     4  2.69  1.48  0.138   
#>  7        11          8    -3  2.69 -1.11  0.266   
#>  8        11         15     4  2.69  1.48  0.138   
#>  9         3          3     0  2.69  0     1       
#> 10         4          8     4  2.69  1.48  0.138   
#> # ℹ 990 more rows

# Supplying only the total scores
TS_pre <- matrix(rowSums(dat_pre))
TS_post <- matrix(rowSums(dat_post))
RCI(predat = TS_pre, postdat = TS_post,
   SEM.pre=SEM.alpha, SEM.post=SEM.alpha.post)
#> # A tibble: 1,000 × 6
#>    pre.score post.score  diff    SE      z        p
#>        <dbl>      <dbl> <dbl> <dbl>  <dbl>    <dbl>
#>  1        18         19     1  2.69  0.371 0.711   
#>  2        18          9    -9  2.69 -3.34  0.000836
#>  3        15          9    -6  2.69 -2.23  0.0260  
#>  4        11         10    -1  2.69 -0.371 0.711   
#>  5         4          5     1  2.69  0.371 0.711   
#>  6         4          8     4  2.69  1.48  0.138   
#>  7        11          8    -3  2.69 -1.11  0.266   
#>  8        11         15     4  2.69  1.48  0.138   
#>  9         3          3     0  2.69  0     1       
#> 10         4          8     4  2.69  1.48  0.138   
#> # ℹ 990 more rows

######

# interactive shiny interfaces for live scoring
mod_pre <- mirt(Science)
#> 

# (optional) setup mod_post to have medium effect size change (d = 0.5)
sv <- mod2values(mod_pre)
sv$value[sv$name == 'MEAN_1'] <- 0.5
mod_post <- mirt(Science, pars=sv, TOL=NA)

# only use pre-test model for scoring
if(interactive()){
    RCI(mod_pre=mod_pre, shiny=TRUE)

    # use both pre-test and post-test models for including empirical priors
    RCI(mod_pre=mod_pre, mod_post=mod_post, shiny=TRUE,
        main='Perceptions of Science and Technology')
 }

############################
# Example where individuals take completely different item set pre-post
#   but prior calibration has been performed to equate the items

dat <- key2binary(SAT12,
  key = c(1,4,5,2,3,1,2,1,3,1,2,4,2,1,5,3,4,4,1,4,3,3,4,1,3,5,1,3,1,5,4,5))

mod <- mirt(dat)
#> 

# with N=5 individuals under investigation
predat <- postdat <- dat[1:5,]
predat[, 17:32] <- NA
postdat[, 1:16] <- NA

predat
#> # A tibble: 5 × 32
#>   Item.1 Item.2 Item.3 Item.4 Item.5 Item.6 Item.7 Item.8 Item.9 Item.10 Item.11
#>    <int>  <int>  <int>  <int>  <int>  <int>  <int>  <int>  <int>   <int>   <int>
#> 1      1      1      1      1      1      1      1      1      1       1       1
#> 2      0      1      0      0      1      0      1      0      1       1       1
#> 3      1      1      1      0      1      0      1      0      1       0       1
#> 4      0      1      0      1      1      0      1      0      1       0       1
#> 5      0      1      1      1      1      0      1      1      0       0       1
#> # ℹ 21 more variables: Item.12 <int>, Item.13 <int>, Item.14 <int>,
#> #   Item.15 <int>, Item.16 <int>, Item.17 <lgl>, Item.18 <lgl>, Item.19 <lgl>,
#> #   Item.20 <lgl>, Item.21 <lgl>, Item.22 <lgl>, Item.23 <lgl>, Item.24 <lgl>,
#> #   Item.25 <lgl>, Item.26 <lgl>, Item.27 <lgl>, Item.28 <lgl>, Item.29 <lgl>,
#> #   Item.30 <lgl>, Item.31 <lgl>, Item.32 <lgl>
postdat
#> # A tibble: 5 × 32
#>   Item.1 Item.2 Item.3 Item.4 Item.5 Item.6 Item.7 Item.8 Item.9 Item.10 Item.11
#>   <lgl>  <lgl>  <lgl>  <lgl>  <lgl>  <lgl>  <lgl>  <lgl>  <lgl>  <lgl>   <lgl>  
#> 1 NA     NA     NA     NA     NA     NA     NA     NA     NA     NA      NA     
#> 2 NA     NA     NA     NA     NA     NA     NA     NA     NA     NA      NA     
#> 3 NA     NA     NA     NA     NA     NA     NA     NA     NA     NA      NA     
#> 4 NA     NA     NA     NA     NA     NA     NA     NA     NA     NA      NA     
#> 5 NA     NA     NA     NA     NA     NA     NA     NA     NA     NA      NA     
#> # ℹ 21 more variables: Item.12 <lgl>, Item.13 <lgl>, Item.14 <lgl>,
#> #   Item.15 <lgl>, Item.16 <lgl>, Item.17 <int>, Item.18 <int>, Item.19 <int>,
#> #   Item.20 <int>, Item.21 <int>, Item.22 <int>, Item.23 <int>, Item.24 <int>,
#> #   Item.25 <int>, Item.26 <int>, Item.27 <int>, Item.28 <int>, Item.29 <int>,
#> #   Item.30 <int>, Item.31 <int>, Item.32 <int>

RCI(mod, predat, postdat)
#> Error in pnorm(abs(z), lower.tail = FALSE): Non-numeric argument to mathematical function

######
# Two-dimensional IRT model for each time point, 20 items (no DIF)

J <- 20
N <- 500
slopes <- rlnorm(J, .2, .2)
a <- matrix(c(slopes, numeric(J*2), slopes),J*2)
ints <- rnorm(J)
d <- matrix(c(ints, ints), ncol=1)
data.frame(a=a, d=d)
#>          a.1       a.2           d
#> 1  1.1552393 0.0000000 -0.63265649
#> 2  0.7725387 0.0000000 -0.23498495
#> 3  1.1911807 0.0000000 -0.23365145
#> 4  1.5201541 0.0000000 -0.09858223
#> 5  1.1752077 0.0000000  0.83519448
#> 6  1.0858223 0.0000000 -0.43051228
#> 7  1.6628731 0.0000000  0.41443651
#> 8  1.1012075 0.0000000  1.41245283
#> 9  1.0354140 0.0000000  0.84976955
#> 10 1.3580607 0.0000000  0.17070288
#> 11 0.9164153 0.0000000 -0.24230045
#> 12 1.1904205 0.0000000 -1.66708924
#> 13 1.5486689 0.0000000 -1.74009514
#> 14 1.1839185 0.0000000  0.84319987
#> 15 1.2910372 0.0000000 -1.96094872
#> 16 1.3247273 0.0000000 -0.33889244
#> 17 1.3673442 0.0000000  1.84475886
#> 18 1.2868859 0.0000000  0.51631728
#> 19 0.9999844 0.0000000 -0.73065644
#> 20 1.1269914 0.0000000  1.68359583
#> 21 0.0000000 1.1552393 -0.63265649
#> 22 0.0000000 0.7725387 -0.23498495
#> 23 0.0000000 1.1911807 -0.23365145
#> 24 0.0000000 1.5201541 -0.09858223
#> 25 0.0000000 1.1752077  0.83519448
#> 26 0.0000000 1.0858223 -0.43051228
#> 27 0.0000000 1.6628731  0.41443651
#> 28 0.0000000 1.1012075  1.41245283
#> 29 0.0000000 1.0354140  0.84976955
#> 30 0.0000000 1.3580607  0.17070288
#> 31 0.0000000 0.9164153 -0.24230045
#> 32 0.0000000 1.1904205 -1.66708924
#> 33 0.0000000 1.5486689 -1.74009514
#> 34 0.0000000 1.1839185  0.84319987
#> 35 0.0000000 1.2910372 -1.96094872
#> 36 0.0000000 1.3247273 -0.33889244
#> 37 0.0000000 1.3673442  1.84475886
#> 38 0.0000000 1.2868859  0.51631728
#> 39 0.0000000 0.9999844 -0.73065644
#> 40 0.0000000 1.1269914  1.68359583

# mean effects across time
mu <- c(0, -1/2)
sigma <- matrix(c(1, .7, .7, 1), 2,2)

dat <- simdata(a, d, N, mu=mu, sigma=sigma, itemtype = '2PL')

# build equality constraints across time points
constr <- NULL
for(i in (1:J)){
    constr <- c(constr, paste0("(", i, ',', i+J, ",a1,a2)"))
    constr <- c(constr, paste0("(", i, ',', i+J, ",d)"))
}
constr <- paste0(constr, collapse=',')

# define model where item parameters constrained over time, and
# latent trait has potential scale-location changes (e.g., regression to the
# mean effects)
model <- sprintf("
                  thetapre = 1-%i,
                  thetapost = %i-%i,
                  COV = thetapre*thetapost, thetapost*thetapost
                  MEAN = thetapost
                  CONSTRAIN = %s", J, J+1, 2*J, constr)
cat(model)
#> 
#>                   thetapre = 1-20,
#>                   thetapost = 21-40,
#>                   COV = thetapre*thetapost, thetapost*thetapost
#>                   MEAN = thetapost
#>                   CONSTRAIN = (1,21,a1,a2),(1,21,d),(2,22,a1,a2),(2,22,d),(3,23,a1,a2),(3,23,d),(4,24,a1,a2),(4,24,d),(5,25,a1,a2),(5,25,d),(6,26,a1,a2),(6,26,d),(7,27,a1,a2),(7,27,d),(8,28,a1,a2),(8,28,d),(9,29,a1,a2),(9,29,d),(10,30,a1,a2),(10,30,d),(11,31,a1,a2),(11,31,d),(12,32,a1,a2),(12,32,d),(13,33,a1,a2),(13,33,d),(14,34,a1,a2),(14,34,d),(15,35,a1,a2),(15,35,d),(16,36,a1,a2),(16,36,d),(17,37,a1,a2),(17,37,d),(18,38,a1,a2),(18,38,d),(19,39,a1,a2),(19,39,d),(20,40,a1,a2),(20,40,d)

# fit the model to calibration data
mod <- mirt(dat, model = model, SE=TRUE)
#> 
#> 
#> Calculating information matrix...
coef(mod, printSE=TRUE)
#> $Item_1
#>        a1 a2      d logit(g) logit(u)
#> par 1.252  0 -0.639     -999      999
#> SE  0.117 NA  0.097       NA       NA
#> 
#> $Item_2
#>        a1 a2      d logit(g) logit(u)
#> par 0.702  0 -0.268     -999      999
#> SE  0.084 NA  0.077       NA       NA
#> 
#> $Item_3
#>        a1 a2      d logit(g) logit(u)
#> par 1.210  0 -0.210     -999      999
#> SE  0.112 NA  0.094       NA       NA
#> 
#> $Item_4
#>        a1 a2      d logit(g) logit(u)
#> par 1.608  0 -0.178     -999      999
#> SE  0.140 NA  0.110       NA       NA
#> 
#> $Item_5
#>        a1 a2     d logit(g) logit(u)
#> par 1.237  0 0.770     -999      999
#> SE  0.116 NA 0.103       NA       NA
#> 
#> $Item_6
#>        a1 a2      d logit(g) logit(u)
#> par 1.296  0 -0.445     -999      999
#> SE  0.118 NA  0.097       NA       NA
#> 
#> $Item_7
#>        a1 a2     d logit(g) logit(u)
#> par 1.856  0 0.377     -999      999
#> SE  0.161 NA 0.124       NA       NA
#> 
#> $Item_8
#>        a1 a2     d logit(g) logit(u)
#> par 1.125  0 1.368     -999      999
#> SE  0.115 NA 0.113       NA       NA
#> 
#> $Item_9
#>        a1 a2     d logit(g) logit(u)
#> par 0.961  0 0.717     -999      999
#> SE  0.099 NA 0.092       NA       NA
#> 
#> $Item_10
#>        a1 a2     d logit(g) logit(u)
#> par 1.517  0 0.150     -999      999
#> SE  0.133 NA 0.107       NA       NA
#> 
#> $Item_11
#>        a1 a2      d logit(g) logit(u)
#> par 0.855  0 -0.271     -999      999
#> SE  0.092 NA  0.081       NA       NA
#> 
#> $Item_12
#>        a1 a2      d logit(g) logit(u)
#> par 1.056  0 -1.475     -999      999
#> SE  0.115 NA  0.104       NA       NA
#> 
#> $Item_13
#>        a1 a2      d logit(g) logit(u)
#> par 1.745  0 -2.061     -999      999
#> SE  0.173 NA  0.157       NA       NA
#> 
#> $Item_14
#>        a1 a2     d logit(g) logit(u)
#> par 1.333  0 0.860     -999      999
#> SE  0.123 NA 0.109       NA       NA
#> 
#> $Item_15
#>        a1 a2      d logit(g) logit(u)
#> par 1.488  0 -2.028     -999      999
#> SE  0.153 NA  0.142       NA       NA
#> 
#> $Item_16
#>        a1 a2      d logit(g) logit(u)
#> par 1.409  0 -0.359     -999      999
#> SE  0.125 NA  0.102       NA       NA
#> 
#> $Item_17
#>        a1 a2     d logit(g) logit(u)
#> par 1.378  0 1.715     -999      999
#> SE  0.136 NA 0.135       NA       NA
#> 
#> $Item_18
#>        a1 a2     d logit(g) logit(u)
#> par 1.202  0 0.321     -999      999
#> SE  0.111 NA 0.096       NA       NA
#> 
#> $Item_19
#>        a1 a2      d logit(g) logit(u)
#> par 1.061  0 -0.807     -999      999
#> SE  0.106 NA  0.092       NA       NA
#> 
#> $Item_20
#>        a1 a2     d logit(g) logit(u)
#> par 1.146  0 1.724     -999      999
#> SE  0.121 NA 0.126       NA       NA
#> 
#> $Item_21
#>     a1    a2      d logit(g) logit(u)
#> par  0 1.252 -0.639     -999      999
#> SE  NA 0.117  0.097       NA       NA
#> 
#> $Item_22
#>     a1    a2      d logit(g) logit(u)
#> par  0 0.702 -0.268     -999      999
#> SE  NA 0.084  0.077       NA       NA
#> 
#> $Item_23
#>     a1    a2      d logit(g) logit(u)
#> par  0 1.210 -0.210     -999      999
#> SE  NA 0.112  0.094       NA       NA
#> 
#> $Item_24
#>     a1    a2      d logit(g) logit(u)
#> par  0 1.608 -0.178     -999      999
#> SE  NA 0.140  0.110       NA       NA
#> 
#> $Item_25
#>     a1    a2     d logit(g) logit(u)
#> par  0 1.237 0.770     -999      999
#> SE  NA 0.116 0.103       NA       NA
#> 
#> $Item_26
#>     a1    a2      d logit(g) logit(u)
#> par  0 1.296 -0.445     -999      999
#> SE  NA 0.118  0.097       NA       NA
#> 
#> $Item_27
#>     a1    a2     d logit(g) logit(u)
#> par  0 1.856 0.377     -999      999
#> SE  NA 0.161 0.124       NA       NA
#> 
#> $Item_28
#>     a1    a2     d logit(g) logit(u)
#> par  0 1.125 1.368     -999      999
#> SE  NA 0.115 0.113       NA       NA
#> 
#> $Item_29
#>     a1    a2     d logit(g) logit(u)
#> par  0 0.961 0.717     -999      999
#> SE  NA 0.099 0.092       NA       NA
#> 
#> $Item_30
#>     a1    a2     d logit(g) logit(u)
#> par  0 1.517 0.150     -999      999
#> SE  NA 0.133 0.107       NA       NA
#> 
#> $Item_31
#>     a1    a2      d logit(g) logit(u)
#> par  0 0.855 -0.271     -999      999
#> SE  NA 0.092  0.081       NA       NA
#> 
#> $Item_32
#>     a1    a2      d logit(g) logit(u)
#> par  0 1.056 -1.475     -999      999
#> SE  NA 0.115  0.104       NA       NA
#> 
#> $Item_33
#>     a1    a2      d logit(g) logit(u)
#> par  0 1.745 -2.061     -999      999
#> SE  NA 0.173  0.157       NA       NA
#> 
#> $Item_34
#>     a1    a2     d logit(g) logit(u)
#> par  0 1.333 0.860     -999      999
#> SE  NA 0.123 0.109       NA       NA
#> 
#> $Item_35
#>     a1    a2      d logit(g) logit(u)
#> par  0 1.488 -2.028     -999      999
#> SE  NA 0.153  0.142       NA       NA
#> 
#> $Item_36
#>     a1    a2      d logit(g) logit(u)
#> par  0 1.409 -0.359     -999      999
#> SE  NA 0.125  0.102       NA       NA
#> 
#> $Item_37
#>     a1    a2     d logit(g) logit(u)
#> par  0 1.378 1.715     -999      999
#> SE  NA 0.136 0.135       NA       NA
#> 
#> $Item_38
#>     a1    a2     d logit(g) logit(u)
#> par  0 1.202 0.321     -999      999
#> SE  NA 0.111 0.096       NA       NA
#> 
#> $Item_39
#>     a1    a2      d logit(g) logit(u)
#> par  0 1.061 -0.807     -999      999
#> SE  NA 0.106  0.092       NA       NA
#> 
#> $Item_40
#>     a1    a2     d logit(g) logit(u)
#> par  0 1.146 1.724     -999      999
#> SE  NA 0.121 0.126       NA       NA
#> 
#> $GroupPars
#>     MEAN_1 MEAN_2 COV_11 COV_21 COV_22
#> par      0 -0.519      1  0.691  0.971
#> SE      NA  0.049     NA  0.047  0.096
#> 
coef(mod, simplify=TRUE)
#> $items
#>            a1    a2      d g u
#> Item_1  1.252 0.000 -0.639 0 1
#> Item_2  0.702 0.000 -0.268 0 1
#> Item_3  1.210 0.000 -0.210 0 1
#> Item_4  1.608 0.000 -0.178 0 1
#> Item_5  1.237 0.000  0.770 0 1
#> Item_6  1.296 0.000 -0.445 0 1
#> Item_7  1.856 0.000  0.377 0 1
#> Item_8  1.125 0.000  1.368 0 1
#> Item_9  0.961 0.000  0.717 0 1
#> Item_10 1.517 0.000  0.150 0 1
#> Item_11 0.855 0.000 -0.271 0 1
#> Item_12 1.056 0.000 -1.475 0 1
#> Item_13 1.745 0.000 -2.061 0 1
#> Item_14 1.333 0.000  0.860 0 1
#> Item_15 1.488 0.000 -2.028 0 1
#> Item_16 1.409 0.000 -0.359 0 1
#> Item_17 1.378 0.000  1.715 0 1
#> Item_18 1.202 0.000  0.321 0 1
#> Item_19 1.061 0.000 -0.807 0 1
#> Item_20 1.146 0.000  1.724 0 1
#> Item_21 0.000 1.252 -0.639 0 1
#> Item_22 0.000 0.702 -0.268 0 1
#> Item_23 0.000 1.210 -0.210 0 1
#> Item_24 0.000 1.608 -0.178 0 1
#> Item_25 0.000 1.237  0.770 0 1
#> Item_26 0.000 1.296 -0.445 0 1
#> Item_27 0.000 1.856  0.377 0 1
#> Item_28 0.000 1.125  1.368 0 1
#> Item_29 0.000 0.961  0.717 0 1
#> Item_30 0.000 1.517  0.150 0 1
#> Item_31 0.000 0.855 -0.271 0 1
#> Item_32 0.000 1.056 -1.475 0 1
#> Item_33 0.000 1.745 -2.061 0 1
#> Item_34 0.000 1.333  0.860 0 1
#> Item_35 0.000 1.488 -2.028 0 1
#> Item_36 0.000 1.409 -0.359 0 1
#> Item_37 0.000 1.378  1.715 0 1
#> Item_38 0.000 1.202  0.321 0 1
#> Item_39 0.000 1.061 -0.807 0 1
#> Item_40 0.000 1.146  1.724 0 1
#> 
#> $means
#>  thetapre thetapost 
#>     0.000    -0.519 
#> 
#> $cov
#>           thetapre thetapost
#> thetapre     1.000     0.691
#> thetapost    0.691     0.971
#> 
summary(mod)
#>         thetapre thetapost    h2
#> Item_1     0.592           0.351
#> Item_2     0.381           0.145
#> Item_3     0.579           0.336
#> Item_4     0.687           0.472
#> Item_5     0.588           0.346
#> Item_6     0.606           0.367
#> Item_7     0.737           0.543
#> Item_8     0.551           0.304
#> Item_9     0.492           0.242
#> Item_10    0.665           0.443
#> Item_11    0.449           0.202
#> Item_12    0.527           0.278
#> Item_13    0.716           0.512
#> Item_14    0.617           0.380
#> Item_15    0.658           0.433
#> Item_16    0.638           0.407
#> Item_17    0.629           0.396
#> Item_18    0.577           0.333
#> Item_19    0.529           0.280
#> Item_20    0.558           0.312
#> Item_21              0.587 0.344
#> Item_22              0.377 0.142
#> Item_23              0.574 0.329
#> Item_24              0.682 0.465
#> Item_25              0.582 0.339
#> Item_26              0.600 0.360
#> Item_27              0.732 0.536
#> Item_28              0.546 0.298
#> Item_29              0.486 0.236
#> Item_30              0.660 0.435
#> Item_31              0.444 0.197
#> Item_32              0.522 0.272
#> Item_33              0.711 0.505
#> Item_34              0.611 0.374
#> Item_35              0.653 0.426
#> Item_36              0.632 0.400
#> Item_37              0.624 0.389
#> Item_38              0.571 0.326
#> Item_39              0.524 0.274
#> Item_40              0.553 0.306
#> 
#>         SE.thetapre SE.thetapost
#> Item_1        0.036             
#> Item_2        0.039             
#> Item_3        0.036             
#> Item_4        0.032             
#> Item_5        0.036             
#> Item_6        0.035             
#> Item_7        0.029             
#> Item_8        0.039             
#> Item_9        0.038             
#> Item_10       0.033             
#> Item_11       0.039             
#> Item_12       0.041             
#> Item_13       0.035             
#> Item_14       0.035             
#> Item_15       0.038             
#> Item_16       0.034             
#> Item_17       0.037             
#> Item_18       0.036             
#> Item_19       0.038             
#> Item_20       0.041             
#> Item_21                    0.036
#> Item_22                    0.039
#> Item_23                    0.036
#> Item_24                    0.032
#> Item_25                    0.036
#> Item_26                    0.035
#> Item_27                    0.029
#> Item_28                    0.039
#> Item_29                    0.038
#> Item_30                    0.033
#> Item_31                    0.038
#> Item_32                    0.041
#> Item_33                    0.035
#> Item_34                    0.035
#> Item_35                    0.038
#> Item_36                    0.034
#> Item_37                    0.038
#> Item_38                    0.036
#> Item_39                    0.038
#> Item_40                    0.041
#> 
#> SS loadings:  7.081 6.954 
#> Proportion Var:  0.177 0.174 
#> 
#> Factor correlations: 
#> 
#>           thetapre thetapost
#> thetapre     1.000          
#> thetapost    0.701         1

# test data
Theta <- cbind(c(0, 1, 2), c(0,1,2))
nochange <- simdata(a, d, itemtype = '2PL', Theta = Theta)
change <- simdata(a, d, itemtype = '2PL', Theta = Theta +
                      cbind(0, c(-1, -1, -1)))

# total score differences
data.frame(pre=rowSums(nochange[,1:J]),
           post=rowSums(nochange[,1:J + J]))
#>   pre post
#> 1  10    8
#> 2  12   15
#> 3  18   18
data.frame(pre=rowSums(change[,1:J]),
           post=rowSums(change[,1:J + J]))
#>   pre post
#> 1  12    5
#> 2  14   10
#> 3  19   17

RCI(mod, predat = nochange)
#> Error in pnorm(abs(z), lower.tail = FALSE): Non-numeric argument to mathematical function
RCI(mod, predat = change)
#> Error in pnorm(abs(z), lower.tail = FALSE): Non-numeric argument to mathematical function

# }