SAT12 {mirt}	R Documentation

Description of SAT12 data

Description

Data obtained from the TESTFACT (Woods et al., 2003) manual, with 32 response pattern scored items for a grade 12 science assessment test (SAT) measuring topics of chemistry, biology, and physics. The scoring key for these data is [1, 4, 5, 2, 3, 1, 2, 1, 3, 1, 2, 4, 2, 1, 5, 3, 4, 4, 1, 4, 3, 3, 4, 1, 3, 5, 1, 3, 1, 5, 4, 5], respectively. However, careful analysis using the nominal response model suggests that the scoring key for item 32 may be incorrect, and should be changed from 5 to 3.

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

References

Wood, R., Wilson, D. T., Gibbons, R. D., Schilling, S. G., Muraki, E., & Bock, R. D. (2003). TESTFACT 4 for Windows: Test Scoring, Item Statistics, and Full-information Item Factor Analysis [Computer software]. Lincolnwood, IL: Scientific Software International.

Examples

## No test: 
#score the data (missing scored as 0)
head(SAT12)

##   Item.1 Item.2 Item.3 Item.4 Item.5 Item.6 Item.7 Item.8 Item.9 Item.10
## 1      1      4      5      2      3      1      2      1      3       1
## 2      3      4      2      8      3      3      2      8      3       1
## 3      1      4      5      4      3      2      2      3      3       2
## 4      2      4      4      2      3      3      2      4      3       2
## 5      2      4      5      2      3      2      2      1      1       2
## 6      1      4      3      1      3      2      2      3      3       1
##   Item.11 Item.12 Item.13 Item.14 Item.15 Item.16 Item.17 Item.18 Item.19
## 1       2       4       2       1       5       3       4       4       1
## 2       2       8       2       1       5       2       4       1       1
## 3       2       1       3       1       5       5       4       1       3
## 4       2       4       2       1       5       2       4       1       3
## 5       2       4       2       1       5       4       4       5       1
## 6       2       3       2       1       5       5       4       4       1
##   Item.20 Item.21 Item.22 Item.23 Item.24 Item.25 Item.26 Item.27 Item.28
## 1       4       3       3       4       1       3       5       1       3
## 2       4       3       3       8       1       8       4       1       4
## 3       4       3       3       1       1       3       4       1       3
## 4       4       3       1       5       2       5       4       1       3
## 5       4       3       3       3       1       1       5       1       3
## 6       4       3       3       4       1       1       4       1       4
##   Item.29 Item.30 Item.31 Item.32
## 1       1       5       4       5
## 2       5       8       4       8
## 3       4       4       4       1
## 4       4       2       4       2
## 5       1       2       4       1
## 6       2       3       4       3

data <- key2binary(SAT12,
    key = c(1,4,5,2,3,1,2,1,3,1,2,4,2,1,5,3,4,4,1,4,3,3,4,1,3,5,1,3,1,5,4,5))
head(data)

##      Item.1 Item.2 Item.3 Item.4 Item.5 Item.6 Item.7 Item.8 Item.9
## [1,]      1      1      1      1      1      1      1      1      1
## [2,]      0      1      0      0      1      0      1      0      1
## [3,]      1      1      1      0      1      0      1      0      1
## [4,]      0      1      0      1      1      0      1      0      1
## [5,]      0      1      1      1      1      0      1      1      0
## [6,]      1      1      0      0      1      0      1      0      1
##      Item.10 Item.11 Item.12 Item.13 Item.14 Item.15 Item.16 Item.17
## [1,]       1       1       1       1       1       1       1       1
## [2,]       1       1       0       1       1       1       0       1
## [3,]       0       1       0       0       1       1       0       1
## [4,]       0       1       1       1       1       1       0       1
## [5,]       0       1       1       1       1       1       0       1
## [6,]       1       1       0       1       1       1       0       1
##      Item.18 Item.19 Item.20 Item.21 Item.22 Item.23 Item.24 Item.25
## [1,]       1       1       1       1       1       1       1       1
## [2,]       0       1       1       1       1       0       1       0
## [3,]       0       0       1       1       1       0       1       1
## [4,]       0       0       1       1       0       0       0       0
## [5,]       0       1       1       1       1       0       1       0
## [6,]       1       1       1       1       1       1       1       0
##      Item.26 Item.27 Item.28 Item.29 Item.30 Item.31 Item.32
## [1,]       1       1       1       1       1       1       1
## [2,]       0       1       0       0       0       1       0
## [3,]       0       1       1       0       0       1       0
## [4,]       0       1       1       0       0       1       0
## [5,]       1       1       1       1       0       1       0
## [6,]       0       1       0       0       0       1       0

#score the data, missing (value of 8) treated as NA
SAT12missing <- SAT12
SAT12missing[SAT12missing == 8] <- NA
data <- key2binary(SAT12missing,
    key = c(1,4,5,2,3,1,2,1,3,1,2,4,2,1,5,3,4,4,1,4,3,3,4,1,3,5,1,3,1,5,4,5))
head(data)

##      Item.1 Item.2 Item.3 Item.4 Item.5 Item.6 Item.7 Item.8 Item.9
## [1,]      1      1      1      1      1      1      1      1      1
## [2,]      0      1      0      0      1      0      1      0      1
## [3,]      1      1      1      0      1      0      1      0      1
## [4,]      0      1      0      1      1      0      1      0      1
## [5,]      0      1      1      1      1      0      1      1      0
## [6,]      1      1      0      0      1      0      1      0      1
##      Item.10 Item.11 Item.12 Item.13 Item.14 Item.15 Item.16 Item.17
## [1,]       1       1       1       1       1       1       1       1
## [2,]       1       1       0       1       1       1       0       1
## [3,]       0       1       0       0       1       1       0       1
## [4,]       0       1       1       1       1       1       0       1
## [5,]       0       1       1       1       1       1       0       1
## [6,]       1       1       0       1       1       1       0       1
##      Item.18 Item.19 Item.20 Item.21 Item.22 Item.23 Item.24 Item.25
## [1,]       1       1       1       1       1       1       1       1
## [2,]       0       1       1       1       1       0       1       0
## [3,]       0       0       1       1       1       0       1       1
## [4,]       0       0       1       1       0       0       0       0
## [5,]       0       1       1       1       1       0       1       0
## [6,]       1       1       1       1       1       1       1       0
##      Item.26 Item.27 Item.28 Item.29 Item.30 Item.31 Item.32
## [1,]       1       1       1       1       1       1       1
## [2,]       0       1       0       0       0       1       0
## [3,]       0       1       1       0       0       1       0
## [4,]       0       1       1       0       0       1       0
## [5,]       1       1       1       1       0       1       0
## [6,]       0       1       0       0       0       1       0

#potentially better scoring for item 32 (based on nominal model finding)
data <- key2binary(SAT12,
    key = c(1,4,5,2,3,1,2,1,3,1,2,4,2,1,5,3,4,4,1,4,3,3,4,1,3,5,1,3,1,5,4,3))
## End(No test)

[Package mirt version 1.16 Index]