SIBTEST {mirt}	R Documentation

Simultaneous Item Bias Test (SIBTEST)

Description

Classical test theory approach to detecting unidirectional and bidirectional (with one crossing location) DIF. This family of statistics is intended for unidimensional tests, and applies a regression-corrected matched-total score approach to quantify the response bias between two groups. Can be used for DIF, DBF, and DTF testing.

Usage

SIBTEST(dat, group, suspect_set, match_set, focal_name, guess_correction = 0,
  Jmin = 2, pk_focal = FALSE, correction = TRUE, details = FALSE)

Arguments

dat
integer dataset to be tested containing dichotomous or polytomous responses

group
a vector indicating group membership

suspect_set
an integer vector indicating which items to inspect with SIBTEST. Including only one value will perform a DIF test, while including more than one will perform a simultaneous bundle test (DBF); including all non-matched items will perform DTF. If missing, a simultaneous test using all the items not listed in match_set will be used (i.e., DTF)

match_set
an integer vector indicating which items to use as the items which are matched (i.e., contain no DIF). These are analogous to 'achor' items in the likelihood method to locate DIF. If missing, all items other than the items found in the suspect_set will be used

focal_name
name of the focal group; e.g., 'focal'. If not specified then one will be selected automatically

guess_correction
a vector of numbers from 0 to 1 indicating how much to correct the items for guessing. It's length should be the same as ncol(dat)

Jmin
the minimum number of observations required when splitting the data into focal and reference groups conditioned on the matched set

pk_focal
logical; using the group weights from the focal group instead of the total sample? Default is FALSE as per Shealy and Stout's recommendation

correction
logical; apply the composite correction for the difference between focal composite scores using the true-score regression technique? Default is TRUE, reflecting Shealy and Stout's method

details
logical; return a data.frame containing the details required to compute SIBTEST?

Details

SIBTEST is similar to the Mantel-Haenszel approach for detecting DIF but uses a regression correction based on the KR-20/coefficient alpha reliability index to correct the observed differences when the latent trait distributions are not equal. Function supports the standard SIBTEST for dichotomous and polytomous data (compensatory) and supports crossing DIF testing (i.e., non-compensatory/non-uniform) using the asymptotic sampling distribution version of the Crossing-SIBTEST (CSIBTEST) statistic described by Chalmers (accepted).

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

References

Chalmers, R. P. (accepted). Improving the Crossing-SIBTEST statistic for detecting non-uniform DIF. Psychometrika.

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi: 10.18637/jss.v048.i06

Chang, H. H., Mazzeo, J. & Roussos, L. (1996). DIF for Polytomously Scored Items: An Adaptation of the SIBTEST Procedure. Journal of Educational Measurement, 33, 333-353.

Li, H.-H. & Stout, W. (1996). A new procedure for detection of crossing DIF. Psychometrika, 61, 647-677.

Shealy, R. & Stout, W. (1993). A model-based standardization approach that separates true bias/DIF from group ability differences and detect test bias/DTF as well as item bias/DIF. Psychometrika, 58, 159-194.

Examples

## No test: 

set.seed(1234)
n <- 30
N <- 500
a <- matrix(1, n)
d <- matrix(rnorm(n), n)
group <- c(rep('reference', N), rep('focal', N*2))

## -------------
# groups completely equal
dat1 <- simdata(a, d, N, itemtype = 'dich')
dat2 <- simdata(a, d, N*2, itemtype = 'dich')
dat <- rbind(dat1, dat2)

#DIF (all other items as anchors)
SIBTEST(dat, group, suspect_set = 6)

##          focal_group n_matched_set n_suspect_set   beta    SE    X2 df
## SIBTEST        focal            29             1 -0.029 0.027 1.165  1
## CSIBTEST       focal            29             1  0.031    NA 1.302  2
##              p
## SIBTEST  0.280
## CSIBTEST 0.522

#DIF (specific anchors)
SIBTEST(dat, group, match_set = 1:5, suspect_set = 6)

##          focal_group n_matched_set n_suspect_set   beta    SE    X2 df
## SIBTEST        focal             5             1 -0.010 0.028 0.120  1
## CSIBTEST       focal             5             1  0.026    NA 1.323  2
##              p
## SIBTEST  0.730
## CSIBTEST 0.516

# DBF (all and specific anchors, respectively)
SIBTEST(dat, group, suspect_set = 11:30)

##          focal_group n_matched_set n_suspect_set  beta    SE    X2 df
## SIBTEST        focal            10            20 0.329 0.162 4.153  1
## CSIBTEST       focal            10            20 0.329    NA 4.153  1
##              p
## SIBTEST  0.042
## CSIBTEST 0.042

SIBTEST(dat, group, match_set = 1:5, suspect_set = 11:30)

##          focal_group n_matched_set n_suspect_set  beta    SE    X2 df
## SIBTEST        focal             5            20 0.524 0.194 7.293  1
## CSIBTEST       focal             5            20 0.524    NA 7.293  1
##              p
## SIBTEST  0.007
## CSIBTEST 0.007

#DTF
SIBTEST(dat, group, suspect_set = 11:30)

##          focal_group n_matched_set n_suspect_set  beta    SE    X2 df
## SIBTEST        focal            10            20 0.329 0.162 4.153  1
## CSIBTEST       focal            10            20 0.329    NA 4.153  1
##              p
## SIBTEST  0.042
## CSIBTEST 0.042

SIBTEST(dat, group, match_set = 1:10) #equivalent

##          focal_group n_matched_set n_suspect_set  beta    SE    X2 df
## SIBTEST        focal            10            20 0.329 0.162 4.153  1
## CSIBTEST       focal            10            20 0.329    NA 4.153  1
##              p
## SIBTEST  0.042
## CSIBTEST 0.042

# different hyper pars
dat1 <- simdata(a, d, N, itemtype = 'dich')
dat2 <- simdata(a, d, N*2, itemtype = 'dich', mu = .5, sigma = matrix(1.5))
dat <- rbind(dat1, dat2)
SIBTEST(dat, group, 6:30)

##          focal_group n_matched_set n_suspect_set   beta    SE     X2 df
## SIBTEST        focal             5            25 -0.443 0.254  3.039  1
## CSIBTEST       focal             5            25  1.030    NA 17.787  2
##              p
## SIBTEST  0.081
## CSIBTEST 0.000

SIBTEST(dat, group, 11:30)

##          focal_group n_matched_set n_suspect_set   beta    SE    X2 df
## SIBTEST        focal            10            20 -0.153 0.172 0.793  1
## CSIBTEST       focal            10            20  0.257    NA 3.115  2
##              p
## SIBTEST  0.373
## CSIBTEST 0.211

#DIF testing with anchors 1 through 5
SIBTEST(dat, group, 6, match_set = 1:5)

##          focal_group n_matched_set n_suspect_set   beta    SE    X2 df
## SIBTEST        focal             5             1 -0.023 0.028 0.657  1
## CSIBTEST       focal             5             1  0.053    NA 4.257  2
##              p
## SIBTEST  0.418
## CSIBTEST 0.119

SIBTEST(dat, group, 7, match_set = 1:5)

##          focal_group n_matched_set n_suspect_set  beta    SE    X2 df   p
## SIBTEST        focal             5             1 0.037 0.029 1.641  1 0.2
## CSIBTEST       focal             5             1 0.037    NA 1.641  1 0.2

SIBTEST(dat, group, 8, match_set = 1:5)

##          focal_group n_matched_set n_suspect_set   beta    SE    X2 df
## SIBTEST        focal             5             1 -0.056 0.029 3.810  1
## CSIBTEST       focal             5             1  0.083    NA 8.443  2
##              p
## SIBTEST  0.051
## CSIBTEST 0.015

#DIF testing with all other items as anchors
SIBTEST(dat, group, 6)

##          focal_group n_matched_set n_suspect_set   beta    SE    X2 df
## SIBTEST        focal            29             1 -0.017 0.027 0.394  1
## CSIBTEST       focal            29             1  0.017    NA 0.394  1
##             p
## SIBTEST  0.53
## CSIBTEST 0.53

SIBTEST(dat, group, 7)

##          focal_group n_matched_set n_suspect_set beta    SE    X2 df     p
## SIBTEST        focal            29             1 0.05 0.029 2.993  1 0.084
## CSIBTEST       focal            29             1 0.05    NA 2.993  1 0.084

SIBTEST(dat, group, 8)

##          focal_group n_matched_set n_suspect_set   beta    SE    X2 df
## SIBTEST        focal            29             1 -0.035 0.029 1.398  1
## CSIBTEST       focal            29             1  0.081    NA 8.288  2
##              p
## SIBTEST  0.237
## CSIBTEST 0.016

## -------------
## systematic differing slopes and intercepts (clear DTF)
dat1 <- simdata(a, d, N, itemtype = 'dich')
dat2 <- simdata(a + c(numeric(15), rnorm(n-15, 1, .25)), d + c(numeric(15), rnorm(n-15, 1, 1)),
  N*2, itemtype = 'dich')
dat <- rbind(dat1, dat2)
SIBTEST(dat, group, 6:30)

##          focal_group n_matched_set n_suspect_set   beta   SE     X2 df p
## SIBTEST        focal             5            25 -1.659 0.25 43.937  1 0
## CSIBTEST       focal             5            25  1.659   NA 43.937  1 0

SIBTEST(dat, group, 11:30)

##          focal_group n_matched_set n_suspect_set  beta    SE      X2 df p
## SIBTEST        focal            10            20 -1.89 0.178 112.806  1 0
## CSIBTEST       focal            10            20  1.89    NA 112.806  1 0

#DIF testing using valid anchors
SIBTEST(dat, group, suspect_set = 6, match_set = 1:5)

##          focal_group n_matched_set n_suspect_set  beta    SE    X2 df
## SIBTEST        focal             5             1 0.007 0.027 0.059  1
## CSIBTEST       focal             5             1 0.024    NA 1.166  2
##              p
## SIBTEST  0.808
## CSIBTEST 0.558

SIBTEST(dat, group, suspect_set = 7, match_set = 1:5)

##          focal_group n_matched_set n_suspect_set  beta    SE    X2 df
## SIBTEST        focal             5             1 0.001 0.028 0.001  1
## CSIBTEST       focal             5             1 0.012    NA 0.305  2
##              p
## SIBTEST  0.976
## CSIBTEST 0.859

SIBTEST(dat, group, suspect_set = 30, match_set = 1:5)

##          focal_group n_matched_set n_suspect_set  beta    SE     X2 df p
## SIBTEST        focal             5             1 -0.19 0.027 48.204  1 0
## CSIBTEST       focal             5             1  0.19    NA 48.204  1 0

## End(No test)

[Package mirt version 1.27 Index]