SIBTEST {mirt}	R Documentation

Simultaneous Item Bias Test (SIBTEST)

Description

Classical test theory approach to detecting DIF for unidimensional tests by applying a regression-corrected matched-total score approach. SIBTEST is similar to the Mantel-Haenszel approach for detecting DIF but uses a regression correction based on the KR-20/coefficient alpha reliability index to correct the observed differences when the latent trait distributions are not equal. Function supports the standard SIBTEST for dichotomous and poltomous data (compensatory) and also supports crossed DIF testing (i.e., non-compensatory).

Usage

SIBTEST(dat, group, focal_set, match_set, focal_name, guess_correction = 0,
  Jmin = 2, cross = FALSE, permute = 1000, pk_focal = FALSE,
  correction = TRUE, details = FALSE)

Arguments

dat
integer dataset to be tested containing dichotomous or polytomous responses

group
a vector indicating group membership

focal_set
an integer vector indicating which items to inspect with SIBTEST. Including only one value will perform a DIF test, while including more than one will perform a simultaneous bundle test (DBF); including all non-matched items will perform DTF. If missing, a simultaneous test using all the items not listed in match_set will be used (i.e., DTF)

match_set
an integer vector indicating which items to use as the items which are matched (i.e., contain no DIF). These are analogous to 'achor' items in the likelihood method to locate DIF. If missing, all items other than the items found in the focal_set will be used

focal_name
name of the focal group; e.g., 'focal'. If not specified then one will be selected automatically

guess_correction
a vector of numbers from 0 to 1 indicating how much to correct the items for guessing. It's length should be the same as ncol(dat)

Jmin
the minimum number of observations required when splitting the data into focal and reference groups conditioned on the matched set

cross
logical; perform the crossing test for non-compensatory bias? Default is FALSE

permute
number of permutations to perform when cross = TRUE. Default is 1000

pk_focal
logical; using the group weights from the focal group instead of the total sample? Default is FALSE as per Shealy and Stout's recommendation

correction
logical; apply the composite correction for the difference between focal composite scores using the true-score regression technique? Default is TRUE, reflecting Shealy and Stout's method

details
logical; return a data.frame containing the details required to compute SIBTEST?

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

References

Chang, H. H., Mazzeo, J. & Roussos, L. (1996). DIF for Polytomously Scored Items: An Adaptation of the SIBTEST Procedure. Journal of Educational Measurement, 33, 333-353.

Li, H.-H. & Stout, W. (1996). A new procedure for detetion of crossing DIF. Psychometrika, 61, 647-677.

Shealy, R. & Stout, W. (1993). A model-based standardization approach that separates true bias/DIF from group ability differences and ddetect test bias/DTF as well as item bias/DIF. Psychometrika, 58, 159-194.

Examples

## No test: 

library(mirt)

set.seed(1234)
n <- 30
N <- 500
a <- matrix(1, n)
d <- matrix(rnorm(n), n)
group <- c(rep('reference', N), rep('focal', N*2))

## -------------
# groups completely equal
dat1 <- simdata(a, d, N, itemtype = '2PL')
dat2 <- simdata(a, d, N*2, itemtype = '2PL')
dat <- rbind(dat1, dat2)

#DIF (all other items as anchors)
SIBTEST(dat, group, focal_set = 6)

##         focal_group n_matched_set n_focal_set        beta         z
## SIBTEST       focal            29           1 -0.02895891 -1.079226
##                 p
## SIBTEST 0.2804868

#DIF (specific anchors)
SIBTEST(dat, group, match_set = 1:5, focal_set = 6)

##         focal_group n_matched_set n_focal_set         beta          z
## SIBTEST       focal             5           1 -0.009530775 -0.3457252
##                 p
## SIBTEST 0.7295492

# DBF (all and specific anchors, respectively)
SIBTEST(dat, group, focal_set = 11:30)

##         focal_group n_matched_set n_focal_set      beta        z
## SIBTEST       focal            10          20 0.3294495 2.037803
##                  p
## SIBTEST 0.04156966

SIBTEST(dat, group, match_set = 1:5, focal_set = 11:30)

##         focal_group n_matched_set n_focal_set      beta        z
## SIBTEST       focal             5          20 0.5238113 2.700537
##                   p
## SIBTEST 0.006922767

#DTF
SIBTEST(dat, group, focal_set = 11:30)

##         focal_group n_matched_set n_focal_set      beta        z
## SIBTEST       focal            10          20 0.3294495 2.037803
##                  p
## SIBTEST 0.04156966

SIBTEST(dat, group, match_set = 1:10) #equivalent

##         focal_group n_matched_set n_focal_set      beta        z
## SIBTEST       focal            10          20 0.3294495 2.037803
##                  p
## SIBTEST 0.04156966

# different hyper pars
dat1 <- simdata(a, d, N, itemtype = '2PL')
dat2 <- simdata(a, d, N*2, itemtype = '2PL', mu = .5, sigma = matrix(1.5))
dat <- rbind(dat1, dat2)
SIBTEST(dat, group, 6:30)

##         focal_group n_matched_set n_focal_set       beta        z
## SIBTEST       focal             5          25 -0.4426694 -1.74327
##                  p
## SIBTEST 0.08128638

SIBTEST(dat, group, 11:30)

##         focal_group n_matched_set n_focal_set       beta          z
## SIBTEST       focal            10          20 -0.1533017 -0.8902793
##                p
## SIBTEST 0.373316

#DIF testing with anchors 1 through 5
SIBTEST(dat, group, 6, match_set = 1:5)

##         focal_group n_matched_set n_focal_set        beta          z
## SIBTEST       focal             5           1 -0.02260609 -0.8105353
##                 p
## SIBTEST 0.4176326

SIBTEST(dat, group, 7, match_set = 1:5)

##         focal_group n_matched_set n_focal_set       beta        z
## SIBTEST       focal             5           1 0.03675611 1.281086
##                 p
## SIBTEST 0.2001635

SIBTEST(dat, group, 8, match_set = 1:5)

##         focal_group n_matched_set n_focal_set        beta         z
## SIBTEST       focal             5           1 -0.05599984 -1.951795
##                  p
## SIBTEST 0.05096254

#DIF testing with all other items as anchors
SIBTEST(dat, group, 6)

##         focal_group n_matched_set n_focal_set       beta          z
## SIBTEST       focal            29           1 -0.0166756 -0.6280739
##                 p
## SIBTEST 0.5299555

SIBTEST(dat, group, 7)

##         focal_group n_matched_set n_focal_set       beta        z
## SIBTEST       focal            29           1 0.04995383 1.730141
##                  p
## SIBTEST 0.08360504

SIBTEST(dat, group, 8)

##         focal_group n_matched_set n_focal_set        beta         z
## SIBTEST       focal            29           1 -0.03455412 -1.182194
##                 p
## SIBTEST 0.2371289

#crossed SIBTEST
SIBTEST(dat, group, 6, match_set = 1:5, cross=TRUE)

##                 focal_group n_matched_set n_focal_set       beta        z
## Crossed_SIBTEST       focal             5           1 0.03540281 1.504887
##                    p
## Crossed_SIBTEST 0.37

SIBTEST(dat, group, 7, match_set = 1:5, cross=TRUE)

##                 focal_group n_matched_set n_focal_set       beta        z
## Crossed_SIBTEST       focal             5           1 0.03675611 1.281086
##                     p
## Crossed_SIBTEST 0.506

SIBTEST(dat, group, 8, match_set = 1:5, cross=TRUE)

##                 focal_group n_matched_set n_focal_set      beta        z p
## Crossed_SIBTEST       focal             5           1 0.0874903 3.607364 0

## -------------
## systematic differing slopes and intercepts (clear DTF)
dat1 <- simdata(a, d, N, itemtype = '2PL')
dat2 <- simdata(a + c(numeric(15), rnorm(n-15, 1, .25)), d + c(numeric(15), rnorm(n-15, 1, 1)),
  N*2, itemtype = '2PL')
dat <- rbind(dat1, dat2)
SIBTEST(dat, group, 6:30)

##         focal_group n_matched_set n_focal_set      beta         z p
## SIBTEST       focal             5          25 -2.800422 -10.64853 0

SIBTEST(dat, group, 11:30)

##         focal_group n_matched_set n_focal_set      beta         z p
## SIBTEST       focal            10          20 -2.799902 -15.68929 0

#DIF testing using valid anchors
SIBTEST(dat, group, focal_set = 6, match_set = 1:5)

##         focal_group n_matched_set n_focal_set        beta         z
## SIBTEST       focal             5           1 -0.04568947 -1.625955
##                 p
## SIBTEST 0.1039593

SIBTEST(dat, group, focal_set = 7, match_set = 1:5)

##         focal_group n_matched_set n_focal_set       beta       z         p
## SIBTEST       focal             5           1 0.04047954 1.43352 0.1517093

SIBTEST(dat, group, focal_set = 30, match_set = 1:5)

##         focal_group n_matched_set n_focal_set       beta         z
## SIBTEST       focal             5           1 -0.1472457 -5.580823
##                    p
## SIBTEST 2.393826e-08

SIBTEST(dat, group, focal_set = 11, match_set = 1:10, cross=TRUE)

##                 focal_group n_matched_set n_focal_set       beta         z
## Crossed_SIBTEST       focal            10           1 0.02326573 0.9664296
##                     p
## Crossed_SIBTEST 0.355

SIBTEST(dat, group, focal_set = 30, match_set = 1:15, cross=TRUE)

##                 focal_group n_matched_set n_focal_set      beta        z
## Crossed_SIBTEST       focal            15           1 0.1753922 7.638039
##                     p
## Crossed_SIBTEST 0.001

## End(No test)

[Package mirt version 1.23 Index]