SingleGroupClass-class {mirt}	R Documentation

Class "SingleGroupClass"

Description

Defines the object returned from mirt when model is exploratory.

Slots

Call:

function call

Data:

list of data, sometimes in different forms

Options:

list of estimation options

Fit:

a list of fit information

Model:

a list of model-based information

ParObjects:

a list of the S4 objects used during estimation

OptimInfo:

a list of arguments from the optimization process

Internals:

a list of internal arguments for secondary computations (inspecting this object is generally not required)

vcov:

a matrix represented the asymtotic covariance matrix of the parameter estimates

time:

a data.frame indicating the breakdown of computation times in seconds

Methods

anova

signature(object = "SingleGroupClass")

coef

signature(object = "SingleGroupClass")

plot

signature(x = "SingleGroupClass", y = "missing")

print

signature(x = "SingleGroupClass")

residuals

signature(object = "SingleGroupClass")

show

signature(object = "SingleGroupClass")

summary

signature(object = "SingleGroupClass")

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

[Package mirt version 1.20 Index]