Compare nested models with likelihood-based statistics

Compare nested models using likelihood ratio test (X2), Akaike Information Criterion (AIC), sample size adjusted AIC (AICc), Bayesian Information Criterion (BIC), Sample-Size Adjusted BIC (SABIC), and Hannan-Quinn (HQ) Criterion.

# S4 method for SingleGroupClass
anova(object, object2, bounded = FALSE, mix = 0.5, verbose = TRUE)

Arguments

object
an object of class SingleGroupClass, MultipleGroupClass, or MixedClass

object2
a second model estimated from any of the mirt package estimation methods

bounded
logical; are the two models comparing a bounded parameter (e.g., comparing a single 2PL and 3PL model with 1 df)? If TRUE then a 50:50 mix of chi-squared distributions is used to obtain the p-value

mix
proportion of chi-squared mixtures. Default is 0.5

verbose
logical; print additional information to console?

Value

a data.frame/mirt_df object

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi: 10.18637/jss.v048.i06

Examples

# \donttest{
x <- mirt(Science, 1)
x2 <- mirt(Science, 2)
anova(x, x2)

#> 
#> Model 1: mirt(data = Science, model = 1)
#> Model 2: mirt(data = Science, model = 2)
#> 

#>        AIC     AICc    SABIC       HQ      BIC    logLik     X2  df     p
#> 1 3249.739 3251.190 3262.512 3274.922 3313.279 -1608.870    NaN NaN   NaN
#> 2 3241.938 3243.981 3257.106 3271.843 3317.392 -1601.969 13.801   3 0.003

# in isolation
anova(x)

#>        AIC    AICc    SABIC       HQ      BIC   logLik
#> 1 3249.739 3251.19 3262.512 3274.922 3313.279 -1608.87

# bounded parameter
dat <- expand.table(LSAT7)
mod <- mirt(dat, 1)
mod2 <- mirt(dat, 1, itemtype = c(rep('2PL', 4), '3PL'))
anova(mod, mod2) #unbounded test

#> 
#> Model 1: mirt(data = dat, model = 1)
#> Model 2: mirt(data = dat, model = 1, itemtype = c(rep("2PL", 4), "3PL"))
#> 

#>        AIC     AICc    SABIC       HQ      BIC    logLik    X2  df    p
#> 1 5337.610 5337.833 5354.927 5356.263 5386.688 -2658.805   NaN NaN  NaN
#> 2 5339.587 5339.855 5358.636 5360.106 5393.573 -2658.794 0.023   1 0.88

anova(mod, mod2, bounded = TRUE) #bounded

#> 
#> Model 1: mirt(data = dat, model = 1)
#> Model 2: mirt(data = dat, model = 1, itemtype = c(rep("2PL", 4), "3PL"))
#> 

#>        AIC     AICc    SABIC       HQ      BIC    logLik    X2  df    p
#> 1 5337.610 5337.833 5354.927 5356.263 5386.688 -2658.805   NaN NaN  NaN
#> 2 5339.587 5339.855 5358.636 5360.106 5393.573 -2658.794 0.023   1 0.44

# priors
model <- 'F = 1-5
          PRIOR = (5, g, norm, -1, 1)'
mod1b <- mirt(dat, model, itemtype = c(rep('2PL', 4), '3PL'))
anova(mod1b)

#>        AIC     AICc   SABIC       HQ      BIC    logLik      DIC   logPost
#> 1 5339.571 5339.838 5358.62 5360.089 5393.557 -2658.786 5341.409 -2659.705

model2 <- 'F = 1-5
          PRIOR = (1-5, g, norm, -1, 1)'
mod2b <- mirt(dat, model2, itemtype = '3PL')
anova(mod1b, mod2b)

#> 
#> Model 1: mirt(data = dat, model = model, itemtype = c(rep("2PL", 4), "3PL"))
#> Model 2: mirt(data = dat, model = model2, itemtype = "3PL")
#> 

#>        AIC     AICc    SABIC       HQ      BIC      DIC    logLik   logPost  df
#> 1 5339.571 5339.838 5358.620 5360.089 5393.557 5341.409 -2658.786 -2659.705 NaN
#> 2 5348.306 5348.794 5374.282 5376.286 5421.923 5358.015 -2659.153 -2664.008   4
#>   Bayes_Factor
#> 1           NA
#> 2       73.927

# }