anova-method {mirt}	R Documentation

Compare nested models with likelihood-based statistics

Description

Compare nested models using likelihood ratio, AIC, BIC, etc.

Usage

## S4 method for signature 'SingleGroupClass'
anova(object, object2, bounded = FALSE,
  mix = 0.5, verbose = TRUE)

Arguments

object
an object of class SingleGroupClass, MultipleGroupClass, or MixedClass

object2
a second model estimated from any of the mirt package estimation methods

bounded
logical; are the two models comparing a bounded parameter (e.g., comparing a single 2PL and 3PL model with 1 df)? If TRUE then a 50:50 mix of chi-squared distributions is used to obtain the p-value

mix
proportion of chi-squared mixtures. Default is 0.5

verbose
logical; print additional information to console?

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi: 10.18637/jss.v048.i06

Examples

## No test: 
x <- mirt(Science, 1)
x2 <- mirt(Science, 2)
anova(x, x2)

## 
## Model 1: mirt(data = Science, model = 1)
## Model 2: mirt(data = Science, model = 2)

##        AIC     AICc    SABIC      BIC    logLik     X2  df     p
## 1 3249.739 3251.190 3262.512 3313.279 -1608.870    NaN NaN   NaN
## 2 3241.938 3243.981 3257.106 3317.392 -1601.969 13.801   3 0.003

# bounded parameter
dat <- expand.table(LSAT7)
mod <- mirt(dat, 1)
mod2 <- mirt(dat, 1, itemtype = c(rep('2PL', 4), '3PL'))
anova(mod, mod2) #unbounded test

## 
## Model 1: mirt(data = dat, model = 1)
## Model 2: mirt(data = dat, model = 1, itemtype = c(rep("2PL", 4), "3PL"))

##        AIC     AICc    SABIC      BIC    logLik    X2  df     p
## 1 5337.610 5337.833 5354.927 5386.688 -2658.805   NaN NaN   NaN
## 2 5339.587 5339.855 5358.636 5393.573 -2658.794 0.023   1 0.879

anova(mod, mod2, bounded = TRUE) #bounded

## 
## Model 1: mirt(data = dat, model = 1)
## Model 2: mirt(data = dat, model = 1, itemtype = c(rep("2PL", 4), "3PL"))

##        AIC     AICc    SABIC      BIC    logLik    X2  df    p
## 1 5337.610 5337.833 5354.927 5386.688 -2658.805   NaN NaN  NaN
## 2 5339.587 5339.855 5358.636 5393.573 -2658.794 0.023   1 0.44

## End(No test)

[Package mirt version 1.25 Index]