Compare nested models with likelihood-based statistics

Compare nested models using likelihood ratio test (X2), Akaike Information Criterion (AIC), Bayesian Information Criterion (BIC), Sample-Size Adjusted BIC (SABIC), and Hannan-Quinn (HQ) Criterion. When given a sequence of objects, anova tests the models against one another in the order specified. Note that the object inputs should be ordered in terms of most constrained model to least constrained.

Usage

# S4 method for class 'SingleGroupClass'
anova(
  object,
  object2,
  ...,
  bounded = FALSE,
  mix = 0.5,
  frame = 1,
  verbose = FALSE
)

Arguments

object

an object of class SingleGroupClass, MultipleGroupClass, or MixedClass, reflecting the most constrained model fitted

object2

a second model estimated from any of the mirt package estimation methods

...

additional less constrained model objects to be compared sequentially to the previous model

bounded

logical; are the two models comparing a bounded parameter (e.g., comparing a single 2PL and 3PL model with 1 df)? If TRUE then a 50:50 mix of chi-squared distributions is used to obtain the p-value

mix

proportion of chi-squared mixtures. Default is 0.5

frame

(internal parameter not for standard use)

verbose

(deprecated argument)

Value

a data.frame/mirt_df object

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi:10.18637/jss.v048.i06

Examples

# \donttest{
x <- mirt(Science, 1)
#> Iteration: 1, Log-Lik: -1629.361, Max-Change: 0.50660Iteration: 2, Log-Lik: -1617.374, Max-Change: 0.25442Iteration: 3, Log-Lik: -1612.894, Max-Change: 0.16991Iteration: 4, Log-Lik: -1610.306, Max-Change: 0.10461Iteration: 5, Log-Lik: -1609.814, Max-Change: 0.09162Iteration: 6, Log-Lik: -1609.534, Max-Change: 0.07363Iteration: 7, Log-Lik: -1609.030, Max-Change: 0.03677Iteration: 8, Log-Lik: -1608.988, Max-Change: 0.03200Iteration: 9, Log-Lik: -1608.958, Max-Change: 0.02754Iteration: 10, Log-Lik: -1608.878, Max-Change: 0.01443Iteration: 11, Log-Lik: -1608.875, Max-Change: 0.00847Iteration: 12, Log-Lik: -1608.873, Max-Change: 0.00515Iteration: 13, Log-Lik: -1608.872, Max-Change: 0.00550Iteration: 14, Log-Lik: -1608.872, Max-Change: 0.00318Iteration: 15, Log-Lik: -1608.871, Max-Change: 0.00462Iteration: 16, Log-Lik: -1608.871, Max-Change: 0.00277Iteration: 17, Log-Lik: -1608.870, Max-Change: 0.00145Iteration: 18, Log-Lik: -1608.870, Max-Change: 0.00175Iteration: 19, Log-Lik: -1608.870, Max-Change: 0.00126Iteration: 20, Log-Lik: -1608.870, Max-Change: 0.00025Iteration: 21, Log-Lik: -1608.870, Max-Change: 0.00285Iteration: 22, Log-Lik: -1608.870, Max-Change: 0.00108Iteration: 23, Log-Lik: -1608.870, Max-Change: 0.00022Iteration: 24, Log-Lik: -1608.870, Max-Change: 0.00059Iteration: 25, Log-Lik: -1608.870, Max-Change: 0.00014Iteration: 26, Log-Lik: -1608.870, Max-Change: 0.00068Iteration: 27, Log-Lik: -1608.870, Max-Change: 0.00065Iteration: 28, Log-Lik: -1608.870, Max-Change: 0.00019Iteration: 29, Log-Lik: -1608.870, Max-Change: 0.00061Iteration: 30, Log-Lik: -1608.870, Max-Change: 0.00012Iteration: 31, Log-Lik: -1608.870, Max-Change: 0.00012Iteration: 32, Log-Lik: -1608.870, Max-Change: 0.00058Iteration: 33, Log-Lik: -1608.870, Max-Change: 0.00055Iteration: 34, Log-Lik: -1608.870, Max-Change: 0.00015Iteration: 35, Log-Lik: -1608.870, Max-Change: 0.00052Iteration: 36, Log-Lik: -1608.870, Max-Change: 0.00010
x2 <- mirt(Science, 2)
#> Iteration: 1, Log-Lik: -1634.562, Max-Change: 0.49365Iteration: 2, Log-Lik: -1608.496, Max-Change: 0.17069Iteration: 3, Log-Lik: -1604.784, Max-Change: 0.08523Iteration: 4, Log-Lik: -1604.195, Max-Change: 0.05345Iteration: 5, Log-Lik: -1603.777, Max-Change: 0.03491Iteration: 6, Log-Lik: -1603.581, Max-Change: 0.02948Iteration: 7, Log-Lik: -1603.263, Max-Change: 0.02614Iteration: 8, Log-Lik: -1603.177, Max-Change: 0.02281Iteration: 9, Log-Lik: -1603.099, Max-Change: 0.02311Iteration: 10, Log-Lik: -1602.717, Max-Change: 0.01961Iteration: 11, Log-Lik: -1602.673, Max-Change: 0.01556Iteration: 12, Log-Lik: -1602.635, Max-Change: 0.01694Iteration: 13, Log-Lik: -1602.443, Max-Change: 0.01534Iteration: 14, Log-Lik: -1602.421, Max-Change: 0.01311Iteration: 15, Log-Lik: -1602.401, Max-Change: 0.01384Iteration: 16, Log-Lik: -1602.303, Max-Change: 0.00987Iteration: 17, Log-Lik: -1602.288, Max-Change: 0.01006Iteration: 18, Log-Lik: -1602.275, Max-Change: 0.00939Iteration: 19, Log-Lik: -1602.206, Max-Change: 0.00972Iteration: 20, Log-Lik: -1602.198, Max-Change: 0.01016Iteration: 21, Log-Lik: -1602.189, Max-Change: 0.00670Iteration: 22, Log-Lik: -1602.165, Max-Change: 0.00679Iteration: 23, Log-Lik: -1602.158, Max-Change: 0.00694Iteration: 24, Log-Lik: -1602.152, Max-Change: 0.00555Iteration: 25, Log-Lik: -1602.116, Max-Change: 0.00689Iteration: 26, Log-Lik: -1602.112, Max-Change: 0.00465Iteration: 27, Log-Lik: -1602.107, Max-Change: 0.00478Iteration: 28, Log-Lik: -1602.082, Max-Change: 0.00389Iteration: 29, Log-Lik: -1602.078, Max-Change: 0.00745Iteration: 30, Log-Lik: -1602.074, Max-Change: 0.00380Iteration: 31, Log-Lik: -1602.069, Max-Change: 0.00373Iteration: 32, Log-Lik: -1602.066, Max-Change: 0.00792Iteration: 33, Log-Lik: -1602.063, Max-Change: 0.00397Iteration: 34, Log-Lik: -1602.060, Max-Change: 0.00364Iteration: 35, Log-Lik: -1602.057, Max-Change: 0.00808Iteration: 36, Log-Lik: -1602.055, Max-Change: 0.00332Iteration: 37, Log-Lik: -1602.052, Max-Change: 0.00613Iteration: 38, Log-Lik: -1602.049, Max-Change: 0.00319Iteration: 39, Log-Lik: -1602.047, Max-Change: 0.00596Iteration: 40, Log-Lik: -1602.043, Max-Change: 0.00303Iteration: 41, Log-Lik: -1602.040, Max-Change: 0.00302Iteration: 42, Log-Lik: -1602.038, Max-Change: 0.00611Iteration: 43, Log-Lik: -1602.034, Max-Change: 0.00392Iteration: 44, Log-Lik: -1602.033, Max-Change: 0.00329Iteration: 45, Log-Lik: -1602.030, Max-Change: 0.00406Iteration: 46, Log-Lik: -1602.026, Max-Change: 0.00303Iteration: 47, Log-Lik: -1602.025, Max-Change: 0.00261Iteration: 48, Log-Lik: -1602.023, Max-Change: 0.00457Iteration: 49, Log-Lik: -1602.021, Max-Change: 0.00265Iteration: 50, Log-Lik: -1602.020, Max-Change: 0.00290Iteration: 51, Log-Lik: -1602.018, Max-Change: 0.00268Iteration: 52, Log-Lik: -1602.010, Max-Change: 0.00310Iteration: 53, Log-Lik: -1602.008, Max-Change: 0.00361Iteration: 54, Log-Lik: -1602.007, Max-Change: 0.00238Iteration: 55, Log-Lik: -1602.006, Max-Change: 0.00327Iteration: 56, Log-Lik: -1602.005, Max-Change: 0.00269Iteration: 57, Log-Lik: -1602.004, Max-Change: 0.00324Iteration: 58, Log-Lik: -1602.002, Max-Change: 0.00218Iteration: 59, Log-Lik: -1602.001, Max-Change: 0.00198Iteration: 60, Log-Lik: -1602.000, Max-Change: 0.00346Iteration: 61, Log-Lik: -1601.999, Max-Change: 0.00226Iteration: 62, Log-Lik: -1601.998, Max-Change: 0.00247Iteration: 63, Log-Lik: -1601.997, Max-Change: 0.00264Iteration: 64, Log-Lik: -1601.996, Max-Change: 0.00380Iteration: 65, Log-Lik: -1601.995, Max-Change: 0.00236Iteration: 66, Log-Lik: -1601.994, Max-Change: 0.00244Iteration: 67, Log-Lik: -1601.992, Max-Change: 0.00164Iteration: 68, Log-Lik: -1601.992, Max-Change: 0.00214Iteration: 69, Log-Lik: -1601.991, Max-Change: 0.00193Iteration: 70, Log-Lik: -1601.990, Max-Change: 0.00267Iteration: 71, Log-Lik: -1601.989, Max-Change: 0.00227Iteration: 72, Log-Lik: -1601.989, Max-Change: 0.00232Iteration: 73, Log-Lik: -1601.988, Max-Change: 0.00172Iteration: 74, Log-Lik: -1601.987, Max-Change: 0.00149Iteration: 75, Log-Lik: -1601.987, Max-Change: 0.00245Iteration: 76, Log-Lik: -1601.986, Max-Change: 0.00149Iteration: 77, Log-Lik: -1601.985, Max-Change: 0.00215Iteration: 78, Log-Lik: -1601.985, Max-Change: 0.00206Iteration: 79, Log-Lik: -1601.984, Max-Change: 0.00223Iteration: 80, Log-Lik: -1601.983, Max-Change: 0.00190Iteration: 81, Log-Lik: -1601.983, Max-Change: 0.00181Iteration: 82, Log-Lik: -1601.982, Max-Change: 0.00125Iteration: 83, Log-Lik: -1601.981, Max-Change: 0.00136Iteration: 84, Log-Lik: -1601.981, Max-Change: 0.00207Iteration: 85, Log-Lik: -1601.981, Max-Change: 0.00124Iteration: 86, Log-Lik: -1601.980, Max-Change: 0.00187Iteration: 87, Log-Lik: -1601.980, Max-Change: 0.00123Iteration: 88, Log-Lik: -1601.980, Max-Change: 0.00193Iteration: 89, Log-Lik: -1601.980, Max-Change: 0.00123Iteration: 90, Log-Lik: -1601.979, Max-Change: 0.00198Iteration: 91, Log-Lik: -1601.979, Max-Change: 0.00118Iteration: 92, Log-Lik: -1601.979, Max-Change: 0.00121Iteration: 93, Log-Lik: -1601.979, Max-Change: 0.00196Iteration: 94, Log-Lik: -1601.978, Max-Change: 0.00112Iteration: 95, Log-Lik: -1601.978, Max-Change: 0.00117Iteration: 96, Log-Lik: -1601.978, Max-Change: 0.00193Iteration: 97, Log-Lik: -1601.978, Max-Change: 0.00106Iteration: 98, Log-Lik: -1601.977, Max-Change: 0.00112Iteration: 99, Log-Lik: -1601.977, Max-Change: 0.00187Iteration: 100, Log-Lik: -1601.977, Max-Change: 0.00103Iteration: 101, Log-Lik: -1601.977, Max-Change: 0.00109Iteration: 102, Log-Lik: -1601.977, Max-Change: 0.00127Iteration: 103, Log-Lik: -1601.976, Max-Change: 0.00118Iteration: 104, Log-Lik: -1601.976, Max-Change: 0.00115Iteration: 105, Log-Lik: -1601.976, Max-Change: 0.00115Iteration: 106, Log-Lik: -1601.975, Max-Change: 0.00108Iteration: 107, Log-Lik: -1601.975, Max-Change: 0.00108Iteration: 108, Log-Lik: -1601.975, Max-Change: 0.00107Iteration: 109, Log-Lik: -1601.974, Max-Change: 0.00103Iteration: 110, Log-Lik: -1601.974, Max-Change: 0.00102Iteration: 111, Log-Lik: -1601.974, Max-Change: 0.00101Iteration: 112, Log-Lik: -1601.974, Max-Change: 0.00083Iteration: 113, Log-Lik: -1601.974, Max-Change: 0.00088Iteration: 114, Log-Lik: -1601.974, Max-Change: 0.00090Iteration: 115, Log-Lik: -1601.973, Max-Change: 0.00108Iteration: 116, Log-Lik: -1601.973, Max-Change: 0.00037Iteration: 117, Log-Lik: -1601.973, Max-Change: 0.00138Iteration: 118, Log-Lik: -1601.973, Max-Change: 0.00320Iteration: 119, Log-Lik: -1601.973, Max-Change: 0.00085Iteration: 120, Log-Lik: -1601.973, Max-Change: 0.00039Iteration: 121, Log-Lik: -1601.973, Max-Change: 0.00034Iteration: 122, Log-Lik: -1601.973, Max-Change: 0.00026Iteration: 123, Log-Lik: -1601.973, Max-Change: 0.00109Iteration: 124, Log-Lik: -1601.973, Max-Change: 0.00128Iteration: 125, Log-Lik: -1601.973, Max-Change: 0.00113Iteration: 126, Log-Lik: -1601.973, Max-Change: 0.00094Iteration: 127, Log-Lik: -1601.973, Max-Change: 0.00080Iteration: 128, Log-Lik: -1601.972, Max-Change: 0.00028Iteration: 129, Log-Lik: -1601.972, Max-Change: 0.00105Iteration: 130, Log-Lik: -1601.972, Max-Change: 0.00291Iteration: 131, Log-Lik: -1601.972, Max-Change: 0.00082Iteration: 132, Log-Lik: -1601.972, Max-Change: 0.00034Iteration: 133, Log-Lik: -1601.972, Max-Change: 0.00029Iteration: 134, Log-Lik: -1601.972, Max-Change: 0.00112Iteration: 135, Log-Lik: -1601.972, Max-Change: 0.00244Iteration: 136, Log-Lik: -1601.972, Max-Change: 0.00039Iteration: 137, Log-Lik: -1601.972, Max-Change: 0.00027Iteration: 138, Log-Lik: -1601.972, Max-Change: 0.00022Iteration: 139, Log-Lik: -1601.972, Max-Change: 0.00112Iteration: 140, Log-Lik: -1601.972, Max-Change: 0.00024Iteration: 141, Log-Lik: -1601.972, Max-Change: 0.00098Iteration: 142, Log-Lik: -1601.972, Max-Change: 0.00261Iteration: 143, Log-Lik: -1601.972, Max-Change: 0.00031Iteration: 144, Log-Lik: -1601.972, Max-Change: 0.00023Iteration: 145, Log-Lik: -1601.972, Max-Change: 0.00208Iteration: 146, Log-Lik: -1601.972, Max-Change: 0.00074Iteration: 147, Log-Lik: -1601.972, Max-Change: 0.00031Iteration: 148, Log-Lik: -1601.972, Max-Change: 0.00027Iteration: 149, Log-Lik: -1601.971, Max-Change: 0.00104Iteration: 150, Log-Lik: -1601.971, Max-Change: 0.00114Iteration: 151, Log-Lik: -1601.971, Max-Change: 0.00026Iteration: 152, Log-Lik: -1601.971, Max-Change: 0.00099Iteration: 153, Log-Lik: -1601.971, Max-Change: 0.00235Iteration: 154, Log-Lik: -1601.971, Max-Change: 0.00028Iteration: 155, Log-Lik: -1601.971, Max-Change: 0.00022Iteration: 156, Log-Lik: -1601.971, Max-Change: 0.00019Iteration: 157, Log-Lik: -1601.971, Max-Change: 0.00104Iteration: 158, Log-Lik: -1601.971, Max-Change: 0.00023Iteration: 159, Log-Lik: -1601.971, Max-Change: 0.00088Iteration: 160, Log-Lik: -1601.971, Max-Change: 0.00085Iteration: 161, Log-Lik: -1601.971, Max-Change: 0.00030Iteration: 162, Log-Lik: -1601.971, Max-Change: 0.00019Iteration: 163, Log-Lik: -1601.971, Max-Change: 0.00075Iteration: 164, Log-Lik: -1601.971, Max-Change: 0.00106Iteration: 165, Log-Lik: -1601.971, Max-Change: 0.00093Iteration: 166, Log-Lik: -1601.971, Max-Change: 0.00081Iteration: 167, Log-Lik: -1601.971, Max-Change: 0.00068Iteration: 168, Log-Lik: -1601.971, Max-Change: 0.00022Iteration: 169, Log-Lik: -1601.971, Max-Change: 0.00020Iteration: 170, Log-Lik: -1601.971, Max-Change: 0.00080Iteration: 171, Log-Lik: -1601.971, Max-Change: 0.00104Iteration: 172, Log-Lik: -1601.971, Max-Change: 0.00026Iteration: 173, Log-Lik: -1601.971, Max-Change: 0.00019Iteration: 174, Log-Lik: -1601.971, Max-Change: 0.00078Iteration: 175, Log-Lik: -1601.971, Max-Change: 0.00103Iteration: 176, Log-Lik: -1601.971, Max-Change: 0.00071Iteration: 177, Log-Lik: -1601.971, Max-Change: 0.00014Iteration: 178, Log-Lik: -1601.971, Max-Change: 0.00066Iteration: 179, Log-Lik: -1601.971, Max-Change: 0.00021Iteration: 180, Log-Lik: -1601.971, Max-Change: 0.00065Iteration: 181, Log-Lik: -1601.971, Max-Change: 0.00147Iteration: 182, Log-Lik: -1601.970, Max-Change: 0.00020Iteration: 183, Log-Lik: -1601.970, Max-Change: 0.00017Iteration: 184, Log-Lik: -1601.970, Max-Change: 0.00062Iteration: 185, Log-Lik: -1601.970, Max-Change: 0.00058Iteration: 186, Log-Lik: -1601.970, Max-Change: 0.00064Iteration: 187, Log-Lik: -1601.970, Max-Change: 0.00105Iteration: 188, Log-Lik: -1601.970, Max-Change: 0.00019Iteration: 189, Log-Lik: -1601.970, Max-Change: 0.00077Iteration: 190, Log-Lik: -1601.970, Max-Change: 0.00071Iteration: 191, Log-Lik: -1601.970, Max-Change: 0.00021Iteration: 192, Log-Lik: -1601.970, Max-Change: 0.00077Iteration: 193, Log-Lik: -1601.970, Max-Change: 0.00081Iteration: 194, Log-Lik: -1601.970, Max-Change: 0.00028Iteration: 195, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 196, Log-Lik: -1601.970, Max-Change: 0.00069Iteration: 197, Log-Lik: -1601.970, Max-Change: 0.00063Iteration: 198, Log-Lik: -1601.970, Max-Change: 0.00020Iteration: 199, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 200, Log-Lik: -1601.970, Max-Change: 0.00068Iteration: 201, Log-Lik: -1601.970, Max-Change: 0.00064Iteration: 202, Log-Lik: -1601.970, Max-Change: 0.00023Iteration: 203, Log-Lik: -1601.970, Max-Change: 0.00016Iteration: 204, Log-Lik: -1601.970, Max-Change: 0.00062Iteration: 205, Log-Lik: -1601.970, Max-Change: 0.00064Iteration: 206, Log-Lik: -1601.970, Max-Change: 0.00021Iteration: 207, Log-Lik: -1601.970, Max-Change: 0.00072Iteration: 208, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 209, Log-Lik: -1601.970, Max-Change: 0.00061Iteration: 210, Log-Lik: -1601.970, Max-Change: 0.00020Iteration: 211, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 212, Log-Lik: -1601.970, Max-Change: 0.00067Iteration: 213, Log-Lik: -1601.970, Max-Change: 0.00065Iteration: 214, Log-Lik: -1601.970, Max-Change: 0.00021Iteration: 215, Log-Lik: -1601.970, Max-Change: 0.00014Iteration: 216, Log-Lik: -1601.970, Max-Change: 0.00058Iteration: 217, Log-Lik: -1601.970, Max-Change: 0.00060Iteration: 218, Log-Lik: -1601.970, Max-Change: 0.00020Iteration: 219, Log-Lik: -1601.970, Max-Change: 0.00068Iteration: 220, Log-Lik: -1601.970, Max-Change: 0.00019Iteration: 221, Log-Lik: -1601.970, Max-Change: 0.00059Iteration: 222, Log-Lik: -1601.970, Max-Change: 0.00019Iteration: 223, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 224, Log-Lik: -1601.970, Max-Change: 0.00064Iteration: 225, Log-Lik: -1601.970, Max-Change: 0.00013Iteration: 226, Log-Lik: -1601.970, Max-Change: 0.00059Iteration: 227, Log-Lik: -1601.970, Max-Change: 0.00020Iteration: 228, Log-Lik: -1601.970, Max-Change: 0.00013Iteration: 229, Log-Lik: -1601.970, Max-Change: 0.00054Iteration: 230, Log-Lik: -1601.970, Max-Change: 0.00052Iteration: 231, Log-Lik: -1601.970, Max-Change: 0.00015Iteration: 232, Log-Lik: -1601.970, Max-Change: 0.00014Iteration: 233, Log-Lik: -1601.970, Max-Change: 0.00055Iteration: 234, Log-Lik: -1601.970, Max-Change: 0.00052Iteration: 235, Log-Lik: -1601.970, Max-Change: 0.00019Iteration: 236, Log-Lik: -1601.970, Max-Change: 0.00013Iteration: 237, Log-Lik: -1601.970, Max-Change: 0.00053Iteration: 238, Log-Lik: -1601.970, Max-Change: 0.00054Iteration: 239, Log-Lik: -1601.970, Max-Change: 0.00017Iteration: 240, Log-Lik: -1601.970, Max-Change: 0.00062Iteration: 241, Log-Lik: -1601.970, Max-Change: 0.00017Iteration: 242, Log-Lik: -1601.970, Max-Change: 0.00053Iteration: 243, Log-Lik: -1601.970, Max-Change: 0.00017Iteration: 244, Log-Lik: -1601.970, Max-Change: 0.00016Iteration: 245, Log-Lik: -1601.970, Max-Change: 0.00058Iteration: 246, Log-Lik: -1601.970, Max-Change: 0.00011Iteration: 247, Log-Lik: -1601.970, Max-Change: 0.00054Iteration: 248, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 249, Log-Lik: -1601.970, Max-Change: 0.00062Iteration: 250, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 251, Log-Lik: -1601.970, Max-Change: 0.00054Iteration: 252, Log-Lik: -1601.970, Max-Change: 0.00017Iteration: 253, Log-Lik: -1601.970, Max-Change: 0.00016Iteration: 254, Log-Lik: -1601.970, Max-Change: 0.00058Iteration: 255, Log-Lik: -1601.970, Max-Change: 0.00012Iteration: 256, Log-Lik: -1601.970, Max-Change: 0.00055Iteration: 257, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 258, Log-Lik: -1601.970, Max-Change: 0.00012Iteration: 259, Log-Lik: -1601.970, Max-Change: 0.00050Iteration: 260, Log-Lik: -1601.969, Max-Change: 0.00048Iteration: 261, Log-Lik: -1601.969, Max-Change: 0.00014Iteration: 262, Log-Lik: -1601.969, Max-Change: 0.00013Iteration: 263, Log-Lik: -1601.969, Max-Change: 0.00051Iteration: 264, Log-Lik: -1601.969, Max-Change: 0.00049Iteration: 265, Log-Lik: -1601.969, Max-Change: 0.00017Iteration: 266, Log-Lik: -1601.969, Max-Change: 0.00059Iteration: 267, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 268, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 269, Log-Lik: -1601.969, Max-Change: 0.00047Iteration: 270, Log-Lik: -1601.969, Max-Change: 0.00013Iteration: 271, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 272, Log-Lik: -1601.969, Max-Change: 0.00049Iteration: 273, Log-Lik: -1601.969, Max-Change: 0.00047Iteration: 274, Log-Lik: -1601.969, Max-Change: 0.00016Iteration: 275, Log-Lik: -1601.969, Max-Change: 0.00058Iteration: 276, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 277, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 278, Log-Lik: -1601.969, Max-Change: 0.00046Iteration: 279, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 280, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 281, Log-Lik: -1601.969, Max-Change: 0.00048Iteration: 282, Log-Lik: -1601.969, Max-Change: 0.00047Iteration: 283, Log-Lik: -1601.969, Max-Change: 0.00016Iteration: 284, Log-Lik: -1601.969, Max-Change: 0.00056Iteration: 285, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 286, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 287, Log-Lik: -1601.969, Max-Change: 0.00045Iteration: 288, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 289, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 290, Log-Lik: -1601.969, Max-Change: 0.00047Iteration: 291, Log-Lik: -1601.969, Max-Change: 0.00046Iteration: 292, Log-Lik: -1601.969, Max-Change: 0.00016Iteration: 293, Log-Lik: -1601.969, Max-Change: 0.00055Iteration: 294, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 295, Log-Lik: -1601.969, Max-Change: 0.00010Iteration: 296, Log-Lik: -1601.969, Max-Change: 0.00044Iteration: 297, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 298, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 299, Log-Lik: -1601.969, Max-Change: 0.00046Iteration: 300, Log-Lik: -1601.969, Max-Change: 0.00045Iteration: 301, Log-Lik: -1601.969, Max-Change: 0.00015Iteration: 302, Log-Lik: -1601.969, Max-Change: 0.00054Iteration: 303, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 304, Log-Lik: -1601.969, Max-Change: 0.00010Iteration: 305, Log-Lik: -1601.969, Max-Change: 0.00043Iteration: 306, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 307, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 308, Log-Lik: -1601.969, Max-Change: 0.00045Iteration: 309, Log-Lik: -1601.969, Max-Change: 0.00044Iteration: 310, Log-Lik: -1601.969, Max-Change: 0.00015Iteration: 311, Log-Lik: -1601.969, Max-Change: 0.00053Iteration: 312, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 313, Log-Lik: -1601.969, Max-Change: 0.00010
anova(x, x2)
#>         AIC    SABIC       HQ      BIC    logLik     X2 df     p
#> x  3249.739 3262.512 3274.922 3313.279 -1608.870                
#> x2 3241.938 3257.106 3271.843 3317.392 -1601.969 13.801  3 0.003

# compare three models sequentially (X2 not always meaningful)
x3 <- mirt(Science, 1, 'gpcm')
#> Iteration: 1, Log-Lik: -1689.735, Max-Change: 1.30401Iteration: 2, Log-Lik: -1618.015, Max-Change: 0.28548Iteration: 3, Log-Lik: -1615.635, Max-Change: 0.27916Iteration: 4, Log-Lik: -1613.549, Max-Change: 0.13014Iteration: 5, Log-Lik: -1613.306, Max-Change: 0.08712Iteration: 6, Log-Lik: -1613.172, Max-Change: 0.09075Iteration: 7, Log-Lik: -1612.802, Max-Change: 0.06897Iteration: 8, Log-Lik: -1612.774, Max-Change: 0.05440Iteration: 9, Log-Lik: -1612.756, Max-Change: 0.03767Iteration: 10, Log-Lik: -1612.733, Max-Change: 0.03981Iteration: 11, Log-Lik: -1612.724, Max-Change: 0.02653Iteration: 12, Log-Lik: -1612.718, Max-Change: 0.03424Iteration: 13, Log-Lik: -1612.700, Max-Change: 0.00646Iteration: 14, Log-Lik: -1612.697, Max-Change: 0.01774Iteration: 15, Log-Lik: -1612.695, Max-Change: 0.01250Iteration: 16, Log-Lik: -1612.692, Max-Change: 0.01716Iteration: 17, Log-Lik: -1612.690, Max-Change: 0.01090Iteration: 18, Log-Lik: -1612.689, Max-Change: 0.01170Iteration: 19, Log-Lik: -1612.687, Max-Change: 0.00599Iteration: 20, Log-Lik: -1612.687, Max-Change: 0.00160Iteration: 21, Log-Lik: -1612.687, Max-Change: 0.00208Iteration: 22, Log-Lik: -1612.687, Max-Change: 0.02543Iteration: 23, Log-Lik: -1612.685, Max-Change: 0.00161Iteration: 24, Log-Lik: -1612.685, Max-Change: 0.00099Iteration: 25, Log-Lik: -1612.685, Max-Change: 0.00044Iteration: 26, Log-Lik: -1612.685, Max-Change: 0.00027Iteration: 27, Log-Lik: -1612.685, Max-Change: 0.00102Iteration: 28, Log-Lik: -1612.685, Max-Change: 0.00049Iteration: 29, Log-Lik: -1612.685, Max-Change: 0.00020Iteration: 30, Log-Lik: -1612.685, Max-Change: 0.00019Iteration: 31, Log-Lik: -1612.685, Max-Change: 0.01678Iteration: 32, Log-Lik: -1612.684, Max-Change: 0.00133Iteration: 33, Log-Lik: -1612.684, Max-Change: 0.00078Iteration: 34, Log-Lik: -1612.684, Max-Change: 0.00030Iteration: 35, Log-Lik: -1612.684, Max-Change: 0.00090Iteration: 36, Log-Lik: -1612.684, Max-Change: 0.00027Iteration: 37, Log-Lik: -1612.684, Max-Change: 0.00014Iteration: 38, Log-Lik: -1612.684, Max-Change: 0.00083Iteration: 39, Log-Lik: -1612.684, Max-Change: 0.00024Iteration: 40, Log-Lik: -1612.684, Max-Change: 0.00012Iteration: 41, Log-Lik: -1612.684, Max-Change: 0.01511Iteration: 42, Log-Lik: -1612.683, Max-Change: 0.00108Iteration: 43, Log-Lik: -1612.683, Max-Change: 0.00105Iteration: 44, Log-Lik: -1612.683, Max-Change: 0.00038Iteration: 45, Log-Lik: -1612.683, Max-Change: 0.00111Iteration: 46, Log-Lik: -1612.683, Max-Change: 0.00022Iteration: 47, Log-Lik: -1612.683, Max-Change: 0.00015Iteration: 48, Log-Lik: -1612.683, Max-Change: 0.00063Iteration: 49, Log-Lik: -1612.683, Max-Change: 0.00021Iteration: 50, Log-Lik: -1612.683, Max-Change: 0.00009
x4 <- mirt(Science, 1, 'nominal')
#> Iteration: 1, Log-Lik: -2231.749, Max-Change: 3.43133Iteration: 2, Log-Lik: -1647.755, Max-Change: 0.78640Iteration: 3, Log-Lik: -1629.925, Max-Change: 0.48800Iteration: 4, Log-Lik: -1621.546, Max-Change: 0.44557Iteration: 5, Log-Lik: -1616.776, Max-Change: 0.30268Iteration: 6, Log-Lik: -1613.789, Max-Change: 0.29869Iteration: 7, Log-Lik: -1610.323, Max-Change: 0.20615Iteration: 8, Log-Lik: -1609.697, Max-Change: 0.28886Iteration: 9, Log-Lik: -1609.348, Max-Change: 0.13388Iteration: 10, Log-Lik: -1609.105, Max-Change: 0.11615Iteration: 11, Log-Lik: -1608.975, Max-Change: 0.10205Iteration: 12, Log-Lik: -1608.881, Max-Change: 0.09345Iteration: 13, Log-Lik: -1608.596, Max-Change: 0.06321Iteration: 14, Log-Lik: -1608.565, Max-Change: 0.03301Iteration: 15, Log-Lik: -1608.548, Max-Change: 0.02973Iteration: 16, Log-Lik: -1608.507, Max-Change: 0.01790Iteration: 17, Log-Lik: -1608.500, Max-Change: 0.02676Iteration: 18, Log-Lik: -1608.494, Max-Change: 0.02509Iteration: 19, Log-Lik: -1608.472, Max-Change: 0.00996Iteration: 20, Log-Lik: -1608.468, Max-Change: 0.00382Iteration: 21, Log-Lik: -1608.467, Max-Change: 0.00338Iteration: 22, Log-Lik: -1608.464, Max-Change: 0.03036Iteration: 23, Log-Lik: -1608.462, Max-Change: 0.00312Iteration: 24, Log-Lik: -1608.462, Max-Change: 0.00244Iteration: 25, Log-Lik: -1608.461, Max-Change: 0.00209Iteration: 26, Log-Lik: -1608.460, Max-Change: 0.00238Iteration: 27, Log-Lik: -1608.460, Max-Change: 0.00206Iteration: 28, Log-Lik: -1608.460, Max-Change: 0.02332Iteration: 29, Log-Lik: -1608.458, Max-Change: 0.00177Iteration: 30, Log-Lik: -1608.458, Max-Change: 0.00144Iteration: 31, Log-Lik: -1608.458, Max-Change: 0.00192Iteration: 32, Log-Lik: -1608.458, Max-Change: 0.00042Iteration: 33, Log-Lik: -1608.458, Max-Change: 0.00168Iteration: 34, Log-Lik: -1608.458, Max-Change: 0.00024Iteration: 35, Log-Lik: -1608.458, Max-Change: 0.00416Iteration: 36, Log-Lik: -1608.458, Max-Change: 0.00063Iteration: 37, Log-Lik: -1608.457, Max-Change: 0.00059Iteration: 38, Log-Lik: -1608.457, Max-Change: 0.00021Iteration: 39, Log-Lik: -1608.457, Max-Change: 0.00124Iteration: 40, Log-Lik: -1608.457, Max-Change: 0.00057Iteration: 41, Log-Lik: -1608.457, Max-Change: 0.00036Iteration: 42, Log-Lik: -1608.457, Max-Change: 0.00023Iteration: 43, Log-Lik: -1608.457, Max-Change: 0.02043Iteration: 44, Log-Lik: -1608.456, Max-Change: 0.00064Iteration: 45, Log-Lik: -1608.456, Max-Change: 0.00021Iteration: 46, Log-Lik: -1608.456, Max-Change: 0.00104Iteration: 47, Log-Lik: -1608.456, Max-Change: 0.00061Iteration: 48, Log-Lik: -1608.456, Max-Change: 0.00021Iteration: 49, Log-Lik: -1608.456, Max-Change: 0.00017Iteration: 50, Log-Lik: -1608.456, Max-Change: 0.00168Iteration: 51, Log-Lik: -1608.456, Max-Change: 0.00018Iteration: 52, Log-Lik: -1608.456, Max-Change: 0.00017Iteration: 53, Log-Lik: -1608.456, Max-Change: 0.00250Iteration: 54, Log-Lik: -1608.456, Max-Change: 0.00019Iteration: 55, Log-Lik: -1608.456, Max-Change: 0.00018Iteration: 56, Log-Lik: -1608.456, Max-Change: 0.02338Iteration: 57, Log-Lik: -1608.455, Max-Change: 0.00081Iteration: 58, Log-Lik: -1608.455, Max-Change: 0.00080Iteration: 59, Log-Lik: -1608.455, Max-Change: 0.00022Iteration: 60, Log-Lik: -1608.455, Max-Change: 0.00084Iteration: 61, Log-Lik: -1608.455, Max-Change: 0.00040Iteration: 62, Log-Lik: -1608.455, Max-Change: 0.00033Iteration: 63, Log-Lik: -1608.455, Max-Change: 0.00026Iteration: 64, Log-Lik: -1608.455, Max-Change: 0.00126Iteration: 65, Log-Lik: -1608.455, Max-Change: 0.00015Iteration: 66, Log-Lik: -1608.455, Max-Change: 0.00083Iteration: 67, Log-Lik: -1608.455, Max-Change: 0.00022Iteration: 68, Log-Lik: -1608.455, Max-Change: 0.00018Iteration: 69, Log-Lik: -1608.455, Max-Change: 0.00014Iteration: 70, Log-Lik: -1608.455, Max-Change: 0.00151Iteration: 71, Log-Lik: -1608.455, Max-Change: 0.00007
anova(x, x2, x3, x4)
#>         AIC    SABIC       HQ      BIC    logLik      X2 df     p
#> x  3249.739 3262.512 3274.922 3313.279 -1608.870                 
#> x2 3241.938 3257.106 3271.843 3317.392 -1601.969  13.801  3 0.003
#> x3 3257.366 3270.139 3282.549 3320.906 -1612.683 -21.428 -3   NaN
#> x4 3264.910 3284.069 3302.684 3360.220 -1608.455   8.456  8  0.39

# in isolation
anova(x)
#>        AIC    SABIC       HQ      BIC   logLik
#> x 3249.739 3262.512 3274.922 3313.279 -1608.87

# with priors on first model
model <- "Theta = 1-4
          PRIOR = (1-4, a1, lnorm, 0, 10)"
xp <- mirt(Science, model)
#> Iteration: 1, Log-Lik: -1634.928, Max-Change: 0.54839Iteration: 2, Log-Lik: -1627.674, Max-Change: 0.22260Iteration: 3, Log-Lik: -1625.422, Max-Change: 0.15388Iteration: 4, Log-Lik: -1623.848, Max-Change: 0.08425Iteration: 5, Log-Lik: -1623.573, Max-Change: 0.07215Iteration: 6, Log-Lik: -1623.386, Max-Change: 0.06194Iteration: 7, Log-Lik: -1622.910, Max-Change: 0.02298Iteration: 8, Log-Lik: -1622.901, Max-Change: 0.01549Iteration: 9, Log-Lik: -1622.895, Max-Change: 0.01268Iteration: 10, Log-Lik: -1622.884, Max-Change: 0.00471Iteration: 11, Log-Lik: -1622.883, Max-Change: 0.00444Iteration: 12, Log-Lik: -1622.883, Max-Change: 0.00403Iteration: 13, Log-Lik: -1622.881, Max-Change: 0.00138Iteration: 14, Log-Lik: -1622.881, Max-Change: 0.00122Iteration: 15, Log-Lik: -1622.881, Max-Change: 0.00100Iteration: 16, Log-Lik: -1622.881, Max-Change: 0.00051Iteration: 17, Log-Lik: -1622.881, Max-Change: 0.00050Iteration: 18, Log-Lik: -1622.881, Max-Change: 0.00042Iteration: 19, Log-Lik: -1622.881, Max-Change: 0.00021Iteration: 20, Log-Lik: -1622.881, Max-Change: 0.00013Iteration: 21, Log-Lik: -1622.881, Max-Change: 0.00012Iteration: 22, Log-Lik: -1622.881, Max-Change: 0.00005
anova(xp, x2)
#>         AIC    SABIC       HQ      BIC    logLik   logPost df
#> xp 3249.829 3262.602 3275.012 3313.369 -1608.914 -1622.881 NA
#> x2 3241.938 3257.106 3271.843 3317.392 -1601.969 -1601.969  3
anova(xp)
#>         AIC    SABIC       HQ      BIC    logLik   logPost
#> xp 3249.829 3262.602 3275.012 3313.369 -1608.914 -1622.881

# bounded parameter
dat <- expand.table(LSAT7)
mod <- mirt(dat, 1)
#> Iteration: 1, Log-Lik: -2668.786, Max-Change: 0.18243Iteration: 2, Log-Lik: -2663.691, Max-Change: 0.13637Iteration: 3, Log-Lik: -2661.454, Max-Change: 0.10231Iteration: 4, Log-Lik: -2659.430, Max-Change: 0.04181Iteration: 5, Log-Lik: -2659.241, Max-Change: 0.03417Iteration: 6, Log-Lik: -2659.113, Max-Change: 0.02911Iteration: 7, Log-Lik: -2658.812, Max-Change: 0.00456Iteration: 8, Log-Lik: -2658.809, Max-Change: 0.00363Iteration: 9, Log-Lik: -2658.808, Max-Change: 0.00273Iteration: 10, Log-Lik: -2658.806, Max-Change: 0.00144Iteration: 11, Log-Lik: -2658.806, Max-Change: 0.00118Iteration: 12, Log-Lik: -2658.806, Max-Change: 0.00101Iteration: 13, Log-Lik: -2658.805, Max-Change: 0.00042Iteration: 14, Log-Lik: -2658.805, Max-Change: 0.00025Iteration: 15, Log-Lik: -2658.805, Max-Change: 0.00026Iteration: 16, Log-Lik: -2658.805, Max-Change: 0.00023Iteration: 17, Log-Lik: -2658.805, Max-Change: 0.00023Iteration: 18, Log-Lik: -2658.805, Max-Change: 0.00021Iteration: 19, Log-Lik: -2658.805, Max-Change: 0.00019Iteration: 20, Log-Lik: -2658.805, Max-Change: 0.00017Iteration: 21, Log-Lik: -2658.805, Max-Change: 0.00017Iteration: 22, Log-Lik: -2658.805, Max-Change: 0.00015Iteration: 23, Log-Lik: -2658.805, Max-Change: 0.00015Iteration: 24, Log-Lik: -2658.805, Max-Change: 0.00013Iteration: 25, Log-Lik: -2658.805, Max-Change: 0.00013Iteration: 26, Log-Lik: -2658.805, Max-Change: 0.00011Iteration: 27, Log-Lik: -2658.805, Max-Change: 0.00011Iteration: 28, Log-Lik: -2658.805, Max-Change: 0.00010
mod2 <- mirt(dat, 1, itemtype = c(rep('2PL', 4), '3PL'))
#> Iteration: 1, Log-Lik: -2672.054, Max-Change: 0.24544Iteration: 2, Log-Lik: -2663.819, Max-Change: 0.13674Iteration: 3, Log-Lik: -2661.462, Max-Change: 0.10270Iteration: 4, Log-Lik: -2659.415, Max-Change: 0.04081Iteration: 5, Log-Lik: -2659.229, Max-Change: 0.03430Iteration: 6, Log-Lik: -2659.100, Max-Change: 0.02876Iteration: 7, Log-Lik: -2658.813, Max-Change: 0.00848Iteration: 8, Log-Lik: -2658.806, Max-Change: 0.00676Iteration: 9, Log-Lik: -2658.803, Max-Change: 0.00505Iteration: 10, Log-Lik: -2658.799, Max-Change: 0.00378Iteration: 11, Log-Lik: -2658.798, Max-Change: 0.00349Iteration: 12, Log-Lik: -2658.797, Max-Change: 0.00290Iteration: 13, Log-Lik: -2658.796, Max-Change: 0.00239Iteration: 14, Log-Lik: -2658.795, Max-Change: 0.00188Iteration: 15, Log-Lik: -2658.795, Max-Change: 0.00163Iteration: 16, Log-Lik: -2658.794, Max-Change: 0.00101Iteration: 17, Log-Lik: -2658.794, Max-Change: 0.00057Iteration: 18, Log-Lik: -2658.794, Max-Change: 0.00043Iteration: 19, Log-Lik: -2658.794, Max-Change: 0.00044Iteration: 20, Log-Lik: -2658.794, Max-Change: 0.00040Iteration: 21, Log-Lik: -2658.794, Max-Change: 0.00038Iteration: 22, Log-Lik: -2658.794, Max-Change: 0.00031Iteration: 23, Log-Lik: -2658.794, Max-Change: 0.00032Iteration: 24, Log-Lik: -2658.794, Max-Change: 0.00029Iteration: 25, Log-Lik: -2658.794, Max-Change: 0.00028Iteration: 26, Log-Lik: -2658.794, Max-Change: 0.00025Iteration: 27, Log-Lik: -2658.794, Max-Change: 0.00024Iteration: 28, Log-Lik: -2658.794, Max-Change: 0.00021Iteration: 29, Log-Lik: -2658.794, Max-Change: 0.00021Iteration: 30, Log-Lik: -2658.794, Max-Change: 0.00019Iteration: 31, Log-Lik: -2658.794, Max-Change: 0.00018Iteration: 32, Log-Lik: -2658.794, Max-Change: 0.00016Iteration: 33, Log-Lik: -2658.794, Max-Change: 0.00016Iteration: 34, Log-Lik: -2658.794, Max-Change: 0.00014Iteration: 35, Log-Lik: -2658.794, Max-Change: 0.00014Iteration: 36, Log-Lik: -2658.794, Max-Change: 0.00063Iteration: 37, Log-Lik: -2658.794, Max-Change: 0.00039Iteration: 38, Log-Lik: -2658.794, Max-Change: 0.00028Iteration: 39, Log-Lik: -2658.794, Max-Change: 0.00019Iteration: 40, Log-Lik: -2658.794, Max-Change: 0.00039Iteration: 41, Log-Lik: -2658.794, Max-Change: 0.00028Iteration: 42, Log-Lik: -2658.794, Max-Change: 0.00020Iteration: 43, Log-Lik: -2658.794, Max-Change: 0.00008
anova(mod, mod2) #unbounded test
#>           AIC    SABIC       HQ      BIC    logLik    X2 df    p
#> mod  5337.610 5354.927 5356.263 5386.688 -2658.805              
#> mod2 5339.587 5358.636 5360.106 5393.573 -2658.794 0.023  1 0.88
anova(mod, mod2, bounded = TRUE) #bounded
#>           AIC    SABIC       HQ      BIC    logLik    X2 df    p
#> mod  5337.610 5354.927 5356.263 5386.688 -2658.805              
#> mod2 5339.587 5358.636 5360.106 5393.573 -2658.794 0.023  1 0.44

# priors
model <- 'F = 1-5
          PRIOR = (5, g, norm, -1, 1)'
mod1b <- mirt(dat, model, itemtype = c(rep('2PL', 4), '3PL'))
#> Iteration: 1, Log-Lik: -2679.516, Max-Change: 0.41677Iteration: 2, Log-Lik: -2664.891, Max-Change: 0.13696Iteration: 3, Log-Lik: -2662.401, Max-Change: 0.10292Iteration: 4, Log-Lik: -2660.619, Max-Change: 0.05687Iteration: 5, Log-Lik: -2660.324, Max-Change: 0.04534Iteration: 6, Log-Lik: -2660.134, Max-Change: 0.03724Iteration: 7, Log-Lik: -2659.754, Max-Change: 0.00959Iteration: 8, Log-Lik: -2659.737, Max-Change: 0.00868Iteration: 9, Log-Lik: -2659.728, Max-Change: 0.00782Iteration: 10, Log-Lik: -2659.712, Max-Change: 0.00541Iteration: 11, Log-Lik: -2659.710, Max-Change: 0.00378Iteration: 12, Log-Lik: -2659.709, Max-Change: 0.00346Iteration: 13, Log-Lik: -2659.707, Max-Change: 0.00235Iteration: 14, Log-Lik: -2659.707, Max-Change: 0.00227Iteration: 15, Log-Lik: -2659.706, Max-Change: 0.00196Iteration: 16, Log-Lik: -2659.705, Max-Change: 0.00141Iteration: 17, Log-Lik: -2659.705, Max-Change: 0.00119Iteration: 18, Log-Lik: -2659.705, Max-Change: 0.00108Iteration: 19, Log-Lik: -2659.705, Max-Change: 0.00096Iteration: 20, Log-Lik: -2659.705, Max-Change: 0.00066Iteration: 21, Log-Lik: -2659.705, Max-Change: 0.00045Iteration: 22, Log-Lik: -2659.705, Max-Change: 0.00024Iteration: 23, Log-Lik: -2659.705, Max-Change: 0.00022Iteration: 24, Log-Lik: -2659.705, Max-Change: 0.00021Iteration: 25, Log-Lik: -2659.705, Max-Change: 0.00017Iteration: 26, Log-Lik: -2659.705, Max-Change: 0.00018Iteration: 27, Log-Lik: -2659.705, Max-Change: 0.00016Iteration: 28, Log-Lik: -2659.705, Max-Change: 0.00016Iteration: 29, Log-Lik: -2659.705, Max-Change: 0.00014Iteration: 30, Log-Lik: -2659.705, Max-Change: 0.00014Iteration: 31, Log-Lik: -2659.705, Max-Change: 0.00012Iteration: 32, Log-Lik: -2659.705, Max-Change: 0.00012Iteration: 33, Log-Lik: -2659.705, Max-Change: 0.00011Iteration: 34, Log-Lik: -2659.705, Max-Change: 0.00010Iteration: 35, Log-Lik: -2659.705, Max-Change: 0.00009
anova(mod1b)
#>            AIC   SABIC       HQ      BIC    logLik   logPost
#> mod1b 5339.571 5358.62 5360.089 5393.557 -2658.786 -2659.705

model2 <- 'F = 1-5
          PRIOR = (1-5, g, norm, -1, 1)'
mod2b <- mirt(dat, model2, itemtype = '3PL')
#> Iteration: 1, Log-Lik: -2749.170, Max-Change: 0.61785Iteration: 2, Log-Lik: -2679.057, Max-Change: 0.18138Iteration: 3, Log-Lik: -2669.249, Max-Change: 0.14748Iteration: 4, Log-Lik: -2666.643, Max-Change: 0.11394Iteration: 5, Log-Lik: -2665.604, Max-Change: 0.09133Iteration: 6, Log-Lik: -2665.074, Max-Change: 0.07398Iteration: 7, Log-Lik: -2664.247, Max-Change: 0.02652Iteration: 8, Log-Lik: -2664.176, Max-Change: 0.02518Iteration: 9, Log-Lik: -2664.132, Max-Change: 0.02052Iteration: 10, Log-Lik: -2664.073, Max-Change: 0.01701Iteration: 11, Log-Lik: -2664.058, Max-Change: 0.01404Iteration: 12, Log-Lik: -2664.047, Max-Change: 0.01265Iteration: 13, Log-Lik: -2664.012, Max-Change: 0.00574Iteration: 14, Log-Lik: -2664.011, Max-Change: 0.00366Iteration: 15, Log-Lik: -2664.010, Max-Change: 0.00315Iteration: 16, Log-Lik: -2664.010, Max-Change: 0.00432Iteration: 17, Log-Lik: -2664.009, Max-Change: 0.00232Iteration: 18, Log-Lik: -2664.009, Max-Change: 0.00232Iteration: 19, Log-Lik: -2664.009, Max-Change: 0.00120Iteration: 20, Log-Lik: -2664.008, Max-Change: 0.00033Iteration: 21, Log-Lik: -2664.008, Max-Change: 0.00106Iteration: 22, Log-Lik: -2664.008, Max-Change: 0.00026Iteration: 23, Log-Lik: -2664.008, Max-Change: 0.00094Iteration: 24, Log-Lik: -2664.008, Max-Change: 0.00032Iteration: 25, Log-Lik: -2664.008, Max-Change: 0.00022Iteration: 26, Log-Lik: -2664.008, Max-Change: 0.00084Iteration: 27, Log-Lik: -2664.008, Max-Change: 0.00041Iteration: 28, Log-Lik: -2664.008, Max-Change: 0.00023Iteration: 29, Log-Lik: -2664.008, Max-Change: 0.00081Iteration: 30, Log-Lik: -2664.008, Max-Change: 0.00035Iteration: 31, Log-Lik: -2664.008, Max-Change: 0.00018Iteration: 32, Log-Lik: -2664.008, Max-Change: 0.00076Iteration: 33, Log-Lik: -2664.008, Max-Change: 0.00036Iteration: 34, Log-Lik: -2664.008, Max-Change: 0.00019Iteration: 35, Log-Lik: -2664.008, Max-Change: 0.00073Iteration: 36, Log-Lik: -2664.008, Max-Change: 0.00032Iteration: 37, Log-Lik: -2664.008, Max-Change: 0.00017Iteration: 38, Log-Lik: -2664.008, Max-Change: 0.00070Iteration: 39, Log-Lik: -2664.008, Max-Change: 0.00032Iteration: 40, Log-Lik: -2664.008, Max-Change: 0.00017Iteration: 41, Log-Lik: -2664.008, Max-Change: 0.00067Iteration: 42, Log-Lik: -2664.008, Max-Change: 0.00029Iteration: 43, Log-Lik: -2664.008, Max-Change: 0.00015Iteration: 44, Log-Lik: -2664.008, Max-Change: 0.00064Iteration: 45, Log-Lik: -2664.008, Max-Change: 0.00028Iteration: 46, Log-Lik: -2664.008, Max-Change: 0.00015Iteration: 47, Log-Lik: -2664.008, Max-Change: 0.00062Iteration: 48, Log-Lik: -2664.008, Max-Change: 0.00027Iteration: 49, Log-Lik: -2664.008, Max-Change: 0.00014Iteration: 50, Log-Lik: -2664.008, Max-Change: 0.00060Iteration: 51, Log-Lik: -2664.008, Max-Change: 0.00025Iteration: 52, Log-Lik: -2664.008, Max-Change: 0.00014Iteration: 53, Log-Lik: -2664.008, Max-Change: 0.00058Iteration: 54, Log-Lik: -2664.008, Max-Change: 0.00024Iteration: 55, Log-Lik: -2664.008, Max-Change: 0.00013Iteration: 56, Log-Lik: -2664.008, Max-Change: 0.00055Iteration: 57, Log-Lik: -2664.008, Max-Change: 0.00023Iteration: 58, Log-Lik: -2664.008, Max-Change: 0.00012Iteration: 59, Log-Lik: -2664.008, Max-Change: 0.00053Iteration: 60, Log-Lik: -2664.008, Max-Change: 0.00022Iteration: 61, Log-Lik: -2664.008, Max-Change: 0.00012Iteration: 62, Log-Lik: -2664.008, Max-Change: 0.00051Iteration: 63, Log-Lik: -2664.008, Max-Change: 0.00021Iteration: 64, Log-Lik: -2664.008, Max-Change: 0.00011Iteration: 65, Log-Lik: -2664.008, Max-Change: 0.00049Iteration: 66, Log-Lik: -2664.008, Max-Change: 0.00021Iteration: 67, Log-Lik: -2664.008, Max-Change: 0.00011Iteration: 68, Log-Lik: -2664.008, Max-Change: 0.00048Iteration: 69, Log-Lik: -2664.008, Max-Change: 0.00020Iteration: 70, Log-Lik: -2664.008, Max-Change: 0.00011Iteration: 71, Log-Lik: -2664.008, Max-Change: 0.00046Iteration: 72, Log-Lik: -2664.008, Max-Change: 0.00019Iteration: 73, Log-Lik: -2664.008, Max-Change: 0.00010Iteration: 74, Log-Lik: -2664.008, Max-Change: 0.00044Iteration: 75, Log-Lik: -2664.008, Max-Change: 0.00018Iteration: 76, Log-Lik: -2664.008, Max-Change: 0.00010
anova(mod1b, mod2b)
#>            AIC    SABIC       HQ      BIC    logLik   logPost df
#> mod1b 5339.571 5358.620 5360.089 5393.557 -2658.786 -2659.705 NA
#> mod2b 5348.306 5374.282 5376.286 5421.923 -2659.153 -2664.008  4

# }