boot.LR {mirt}	R Documentation

Parametric bootstrap likleihood-ratio test

Description

Given two fitted models, compute a parametric bootstrap test to determine whether the less restrictive models fits significantly better than the more restricted model. Note that this hypothesis test also works when prior parameter distributions are included for either model. Function can be run in parallel after using a stuitable mirtCluster definition.

Usage

boot.LR(mod, mod2, R = 1000)

Arguments

mod
an estimated model object

mod2
an estimated model object

R
number of parametric bootstraps to use.

Value

a p-value evaluating whether the more restrictive model fits significantly worse than the less restrictive model

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

Examples

## No test: 

#standard
dat <- expand.table(LSAT7)
mod1 <- mirt(dat, 1)
mod2 <- mirt(dat, 1, '3PL')

# standard LR test
anova(mod1, mod2)

## 
## Model 1: mirt(data = dat, model = 1)
## Model 2: mirt(data = dat, model = 1, itemtype = "3PL")

##       AIC     AICc    SABIC      BIC    logLik  X2  df     p
## 1 5337.61 5337.833 5354.927 5386.688 -2658.805 NaN NaN   NaN
## 2 5346.11 5346.598 5372.085 5419.726 -2658.055 1.5   5 0.913

# bootstrap LR test (run in parallel to save time)
mirtCluster()

## mirtCluster() has already been defined

boot.LR(mod1, mod2, R=200)

## [1] 0.278607

## End(No test)

[Package mirt version 1.23.1 Index]