Extract raw coefs from model object

Return a list (or data.frame) of raw item and group level coefficients. Note that while the output to the console is rounded to three digits, the returned list of objects is not. Hence, elements from cfs <- coef(mod); cfs[[1]] will contain the non-rounded results (useful for simulations).

Usage

# S4 method for class 'SingleGroupClass'
coef(
  object,
  CI = 0.95,
  printSE = FALSE,
  rotate = "none",
  Target = NULL,
  IRTpars = FALSE,
  rawug = FALSE,
  as.data.frame = FALSE,
  simplify = FALSE,
  unique = FALSE,
  verbose = TRUE,
  ...
)

Arguments

object

an object of class SingleGroupClass, MultipleGroupClass, or MixedClass

CI

the amount of converged used to compute confidence intervals; default is 95 percent confidence intervals

printSE

logical; print the standard errors instead of the confidence intervals? When IRTpars = TRUE then the delta method will be used to compute the associated standard errors from mirt's default slope-intercept form

rotate

see summary method for details. The default rotation is 'none'

Target

a dummy variable matrix indicting a target rotation pattern

IRTpars

logical; convert slope intercept parameters into traditional IRT parameters? Only applicable to unidimensional models or models with simple structure (i.e., only one non-zero slope). If a suitable ACOV estimate was computed in the fitted model, and printSE = FALSE, then suitable CIs will be included based on the delta method (where applicable)

rawug

logical; return the untransformed internal g and u parameters? If FALSE, g and u's are converted with the original format along with delta standard errors

as.data.frame

logical; convert list output to a data.frame instead?

simplify

logical; if all items have the same parameter names (indicating they are of the same class) then they are collapsed to a matrix, and a list of length 2 is returned containing a matrix of item parameters and group-level estimates

unique

return the vector of uniquely estimated parameters

verbose

logical; allow information to be printed to the console?

...

additional arguments to be passed

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi:10.18637/jss.v048.i06

See also

Examples

# \donttest{
dat <- expand.table(LSAT7)
x <- mirt(dat, 1)
#> Iteration: 1, Log-Lik: -2668.786, Max-Change: 0.18243Iteration: 2, Log-Lik: -2663.691, Max-Change: 0.13637Iteration: 3, Log-Lik: -2661.454, Max-Change: 0.10231Iteration: 4, Log-Lik: -2659.430, Max-Change: 0.04181Iteration: 5, Log-Lik: -2659.241, Max-Change: 0.03417Iteration: 6, Log-Lik: -2659.113, Max-Change: 0.02911Iteration: 7, Log-Lik: -2658.812, Max-Change: 0.00456Iteration: 8, Log-Lik: -2658.809, Max-Change: 0.00363Iteration: 9, Log-Lik: -2658.808, Max-Change: 0.00273Iteration: 10, Log-Lik: -2658.806, Max-Change: 0.00144Iteration: 11, Log-Lik: -2658.806, Max-Change: 0.00118Iteration: 12, Log-Lik: -2658.806, Max-Change: 0.00101Iteration: 13, Log-Lik: -2658.805, Max-Change: 0.00042Iteration: 14, Log-Lik: -2658.805, Max-Change: 0.00025Iteration: 15, Log-Lik: -2658.805, Max-Change: 0.00026Iteration: 16, Log-Lik: -2658.805, Max-Change: 0.00023Iteration: 17, Log-Lik: -2658.805, Max-Change: 0.00023Iteration: 18, Log-Lik: -2658.805, Max-Change: 0.00021Iteration: 19, Log-Lik: -2658.805, Max-Change: 0.00019Iteration: 20, Log-Lik: -2658.805, Max-Change: 0.00017Iteration: 21, Log-Lik: -2658.805, Max-Change: 0.00017Iteration: 22, Log-Lik: -2658.805, Max-Change: 0.00015Iteration: 23, Log-Lik: -2658.805, Max-Change: 0.00015Iteration: 24, Log-Lik: -2658.805, Max-Change: 0.00013Iteration: 25, Log-Lik: -2658.805, Max-Change: 0.00013Iteration: 26, Log-Lik: -2658.805, Max-Change: 0.00011Iteration: 27, Log-Lik: -2658.805, Max-Change: 0.00011Iteration: 28, Log-Lik: -2658.805, Max-Change: 0.00010
coef(x)
#> $Item.1
#>        a1     d g u
#> par 0.988 1.856 0 1
#> 
#> $Item.2
#>        a1     d g u
#> par 1.081 0.808 0 1
#> 
#> $Item.3
#>        a1     d g u
#> par 1.706 1.804 0 1
#> 
#> $Item.4
#>        a1     d g u
#> par 0.765 0.486 0 1
#> 
#> $Item.5
#>        a1     d g u
#> par 0.736 1.855 0 1
#> 
#> $GroupPars
#>     MEAN_1 COV_11
#> par      0      1
#> 
coef(x, IRTpars = TRUE)
#> $Item.1
#>         a      b g u
#> par 0.988 -1.879 0 1
#> 
#> $Item.2
#>         a      b g u
#> par 1.081 -0.748 0 1
#> 
#> $Item.3
#>         a      b g u
#> par 1.706 -1.058 0 1
#> 
#> $Item.4
#>         a      b g u
#> par 0.765 -0.635 0 1
#> 
#> $Item.5
#>         a     b g u
#> par 0.736 -2.52 0 1
#> 
#> $GroupPars
#>     MEAN_1 COV_11
#> par      0      1
#> 
coef(x, simplify = TRUE)
#> $items
#>           a1     d g u
#> Item.1 0.988 1.856 0 1
#> Item.2 1.081 0.808 0 1
#> Item.3 1.706 1.804 0 1
#> Item.4 0.765 0.486 0 1
#> Item.5 0.736 1.855 0 1
#> 
#> $means
#> F1 
#>  0 
#> 
#> $cov
#>    F1
#> F1  1
#> 

#with computed information matrix
x <- mirt(dat, 1, SE = TRUE)
#> Iteration: 1, Log-Lik: -2668.786, Max-Change: 0.18243Iteration: 2, Log-Lik: -2663.691, Max-Change: 0.13637Iteration: 3, Log-Lik: -2661.454, Max-Change: 0.10231Iteration: 4, Log-Lik: -2659.430, Max-Change: 0.04181Iteration: 5, Log-Lik: -2659.241, Max-Change: 0.03417Iteration: 6, Log-Lik: -2659.113, Max-Change: 0.02911Iteration: 7, Log-Lik: -2658.812, Max-Change: 0.00456Iteration: 8, Log-Lik: -2658.809, Max-Change: 0.00363Iteration: 9, Log-Lik: -2658.808, Max-Change: 0.00273Iteration: 10, Log-Lik: -2658.806, Max-Change: 0.00144Iteration: 11, Log-Lik: -2658.806, Max-Change: 0.00118Iteration: 12, Log-Lik: -2658.806, Max-Change: 0.00101Iteration: 13, Log-Lik: -2658.805, Max-Change: 0.00042Iteration: 14, Log-Lik: -2658.805, Max-Change: 0.00025Iteration: 15, Log-Lik: -2658.805, Max-Change: 0.00026Iteration: 16, Log-Lik: -2658.805, Max-Change: 0.00023Iteration: 17, Log-Lik: -2658.805, Max-Change: 0.00023Iteration: 18, Log-Lik: -2658.805, Max-Change: 0.00021Iteration: 19, Log-Lik: -2658.805, Max-Change: 0.00019Iteration: 20, Log-Lik: -2658.805, Max-Change: 0.00017Iteration: 21, Log-Lik: -2658.805, Max-Change: 0.00017Iteration: 22, Log-Lik: -2658.805, Max-Change: 0.00015Iteration: 23, Log-Lik: -2658.805, Max-Change: 0.00015Iteration: 24, Log-Lik: -2658.805, Max-Change: 0.00013Iteration: 25, Log-Lik: -2658.805, Max-Change: 0.00013Iteration: 26, Log-Lik: -2658.805, Max-Change: 0.00011Iteration: 27, Log-Lik: -2658.805, Max-Change: 0.00011Iteration: 28, Log-Lik: -2658.805, Max-Change: 0.00010
#> 
#> Calculating information matrix...
coef(x)
#> $Item.1
#>            a1     d  g  u
#> par     0.988 1.856  0  1
#> CI_2.5  0.641 1.598 NA NA
#> CI_97.5 1.335 2.114 NA NA
#> 
#> $Item.2
#>            a1     d  g  u
#> par     1.081 0.808  0  1
#> CI_2.5  0.750 0.629 NA NA
#> CI_97.5 1.412 0.987 NA NA
#> 
#> $Item.3
#>            a1     d  g  u
#> par     1.706 1.804  0  1
#> CI_2.5  1.078 1.404 NA NA
#> CI_97.5 2.334 2.205 NA NA
#> 
#> $Item.4
#>            a1     d  g  u
#> par     0.765 0.486  0  1
#> CI_2.5  0.502 0.339 NA NA
#> CI_97.5 1.028 0.633 NA NA
#> 
#> $Item.5
#>            a1     d  g  u
#> par     0.736 1.855  0  1
#> CI_2.5  0.440 1.630 NA NA
#> CI_97.5 1.032 2.079 NA NA
#> 
#> $GroupPars
#>         MEAN_1 COV_11
#> par          0      1
#> CI_2.5      NA     NA
#> CI_97.5     NA     NA
#> 
coef(x, printSE = TRUE)
#> $Item.1
#>        a1     d logit(g) logit(u)
#> par 0.988 1.856     -999      999
#> SE  0.177 0.131       NA       NA
#> 
#> $Item.2
#>        a1     d logit(g) logit(u)
#> par 1.081 0.808     -999      999
#> SE  0.169 0.091       NA       NA
#> 
#> $Item.3
#>        a1     d logit(g) logit(u)
#> par 1.706 1.804     -999      999
#> SE  0.320 0.204       NA       NA
#> 
#> $Item.4
#>        a1     d logit(g) logit(u)
#> par 0.765 0.486     -999      999
#> SE  0.134 0.075       NA       NA
#> 
#> $Item.5
#>        a1     d logit(g) logit(u)
#> par 0.736 1.855     -999      999
#> SE  0.151 0.114       NA       NA
#> 
#> $GroupPars
#>     MEAN_1 COV_11
#> par      0      1
#> SE      NA     NA
#> 
coef(x, as.data.frame = TRUE)
#>                        par    CI_2.5   CI_97.5
#> Item.1.a1        0.9879254 0.6405319 1.3353189
#> Item.1.d         1.8560605 1.5983450 2.1137759
#> Item.1.g         0.0000000        NA        NA
#> Item.1.u         1.0000000        NA        NA
#> Item.2.a1        1.0808847 0.7500334 1.4117360
#> Item.2.d         0.8079786 0.6291264 0.9868309
#> Item.2.g         0.0000000        NA        NA
#> Item.2.u         1.0000000        NA        NA
#> Item.3.a1        1.7058006 1.0778209 2.3337803
#> Item.3.d         1.8042187 1.4035692 2.2048683
#> Item.3.g         0.0000000        NA        NA
#> Item.3.u         1.0000000        NA        NA
#> Item.4.a1        0.7651853 0.5022681 1.0281025
#> Item.4.d         0.4859966 0.3391601 0.6328331
#> Item.4.g         0.0000000        NA        NA
#> Item.4.u         1.0000000        NA        NA
#> Item.5.a1        0.7357980 0.4395386 1.0320574
#> Item.5.d         1.8545127 1.6302516 2.0787739
#> Item.5.g         0.0000000        NA        NA
#> Item.5.u         1.0000000        NA        NA
#> GroupPars.MEAN_1 0.0000000        NA        NA
#> GroupPars.COV_11 1.0000000        NA        NA

#two factors
x2 <- mirt(Science, 2)
#> Iteration: 1, Log-Lik: -1634.562, Max-Change: 0.49365Iteration: 2, Log-Lik: -1608.496, Max-Change: 0.17069Iteration: 3, Log-Lik: -1604.784, Max-Change: 0.08523Iteration: 4, Log-Lik: -1604.195, Max-Change: 0.05345Iteration: 5, Log-Lik: -1603.777, Max-Change: 0.03491Iteration: 6, Log-Lik: -1603.581, Max-Change: 0.02948Iteration: 7, Log-Lik: -1603.263, Max-Change: 0.02614Iteration: 8, Log-Lik: -1603.177, Max-Change: 0.02281Iteration: 9, Log-Lik: -1603.099, Max-Change: 0.02311Iteration: 10, Log-Lik: -1602.717, Max-Change: 0.01961Iteration: 11, Log-Lik: -1602.673, Max-Change: 0.01556Iteration: 12, Log-Lik: -1602.635, Max-Change: 0.01694Iteration: 13, Log-Lik: -1602.443, Max-Change: 0.01534Iteration: 14, Log-Lik: -1602.421, Max-Change: 0.01311Iteration: 15, Log-Lik: -1602.401, Max-Change: 0.01384Iteration: 16, Log-Lik: -1602.303, Max-Change: 0.00987Iteration: 17, Log-Lik: -1602.288, Max-Change: 0.01006Iteration: 18, Log-Lik: -1602.275, Max-Change: 0.00939Iteration: 19, Log-Lik: -1602.206, Max-Change: 0.00972Iteration: 20, Log-Lik: -1602.198, Max-Change: 0.01016Iteration: 21, Log-Lik: -1602.189, Max-Change: 0.00670Iteration: 22, Log-Lik: -1602.165, Max-Change: 0.00679Iteration: 23, Log-Lik: -1602.158, Max-Change: 0.00694Iteration: 24, Log-Lik: -1602.152, Max-Change: 0.00555Iteration: 25, Log-Lik: -1602.116, Max-Change: 0.00689Iteration: 26, Log-Lik: -1602.112, Max-Change: 0.00465Iteration: 27, Log-Lik: -1602.107, Max-Change: 0.00478Iteration: 28, Log-Lik: -1602.082, Max-Change: 0.00389Iteration: 29, Log-Lik: -1602.078, Max-Change: 0.00745Iteration: 30, Log-Lik: -1602.074, Max-Change: 0.00380Iteration: 31, Log-Lik: -1602.069, Max-Change: 0.00373Iteration: 32, Log-Lik: -1602.066, Max-Change: 0.00792Iteration: 33, Log-Lik: -1602.063, Max-Change: 0.00397Iteration: 34, Log-Lik: -1602.060, Max-Change: 0.00364Iteration: 35, Log-Lik: -1602.057, Max-Change: 0.00808Iteration: 36, Log-Lik: -1602.055, Max-Change: 0.00332Iteration: 37, Log-Lik: -1602.052, Max-Change: 0.00613Iteration: 38, Log-Lik: -1602.049, Max-Change: 0.00319Iteration: 39, Log-Lik: -1602.047, Max-Change: 0.00596Iteration: 40, Log-Lik: -1602.043, Max-Change: 0.00303Iteration: 41, Log-Lik: -1602.040, Max-Change: 0.00302Iteration: 42, Log-Lik: -1602.038, Max-Change: 0.00611Iteration: 43, Log-Lik: -1602.034, Max-Change: 0.00392Iteration: 44, Log-Lik: -1602.033, Max-Change: 0.00329Iteration: 45, Log-Lik: -1602.030, Max-Change: 0.00406Iteration: 46, Log-Lik: -1602.026, Max-Change: 0.00303Iteration: 47, Log-Lik: -1602.025, Max-Change: 0.00261Iteration: 48, Log-Lik: -1602.023, Max-Change: 0.00457Iteration: 49, Log-Lik: -1602.021, Max-Change: 0.00265Iteration: 50, Log-Lik: -1602.020, Max-Change: 0.00290Iteration: 51, Log-Lik: -1602.018, Max-Change: 0.00268Iteration: 52, Log-Lik: -1602.010, Max-Change: 0.00310Iteration: 53, Log-Lik: -1602.008, Max-Change: 0.00361Iteration: 54, Log-Lik: -1602.007, Max-Change: 0.00238Iteration: 55, Log-Lik: -1602.006, Max-Change: 0.00327Iteration: 56, Log-Lik: -1602.005, Max-Change: 0.00269Iteration: 57, Log-Lik: -1602.004, Max-Change: 0.00324Iteration: 58, Log-Lik: -1602.002, Max-Change: 0.00218Iteration: 59, Log-Lik: -1602.001, Max-Change: 0.00198Iteration: 60, Log-Lik: -1602.000, Max-Change: 0.00346Iteration: 61, Log-Lik: -1601.999, Max-Change: 0.00226Iteration: 62, Log-Lik: -1601.998, Max-Change: 0.00247Iteration: 63, Log-Lik: -1601.997, Max-Change: 0.00264Iteration: 64, Log-Lik: -1601.996, Max-Change: 0.00380Iteration: 65, Log-Lik: -1601.995, Max-Change: 0.00236Iteration: 66, Log-Lik: -1601.994, Max-Change: 0.00244Iteration: 67, Log-Lik: -1601.992, Max-Change: 0.00164Iteration: 68, Log-Lik: -1601.992, Max-Change: 0.00214Iteration: 69, Log-Lik: -1601.991, Max-Change: 0.00193Iteration: 70, Log-Lik: -1601.990, Max-Change: 0.00267Iteration: 71, Log-Lik: -1601.989, Max-Change: 0.00227Iteration: 72, Log-Lik: -1601.989, Max-Change: 0.00232Iteration: 73, Log-Lik: -1601.988, Max-Change: 0.00172Iteration: 74, Log-Lik: -1601.987, Max-Change: 0.00149Iteration: 75, Log-Lik: -1601.987, Max-Change: 0.00245Iteration: 76, Log-Lik: -1601.986, Max-Change: 0.00149Iteration: 77, Log-Lik: -1601.985, Max-Change: 0.00215Iteration: 78, Log-Lik: -1601.985, Max-Change: 0.00206Iteration: 79, Log-Lik: -1601.984, Max-Change: 0.00223Iteration: 80, Log-Lik: -1601.983, Max-Change: 0.00190Iteration: 81, Log-Lik: -1601.983, Max-Change: 0.00181Iteration: 82, Log-Lik: -1601.982, Max-Change: 0.00125Iteration: 83, Log-Lik: -1601.981, Max-Change: 0.00136Iteration: 84, Log-Lik: -1601.981, Max-Change: 0.00207Iteration: 85, Log-Lik: -1601.981, Max-Change: 0.00124Iteration: 86, Log-Lik: -1601.980, Max-Change: 0.00187Iteration: 87, Log-Lik: -1601.980, Max-Change: 0.00123Iteration: 88, Log-Lik: -1601.980, Max-Change: 0.00193Iteration: 89, Log-Lik: -1601.980, Max-Change: 0.00123Iteration: 90, Log-Lik: -1601.979, Max-Change: 0.00198Iteration: 91, Log-Lik: -1601.979, Max-Change: 0.00118Iteration: 92, Log-Lik: -1601.979, Max-Change: 0.00121Iteration: 93, Log-Lik: -1601.979, Max-Change: 0.00196Iteration: 94, Log-Lik: -1601.978, Max-Change: 0.00112Iteration: 95, Log-Lik: -1601.978, Max-Change: 0.00117Iteration: 96, Log-Lik: -1601.978, Max-Change: 0.00193Iteration: 97, Log-Lik: -1601.978, Max-Change: 0.00106Iteration: 98, Log-Lik: -1601.977, Max-Change: 0.00112Iteration: 99, Log-Lik: -1601.977, Max-Change: 0.00187Iteration: 100, Log-Lik: -1601.977, Max-Change: 0.00103Iteration: 101, Log-Lik: -1601.977, Max-Change: 0.00109Iteration: 102, Log-Lik: -1601.977, Max-Change: 0.00127Iteration: 103, Log-Lik: -1601.976, Max-Change: 0.00118Iteration: 104, Log-Lik: -1601.976, Max-Change: 0.00115Iteration: 105, Log-Lik: -1601.976, Max-Change: 0.00115Iteration: 106, Log-Lik: -1601.975, Max-Change: 0.00108Iteration: 107, Log-Lik: -1601.975, Max-Change: 0.00108Iteration: 108, Log-Lik: -1601.975, Max-Change: 0.00107Iteration: 109, Log-Lik: -1601.974, Max-Change: 0.00103Iteration: 110, Log-Lik: -1601.974, Max-Change: 0.00102Iteration: 111, Log-Lik: -1601.974, Max-Change: 0.00101Iteration: 112, Log-Lik: -1601.974, Max-Change: 0.00083Iteration: 113, Log-Lik: -1601.974, Max-Change: 0.00088Iteration: 114, Log-Lik: -1601.974, Max-Change: 0.00090Iteration: 115, Log-Lik: -1601.973, Max-Change: 0.00108Iteration: 116, Log-Lik: -1601.973, Max-Change: 0.00037Iteration: 117, Log-Lik: -1601.973, Max-Change: 0.00138Iteration: 118, Log-Lik: -1601.973, Max-Change: 0.00320Iteration: 119, Log-Lik: -1601.973, Max-Change: 0.00085Iteration: 120, Log-Lik: -1601.973, Max-Change: 0.00039Iteration: 121, Log-Lik: -1601.973, Max-Change: 0.00034Iteration: 122, Log-Lik: -1601.973, Max-Change: 0.00026Iteration: 123, Log-Lik: -1601.973, Max-Change: 0.00109Iteration: 124, Log-Lik: -1601.973, Max-Change: 0.00128Iteration: 125, Log-Lik: -1601.973, Max-Change: 0.00113Iteration: 126, Log-Lik: -1601.973, Max-Change: 0.00094Iteration: 127, Log-Lik: -1601.973, Max-Change: 0.00080Iteration: 128, Log-Lik: -1601.972, Max-Change: 0.00028Iteration: 129, Log-Lik: -1601.972, Max-Change: 0.00105Iteration: 130, Log-Lik: -1601.972, Max-Change: 0.00291Iteration: 131, Log-Lik: -1601.972, Max-Change: 0.00082Iteration: 132, Log-Lik: -1601.972, Max-Change: 0.00034Iteration: 133, Log-Lik: -1601.972, Max-Change: 0.00029Iteration: 134, Log-Lik: -1601.972, Max-Change: 0.00112Iteration: 135, Log-Lik: -1601.972, Max-Change: 0.00244Iteration: 136, Log-Lik: -1601.972, Max-Change: 0.00039Iteration: 137, Log-Lik: -1601.972, Max-Change: 0.00027Iteration: 138, Log-Lik: -1601.972, Max-Change: 0.00022Iteration: 139, Log-Lik: -1601.972, Max-Change: 0.00112Iteration: 140, Log-Lik: -1601.972, Max-Change: 0.00024Iteration: 141, Log-Lik: -1601.972, Max-Change: 0.00098Iteration: 142, Log-Lik: -1601.972, Max-Change: 0.00261Iteration: 143, Log-Lik: -1601.972, Max-Change: 0.00031Iteration: 144, Log-Lik: -1601.972, Max-Change: 0.00023Iteration: 145, Log-Lik: -1601.972, Max-Change: 0.00208Iteration: 146, Log-Lik: -1601.972, Max-Change: 0.00074Iteration: 147, Log-Lik: -1601.972, Max-Change: 0.00031Iteration: 148, Log-Lik: -1601.972, Max-Change: 0.00027Iteration: 149, Log-Lik: -1601.971, Max-Change: 0.00104Iteration: 150, Log-Lik: -1601.971, Max-Change: 0.00114Iteration: 151, Log-Lik: -1601.971, Max-Change: 0.00026Iteration: 152, Log-Lik: -1601.971, Max-Change: 0.00099Iteration: 153, Log-Lik: -1601.971, Max-Change: 0.00235Iteration: 154, Log-Lik: -1601.971, Max-Change: 0.00028Iteration: 155, Log-Lik: -1601.971, Max-Change: 0.00022Iteration: 156, Log-Lik: -1601.971, Max-Change: 0.00019Iteration: 157, Log-Lik: -1601.971, Max-Change: 0.00104Iteration: 158, Log-Lik: -1601.971, Max-Change: 0.00023Iteration: 159, Log-Lik: -1601.971, Max-Change: 0.00088Iteration: 160, Log-Lik: -1601.971, Max-Change: 0.00085Iteration: 161, Log-Lik: -1601.971, Max-Change: 0.00030Iteration: 162, Log-Lik: -1601.971, Max-Change: 0.00019Iteration: 163, Log-Lik: -1601.971, Max-Change: 0.00075Iteration: 164, Log-Lik: -1601.971, Max-Change: 0.00106Iteration: 165, Log-Lik: -1601.971, Max-Change: 0.00093Iteration: 166, Log-Lik: -1601.971, Max-Change: 0.00081Iteration: 167, Log-Lik: -1601.971, Max-Change: 0.00068Iteration: 168, Log-Lik: -1601.971, Max-Change: 0.00022Iteration: 169, Log-Lik: -1601.971, Max-Change: 0.00020Iteration: 170, Log-Lik: -1601.971, Max-Change: 0.00080Iteration: 171, Log-Lik: -1601.971, Max-Change: 0.00104Iteration: 172, Log-Lik: -1601.971, Max-Change: 0.00026Iteration: 173, Log-Lik: -1601.971, Max-Change: 0.00019Iteration: 174, Log-Lik: -1601.971, Max-Change: 0.00078Iteration: 175, Log-Lik: -1601.971, Max-Change: 0.00103Iteration: 176, Log-Lik: -1601.971, Max-Change: 0.00071Iteration: 177, Log-Lik: -1601.971, Max-Change: 0.00014Iteration: 178, Log-Lik: -1601.971, Max-Change: 0.00066Iteration: 179, Log-Lik: -1601.971, Max-Change: 0.00021Iteration: 180, Log-Lik: -1601.971, Max-Change: 0.00065Iteration: 181, Log-Lik: -1601.971, Max-Change: 0.00147Iteration: 182, Log-Lik: -1601.970, Max-Change: 0.00020Iteration: 183, Log-Lik: -1601.970, Max-Change: 0.00017Iteration: 184, Log-Lik: -1601.970, Max-Change: 0.00062Iteration: 185, Log-Lik: -1601.970, Max-Change: 0.00058Iteration: 186, Log-Lik: -1601.970, Max-Change: 0.00064Iteration: 187, Log-Lik: -1601.970, Max-Change: 0.00105Iteration: 188, Log-Lik: -1601.970, Max-Change: 0.00019Iteration: 189, Log-Lik: -1601.970, Max-Change: 0.00077Iteration: 190, Log-Lik: -1601.970, Max-Change: 0.00071Iteration: 191, Log-Lik: -1601.970, Max-Change: 0.00021Iteration: 192, Log-Lik: -1601.970, Max-Change: 0.00077Iteration: 193, Log-Lik: -1601.970, Max-Change: 0.00081Iteration: 194, Log-Lik: -1601.970, Max-Change: 0.00028Iteration: 195, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 196, Log-Lik: -1601.970, Max-Change: 0.00069Iteration: 197, Log-Lik: -1601.970, Max-Change: 0.00063Iteration: 198, Log-Lik: -1601.970, Max-Change: 0.00020Iteration: 199, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 200, Log-Lik: -1601.970, Max-Change: 0.00068Iteration: 201, Log-Lik: -1601.970, Max-Change: 0.00064Iteration: 202, Log-Lik: -1601.970, Max-Change: 0.00023Iteration: 203, Log-Lik: -1601.970, Max-Change: 0.00016Iteration: 204, Log-Lik: -1601.970, Max-Change: 0.00062Iteration: 205, Log-Lik: -1601.970, Max-Change: 0.00064Iteration: 206, Log-Lik: -1601.970, Max-Change: 0.00021Iteration: 207, Log-Lik: -1601.970, Max-Change: 0.00072Iteration: 208, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 209, Log-Lik: -1601.970, Max-Change: 0.00061Iteration: 210, Log-Lik: -1601.970, Max-Change: 0.00020Iteration: 211, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 212, Log-Lik: -1601.970, Max-Change: 0.00067Iteration: 213, Log-Lik: -1601.970, Max-Change: 0.00065Iteration: 214, Log-Lik: -1601.970, Max-Change: 0.00021Iteration: 215, Log-Lik: -1601.970, Max-Change: 0.00014Iteration: 216, Log-Lik: -1601.970, Max-Change: 0.00058Iteration: 217, Log-Lik: -1601.970, Max-Change: 0.00060Iteration: 218, Log-Lik: -1601.970, Max-Change: 0.00020Iteration: 219, Log-Lik: -1601.970, Max-Change: 0.00068Iteration: 220, Log-Lik: -1601.970, Max-Change: 0.00019Iteration: 221, Log-Lik: -1601.970, Max-Change: 0.00059Iteration: 222, Log-Lik: -1601.970, Max-Change: 0.00019Iteration: 223, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 224, Log-Lik: -1601.970, Max-Change: 0.00064Iteration: 225, Log-Lik: -1601.970, Max-Change: 0.00013Iteration: 226, Log-Lik: -1601.970, Max-Change: 0.00059Iteration: 227, Log-Lik: -1601.970, Max-Change: 0.00020Iteration: 228, Log-Lik: -1601.970, Max-Change: 0.00013Iteration: 229, Log-Lik: -1601.970, Max-Change: 0.00054Iteration: 230, Log-Lik: -1601.970, Max-Change: 0.00052Iteration: 231, Log-Lik: -1601.970, Max-Change: 0.00015Iteration: 232, Log-Lik: -1601.970, Max-Change: 0.00014Iteration: 233, Log-Lik: -1601.970, Max-Change: 0.00055Iteration: 234, Log-Lik: -1601.970, Max-Change: 0.00052Iteration: 235, Log-Lik: -1601.970, Max-Change: 0.00019Iteration: 236, Log-Lik: -1601.970, Max-Change: 0.00013Iteration: 237, Log-Lik: -1601.970, Max-Change: 0.00053Iteration: 238, Log-Lik: -1601.970, Max-Change: 0.00054Iteration: 239, Log-Lik: -1601.970, Max-Change: 0.00017Iteration: 240, Log-Lik: -1601.970, Max-Change: 0.00062Iteration: 241, Log-Lik: -1601.970, Max-Change: 0.00017Iteration: 242, Log-Lik: -1601.970, Max-Change: 0.00053Iteration: 243, Log-Lik: -1601.970, Max-Change: 0.00017Iteration: 244, Log-Lik: -1601.970, Max-Change: 0.00016Iteration: 245, Log-Lik: -1601.970, Max-Change: 0.00058Iteration: 246, Log-Lik: -1601.970, Max-Change: 0.00011Iteration: 247, Log-Lik: -1601.970, Max-Change: 0.00054Iteration: 248, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 249, Log-Lik: -1601.970, Max-Change: 0.00062Iteration: 250, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 251, Log-Lik: -1601.970, Max-Change: 0.00054Iteration: 252, Log-Lik: -1601.970, Max-Change: 0.00017Iteration: 253, Log-Lik: -1601.970, Max-Change: 0.00016Iteration: 254, Log-Lik: -1601.970, Max-Change: 0.00058Iteration: 255, Log-Lik: -1601.970, Max-Change: 0.00012Iteration: 256, Log-Lik: -1601.970, Max-Change: 0.00055Iteration: 257, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 258, Log-Lik: -1601.970, Max-Change: 0.00012Iteration: 259, Log-Lik: -1601.970, Max-Change: 0.00050Iteration: 260, Log-Lik: -1601.969, Max-Change: 0.00048Iteration: 261, Log-Lik: -1601.969, Max-Change: 0.00014Iteration: 262, Log-Lik: -1601.969, Max-Change: 0.00013Iteration: 263, Log-Lik: -1601.969, Max-Change: 0.00051Iteration: 264, Log-Lik: -1601.969, Max-Change: 0.00049Iteration: 265, Log-Lik: -1601.969, Max-Change: 0.00017Iteration: 266, Log-Lik: -1601.969, Max-Change: 0.00059Iteration: 267, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 268, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 269, Log-Lik: -1601.969, Max-Change: 0.00047Iteration: 270, Log-Lik: -1601.969, Max-Change: 0.00013Iteration: 271, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 272, Log-Lik: -1601.969, Max-Change: 0.00049Iteration: 273, Log-Lik: -1601.969, Max-Change: 0.00047Iteration: 274, Log-Lik: -1601.969, Max-Change: 0.00016Iteration: 275, Log-Lik: -1601.969, Max-Change: 0.00058Iteration: 276, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 277, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 278, Log-Lik: -1601.969, Max-Change: 0.00046Iteration: 279, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 280, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 281, Log-Lik: -1601.969, Max-Change: 0.00048Iteration: 282, Log-Lik: -1601.969, Max-Change: 0.00047Iteration: 283, Log-Lik: -1601.969, Max-Change: 0.00016Iteration: 284, Log-Lik: -1601.969, Max-Change: 0.00056Iteration: 285, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 286, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 287, Log-Lik: -1601.969, Max-Change: 0.00045Iteration: 288, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 289, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 290, Log-Lik: -1601.969, Max-Change: 0.00047Iteration: 291, Log-Lik: -1601.969, Max-Change: 0.00046Iteration: 292, Log-Lik: -1601.969, Max-Change: 0.00016Iteration: 293, Log-Lik: -1601.969, Max-Change: 0.00055Iteration: 294, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 295, Log-Lik: -1601.969, Max-Change: 0.00010Iteration: 296, Log-Lik: -1601.969, Max-Change: 0.00044Iteration: 297, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 298, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 299, Log-Lik: -1601.969, Max-Change: 0.00046Iteration: 300, Log-Lik: -1601.969, Max-Change: 0.00045Iteration: 301, Log-Lik: -1601.969, Max-Change: 0.00015Iteration: 302, Log-Lik: -1601.969, Max-Change: 0.00054Iteration: 303, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 304, Log-Lik: -1601.969, Max-Change: 0.00010Iteration: 305, Log-Lik: -1601.969, Max-Change: 0.00043Iteration: 306, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 307, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 308, Log-Lik: -1601.969, Max-Change: 0.00045Iteration: 309, Log-Lik: -1601.969, Max-Change: 0.00044Iteration: 310, Log-Lik: -1601.969, Max-Change: 0.00015Iteration: 311, Log-Lik: -1601.969, Max-Change: 0.00053Iteration: 312, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 313, Log-Lik: -1601.969, Max-Change: 0.00010
coef(x2)
#> $Comfort
#>         a1    a2    d1    d2     d3
#> par -1.335 0.097 5.211 2.866 -1.603
#> 
#> $Work
#>         a1    a2    d1    d2     d3
#> par -0.879 1.853 3.704 1.153 -2.904
#> 
#> $Future
#>        a1    a2    d1    d2     d3
#> par -1.47 1.165 4.663 1.957 -1.736
#> 
#> $Benefit
#>         a1 a2    d1    d2     d3
#> par -1.722  0 3.989 1.195 -2.044
#> 
#> $GroupPars
#>     MEAN_1 MEAN_2 COV_11 COV_21 COV_22
#> par      0      0      1      0      1
#> 
coef(x2, rotate = 'varimax')
#> 
#> Rotation:  varimax 
#> 
#> $Comfort
#>        a1    a2    d1    d2     d3
#> par 1.254 0.468 5.211 2.866 -1.603
#> 
#> $Work
#>        a1    a2    d1    d2     d3
#> par 0.323 2.025 3.704 1.153 -2.904
#> 
#> $Future
#>        a1    a2    d1    d2     d3
#> par 1.083 1.531 4.663 1.957 -1.736
#> 
#> $Benefit
#>        a1    a2    d1    d2     d3
#> par 1.653 0.484 3.989 1.195 -2.044
#> 
#> $GroupPars
#>     MEAN_1 MEAN_2 COV_11 COV_21 COV_22
#> par      0      0      1      0      1
#> 

# }