Extract raw coefs from model object

Return a list (or data.frame) of raw item and group level coefficients. Note that while the output to the console is rounded to three digits, the returned list of objects is not. Hence, elements from cfs <- coef(mod); cfs[[1]] will contain the non-rounded results (useful for simulations).

Usage

# S4 method for class 'SingleGroupClass'
coef(
  object,
  CI = 0.95,
  printSE = FALSE,
  rotate = "none",
  Target = NULL,
  IRTpars = FALSE,
  rawug = FALSE,
  as.data.frame = FALSE,
  simplify = FALSE,
  unique = FALSE,
  verbose = TRUE,
  ...
)

Arguments

object

an object of class SingleGroupClass, MultipleGroupClass, or MixedClass

CI

the amount of converged used to compute confidence intervals; default is 95 percent confidence intervals

printSE

logical; print the standard errors instead of the confidence intervals? When IRTpars = TRUE then the delta method will be used to compute the associated standard errors from mirt's default slope-intercept form

rotate

see summary method for details. The default rotation is 'none'

Target

a dummy variable matrix indicting a target rotation pattern

IRTpars

logical; convert slope intercept parameters into traditional IRT parameters? Only applicable to unidimensional models or models with simple structure (i.e., only one non-zero slope). If a suitable ACOV estimate was computed in the fitted model, and printSE = FALSE, then suitable CIs will be included based on the delta method (where applicable)

rawug

logical; return the untransformed internal g and u parameters? If FALSE, g and u's are converted with the original format along with delta standard errors

as.data.frame

logical; convert list output to a data.frame instead?

simplify

logical; if all items have the same parameter names (indicating they are of the same class) then they are collapsed to a matrix, and a list of length 2 is returned containing a matrix of item parameters and group-level estimates

unique

return the vector of uniquely estimated parameters

verbose

logical; allow information to be printed to the console?

...

additional arguments to be passed

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi:10.18637/jss.v048.i06

See also

Examples

if (FALSE) { # \dontrun{
dat <- expand.table(LSAT7)
x <- mirt(dat, 1)
coef(x)
coef(x, IRTpars = TRUE)
coef(x, simplify = TRUE)

#with computed information matrix
x <- mirt(dat, 1, SE = TRUE)
coef(x)
coef(x, printSE = TRUE)
coef(x, as.data.frame = TRUE)

#two factors
x2 <- mirt(Science, 2)
coef(x2)
coef(x2, rotate = 'varimax')

} # }