createGroup {mirt}	R Documentation

Create a user defined group-level object with correct generic functions

Description

Initializes the proper S4 class and methods necessary for mirt functions to use in estimation for definiting customized group-level functions. To use the defined objects pass to the mirt(..., customGroup = OBJECT) command, and ensure that the class parameters are properly labeled.

Usage

createGroup(par, est, den, nfact, gr = NULL, hss = NULL, gen = NULL,
  lbound = NULL, ubound = NULL, derivType = "Richardson")

Arguments

par
a named vector of the starting values for the parameters

est
a logical vector indicating which parameters should be freely estimated by default

den
the probability density function given the Theta/ability values. First input contains a vector of all the defined parameters and the second input must be a matrix called Theta. Function also must return a numeric vector object corresponding to the associated densities for each row in the Theta input

nfact
number of factors required for the model. E.g., for unidimensional models with only one dimension of integration nfact = 1

gr
gradient function (vector of first derivatives) of the log-likelihood used in estimation. The function must be of the form gr(x, Theta), where x is the object defined by createGroup() and Theta is a matrix of latent trait parameters

hss
Hessian function (matrix of second derivatives) of the log-likelihood used in estimation. If not specified a numeric approximation will be used. The input is idential to the gr argument

gen
a function used when GenRandomPars = TRUE is passed to the estimation function to generate random starting values. Function must be of the form function(object) ... and must return a vector with properties equivalent to the par object. If NULL, parameters will remain at the defined starting values by default

lbound
optional vector indicating the lower bounds of the parameters. If not specified then the bounds will be set to -Inf

ubound
optional vector indicating the lower bounds of the parameters. If not specified then the bounds will be set to Inf

derivType
if the gr or hss terms are not specified this type will be used to obtain them numerically. Default is 'Richardson'

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

Examples

## No test: 

# normal density example
den <- function(obj, Theta) dnorm(Theta, obj@par[1], obj@par[2])
par <- c(mu = 0, sigma = 1)
est <- c(FALSE, TRUE)
lbound <- c(-Inf, 0)
grp <- createGroup(par, est, den, nfact = 1, lbound=lbound)

mod <- mirt(Science, 1, 'Rasch')
modcustom <- mirt(Science, 1, 'Rasch', customGroup=grp)

## Warning in dnorm(Theta, obj@par[1], obj@par[2]): NaNs produced

## Warning in dnorm(Theta, obj@par[1], obj@par[2]): NaNs produced

## Warning in dnorm(Theta, obj@par[1], obj@par[2]): NaNs produced

## Warning in dnorm(Theta, obj@par[1], obj@par[2]): NaNs produced

## Warning in dnorm(Theta, obj@par[1], obj@par[2]): NaNs produced

## Warning in dnorm(Theta, obj@par[1], obj@par[2]): NaNs produced

## Warning in dnorm(Theta, obj@par[1], obj@par[2]): NaNs produced

## Warning in dnorm(Theta, obj@par[1], obj@par[2]): NaNs produced

coef(mod)

## $Comfort
##     a1 ak0 ak1 ak2 ak3 d0    d1  d2    d3
## par  1   0   1   2   3  0 3.102 5.7 4.311
## 
## $Work
##     a1 ak0 ak1 ak2 ak3 d0  d1    d2    d3
## par  1   0   1   2   3  0 1.9 2.812 0.953
## 
## $Future
##     a1 ak0 ak1 ak2 ak3 d0    d1    d2    d3
## par  1   0   1   2   3  0 2.654 4.076 2.941
## 
## $Benefit
##     a1 ak0 ak1 ak2 ak3 d0    d1    d2    d3
## par  1   0   1   2   3  0 2.454 3.355 1.996
## 
## $GroupPars
##     MEAN_1 COV_11
## par      0  1.006

coef(modcustom)

## $Comfort
##     a1 ak0 ak1 ak2 ak3 d0    d1    d2    d3
## par  1   0   1   2   3  0 3.073 5.666 4.277
## 
## $Work
##     a1 ak0 ak1 ak2 ak3 d0   d1    d2   d3
## par  1   0   1   2   3  0 1.89 2.799 0.94
## 
## $Future
##     a1 ak0 ak1 ak2 ak3 d0    d1    d2    d3
## par  1   0   1   2   3  0 2.639 4.058 2.922
## 
## $Benefit
##     a1 ak0 ak1 ak2 ak3 d0    d1   d2    d3
## par  1   0   1   2   3  0 2.442 3.34 1.982
## 
## $GroupPars
##     mu sigma
## par  0     1

## End(No test)

[Package mirt version 1.19 Index]