Draw plausible parameter instantiations from a given model

Draws plausible parameters from a model using parametric sampling (if the information matrix was computed) or via boostrap sampling. Primarily for use with the DRF function.

draw_parameters(
  mod,
  draws,
  method = c("parametric", "boostrap"),
  redraws = 20,
  ...
)

Arguments

mod
estimated single or multiple-group model

draws
number of draws to obtain

method
type of plausible values to obtain. Can be 'parametric', for the parametric sampling scheme which uses the estimated information matrix, or 'boostrap' to obtain values from the boot function. Default is 'parametric'

redraws
number of redraws to perform when the given parameteric sample does not satisfy the upper and lower parameter bounds. If a valid set cannot be found within this number of draws then an error will be thrown

...
additional arguments to be passed

Value

returns a draws x p matrix of plausible parameters, where each row correspeonds to a single set

Examples

# \donttest{
set.seed(1234)
n <- 40
N <- 500

# only first 5 items as anchors
model <- 'F = 1-40
          CONSTRAINB = (1-5, a1), (1-5, d)'

a <- matrix(1, n)
d <- matrix(rnorm(n), n)
group <- c(rep('Group_1', N), rep('Group_2', N))

## -------------
# groups completely equal
dat1 <- simdata(a, d, N, itemtype = 'dich')
dat2 <- simdata(a, d, N, itemtype = 'dich')
dat <- rbind(dat1, dat2)
mod <- multipleGroup(dat, model, group=group, SE=TRUE,
                     invariance=c('free_means', 'free_var'))

param_set <- draw_parameters(mod, 100)
head(param_set)

#>           [,1]      [,2] [,3] [,4]     [,5]      [,6] [,7] [,8]      [,9]
#> [1,] 0.9508082 -1.090240    0    1 1.131674 0.5995145    0    1 1.4946106
#> [2,] 1.1062395 -1.381437    0    1 1.210205 0.3462211    0    1 1.2580931
#> [3,] 1.1103972 -1.287658    0    1 1.069434 0.3799064    0    1 1.0885594
#> [4,] 1.2048936 -1.182457    0    1 1.037757 0.3979255    0    1 0.9920802
#> [5,] 1.0694307 -1.118175    0    1 1.037058 0.3434790    0    1 1.0698760
#> [6,] 0.9530908 -1.102341    0    1 1.181057 0.4442564    0    1 1.2534737
#>         [,10] [,11] [,12]     [,13]     [,14] [,15] [,16]     [,17]     [,18]
#> [1,] 1.293120     0     1 1.1244438 -2.132220     0     1 0.8305799 0.4797304
#> [2,] 1.141553     0     1 1.0044415 -2.216705     0     1 1.0487837 0.5045070
#> [3,] 1.057657     0     1 0.9973533 -2.065743     0     1 0.9218030 0.4400104
#> [4,] 1.097563     0     1 1.2481544 -2.300055     0     1 0.8604047 0.4677333
#> [5,] 1.257874     0     1 1.1101692 -2.312099     0     1 0.7024525 0.3758068
#> [6,] 1.277132     0     1 1.2961896 -2.058542     0     1 0.8608754 0.5196577
#>      [,19] [,20]    [,21]     [,22] [,23] [,24]     [,25]      [,26] [,27]
#> [1,]     0     1 1.048600 0.5988930     0     1 1.4300115 -0.7023182     0
#> [2,]     0     1 1.101551 0.4871206     0     1 1.0750734 -0.4635880     0
#> [3,]     0     1 1.309718 0.4591179     0     1 1.2205973 -0.3580720     0
#> [4,]     0     1 1.206620 0.5392703     0     1 1.0807866 -0.5841053     0
#> [5,]     0     1 1.148354 0.5250229     0     1 0.8442819 -0.4644225     0
#> [6,]     0     1 1.074708 0.4160709     0     1 1.3004257 -0.3850059     0
#>      [,28]    [,29]      [,30] [,31] [,32]     [,33]      [,34] [,35] [,36]
#> [1,]     1 1.404012 -0.6185119     0     1 0.9490040 -0.4020894     0     1
#> [2,]     1 1.139388 -0.5199345     0     1 0.9521555 -0.6784563     0     1
#> [3,]     1 1.108187 -0.6036393     0     1 0.7753608 -0.5774887     0     1
#> [4,]     1 1.284275 -0.7277407     0     1 0.7482995 -0.3126502     0     1
#> [5,]     1 1.252585 -0.6637791     0     1 1.0148150 -0.4759510     0     1
#> [6,]     1 1.056796 -0.4669908     0     1 1.0752973 -0.5471399     0     1
#>          [,37]      [,38] [,39] [,40]     [,41]      [,42] [,43] [,44]
#> [1,] 1.2055754 -0.8512938     0     1 1.1508871 -0.1419016     0     1
#> [2,] 1.1672176 -1.0454200     0     1 1.2410373 -0.3856664     0     1
#> [3,] 0.9626271 -0.8724353     0     1 1.1468847 -0.2796489     0     1
#> [4,] 1.0410570 -0.9413548     0     1 1.1496565 -0.3770805     0     1
#> [5,] 1.0545005 -0.9814805     0     1 0.9714512 -0.5829424     0     1
#> [6,] 1.0888333 -0.8761581     0     1 0.8716506 -0.3504870     0     1
#>          [,45]      [,46] [,47] [,48]    [,49]      [,50] [,51] [,52]     [,53]
#> [1,] 1.2923583 -0.9998531     0     1 1.320256 -0.8944858     0     1 1.2219670
#> [2,] 0.9802955 -1.0359681     0     1 1.306339 -0.9610244     0     1 0.9268949
#> [3,] 1.1223813 -0.7922604     0     1 1.075357 -0.9222204     0     1 0.9725781
#> [4,] 0.8735673 -0.8156587     0     1 1.133964 -0.9482200     0     1 0.8512548
#> [5,] 1.0512766 -0.9890253     0     1 1.220752 -0.7655182     0     1 0.6345973
#> [6,] 0.9520499 -0.8391847     0     1 1.258394 -0.7322055     0     1 1.0116317
#>           [,54] [,55] [,56]    [,57]    [,58] [,59] [,60]     [,61]       [,62]
#> [1,] 0.12657328     0     1 1.490720 1.105207     0     1 1.0308247  0.03325259
#> [2,] 0.01463468     0     1 1.315260 1.102609     0     1 1.0924898 -0.19475849
#> [3,] 0.16705437     0     1 1.461502 1.229672     0     1 1.1170185 -0.14611165
#> [4,] 0.10902831     0     1 1.328143 1.207589     0     1 0.9973009 -0.17258154
#> [5,] 0.18144665     0     1 1.182052 1.127486     0     1 0.9167619 -0.12238736
#> [6,] 0.21350013     0     1 1.166695 1.072812     0     1 0.8786214 -0.18042354
#>      [,63] [,64]     [,65]      [,66] [,67] [,68]     [,69]      [,70] [,71]
#> [1,]     0     1 0.9038715 -0.2602564     0     1 1.0683772 -0.7204854     0
#> [2,]     0     1 0.9351548 -0.5092514     0     1 1.0905007 -1.0631623     0
#> [3,]     0     1 1.2431841 -0.3172786     0     1 1.3224243 -1.0061128     0
#> [4,]     0     1 1.0973202 -0.4747361     0     1 0.8725659 -1.0184864     0
#> [5,]     0     1 1.0851715 -0.4373217     0     1 1.3505017 -1.0147292     0
#> [6,]     0     1 1.0293762 -0.2263010     0     1 1.3174595 -0.9554100     0
#>      [,72]     [,73]      [,74] [,75] [,76]    [,77]    [,78] [,79] [,80]
#> [1,]     1 0.7957660 -0.3977214     0     1 1.281399 2.644462     0     1
#> [2,]     1 1.0783181 -0.5448083     0     1 1.513309 2.754327     0     1
#> [3,]     1 1.1177004 -0.6616375     0     1 1.878915 2.845823     0     1
#> [4,]     1 1.1214052 -0.6979218     0     1 1.503532 2.914848     0     1
#> [5,]     1 0.7529081 -0.4411806     0     1 1.327589 2.785122     0     1
#> [6,]     1 1.0517254 -0.5178313     0     1 1.666037 3.054605     0     1
#>          [,81]        [,82] [,83] [,84]    [,85]      [,86] [,87] [,88]
#> [1,] 1.0414954  0.274282245     0     1 1.142956 -0.3758635     0     1
#> [2,] 1.0916483  0.106565673     0     1 1.017860 -0.4444722     0     1
#> [3,] 0.9801698  0.007121071     0     1 1.186623 -0.5527741     0     1
#> [4,] 0.9966010 -0.048008380     0     1 1.009427 -0.5731822     0     1
#> [5,] 0.7730778  0.162600257     0     1 1.162453 -0.5023252     0     1
#> [6,] 1.2069635  0.112967427     0     1 1.194206 -0.6327296     0     1
#>          [,89]      [,90] [,91] [,92]     [,93]     [,94] [,95] [,96]     [,97]
#> [1,] 0.7858014 -0.2560689     0     1 1.0773261 0.4218586     0     1 1.0902891
#> [2,] 1.0158681 -0.5223369     0     1 1.1446828 0.4550389     0     1 1.0046447
#> [3,] 0.7773660 -0.4173215     0     1 0.7006403 0.6246165     0     1 0.9905861
#> [4,] 0.9671124 -0.2472613     0     1 1.0088338 0.6031528     0     1 1.1443702
#> [5,] 0.6290454 -0.4324590     0     1 0.8217616 0.4917042     0     1 1.2035087
#> [6,] 0.7689329 -0.3837231     0     1 1.4476114 0.3970508     0     1 1.2307077
#>           [,98] [,99] [,100]   [,101]    [,102] [,103] [,104]    [,105]
#> [1,] -0.5864734     0      1 1.185578 -1.227117      0      1 1.3435690
#> [2,] -0.7747869     0      1 1.350796 -1.564017      0      1 1.0749936
#> [3,] -0.5597943     0      1 1.367103 -1.503782      0      1 0.9750391
#> [4,] -0.5435829     0      1 1.064384 -1.326332      0      1 0.8610794
#> [5,] -0.7075321     0      1 1.237309 -1.462118      0      1 1.1064333
#> [6,] -0.7754601     0      1 1.484138 -1.524509      0      1 1.2835361
#>         [,106] [,107] [,108]    [,109]     [,110] [,111] [,112]   [,113]
#> [1,] 0.9929827      0      1 1.0788786 -0.8061388      0      1 1.258350
#> [2,] 0.7038529      0      1 1.1338104 -1.2382333      0      1 1.050970
#> [3,] 0.5767637      0      1 1.0166901 -0.9163397      0      1 1.334494
#> [4,] 0.6363576      0      1 0.9633732 -1.0873617      0      1 1.048905
#> [5,] 0.6501523      0      1 0.9735527 -0.8551775      0      1 1.119480
#> [6,] 1.1062122      0      1 1.2449820 -0.7858268      0      1 1.031900
#>            [,114] [,115] [,116]    [,117]     [,118] [,119] [,120]    [,121]
#> [1,]  0.170505847      0      1 1.0878106 -0.6875603      0      1 0.9973361
#> [2,] -0.203666719      0      1 1.1506334 -0.8979071      0      1 0.9435221
#> [3,] -0.157403651      0      1 1.0165844 -0.6326679      0      1 1.0579007
#> [4,]  0.002851266      0      1 0.8243482 -0.9153533      0      1 1.0558843
#> [5,] -0.154737093      0      1 1.2742153 -0.9941709      0      1 0.8948933
#> [6,] -0.039741655      0      1 0.8740436 -0.9878720      0      1 1.0238616
#>        [,122] [,123] [,124]    [,125]       [,126] [,127] [,128]    [,129]
#> [1,] 1.199695      0      1 1.0296248 -0.254425015      0      1 1.2696636
#> [2,] 1.219630      0      1 1.0729861 -0.429193458      0      1 1.1709322
#> [3,] 1.247084      0      1 0.9351892 -0.009590747      0      1 1.0458483
#> [4,] 1.419982      0      1 0.6490558 -0.479184393      0      1 1.2864697
#> [5,] 1.299701      0      1 0.8293272 -0.101447547      0      1 0.8395305
#> [6,] 1.331767      0      1 0.7444838 -0.265627420      0      1 1.1938280
#>          [,130] [,131] [,132]    [,133]     [,134] [,135] [,136]    [,137]
#> [1,] -0.5897368      0      1 1.0301515 -0.4870232      0      1 0.9042345
#> [2,] -0.7558106      0      1 0.9761142 -0.5800779      0      1 0.9519418
#> [3,] -0.6697879      0      1 1.0101125 -0.6207978      0      1 1.1062432
#> [4,] -0.6419688      0      1 0.8414136 -0.5811859      0      1 1.0145135
#> [5,] -0.5320286      0      1 0.7050669 -0.6880279      0      1 1.1347366
#> [6,] -0.6819696      0      1 1.2168491 -0.4920477      0      1 0.9474490
#>         [,138] [,139] [,140]    [,141]     [,142] [,143] [,144]    [,145]
#> [1,] -1.618417      0      1 0.6514902 -0.7760469      0      1 1.0127751
#> [2,] -1.641739      0      1 0.6988641 -0.8745586      0      1 1.0905396
#> [3,] -1.387974      0      1 0.8616208 -1.0540627      0      1 1.0106507
#> [4,] -1.723162      0      1 0.8537917 -1.0436089      0      1 0.9703324
#> [5,] -1.560989      0      1 0.8946300 -0.8468615      0      1 0.7133619
#> [6,] -1.558725      0      1 0.8941632 -1.1476553      0      1 1.1043900
#>         [,146] [,147] [,148]    [,149]    [,150] [,151] [,152]    [,153]
#> [1,] -2.174250      0      1 0.9108931 -1.190304      0      1 1.1222249
#> [2,] -2.689214      0      1 0.7579149 -1.038343      0      1 0.9439938
#> [3,] -2.084023      0      1 1.2226719 -1.376245      0      1 1.1183569
#> [4,] -1.952863      0      1 1.0898104 -1.551210      0      1 0.8541287
#> [5,] -2.240421      0      1 1.3794366 -1.509661      0      1 0.9375943
#> [6,] -2.188792      0      1 1.4050985 -1.399785      0      1 0.8760451
#>          [,154] [,155] [,156]    [,157]     [,158] [,159] [,160] [,161] [,162]
#> [1,] -0.1581998      0      1 1.1024153 -0.3447116      0      1      0      1
#> [2,] -0.3465350      0      1 1.0456951 -0.5979756      0      1      0      1
#> [3,] -0.1622324      0      1 1.1514504 -0.3475905      0      1      0      1
#> [4,] -0.1688018      0      1 0.9571307 -0.3239786      0      1      0      1
#> [5,] -0.1596477      0      1 0.8469288 -0.4320925      0      1      0      1
#> [6,] -0.1841768      0      1 0.9512729 -0.4609813      0      1      0      1
#>         [,163]    [,164] [,165] [,166]   [,167]    [,168] [,169] [,170]
#> [1,] 0.9508082 -1.090240      0      1 1.131674 0.5995145      0      1
#> [2,] 1.1062395 -1.381437      0      1 1.210205 0.3462211      0      1
#> [3,] 1.1103972 -1.287658      0      1 1.069434 0.3799064      0      1
#> [4,] 1.2048936 -1.182457      0      1 1.037757 0.3979255      0      1
#> [5,] 1.0694307 -1.118175      0      1 1.037058 0.3434790      0      1
#> [6,] 0.9530908 -1.102341      0      1 1.181057 0.4442564      0      1
#>         [,171]   [,172] [,173] [,174]    [,175]    [,176] [,177] [,178]
#> [1,] 1.4946106 1.293120      0      1 1.1244438 -2.132220      0      1
#> [2,] 1.2580931 1.141553      0      1 1.0044415 -2.216705      0      1
#> [3,] 1.0885594 1.057657      0      1 0.9973533 -2.065743      0      1
#> [4,] 0.9920802 1.097563      0      1 1.2481544 -2.300055      0      1
#> [5,] 1.0698760 1.257874      0      1 1.1101692 -2.312099      0      1
#> [6,] 1.2534737 1.277132      0      1 1.2961896 -2.058542      0      1
#>         [,179]    [,180] [,181] [,182]    [,183]    [,184] [,185] [,186]
#> [1,] 0.8305799 0.4797304      0      1 0.9013272 0.5683280      0      1
#> [2,] 1.0487837 0.5045070      0      1 0.8144625 0.2624203      0      1
#> [3,] 0.9218030 0.4400104      0      1 0.9214203 0.3744210      0      1
#> [4,] 0.8604047 0.4677333      0      1 0.8041635 0.3776195      0      1
#> [5,] 0.7024525 0.3758068      0      1 0.9500476 0.5424239      0      1
#> [6,] 0.8608754 0.5196577      0      1 0.9041244 0.5478380      0      1
#>         [,187]     [,188] [,189] [,190]    [,191]     [,192] [,193] [,194]
#> [1,] 0.9666922 -0.2655062      0      1 1.0130509 -0.3232058      0      1
#> [2,] 0.9463856 -0.4571142      0      1 0.8369106 -0.2691143      0      1
#> [3,] 0.8657511 -0.5476008      0      1 0.9019815 -0.4428766      0      1
#> [4,] 1.2148126 -0.3710019      0      1 0.9658424 -0.5175862      0      1
#> [5,] 0.9962038 -0.3201038      0      1 1.1932079 -0.4424748      0      1
#> [6,] 1.0830407 -0.4054991      0      1 0.9486247 -0.5486140      0      1
#>         [,195]     [,196] [,197] [,198]    [,199]     [,200] [,201] [,202]
#> [1,] 0.8420296 -0.4483782      0      1 1.0290149 -0.6065393      0      1
#> [2,] 0.8499383 -0.5087451      0      1 1.0550301 -0.8818950      0      1
#> [3,] 0.9219106 -0.6334462      0      1 1.0480751 -0.7220186      0      1
#> [4,] 0.9801788 -0.6329010      0      1 0.9383057 -0.8523720      0      1
#> [5,] 1.0271645 -0.5102411      0      1 1.0630001 -0.5765267      0      1
#> [6,] 0.8208585 -0.3491054      0      1 1.0806057 -0.5442712      0      1
#>        [,203]     [,204] [,205] [,206]    [,207]     [,208] [,209] [,210]
#> [1,] 1.031434 -0.6324790      0      1 1.0457666 -0.7084809      0      1
#> [2,] 1.281079 -0.6415990      0      1 1.3233145 -0.9462706      0      1
#> [3,] 1.280514 -0.5879088      0      1 1.4753070 -0.9817803      0      1
#> [4,] 1.129901 -0.5392947      0      1 1.2859273 -0.8973160      0      1
#> [5,] 1.101856 -0.4669393      0      1 0.8556751 -0.7729075      0      1
#> [6,] 1.028506 -0.4677455      0      1 1.1244478 -0.8492876      0      1
#>         [,211]     [,212] [,213] [,214]    [,215]    [,216] [,217] [,218]
#> [1,] 0.9562316 -0.7765237      0      1 0.9985169 0.3495552      0      1
#> [2,] 0.9406724 -0.7176055      0      1 1.1504400 0.1439441      0      1
#> [3,] 1.0627731 -0.5798559      0      1 0.9714495 0.2097078      0      1
#> [4,] 1.0373478 -0.8557307      0      1 0.9668935 0.2472405      0      1
#> [5,] 0.8348874 -0.6430370      0      1 0.8185530 0.1383656      0      1
#> [6,] 1.0065657 -0.7423843      0      1 1.0161722 0.2584311      0      1
#>         [,219]    [,220] [,221] [,222]    [,223]      [,224] [,225] [,226]
#> [1,] 1.0828015 1.4622017      0      1 0.9547353  0.04459326      0      1
#> [2,] 0.9671356 0.9261104      0      1 1.1156551 -0.25882563      0      1
#> [3,] 1.1701292 1.1060436      0      1 1.0339354 -0.05280801      0      1
#> [4,] 1.3060864 1.2389561      0      1 1.0054276 -0.09243620      0      1
#> [5,] 1.0486564 1.1198365      0      1 0.9952094 -0.13967530      0      1
#> [6,] 1.3953387 1.3999084      0      1 1.0261271 -0.11916432      0      1
#>         [,227]     [,228] [,229] [,230]    [,231]     [,232] [,233] [,234]
#> [1,] 0.9878409 -0.1895908      0      1 0.6441882 -0.5571625      0      1
#> [2,] 1.2261610 -0.2972543      0      1 0.8289070 -0.7466779      0      1
#> [3,] 1.2488866 -0.3531854      0      1 0.8166856 -0.8623090      0      1
#> [4,] 1.4632307 -0.3088658      0      1 0.6935948 -0.8012698      0      1
#> [5,] 0.9255100 -0.2853635      0      1 0.9884424 -0.8228380      0      1
#> [6,] 0.9472783 -0.1342560      0      1 0.6279410 -0.7873138      0      1
#>         [,235]     [,236] [,237] [,238]    [,239]   [,240] [,241] [,242]
#> [1,] 1.0569659 -0.6364697      0      1 1.1326394 2.918423      0      1
#> [2,] 1.0204416 -0.6482608      0      1 1.0313263 2.445498      0      1
#> [3,] 0.7030436 -0.6477364      0      1 1.3710465 2.484449      0      1
#> [4,] 0.7512808 -0.6552009      0      1 1.1249090 2.637560      0      1
#> [5,] 0.9365033 -0.4883961      0      1 0.9182891 2.365203      0      1
#> [6,] 1.0992436 -0.5724832      0      1 1.0952615 2.632716      0      1
#>         [,243]     [,244] [,245] [,246]   [,247]       [,248] [,249] [,250]
#> [1,] 0.9476881 0.15660197      0      1 1.405174 -0.001951073      0      1
#> [2,] 0.8464164 0.11486558      0      1 1.319665 -0.458588795      0      1
#> [3,] 0.9464964 0.07573852      0      1 1.179516 -0.101513698      0      1
#> [4,] 1.0251921 0.25574479      0      1 1.197509 -0.168640277      0      1
#> [5,] 0.9782002 0.29106769      0      1 1.140322 -0.043980732      0      1
#> [6,] 0.9512229 0.27109418      0      1 1.560881 -0.171698588      0      1
#>         [,251]     [,252] [,253] [,254]   [,255]    [,256] [,257] [,258]
#> [1,] 0.8355067 -0.2748009      0      1 1.030028 0.9774675      0      1
#> [2,] 1.0683567 -0.4313862      0      1 1.234744 0.5876636      0      1
#> [3,] 1.0723264 -0.3895034      0      1 1.200005 1.0317878      0      1
#> [4,] 1.0724920 -0.4825183      0      1 1.314280 0.7996089      0      1
#> [5,] 0.9818229 -0.1749088      0      1 1.120307 0.8345080      0      1
#> [6,] 1.0382559 -0.3845457      0      1 1.316979 0.9654542      0      1
#>         [,259]     [,260] [,261] [,262]    [,263]    [,264] [,265] [,266]
#> [1,] 0.6890765 -0.2065104      0      1 1.0246358 -1.210769      0      1
#> [2,] 1.2797117 -0.5276152      0      1 0.7767512 -1.124416      0      1
#> [3,] 1.0108947 -0.4118439      0      1 1.0065608 -1.381070      0      1
#> [4,] 1.1085288 -0.4351863      0      1 0.9773176 -1.436431      0      1
#> [5,] 1.0026344 -0.4275693      0      1 1.0698491 -1.329545      0      1
#> [6,] 0.8566996 -0.5074841      0      1 0.8202270 -1.207524      0      1
#>         [,267]    [,268] [,269] [,270]    [,271]     [,272] [,273] [,274]
#> [1,] 1.2907082 0.9085525      0      1 0.8116268 -0.7206798      0      1
#> [2,] 1.1468531 0.5757390      0      1 0.9811042 -0.8279592      0      1
#> [3,] 1.2455196 0.8599447      0      1 1.1592145 -0.9234199      0      1
#> [4,] 1.1629557 0.6807862      0      1 1.0000083 -0.8068264      0      1
#> [5,] 0.9972136 0.7112144      0      1 1.2557261 -0.9080728      0      1
#> [6,] 1.1767704 0.8386689      0      1 1.0058563 -0.7708327      0      1
#>        [,275]    [,276] [,277] [,278]    [,279]     [,280] [,281] [,282]
#> [1,] 1.051278 0.4383378      0      1 0.8611985 -0.6719143      0      1
#> [2,] 0.940244 0.1401959      0      1 1.0086634 -0.9466725      0      1
#> [3,] 1.091937 0.1038293      0      1 1.0642135 -0.8734024      0      1
#> [4,] 1.190358 0.1225179      0      1 0.9497107 -0.8952337      0      1
#> [5,] 1.239023 0.1388828      0      1 0.9224000 -0.8770794      0      1
#> [6,] 1.047239 0.2795890      0      1 0.9907808 -0.7047977      0      1
#>         [,283]   [,284] [,285] [,286]    [,287]     [,288] [,289] [,290]
#> [1,] 0.8401128 1.141069      0      1 0.9251503 -0.3962175      0      1
#> [2,] 0.8709718 1.086084      0      1 1.4479548 -0.5132071      0      1
#> [3,] 1.1038022 1.263652      0      1 1.2543885 -0.3480903      0      1
#> [4,] 1.0996983 1.278973      0      1 1.0246877 -0.4193097      0      1
#> [5,] 1.1218696 1.228908      0      1 1.2164529 -0.2475104      0      1
#> [6,] 1.2476651 1.377301      0      1 1.4011289 -0.3169907      0      1
#>         [,291]     [,292] [,293] [,294]   [,295]     [,296] [,297] [,298]
#> [1,] 0.7929059 -0.5693313      0      1 1.005761 -0.1053292      0      1
#> [2,] 0.7835427 -0.6736538      0      1 1.102319 -0.3603848      0      1
#> [3,] 1.0014968 -0.6245642      0      1 1.054083 -0.5073574      0      1
#> [4,] 0.9326631 -0.5229844      0      1 1.011851 -0.1159149      0      1
#> [5,] 1.0264894 -0.5103861      0      1 0.881811 -0.1900152      0      1
#> [6,] 0.9496131 -0.6102716      0      1 1.086935 -0.3581252      0      1
#>         [,299]    [,300] [,301] [,302]    [,303]     [,304] [,305] [,306]
#> [1,] 1.1363058 -1.570124      0      1 0.9926068 -1.1639815      0      1
#> [2,] 1.2050101 -1.898106      0      1 0.8264989 -1.2051940      0      1
#> [3,] 1.2703297 -1.676416      0      1 1.1532300 -1.1694239      0      1
#> [4,] 1.1007493 -1.519405      0      1 0.8607271 -1.0156976      0      1
#> [5,] 0.9502208 -1.754312      0      1 0.9040017 -1.3028490      0      1
#> [6,] 1.0372421 -1.441296      0      1 0.8948591 -0.9882351      0      1
#>         [,307]    [,308] [,309] [,310]    [,311]     [,312] [,313] [,314]
#> [1,] 0.6759833 -1.730643      0      1 0.8245516 -0.8993679      0      1
#> [2,] 1.2759251 -2.452676      0      1 1.0545983 -1.2412230      0      1
#> [3,] 1.1873273 -2.144179      0      1 1.2329414 -1.0836517      0      1
#> [4,] 0.8036841 -2.220864      0      1 0.9058524 -1.1401327      0      1
#> [5,] 1.1496303 -2.164663      0      1 0.9772203 -0.9998511      0      1
#> [6,] 1.0606886 -1.968724      0      1 0.8837085 -0.8068005      0      1
#>         [,315]       [,316] [,317] [,318]    [,319]     [,320] [,321] [,322]
#> [1,] 1.1098827 -0.006424539      0      1 1.0649269 -0.3596501      0      1
#> [2,] 0.9287809 -0.272656076      0      1 0.9492123 -0.4912140      0      1
#> [3,] 1.2039610 -0.085502081      0      1 1.0937925 -0.3499363      0      1
#> [4,] 0.9663910 -0.080146508      0      1 0.7806662 -0.3570604      0      1
#> [5,] 1.0157667 -0.096140175      0      1 1.0117457 -0.1738891      0      1
#> [6,] 1.2944196 -0.027248379      0      1 0.8357723 -0.3652937      0      1
#>           [,323]    [,324]
#> [1,] -0.34398260 1.0489868
#> [2,] -0.05889772 0.9265268
#> [3,] -0.03114239 0.8609106
#> [4,] -0.14550858 0.8264699
#> [5,] -0.16478766 0.9903766
#> [6,] -0.26087629 0.7742060

# }