draw_parameters {mirt}	R Documentation

Draw plausible parameter instantiations from a given model

Description

Draws plausible parameters from a model using parametric sampling (if the information matrix was computed) or via bootstrap sampling. Primarily for use with the DRF function.

Usage

draw_parameters(
  mod,
  draws,
  method = c("parametric", "boostrap"),
  redraws = 20,
  verbose = FALSE,
  ...
)

Arguments

mod
estimated single or multiple-group model

draws
number of draws to obtain

method
type of plausible values to obtain. Can be 'parametric', for the parametric sampling scheme which uses the estimated information matrix, or 'boostrap' to obtain values from the boot function. Default is 'parametric'

redraws
number of redraws to perform when the given parameteric sample does not satisfy the upper and lower parameter bounds. If a valid set cannot be found within this number of draws then an error will be thrown

verbose
logical; include additional information in the console?

...
additional arguments to be passed

Value

returns a draws x p matrix of plausible parameters, where each row correspeonds to a single set

Examples

## No test: 
set.seed(1234)
n <- 40
N <- 500

# only first 5 items as anchors
model <- 'F = 1-40
          CONSTRAINB = (1-5, a1), (1-5, d)'

a <- matrix(1, n)
d <- matrix(rnorm(n), n)
group <- c(rep('Group_1', N), rep('Group_2', N))

## -------------
# groups completely equal
dat1 <- simdata(a, d, N, itemtype = 'dich')
dat2 <- simdata(a, d, N, itemtype = 'dich')
dat <- rbind(dat1, dat2)
mod <- multipleGroup(dat, model, group=group, SE=TRUE,
                     invariance=c('free_means', 'free_var'))

param_set <- draw_parameters(mod, 100)
head(param_set)

##           [,1]      [,2] [,3] [,4]      [,5]      [,6] [,7] [,8]      [,9]
## [1,] 1.0932743 -1.068591    0    1 1.0304037 0.4502467    0    1 1.0493886
## [2,] 1.1054115 -1.160225    0    1 1.0101008 0.4327771    0    1 0.9969425
## [3,] 1.0588536 -1.043460    0    1 0.9338291 0.5507912    0    1 1.1726077
## [4,] 1.0641110 -1.267074    0    1 1.1917589 0.5088504    0    1 1.1761461
## [5,] 1.2415625 -1.347729    0    1 0.9419911 0.3352369    0    1 1.0874302
## [6,] 0.9296441 -1.156598    0    1 1.1481915 0.3428099    0    1 1.2360327
##         [,10] [,11] [,12]     [,13]     [,14] [,15] [,16]     [,17]     [,18]
## [1,] 1.154459     0     1 1.3565442 -2.319530     0     1 0.9527591 0.5626412
## [2,] 1.164268     0     1 1.0189394 -2.139889     0     1 0.7642930 0.5358643
## [3,] 1.272867     0     1 1.2890713 -2.144469     0     1 0.8298928 0.6190689
## [4,] 1.188510     0     1 1.0034363 -2.264064     0     1 0.9192414 0.5393377
## [5,] 1.040352     0     1 0.9625683 -2.227885     0     1 0.9023737 0.5070200
## [6,] 1.299036     0     1 1.2193637 -2.283865     0     1 0.7739792 0.3709170
##      [,19] [,20]    [,21]     [,22] [,23] [,24]     [,25]      [,26] [,27]
## [1,]     0     1 1.072320 0.4995215     0     1 1.0932079 -0.4469251     0
## [2,]     0     1 1.248633 0.4804203     0     1 0.9357223 -0.6628563     0
## [3,]     0     1 1.326394 0.4463118     0     1 1.0975333 -0.4262287     0
## [4,]     0     1 1.350436 0.2500279     0     1 1.2905898 -0.4520652     0
## [5,]     0     1 1.444609 0.5100192     0     1 1.0291162 -0.5460004     0
## [6,]     0     1 0.956844 0.1510659     0     1 1.1902397 -0.5464074     0
##      [,28]     [,29]      [,30] [,31] [,32]     [,33]      [,34] [,35] [,36]
## [1,]     1 0.8829600 -0.4263604     0     1 0.8922963 -0.2598440     0     1
## [2,]     1 1.0030996 -0.7802953     0     1 0.7341051 -0.5136644     0     1
## [3,]     1 0.9248617 -0.5816143     0     1 0.6037833 -0.3880138     0     1
## [4,]     1 1.3001149 -0.4701872     0     1 0.7644687 -0.3982218     0     1
## [5,]     1 1.3070933 -0.5836171     0     1 0.7874720 -0.4615814     0     1
## [6,]     1 0.9381240 -0.6536426     0     1 0.6134178 -0.4913699     0     1
##          [,37]      [,38] [,39] [,40]     [,41]      [,42] [,43] [,44]    [,45]
## [1,] 0.8429839 -0.9134404     0     1 1.1763025 -0.3159328     0     1 1.056970
## [2,] 0.8715442 -0.7559496     0     1 1.3142568 -0.4368640     0     1 1.129900
## [3,] 1.0102379 -0.8705496     0     1 0.9724623 -0.2052890     0     1 1.120864
## [4,] 1.0163318 -0.9024250     0     1 1.1937872 -0.4859021     0     1 1.077213
## [5,] 1.0219814 -1.0199597     0     1 1.2593303 -0.4458798     0     1 1.101842
## [6,] 0.9313696 -0.9216369     0     1 1.2909519 -0.4340186     0     1 1.021393
##           [,46] [,47] [,48]    [,49]      [,50] [,51] [,52]     [,53]
## [1,] -0.8829651     0     1 1.009843 -0.9162045     0     1 0.8509561
## [2,] -1.0435554     0     1 1.348743 -1.0346809     0     1 1.0337138
## [3,] -0.9352326     0     1 1.074015 -0.7817479     0     1 1.0465153
## [4,] -1.0285902     0     1 1.144905 -0.8775969     0     1 0.6738345
## [5,] -1.0962856     0     1 1.021204 -0.8736799     0     1 0.8979971
## [6,] -1.0199801     0     1 1.032786 -1.1194219     0     1 0.8641428
##            [,54] [,55] [,56]     [,57]     [,58] [,59] [,60]     [,61]
## [1,]  0.28520044     0     1 1.3960836 1.2745237     0     1 0.9049768
## [2,]  0.14875216     0     1 0.9584291 0.8691768     0     1 0.8528032
## [3,]  0.32571452     0     1 1.4840633 1.1483469     0     1 1.0617626
## [4,]  0.05481774     0     1 1.3234433 1.0170065     0     1 0.9622920
## [5,]  0.13632445     0     1 1.4834053 1.1759660     0     1 0.8809452
## [6,] -0.19261769     0     1 0.9749428 0.8612682     0     1 0.7263630
##            [,62] [,63] [,64]     [,65]      [,66] [,67] [,68]     [,69]
## [1,]  0.23960582     0     1 1.1147295 -0.3944717     0     1 0.9280099
## [2,] -0.14709870     0     1 0.8732223 -0.3952885     0     1 0.9353382
## [3,]  0.05050045     0     1 1.3153315 -0.5048765     0     1 1.2577706
## [4,] -0.10373475     0     1 1.0584595 -0.5585976     0     1 1.2583345
## [5,] -0.21265216     0     1 1.0002266 -0.5300889     0     1 1.1044549
## [6,] -0.18931782     0     1 0.9472484 -0.4163406     0     1 1.0140788
##           [,70] [,71] [,72]     [,73]      [,74] [,75] [,76]    [,77]    [,78]
## [1,] -0.9186439     0     1 0.8843822 -0.6292826     0     1 1.201452 2.689813
## [2,] -0.9015860     0     1 0.9377831 -0.5722138     0     1 1.201948 2.870918
## [3,] -0.9026512     0     1 1.0352960 -0.8258879     0     1 1.157653 2.404312
## [4,] -0.7682697     0     1 0.8894741 -0.6065435     0     1 1.564675 3.165334
## [5,] -0.9848604     0     1 0.9941136 -0.6344542     0     1 1.521427 2.707065
## [6,] -0.9584903     0     1 0.9469128 -0.8272694     0     1 1.379037 2.923075
##      [,79] [,80]     [,81]       [,82] [,83] [,84]    [,85]      [,86] [,87]
## [1,]     0     1 0.8266239  0.23942707     0     1 0.935058 -0.6100174     0
## [2,]     0     1 0.8547433  0.04245084     0     1 1.186181 -0.7336875     0
## [3,]     0     1 0.9065702  0.09851040     0     1 1.226667 -0.5264532     0
## [4,]     0     1 0.8637495  0.06628946     0     1 1.169432 -0.6563654     0
## [5,]     0     1 1.0705687 -0.18440171     0     1 1.328909 -0.5819601     0
## [6,]     0     1 1.1421833  0.02158851     0     1 1.212900 -0.6098381     0
##      [,88]     [,89]      [,90] [,91] [,92]     [,93]     [,94] [,95] [,96]
## [1,]     1 0.9308422 -0.3048860     0     1 1.1613751 0.7601815     0     1
## [2,]     1 0.8148861 -0.5811512     0     1 1.0201076 0.3791420     0     1
## [3,]     1 0.9837659 -0.2858480     0     1 1.1527111 0.5390425     0     1
## [4,]     1 0.9114794 -0.1672388     0     1 0.8148990 0.4776476     0     1
## [5,]     1 0.9949380 -0.2842741     0     1 1.0815155 0.3933887     0     1
## [6,]     1 0.6864495 -0.5532132     0     1 0.9284866 0.4878206     0     1
##          [,97]      [,98] [,99] [,100]   [,101]    [,102] [,103] [,104]
## [1,] 1.1009475 -0.3954280     0      1 1.519014 -1.445092      0      1
## [2,] 1.1334879 -0.5945430     0      1 1.331797 -1.571551      0      1
## [3,] 1.1310255 -0.4695673     0      1 1.623967 -1.440331      0      1
## [4,] 0.9709377 -0.5298882     0      1 1.253495 -1.506501      0      1
## [5,] 1.2333767 -0.5771723     0      1 1.665513 -1.650269      0      1
## [6,] 1.0679226 -0.6596770     0      1 1.464518 -1.574170      0      1
##        [,105]    [,106] [,107] [,108]    [,109]     [,110] [,111] [,112]
## [1,] 1.339280 0.9404397      0      1 1.2069977 -1.0166139      0      1
## [2,] 1.135063 0.8440658      0      1 1.0806136 -0.9638710      0      1
## [3,] 1.281008 0.8551403      0      1 1.0961011 -0.8558904      0      1
## [4,] 1.424057 0.6503226      0      1 1.0103731 -0.9793761      0      1
## [5,] 1.785416 0.6975291      0      1 1.0744588 -1.0697066      0      1
## [6,] 1.465013 0.4393383      0      1 0.9689445 -0.9601775      0      1
##         [,113]      [,114] [,115] [,116]    [,117]     [,118] [,119] [,120]
## [1,] 1.1625647 -0.03367786      0      1 0.9259424 -0.9350858      0      1
## [2,] 1.2874586 -0.05672349      0      1 0.7413955 -0.8069379      0      1
## [3,] 0.9606102 -0.03049809      0      1 0.8944237 -0.6369680      0      1
## [4,] 0.8726023 -0.06699188      0      1 0.9442333 -0.8848880      0      1
## [5,] 1.2591267 -0.28188740      0      1 1.1114134 -1.0469899      0      1
## [6,] 1.1443778 -0.11468200      0      1 0.9545939 -0.9492122      0      1
##         [,121]   [,122] [,123] [,124]    [,125]     [,126] [,127] [,128]
## [1,] 0.8883158 1.425635      0      1 1.0435934 -0.3317235      0      1
## [2,] 0.8614355 1.376445      0      1 0.7767570 -0.1606989      0      1
## [3,] 1.2769722 1.272186      0      1 0.7268816 -0.2477991      0      1
## [4,] 1.0490126 1.272040      0      1 0.8687849 -0.1414372      0      1
## [5,] 1.1223556 1.302536      0      1 0.9898519 -0.2702985      0      1
## [6,] 1.1467092 1.164615      0      1 0.8325984 -0.1967782      0      1
##         [,129]     [,130] [,131] [,132]    [,133]     [,134] [,135] [,136]
## [1,] 1.1431107 -0.5499875      0      1 0.6556906 -0.4840821      0      1
## [2,] 0.9589297 -0.7708177      0      1 0.8864223 -0.7618080      0      1
## [3,] 1.0464618 -0.5331965      0      1 0.8139787 -0.4494785      0      1
## [4,] 0.9531351 -0.6118886      0      1 0.9716278 -0.3986095      0      1
## [5,] 1.0035153 -0.6666991      0      1 0.8028098 -0.5377433      0      1
## [6,] 1.0437810 -0.6799317      0      1 1.1775783 -0.6415326      0      1
##         [,137]    [,138] [,139] [,140]    [,141]     [,142] [,143] [,144]
## [1,] 0.7610125 -1.435648      0      1 0.8676006 -1.0160550      0      1
## [2,] 1.3602392 -1.893464      0      1 0.7780700 -0.9542235      0      1
## [3,] 0.8649010 -1.499841      0      1 0.6555519 -0.8602109      0      1
## [4,] 1.1170583 -1.906619      0      1 0.7778158 -0.8488345      0      1
## [5,] 1.2506487 -1.547159      0      1 0.8809402 -0.9284446      0      1
## [6,] 0.9308590 -1.628085      0      1 0.6659999 -0.9640313      0      1
##         [,145]    [,146] [,147] [,148]    [,149]    [,150] [,151] [,152]
## [1,] 1.1296474 -2.631746      0      1 0.9000865 -1.190730      0      1
## [2,] 1.0474917 -2.410147      0      1 0.9945276 -1.592852      0      1
## [3,] 1.0084517 -2.239005      0      1 0.8192665 -1.134606      0      1
## [4,] 1.1201575 -2.241090      0      1 1.1315720 -1.523619      0      1
## [5,] 0.7311578 -2.084342      0      1 1.3030463 -1.780122      0      1
## [6,] 1.2266689 -2.616580      0      1 0.8653552 -1.480079      0      1
##         [,153]       [,154] [,155] [,156]    [,157]     [,158] [,159] [,160]
## [1,] 0.7532886 -0.002891249      0      1 1.1155239 -0.4418437      0      1
## [2,] 0.9539449 -0.230533530      0      1 0.9045480 -0.5495541      0      1
## [3,] 0.8872304 -0.066013035      0      1 1.0179579 -0.3207985      0      1
## [4,] 0.9423585 -0.243083417      0      1 0.9020893 -0.3614155      0      1
## [5,] 1.0920962 -0.101092772      0      1 0.8598871 -0.3301636      0      1
## [6,] 0.8109871 -0.216604364      0      1 0.9446274 -0.3801080      0      1
##      [,161] [,162]    [,163]    [,164] [,165] [,166]    [,167]    [,168] [,169]
## [1,]      0      1 1.0932743 -1.068591      0      1 1.0304037 0.4502467      0
## [2,]      0      1 1.1054115 -1.160225      0      1 1.0101008 0.4327771      0
## [3,]      0      1 1.0588536 -1.043460      0      1 0.9338291 0.5507912      0
## [4,]      0      1 1.0641110 -1.267074      0      1 1.1917589 0.5088504      0
## [5,]      0      1 1.2415625 -1.347729      0      1 0.9419911 0.3352369      0
## [6,]      0      1 0.9296441 -1.156598      0      1 1.1481915 0.3428099      0
##      [,170]    [,171]   [,172] [,173] [,174]    [,175]    [,176] [,177] [,178]
## [1,]      1 1.0493886 1.154459      0      1 1.3565442 -2.319530      0      1
## [2,]      1 0.9969425 1.164268      0      1 1.0189394 -2.139889      0      1
## [3,]      1 1.1726077 1.272867      0      1 1.2890713 -2.144469      0      1
## [4,]      1 1.1761461 1.188510      0      1 1.0034363 -2.264064      0      1
## [5,]      1 1.0874302 1.040352      0      1 0.9625683 -2.227885      0      1
## [6,]      1 1.2360327 1.299036      0      1 1.2193637 -2.283865      0      1
##         [,179]    [,180] [,181] [,182]    [,183]    [,184] [,185] [,186]
## [1,] 0.9527591 0.5626412      0      1 1.2024594 0.5422886      0      1
## [2,] 0.7642930 0.5358643      0      1 0.8156347 0.4825597      0      1
## [3,] 0.8298928 0.6190689      0      1 0.9153025 0.6947908      0      1
## [4,] 0.9192414 0.5393377      0      1 0.7984201 0.6321736      0      1
## [5,] 0.9023737 0.5070200      0      1 0.8354295 0.4091115      0      1
## [6,] 0.7739792 0.3709170      0      1 1.0937001 0.4247664      0      1
##         [,187]     [,188] [,189] [,190]    [,191]     [,192] [,193] [,194]
## [1,] 1.1340813 -0.4749462      0      1 1.1389186 -0.3571415      0      1
## [2,] 0.9844329 -0.3898828      0      1 0.7207867 -0.3600146      0      1
## [3,] 1.3987281 -0.4008643      0      1 1.0502738 -0.3025948      0      1
## [4,] 0.9334970 -0.4722491      0      1 0.6783510 -0.5178193      0      1
## [5,] 1.0979281 -0.8105320      0      1 0.9151681 -0.4235957      0      1
## [6,] 0.6524029 -0.6254205      0      1 0.8977967 -0.4359027      0      1
##         [,195]     [,196] [,197] [,198]    [,199]     [,200] [,201] [,202]
## [1,] 0.7496703 -0.4577155      0      1 1.2777396 -0.8828141      0      1
## [2,] 0.6528497 -0.4518765      0      1 0.7574333 -0.5594043      0      1
## [3,] 0.9311817 -0.3270386      0      1 1.1557938 -0.6736464      0      1
## [4,] 0.5302274 -0.4579942      0      1 0.8977829 -0.6556707      0      1
## [5,] 0.9511756 -0.6439639      0      1 0.9525434 -1.0408119      0      1
## [6,] 0.6920225 -0.6498309      0      1 1.1227448 -0.7524430      0      1
##         [,203]     [,204] [,205] [,206]    [,207]     [,208] [,209] [,210]
## [1,] 1.3224686 -0.4294405      0      1 1.1065756 -0.9673295      0      1
## [2,] 0.9102370 -0.3931931      0      1 0.8306301 -0.5328692      0      1
## [3,] 1.2223808 -0.4655124      0      1 1.1982666 -0.6601950      0      1
## [4,] 0.9289631 -0.7532265      0      1 1.1025128 -0.8572025      0      1
## [5,] 0.8076755 -0.5855607      0      1 1.1911164 -0.9893750      0      1
## [6,] 1.2323009 -0.6760859      0      1 1.1389330 -0.8667753      0      1
##         [,211]     [,212] [,213] [,214]    [,215]    [,216] [,217] [,218]
## [1,] 1.1927142 -0.8848540      0      1 1.1247468 0.1569238      0      1
## [2,] 1.1686887 -0.5083389      0      1 0.8028731 0.3343955      0      1
## [3,] 1.0427030 -0.4504786      0      1 1.1091656 0.3329796      0      1
## [4,] 1.0526449 -0.7742448      0      1 1.0311182 0.2750996      0      1
## [5,] 0.7923241 -0.7988097      0      1 0.8317697 0.1203479      0      1
## [6,] 1.0913429 -0.8996382      0      1 0.9551971 0.1039048      0      1
##         [,219]    [,220] [,221] [,222]    [,223]      [,224] [,225] [,226]
## [1,] 0.9562003 0.8614903      0      1 0.9347868 -0.19035596      0      1
## [2,] 0.8949071 1.1737094      0      1 0.8287547 -0.07982884      0      1
## [3,] 0.9824445 1.1691857      0      1 0.9987563 -0.03160118      0      1
## [4,] 1.0461366 1.1130158      0      1 0.9599269 -0.13636627      0      1
## [5,] 0.9480721 0.8924371      0      1 0.7387001 -0.40461407      0      1
## [6,] 1.1650517 1.0469950      0      1 0.8575241 -0.15158950      0      1
##         [,227]      [,228] [,229] [,230]    [,231]     [,232] [,233] [,234]
## [1,] 1.1923981 -0.13457801      0      1 1.1048783 -0.8580522      0      1
## [2,] 0.9259563 -0.23499546      0      1 0.6516027 -0.7789132      0      1
## [3,] 1.1888008  0.08483676      0      1 1.1344235 -0.6034025      0      1
## [4,] 0.8050468 -0.29486181      0      1 0.6841821 -0.8050509      0      1
## [5,] 1.0115971 -0.47606401      0      1 0.6414706 -0.9006422      0      1
## [6,] 1.0481937 -0.43783505      0      1 0.9198174 -0.8428986      0      1
##         [,235]     [,236] [,237] [,238]   [,239]   [,240] [,241] [,242]
## [1,] 0.9160278 -0.5752614      0      1 1.174354 2.666040      0      1
## [2,] 0.7063287 -0.5626455      0      1 1.177802 2.710018      0      1
## [3,] 0.9131294 -0.4514593      0      1 1.305846 2.902359      0      1
## [4,] 0.8166246 -0.5851632      0      1 0.908682 2.668180      0      1
## [5,] 0.8798593 -0.8775316      0      1 0.779324 2.165681      0      1
## [6,] 0.9360799 -0.7188424      0      1 1.041410 2.491484      0      1
##         [,243]      [,244] [,245] [,246]   [,247]      [,248] [,249] [,250]
## [1,] 1.0583665  0.27108054      0      1 1.637879 -0.22115676      0      1
## [2,] 0.7165071  0.27463497      0      1 1.243729  0.05359824      0      1
## [3,] 1.0256930  0.28262046      0      1 1.359175  0.02608565      0      1
## [4,] 0.7749699  0.10605954      0      1 1.122507 -0.28542762      0      1
## [5,] 0.6949670 -0.07076239      0      1 1.477874 -0.25966007      0      1
## [6,] 0.9399259  0.14968395      0      1 1.373542 -0.34071317      0      1
##         [,251]     [,252] [,253] [,254]   [,255]    [,256] [,257] [,258]
## [1,] 0.9187056 -0.3566940      0      1 1.329434 0.9796426      0      1
## [2,] 0.9317980 -0.3641277      0      1 1.168574 1.0907550      0      1
## [3,] 1.1387606 -0.2939781      0      1 1.373722 1.2075709      0      1
## [4,] 0.7376370 -0.2814853      0      1 1.218199 0.9094240      0      1
## [5,] 0.7814198 -0.2855973      0      1 1.375830 0.9420577      0      1
## [6,] 0.9490407 -0.7161866      0      1 1.192153 0.8763603      0      1
##         [,259]     [,260] [,261] [,262]    [,263]    [,264] [,265] [,266]
## [1,] 0.9824335 -0.3661656      0      1 1.0877602 -1.467317      0      1
## [2,] 1.0276314 -0.4188786      0      1 0.9094788 -1.334707      0      1
## [3,] 0.9750434 -0.2791284      0      1 1.0743956 -1.319516      0      1
## [4,] 0.9512470 -0.7616715      0      1 0.8695797 -1.407272      0      1
## [5,] 0.8345820 -0.3958866      0      1 0.7894384 -1.345224      0      1
## [6,] 0.9722109 -0.5150122      0      1 1.1266870 -1.538534      0      1
##         [,267]    [,268] [,269] [,270]    [,271]     [,272] [,273] [,274]
## [1,] 1.0638303 0.6696183      0      1 0.6967309 -1.0746257      0      1
## [2,] 1.0658119 0.9005370      0      1 0.8394601 -1.0215187      0      1
## [3,] 1.3604434 0.8048931      0      1 1.0036252 -0.6163951      0      1
## [4,] 0.8865568 0.4923959      0      1 1.2842510 -1.0323700      0      1
## [5,] 0.9786950 0.4783595      0      1 1.0321751 -1.0343902      0      1
## [6,] 1.1506494 0.4766109      0      1 0.9441992 -0.9376986      0      1
##         [,275]     [,276] [,277] [,278]    [,279]     [,280] [,281] [,282]
## [1,] 1.3083693 0.12925243      0      1 1.0617989 -0.9978808      0      1
## [2,] 1.1479359 0.23276069      0      1 0.7961228 -0.9145039      0      1
## [3,] 1.1203828 0.33167591      0      1 0.9122386 -0.7241871      0      1
## [4,] 1.0536380 0.36664746      0      1 0.9727811 -1.0453773      0      1
## [5,] 1.1361198 0.16056387      0      1 0.9831541 -1.1352989      0      1
## [6,] 0.9857785 0.07735063      0      1 0.8574289 -1.0566909      0      1
##         [,283]    [,284] [,285] [,286]    [,287]     [,288] [,289] [,290]
## [1,] 1.1918508 1.3016980      0      1 1.3066930 -0.5461669      0      1
## [2,] 0.8914084 1.1534769      0      1 0.9626064 -0.3116573      0      1
## [3,] 1.1604658 1.1209385      0      1 1.2384079 -0.2605566      0      1
## [4,] 0.9861808 1.0519875      0      1 1.2244967 -0.5548135      0      1
## [5,] 0.8477871 1.0086655      0      1 0.9654957 -0.5311283      0      1
## [6,] 1.1448511 0.9740737      0      1 0.9486223 -0.5613192      0      1
##         [,291]     [,292] [,293] [,294]    [,295]     [,296] [,297] [,298]
## [1,] 0.8543815 -0.5298696      0      1 1.3449990 -0.3525254      0      1
## [2,] 0.8130169 -0.4966404      0      1 0.8587591 -0.3956026      0      1
## [3,] 1.0682920 -0.3439448      0      1 0.9909866 -0.2972000      0      1
## [4,] 0.7269867 -0.5210504      0      1 0.9559495 -0.3523335      0      1
## [5,] 0.9335176 -0.6651471      0      1 0.9184139 -0.2616903      0      1
## [6,] 0.7928735 -0.7776049      0      1 1.1255232 -0.5772838      0      1
##        [,299]    [,300] [,301] [,302]    [,303]    [,304] [,305] [,306]
## [1,] 1.034894 -1.699261      0      1 0.8269910 -1.340526      0      1
## [2,] 1.481448 -1.599038      0      1 0.9405661 -1.008521      0      1
## [3,] 1.227312 -1.319597      0      1 1.3291097 -1.008094      0      1
## [4,] 1.147980 -1.647176      0      1 0.9051996 -1.227721      0      1
## [5,] 1.171857 -1.901298      0      1 0.6080092 -1.208099      0      1
## [6,] 1.062384 -1.988857      0      1 0.9805426 -1.273074      0      1
##        [,307]    [,308] [,309] [,310]    [,311]     [,312] [,313] [,314]
## [1,] 1.109798 -2.077264      0      1 0.8659090 -1.2518730      0      1
## [2,] 1.149921 -2.108714      0      1 0.9968357 -0.8567747      0      1
## [3,] 1.076467 -2.001783      0      1 1.1171761 -0.8583764      0      1
## [4,] 1.024417 -2.062564      0      1 1.1173261 -1.0066144      0      1
## [5,] 1.182852 -2.359265      0      1 0.9599168 -1.2709238      0      1
## [6,] 0.822313 -2.377744      0      1 1.1797473 -1.3337544      0      1
##         [,315]      [,316] [,317] [,318]    [,319]     [,320] [,321] [,322]
## [1,] 0.8198482 -0.07561636      0      1 1.0259215 -0.2782652      0      1
## [2,] 0.9066170 -0.12383739      0      1 0.5988081 -0.1837468      0      1
## [3,] 1.1520987 -0.11644538      0      1 0.9983862 -0.2807078      0      1
## [4,] 1.0116487 -0.25403786      0      1 0.7988377 -0.2591919      0      1
## [5,] 0.9307391 -0.23234651      0      1 0.9092582 -0.6198339      0      1
## [6,] 1.0840434 -0.27668335      0      1 0.8241558 -0.2133823      0      1
##           [,323]    [,324]
## [1,] -0.14294123 0.8052436
## [2,] -0.31764526 1.1646612
## [3,] -0.31798755 0.7748365
## [4,] -0.15880406 1.0902159
## [5,] -0.11690219 0.9898398
## [6,] -0.08002515 0.9635314

## End(No test)

[Package mirt version 1.43 Index]