draw_parameters {mirt}	R Documentation

Draw plausible parameter instantiations from a given model

Description

Draws plausible parameters from a model using parametric sampling (if the information matrix was computed) or via boostrap sampling. Primarily for use with the DRF function.

Usage

draw_parameters(
  mod,
  draws,
  method = c("parametric", "boostrap"),
  redraws = 20,
  verbose = FALSE,
  ...
)

Arguments

mod
estimated single or multiple-group model

draws
number of draws to obtain

method
type of plausible values to obtain. Can be 'parametric', for the parametric sampling scheme which uses the estimated information matrix, or 'boostrap' to obtain values from the boot function. Default is 'parametric'

redraws
number of redraws to perform when the given parameteric sample does not satisfy the upper and lower parameter bounds. If a valid set cannot be found within this number of draws then an error will be thrown

verbose
logical; include additional information in the console?

...
additional arguments to be passed

Value

returns a draws x p matrix of plausible parameters, where each row correspeonds to a single set

Examples

## No test: 
set.seed(1234)
n <- 40
N <- 500

# only first 5 items as anchors
model <- 'F = 1-40
          CONSTRAINB = (1-5, a1), (1-5, d)'

a <- matrix(1, n)
d <- matrix(rnorm(n), n)
group <- c(rep('Group_1', N), rep('Group_2', N))

## -------------
# groups completely equal
dat1 <- simdata(a, d, N, itemtype = 'dich')
dat2 <- simdata(a, d, N, itemtype = 'dich')
dat <- rbind(dat1, dat2)
mod <- multipleGroup(dat, model, group=group, SE=TRUE,
                     invariance=c('free_means', 'free_var'))

param_set <- draw_parameters(mod, 100)
head(param_set)

##          [,1]      [,2] [,3] [,4]      [,5]      [,6] [,7] [,8]     [,9]
## [1,] 1.208328 -1.145969    0    1 1.0220734 0.5617959    0    1 1.066744
## [2,] 1.357488 -1.154840    0    1 1.0894010 0.4917074    0    1 1.102897
## [3,] 1.079252 -1.235091    0    1 1.1191024 0.2513951    0    1 1.134195
## [4,] 1.098925 -1.159577    0    1 1.0510720 0.4815690    0    1 1.201709
## [5,] 1.014480 -1.149454    0    1 1.1087850 0.4871285    0    1 1.154804
## [6,] 1.012945 -1.053866    0    1 0.9684608 0.3778970    0    1 1.147033
##         [,10] [,11] [,12]    [,13]     [,14] [,15] [,16]     [,17]     [,18]
## [1,] 1.193752     0     1 1.151657 -2.243354     0     1 0.8753515 0.4793675
## [2,] 1.178375     0     1 1.262503 -2.361813     0     1 0.8840302 0.6435631
## [3,] 1.072675     0     1 1.146907 -2.377617     0     1 0.8126080 0.3947152
## [4,] 1.315858     0     1 1.252381 -2.355673     0     1 0.8426679 0.4966367
## [5,] 1.210320     0     1 1.188245 -2.452722     0     1 0.9260490 0.5970086
## [6,] 1.284190     0     1 0.986116 -2.263431     0     1 0.7214550 0.4794008
##      [,19] [,20]     [,21]     [,22] [,23] [,24]     [,25]      [,26] [,27]
## [1,]     0     1 0.8951190 0.4980805     0     1 0.8551975 -0.2392818     0
## [2,]     0     1 1.4036981 0.5887119     0     1 1.3140167 -0.4763780     0
## [3,]     0     1 1.2633817 0.3096849     0     1 1.1705688 -0.2674143     0
## [4,]     0     1 0.9379468 0.2364438     0     1 1.1213767 -0.5046142     0
## [5,]     0     1 1.1009072 0.5501442     0     1 0.9743357 -0.2606921     0
## [6,]     0     1 1.3348026 0.3531976     0     1 1.1765122 -0.3261169     0
##      [,28]     [,29]      [,30] [,31] [,32]     [,33]      [,34] [,35] [,36]
## [1,]     1 0.9323392 -0.4476268     0     1 0.8879389 -0.4853375     0     1
## [2,]     1 1.1542200 -0.3544660     0     1 1.0288262 -0.4018358     0     1
## [3,]     1 1.1702705 -0.6840959     0     1 0.8305765 -0.3872699     0     1
## [4,]     1 1.1739225 -0.6831728     0     1 0.7804097 -0.2932243     0     1
## [5,]     1 1.1920185 -0.6399873     0     1 0.8915721 -0.4930857     0     1
## [6,]     1 1.0964068 -0.4162887     0     1 0.9625782 -0.3104983     0     1
##          [,37]      [,38] [,39] [,40]    [,41]      [,42] [,43] [,44]     [,45]
## [1,] 1.0219504 -0.8779007     0     1 1.297686 -0.2915698     0     1 1.1226593
## [2,] 1.2415569 -0.7659166     0     1 1.043310 -0.4154708     0     1 1.1593001
## [3,] 0.8760841 -0.8630848     0     1 1.171696 -0.5292467     0     1 1.0014933
## [4,] 0.9286273 -0.8865461     0     1 1.236403 -0.4307159     0     1 1.2822403
## [5,] 1.2550914 -0.9802652     0     1 1.036048 -0.4428513     0     1 0.9250708
## [6,] 1.2441688 -0.8560418     0     1 1.051719 -0.5015494     0     1 0.8685676
##           [,46] [,47] [,48]     [,49]      [,50] [,51] [,52]     [,53]
## [1,] -1.0693881     0     1 1.1126882 -0.7413017     0     1 0.7198128
## [2,] -0.9132021     0     1 1.1589746 -0.6620759     0     1 0.9810661
## [3,] -1.0094424     0     1 1.3469976 -0.8687497     0     1 0.8408400
## [4,] -1.1147395     0     1 1.0711821 -0.8140710     0     1 1.0207238
## [5,] -0.8432816     0     1 1.3794393 -0.8869437     0     1 0.6689462
## [6,] -0.8713375     0     1 0.8916763 -0.7080190     0     1 0.8436802
##            [,54] [,55] [,56]    [,57]     [,58] [,59] [,60]     [,61]
## [1,]  0.18832036     0     1 1.093495 0.8741494     0     1 0.8924900
## [2,]  0.27017673     0     1 1.430718 1.2202757     0     1 1.1499376
## [3,] -0.09797619     0     1 1.341057 0.8220261     0     1 1.0427619
## [4,]  0.08858861     0     1 1.153645 0.9783818     0     1 0.9315087
## [5,] -0.08051257     0     1 1.202665 1.1220019     0     1 1.2351637
## [6,]  0.01207303     0     1 1.138584 1.1166813     0     1 1.0143068
##             [,62] [,63] [,64]     [,65]      [,66] [,67] [,68]     [,69]
## [1,]  0.004494797     0     1 0.9862407 -0.3360680     0     1 0.8957561
## [2,]  0.073407448     0     1 0.8850150 -0.3078419     0     1 0.9937174
## [3,] -0.280875134     0     1 0.9278487 -0.5838062     0     1 1.2581990
## [4,] -0.119328612     0     1 1.0108382 -0.4011500     0     1 0.8902086
## [5,] -0.024930270     0     1 0.9168654 -0.3991393     0     1 0.9859686
## [6,]  0.129016882     0     1 1.1180479 -0.4210222     0     1 1.2240662
##           [,70] [,71] [,72]     [,73]      [,74] [,75] [,76]    [,77]    [,78]
## [1,] -0.9522470     0     1 1.0398228 -0.6585694     0     1 1.652558 3.246902
## [2,] -0.7974597     0     1 1.2789071 -0.7127028     0     1 1.483011 2.928691
## [3,] -0.9788969     0     1 0.9774792 -0.6193762     0     1 1.324311 2.597591
## [4,] -0.8817614     0     1 1.1353683 -0.5934303     0     1 1.348144 2.756988
## [5,] -0.9286965     0     1 1.2958833 -0.7245797     0     1 1.446325 2.896458
## [6,] -0.8852584     0     1 0.7463956 -0.4251890     0     1 1.245359 2.787335
##      [,79] [,80]     [,81]       [,82] [,83] [,84]    [,85]      [,86] [,87]
## [1,]     0     1 0.8825906  0.10024956     0     1 1.180183 -0.5434231     0
## [2,]     0     1 1.0451077  0.21986164     0     1 1.238010 -0.5963875     0
## [3,]     0     1 1.0272069 -0.04269200     0     1 1.243552 -0.7512982     0
## [4,]     0     1 0.8616179  0.15763298     0     1 1.360949 -0.5983241     0
## [5,]     0     1 0.8172467  0.12555087     0     1 1.049137 -0.5584484     0
## [6,]     0     1 1.0659170  0.05536605     0     1 1.470953 -0.5710640     0
##      [,88]     [,89]      [,90] [,91] [,92]     [,93]     [,94] [,95] [,96]
## [1,]     1 0.7062902 -0.3059541     0     1 1.1425039 0.5024045     0     1
## [2,]     1 0.9973814 -0.3453970     0     1 1.0252043 0.5108164     0     1
## [3,]     1 0.7973911 -0.4675832     0     1 0.8376975 0.3320440     0     1
## [4,]     1 0.8640228 -0.3718976     0     1 1.0046094 0.4774511     0     1
## [5,]     1 0.7749281 -0.3839533     0     1 1.2342228 0.4311545     0     1
## [6,]     1 1.0009354 -0.5099087     0     1 0.9774938 0.5281460     0     1
##          [,97]      [,98] [,99] [,100]   [,101]    [,102] [,103] [,104]
## [1,] 1.0382544 -0.3797200     0      1 1.186205 -1.301870      0      1
## [2,] 0.7804827 -0.5455762     0      1 1.599199 -1.506444      0      1
## [3,] 1.1509683 -0.7532208     0      1 1.369422 -1.598787      0      1
## [4,] 1.1551577 -0.6013286     0      1 1.383151 -1.500246      0      1
## [5,] 0.9644255 -0.5252820     0      1 1.316722 -1.192870      0      1
## [6,] 1.4024447 -0.6596732     0      1 1.265880 -1.428381      0      1
##        [,105]    [,106] [,107] [,108]    [,109]     [,110] [,111] [,112]
## [1,] 1.414032 0.8720546      0      1 0.8838262 -0.8503123      0      1
## [2,] 1.478024 0.8868632      0      1 1.2094251 -0.6769940      0      1
## [3,] 1.521905 0.7442979      0      1 1.3919575 -1.1966637      0      1
## [4,] 1.295021 0.9537243      0      1 1.1081865 -1.0165162      0      1
## [5,] 1.561583 0.9062210      0      1 1.1610639 -0.9727643      0      1
## [6,] 1.124799 0.8168231      0      1 0.9968335 -0.7764498      0      1
##         [,113]      [,114] [,115] [,116]    [,117]     [,118] [,119] [,120]
## [1,] 0.8012165 -0.06078404      0      1 0.8476663 -0.9031918      0      1
## [2,] 1.0989652 -0.11996229      0      1 0.7350352 -0.6418122      0      1
## [3,] 1.2451726 -0.04208152      0      1 1.0340934 -0.9691431      0      1
## [4,] 1.2119157 -0.28661520      0      1 0.9847756 -0.8111893      0      1
## [5,] 1.3629172 -0.07272020      0      1 1.2238703 -1.0409305      0      1
## [6,] 1.0689173 -0.11161532      0      1 1.0164399 -0.7862788      0      1
##         [,121]   [,122] [,123] [,124]    [,125]      [,126] [,127] [,128]
## [1,] 1.0421818 1.367278      0      1 0.8709250 -0.17890430      0      1
## [2,] 1.2228282 1.334115      0      1 0.8335159 -0.05934307      0      1
## [3,] 0.9029838 1.034466      0      1 0.8605289 -0.44305659      0      1
## [4,] 0.9368265 1.367077      0      1 1.0257879 -0.31998293      0      1
## [5,] 1.1872583 1.387492      0      1 1.0651525 -0.29739227      0      1
## [6,] 1.2745153 1.457552      0      1 0.8356135 -0.03990346      0      1
##         [,129]     [,130] [,131] [,132]    [,133]     [,134] [,135] [,136]
## [1,] 0.9643750 -0.5704161      0      1 0.6998825 -0.5070721      0      1
## [2,] 1.0471689 -0.5984069      0      1 1.0706775 -0.4966106      0      1
## [3,] 0.9137426 -0.5662779      0      1 1.0160797 -0.6155874      0      1
## [4,] 1.1203774 -0.7744783      0      1 0.7022529 -0.4264124      0      1
## [5,] 1.0561309 -0.4259370      0      1 0.9224881 -0.4877425      0      1
## [6,] 1.0657937 -0.6863912      0      1 0.7960620 -0.4327773      0      1
##        [,137]    [,138] [,139] [,140]    [,141]     [,142] [,143] [,144]
## [1,] 0.724052 -1.320014      0      1 0.8571693 -0.8544592      0      1
## [2,] 1.148410 -1.559861      0      1 0.8634515 -0.9405187      0      1
## [3,] 1.345787 -1.699282      0      1 1.0343118 -0.9665723      0      1
## [4,] 1.099991 -1.709910      0      1 0.9520824 -0.9198076      0      1
## [5,] 1.362509 -2.003395      0      1 0.8422713 -1.0225044      0      1
## [6,] 1.109800 -1.680059      0      1 0.9547631 -0.9065785      0      1
##         [,145]    [,146] [,147] [,148]    [,149]    [,150] [,151] [,152]
## [1,] 1.0611473 -2.214885      0      1 0.8168199 -1.386453      0      1
## [2,] 0.8647644 -2.212662      0      1 1.0597244 -1.502428      0      1
## [3,] 0.6552830 -2.295440      0      1 1.0428030 -1.524685      0      1
## [4,] 1.2638700 -2.380488      0      1 1.2832353 -1.582966      0      1
## [5,] 0.7769222 -2.161878      0      1 1.0052512 -1.426675      0      1
## [6,] 1.2545112 -2.544281      0      1 1.1146305 -1.306924      0      1
##         [,153]      [,154] [,155] [,156]    [,157]      [,158] [,159] [,160]
## [1,] 0.8678954 -0.14741258      0      1 0.9636552 -0.40921312      0      1
## [2,] 1.1278916 -0.02042404      0      1 1.2169989 -0.08123524      0      1
## [3,] 1.1308787 -0.18923966      0      1 1.1768076 -0.36784776      0      1
## [4,] 0.7027389 -0.09172143      0      1 0.9985754 -0.47819981      0      1
## [5,] 0.6289662 -0.03615891      0      1 0.8817627 -0.42687380      0      1
## [6,] 0.6522787 -0.36915438      0      1 1.1017273 -0.46636705      0      1
##      [,161] [,162]   [,163]    [,164] [,165] [,166]    [,167]    [,168] [,169]
## [1,]      0      1 1.208328 -1.145969      0      1 1.0220734 0.5617959      0
## [2,]      0      1 1.357488 -1.154840      0      1 1.0894010 0.4917074      0
## [3,]      0      1 1.079252 -1.235091      0      1 1.1191024 0.2513951      0
## [4,]      0      1 1.098925 -1.159577      0      1 1.0510720 0.4815690      0
## [5,]      0      1 1.014480 -1.149454      0      1 1.1087850 0.4871285      0
## [6,]      0      1 1.012945 -1.053866      0      1 0.9684608 0.3778970      0
##      [,170]   [,171]   [,172] [,173] [,174]   [,175]    [,176] [,177] [,178]
## [1,]      1 1.066744 1.193752      0      1 1.151657 -2.243354      0      1
## [2,]      1 1.102897 1.178375      0      1 1.262503 -2.361813      0      1
## [3,]      1 1.134195 1.072675      0      1 1.146907 -2.377617      0      1
## [4,]      1 1.201709 1.315858      0      1 1.252381 -2.355673      0      1
## [5,]      1 1.154804 1.210320      0      1 1.188245 -2.452722      0      1
## [6,]      1 1.147033 1.284190      0      1 0.986116 -2.263431      0      1
##         [,179]    [,180] [,181] [,182]    [,183]    [,184] [,185] [,186]
## [1,] 0.8753515 0.4793675      0      1 1.1349700 0.6834892      0      1
## [2,] 0.8840302 0.6435631      0      1 1.3176678 0.5927141      0      1
## [3,] 0.8126080 0.3947152      0      1 1.1848752 0.2839693      0      1
## [4,] 0.8426679 0.4966367      0      1 0.7648917 0.5132986      0      1
## [5,] 0.9260490 0.5970086      0      1 0.8896375 0.3446760      0      1
## [6,] 0.7214550 0.4794008      0      1 0.9741958 0.3955611      0      1
##         [,187]      [,188] [,189] [,190]    [,191]     [,192] [,193] [,194]
## [1,] 0.9914191 -0.27944864      0      1 1.0001631 -0.3633372      0      1
## [2,] 1.1477135 -0.04372981      0      1 0.9142705 -0.3590650      0      1
## [3,] 1.5303164 -0.51586703      0      1 0.6984767 -0.4774353      0      1
## [4,] 0.8434871 -0.50809367      0      1 0.7519772 -0.4221966      0      1
## [5,] 1.0250116 -0.54060893      0      1 0.9027272 -0.2190366      0      1
## [6,] 0.9470696 -0.36207742      0      1 1.1146928 -0.2665331      0      1
##         [,195]     [,196] [,197] [,198]    [,199]     [,200] [,201] [,202]
## [1,] 0.7491077 -0.4112571      0      1 0.8337765 -0.6404417      0      1
## [2,] 1.1574587 -0.4195545      0      1 1.2825745 -0.6730658      0      1
## [3,] 1.0626187 -0.4358063      0      1 1.0249484 -0.6600063      0      1
## [4,] 0.6120954 -0.5499848      0      1 1.0661130 -0.7611174      0      1
## [5,] 0.8501618 -0.3653878      0      1 0.7719732 -0.8536733      0      1
## [6,] 0.9092382 -0.4739351      0      1 0.9977273 -0.7791483      0      1
##         [,203]     [,204] [,205] [,206]    [,207]     [,208] [,209] [,210]
## [1,] 1.0991489 -0.3621456      0      1 1.1512788 -0.8747346      0      1
## [2,] 1.1541681 -0.6042400      0      1 1.6054416 -0.8216246      0      1
## [3,] 1.1557338 -0.4897813      0      1 1.3585469 -1.0607820      0      1
## [4,] 0.9562807 -0.3736347      0      1 1.0434438 -0.6887592      0      1
## [5,] 1.0027419 -0.5811291      0      1 0.8067972 -0.8635907      0      1
## [6,] 0.9401655 -0.5568498      0      1 1.2285736 -0.9173137      0      1
##         [,211]     [,212] [,213] [,214]    [,215]    [,216] [,217] [,218]
## [1,] 0.8093670 -0.6817949      0      1 0.9992066 0.5007050      0      1
## [2,] 1.1313071 -0.5465560      0      1 1.1591442 0.4758734      0      1
## [3,] 1.1478159 -0.7316758      0      1 1.1130478 0.1657295      0      1
## [4,] 1.0698328 -0.5460588      0      1 0.7588970 0.3010107      0      1
## [5,] 1.1246489 -0.7701388      0      1 0.9375081 0.3002141      0      1
## [6,] 0.7572978 -0.5496114      0      1 0.7563795 0.2776114      0      1
##         [,219]   [,220] [,221] [,222]    [,223]      [,224] [,225] [,226]
## [1,] 0.9552971 1.087054      0      1 0.8603082 -0.31409406      0      1
## [2,] 1.4037342 1.371514      0      1 1.0916765 -0.06825958      0      1
## [3,] 1.3365044 1.053080      0      1 0.9522338 -0.02919940      0      1
## [4,] 1.2175456 1.309529      0      1 0.8460995 -0.13073831      0      1
## [5,] 1.0807113 1.156740      0      1 0.8966635 -0.34943574      0      1
## [6,] 1.0802083 1.036824      0      1 0.9377517 -0.28685611      0      1
##         [,227]     [,228] [,229] [,230]    [,231]     [,232] [,233] [,234]
## [1,] 0.9532853 -0.1174120      0      1 0.6875382 -0.7562324      0      1
## [2,] 1.2384723 -0.1674194      0      1 1.0118600 -0.6988860      0      1
## [3,] 1.0870806 -0.3385235      0      1 0.6965977 -0.8311298      0      1
## [4,] 1.0746443 -0.3206087      0      1 0.8439035 -0.6482173      0      1
## [5,] 0.9060625 -0.3373859      0      1 0.7936639 -0.7811721      0      1
## [6,] 1.1054253 -0.1436248      0      1 1.0083235 -0.8347944      0      1
##         [,235]     [,236] [,237] [,238]    [,239]   [,240] [,241] [,242]
## [1,] 0.9310964 -0.5789449      0      1 0.9930800 2.698593      0      1
## [2,] 0.9480607 -0.4024566      0      1 1.1594594 2.544428      0      1
## [3,] 1.1646181 -0.6680155      0      1 1.3461972 2.531761      0      1
## [4,] 0.8012584 -0.5699412      0      1 1.2961720 3.021190      0      1
## [5,] 0.8140154 -0.7954179      0      1 0.9468783 2.493810      0      1
## [6,] 0.8254953 -0.4412170      0      1 1.4346863 2.866720      0      1
##         [,243]    [,244] [,245] [,246]   [,247]     [,248] [,249] [,250]
## [1,] 0.8754900 0.2091783      0      1 1.273474 -0.2199456      0      1
## [2,] 1.1550173 0.1788473      0      1 1.363651 -0.1046213      0      1
## [3,] 0.8082708 0.1275906      0      1 1.100474 -0.4650655      0      1
## [4,] 1.0117701 0.2126593      0      1 1.211854 -0.0804176      0      1
## [5,] 1.0052219 0.1960512      0      1 1.319178 -0.1510292      0      1
## [6,] 0.9269398 0.1745739      0      1 1.115034 -0.2175610      0      1
##         [,251]     [,252] [,253] [,254]   [,255]    [,256] [,257] [,258]
## [1,] 0.6575344 -0.4040546      0      1 1.352146 1.0296231      0      1
## [2,] 1.4429013 -0.1902387      0      1 1.692811 0.8697716      0      1
## [3,] 1.1937298 -0.3219946      0      1 1.095376 0.6893026      0      1
## [4,] 0.9207533 -0.3555472      0      1 1.226714 0.9337587      0      1
## [5,] 0.8491640 -0.3568146      0      1 1.309554 1.0893918      0      1
## [6,] 0.9360953 -0.3004467      0      1 1.235761 1.0060269      0      1
##         [,259]     [,260] [,261] [,262]    [,263]    [,264] [,265] [,266]
## [1,] 0.8049783 -0.2162008      0      1 1.0745686 -1.304194      0      1
## [2,] 1.2437820 -0.4263764      0      1 1.2562741 -1.178927      0      1
## [3,] 1.2059105 -0.4380052      0      1 1.0196214 -1.437674      0      1
## [4,] 1.0806885 -0.2632171      0      1 0.8022939 -1.418270      0      1
## [5,] 0.9064721 -0.4709006      0      1 1.0099683 -1.222115      0      1
## [6,] 0.8879293 -0.3284822      0      1 1.0907279 -1.525581      0      1
##         [,267]    [,268] [,269] [,270]    [,271]     [,272] [,273] [,274]
## [1,] 0.9896618 0.7740420      0      1 1.0895442 -0.9422603      0      1
## [2,] 1.3497119 0.6594897      0      1 1.0429883 -0.8106312      0      1
## [3,] 1.3644923 0.7163551      0      1 0.8661628 -0.7346780      0      1
## [4,] 1.1794542 0.6974463      0      1 0.8584152 -0.8039991      0      1
## [5,] 1.2706773 0.7122730      0      1 0.9892412 -0.7461964      0      1
## [6,] 0.9687638 0.5516726      0      1 0.7254638 -0.8328889      0      1
##        [,275]    [,276] [,277] [,278]   [,279]     [,280] [,281] [,282]
## [1,] 1.142470 0.1395470      0      1 1.156773 -0.9063993      0      1
## [2,] 1.349938 0.3006798      0      1 1.210994 -0.7432533      0      1
## [3,] 1.310042 0.1828768      0      1 1.177559 -0.8240892      0      1
## [4,] 1.044825 0.3252080      0      1 1.048052 -0.7501967      0      1
## [5,] 1.216028 0.2391018      0      1 1.069954 -0.9077939      0      1
## [6,] 0.987707 0.1855100      0      1 1.042416 -0.7491577      0      1
##         [,283]    [,284] [,285] [,286]   [,287]     [,288] [,289] [,290]
## [1,] 0.8881697 1.1182687      0      1 1.113978 -0.3071439      0      1
## [2,] 1.3416960 1.3907432      0      1 1.511881 -0.2207411      0      1
## [3,] 0.8737486 0.9048685      0      1 1.426395 -0.5665551      0      1
## [4,] 1.0039445 1.3194599      0      1 1.088197 -0.3227087      0      1
## [5,] 0.9265984 1.0073084      0      1 1.166144 -0.6339095      0      1
## [6,] 1.0059509 1.1770385      0      1 0.983238 -0.4815646      0      1
##         [,291]     [,292] [,293] [,294]    [,295]     [,296] [,297] [,298]
## [1,] 0.8670051 -0.2637136      0      1 1.0803591 -0.2373602      0      1
## [2,] 1.1205309 -0.3329890      0      1 1.3249704 -0.3718229      0      1
## [3,] 0.8408910 -0.5107872      0      1 1.0190539 -0.3189411      0      1
## [4,] 0.9390287 -0.5171743      0      1 1.0102302 -0.2787683      0      1
## [5,] 0.8577580 -0.7961129      0      1 1.1313184 -0.5170754      0      1
## [6,] 0.8347374 -0.5829431      0      1 0.8597526 -0.3373140      0      1
##         [,299]    [,300] [,301] [,302]    [,303]     [,304] [,305] [,306]
## [1,] 0.8179855 -1.657451      0      1 1.1675058 -1.3122676      0      1
## [2,] 1.4712512 -1.524051      0      1 0.9501566 -1.1900599      0      1
## [3,] 1.1178217 -1.641700      0      1 0.8918133 -1.1263637      0      1
## [4,] 0.9695901 -1.527646      0      1 1.0079680 -0.9459712      0      1
## [5,] 1.1881281 -1.733476      0      1 0.9233112 -1.1268525      0      1
## [6,] 0.8300903 -1.524680      0      1 0.7900015 -1.0605794      0      1
##         [,307]    [,308] [,309] [,310]    [,311]     [,312] [,313] [,314]
## [1,] 1.2117505 -2.246303      0      1 1.0258488 -1.1119970      0      1
## [2,] 1.1631032 -1.969912      0      1 1.1444696 -0.9233748      0      1
## [3,] 1.0141935 -2.484114      0      1 1.1831401 -1.1633868      0      1
## [4,] 1.0597010 -1.919837      0      1 0.9807682 -0.8123467      0      1
## [5,] 1.0987847 -2.344244      0      1 1.0070393 -1.1778866      0      1
## [6,] 0.9802237 -2.162652      0      1 1.0012725 -1.1158564      0      1
##         [,315]       [,316] [,317] [,318]    [,319]     [,320] [,321] [,322]
## [1,] 1.0843047 -0.046297238      0      1 0.6804833 -0.3641342      0      1
## [2,] 1.1480739 -0.007395804      0      1 1.1109133 -0.2706401      0      1
## [3,] 0.9967081 -0.215519821      0      1 1.1225246 -0.2761664      0      1
## [4,] 1.0813188 -0.192917363      0      1 0.7914098 -0.2669496      0      1
## [5,] 0.8940663 -0.031947946      0      1 0.6420401 -0.5589066      0      1
## [6,] 0.9778156 -0.136574723      0      1 0.7649839 -0.2924828      0      1
##           [,323]    [,324]
## [1,] -0.19216327 1.0426532
## [2,] -0.30204234 0.4537438
## [3,] -0.08920757 0.7441069
## [4,] -0.32530493 1.0474247
## [5,] -0.11838596 1.0381521
## [6,] -0.19846953 1.1226812

## End(No test)

[Package mirt version 1.36.2 Index]