Draw plausible parameter instantiations from a given model

Draws plausible parameters from a model using parametric sampling (if the information matrix was computed) or via bootstrap sampling. Primarily for use with the DRF function.

Usage

draw_parameters(
  mod,
  draws,
  method = c("parametric", "boostrap"),
  redraws = 20,
  verbose = FALSE,
  ...
)

Arguments

mod

estimated single or multiple-group model

draws

number of draws to obtain

method

type of plausible values to obtain. Can be 'parametric', for the parametric sampling scheme which uses the estimated information matrix, or 'boostrap' to obtain values from the boot function. Default is 'parametric'

redraws

number of redraws to perform when the given parameteric sample does not satisfy the upper and lower parameter bounds. If a valid set cannot be found within this number of draws then an error will be thrown

verbose

logical; include additional information in the console?

...

additional arguments to be passed

Value

returns a draws x p matrix of plausible parameters, where each row correspeonds to a single set

Examples

# \donttest{
set.seed(1234)
n <- 40
N <- 500

# only first 5 items as anchors
model <- 'F = 1-40
          CONSTRAINB = (1-5, a1), (1-5, d)'

a <- matrix(1, n)
d <- matrix(rnorm(n), n)
group <- c(rep('Group_1', N), rep('Group_2', N))

## -------------
# groups completely equal
dat1 <- simdata(a, d, N, itemtype = 'dich')
dat2 <- simdata(a, d, N, itemtype = 'dich')
dat <- rbind(dat1, dat2)
mod <- multipleGroup(dat, model, group=group, SE=TRUE,
                     invariance=c('free_means', 'free_var'))

param_set <- draw_parameters(mod, 100)
head(param_set)
#>           [,1]      [,2] [,3] [,4]      [,5]      [,6] [,7] [,8]     [,9]
#> [1,] 1.1391466 -1.262732    0    1 1.0429557 0.4288456    0    1 1.426606
#> [2,] 1.0484950 -1.270227    0    1 1.2642244 0.4139766    0    1 1.055069
#> [3,] 0.9015547 -1.120840    0    1 1.1142845 0.2913125    0    1 1.179313
#> [4,] 1.2083648 -1.252644    0    1 1.2302610 0.3839690    0    1 1.282216
#> [5,] 1.0422811 -1.129911    0    1 1.0267379 0.4106248    0    1 1.109888
#> [6,] 0.9569274 -1.073820    0    1 0.8923374 0.5172596    0    1 1.122108
#>          [,10] [,11] [,12]    [,13]     [,14] [,15] [,16]     [,17]     [,18]
#> [1,] 1.3571774     0     1 1.322894 -2.259730     0     1 0.7230033 0.4307521
#> [2,] 0.9838274     0     1 1.190297 -2.265337     0     1 0.7787838 0.6617068
#> [3,] 1.1912066     0     1 1.182237 -2.236729     0     1 0.9837763 0.5259850
#> [4,] 0.9824323     0     1 1.077731 -2.368946     0     1 0.8211142 0.3825557
#> [5,] 1.0501565     0     1 1.193853 -2.170215     0     1 0.7517755 0.3950838
#> [6,] 1.3031512     0     1 0.931496 -2.084264     0     1 0.8255580 0.5518470
#>      [,19] [,20]    [,21]     [,22] [,23] [,24]    [,25]      [,26] [,27] [,28]
#> [1,]     0     1 1.163286 0.4935818     0     1 1.138407 -0.4730955     0     1
#> [2,]     0     1 1.295991 0.2962792     0     1 1.277705 -0.3843502     0     1
#> [3,]     0     1 1.393148 0.4957510     0     1 1.207929 -0.5509626     0     1
#> [4,]     0     1 1.221361 0.2570869     0     1 1.050326 -0.5914608     0     1
#> [5,]     0     1 1.051666 0.1174212     0     1 1.302897 -0.4298328     0     1
#> [6,]     0     1 1.066978 0.3894483     0     1 1.192808 -0.4857308     0     1
#>          [,29]      [,30] [,31] [,32]     [,33]      [,34] [,35] [,36]
#> [1,] 1.3527922 -0.5621641     0     1 0.8924486 -0.5117360     0     1
#> [2,] 1.1797521 -0.5450939     0     1 0.9312757 -0.4114287     0     1
#> [3,] 1.4737985 -0.6303789     0     1 0.8593667 -0.4383548     0     1
#> [4,] 1.0532488 -0.6968822     0     1 1.0448985 -0.4547601     0     1
#> [5,] 0.9259634 -0.5687442     0     1 0.8781226 -0.6390869     0     1
#> [6,] 1.4960321 -0.4478337     0     1 0.8747682 -0.4139592     0     1
#>          [,37]      [,38] [,39] [,40]     [,41]      [,42] [,43] [,44]
#> [1,] 1.0394548 -0.9996516     0     1 1.3385254 -0.4398966     0     1
#> [2,] 0.8843197 -0.9601142     0     1 1.0098375 -0.4068907     0     1
#> [3,] 1.0472866 -0.8472970     0     1 1.0328670 -0.4165789     0     1
#> [4,] 1.0481752 -1.1208752     0     1 1.5198534 -0.3905801     0     1
#> [5,] 0.8323410 -0.7925511     0     1 0.9477634 -0.4690860     0     1
#> [6,] 0.8903161 -0.8060024     0     1 1.1486373 -0.5070982     0     1
#>          [,45]      [,46] [,47] [,48]     [,49]      [,50] [,51] [,52]
#> [1,] 1.0213154 -1.0562481     0     1 1.2856200 -0.9979774     0     1
#> [2,] 1.0863028 -0.9086137     0     1 1.1838080 -0.6951990     0     1
#> [3,] 1.3031162 -1.0691690     0     1 1.0733000 -0.7782116     0     1
#> [4,] 1.2671260 -1.2655032     0     1 1.0755919 -0.8150880     0     1
#> [5,] 0.9003790 -0.8236885     0     1 1.0510653 -0.7829117     0     1
#> [6,] 0.7936885 -0.8023493     0     1 0.8012045 -0.7711537     0     1
#>          [,53]      [,54] [,55] [,56]    [,57]     [,58] [,59] [,60]     [,61]
#> [1,] 0.8176164 0.14319037     0     1 1.300684 1.0884990     0     1 0.8592558
#> [2,] 0.9589931 0.23632296     0     1 1.465501 1.0258831     0     1 1.1314653
#> [3,] 0.8126470 0.21273306     0     1 1.384536 1.0716996     0     1 0.8341388
#> [4,] 0.9141173 0.12657448     0     1 1.376906 0.9620413     0     1 0.9868449
#> [5,] 0.9342269 0.07781069     0     1 1.084790 1.0487980     0     1 0.9410798
#> [6,] 1.0545782 0.01144942     0     1 1.368823 1.3333786     0     1 1.0745860
#>            [,62] [,63] [,64]     [,65]      [,66] [,67] [,68]    [,69]
#> [1,] -0.11398877     0     1 0.8092358 -0.5233629     0     1 1.313813
#> [2,] -0.02998945     0     1 1.0429868 -0.5469495     0     1 1.241366
#> [3,] -0.11668911     0     1 1.2888691 -0.5121931     0     1 1.274843
#> [4,] -0.21986683     0     1 1.1642869 -0.7154799     0     1 1.077223
#> [5,]  0.06666392     0     1 0.9142211 -0.4847572     0     1 1.226570
#> [6,]  0.11548229     0     1 0.8656716 -0.3201641     0     1 1.139974
#>           [,70] [,71] [,72]     [,73]      [,74] [,75] [,76]    [,77]    [,78]
#> [1,] -0.9808318     0     1 1.0888390 -0.7741621     0     1 1.178652 2.567551
#> [2,] -0.9901069     0     1 1.0017602 -0.5739475     0     1 1.713045 3.165313
#> [3,] -1.1133857     0     1 0.8404786 -0.5877344     0     1 1.559142 2.637754
#> [4,] -0.9457003     0     1 0.9973522 -0.5775370     0     1 1.482782 2.682693
#> [5,] -1.0808252     0     1 1.0079993 -0.6276279     0     1 1.150029 2.552865
#> [6,] -0.7694768     0     1 1.3709517 -0.6365555     0     1 1.572618 2.960952
#>      [,79] [,80]     [,81]         [,82] [,83] [,84]     [,85]      [,86] [,87]
#> [1,]     0     1 0.9749301  0.1262439618     0     1 1.2569353 -0.6753267     0
#> [2,]     0     1 0.6378720 -0.0416330361     0     1 1.3503415 -0.6989660     0
#> [3,]     0     1 1.0769427  0.0747546849     0     1 1.1304517 -0.4273359     0
#> [4,]     0     1 0.9466747  0.0001041858     0     1 1.2338614 -0.7305844     0
#> [5,]     0     1 0.9646444  0.0222671210     0     1 0.8675036 -0.5533653     0
#> [6,]     0     1 0.8953225  0.0550663892     0     1 0.9862144 -0.5061771     0
#>      [,88]     [,89]      [,90] [,91] [,92]     [,93]     [,94] [,95] [,96]
#> [1,]     1 0.8943716 -0.3695973     0     1 1.1414393 0.4496626     0     1
#> [2,]     1 0.8856731 -0.3901232     0     1 1.1720901 0.4237437     0     1
#> [3,]     1 0.7230544 -0.4177318     0     1 1.1242041 0.3721107     0     1
#> [4,]     1 0.7502930 -0.4709267     0     1 1.1538620 0.5534917     0     1
#> [5,]     1 0.8277908 -0.5723135     0     1 0.9200626 0.4596049     0     1
#> [6,]     1 0.8063402 -0.2831992     0     1 0.9777809 0.4521513     0     1
#>          [,97]      [,98] [,99] [,100]   [,101]    [,102] [,103] [,104]
#> [1,] 0.9558488 -0.6142072     0      1 1.345435 -1.290429      0      1
#> [2,] 1.1158181 -0.5876532     0      1 1.575488 -1.563279      0      1
#> [3,] 1.2789157 -0.6620498     0      1 1.565973 -1.553428      0      1
#> [4,] 1.0295186 -0.6704548     0      1 1.205469 -1.614798      0      1
#> [5,] 0.9924571 -0.5624844     0      1 1.092117 -1.555363      0      1
#> [6,] 1.0533409 -0.3613616     0      1 1.402935 -1.392784      0      1
#>        [,105]    [,106] [,107] [,108]    [,109]     [,110] [,111] [,112]
#> [1,] 1.371332 0.8212192      0      1 1.0591467 -0.7177679      0      1
#> [2,] 1.387873 0.6305891      0      1 1.2003530 -0.9511680      0      1
#> [3,] 1.236605 0.5940224      0      1 0.9270752 -0.9554079      0      1
#> [4,] 1.229710 0.5414319      0      1 0.8022843 -0.8891023      0      1
#> [5,] 1.279885 0.9269743      0      1 1.0720163 -1.0762102      0      1
#> [6,] 1.553092 1.0841694      0      1 1.3316021 -0.9780298      0      1
#>         [,113]      [,114] [,115] [,116]    [,117]     [,118] [,119] [,120]
#> [1,] 1.3841749 -0.11060560      0      1 0.9359197 -0.9057243      0      1
#> [2,] 1.2721097 -0.12842702      0      1 1.1714799 -0.9517767      0      1
#> [3,] 0.9154542 -0.25247084      0      1 1.2103053 -0.9697574      0      1
#> [4,] 1.2261741 -0.23379122      0      1 1.0271744 -0.7794422      0      1
#> [5,] 1.3606449 -0.11402170      0      1 0.9669167 -0.8006606      0      1
#> [6,] 1.1510462 -0.05004277      0      1 0.8964732 -0.7972518      0      1
#>         [,121]   [,122] [,123] [,124]    [,125]      [,126] [,127] [,128]
#> [1,] 1.0723194 1.231365      0      1 0.9407797 -0.32888606      0      1
#> [2,] 1.0880419 1.285116      0      1 1.0320999 -0.34028824      0      1
#> [3,] 1.3211029 1.525652      0      1 0.9438048 -0.27036418      0      1
#> [4,] 0.7331607 1.299448      0      1 1.0681650 -0.30187500      0      1
#> [5,] 1.1344798 1.469862      0      1 0.7784911 -0.10341601      0      1
#> [6,] 0.9448849 1.203272      0      1 0.8675712 -0.06169106      0      1
#>         [,129]     [,130] [,131] [,132]    [,133]     [,134] [,135] [,136]
#> [1,] 0.8582871 -0.5372003      0      1 1.0455377 -0.5890003      0      1
#> [2,] 1.0358110 -0.5796346      0      1 0.7891642 -0.4892284      0      1
#> [3,] 1.0376525 -0.5784201      0      1 1.0742263 -0.7452423      0      1
#> [4,] 0.9488971 -0.7974820      0      1 0.8301222 -0.6310389      0      1
#> [5,] 0.7512008 -0.7169416      0      1 0.8355657 -0.4990195      0      1
#> [6,] 1.2407197 -0.6411223      0      1 1.0449230 -0.4900294      0      1
#>         [,137]    [,138] [,139] [,140]    [,141]     [,142] [,143] [,144]
#> [1,] 1.2294915 -1.652614      0      1 0.8113574 -1.0591062      0      1
#> [2,] 0.9020176 -1.372767      0      1 0.9551096 -1.1610791      0      1
#> [3,] 1.0222266 -1.510695      0      1 0.8123734 -0.9439125      0      1
#> [4,] 0.7451934 -1.646803      0      1 0.9577524 -1.1057917      0      1
#> [5,] 0.9055043 -1.493830      0      1 0.8405727 -0.9015816      0      1
#> [6,] 1.0238202 -1.606560      0      1 0.5875165 -0.6929276      0      1
#>         [,145]    [,146] [,147] [,148]   [,149]    [,150] [,151] [,152]
#> [1,] 0.9904558 -2.317326      0      1 1.117666 -1.358480      0      1
#> [2,] 1.1253020 -2.307847      0      1 1.051013 -1.437486      0      1
#> [3,] 1.1460397 -2.378568      0      1 1.337640 -1.418851      0      1
#> [4,] 1.2316899 -2.440185      0      1 1.103262 -1.415160      0      1
#> [5,] 1.2664408 -2.624069      0      1 1.225434 -1.484991      0      1
#> [6,] 1.1905025 -2.587986      0      1 1.003452 -1.142933      0      1
#>         [,153]     [,154] [,155] [,156]    [,157]     [,158] [,159] [,160]
#> [1,] 1.0051991 -0.1446542      0      1 1.2271215 -0.2833808      0      1
#> [2,] 0.7331591 -0.1336433      0      1 0.8868182 -0.3848594      0      1
#> [3,] 0.7006685 -0.1968367      0      1 1.1800768 -0.3825686      0      1
#> [4,] 0.7622291 -0.2943376      0      1 1.0171921 -0.2595964      0      1
#> [5,] 0.7589868 -0.3463511      0      1 0.9470741 -0.4106837      0      1
#> [6,] 0.8812604 -0.1017550      0      1 1.1799061 -0.2237307      0      1
#>      [,161] [,162]    [,163]    [,164] [,165] [,166]    [,167]    [,168] [,169]
#> [1,]      0      1 1.1391466 -1.262732      0      1 1.0429557 0.4288456      0
#> [2,]      0      1 1.0484950 -1.270227      0      1 1.2642244 0.4139766      0
#> [3,]      0      1 0.9015547 -1.120840      0      1 1.1142845 0.2913125      0
#> [4,]      0      1 1.2083648 -1.252644      0      1 1.2302610 0.3839690      0
#> [5,]      0      1 1.0422811 -1.129911      0      1 1.0267379 0.4106248      0
#> [6,]      0      1 0.9569274 -1.073820      0      1 0.8923374 0.5172596      0
#>      [,170]   [,171]    [,172] [,173] [,174]   [,175]    [,176] [,177] [,178]
#> [1,]      1 1.426606 1.3571774      0      1 1.322894 -2.259730      0      1
#> [2,]      1 1.055069 0.9838274      0      1 1.190297 -2.265337      0      1
#> [3,]      1 1.179313 1.1912066      0      1 1.182237 -2.236729      0      1
#> [4,]      1 1.282216 0.9824323      0      1 1.077731 -2.368946      0      1
#> [5,]      1 1.109888 1.0501565      0      1 1.193853 -2.170215      0      1
#> [6,]      1 1.122108 1.3031512      0      1 0.931496 -2.084264      0      1
#>         [,179]    [,180] [,181] [,182]    [,183]    [,184] [,185] [,186]
#> [1,] 0.7230033 0.4307521      0      1 0.5769388 0.3785589      0      1
#> [2,] 0.7787838 0.6617068      0      1 0.9951119 0.5035140      0      1
#> [3,] 0.9837763 0.5259850      0      1 0.6653222 0.2314168      0      1
#> [4,] 0.8211142 0.3825557      0      1 0.9269222 0.3332792      0      1
#> [5,] 0.7517755 0.3950838      0      1 0.6941894 0.4616587      0      1
#> [6,] 0.8255580 0.5518470      0      1 0.8832418 0.6587158      0      1
#>         [,187]     [,188] [,189] [,190]    [,191]     [,192] [,193] [,194]
#> [1,] 1.1773518 -0.5980202      0      1 0.8270720 -0.3566062      0      1
#> [2,] 1.2006452 -0.4580316      0      1 1.2312009 -0.4468917      0      1
#> [3,] 0.9915531 -0.4397262      0      1 0.7832415 -0.3698774      0      1
#> [4,] 0.8631275 -0.6454301      0      1 0.9260350 -0.5538107      0      1
#> [5,] 1.1585960 -0.5414431      0      1 0.9432135 -0.4511658      0      1
#> [6,] 1.1344602 -0.1556356      0      1 0.8387838 -0.2676410      0      1
#>         [,195]     [,196] [,197] [,198]    [,199]     [,200] [,201] [,202]
#> [1,] 0.7955679 -0.6940780      0      1 1.1881213 -0.8672479      0      1
#> [2,] 1.0363178 -0.6676125      0      1 0.9175609 -0.8380504      0      1
#> [3,] 0.8495408 -0.4471783      0      1 1.0530105 -0.7772059      0      1
#> [4,] 0.6804507 -0.7378905      0      1 0.9491294 -0.8842631      0      1
#> [5,] 0.8446017 -0.3752494      0      1 0.7418588 -0.4362904      0      1
#> [6,] 0.7291264 -0.3666279      0      1 0.8138688 -0.5491412      0      1
#>         [,203]     [,204] [,205] [,206]    [,207]     [,208] [,209] [,210]
#> [1,] 1.1721466 -0.5048695      0      1 1.4271223 -0.8890825      0      1
#> [2,] 0.9024845 -0.3658077      0      1 1.3371508 -0.7301239      0      1
#> [3,] 0.8207079 -0.6218742      0      1 1.2301943 -0.8838944      0      1
#> [4,] 1.1753149 -0.8479801      0      1 1.0316858 -1.2796658      0      1
#> [5,] 0.9550714 -0.3683115      0      1 1.5088142 -0.9401337      0      1
#> [6,] 0.7787619 -0.1532719      0      1 0.6774216 -0.6721829      0      1
#>         [,211]     [,212] [,213] [,214]    [,215]     [,216] [,217] [,218]
#> [1,] 1.0828452 -0.5861556      0      1 0.9437336 0.09630676      0      1
#> [2,] 0.9179873 -0.6699210      0      1 1.1634014 0.40873175      0      1
#> [3,] 0.8099081 -0.6891363      0      1 0.9437219 0.36898479      0      1
#> [4,] 0.9380594 -0.7708488      0      1 0.9953128 0.24996544      0      1
#> [5,] 0.9065546 -0.5227409      0      1 0.8901220 0.22865416      0      1
#> [6,] 0.8478855 -0.5027797      0      1 0.8493828 0.64925409      0      1
#>        [,219]   [,220] [,221] [,222]    [,223]      [,224] [,225] [,226]
#> [1,] 1.251221 1.266023      0      1 0.8093781 -0.03781391      0      1
#> [2,] 1.275117 1.170511      0      1 0.8337817 -0.08649730      0      1
#> [3,] 1.075977 1.033217      0      1 0.8602997 -0.24081851      0      1
#> [4,] 1.178969 0.896589      0      1 1.0906765 -0.21602028      0      1
#> [5,] 1.465805 1.452843      0      1 0.9667596  0.16206983      0      1
#> [6,] 1.003671 1.281015      0      1 0.9793592  0.08787345      0      1
#>         [,227]     [,228] [,229] [,230]    [,231]     [,232] [,233] [,234]
#> [1,] 1.1241160 -0.2655825      0      1 0.8256853 -0.6485612      0      1
#> [2,] 1.1013481 -0.3204749      0      1 1.0770737 -0.8222168      0      1
#> [3,] 1.2208760 -0.2838568      0      1 0.7337515 -0.7363449      0      1
#> [4,] 1.1255037 -0.5292447      0      1 0.7869515 -1.0085977      0      1
#> [5,] 1.3929428 -0.1996781      0      1 0.7863706 -0.7639819      0      1
#> [6,] 0.9912011 -0.1305225      0      1 0.7348661 -0.7732199      0      1
#>         [,235]     [,236] [,237] [,238]    [,239]   [,240] [,241] [,242]
#> [1,] 1.0720076 -0.4475811      0      1 1.0192269 2.661383      0      1
#> [2,] 0.9165141 -0.5343090      0      1 1.3177828 2.757126      0      1
#> [3,] 0.9368476 -0.6705803      0      1 0.9742970 2.529697      0      1
#> [4,] 0.8131264 -0.5930898      0      1 0.9871156 2.429143      0      1
#> [5,] 1.0643571 -0.4905185      0      1 1.2126903 2.697156      0      1
#> [6,] 0.8418788 -0.4321600      0      1 0.7567289 2.617331      0      1
#>         [,243]      [,244] [,245] [,246]    [,247]      [,248] [,249] [,250]
#> [1,] 0.8355120  0.06781088      0      1 1.4137778 -0.21320620      0      1
#> [2,] 1.1336158  0.25348323      0      1 1.2815124 -0.17706541      0      1
#> [3,] 0.5790890  0.13768966      0      1 1.2638129 -0.27375573      0      1
#> [4,] 0.9006469 -0.10518950      0      1 1.0604838 -0.50714109      0      1
#> [5,] 0.9075214  0.18476248      0      1 1.5191378 -0.14975678      0      1
#> [6,] 0.9369070  0.43921166      0      1 0.9969372  0.06122403      0      1
#>         [,251]     [,252] [,253] [,254]    [,255]    [,256] [,257] [,258]
#> [1,] 1.1006380 -0.2310023      0      1 1.4072849 0.8883015      0      1
#> [2,] 1.0802242 -0.6114732      0      1 1.2684356 0.9621889      0      1
#> [3,] 0.7880489 -0.3299429      0      1 1.1213530 0.9072507      0      1
#> [4,] 0.9676486 -0.5462879      0      1 1.3448592 0.6590060      0      1
#> [5,] 1.2516732 -0.2830268      0      1 1.4669822 1.2250320      0      1
#> [6,] 0.8800127 -0.2295368      0      1 0.9690159 1.0716072      0      1
#>         [,259]     [,260] [,261] [,262]    [,263]    [,264] [,265] [,266]
#> [1,] 1.0834702 -0.2874097      0      1 1.2372478 -1.584904      0      1
#> [2,] 1.0168658 -0.4050202      0      1 0.7296529 -1.262675      0      1
#> [3,] 1.2575103 -0.5186030      0      1 0.8585336 -1.362947      0      1
#> [4,] 0.9251529 -0.7952480      0      1 1.1705157 -1.415849      0      1
#> [5,] 1.0538576 -0.5370151      0      1 0.9277009 -1.212447      0      1
#> [6,] 0.6340682 -0.3535294      0      1 0.8223437 -1.063422      0      1
#>        [,267]    [,268] [,269] [,270]    [,271]     [,272] [,273] [,274]
#> [1,] 1.306094 0.6539572      0      1 0.9254564 -1.0119465      0      1
#> [2,] 1.165126 0.6360270      0      1 0.6041941 -0.8960466      0      1
#> [3,] 1.042477 0.5804425      0      1 0.9719381 -0.9049791      0      1
#> [4,] 1.020742 0.4597298      0      1 0.8853996 -0.8461720      0      1
#> [5,] 1.343506 0.7992435      0      1 1.0889564 -0.7382568      0      1
#> [6,] 0.879895 0.8475420      0      1 0.8833626 -0.9095419      0      1
#>         [,275]     [,276] [,277] [,278]    [,279]     [,280] [,281] [,282]
#> [1,] 1.1094768 0.26996213      0      1 1.2318713 -0.9162525      0      1
#> [2,] 1.1581164 0.22195852      0      1 0.9474639 -0.8550198      0      1
#> [3,] 0.8855225 0.02187362      0      1 1.0470272 -1.0408446      0      1
#> [4,] 1.0652420 0.08396112      0      1 1.0388019 -0.9947843      0      1
#> [5,] 1.2904355 0.37359744      0      1 0.9947425 -0.7320641      0      1
#> [6,] 1.0330122 0.53340093      0      1 1.0023864 -0.6590190      0      1
#>         [,283]   [,284] [,285] [,286]    [,287]     [,288] [,289] [,290]
#> [1,] 1.0127823 1.086734      0      1 1.3016414 -0.5645558      0      1
#> [2,] 0.9325871 1.185045      0      1 1.2797537 -0.3321647      0      1
#> [3,] 0.8565887 1.200679      0      1 1.0858045 -0.5370098      0      1
#> [4,] 0.7655176 1.001384      0      1 0.9557775 -0.6196753      0      1
#> [5,] 0.9698479 1.206207      0      1 1.0974368 -0.3217709      0      1
#> [6,] 0.8846880 1.368723      0      1 0.6796234 -0.4113856      0      1
#>         [,291]     [,292] [,293] [,294]    [,295]      [,296] [,297] [,298]
#> [1,] 0.9665656 -0.4499586      0      1 1.2751136 -0.27242264      0      1
#> [2,] 1.0356769 -0.5721720      0      1 1.1089564 -0.16897548      0      1
#> [3,] 0.9499311 -0.4461398      0      1 1.1750387 -0.44589464      0      1
#> [4,] 0.6212646 -0.6609367      0      1 0.9129271 -0.70055388      0      1
#> [5,] 0.9964668 -0.5005566      0      1 1.1185724 -0.19530289      0      1
#> [6,] 0.7778943 -0.4090571      0      1 1.0454043 -0.08128453      0      1
#>         [,299]    [,300] [,301] [,302]    [,303]    [,304] [,305] [,306]
#> [1,] 0.9457158 -1.694387      0      1 0.8919256 -1.298182      0      1
#> [2,] 1.1671585 -1.568943      0      1 1.1484209 -1.217314      0      1
#> [3,] 0.9104830 -1.588929      0      1 0.7172748 -1.194906      0      1
#> [4,] 0.7951280 -1.910012      0      1 0.9743102 -1.257370      0      1
#> [5,] 0.8852881 -1.377143      0      1 0.9976048 -1.021567      0      1
#> [6,] 0.8919724 -1.526344      0      1 0.7979577 -1.070962      0      1
#>         [,307]    [,308] [,309] [,310]    [,311]    [,312] [,313] [,314]
#> [1,] 0.9776836 -1.917575      0      1 0.7331025 -0.976802      0      1
#> [2,] 0.9623449 -1.886201      0      1 1.0542558 -1.065177      0      1
#> [3,] 0.9433830 -2.146877      0      1 1.0355105 -1.116183      0      1
#> [4,] 1.0764580 -2.357391      0      1 1.0398692 -1.217281      0      1
#> [5,] 1.0682595 -2.334698      0      1 1.1997882 -1.072752      0      1
#> [6,] 0.8298636 -1.838796      0      1 1.0184892 -1.013326      0      1
#>         [,315]        [,316] [,317] [,318]    [,319]     [,320] [,321] [,322]
#> [1,] 1.0091735 -0.2196754000      0      1 1.0090686 -0.2500386      0      1
#> [2,] 1.0190790 -0.0007909197      0      1 0.9135182 -0.2922381      0      1
#> [3,] 1.0794865 -0.1673695662      0      1 0.7342340 -0.3530659      0      1
#> [4,] 0.9524019 -0.4012828584      0      1 0.9772363 -0.2793675      0      1
#> [5,] 1.1289963  0.1365166467      0      1 0.7633138 -0.1245854      0      1
#> [6,] 0.8079592 -0.0061984044      0      1 0.7488025 -0.0674045      0      1
#>           [,323]    [,324]
#> [1,] -0.17453294 0.7971852
#> [2,] -0.23051654 0.8413674
#> [3,] -0.17643803 1.0450901
#> [4,]  0.06442799 1.0789415
#> [5,] -0.27483590 0.8621857
#> [6,] -0.47548109 1.1212373
# }