draw_parameters {mirt}	R Documentation

Draw plausible parameter instantiations from a given model

Description

Draws plausible parameters from a model using parametric sampling (if the information matrix was computed) or via bootstrap sampling. Primarily for use with the DRF function.

Usage

draw_parameters(
  mod,
  draws,
  method = c("parametric", "boostrap"),
  redraws = 20,
  verbose = FALSE,
  ...
)

Arguments

mod
estimated single or multiple-group model

draws
number of draws to obtain

method
type of plausible values to obtain. Can be 'parametric', for the parametric sampling scheme which uses the estimated information matrix, or 'boostrap' to obtain values from the boot function. Default is 'parametric'

redraws
number of redraws to perform when the given parameteric sample does not satisfy the upper and lower parameter bounds. If a valid set cannot be found within this number of draws then an error will be thrown

verbose
logical; include additional information in the console?

...
additional arguments to be passed

Value

returns a draws x p matrix of plausible parameters, where each row correspeonds to a single set

Examples

## No test: 
set.seed(1234)
n <- 40
N <- 500

# only first 5 items as anchors
model <- 'F = 1-40
          CONSTRAINB = (1-5, a1), (1-5, d)'

a <- matrix(1, n)
d <- matrix(rnorm(n), n)
group <- c(rep('Group_1', N), rep('Group_2', N))

## -------------
# groups completely equal
dat1 <- simdata(a, d, N, itemtype = 'dich')
dat2 <- simdata(a, d, N, itemtype = 'dich')
dat <- rbind(dat1, dat2)
mod <- multipleGroup(dat, model, group=group, SE=TRUE,
                     invariance=c('free_means', 'free_var'))

param_set <- draw_parameters(mod, 100)
head(param_set)

##           [,1]       [,2] [,3] [,4]      [,5]      [,6] [,7] [,8]     [,9]
## [1,] 1.0103163 -1.1626794    0    1 1.0574152 0.3014738    0    1 1.166507
## [2,] 1.2785683 -1.1979903    0    1 1.1268941 0.3522184    0    1 1.185616
## [3,] 1.1545179 -1.0676579    0    1 1.0265333 0.4469994    0    1 1.223583
## [4,] 0.9767176 -1.1982696    0    1 1.0911564 0.4433458    0    1 1.135087
## [5,] 1.2498596 -1.2419046    0    1 1.0499027 0.2981972    0    1 1.239663
## [6,] 1.0232032 -0.9749986    0    1 0.9714587 0.5951520    0    1 1.289774
##         [,10] [,11] [,12]    [,13]     [,14] [,15] [,16]     [,17]     [,18]
## [1,] 1.084819     0     1 1.194127 -2.076630     0     1 0.9220800 0.6038229
## [2,] 1.090233     0     1 1.193952 -2.425191     0     1 0.8667636 0.5329941
## [3,] 1.275322     0     1 1.147366 -2.250580     0     1 0.9132967 0.6383999
## [4,] 1.087048     0     1 1.114389 -2.059807     0     1 0.7351379 0.6555570
## [5,] 1.078973     0     1 1.612269 -2.573731     0     1 0.8743595 0.4251740
## [6,] 1.316082     0     1 1.105384 -2.043871     0     1 0.8297941 0.6649090
##      [,19] [,20]     [,21]     [,22] [,23] [,24]     [,25]      [,26] [,27]
## [1,]     0     1 1.3419108 0.4871712     0     1 1.1461505 -0.4933308     0
## [2,]     0     1 0.9907073 0.3243512     0     1 0.9865663 -0.5235888     0
## [3,]     0     1 1.2003941 0.3547740     0     1 1.1850334 -0.5484171     0
## [4,]     0     1 1.1962273 0.6858187     0     1 1.1660580 -0.4254388     0
## [5,]     0     1 1.1512028 0.3649826     0     1 1.0764915 -0.4722607     0
## [6,]     0     1 0.5655467 0.5178048     0     1 1.0297179 -0.5115625     0
##      [,28]     [,29]      [,30] [,31] [,32]     [,33]      [,34] [,35] [,36]
## [1,]     1 1.1083820 -0.5863011     0     1 0.7489231 -0.4929587     0     1
## [2,]     1 1.2131171 -0.7237987     0     1 0.8222562 -0.4678953     0     1
## [3,]     1 1.2436207 -0.6392854     0     1 1.0684792 -0.5014053     0     1
## [4,]     1 1.2059973 -0.7029644     0     1 0.7830607 -0.3434912     0     1
## [5,]     1 0.9212799 -0.5947554     0     1 0.8978260 -0.4364632     0     1
## [6,]     1 1.4231190 -0.5551854     0     1 0.9961318 -0.4610126     0     1
##          [,37]      [,38] [,39] [,40]    [,41]      [,42] [,43] [,44]     [,45]
## [1,] 0.8637237 -1.0801508     0     1 1.000434 -0.2604544     0     1 1.0624291
## [2,] 0.9405760 -0.9594637     0     1 1.122590 -0.4397564     0     1 0.8916968
## [3,] 0.9941571 -0.7366862     0     1 1.104328 -0.3786410     0     1 1.1771663
## [4,] 1.2325688 -1.1197597     0     1 1.147892 -0.4749790     0     1 0.9077350
## [5,] 1.0479080 -1.1285106     0     1 1.189476 -0.5391407     0     1 0.9445327
## [6,] 0.9716449 -0.9442414     0     1 1.121802 -0.3548517     0     1 0.9964929
##          [,46] [,47] [,48]     [,49]      [,50] [,51] [,52]     [,53]
## [1,] -1.185577     0     1 1.0467465 -0.7346138     0     1 1.0588903
## [2,] -1.109418     0     1 1.1546171 -0.9294792     0     1 0.8817898
## [3,] -0.981086     0     1 1.1462204 -0.8126261     0     1 1.0134758
## [4,] -1.128365     0     1 0.9791860 -0.8211259     0     1 0.7610715
## [5,] -0.982239     0     1 0.8755773 -0.8889870     0     1 1.0712022
## [6,] -0.706657     0     1 1.0893074 -0.7250260     0     1 0.8714601
##           [,54] [,55] [,56]    [,57]     [,58] [,59] [,60]     [,61]
## [1,] 0.18824828     0     1 1.345931 0.9148456     0     1 0.9831930
## [2,] 0.11403497     0     1 1.109831 1.1394553     0     1 0.8355639
## [3,] 0.08774318     0     1 1.024235 0.9508767     0     1 1.0336452
## [4,] 0.12986246     0     1 1.154932 1.0750144     0     1 0.8791226
## [5,] 0.16758741     0     1 1.135145 0.8953826     0     1 1.1056766
## [6,] 0.30420193     0     1 1.301079 1.2328739     0     1 1.0819915
##            [,62] [,63] [,64]     [,65]      [,66] [,67] [,68]     [,69]
## [1,] -0.09404233     0     1 0.8348042 -0.5372691     0     1 0.8753129
## [2,] -0.02999197     0     1 1.1348559 -0.4015238     0     1 1.0384113
## [3,] -0.23070585     0     1 1.0376871 -0.4718945     0     1 1.1994372
## [4,] -0.14261073     0     1 1.0255555 -0.5643716     0     1 1.2271250
## [5,]  0.01011741     0     1 0.8866388 -0.3017437     0     1 1.2566888
## [6,] -0.01205956     0     1 0.8969740 -0.2671314     0     1 0.9378357
##           [,70] [,71] [,72]     [,73]      [,74] [,75] [,76]     [,77]    [,78]
## [1,] -0.7919735     0     1 1.2349199 -0.6933985     0     1 0.9150622 2.314158
## [2,] -0.9451285     0     1 1.0227205 -0.6313385     0     1 1.0242769 2.553469
## [3,] -1.1441254     0     1 1.1616125 -0.8835160     0     1 1.3906777 3.004303
## [4,] -1.1079774     0     1 0.9790090 -0.6418385     0     1 1.0849250 2.830739
## [5,] -1.2886087     0     1 1.1045090 -0.6679676     0     1 1.1926862 2.447794
## [6,] -0.6962885     0     1 0.9994566 -0.6087026     0     1 1.0464843 2.757824
##      [,79] [,80]     [,81]        [,82] [,83] [,84]    [,85]      [,86] [,87]
## [1,]     0     1 0.9005401  0.043035704     0     1 1.127683 -0.4996656     0
## [2,]     0     1 1.0086064 -0.000180657     0     1 1.115908 -0.7813938     0
## [3,]     0     1 0.8530129  0.081759560     0     1 1.010127 -0.8723785     0
## [4,]     0     1 0.9132467 -0.043669972     0     1 0.759433 -0.4622772     0
## [5,]     0     1 0.9622858 -0.011666095     0     1 1.201648 -0.6399106     0
## [6,]     0     1 1.0344530  0.202808306     0     1 1.160394 -0.3470463     0
##      [,88]     [,89]      [,90] [,91] [,92]     [,93]     [,94] [,95] [,96]
## [1,]     1 0.8193471 -0.4446095     0     1 0.7976472 0.2889348     0     1
## [2,]     1 0.6586122 -0.3302503     0     1 0.8344476 0.3939533     0     1
## [3,]     1 0.6692271 -0.4801828     0     1 0.9685518 0.4491879     0     1
## [4,]     1 0.6964850 -0.4029312     0     1 0.8016952 0.3687072     0     1
## [5,]     1 0.7985686 -0.3422180     0     1 0.9213467 0.3742333     0     1
## [6,]     1 0.6951376 -0.3145844     0     1 0.9908559 0.5601991     0     1
##          [,97]      [,98] [,99] [,100]   [,101]    [,102] [,103] [,104]
## [1,] 0.8845121 -0.6929273     0      1 1.348276 -1.475411      0      1
## [2,] 1.2331846 -0.8040350     0      1 1.268974 -1.677208      0      1
## [3,] 1.2089820 -0.6839073     0      1 1.278573 -1.608455      0      1
## [4,] 0.9031432 -0.6710719     0      1 1.448540 -1.273003      0      1
## [5,] 0.8087657 -0.6227362     0      1 1.166303 -1.382452      0      1
## [6,] 1.1601957 -0.4231275     0      1 1.279998 -1.456775      0      1
##        [,105]    [,106] [,107] [,108]    [,109]     [,110] [,111] [,112]
## [1,] 1.594150 0.7250143      0      1 1.0453871 -0.9947416      0      1
## [2,] 1.029742 0.7837866      0      1 0.9571789 -0.9895031      0      1
## [3,] 1.182934 0.8175176      0      1 1.0394411 -0.8411525      0      1
## [4,] 1.103954 0.7434528      0      1 1.1913103 -0.8557881      0      1
## [5,] 1.531453 1.0994663      0      1 0.9442853 -0.9118235      0      1
## [6,] 1.205255 0.7838138      0      1 0.9640912 -0.8941962      0      1
##         [,113]      [,114] [,115] [,116]    [,117]     [,118] [,119] [,120]
## [1,] 1.0445487 -0.06417902      0      1 1.1306916 -1.1520581      0      1
## [2,] 1.0922369 -0.30069944      0      1 1.1518975 -1.0396717      0      1
## [3,] 0.9027324 -0.12563169      0      1 0.7149750 -1.0680878      0      1
## [4,] 1.2030864 -0.10012609      0      1 0.8096186 -0.9645334      0      1
## [5,] 1.3293594 -0.18443564      0      1 0.9950675 -1.0735559      0      1
## [6,] 1.1676462 -0.12191578      0      1 1.1033131 -0.6070633      0      1
##         [,121]   [,122] [,123] [,124]    [,125]      [,126] [,127] [,128]
## [1,] 0.9611096 1.150393      0      1 0.9754102 -0.47369043      0      1
## [2,] 0.9946150 1.407722      0      1 1.0117525 -0.27648606      0      1
## [3,] 1.0796933 1.282225      0      1 0.7897721 -0.04982585      0      1
## [4,] 1.1056062 1.268379      0      1 0.7764922 -0.31207145      0      1
## [5,] 1.0634702 1.165443      0      1 1.1406668 -0.17683005      0      1
## [6,] 1.0493929 1.172515      0      1 0.9821823 -0.15458154      0      1
##         [,129]     [,130] [,131] [,132]    [,133]     [,134] [,135] [,136]
## [1,] 0.9686464 -0.6389839      0      1 0.9306497 -0.6531876      0      1
## [2,] 1.0144047 -0.5702806      0      1 0.6425428 -0.5561375      0      1
## [3,] 0.8856001 -0.5952571      0      1 0.7555390 -0.3902578      0      1
## [4,] 0.9152130 -0.7242670      0      1 0.9399757 -0.5763202      0      1
## [5,] 0.7768845 -0.2824460      0      1 1.0368374 -0.5803597      0      1
## [6,] 0.9190227 -0.4077401      0      1 0.9799849 -0.4348379      0      1
##         [,137]    [,138] [,139] [,140]    [,141]     [,142] [,143] [,144]
## [1,] 1.0160331 -1.648671      0      1 0.8744438 -1.1111105      0      1
## [2,] 1.0243889 -1.726666      0      1 0.9166478 -0.8049003      0      1
## [3,] 1.1407933 -1.606800      0      1 0.8359236 -0.9853514      0      1
## [4,] 0.8215601 -1.619735      0      1 0.9621530 -0.8342080      0      1
## [5,] 1.0229215 -1.721685      0      1 0.6826022 -0.8909038      0      1
## [6,] 1.2761779 -1.480587      0      1 0.8593472 -0.7772513      0      1
##         [,145]    [,146] [,147] [,148]    [,149]    [,150] [,151] [,152]
## [1,] 1.1391513 -2.513857      0      1 1.5181009 -1.633439      0      1
## [2,] 1.2106795 -2.316103      0      1 1.0487574 -1.284057      0      1
## [3,] 0.8929120 -2.153402      0      1 1.1283108 -1.577897      0      1
## [4,] 1.0434521 -2.403367      0      1 1.1222194 -1.330144      0      1
## [5,] 0.8817182 -2.149930      0      1 0.9386559 -1.398401      0      1
## [6,] 0.8455696 -1.912390      0      1 0.9871854 -1.292153      0      1
##         [,153]      [,154] [,155] [,156]    [,157]     [,158] [,159] [,160]
## [1,] 0.7196351 -0.05354159      0      1 0.9701218 -0.2448963      0      1
## [2,] 0.9665343 -0.08818645      0      1 1.1486980 -0.5065770      0      1
## [3,] 0.8832648 -0.15459286      0      1 0.9509175 -0.3223980      0      1
## [4,] 0.8267997 -0.17920663      0      1 0.8274109 -0.3634138      0      1
## [5,] 0.9900635 -0.28211823      0      1 1.0302332 -0.5655320      0      1
## [6,] 0.9926564 -0.18880467      0      1 0.9232218 -0.2247155      0      1
##      [,161] [,162]    [,163]     [,164] [,165] [,166]    [,167]    [,168]
## [1,]      0      1 1.0103163 -1.1626794      0      1 1.0574152 0.3014738
## [2,]      0      1 1.2785683 -1.1979903      0      1 1.1268941 0.3522184
## [3,]      0      1 1.1545179 -1.0676579      0      1 1.0265333 0.4469994
## [4,]      0      1 0.9767176 -1.1982696      0      1 1.0911564 0.4433458
## [5,]      0      1 1.2498596 -1.2419046      0      1 1.0499027 0.2981972
## [6,]      0      1 1.0232032 -0.9749986      0      1 0.9714587 0.5951520
##      [,169] [,170]   [,171]   [,172] [,173] [,174]   [,175]    [,176] [,177]
## [1,]      0      1 1.166507 1.084819      0      1 1.194127 -2.076630      0
## [2,]      0      1 1.185616 1.090233      0      1 1.193952 -2.425191      0
## [3,]      0      1 1.223583 1.275322      0      1 1.147366 -2.250580      0
## [4,]      0      1 1.135087 1.087048      0      1 1.114389 -2.059807      0
## [5,]      0      1 1.239663 1.078973      0      1 1.612269 -2.573731      0
## [6,]      0      1 1.289774 1.316082      0      1 1.105384 -2.043871      0
##      [,178]    [,179]    [,180] [,181] [,182]    [,183]    [,184] [,185] [,186]
## [1,]      1 0.9220800 0.6038229      0      1 1.0377643 0.5183385      0      1
## [2,]      1 0.8667636 0.5329941      0      1 0.8466209 0.2699329      0      1
## [3,]      1 0.9132967 0.6383999      0      1 0.9421524 0.4521924      0      1
## [4,]      1 0.7351379 0.6555570      0      1 0.7046196 0.5282415      0      1
## [5,]      1 0.8743595 0.4251740      0      1 0.8271896 0.4032405      0      1
## [6,]      1 0.8297941 0.6649090      0      1 0.8560267 0.5125017      0      1
##         [,187]     [,188] [,189] [,190]    [,191]      [,192] [,193] [,194]
## [1,] 1.3248095 -0.2723422      0      1 1.2674431 -0.31883117      0      1
## [2,] 1.1301829 -0.5259065      0      1 0.8665950 -0.44700108      0      1
## [3,] 1.1737353 -0.5295596      0      1 0.9147615 -0.24197065      0      1
## [4,] 0.8231088 -0.5573376      0      1 0.8300713 -0.40783455      0      1
## [5,] 1.0880311 -0.5611380      0      1 0.8074075 -0.35464554      0      1
## [6,] 1.2539066 -0.4520503      0      1 0.8763732 -0.07250134      0      1
##         [,195]     [,196] [,197] [,198]    [,199]     [,200] [,201] [,202]
## [1,] 0.8601484 -0.4734450      0      1 1.2213748 -0.7780836      0      1
## [2,] 0.7971635 -0.5677240      0      1 1.0202127 -0.9662832      0      1
## [3,] 0.8859438 -0.4759356      0      1 0.9692751 -0.5380371      0      1
## [4,] 0.7715596 -0.4509772      0      1 0.8420804 -0.6439326      0      1
## [5,] 0.6473789 -0.4192774      0      1 1.2008947 -0.7755910      0      1
## [6,] 1.0088255 -0.4661847      0      1 1.0188623 -0.6395350      0      1
##         [,203]     [,204] [,205] [,206]    [,207]     [,208] [,209] [,210]
## [1,] 1.0484395 -0.3877228      0      1 1.3299263 -0.9198142      0      1
## [2,] 1.2024183 -0.6439990      0      1 0.9354612 -1.0631734      0      1
## [3,] 1.1836744 -0.5101770      0      1 1.3201245 -0.6586225      0      1
## [4,] 0.9332238 -0.4220321      0      1 0.9671346 -0.7430915      0      1
## [5,] 1.1624484 -0.7411128      0      1 1.0813659 -0.8804587      0      1
## [6,] 1.2838292 -0.3634252      0      1 1.3074097 -0.6789278      0      1
##         [,211]     [,212] [,213] [,214]    [,215]     [,216] [,217] [,218]
## [1,] 1.0204270 -0.8487723      0      1 1.4316651 0.26360835      0      1
## [2,] 0.9312207 -0.6815240      0      1 0.8537644 0.01289777      0      1
## [3,] 0.9454534 -0.7179334      0      1 1.2520907 0.44207797      0      1
## [4,] 0.7995777 -0.4955038      0      1 1.0929203 0.32815338      0      1
## [5,] 0.7427395 -0.8531450      0      1 0.8822173 0.06342788      0      1
## [6,] 0.7403708 -0.5282772      0      1 1.0227594 0.29302623      0      1
##        [,219]   [,220] [,221] [,222]    [,223]      [,224] [,225] [,226]
## [1,] 1.370266 1.039543      0      1 1.2544502 -0.09066915      0      1
## [2,] 1.172820 1.024917      0      1 1.0680640 -0.34981106      0      1
## [3,] 1.233975 1.260408      0      1 0.9367682 -0.14098385      0      1
## [4,] 1.181572 1.042293      0      1 1.0976328 -0.12723222      0      1
## [5,] 1.223149 1.119763      0      1 0.9036701 -0.09074331      0      1
## [6,] 1.051230 1.263095      0      1 0.7374434  0.11779866      0      1
##        [,227]     [,228] [,229] [,230]    [,231]     [,232] [,233] [,234]
## [1,] 1.332815 -0.1782603      0      1 0.7337923 -0.6527244      0      1
## [2,] 1.156580 -0.2673808      0      1 0.7981640 -0.8217517      0      1
## [3,] 1.237110 -0.2239011      0      1 0.8219741 -0.7386222      0      1
## [4,] 1.131655 -0.1573622      0      1 0.7734508 -0.6096068      0      1
## [5,] 1.198981 -0.2760732      0      1 1.0273908 -0.7275090      0      1
## [6,] 1.154124 -0.1838478      0      1 0.9373237 -0.6715142      0      1
##         [,235]     [,236] [,237] [,238]    [,239]   [,240] [,241] [,242]
## [1,] 1.0370028 -0.5163893      0      1 1.2155243 2.653787      0      1
## [2,] 1.0568255 -0.8169578      0      1 0.8470493 2.334884      0      1
## [3,] 0.7974233 -0.4410974      0      1 1.2073891 2.785959      0      1
## [4,] 1.0298159 -0.6814080      0      1 0.8168776 2.224117      0      1
## [5,] 1.2165558 -0.5710534      0      1 1.1217068 2.219793      0      1
## [6,] 0.7988188 -0.4457367      0      1 0.9845684 2.544195      0      1
##         [,243]    [,244] [,245] [,246]   [,247]      [,248] [,249] [,250]
## [1,] 0.9290075 0.1645451      0      1 1.466858 -0.26508355      0      1
## [2,] 1.0750437 0.2950070      0      1 1.251991 -0.42734622      0      1
## [3,] 0.8963452 0.2260039      0      1 1.183988 -0.21176115      0      1
## [4,] 1.0040446 0.2627031      0      1 1.172948 -0.20960761      0      1
## [5,] 0.8949451 0.2145739      0      1 1.332218 -0.38126595      0      1
## [6,] 0.9214727 0.1359430      0      1 1.472388  0.07401777      0      1
##         [,251]     [,252] [,253] [,254]   [,255]    [,256] [,257] [,258]
## [1,] 1.1372950 -0.2893520      0      1 1.392285 1.0037440      0      1
## [2,] 1.0646027 -0.4813209      0      1 1.269597 0.8683177      0      1
## [3,] 1.1468099 -0.6411869      0      1 1.426072 1.0787682      0      1
## [4,] 0.7916215 -0.2454461      0      1 1.002984 0.9925174      0      1
## [5,] 0.9982918 -0.6602266      0      1 1.192494 0.6962642      0      1
## [6,] 1.1414140 -0.2465584      0      1 1.424394 1.3153260      0      1
##         [,259]     [,260] [,261] [,262]    [,263]    [,264] [,265] [,266]
## [1,] 1.3437666 -0.3411875      0      1 1.0279704 -1.300336      0      1
## [2,] 0.8230952 -0.3812643      0      1 1.1056653 -1.488101      0      1
## [3,] 1.2188444 -0.2562332      0      1 0.9656061 -1.310899      0      1
## [4,] 1.0918589 -0.3122069      0      1 1.0519957 -1.226295      0      1
## [5,] 1.0058661 -0.4147772      0      1 0.8472389 -1.526059      0      1
## [6,] 1.1460696 -0.2658641      0      1 0.9564560 -1.244137      0      1
##        [,267]    [,268] [,269] [,270]    [,271]     [,272] [,273] [,274]
## [1,] 1.210815 0.7944739      0      1 1.0867546 -0.7785317      0      1
## [2,] 1.383520 0.5719057      0      1 1.2020481 -1.0478782      0      1
## [3,] 1.185317 0.6038661      0      1 0.9905464 -0.8310819      0      1
## [4,] 1.015116 0.7239486      0      1 0.8126954 -0.6249734      0      1
## [5,] 1.150246 0.6455967      0      1 0.8006963 -1.0618577      0      1
## [6,] 1.163512 0.8988207      0      1 0.8399223 -0.6772673      0      1
##        [,275]    [,276] [,277] [,278]    [,279]     [,280] [,281] [,282]
## [1,] 1.373163 0.2190797      0      1 1.0057201 -0.7564578      0      1
## [2,] 1.148344 0.2751480      0      1 0.9502727 -1.0397739      0      1
## [3,] 1.080075 0.3490905      0      1 0.9825233 -0.7417200      0      1
## [4,] 1.350424 0.2429885      0      1 0.8519826 -0.8679425      0      1
## [5,] 1.257718 0.0382402      0      1 0.9637481 -1.0014728      0      1
## [6,] 1.124451 0.1703340      0      1 1.0042486 -0.7915179      0      1
##         [,283]   [,284] [,285] [,286]    [,287]     [,288] [,289] [,290]
## [1,] 1.1049121 1.049772      0      1 1.4106509 -0.4907262      0      1
## [2,] 1.0698655 0.947455      0      1 1.0592160 -0.4002302      0      1
## [3,] 0.9821586 1.220763      0      1 0.9289291 -0.2632935      0      1
## [4,] 0.7250467 1.179390      0      1 1.2507272 -0.4502028      0      1
## [5,] 0.7952042 1.106545      0      1 1.2560516 -0.6229167      0      1
## [6,] 1.0646790 1.212875      0      1 1.1514727 -0.2629068      0      1
##         [,291]     [,292] [,293] [,294]    [,295]     [,296] [,297] [,298]
## [1,] 0.8331861 -0.4926722      0      1 1.3309610 -0.2834998      0      1
## [2,] 0.6963885 -0.6198154      0      1 1.1642215 -0.3382263      0      1
## [3,] 1.2074998 -0.4139151      0      1 1.0099392 -0.1364204      0      1
## [4,] 1.0274623 -0.6038162      0      1 0.8843506 -0.4975451      0      1
## [5,] 0.7695484 -0.5532640      0      1 1.3183484 -0.4470757      0      1
## [6,] 0.8758735 -0.3418717      0      1 1.2444317 -0.2605684      0      1
##         [,299]    [,300] [,301] [,302]    [,303]     [,304] [,305] [,306]
## [1,] 1.1024197 -1.701910      0      1 0.9947516 -1.0405959      0      1
## [2,] 1.1251106 -1.738333      0      1 0.9932373 -1.4409990      0      1
## [3,] 1.0975840 -1.505298      0      1 0.9459809 -0.9309363      0      1
## [4,] 1.0872705 -1.453313      0      1 0.8443252 -1.0883551      0      1
## [5,] 0.9180244 -1.448512      0      1 0.8943656 -1.1185505      0      1
## [6,] 1.2155006 -1.442098      0      1 0.9251260 -1.1649215      0      1
##         [,307]    [,308] [,309] [,310]   [,311]     [,312] [,313] [,314]
## [1,] 0.9525355 -1.816990      0      1 1.386857 -0.9123444      0      1
## [2,] 0.6886306 -1.772611      0      1 1.113177 -1.2041370      0      1
## [3,] 1.0588516 -1.977008      0      1 0.952730 -1.1641035      0      1
## [4,] 1.0204812 -2.182421      0      1 1.025535 -1.0117995      0      1
## [5,] 0.8063633 -2.146444      0      1 1.155036 -1.0791553      0      1
## [6,] 1.1772269 -1.932232      0      1 1.057835 -1.0018068      0      1
##        [,315]      [,316] [,317] [,318]    [,319]      [,320] [,321] [,322]
## [1,] 1.139357 -0.28503566      0      1 1.1682862 -0.26620056      0      1
## [2,] 1.101842 -0.18125795      0      1 0.7645465 -0.24885784      0      1
## [3,] 1.291823 -0.15363844      0      1 0.8767453 -0.09329269      0      1
## [4,] 1.043327 -0.06731312      0      1 0.8752505 -0.22265605      0      1
## [5,] 1.128782 -0.07731377      0      1 0.9694675 -0.49869393      0      1
## [6,] 1.015938  0.02346207      0      1 0.9124050 -0.32852840      0      1
##           [,323]    [,324]
## [1,] -0.10676157 0.6593277
## [2,] -0.09992968 0.9279315
## [3,] -0.25675047 0.7733301
## [4,] -0.23374286 1.0034458
## [5,] -0.16334320 0.8160134
## [6,] -0.39532602 0.8949884

## End(No test)

[Package mirt version 1.38.3 Index]