draw_parameters {mirt}	R Documentation

Draw plausible parameter instantiations from a given model

Description

Draws plausible parameters from a model using parametric sampling (if the information matrix was computed) or via bootstrap sampling. Primarily for use with the DRF function.

Usage

draw_parameters(
  mod,
  draws,
  method = c("parametric", "boostrap"),
  redraws = 20,
  verbose = FALSE,
  ...
)

Arguments

mod
estimated single or multiple-group model

draws
number of draws to obtain

method
type of plausible values to obtain. Can be 'parametric', for the parametric sampling scheme which uses the estimated information matrix, or 'boostrap' to obtain values from the boot function. Default is 'parametric'

redraws
number of redraws to perform when the given parameteric sample does not satisfy the upper and lower parameter bounds. If a valid set cannot be found within this number of draws then an error will be thrown

verbose
logical; include additional information in the console?

...
additional arguments to be passed

Value

returns a draws x p matrix of plausible parameters, where each row correspeonds to a single set

Examples

## No test: 
set.seed(1234)
n <- 40
N <- 500

# only first 5 items as anchors
model <- 'F = 1-40
          CONSTRAINB = (1-5, a1), (1-5, d)'

a <- matrix(1, n)
d <- matrix(rnorm(n), n)
group <- c(rep('Group_1', N), rep('Group_2', N))

## -------------
# groups completely equal
dat1 <- simdata(a, d, N, itemtype = 'dich')
dat2 <- simdata(a, d, N, itemtype = 'dich')
dat <- rbind(dat1, dat2)
mod <- multipleGroup(dat, model, group=group, SE=TRUE,
                     invariance=c('free_means', 'free_var'))

param_set <- draw_parameters(mod, 100)
head(param_set)

##           [,1]      [,2] [,3] [,4]      [,5]      [,6] [,7] [,8]      [,9]
## [1,] 0.9640820 -1.025637    0    1 0.9035820 0.3686735    0    1 1.1159592
## [2,] 1.0242350 -1.135846    0    1 1.2700899 0.3947390    0    1 1.1117053
## [3,] 1.1580895 -1.205314    0    1 0.9781231 0.3636032    0    1 1.1340118
## [4,] 1.1288740 -1.233294    0    1 1.0818873 0.4541240    0    1 1.1190510
## [5,] 0.8840735 -1.156058    0    1 0.9764737 0.2324264    0    1 1.1933744
## [6,] 1.2429795 -1.196233    0    1 1.1965378 0.5038754    0    1 0.9889098
##          [,10] [,11] [,12]    [,13]     [,14] [,15] [,16]     [,17]     [,18]
## [1,] 1.2508606     0     1 1.379817 -2.414790     0     1 0.9951578 0.4610141
## [2,] 1.0071835     0     1 1.126177 -2.089796     0     1 1.0102581 0.5549880
## [3,] 0.9652278     0     1 1.163485 -2.293107     0     1 1.0724707 0.3892196
## [4,] 1.0102605     0     1 1.049364 -2.363929     0     1 0.9512160 0.4136668
## [5,] 1.3091908     0     1 1.093297 -2.331091     0     1 0.6631688 0.4421396
## [6,] 1.0277041     0     1 1.256919 -2.300570     0     1 0.8979338 0.3702290
##      [,19] [,20]    [,21]     [,22] [,23] [,24]     [,25]      [,26] [,27]
## [1,]     0     1 1.203618 0.2216977     0     1 0.9533236 -0.7272384     0
## [2,]     0     1 1.388906 0.4492987     0     1 1.3607317 -0.6005731     0
## [3,]     0     1 1.248170 0.3962940     0     1 1.3885415 -0.6463924     0
## [4,]     0     1 1.131894 0.5159175     0     1 1.2140939 -0.4600616     0
## [5,]     0     1 1.126889 0.2883841     0     1 1.0191816 -0.4869150     0
## [6,]     0     1 1.263055 0.5372564     0     1 1.3420745 -0.3351283     0
##      [,28]     [,29]      [,30] [,31] [,32]     [,33]      [,34] [,35] [,36]
## [1,]     1 0.9412949 -0.6907454     0     1 0.7893777 -0.3469677     0     1
## [2,]     1 1.0437567 -0.5674306     0     1 1.0020387 -0.3908561     0     1
## [3,]     1 1.1660816 -0.7906189     0     1 0.9738665 -0.6216756     0     1
## [4,]     1 1.3369578 -0.6053572     0     1 0.7296202 -0.5198516     0     1
## [5,]     1 1.0611240 -0.5738293     0     1 1.0158407 -0.4732140     0     1
## [6,]     1 1.1263624 -0.4911367     0     1 1.0648714 -0.4911520     0     1
##          [,37]      [,38] [,39] [,40]     [,41]      [,42] [,43] [,44]    [,45]
## [1,] 0.9372372 -0.9548933     0     1 0.8613111 -0.4946877     0     1 1.236333
## [2,] 1.1602643 -0.9131340     0     1 1.1955656 -0.2974996     0     1 1.142936
## [3,] 1.0114586 -0.9333135     0     1 1.4418471 -0.4119431     0     1 1.182460
## [4,] 0.9559303 -0.9832482     0     1 1.1293038 -0.4163494     0     1 1.056209
## [5,] 0.8807150 -0.9650714     0     1 1.1533567 -0.5676485     0     1 1.029748
## [6,] 1.1047616 -0.6823101     0     1 1.1008082 -0.4475463     0     1 1.063712
##           [,46] [,47] [,48]     [,49]      [,50] [,51] [,52]     [,53]
## [1,] -0.9867348     0     1 1.3001988 -0.8350623     0     1 0.7229721
## [2,] -0.9141119     0     1 1.1131731 -0.6513477     0     1 0.7332920
## [3,] -0.9492297     0     1 1.1390632 -0.9430970     0     1 0.9601586
## [4,] -0.9026954     0     1 1.1045350 -0.9518095     0     1 1.0452774
## [5,] -1.0089725     0     1 1.2351153 -0.9491126     0     1 0.8533607
## [6,] -0.8728759     0     1 0.9540588 -0.6575739     0     1 0.7496948
##           [,54] [,55] [,56]    [,57]     [,58] [,59] [,60]     [,61]
## [1,] 0.02734016     0     1 1.132112 1.0329441     0     1 1.2742392
## [2,] 0.28521567     0     1 1.332322 1.0076352     0     1 0.9898161
## [3,] 0.11614600     0     1 1.408233 1.0346187     0     1 1.3281305
## [4,] 0.11898908     0     1 1.435560 1.0590079     0     1 0.7513985
## [5,] 0.04327199     0     1 1.301536 0.9688512     0     1 1.0652394
## [6,] 0.05509203     0     1 1.429922 1.2198459     0     1 1.0541488
##            [,62] [,63] [,64]     [,65]      [,66] [,67] [,68]     [,69]
## [1,] -0.26433052     0     1 1.0543317 -0.4308436     0     1 1.0958833
## [2,] -0.12269842     0     1 1.1504107 -0.4181977     0     1 1.2533809
## [3,] -0.09247390     0     1 1.2318887 -0.3751961     0     1 1.2688432
## [4,] -0.05515095     0     1 1.0203617 -0.3493171     0     1 1.2456992
## [5,] -0.13238425     0     1 0.9384362 -0.4320746     0     1 0.9630178
## [6,] -0.10041446     0     1 1.0005794 -0.2608504     0     1 1.1714312
##           [,70] [,71] [,72]     [,73]      [,74] [,75] [,76]    [,77]    [,78]
## [1,] -0.8523670     0     1 0.9789160 -0.7898106     0     1 1.221431 2.515158
## [2,] -0.9387896     0     1 1.1411133 -0.7178407     0     1 1.630843 3.386610
## [3,] -1.2200040     0     1 0.9693962 -0.7070689     0     1 1.514896 2.941436
## [4,] -1.1449223     0     1 1.1983121 -0.7144918     0     1 1.645717 2.952037
## [5,] -0.9623838     0     1 1.2576630 -0.7602197     0     1 1.927658 3.024744
## [6,] -0.6681998     0     1 1.1762679 -0.7516045     0     1 1.195023 2.899430
##      [,79] [,80]     [,81]       [,82] [,83] [,84]    [,85]      [,86] [,87]
## [1,]     0     1 1.0030878  0.07082546     0     1 1.123499 -0.7474898     0
## [2,]     0     1 1.0266160  0.06649106     0     1 1.217378 -0.5146934     0
## [3,]     0     1 1.0007137 -0.22281361     0     1 1.349017 -0.7568006     0
## [4,]     0     1 1.1605404  0.03966971     0     1 1.274191 -0.6693899     0
## [5,]     0     1 0.9414841  0.05579549     0     1 1.144408 -0.6005355     0
## [6,]     0     1 0.9796483  0.17655801     0     1 1.185003 -0.4393498     0
##      [,88]     [,89]      [,90] [,91] [,92]    [,93]     [,94] [,95] [,96]
## [1,]     1 0.9241730 -0.4712688     0     1 1.066984 0.4397899     0     1
## [2,]     1 1.0226690 -0.4599249     0     1 1.044112 0.4670644     0     1
## [3,]     1 0.7636081 -0.5107728     0     1 1.249279 0.3614518     0     1
## [4,]     1 0.9804201 -0.5643643     0     1 1.225210 0.4643570     0     1
## [5,]     1 0.6916716 -0.4412922     0     1 1.005777 0.3376112     0     1
## [6,]     1 1.0335009 -0.3878417     0     1 1.016061 0.5844049     0     1
##         [,97]      [,98] [,99] [,100]   [,101]    [,102] [,103] [,104]   [,105]
## [1,] 1.191629 -0.7714944     0      1 1.453373 -1.769192      0      1 1.174712
## [2,] 1.129955 -0.5695183     0      1 1.533100 -1.508365      0      1 1.532497
## [3,] 1.077325 -0.7643455     0      1 1.640699 -1.677228      0      1 1.314209
## [4,] 1.056101 -0.7309636     0      1 1.550411 -1.543866      0      1 1.651720
## [5,] 1.133609 -0.5738048     0      1 1.673370 -1.631587      0      1 1.389753
## [6,] 1.207281 -0.5685394     0      1 1.262499 -1.204286      0      1 1.122712
##         [,106] [,107] [,108]    [,109]     [,110] [,111] [,112]   [,113]
## [1,] 0.7375485      0      1 0.9737767 -0.9327875      0      1 1.426009
## [2,] 0.6027456      0      1 0.9595318 -0.8305865      0      1 1.261880
## [3,] 0.7250997      0      1 1.0247454 -0.9836496      0      1 1.458761
## [4,] 0.7386333      0      1 1.4921882 -1.0664814      0      1 1.179686
## [5,] 0.6853798      0      1 1.2513214 -0.9880656      0      1 1.223799
## [6,] 0.7859680      0      1 1.1158904 -0.9369780      0      1 1.389671
##           [,114] [,115] [,116]    [,117]     [,118] [,119] [,120]    [,121]
## [1,] -0.14431702      0      1 0.9038358 -0.9280250      0      1 0.9005223
## [2,] -0.05963545      0      1 0.9582369 -0.8565801      0      1 1.1481746
## [3,] -0.02543058      0      1 1.1880893 -0.9503505      0      1 1.1050598
## [4,] -0.09414556      0      1 1.0186286 -1.0450420      0      1 1.3263464
## [5,] -0.13782922      0      1 0.7676872 -0.9476572      0      1 1.1272219
## [6,]  0.09806364      0      1 0.8439073 -0.8227874      0      1 1.0735746
##        [,122] [,123] [,124]    [,125]      [,126] [,127] [,128]    [,129]
## [1,] 1.205065      0      1 0.9598434 -0.34691251      0      1 1.1013446
## [2,] 1.283239      0      1 0.8831997 -0.20594633      0      1 0.9322403
## [3,] 1.160201      0      1 0.9639189 -0.38672934      0      1 1.2948465
## [4,] 1.305506      0      1 1.1022243 -0.09849098      0      1 0.9876400
## [5,] 1.363875      0      1 1.0158947 -0.25992888      0      1 0.9627434
## [6,] 1.234022      0      1 0.9286938 -0.21745213      0      1 1.0414980
##          [,130] [,131] [,132]    [,133]     [,134] [,135] [,136]    [,137]
## [1,] -0.6759261      0      1 0.6397035 -0.5195479      0      1 0.9335323
## [2,] -0.5935043      0      1 0.9956739 -0.6354291      0      1 0.9635602
## [3,] -0.7739484      0      1 1.0811403 -0.5823955      0      1 0.9257000
## [4,] -0.6312913      0      1 1.1240129 -0.4856349      0      1 1.0398865
## [5,] -0.5475188      0      1 0.8270139 -0.5105169      0      1 1.0616191
## [6,] -0.5021141      0      1 0.7519872 -0.4442859      0      1 1.1086191
##         [,138] [,139] [,140]    [,141]     [,142] [,143] [,144]    [,145]
## [1,] -1.521714      0      1 0.8347637 -1.0698743      0      1 0.8588911
## [2,] -1.346388      0      1 0.7650066 -0.9557782      0      1 0.7923117
## [3,] -1.523880      0      1 0.9191801 -0.9723367      0      1 0.8813164
## [4,] -1.684048      0      1 0.9417339 -1.0712104      0      1 1.0200572
## [5,] -1.719817      0      1 0.8310186 -1.0219446      0      1 0.9776516
## [6,] -1.527556      0      1 1.0222811 -1.0003943      0      1 0.8876101
##         [,146] [,147] [,148]    [,149]    [,150] [,151] [,152]    [,153]
## [1,] -2.380606      0      1 1.1368156 -1.491694      0      1 1.0031221
## [2,] -2.167589      0      1 1.2887305 -1.361612      0      1 1.0332442
## [3,] -2.063662      0      1 1.1970118 -1.293184      0      1 0.9726752
## [4,] -2.248993      0      1 0.9702449 -1.124915      0      1 0.8943404
## [5,] -2.308609      0      1 1.1466881 -1.370600      0      1 1.0198570
## [6,] -2.321472      0      1 1.2185160 -1.451182      0      1 0.9941864
##           [,154] [,155] [,156]    [,157]     [,158] [,159] [,160] [,161] [,162]
## [1,] -0.05908923      0      1 0.8973526 -0.4241702      0      1      0      1
## [2,] -0.00646842      0      1 1.2607477 -0.3063825      0      1      0      1
## [3,] -0.31511762      0      1 0.8232387 -0.3226461      0      1      0      1
## [4,] -0.25146291      0      1 1.0460258 -0.3989709      0      1      0      1
## [5,] -0.21369942      0      1 1.2345503 -0.3949182      0      1      0      1
## [6,] -0.19987016      0      1 1.1462810 -0.2100465      0      1      0      1
##         [,163]    [,164] [,165] [,166]    [,167]    [,168] [,169] [,170]
## [1,] 0.9640820 -1.025637      0      1 0.9035820 0.3686735      0      1
## [2,] 1.0242350 -1.135846      0      1 1.2700899 0.3947390      0      1
## [3,] 1.1580895 -1.205314      0      1 0.9781231 0.3636032      0      1
## [4,] 1.1288740 -1.233294      0      1 1.0818873 0.4541240      0      1
## [5,] 0.8840735 -1.156058      0      1 0.9764737 0.2324264      0      1
## [6,] 1.2429795 -1.196233      0      1 1.1965378 0.5038754      0      1
##         [,171]    [,172] [,173] [,174]   [,175]    [,176] [,177] [,178]
## [1,] 1.1159592 1.2508606      0      1 1.379817 -2.414790      0      1
## [2,] 1.1117053 1.0071835      0      1 1.126177 -2.089796      0      1
## [3,] 1.1340118 0.9652278      0      1 1.163485 -2.293107      0      1
## [4,] 1.1190510 1.0102605      0      1 1.049364 -2.363929      0      1
## [5,] 1.1933744 1.3091908      0      1 1.093297 -2.331091      0      1
## [6,] 0.9889098 1.0277041      0      1 1.256919 -2.300570      0      1
##         [,179]    [,180] [,181] [,182]    [,183]    [,184] [,185] [,186]
## [1,] 0.9951578 0.4610141      0      1 1.1930635 0.5277522      0      1
## [2,] 1.0102581 0.5549880      0      1 1.1982221 0.7659801      0      1
## [3,] 1.0724707 0.3892196      0      1 0.8926599 0.4096380      0      1
## [4,] 0.9512160 0.4136668      0      1 0.8578103 0.3349501      0      1
## [5,] 0.6631688 0.4421396      0      1 0.7136576 0.5061333      0      1
## [6,] 0.8979338 0.3702290      0      1 1.0880966 0.6876548      0      1
##         [,187]     [,188] [,189] [,190]    [,191]     [,192] [,193] [,194]
## [1,] 1.1017809 -0.3619835      0      1 0.9061297 -0.5318088      0      1
## [2,] 0.9628829 -0.5081142      0      1 1.1746948 -0.2045436      0      1
## [3,] 0.9353082 -0.4141985      0      1 0.9196448 -0.5435638      0      1
## [4,] 0.7867718 -0.6288581      0      1 0.9825006 -0.3410107      0      1
## [5,] 0.9355397 -0.4448303      0      1 0.8951825 -0.3232816      0      1
## [6,] 1.1244082 -0.5254866      0      1 1.0162109 -0.4505999      0      1
##         [,195]     [,196] [,197] [,198]    [,199]     [,200] [,201] [,202]
## [1,] 0.8315795 -0.5394027      0      1 1.1426322 -0.7964215      0      1
## [2,] 0.9920662 -0.3632719      0      1 0.8663700 -0.6691952      0      1
## [3,] 0.8338684 -0.5995290      0      1 1.0416129 -0.8922108      0      1
## [4,] 0.8730904 -0.4728145      0      1 0.7097182 -0.8521600      0      1
## [5,] 0.7670570 -0.7083480      0      1 0.7579409 -0.7017551      0      1
## [6,] 0.8758857 -0.5683720      0      1 0.9308215 -0.7732169      0      1
##         [,203]     [,204] [,205] [,206]    [,207]     [,208] [,209] [,210]
## [1,] 1.1019109 -0.5555946      0      1 1.0597258 -0.8875151      0      1
## [2,] 1.0730917 -0.3969341      0      1 1.0861836 -0.8934823      0      1
## [3,] 0.8605734 -0.6344528      0      1 1.0478152 -0.9948193      0      1
## [4,] 1.0872577 -0.7701269      0      1 0.9201989 -0.9587004      0      1
## [5,] 1.1079393 -0.5544304      0      1 1.2803035 -1.1571525      0      1
## [6,] 1.2875676 -0.4189356      0      1 1.1615686 -0.9127437      0      1
##         [,211]     [,212] [,213] [,214]    [,215]    [,216] [,217] [,218]
## [1,] 1.0889695 -0.6295935      0      1 1.0565694 0.2227125      0      1
## [2,] 1.0878801 -0.7061697      0      1 1.0469141 0.2931250      0      1
## [3,] 0.7083124 -0.6524373      0      1 0.8999165 0.2722010      0      1
## [4,] 1.1094324 -0.8191550      0      1 0.9666590 0.1805347      0      1
## [5,] 0.9216742 -0.6645401      0      1 0.8750560 0.2211325      0      1
## [6,] 0.9172812 -0.5613998      0      1 1.1060938 0.1238948      0      1
##        [,219]    [,220] [,221] [,222]    [,223]      [,224] [,225] [,226]
## [1,] 1.137790 1.0535952      0      1 0.8693757 -0.04370171      0      1
## [2,] 1.067210 1.2398593      0      1 1.1465834 -0.10773011      0      1
## [3,] 1.035269 0.9190977      0      1 1.1968453 -0.20560713      0      1
## [4,] 1.072794 1.1260293      0      1 0.9140920 -0.22826380      0      1
## [5,] 1.079450 1.1925905      0      1 0.9281398  0.03167921      0      1
## [6,] 1.285726 1.2174669      0      1 1.0041141 -0.16726772      0      1
##         [,227]     [,228] [,229] [,230]    [,231]     [,232] [,233] [,234]
## [1,] 0.8674024 -0.2388714      0      1 0.7773262 -0.8129991      0      1
## [2,] 1.2670352 -0.2221592      0      1 0.8590745 -0.7091874      0      1
## [3,] 0.9055021 -0.4173555      0      1 0.7929374 -1.0482493      0      1
## [4,] 1.0399245 -0.2700501      0      1 0.7714374 -0.9140900      0      1
## [5,] 1.1785379 -0.1692469      0      1 0.5070811 -0.8061039      0      1
## [6,] 1.1895058 -0.3486456      0      1 0.8606189 -0.4953472      0      1
##         [,235]     [,236] [,237] [,238]    [,239]   [,240] [,241] [,242]
## [1,] 1.0042797 -0.6947812      0      1 1.2768146 2.842036      0      1
## [2,] 1.0175650 -0.6209713      0      1 1.0933488 2.743900      0      1
## [3,] 0.7303296 -0.6180578      0      1 0.6389511 2.228030      0      1
## [4,] 0.8170712 -0.7424675      0      1 0.9300473 2.307153      0      1
## [5,] 0.9520234 -0.4522129      0      1 1.0931551 2.599107      0      1
## [6,] 0.9522944 -0.6977222      0      1 1.4159138 2.676266      0      1
##         [,243]     [,244] [,245] [,246]   [,247]     [,248] [,249] [,250]
## [1,] 0.6194490 0.21001338      0      1 1.212713 -0.1902365      0      1
## [2,] 0.7238849 0.19743370      0      1 1.461054 -0.2810847      0      1
## [3,] 0.8163372 0.12829207      0      1 1.156086 -0.5092694      0      1
## [4,] 0.8071731 0.11219728      0      1 1.143764 -0.3392114      0      1
## [5,] 0.8397965 0.07589122      0      1 1.308633 -0.0617330      0      1
## [6,] 0.8937350 0.21775757      0      1 1.594306 -0.2386906      0      1
##         [,251]     [,252] [,253] [,254]   [,255]    [,256] [,257] [,258]
## [1,] 1.2798914 -0.3216938      0      1 1.314921 1.0942167      0      1
## [2,] 0.8578823 -0.2548018      0      1 1.186156 1.0951921      0      1
## [3,] 0.9683975 -0.5494830      0      1 1.139303 0.7716030      0      1
## [4,] 0.7603760 -0.4634884      0      1 1.276301 0.7069885      0      1
## [5,] 0.8767558 -0.1609168      0      1 1.156299 1.1650336      0      1
## [6,] 0.8430840 -0.5420783      0      1 1.452714 0.9773699      0      1
##         [,259]     [,260] [,261] [,262]    [,263]    [,264] [,265] [,266]
## [1,] 1.1664641 -0.3802114      0      1 1.0455443 -1.504635      0      1
## [2,] 1.0414560 -0.5786445      0      1 0.8210235 -1.235932      0      1
## [3,] 1.1149775 -0.6548354      0      1 0.9879845 -1.576589      0      1
## [4,] 0.7952773 -0.5304220      0      1 0.9519388 -1.415213      0      1
## [5,] 0.9490428 -0.6213047      0      1 1.0237461 -1.259001      0      1
## [6,] 1.1306525 -0.1539888      0      1 1.1363055 -1.349354      0      1
##        [,267]    [,268] [,269] [,270]    [,271]     [,272] [,273] [,274]
## [1,] 1.048397 0.5898919      0      1 0.9698498 -0.9436826      0      1
## [2,] 1.121739 0.6728384      0      1 1.0440722 -0.7059202      0      1
## [3,] 1.078931 0.6279446      0      1 0.8228873 -0.8477981      0      1
## [4,] 1.125783 0.6709730      0      1 0.8114809 -0.9342755      0      1
## [5,] 1.088502 0.5543837      0      1 0.8484437 -0.8589147      0      1
## [6,] 1.315544 0.7194345      0      1 1.1462672 -0.8353167      0      1
##         [,275]      [,276] [,277] [,278]    [,279]     [,280] [,281] [,282]
## [1,] 1.2370288  0.28424046      0      1 1.0888182 -0.7805259      0      1
## [2,] 1.1792417  0.33586006      0      1 1.1148860 -0.7424721      0      1
## [3,] 0.9789382  0.22683333      0      1 0.9981865 -1.0482701      0      1
## [4,] 1.0799060 -0.03425254      0      1 1.0565927 -1.1098109      0      1
## [5,] 1.2404790  0.25953970      0      1 1.0902362 -0.9663640      0      1
## [6,] 1.1555241  0.14197584      0      1 0.9511492 -0.8746545      0      1
##         [,283]    [,284] [,285] [,286]    [,287]     [,288] [,289] [,290]
## [1,] 1.1574729 1.2970961      0      1 1.0667046 -0.3765664      0      1
## [2,] 0.9531319 1.1676529      0      1 1.1338697 -0.3841983      0      1
## [3,] 0.9303550 1.1243552      0      1 1.0937180 -0.4753308      0      1
## [4,] 0.4363978 0.7217445      0      1 1.2686669 -0.6017758      0      1
## [5,] 1.0856156 1.2739324      0      1 0.9127929 -0.3962170      0      1
## [6,] 0.7387633 0.9500337      0      1 1.2224968 -0.3276492      0      1
##         [,291]     [,292] [,293] [,294]    [,295]     [,296] [,297] [,298]
## [1,] 1.0824128 -0.5076601      0      1 1.0603385 -0.3737904      0      1
## [2,] 0.6463200 -0.4928935      0      1 0.9984283 -0.2861232      0      1
## [3,] 0.9587616 -0.6209370      0      1 1.3175624 -0.1168393      0      1
## [4,] 0.6212316 -0.6199111      0      1 1.1672517 -0.4626570      0      1
## [5,] 0.9528445 -0.7025005      0      1 0.9990086 -0.4166902      0      1
## [6,] 1.1280788 -0.6718277      0      1 1.2017636 -0.2594483      0      1
##         [,299]    [,300] [,301] [,302]    [,303]    [,304] [,305] [,306]
## [1,] 1.2067865 -1.777704      0      1 1.2345047 -1.351956      0      1
## [2,] 0.8153211 -1.583353      0      1 0.7457487 -1.051984      0      1
## [3,] 0.8555034 -1.742074      0      1 0.8939374 -1.185763      0      1
## [4,] 1.0554229 -1.954610      0      1 0.8990015 -1.056701      0      1
## [5,] 0.8651660 -1.666483      0      1 0.6979789 -1.033004      0      1
## [6,] 1.2040942 -1.424582      0      1 0.8308173 -1.197354      0      1
##         [,307]    [,308] [,309] [,310]    [,311]    [,312] [,313] [,314]
## [1,] 0.8391028 -1.811995      0      1 0.8740229 -1.123906      0      1
## [2,] 1.3726334 -2.369603      0      1 1.0225807 -1.020680      0      1
## [3,] 1.0216983 -2.325158      0      1 1.0151678 -1.104036      0      1
## [4,] 1.2024807 -2.481399      0      1 1.1316883 -1.133383      0      1
## [5,] 0.9300691 -2.043271      0      1 0.9120177 -1.096849      0      1
## [6,] 0.8825337 -2.058409      0      1 1.0195947 -1.099822      0      1
##         [,315]       [,316] [,317] [,318]    [,319]     [,320] [,321] [,322]
## [1,] 0.7992762 -0.164843353      0      1 1.0738564 -0.2394396      0      1
## [2,] 0.8560136  0.009138534      0      1 1.0582012 -0.1327791      0      1
## [3,] 1.1668013 -0.225926648      0      1 0.8494283 -0.3781073      0      1
## [4,] 0.7900661 -0.257345675      0      1 0.6997451 -0.3660770      0      1
## [5,] 1.0934392  0.049617948      0      1 0.7606814 -0.3320898      0      1
## [6,] 0.9176158 -0.067660865      0      1 0.8912570 -0.2777957      0      1
##           [,323]    [,324]
## [1,] -0.14104452 1.0052717
## [2,] -0.21643224 0.9604525
## [3,] -0.08208775 1.0559445
## [4,] -0.11682873 1.1518712
## [5,] -0.17065589 1.0638835
## [6,] -0.15495275 0.7375640

## End(No test)

[Package mirt version 1.40 Index]