Draw plausible parameter instantiations from a given model

Draws plausible parameters from a model using parametric sampling (if the information matrix was computed) or via bootstrap sampling. Primarily for use with the DRF function.

Usage

draw_parameters(
  mod,
  draws,
  method = c("parametric", "boostrap"),
  redraws = 20,
  verbose = FALSE,
  ...
)

Arguments

mod

estimated single or multiple-group model

draws

number of draws to obtain

method

type of plausible values to obtain. Can be 'parametric', for the parametric sampling scheme which uses the estimated information matrix, or 'boostrap' to obtain values from the boot function. Default is 'parametric'

redraws

number of redraws to perform when the given parameteric sample does not satisfy the upper and lower parameter bounds. If a valid set cannot be found within this number of draws then an error will be thrown

verbose

logical; include additional information in the console?

...

additional arguments to be passed

Value

returns a draws x p matrix of plausible parameters, where each row correspeonds to a single set

Examples

# \donttest{
set.seed(1234)
n <- 40
N <- 500

# only first 5 items as anchors
model <- 'F = 1-40
          CONSTRAINB = (1-5, a1), (1-5, d)'

a <- matrix(1, n)
d <- matrix(rnorm(n), n)
group <- c(rep('Group_1', N), rep('Group_2', N))

## -------------
# groups completely equal
dat1 <- simdata(a, d, N, itemtype = 'dich')
dat2 <- simdata(a, d, N, itemtype = 'dich')
dat <- rbind(dat1, dat2)
mod <- multipleGroup(dat, model, group=group, SE=TRUE,
                     invariance=c('free_means', 'free_var'))
#> 
#> 
#> Calculating information matrix...

param_set <- draw_parameters(mod, 100)
head(param_set)
#>           [,1]      [,2] [,3] [,4]      [,5]      [,6] [,7] [,8]     [,9]
#> [1,] 1.1391483 -1.262734    0    1 1.0429570 0.4288435    0    1 1.426608
#> [2,] 1.0484967 -1.270230    0    1 1.2642261 0.4139742    0    1 1.055071
#> [3,] 0.9015559 -1.120842    0    1 1.1142861 0.2913104    0    1 1.179315
#> [4,] 1.2083665 -1.252647    0    1 1.2302625 0.3839666    0    1 1.282218
#> [5,] 1.0422827 -1.129914    0    1 1.0267393 0.4106227    0    1 1.109890
#> [6,] 0.9569286 -1.073822    0    1 0.8923384 0.5172578    0    1 1.122109
#>          [,10] [,11] [,12]     [,13]     [,14] [,15] [,16]     [,17]     [,18]
#> [1,] 1.3571747     0     1 1.3228956 -2.259732     0     1 0.7230044 0.4307506
#> [2,] 0.9838252     0     1 1.1902988 -2.265339     0     1 0.7787847 0.6617051
#> [3,] 1.1912044     0     1 1.1822386 -2.236731     0     1 0.9837777 0.5259832
#> [4,] 0.9824300     0     1 1.0777324 -2.368948     0     1 0.8211154 0.3825540
#> [5,] 1.0501543     0     1 1.1938541 -2.170218     0     1 0.7517767 0.3950822
#> [6,] 1.3031493     0     1 0.9314969 -2.084266     0     1 0.8255591 0.5518456
#>      [,19] [,20]    [,21]     [,22] [,23] [,24]    [,25]      [,26] [,27] [,28]
#> [1,]     0     1 1.163287 0.4935805     0     1 1.138408 -0.4730967     0     1
#> [2,]     0     1 1.295991 0.2962779     0     1 1.277705 -0.3843515     0     1
#> [3,]     0     1 1.393149 0.4957497     0     1 1.207929 -0.5509638     0     1
#> [4,]     0     1 1.221361 0.2570856     0     1 1.050326 -0.5914620     0     1
#> [5,]     0     1 1.051667 0.1174200     0     1 1.302897 -0.4298341     0     1
#> [6,]     0     1 1.066978 0.3894473     0     1 1.192809 -0.4857318     0     1
#>          [,29]      [,30] [,31] [,32]     [,33]      [,34] [,35] [,36]
#> [1,] 1.3527927 -0.5621654     0     1 0.8924490 -0.5117369     0     1
#> [2,] 1.1797525 -0.5450951     0     1 0.9312760 -0.4114297     0     1
#> [3,] 1.4737991 -0.6303803     0     1 0.8593670 -0.4383557     0     1
#> [4,] 1.0532493 -0.6968834     0     1 1.0448990 -0.4547611     0     1
#> [5,] 0.9259638 -0.5687452     0     1 0.8781231 -0.6390878     0     1
#> [6,] 1.4960325 -0.4478350     0     1 0.8747685 -0.4139600     0     1
#>          [,37]      [,38] [,39] [,40]     [,41]      [,42] [,43] [,44]
#> [1,] 1.0394553 -0.9996527     0     1 1.3385259 -0.4398979     0     1
#> [2,] 0.8843201 -0.9601152     0     1 1.0098380 -0.4068918     0     1
#> [3,] 1.0472870 -0.8472980     0     1 1.0328674 -0.4165800     0     1
#> [4,] 1.0481757 -1.1208763     0     1 1.5198539 -0.3905815     0     1
#> [5,] 0.8323413 -0.7925520     0     1 0.9477638 -0.4690871     0     1
#> [6,] 0.8903164 -0.8060033     0     1 1.1486378 -0.5070993     0     1
#>          [,45]      [,46] [,47] [,48]     [,49]      [,50] [,51] [,52]
#> [1,] 1.0213159 -1.0562492     0     1 1.2856206 -0.9979787     0     1
#> [2,] 1.0863032 -0.9086148     0     1 1.1838084 -0.6952002     0     1
#> [3,] 1.3031167 -1.0691703     0     1 1.0733004 -0.7782127     0     1
#> [4,] 1.2671267 -1.2655045     0     1 1.0755923 -0.8150892     0     1
#> [5,] 0.9003792 -0.8236895     0     1 1.0510657 -0.7829129     0     1
#> [6,] 0.7936887 -0.8023501     0     1 0.8012048 -0.7711545     0     1
#>          [,53]      [,54] [,55] [,56]    [,57]     [,58] [,59] [,60]     [,61]
#> [1,] 0.8176167 0.14318947     0     1 1.300685 1.0884976     0     1 0.8592561
#> [2,] 0.9589934 0.23632198     0     1 1.465501 1.0258817     0     1 1.1314657
#> [3,] 0.8126472 0.21273220     0     1 1.384537 1.0716981     0     1 0.8341391
#> [4,] 0.9141176 0.12657353     0     1 1.376907 0.9620398     0     1 0.9868454
#> [5,] 0.9342273 0.07780975     0     1 1.084790 1.0487968     0     1 0.9410801
#> [6,] 1.0545787 0.01144851     0     1 1.368823 1.3333772     0     1 1.0745862
#>            [,62] [,63] [,64]     [,65]      [,66] [,67] [,68]    [,69]
#> [1,] -0.11398977     0     1 0.8092362 -0.5233638     0     1 1.313814
#> [2,] -0.02999061     0     1 1.0429873 -0.5469506     0     1 1.241366
#> [3,] -0.11669005     0     1 1.2888696 -0.5121943     0     1 1.274844
#> [4,] -0.21986790     0     1 1.1642875 -0.7154811     0     1 1.077224
#> [5,]  0.06666290     0     1 0.9142215 -0.4847582     0     1 1.226571
#> [6,]  0.11548131     0     1 0.8656718 -0.3201650     0     1 1.139974
#>           [,70] [,71] [,72]     [,73]      [,74] [,75] [,76]    [,77]    [,78]
#> [1,] -0.9808331     0     1 1.0888395 -0.7741633     0     1 1.178652 2.567549
#> [2,] -0.9901081     0     1 1.0017605 -0.5739486     0     1 1.713045 3.165311
#> [3,] -1.1133870     0     1 0.8404789 -0.5877353     0     1 1.559142 2.637752
#> [4,] -0.9457014     0     1 0.9973526 -0.5775381     0     1 1.482782 2.682691
#> [5,] -1.0808264     0     1 1.0079996 -0.6276290     0     1 1.150029 2.552863
#> [6,] -0.7694778     0     1 1.3709521 -0.6365566     0     1 1.572618 2.960950
#>      [,79] [,80]     [,81]         [,82] [,83] [,84]     [,85]      [,86] [,87]
#> [1,]     0     1 0.9749305  0.1262429236     0     1 1.2569358 -0.6753280     0
#> [2,]     0     1 0.6378724 -0.0416338706     0     1 1.3503421 -0.6989674     0
#> [3,]     0     1 1.0769431  0.0747536323     0     1 1.1304521 -0.4273371     0
#> [4,]     0     1 0.9466751  0.0001031707     0     1 1.2338620 -0.7305857     0
#> [5,]     0     1 0.9646448  0.0222661239     0     1 0.8675040 -0.5533663     0
#> [6,]     0     1 0.8953228  0.0550655416     0     1 0.9862147 -0.5061780     0
#>      [,88]     [,89]      [,90] [,91] [,92]     [,93]     [,94] [,95] [,96]
#> [1,]     1 0.8943719 -0.3695982     0     1 1.1414398 0.4496614     0     1
#> [2,]     1 0.8856734 -0.3901241     0     1 1.1720906 0.4237425     0     1
#> [3,]     1 0.7230547 -0.4177326     0     1 1.1242046 0.3721096     0     1
#> [4,]     1 0.7502933 -0.4709276     0     1 1.1538624 0.5534905     0     1
#> [5,]     1 0.8277912 -0.5723144     0     1 0.9200630 0.4596039     0     1
#> [6,]     1 0.8063404 -0.2832000     0     1 0.9777813 0.4521504     0     1
#>          [,97]      [,98] [,99] [,100]   [,101]    [,102] [,103] [,104]
#> [1,] 0.9558493 -0.6142083     0      1 1.345436 -1.290430      0      1
#> [2,] 1.1158185 -0.5876543     0      1 1.575489 -1.563281      0      1
#> [3,] 1.2789162 -0.6620510     0      1 1.565974 -1.553429      0      1
#> [4,] 1.0295191 -0.6704559     0      1 1.205470 -1.614799      0      1
#> [5,] 0.9924574 -0.5624854     0      1 1.092117 -1.555365      0      1
#> [6,] 1.0533410 -0.3613626     0      1 1.402935 -1.392786      0      1
#>        [,105]    [,106] [,107] [,108]    [,109]     [,110] [,111] [,112]
#> [1,] 1.371332 0.8212177      0      1 1.0591471 -0.7177690      0      1
#> [2,] 1.387874 0.6305876      0      1 1.2003535 -0.9511692      0      1
#> [3,] 1.236606 0.5940211      0      1 0.9270756 -0.9554089      0      1
#> [4,] 1.229710 0.5414306      0      1 0.8022847 -0.8891033      0      1
#> [5,] 1.279885 0.9269728      0      1 1.0720169 -1.0762113      0      1
#> [6,] 1.553092 1.0841679      0      1 1.3316026 -0.9780309      0      1
#>         [,113]      [,114] [,115] [,116]    [,117]     [,118] [,119] [,120]
#> [1,] 1.3841756 -0.11060699      0      1 0.9359201 -0.9057253      0      1
#> [2,] 1.2721103 -0.12842834      0      1 1.1714803 -0.9517779      0      1
#> [3,] 0.9154547 -0.25247189      0      1 1.2103058 -0.9697586      0      1
#> [4,] 1.2261748 -0.23379251      0      1 1.0271747 -0.7794432      0      1
#> [5,] 1.3606454 -0.11402304      0      1 0.9669171 -0.8006616      0      1
#> [6,] 1.1510466 -0.05004385      0      1 0.8964734 -0.7972526      0      1
#>         [,121]   [,122] [,123] [,124]    [,125]      [,126] [,127] [,128]
#> [1,] 1.0723199 1.231363      0      1 0.9407801 -0.32888706      0      1
#> [2,] 1.0880423 1.285115      0      1 1.0321004 -0.34028929      0      1
#> [3,] 1.3211033 1.525651      0      1 0.9438052 -0.27036514      0      1
#> [4,] 0.7331608 1.299447      0      1 1.0681655 -0.30187607      0      1
#> [5,] 1.1344802 1.469861      0      1 0.7784913 -0.10341687      0      1
#> [6,] 0.9448852 1.203271      0      1 0.8675714 -0.06169187      0      1
#>         [,129]     [,130] [,131] [,132]    [,133]     [,134] [,135] [,136]
#> [1,] 0.8582874 -0.5372013      0      1 1.0455382 -0.5890014      0      1
#> [2,] 1.0358113 -0.5796357      0      1 0.7891645 -0.4892293      0      1
#> [3,] 1.0376528 -0.5784212      0      1 1.0742268 -0.7452434      0      1
#> [4,] 0.9488976 -0.7974830      0      1 0.8301226 -0.6310398      0      1
#> [5,] 0.7512012 -0.7169425      0      1 0.8355660 -0.4990204      0      1
#> [6,] 1.2407201 -0.6411234      0      1 1.0449233 -0.4900304      0      1
#>         [,137]    [,138] [,139] [,140]    [,141]     [,142] [,143] [,144]
#> [1,] 1.2294920 -1.652615      0      1 0.8113578 -1.0591070      0      1
#> [2,] 0.9020178 -1.372767      0      1 0.9551101 -1.1610801      0      1
#> [3,] 1.0222270 -1.510696      0      1 0.8123737 -0.9439134      0      1
#> [4,] 0.7451938 -1.646804      0      1 0.9577529 -1.1057927      0      1
#> [5,] 0.9055046 -1.493831      0      1 0.8405730 -0.9015825      0      1
#> [6,] 1.0238206 -1.606561      0      1 0.5875165 -0.6929282      0      1
#>         [,145]    [,146] [,147] [,148]   [,149]    [,150] [,151] [,152]
#> [1,] 0.9904563 -2.317328      0      1 1.117666 -1.358481      0      1
#> [2,] 1.1253025 -2.307848      0      1 1.051013 -1.437488      0      1
#> [3,] 1.1460402 -2.378569      0      1 1.337640 -1.418852      0      1
#> [4,] 1.2316904 -2.440186      0      1 1.103262 -1.415161      0      1
#> [5,] 1.2664414 -2.624070      0      1 1.225435 -1.484993      0      1
#> [6,] 1.1905030 -2.587987      0      1 1.003452 -1.142934      0      1
#>         [,153]     [,154] [,155] [,156]    [,157]     [,158] [,159] [,160]
#> [1,] 1.0051995 -0.1446553      0      1 1.2271219 -0.2833820      0      1
#> [2,] 0.7331594 -0.1336442      0      1 0.8868186 -0.3848605      0      1
#> [3,] 0.7006688 -0.1968375      0      1 1.1800772 -0.3825698      0      1
#> [4,] 0.7622295 -0.2943384      0      1 1.0171924 -0.2595975      0      1
#> [5,] 0.7589873 -0.3463519      0      1 0.9470745 -0.4106848      0      1
#> [6,] 0.8812607 -0.1017558      0      1 1.1799064 -0.2237318      0      1
#>      [,161] [,162]    [,163]    [,164] [,165] [,166]    [,167]    [,168] [,169]
#> [1,]      0      1 1.1391483 -1.262734      0      1 1.0429570 0.4288435      0
#> [2,]      0      1 1.0484967 -1.270230      0      1 1.2642261 0.4139742      0
#> [3,]      0      1 0.9015559 -1.120842      0      1 1.1142861 0.2913104      0
#> [4,]      0      1 1.2083665 -1.252647      0      1 1.2302625 0.3839666      0
#> [5,]      0      1 1.0422827 -1.129914      0      1 1.0267393 0.4106227      0
#> [6,]      0      1 0.9569286 -1.073822      0      1 0.8923384 0.5172578      0
#>      [,170]   [,171]    [,172] [,173] [,174]    [,175]    [,176] [,177] [,178]
#> [1,]      1 1.426608 1.3571747      0      1 1.3228956 -2.259732      0      1
#> [2,]      1 1.055071 0.9838252      0      1 1.1902988 -2.265339      0      1
#> [3,]      1 1.179315 1.1912044      0      1 1.1822386 -2.236731      0      1
#> [4,]      1 1.282218 0.9824300      0      1 1.0777324 -2.368948      0      1
#> [5,]      1 1.109890 1.0501543      0      1 1.1938541 -2.170218      0      1
#> [6,]      1 1.122109 1.3031493      0      1 0.9314969 -2.084266      0      1
#>         [,179]    [,180] [,181] [,182]    [,183]    [,184] [,185] [,186]
#> [1,] 0.7230044 0.4307506      0      1 0.5769400 0.3785568      0      1
#> [2,] 0.7787847 0.6617051      0      1 0.9951140 0.5035113      0      1
#> [3,] 0.9837777 0.5259832      0      1 0.6653239 0.2314147      0      1
#> [4,] 0.8211154 0.3825540      0      1 0.9269246 0.3332767      0      1
#> [5,] 0.7517767 0.3950822      0      1 0.6941908 0.4616564      0      1
#> [6,] 0.8255591 0.5518456      0      1 0.8832430 0.6587134      0      1
#>         [,187]     [,188] [,189] [,190]    [,191]     [,192] [,193] [,194]
#> [1,] 1.1773545 -0.5980234      0      1 0.8270737 -0.3566087      0      1
#> [2,] 1.2006478 -0.4580348      0      1 1.2312037 -0.4468948      0      1
#> [3,] 0.9915550 -0.4397290      0      1 0.7832430 -0.3698797      0      1
#> [4,] 0.8631297 -0.6454328      0      1 0.9260373 -0.5538133      0      1
#> [5,] 1.1585987 -0.5414463      0      1 0.9432157 -0.4511684      0      1
#> [6,] 1.1344615 -0.1556385      0      1 0.8387850 -0.2676433      0      1
#>         [,195]     [,196] [,197] [,198]    [,199]     [,200] [,201] [,202]
#> [1,] 0.7955699 -0.6940803      0      1 1.1881239 -0.8672510      0      1
#> [2,] 1.0363203 -0.6676152      0      1 0.9175631 -0.8380532      0      1
#> [3,] 0.8495424 -0.4471806      0      1 1.0530127 -0.7772087      0      1
#> [4,] 0.6804526 -0.7378925      0      1 0.9491316 -0.8842658      0      1
#> [5,] 0.8446033 -0.3752518      0      1 0.7418598 -0.4362928      0      1
#> [6,] 0.7291273 -0.3666299      0      1 0.8138697 -0.5491436      0      1
#>         [,203]     [,204] [,205] [,206]    [,207]     [,208] [,209] [,210]
#> [1,] 1.1721489 -0.5048728      0      1 1.4271251 -0.8890863      0      1
#> [2,] 0.9024861 -0.3658105      0      1 1.3371533 -0.7301275      0      1
#> [3,] 0.8207097 -0.6218768      0      1 1.2301967 -0.8838977      0      1
#> [4,] 1.1753180 -0.8479833      0      1 1.0316888 -1.2796689      0      1
#> [5,] 0.9550732 -0.3683144      0      1 1.5088175 -0.9401377      0      1
#> [6,] 0.7787624 -0.1532742      0      1 0.6774223 -0.6721852      0      1
#>         [,211]     [,212] [,213] [,214]    [,215]     [,216] [,217] [,218]
#> [1,] 1.0828472 -0.5861585      0      1 0.9437358 0.09630408      0      1
#> [2,] 0.9179892 -0.6699237      0      1 1.1634037 0.40872859      0      1
#> [3,] 0.8099097 -0.6891387      0      1 0.9437236 0.36898206      0      1
#> [4,] 0.9380615 -0.7708514      0      1 0.9953149 0.24996260      0      1
#> [5,] 0.9065562 -0.5227436      0      1 0.8901240 0.22865150      0      1
#> [6,] 0.8478866 -0.5027821      0      1 0.8493834 0.64925158      0      1
#>        [,219]    [,220] [,221] [,222]    [,223]      [,224] [,225] [,226]
#> [1,] 1.251224 1.2660189      0      1 0.8093795 -0.03781640      0      1
#> [2,] 1.275120 1.1705077      0      1 0.8337834 -0.08649986      0      1
#> [3,] 1.075980 1.0332136      0      1 0.8603016 -0.24082097      0      1
#> [4,] 1.178972 0.8965858      0      1 1.0906790 -0.21602316      0      1
#> [5,] 1.465808 1.4528392      0      1 0.9667611  0.16206703      0      1
#> [6,] 1.003672 1.2810119      0      1 0.9793605  0.08787089      0      1
#>         [,227]     [,228] [,229] [,230]    [,231]     [,232] [,233] [,234]
#> [1,] 1.1241183 -0.2655857      0      1 0.8256868 -0.6485636      0      1
#> [2,] 1.1013506 -0.3204781      0      1 1.0770762 -0.8222196      0      1
#> [3,] 1.2208786 -0.2838600      0      1 0.7337529 -0.7363470      0      1
#> [4,] 1.1255067 -0.5292478      0      1 0.7869537 -1.0085999      0      1
#> [5,] 1.3929457 -0.1996818      0      1 0.7863723 -0.7639843      0      1
#> [6,] 0.9912025 -0.1305252      0      1 0.7348674 -0.7732220      0      1
#>         [,235]     [,236] [,237] [,238]    [,239]   [,240] [,241] [,242]
#> [1,] 1.0720096 -0.4475839      0      1 1.0192282 2.661379      0      1
#> [2,] 0.9165160 -0.5343116      0      1 1.3177852 2.757122      0      1
#> [3,] 0.9368496 -0.6705828      0      1 0.9742986 2.529694      0      1
#> [4,] 0.8131281 -0.5930921      0      1 0.9871176 2.429140      0      1
#> [5,] 1.0643592 -0.4905213      0      1 1.2126925 2.697152      0      1
#> [6,] 0.8418799 -0.4321622      0      1 0.7567293 2.617328      0      1
#>         [,243]      [,244] [,245] [,246]    [,247]      [,248] [,249] [,250]
#> [1,] 0.8355139  0.06780849      0      1 1.4137807 -0.21321010      0      1
#> [2,] 1.1336181  0.25348032      0      1 1.2815150 -0.17706914      0      1
#> [3,] 0.5790901  0.13768769      0      1 1.2638156 -0.27375929      0      1
#> [4,] 0.9006494 -0.10519192      0      1 1.0604866 -0.50714435      0      1
#> [5,] 0.9075233  0.18475994      0      1 1.5191410 -0.14976089      0      1
#> [6,] 0.9369082  0.43920923      0      1 0.9969382  0.06122104      0      1
#>         [,251]     [,252] [,253] [,254]    [,255]    [,256] [,257] [,258]
#> [1,] 1.1006399 -0.2310053      0      1 1.4072880 0.8882977      0      1
#> [2,] 1.0802270 -0.6114761      0      1 1.2684383 0.9621852      0      1
#> [3,] 0.7880503 -0.3299453      0      1 1.1213552 0.9072474      0      1
#> [4,] 0.9676510 -0.5462906      0      1 1.3448628 0.6590025      0      1
#> [5,] 1.2516757 -0.2830300      0      1 1.4669848 1.2250279      0      1
#> [6,] 0.8800138 -0.2295392      0      1 0.9690171 1.0716043      0      1
#>         [,259]     [,260] [,261] [,262]    [,263]    [,264] [,265] [,266]
#> [1,] 1.0834722 -0.2874127      0      1 1.2372506 -1.584907      0      1
#> [2,] 1.0168680 -0.4050232      0      1 0.7296544 -1.262677      0      1
#> [3,] 1.2575131 -0.5186062      0      1 0.8585354 -1.362950      0      1
#> [4,] 0.9251556 -0.7952507      0      1 1.1705181 -1.415852      0      1
#> [5,] 1.0538602 -0.5370181      0      1 0.9277027 -1.212450      0      1
#> [6,] 0.6340691 -0.3535315      0      1 0.8223446 -1.063424      0      1
#>        [,267]    [,268] [,269] [,270]    [,271]     [,272] [,273] [,274]
#> [1,] 1.306097 0.6539537      0      1 0.9254586 -1.0119491      0      1
#> [2,] 1.165129 0.6360237      0      1 0.6041956 -0.8960488      0      1
#> [3,] 1.042479 0.5804395      0      1 0.9719401 -0.9049817      0      1
#> [4,] 1.020745 0.4597267      0      1 0.8854014 -0.8461745      0      1
#> [5,] 1.343508 0.7992399      0      1 1.0889584 -0.7382598      0      1
#> [6,] 0.879896 0.8475393      0      1 0.8833642 -0.9095443      0      1
#>         [,275]     [,276] [,277] [,278]    [,279]     [,280] [,281] [,282]
#> [1,] 1.1094789 0.26995895      0      1 1.2318738 -0.9162557      0      1
#> [2,] 1.1581189 0.22195525      0      1 0.9474658 -0.8550226      0      1
#> [3,] 0.8855246 0.02187095      0      1 1.0470295 -1.0408474      0      1
#> [4,] 1.0652445 0.08395807      0      1 1.0388042 -0.9947872      0      1
#> [5,] 1.2904380 0.37359394      0      1 0.9947444 -0.7320669      0      1
#> [6,] 1.0330133 0.53339806      0      1 1.0023877 -0.6590216      0      1
#>         [,283]   [,284] [,285] [,286]    [,287]     [,288] [,289] [,290]
#> [1,] 1.0127844 1.086732      0      1 1.3016443 -0.5645592      0      1
#> [2,] 0.9325888 1.185042      0      1 1.2797563 -0.3321682      0      1
#> [3,] 0.8565901 1.200676      0      1 1.0858069 -0.5370129      0      1
#> [4,] 0.7655193 1.001382      0      1 0.9557799 -0.6196782      0      1
#> [5,] 0.9698498 1.206204      0      1 1.0974389 -0.3217740      0      1
#> [6,] 0.8846890 1.368720      0      1 0.6796244 -0.4113879      0      1
#>         [,291]     [,292] [,293] [,294]    [,295]      [,296] [,297] [,298]
#> [1,] 0.9665674 -0.4499613      0      1 1.2751161 -0.27242604      0      1
#> [2,] 1.0356792 -0.5721748      0      1 1.1089584 -0.16897865      0      1
#> [3,] 0.9499328 -0.4461424      0      1 1.1750413 -0.44589779      0      1
#> [4,] 0.6212662 -0.6609387      0      1 0.9129298 -0.70055663      0      1
#> [5,] 0.9964689 -0.5005593      0      1 1.1185746 -0.19530607      0      1
#> [6,] 0.7778953 -0.4090593      0      1 1.0454056 -0.08128737      0      1
#>         [,299]    [,300] [,301] [,302]    [,303]    [,304] [,305] [,306]
#> [1,] 0.9457179 -1.694390      0      1 0.8919277 -1.298184      0      1
#> [2,] 1.1671609 -1.568946      0      1 1.1484235 -1.217317      0      1
#> [3,] 0.9104849 -1.588931      0      1 0.7172764 -1.194908      0      1
#> [4,] 0.7951303 -1.910015      0      1 0.9743124 -1.257373      0      1
#> [5,] 0.8852897 -1.377146      0      1 0.9976067 -1.021570      0      1
#> [6,] 0.8919738 -1.526346      0      1 0.7979590 -1.070965      0      1
#>         [,307]    [,308] [,309] [,310]    [,311]     [,312] [,313] [,314]
#> [1,] 0.9776854 -1.917578      0      1 0.7331039 -0.9768044      0      1
#> [2,] 0.9623468 -1.886203      0      1 1.0542581 -1.0651804      0      1
#> [3,] 0.9433850 -2.146880      0      1 1.0355126 -1.1161856      0      1
#> [4,] 1.0764607 -2.357394      0      1 1.0398717 -1.2172835      0      1
#> [5,] 1.0682621 -2.334701      0      1 1.1997906 -1.0727553      0      1
#> [6,] 0.8298646 -1.838798      0      1 1.0184908 -1.0133290      0      1
#>         [,315]        [,316] [,317] [,318]    [,319]      [,320] [,321] [,322]
#> [1,] 1.0091757 -0.2196782520      0      1 1.0090705 -0.25004128      0      1
#> [2,] 1.0190809 -0.0007938555      0      1 0.9135201 -0.29224069      0      1
#> [3,] 1.0794888 -0.1673724693      0      1 0.7342355 -0.35306806      0      1
#> [4,] 0.9524044 -0.4012855790      0      1 0.9772383 -0.27937015      0      1
#> [5,] 1.1289982  0.1365135265      0      1 0.7633150 -0.12458776      0      1
#> [6,] 0.8079603 -0.0062007574      0      1 0.7488032 -0.06740662      0      1
#>           [,323]    [,324]
#> [1,] -0.17452981 0.7971819
#> [2,] -0.23051322 0.8413638
#> [3,] -0.17643483 1.0450858
#> [4,]  0.06443079 1.0789363
#> [5,] -0.27483253 0.8621822
#> [6,] -0.47547745 1.1212340
# }