Draw plausible parameter instantiations from a given model

Draws plausible parameters from a model using parametric sampling (if the information matrix was computed) or via bootstrap sampling. Primarily for use with the DRF function.

draw_parameters(
  mod,
  draws,
  method = c("parametric", "boostrap"),
  redraws = 20,
  verbose = FALSE,
  ...
)

Arguments

mod

estimated single or multiple-group model

draws

number of draws to obtain

method

type of plausible values to obtain. Can be 'parametric', for the parametric sampling scheme which uses the estimated information matrix, or 'boostrap' to obtain values from the boot function. Default is 'parametric'

redraws

number of redraws to perform when the given parameteric sample does not satisfy the upper and lower parameter bounds. If a valid set cannot be found within this number of draws then an error will be thrown

verbose

logical; include additional information in the console?

...

additional arguments to be passed

Value

returns a draws x p matrix of plausible parameters, where each row correspeonds to a single set

Examples

# \donttest{
set.seed(1234)
n <- 40
N <- 500

# only first 5 items as anchors
model <- 'F = 1-40
          CONSTRAINB = (1-5, a1), (1-5, d)'

a <- matrix(1, n)
d <- matrix(rnorm(n), n)
group <- c(rep('Group_1', N), rep('Group_2', N))

## -------------
# groups completely equal
dat1 <- simdata(a, d, N, itemtype = 'dich')
dat2 <- simdata(a, d, N, itemtype = 'dich')
dat <- rbind(dat1, dat2)
mod <- multipleGroup(dat, model, group=group, SE=TRUE,
                     invariance=c('free_means', 'free_var'))

param_set <- draw_parameters(mod, 100)
head(param_set)
#>           [,1]      [,2] [,3] [,4]      [,5]      [,6] [,7] [,8]     [,9]
#> [1,] 1.1391733 -1.262820    0    1 1.0429681 0.4287659    0    1 1.426617
#> [2,] 1.0485232 -1.270312    0    1 1.2642454 0.4138879    0    1 1.055090
#> [3,] 0.9015717 -1.120912    0    1 1.1143080 0.2912340    0    1 1.179329
#> [4,] 1.2083866 -1.252729    0    1 1.2302781 0.3838831    0    1 1.282244
#> [5,] 1.0423036 -1.129994    0    1 1.0267542 0.4105462    0    1 1.109908
#> [6,] 0.9569426 -1.073894    0    1 0.8923473 0.5171925    0    1 1.122120
#>          [,10] [,11] [,12]     [,13]     [,14] [,15] [,16]     [,17]     [,18]
#> [1,] 1.3570750     0     1 1.3229055 -2.259822     0     1 0.7230170 0.4306938
#> [2,] 0.9837466     0     1 1.1903106 -2.265426     0     1 0.7787884 0.6616437
#> [3,] 1.1911208     0     1 1.1822470 -2.236812     0     1 0.9837936 0.5259178
#> [4,] 0.9823447     0     1 1.0777503 -2.369026     0     1 0.8211318 0.3824945
#> [5,] 1.0500731     0     1 1.1938625 -2.170304     0     1 0.7517937 0.3950251
#> [6,] 1.3030726     0     1 0.9314959 -2.084336     0     1 0.8255724 0.5517900
#>      [,19] [,20]    [,21]     [,22] [,23] [,24]    [,25]      [,26] [,27] [,28]
#> [1,]     0     1 1.163286 0.4935298     0     1 1.138409 -0.4731452     0     1
#> [2,]     0     1 1.295998 0.2962255     0     1 1.277704 -0.3844025     0     1
#> [3,]     0     1 1.393149 0.4956957     0     1 1.207933 -0.5510122     0     1
#> [4,]     0     1 1.221368 0.2570363     0     1 1.050332 -0.5915070     0     1
#> [5,]     0     1 1.051678 0.1173752     0     1 1.302897 -0.4298850     0     1
#> [6,]     0     1 1.066981 0.3894048     0     1 1.192811 -0.4857772     0     1
#>          [,29]      [,30] [,31] [,32]     [,33]      [,34] [,35] [,36]
#> [1,] 1.3527931 -0.5622184     0     1 0.8924523 -0.5117759     0     1
#> [2,] 1.1797532 -0.5451436     0     1 0.9312762 -0.4114688     0     1
#> [3,] 1.4738014 -0.6304343     0     1 0.8593679 -0.4383919     0     1
#> [4,] 1.0532547 -0.6969281     0     1 1.0449001 -0.4548017     0     1
#> [5,] 0.9259658 -0.5687871     0     1 0.8781297 -0.6391258     0     1
#> [6,] 1.4960287 -0.4478862     0     1 0.8747685 -0.4139944     0     1
#>          [,37]      [,38] [,39] [,40]     [,41]      [,42] [,43] [,44]
#> [1,] 1.0394596 -0.9996972     0     1 1.3385282 -0.4399517     0     1
#> [2,] 0.8843239 -0.9601556     0     1 1.0098399 -0.4069373     0     1
#> [3,] 1.0472867 -0.8473396     0     1 1.0328695 -0.4166243     0     1
#> [4,] 1.0481835 -1.1209198     0     1 1.5198540 -0.3906373     0     1
#> [5,] 0.8323404 -0.7925892     0     1 0.9477676 -0.4691304     0     1
#> [6,] 0.8903154 -0.8060392     0     1 1.1486426 -0.5071443     0     1
#>          [,45]      [,46] [,47] [,48]     [,49]      [,50] [,51] [,52]
#> [1,] 1.0213206 -1.0562947     0     1 1.2856267 -0.9980318     0     1
#> [2,] 1.0863032 -0.9086602     0     1 1.1838055 -0.6952484     0     1
#> [3,] 1.3031199 -1.0692199     0     1 1.0733006 -0.7782572     0     1
#> [4,] 1.2671360 -1.2655549     0     1 1.0755935 -0.8151340     0     1
#> [5,] 0.9003773 -0.8237290     0     1 1.0510661 -0.7829578     0     1
#> [6,] 0.7936868 -0.8023846     0     1 0.8012062 -0.7711908     0     1
#>          [,53]      [,54] [,55] [,56]    [,57]     [,58] [,59] [,60]     [,61]
#> [1,] 0.8176167 0.14315328     0     1 1.300684 1.0884409     0     1 0.8592587
#> [2,] 0.9589909 0.23628262     0     1 1.465505 1.0258226     0     1 1.1314658
#> [3,] 0.8126453 0.21269760     0     1 1.384538 1.0716422     0     1 0.8341418
#> [4,] 0.9141184 0.12653640     0     1 1.376912 0.9619842     0     1 0.9868513
#> [5,] 0.9342297 0.07777191     0     1 1.084789 1.0487468     0     1 0.9410772
#> [6,] 1.0545830 0.01141081     0     1 1.368815 1.3333229     0     1 1.0745821
#>            [,62] [,63] [,64]     [,65]      [,66] [,67] [,68]    [,69]
#> [1,] -0.11402988     0     1 0.8092405 -0.5234029     0     1 1.313815
#> [2,] -0.03003679     0     1 1.0429917 -0.5469948     0     1 1.241368
#> [3,] -0.11672769     0     1 1.2888724 -0.5122424     0     1 1.274849
#> [4,] -0.21990949     0     1 1.1642967 -0.7155273     0     1 1.077225
#> [5,]  0.06662208     0     1 0.9142243 -0.4847983     0     1 1.226575
#> [6,]  0.11544060     0     1 0.8656696 -0.3202004     0     1 1.139970
#>           [,70] [,71] [,72]     [,73]      [,74] [,75] [,76]    [,77]    [,78]
#> [1,] -0.9808861     0     1 1.0888446 -0.7742100     0     1 1.178653 2.567495
#> [2,] -0.9901593     0     1 1.0017598 -0.5739917     0     1 1.713041 3.165237
#> [3,] -1.1134380     0     1 0.8404794 -0.5877732     0     1 1.559150 2.637695
#> [4,] -0.9457465     0     1 0.9973520 -0.5775796     0     1 1.482787 2.682632
#> [5,] -1.0808770     0     1 1.0080005 -0.6276717     0     1 1.150031 2.552812
#> [6,] -0.7695208     0     1 1.3709517 -0.6366048     0     1 1.572617 2.960891
#>      [,79] [,80]     [,81]         [,82] [,83] [,84]     [,85]      [,86] [,87]
#> [1,]     0     1 0.9749299  1.262013e-01     0     1 1.2569399 -0.6753813     0
#> [2,]     0     1 0.6378761 -4.166715e-02     0     1 1.3503466 -0.6990225     0
#> [3,]     0     1 1.0769441  7.471144e-02     0     1 1.1304484 -0.4273836     0
#> [4,]     0     1 0.9466782  6.372824e-05     0     1 1.2338678 -0.7306365     0
#> [5,]     0     1 0.9646472  2.222591e-02     0     1 0.8675055 -0.5534084     0
#> [6,]     0     1 0.8953240  5.503009e-02     0     1 0.9862144 -0.5062192     0
#>      [,88]     [,89]      [,90] [,91] [,92]     [,93]     [,94] [,95] [,96]
#> [1,]     1 0.8943722 -0.3696351     0     1 1.1414429 0.4496148     0     1
#> [2,]     1 0.8856739 -0.3901606     0     1 1.1720946 0.4236956     0     1
#> [3,]     1 0.7230564 -0.4177643     0     1 1.1242100 0.3720655     0     1
#> [4,]     1 0.7502966 -0.4709601     0     1 1.1538625 0.5534444     0     1
#> [5,]     1 0.8277971 -0.5723495     0     1 0.9200647 0.4595635     0     1
#> [6,]     1 0.8063382 -0.2832313     0     1 0.9777832 0.4521123     0     1
#>          [,97]      [,98] [,99] [,100]   [,101]    [,102] [,103] [,104]
#> [1,] 0.9558518 -0.6142525     0      1 1.345433 -1.290486      0      1
#> [2,] 1.1158192 -0.5877017     0      1 1.575491 -1.563344      0      1
#> [3,] 1.2789187 -0.6621007     0      1 1.565977 -1.553490      0      1
#> [4,] 1.0295226 -0.6705001     0      1 1.205478 -1.614853      0      1
#> [5,] 0.9924577 -0.5625288     0      1 1.092125 -1.555417      0      1
#> [6,] 1.0533350 -0.3614032     0      1 1.402934 -1.392839      0      1
#>        [,105]    [,106] [,107] [,108]    [,109]     [,110] [,111] [,112]
#> [1,] 1.371332 0.8211582      0      1 1.0591426 -0.7178136      0      1
#> [2,] 1.387880 0.6305299      0      1 1.2003554 -0.9512186      0      1
#> [3,] 1.236612 0.5939691      0      1 0.9270792 -0.9554504      0      1
#> [4,] 1.229718 0.5413782      0      1 0.8022873 -0.8891415      0      1
#> [5,] 1.279880 0.9269160      0      1 1.0720228 -1.0762579      0      1
#> [6,] 1.553084 1.0841084      0      1 1.3316037 -0.9780793      0      1
#>         [,113]      [,114] [,115] [,116]    [,117]     [,118] [,119] [,120]
#> [1,] 1.3841784 -0.11066277      0      1 0.9359219 -0.9057657      0      1
#> [2,] 1.2721137 -0.12848081      0      1 1.1714822 -0.9518233      0      1
#> [3,] 0.9154618 -0.25251407      0      1 1.2103078 -0.9698035      0      1
#> [4,] 1.2261814 -0.23384254      0      1 1.0271726 -0.7794828      0      1
#> [5,] 1.3606482 -0.11407690      0      1 0.9669156 -0.8007007      0      1
#> [6,] 1.1510474 -0.05008875      0      1 0.8964719 -0.7972877      0      1
#>         [,121]   [,122] [,123] [,124]    [,125]      [,126] [,127] [,128]
#> [1,] 1.0723225 1.231318      0      1 0.9407839 -0.32892724      0      1
#> [2,] 1.0880436 1.285069      0      1 1.0321043 -0.34033120      0      1
#> [3,] 1.3211002 1.525599      0      1 0.9438071 -0.27040348      0      1
#> [4,] 0.7331573 1.299408      0      1 1.0681686 -0.30191764      0      1
#> [5,] 1.1344772 1.469813      0      1 0.7784884 -0.10345169      0      1
#> [6,] 0.9448876 1.203234      0      1 0.8675671 -0.06172586      0      1
#>         [,129]     [,130] [,131] [,132]    [,133]     [,134] [,135] [,136]
#> [1,] 0.8582868 -0.5372406      0      1 1.0455409 -0.5890446      0      1
#> [2,] 1.0358113 -0.5796790      0      1 0.7891652 -0.4892658      0      1
#> [3,] 1.0376526 -0.5784629      0      1 1.0742342 -0.7452859      0      1
#> [4,] 0.9489044 -0.7975239      0      1 0.8301274 -0.6310764      0      1
#> [5,] 0.7512060 -0.7169792      0      1 0.8355666 -0.4990572      0      1
#> [6,] 1.2407208 -0.6411680      0      1 1.0449231 -0.4900693      0      1
#>         [,137]    [,138] [,139] [,140]    [,141]     [,142] [,143] [,144]
#> [1,] 1.2294942 -1.652664      0      1 0.8113621 -1.0591429      0      1
#> [2,] 0.9020156 -1.372806      0      1 0.9551166 -1.1611195      0      1
#> [3,] 1.0222276 -1.510737      0      1 0.8123751 -0.9439469      0      1
#> [4,] 0.7452007 -1.646841      0      1 0.9577579 -1.1058304      0      1
#> [5,] 0.9055052 -1.493870      0      1 0.8405731 -0.9016169      0      1
#> [6,] 1.0238228 -1.606602      0      1 0.5875118 -0.6929535      0      1
#>         [,145]    [,146] [,147] [,148]   [,149]    [,150] [,151] [,152]
#> [1,] 0.9904598 -2.317371      0      1 1.117666 -1.358529      0      1
#> [2,] 1.1253041 -2.307893      0      1 1.051017 -1.437534      0      1
#> [3,] 1.1460432 -2.378614      0      1 1.337640 -1.418902      0      1
#> [4,] 1.2316940 -2.440233      0      1 1.103265 -1.415206      0      1
#> [5,] 1.2664484 -2.624121      0      1 1.225437 -1.485042      0      1
#> [6,] 1.1905095 -2.588034      0      1 1.003447 -1.142972      0      1
#>         [,153]     [,154] [,155] [,156]    [,157]     [,158] [,159] [,160]
#> [1,] 1.0052002 -0.1446961      0      1 1.2271201 -0.2834317      0      1
#> [2,] 0.7331597 -0.1336782      0      1 0.8868209 -0.3849022      0      1
#> [3,] 0.7006706 -0.1968696      0      1 1.1800788 -0.3826167      0      1
#> [4,] 0.7622345 -0.2943723      0      1 1.0171900 -0.2596404      0      1
#> [5,] 0.7589936 -0.3463861      0      1 0.9470770 -0.4107271      0      1
#> [6,] 0.8812599 -0.1017898      0      1 1.1799026 -0.2237757      0      1
#>      [,161] [,162]    [,163]    [,164] [,165] [,166]    [,167]    [,168] [,169]
#> [1,]      0      1 1.1391733 -1.262820      0      1 1.0429681 0.4287659      0
#> [2,]      0      1 1.0485232 -1.270312      0      1 1.2642454 0.4138879      0
#> [3,]      0      1 0.9015717 -1.120912      0      1 1.1143080 0.2912340      0
#> [4,]      0      1 1.2083866 -1.252729      0      1 1.2302781 0.3838831      0
#> [5,]      0      1 1.0423036 -1.129994      0      1 1.0267542 0.4105462      0
#> [6,]      0      1 0.9569426 -1.073894      0      1 0.8923473 0.5171925      0
#>      [,170]   [,171]    [,172] [,173] [,174]    [,175]    [,176] [,177] [,178]
#> [1,]      1 1.426617 1.3570750      0      1 1.3229055 -2.259822      0      1
#> [2,]      1 1.055090 0.9837466      0      1 1.1903106 -2.265426      0      1
#> [3,]      1 1.179329 1.1911208      0      1 1.1822470 -2.236812      0      1
#> [4,]      1 1.282244 0.9823447      0      1 1.0777503 -2.369026      0      1
#> [5,]      1 1.109908 1.0500731      0      1 1.1938625 -2.170304      0      1
#> [6,]      1 1.122120 1.3030726      0      1 0.9314959 -2.084336      0      1
#>         [,179]    [,180] [,181] [,182]    [,183]    [,184] [,185] [,186]
#> [1,] 0.7230170 0.4306938      0      1 0.5769491 0.3784814      0      1
#> [2,] 0.7787884 0.6616437      0      1 0.9951315 0.5034109      0      1
#> [3,] 0.9837936 0.5259178      0      1 0.6653489 0.2313406      0      1
#> [4,] 0.8211318 0.3824945      0      1 0.9269538 0.3331860      0      1
#> [5,] 0.7517937 0.3950251      0      1 0.6941987 0.4615722      0      1
#> [6,] 0.8255724 0.5517900      0      1 0.8832421 0.6586239      0      1
#>         [,187]     [,188] [,189] [,190]    [,191]     [,192] [,193] [,194]
#> [1,] 1.1773834 -0.5981404      0      1 0.8270861 -0.3566986      0      1
#> [2,] 1.2006684 -0.4581539      0      1 1.2312320 -0.4470088      0      1
#> [3,] 0.9915690 -0.4398309      0      1 0.7832556 -0.3699639      0      1
#> [4,] 0.8631583 -0.6455257      0      1 0.9260677 -0.5539042      0      1
#> [5,] 1.1586244 -0.5415647      0      1 0.9432377 -0.4512672      0      1
#> [6,] 1.1344523 -0.1557483      0      1 0.8387873 -0.2677301      0      1
#>         [,195]     [,196] [,197] [,198]    [,199]     [,200] [,201] [,202]
#> [1,] 0.7955953 -0.6941641      0      1 1.1881512 -0.8673675      0      1
#> [2,] 1.0363500 -0.6677137      0      1 0.9175865 -0.8381550      0      1
#> [3,] 0.8495524 -0.4472644      0      1 1.0530314 -0.7773129      0      1
#> [4,] 0.6804811 -0.7379656      0      1 0.9491587 -0.8843622      0      1
#> [5,] 0.8446075 -0.3753405      0      1 0.7418503 -0.4363831      0      1
#> [6,] 0.7291248 -0.3667064      0      1 0.8138640 -0.5492331      0      1
#>         [,203]     [,204] [,205] [,206]    [,207]     [,208] [,209] [,210]
#> [1,] 1.1721655 -0.5049901      0      1 1.4271475 -0.8892223      0      1
#> [2,] 0.9024895 -0.3659127      0      1 1.3371642 -0.7302579      0      1
#> [3,] 0.8207300 -0.6219701      0      1 1.2302162 -0.8840178      0      1
#> [4,] 1.1753623 -0.8480960      0      1 1.0317380 -1.2797787      0      1
#> [5,] 0.9550778 -0.3684216      0      1 1.5088457 -0.9402829      0      1
#> [6,] 0.7787419 -0.1533637      0      1 0.6774169 -0.6722728      0      1
#>         [,211]     [,212] [,213] [,214]    [,215]    [,216] [,217] [,218]
#> [1,] 1.0828575 -0.5862648      0      1 0.9437633 0.0962067      0      1
#> [2,] 0.9180048 -0.6700223      0      1 1.1634167 0.4086137      0      1
#> [3,] 0.8099235 -0.6892266      0      1 0.9437323 0.3688838      0      1
#> [4,] 0.9380843 -0.7709447      0      1 0.9953350 0.2498628      0      1
#> [5,] 0.9065601 -0.5228424      0      1 0.8901406 0.2285538      0      1
#> [6,] 0.8478834 -0.5028715      0      1 0.8493648 0.6491576      0      1
#>        [,219]    [,220] [,221] [,222]    [,223]      [,224] [,225] [,226]
#> [1,] 1.251234 1.2658886      0      1 0.8093858 -0.03790735      0      1
#> [2,] 1.275142 1.1703791      0      1 0.8337950 -0.08659292      0      1
#> [3,] 1.076003 1.0331035      0      1 0.8603245 -0.24091030      0      1
#> [4,] 1.179012 0.8964719      0      1 1.0907076 -0.21612512      0      1
#> [5,] 1.465817 1.4526927      0      1 0.9667567  0.16196365      0      1
#> [6,] 1.003670 1.2809047      0      1 0.9793593  0.08777435      0      1
#>         [,227]     [,228] [,229] [,230]    [,231]     [,232] [,233] [,234]
#> [1,] 1.1241366 -0.2657009      0      1 0.8256937 -0.6486497      0      1
#> [2,] 1.1013752 -0.3205930      0      1 1.0771003 -0.8223220      0      1
#> [3,] 1.2209014 -0.2839774      0      1 0.7337646 -0.7364254      0      1
#> [4,] 1.1255481 -0.5293575      0      1 0.7869873 -1.0086808      0      1
#> [5,] 1.3929660 -0.1998158      0      1 0.7863870 -0.7640716      0      1
#> [6,] 0.9912053 -0.1306283      0      1 0.7348774 -0.7733009      0      1
#>         [,235]     [,236] [,237] [,238]    [,239]   [,240] [,241] [,242]
#> [1,] 1.0720175 -0.4476852      0      1 1.0192340 2.661261      0      1
#> [2,] 0.9165282 -0.5344050      0      1 1.3178091 2.756994      0      1
#> [3,] 0.9368710 -0.6706738      0      1 0.9743164 2.529591      0      1
#> [4,] 0.8131438 -0.5931740      0      1 0.9871451 2.429035      0      1
#> [5,] 1.0643698 -0.4906245      0      1 1.2127107 2.697031      0      1
#> [6,] 0.8418784 -0.4322469      0      1 0.7567214 2.617236      0      1
#>         [,243]      [,244] [,245] [,246]    [,247]      [,248] [,249] [,250]
#> [1,] 0.8355340  0.06772162      0      1 1.4138044 -0.21335226      0      1
#> [2,] 1.1336368  0.25337404      0      1 1.2815356 -0.17720487      0      1
#> [3,] 0.5790983  0.13761689      0      1 1.2638424 -0.27388825      0      1
#> [4,] 0.9006870 -0.10527717      0      1 1.0605285 -0.50725867      0      1
#> [5,] 0.9075384  0.18466623      0      1 1.5191647 -0.14991268      0      1
#> [6,] 0.9369036  0.43911745      0      1 0.9969278  0.06110801      0      1
#>         [,251]     [,252] [,253] [,254]    [,255]    [,256] [,257] [,258]
#> [1,] 1.1006458 -0.2311133      0      1 1.4073203 0.8881620      0      1
#> [2,] 1.0802621 -0.6115840      0      1 1.2684601 0.9620531      0      1
#> [3,] 0.7880575 -0.3300319      0      1 1.1213746 0.9071288      0      1
#> [4,] 0.9676795 -0.5463861      0      1 1.3449152 0.6588780      0      1
#> [5,] 1.2516888 -0.2831493      0      1 1.4669935 1.2248778      0      1
#> [6,] 0.8800113 -0.2296308      0      1 0.9690143 1.0714963      0      1
#>         [,259]     [,260] [,261] [,262]    [,263]    [,264] [,265] [,266]
#> [1,] 1.0834802 -0.2875225      0      1 1.2372837 -1.585027      0      1
#> [2,] 1.0168853 -0.4051299      0      1 0.7296664 -1.262766      0      1
#> [3,] 1.2575420 -0.5187227      0      1 0.8585518 -1.363042      0      1
#> [4,] 0.9252004 -0.7953462      0      1 1.1705405 -1.415958      0      1
#> [5,] 1.0538871 -0.5371288      0      1 0.9277104 -1.212550      0      1
#> [6,] 0.6340702 -0.3536110      0      1 0.8223351 -1.063511      0      1
#>         [,267]    [,268] [,269] [,270]    [,271]     [,272] [,273] [,274]
#> [1,] 1.3061236 0.6538265      0      1 0.9254825 -1.0120478      0      1
#> [2,] 1.1651533 0.6359037      0      1 0.6042100 -0.8961290      0      1
#> [3,] 1.0425028 0.5803314      0      1 0.9719572 -0.9050785      0      1
#> [4,] 1.0207775 0.4596206      0      1 0.8854151 -0.8462630      0      1
#> [5,] 1.3435261 0.7991064      0      1 1.0889661 -0.7383682      0      1
#> [6,] 0.8798895 0.8474392      0      1 0.8833759 -0.9096373      0      1
#>         [,275]     [,276] [,277] [,278]    [,279]     [,280] [,281] [,282]
#> [1,] 1.1094924 0.26984316      0      1 1.2318932 -0.9163726      0      1
#> [2,] 1.1581401 0.22183635      0      1 0.9474796 -0.8551239      0      1
#> [3,] 0.8855515 0.02177485      0      1 1.0470557 -1.0409513      0      1
#> [4,] 1.0652738 0.08385080      0      1 1.0388297 -0.9948871      0      1
#> [5,] 1.2904505 0.37346468      0      1 0.9947491 -0.7321717      0      1
#> [6,] 1.0330021 0.53328989      0      1 1.0023814 -0.6591202      0      1
#>         [,283]   [,284] [,285] [,286]    [,287]     [,288] [,289] [,290]
#> [1,] 1.0128029 1.086628      0      1 1.3016749 -0.5646866      0      1
#> [2,] 0.9325979 1.184938      0      1 1.2797706 -0.3322953      0      1
#> [3,] 0.8565952 1.200579      0      1 1.0858322 -0.5371236      0      1
#> [4,] 0.7655360 1.001292      0      1 0.9558104 -0.6197793      0      1
#> [5,] 0.9698579 1.206098      0      1 1.0974484 -0.3218921      0      1
#> [6,] 0.8846806 1.368624      0      1 0.6796287 -0.4114744      0      1
#>         [,291]     [,292] [,293] [,294]    [,295]      [,296] [,297] [,298]
#> [1,] 0.9665767 -0.4500584      0      1 1.2751350 -0.27255013      0      1
#> [2,] 1.0356999 -0.5722772      0      1 1.1089670 -0.16909395      0      1
#> [3,] 0.9499417 -0.4462351      0      1 1.1750721 -0.44601214      0      1
#> [4,] 0.6212870 -0.6610116      0      1 0.9129748 -0.70065364      0      1
#> [5,] 0.9964826 -0.5006607      0      1 1.1185854 -0.19542343      0      1
#> [6,] 0.7778939 -0.4091420      0      1 1.0454018 -0.08139411      0      1
#>         [,299]    [,300] [,301] [,302]    [,303]    [,304] [,305] [,306]
#> [1,] 0.9457425 -1.694496      0      1 0.8919544 -1.298281      0      1
#> [2,] 1.1671781 -1.569062      0      1 1.1484472 -1.217428      0      1
#> [3,] 0.9105014 -1.589028      0      1 0.7172947 -1.194990      0      1
#> [4,] 0.7951716 -1.910108      0      1 0.9743384 -1.257467      0      1
#> [5,] 0.8852933 -1.377246      0      1 0.9976156 -1.021672      0      1
#> [6,] 0.8919794 -1.526444      0      1 0.7979643 -1.071051      0      1
#>         [,307]    [,308] [,309] [,310]    [,311]     [,312] [,313] [,314]
#> [1,] 0.9776926 -1.917675      0      1 0.7331121 -0.9768939      0      1
#> [2,] 0.9623532 -1.886301      0      1 1.0542755 -1.0652908      0      1
#> [3,] 0.9434050 -2.146977      0      1 1.0355315 -1.1162900      0      1
#> [4,] 1.0764964 -2.357499      0      1 1.0399000 -1.2173863      0      1
#> [5,] 1.0682937 -2.334819      0      1 1.1998077 -1.0728763      0      1
#> [6,] 0.8298586 -1.838885      0      1 1.0184965 -1.0134333      0      1
#>         [,315]        [,316] [,317] [,318]    [,319]      [,320] [,321] [,322]
#> [1,] 1.0091983 -0.2197823576      0      1 1.0090839 -0.25014011      0      1
#> [2,] 1.0190892 -0.0009007269      0      1 0.9135354 -0.29233457      0      1
#> [3,] 1.0795103 -0.1674778046      0      1 0.7342503 -0.35314781      0      1
#> [4,] 0.9524419 -0.4013812735      0      1 0.9772553 -0.27946268      0      1
#> [5,] 1.1289974  0.1363981641      0      1 0.7633143 -0.12467480      0      1
#> [6,] 0.8079596 -0.0062898392      0      1 0.7487938 -0.06748691      0      1
#>           [,323]    [,324]
#> [1,] -0.17442566 0.7971556
#> [2,] -0.23040635 0.8413351
#> [3,] -0.17632929 1.0450458
#> [4,]  0.06453193 1.0788725
#> [5,] -0.27472444 0.8621631
#> [6,] -0.47536681 1.1212272
# }