draw_parameters {mirt}	R Documentation

Draw plausible parameter instantiations from a given model

Description

Draws plausible parameters from a model using parametric sampling (if the information matrix was computed) or via boostrap sampling. Primarily for use with the DRF function.

Usage

draw_parameters(
  mod,
  draws,
  method = c("parametric", "boostrap"),
  redraws = 20,
  ...
)

Arguments

mod
estimated single or multiple-group model

draws
number of draws to obtain

method
type of plausible values to obtain. Can be 'parametric', for the parametric sampling scheme which uses the estimated information matrix, or 'boostrap' to obtain values from the boot function. Default is 'parametric'

redraws
number of redraws to perform when the given parameteric sample does not satisfy the upper and lower parameter bounds. If a valid set cannot be found within this number of draws then an error will be thrown

...
additional arguments to be passed

Value

returns a draws x p matrix of plausible parameters, where each row correspeonds to a single set

Examples

## No test: 
set.seed(1234)
n <- 40
N <- 500

# only first 5 items as anchors
model <- 'F = 1-40
          CONSTRAINB = (1-5, a1), (1-5, d)'

a <- matrix(1, n)
d <- matrix(rnorm(n), n)
group <- c(rep('Group_1', N), rep('Group_2', N))

## -------------
# groups completely equal
dat1 <- simdata(a, d, N, itemtype = 'dich')
dat2 <- simdata(a, d, N, itemtype = 'dich')
dat <- rbind(dat1, dat2)
mod <- multipleGroup(dat, model, group=group, SE=TRUE,
                     invariance=c('free_means', 'free_var'))

param_set <- draw_parameters(mod, 100)
head(param_set)

##          [,1]      [,2] [,3] [,4]     [,5]      [,6] [,7] [,8]     [,9]
## [1,] 1.229930 -1.127495    0    1 1.165967 0.5236504    0    1 1.245661
## [2,] 1.063362 -1.335845    0    1 1.267425 0.2638424    0    1 1.103437
## [3,] 1.040311 -1.268552    0    1 1.181595 0.2678553    0    1 1.245735
## [4,] 1.114719 -1.271659    0    1 1.056160 0.3051946    0    1 1.239188
## [5,] 1.059611 -1.212607    0    1 1.114294 0.4793085    0    1 1.101774
## [6,] 1.104261 -1.293510    0    1 1.067493 0.4266354    0    1 1.134587
##         [,10] [,11] [,12]     [,13]     [,14] [,15] [,16]     [,17]     [,18]
## [1,] 1.347222     0     1 1.3261315 -2.287505     0     1 0.8823766 0.5877615
## [2,] 1.050612     0     1 1.2391720 -2.208921     0     1 0.9384455 0.4510176
## [3,] 1.028923     0     1 1.2463804 -2.359680     0     1 0.8977385 0.4318826
## [4,] 1.163602     0     1 0.8765951 -2.195838     0     1 0.7692732 0.4968675
## [5,] 1.191674     0     1 1.0885062 -2.143565     0     1 0.8689042 0.4819905
## [6,] 1.241187     0     1 1.0971045 -2.029670     0     1 0.9247818 0.4814218
##      [,19] [,20]     [,21]     [,22] [,23] [,24]     [,25]      [,26] [,27]
## [1,]     0     1 1.2932910 0.4332160     0     1 1.2087286 -0.4018337     0
## [2,]     0     1 1.2964323 0.3950519     0     1 0.9924687 -0.7018250     0
## [3,]     0     1 1.1776420 0.2042382     0     1 1.2023037 -0.4529604     0
## [4,]     0     1 1.0777628 0.4491213     0     1 0.9408641 -0.4994771     0
## [5,]     0     1 0.9282249 0.3680442     0     1 1.1899908 -0.5997108     0
## [6,]     0     1 1.4128534 0.4849459     0     1 0.9817304 -0.2737389     0
##      [,28]    [,29]      [,30] [,31] [,32]     [,33]      [,34] [,35] [,36]
## [1,]     1 1.121547 -0.4739862     0     1 0.9575984 -0.3781598     0     1
## [2,]     1 1.295395 -0.6455588     0     1 0.7579322 -0.4642891     0     1
## [3,]     1 1.071237 -0.6004781     0     1 0.7841706 -0.5233419     0     1
## [4,]     1 1.136338 -0.7398158     0     1 0.8931545 -0.4946406     0     1
## [5,]     1 1.000148 -0.6832509     0     1 0.8662252 -0.4995271     0     1
## [6,]     1 1.260210 -0.5910757     0     1 1.0063608 -0.3583589     0     1
##          [,37]      [,38] [,39] [,40]     [,41]      [,42] [,43] [,44]
## [1,] 0.9080870 -0.8761633     0     1 1.1585296 -0.1990391     0     1
## [2,] 1.3740707 -1.0025605     0     1 1.4997500 -0.6755151     0     1
## [3,] 0.9726962 -0.7459558     0     1 1.1283466 -0.4511926     0     1
## [4,] 0.9278879 -0.9510917     0     1 1.3359138 -0.5477567     0     1
## [5,] 1.0222057 -0.9222881     0     1 1.1172190 -0.2524260     0     1
## [6,] 1.1387540 -0.9551479     0     1 0.8443457 -0.3071329     0     1
##          [,45]      [,46] [,47] [,48]     [,49]      [,50] [,51] [,52]
## [1,] 0.9649551 -0.9246309     0     1 0.8630809 -0.9553373     0     1
## [2,] 1.3421147 -1.2592768     0     1 1.1565610 -1.0689234     0     1
## [3,] 0.9656864 -1.0651990     0     1 1.2626160 -0.9862924     0     1
## [4,] 1.1620523 -1.1487465     0     1 1.2844062 -0.7783344     0     1
## [5,] 0.9735839 -0.9678074     0     1 1.1329034 -0.7336921     0     1
## [6,] 1.3170550 -0.8939729     0     1 0.9874061 -0.8939469     0     1
##          [,53]       [,54] [,55] [,56]    [,57]     [,58] [,59] [,60]     [,61]
## [1,] 0.8712516  0.18175586     0     1 1.297115 0.9531491     0     1 0.8935261
## [2,] 0.9227284  0.05210759     0     1 1.176491 0.6987984     0     1 1.0148495
## [3,] 1.0346675  0.01828623     0     1 1.203491 0.9470125     0     1 0.8866179
## [4,] 0.9785408  0.17732064     0     1 1.060233 1.0000935     0     1 1.1483633
## [5,] 1.0766817 -0.02713271     0     1 1.310869 1.0033090     0     1 0.9796655
## [6,] 0.8989605  0.03363983     0     1 1.288271 1.2077124     0     1 1.0221598
##            [,62] [,63] [,64]    [,65]      [,66] [,67] [,68]     [,69]
## [1,] -0.01766144     0     1 1.202815 -0.4045610     0     1 0.9016037
## [2,] -0.31661644     0     1 1.136262 -0.5484247     0     1 1.2900842
## [3,] -0.14041485     0     1 1.015469 -0.5007208     0     1 1.2080163
## [4,]  0.09257785     0     1 1.236729 -0.4815952     0     1 1.0946077
## [5,] -0.20331280     0     1 1.120985 -0.4810076     0     1 1.0869767
## [6,] -0.22338446     0     1 1.076216 -0.4800945     0     1 1.0954424
##           [,70] [,71] [,72]     [,73]      [,74] [,75] [,76]    [,77]    [,78]
## [1,] -0.9339339     0     1 0.9230030 -0.6797646     0     1 1.722167 2.919241
## [2,] -1.1064851     0     1 1.0889798 -0.8488084     0     1 1.011962 2.341640
## [3,] -1.1567000     0     1 1.1571998 -0.7307210     0     1 1.075313 2.535092
## [4,] -0.9526493     0     1 1.0014398 -0.6655252     0     1 1.179257 2.861564
## [5,] -0.8968356     0     1 0.9458934 -0.8581025     0     1 1.200542 2.535631
## [6,] -1.0456343     0     1 1.0197648 -0.7536244     0     1 1.776275 3.501516
##      [,79] [,80]     [,81]       [,82] [,83] [,84]    [,85]      [,86] [,87]
## [1,]     0     1 0.9032435  0.02678452     0     1 1.518969 -0.7720860     0
## [2,]     0     1 0.9269503  0.02741822     0     1 1.345401 -0.8323135     0
## [3,]     0     1 1.0660583  0.12497679     0     1 1.315820 -0.6583921     0
## [4,]     0     1 0.6967074 -0.05295327     0     1 1.167834 -0.6289775     0
## [5,]     0     1 1.0606798  0.05431215     0     1 1.307343 -0.6154079     0
## [6,]     0     1 0.8637178  0.11433310     0     1 1.325652 -0.6887727     0
##      [,88]     [,89]      [,90] [,91] [,92]     [,93]     [,94] [,95] [,96]
## [1,]     1 0.7254718 -0.3179824     0     1 1.0680851 0.5838319     0     1
## [2,]     1 0.9181833 -0.5738278     0     1 0.9396883 0.3650090     0     1
## [3,]     1 0.6506578 -0.5320394     0     1 1.1041661 0.3291333     0     1
## [4,]     1 0.9302175 -0.5856709     0     1 1.0645499 0.3674973     0     1
## [5,]     1 0.9856584 -0.5095434     0     1 1.0356904 0.4482844     0     1
## [6,]     1 0.9154819 -0.3636087     0     1 1.1894177 0.7081647     0     1
##          [,97]      [,98] [,99] [,100]   [,101]    [,102] [,103] [,104]
## [1,] 1.1554730 -0.7079733     0      1 1.368932 -1.440752      0      1
## [2,] 0.9272031 -0.5893464     0      1 1.486069 -1.426476      0      1
## [3,] 1.1823199 -0.6222340     0      1 1.357679 -1.781583      0      1
## [4,] 1.0157046 -0.6493437     0      1 1.512646 -1.656810      0      1
## [5,] 1.0402589 -0.6145964     0      1 1.165508 -1.468848      0      1
## [6,] 1.3112966 -0.5957133     0      1 1.197796 -1.385740      0      1
##        [,105]    [,106] [,107] [,108]    [,109]     [,110] [,111] [,112]
## [1,] 1.461919 0.9607783      0      1 0.9754061 -0.9518514      0      1
## [2,] 1.509717 0.6794397      0      1 1.0812951 -0.9664172      0      1
## [3,] 1.316176 0.6122943      0      1 1.2812581 -1.2067870      0      1
## [4,] 1.441310 0.6420162      0      1 1.1336401 -0.9872395      0      1
## [5,] 1.292809 0.7085636      0      1 0.9009701 -0.8866273      0      1
## [6,] 1.226192 0.7610155      0      1 1.1116975 -1.0290203      0      1
##        [,113]      [,114] [,115] [,116]    [,117]     [,118] [,119] [,120]
## [1,] 1.219515 -0.07041117      0      1 1.2114968 -0.9946331      0      1
## [2,] 1.095809 -0.13293564      0      1 0.9709721 -0.8730477      0      1
## [3,] 1.174081 -0.28014099      0      1 0.9077266 -1.0579273      0      1
## [4,] 1.213625 -0.10446938      0      1 0.7836658 -0.7849504      0      1
## [5,] 1.298495 -0.08585412      0      1 0.8429664 -1.0422569      0      1
## [6,] 1.184827  0.15485392      0      1 1.0501176 -1.0356493      0      1
##         [,121]   [,122] [,123] [,124]    [,125]      [,126] [,127] [,128]
## [1,] 1.0048365 1.301446      0      1 0.9360415 -0.29873793      0      1
## [2,] 1.0845222 1.210790      0      1 0.6575700 -0.31944892      0      1
## [3,] 0.8436727 1.185328      0      1 0.9832782 -0.28466148      0      1
## [4,] 1.1238596 1.147141      0      1 0.8298457 -0.13306278      0      1
## [5,] 1.0541176 1.356740      0      1 0.8059054 -0.27161771      0      1
## [6,] 0.9684941 1.366075      0      1 1.0158563 -0.08061208      0      1
##         [,129]     [,130] [,131] [,132]    [,133]     [,134] [,135] [,136]
## [1,] 0.9275996 -0.5132244      0      1 1.0393471 -0.7425943      0      1
## [2,] 1.0997635 -0.6464280      0      1 1.0619770 -0.6426548      0      1
## [3,] 0.9138039 -0.5939197      0      1 1.1422550 -0.4838695      0      1
## [4,] 1.0400511 -0.6265943      0      1 1.0096545 -0.5732995      0      1
## [5,] 0.9294338 -0.6062455      0      1 1.0127096 -0.5827905      0      1
## [6,] 1.1843845 -0.6568862      0      1 0.8691991 -0.6217917      0      1
##         [,137]    [,138] [,139] [,140]    [,141]     [,142] [,143] [,144]
## [1,] 0.9391242 -1.309725      0      1 0.9574771 -0.8800343      0      1
## [2,] 0.9942921 -1.734926      0      1 0.6301034 -1.0993767      0      1
## [3,] 0.9194545 -1.646020      0      1 0.8256445 -1.0586811      0      1
## [4,] 1.0048017 -1.522371      0      1 0.9739577 -1.0617487      0      1
## [5,] 0.9813246 -1.610631      0      1 0.8074912 -0.8667388      0      1
## [6,] 0.9355722 -1.548664      0      1 1.0067750 -0.9646413      0      1
##         [,145]    [,146] [,147] [,148]    [,149]    [,150] [,151] [,152]
## [1,] 0.8954707 -2.470278      0      1 1.3357712 -1.305617      0      1
## [2,] 1.0810594 -2.489140      0      1 1.2462269 -1.453840      0      1
## [3,] 1.0891526 -2.536895      0      1 1.0833284 -1.499349      0      1
## [4,] 1.1301594 -2.274242      0      1 0.9583628 -1.352313      0      1
## [5,] 0.7156666 -2.466448      0      1 1.0570363 -1.474438      0      1
## [6,] 1.0119707 -2.386474      0      1 1.0607167 -1.344934      0      1
##         [,153]      [,154] [,155] [,156]    [,157]     [,158] [,159] [,160]
## [1,] 0.6856700 -0.13777419      0      1 1.0988028 -0.2729649      0      1
## [2,] 1.0370674 -0.18503024      0      1 1.0262733 -0.4623058      0      1
## [3,] 1.0042383 -0.17785398      0      1 1.1941970 -0.5518198      0      1
## [4,] 0.7560564 -0.12238919      0      1 1.0898491 -0.5403063      0      1
## [5,] 0.8477467 -0.01753896      0      1 0.9841364 -0.4349598      0      1
## [6,] 0.8023859 -0.29487262      0      1 0.8441605 -0.2929377      0      1
##      [,161] [,162]   [,163]    [,164] [,165] [,166]   [,167]    [,168] [,169]
## [1,]      0      1 1.229930 -1.127495      0      1 1.165967 0.5236504      0
## [2,]      0      1 1.063362 -1.335845      0      1 1.267425 0.2638424      0
## [3,]      0      1 1.040311 -1.268552      0      1 1.181595 0.2678553      0
## [4,]      0      1 1.114719 -1.271659      0      1 1.056160 0.3051946      0
## [5,]      0      1 1.059611 -1.212607      0      1 1.114294 0.4793085      0
## [6,]      0      1 1.104261 -1.293510      0      1 1.067493 0.4266354      0
##      [,170]   [,171]   [,172] [,173] [,174]    [,175]    [,176] [,177] [,178]
## [1,]      1 1.245661 1.347222      0      1 1.3261315 -2.287505      0      1
## [2,]      1 1.103437 1.050612      0      1 1.2391720 -2.208921      0      1
## [3,]      1 1.245735 1.028923      0      1 1.2463804 -2.359680      0      1
## [4,]      1 1.239188 1.163602      0      1 0.8765951 -2.195838      0      1
## [5,]      1 1.101774 1.191674      0      1 1.0885062 -2.143565      0      1
## [6,]      1 1.134587 1.241187      0      1 1.0971045 -2.029670      0      1
##         [,179]    [,180] [,181] [,182]    [,183]    [,184] [,185] [,186]
## [1,] 0.8823766 0.5877615      0      1 1.1253340 0.5472914      0      1
## [2,] 0.9384455 0.4510176      0      1 0.9467241 0.3365449      0      1
## [3,] 0.8977385 0.4318826      0      1 0.9034866 0.3283386      0      1
## [4,] 0.7692732 0.4968675      0      1 0.9939071 0.4135525      0      1
## [5,] 0.8689042 0.4819905      0      1 0.7016910 0.5091546      0      1
## [6,] 0.9247818 0.4814218      0      1 0.7265227 0.6034032      0      1
##         [,187]     [,188] [,189] [,190]    [,191]     [,192] [,193] [,194]
## [1,] 1.2501870 -0.3061878      0      1 1.1483143 -0.2396395      0      1
## [2,] 1.1789698 -0.7372866      0      1 1.1233467 -0.6168749      0      1
## [3,] 1.1722055 -0.7868991      0      1 1.0687784 -0.5901883      0      1
## [4,] 1.0729541 -0.6158544      0      1 0.8803137 -0.4927484      0      1
## [5,] 1.2128619 -0.5614407      0      1 0.9644047 -0.4279118      0      1
## [6,] 0.9161552 -0.3322778      0      1 0.7466370 -0.3395232      0      1
##         [,195]     [,196] [,197] [,198]    [,199]     [,200] [,201] [,202]
## [1,] 0.9161332 -0.4109735      0      1 1.2504545 -0.7118599      0      1
## [2,] 0.7838457 -0.3704741      0      1 1.1738950 -0.7926738      0      1
## [3,] 0.8222540 -0.6112330      0      1 1.0960506 -1.2434107      0      1
## [4,] 0.7458998 -0.7217136      0      1 1.0410311 -0.8712763      0      1
## [5,] 0.7032734 -0.3068244      0      1 0.8561249 -0.7593385      0      1
## [6,] 0.8648759 -0.3208848      0      1 1.0790674 -0.7292453      0      1
##         [,203]     [,204] [,205] [,206]    [,207]     [,208] [,209] [,210]
## [1,] 1.3744336 -0.4116127      0      1 1.3858901 -0.6785085      0      1
## [2,] 1.4043496 -0.6576042      0      1 1.1921293 -0.9530600      0      1
## [3,] 1.2108362 -0.7320278      0      1 1.1949814 -1.0859109      0      1
## [4,] 1.2865560 -0.7459333      0      1 1.1621434 -0.9772437      0      1
## [5,] 0.8696047 -0.5965363      0      1 1.1293511 -1.0974765      0      1
## [6,] 0.8597750 -0.4523606      0      1 0.9436415 -0.8969150      0      1
##         [,211]     [,212] [,213] [,214]    [,215]     [,216] [,217] [,218]
## [1,] 1.0103957 -0.4222871      0      1 1.1804849 0.50475295      0      1
## [2,] 0.8388468 -0.7363894      0      1 1.0091384 0.03638981      0      1
## [3,] 0.9067109 -1.0127003      0      1 0.8056815 0.03628504      0      1
## [4,] 0.8974697 -0.7186863      0      1 0.8739937 0.13928905      0      1
## [5,] 0.7724952 -0.6667995      0      1 0.9710937 0.13167401      0      1
## [6,] 1.0268297 -0.7397010      0      1 0.6440537 0.18062099      0      1
##        [,219]    [,220] [,221] [,222]    [,223]      [,224] [,225] [,226]
## [1,] 1.455234 1.4031976      0      1 1.1041221  0.08121849      0      1
## [2,] 1.220944 0.9575576      0      1 1.1979703 -0.13509180      0      1
## [3,] 1.175729 0.7297372      0      1 0.9300398 -0.17030652      0      1
## [4,] 1.248559 1.0507178      0      1 1.1456343 -0.30841959      0      1
## [5,] 1.158655 1.1854453      0      1 0.8556522 -0.26787207      0      1
## [6,] 1.213460 1.2306503      0      1 0.8339401 -0.16734236      0      1
##        [,227]     [,228] [,229] [,230]    [,231]     [,232] [,233] [,234]
## [1,] 1.247586 -0.1589762      0      1 1.0036026 -0.7880439      0      1
## [2,] 1.291007 -0.2584223      0      1 0.8692916 -0.7505884      0      1
## [3,] 1.310076 -0.5706410      0      1 0.7012516 -0.8845068      0      1
## [4,] 1.197343 -0.3081771      0      1 1.0314503 -0.8063905      0      1
## [5,] 1.052098 -0.3883791      0      1 0.7402240 -0.8354265      0      1
## [6,] 1.162991 -0.1691276      0      1 0.7201893 -0.7867833      0      1
##         [,235]     [,236] [,237] [,238]    [,239]   [,240] [,241] [,242]
## [1,] 0.9919418 -0.2835624      0      1 1.0808431 2.875704      0      1
## [2,] 0.9305298 -0.7559283      0      1 0.9990786 2.319479      0      1
## [3,] 0.9434192 -0.7173457      0      1 0.8732633 2.217805      0      1
## [4,] 0.9593247 -0.9771006      0      1 1.3201969 2.682981      0      1
## [5,] 0.8531775 -0.6683132      0      1 0.8634220 2.643137      0      1
## [6,] 0.8165902 -0.7647309      0      1 0.8965803 2.692163      0      1
##         [,243]      [,244] [,245] [,246]   [,247]      [,248] [,249] [,250]
## [1,] 0.9219481  0.21333416      0      1 1.511461 -0.20378974      0      1
## [2,] 0.7446521  0.27751067      0      1 1.543579 -0.22225896      0      1
## [3,] 1.0523149  0.09629859      0      1 1.208054 -0.35460565      0      1
## [4,] 0.6576203 -0.08628687      0      1 1.229619 -0.29634905      0      1
## [5,] 1.1344807  0.25667185      0      1 1.401293 -0.35610317      0      1
## [6,] 1.0122016  0.45308709      0      1 1.551390 -0.08205365      0      1
##         [,251]      [,252] [,253] [,254]   [,255]    [,256] [,257] [,258]
## [1,] 1.2747581 -0.06836194      0      1 1.346462 1.0521773      0      1
## [2,] 1.0079972 -0.48762041      0      1 1.383905 0.8282617      0      1
## [3,] 0.9973420 -0.55465202      0      1 1.262955 0.8153978      0      1
## [4,] 1.1844758 -0.23235658      0      1 1.297598 0.8639890      0      1
## [5,] 0.7870156 -0.26308791      0      1 1.185551 0.9172172      0      1
## [6,] 0.9098969 -0.43366149      0      1 1.565785 1.2216677      0      1
##         [,259]     [,260] [,261] [,262]    [,263]    [,264] [,265] [,266]
## [1,] 1.1372327 -0.4209584      0      1 1.4721452 -1.226322      0      1
## [2,] 1.0231270 -0.3592172      0      1 0.8273685 -1.440528      0      1
## [3,] 1.0621767 -0.7248795      0      1 0.8991617 -1.475028      0      1
## [4,] 0.9926731 -0.4808456      0      1 0.9307990 -1.094850      0      1
## [5,] 0.9799780 -0.4230738      0      1 0.9917669 -1.543039      0      1
## [6,] 0.9500567 -0.3651689      0      1 0.8391095 -1.305936      0      1
##        [,267]    [,268] [,269] [,270]    [,271]     [,272] [,273] [,274]
## [1,] 1.205067 0.7964343      0      1 0.8926905 -0.6823605      0      1
## [2,] 1.280810 0.6660007      0      1 0.8092837 -0.9154516      0      1
## [3,] 1.186600 0.6228647      0      1 0.8237914 -1.0950566      0      1
## [4,] 1.222301 0.5197987      0      1 1.0004272 -0.6944473      0      1
## [5,] 1.005269 0.5520480      0      1 0.7989710 -0.9495722      0      1
## [6,] 1.225330 0.8354920      0      1 1.1097550 -0.7629649      0      1
##         [,275]    [,276] [,277] [,278]    [,279]     [,280] [,281] [,282]
## [1,] 1.1805054 0.2358918      0      1 1.1704542 -0.6275024      0      1
## [2,] 1.1610257 0.2733002      0      1 1.3022135 -1.0651694      0      1
## [3,] 1.1569232 0.1023539      0      1 0.9729064 -1.0178017      0      1
## [4,] 1.1523881 0.2665125      0      1 0.9384864 -0.9013139      0      1
## [5,] 0.9251645 0.2194291      0      1 0.8352059 -0.8094770      0      1
## [6,] 1.0369120 0.3832051      0      1 0.9993715 -0.9882381      0      1
##         [,283]    [,284] [,285] [,286]    [,287]     [,288] [,289] [,290]
## [1,] 0.8693572 1.0162882      0      1 1.1826821 -0.3739144      0      1
## [2,] 1.1540911 1.0232604      0      1 1.1790770 -0.4347602      0      1
## [3,] 1.0942803 0.9836291      0      1 1.0292283 -0.5950219      0      1
## [4,] 0.8754138 1.0350679      0      1 0.7948062 -0.4137812      0      1
## [5,] 1.1306492 1.1442711      0      1 1.2185474 -0.5016048      0      1
## [6,] 1.2151630 1.3139246      0      1 0.8887096 -0.3185717      0      1
##         [,291]     [,292] [,293] [,294]    [,295]      [,296] [,297] [,298]
## [1,] 1.2341532 -0.4764232      0      1 1.4193763 -0.12034777      0      1
## [2,] 0.8711874 -0.5727206      0      1 1.2376507 -0.33794262      0      1
## [3,] 0.7728545 -0.7015813      0      1 1.1471800 -0.50438604      0      1
## [4,] 0.8797628 -0.5649026      0      1 1.1520023 -0.32261951      0      1
## [5,] 0.6242278 -0.6624442      0      1 0.8963058 -0.07677425      0      1
## [6,] 0.7699177 -0.5740354      0      1 1.0746791 -0.13230364      0      1
##         [,299]    [,300] [,301] [,302]    [,303]     [,304] [,305] [,306]
## [1,] 1.3617469 -1.602292      0      1 1.1858247 -0.8665673      0      1
## [2,] 0.9153259 -1.785685      0      1 1.4663810 -1.2983530      0      1
## [3,] 0.8366196 -1.750672      0      1 0.9596264 -1.5515404      0      1
## [4,] 1.3004220 -1.620950      0      1 1.0113779 -1.2229281      0      1
## [5,] 1.2686478 -1.596639      0      1 0.8195389 -1.2838053      0      1
## [6,] 1.2538738 -1.620753      0      1 0.8487585 -0.9298624      0      1
##        [,307]    [,308] [,309] [,310]   [,311]    [,312] [,313] [,314]
## [1,] 1.223143 -1.839042      0      1 1.312791 -1.057050      0      1
## [2,] 1.498743 -2.424702      0      1 0.995876 -1.084756      0      1
## [3,] 1.071954 -2.181931      0      1 1.167589 -1.192724      0      1
## [4,] 1.260491 -2.316369      0      1 1.228326 -1.230469      0      1
## [5,] 1.205663 -1.959447      0      1 1.169586 -1.157363      0      1
## [6,] 1.305295 -2.317913      0      1 1.050236 -1.178591      0      1
##         [,315]       [,316] [,317] [,318]    [,319]     [,320] [,321] [,322]
## [1,] 1.3188149 -0.075813541      0      1 1.0950163 -0.3182027      0      1
## [2,] 1.0802966 -0.232390330      0      1 1.0749034 -0.2861833      0      1
## [3,] 1.0403979 -0.258166277      0      1 0.8897723 -0.4393683      0      1
## [4,] 1.2489167 -0.021144364      0      1 1.0495681 -0.2523485      0      1
## [5,] 0.9110763 -0.282424473      0      1 0.7329997 -0.4071027      0      1
## [6,] 0.9108462 -0.007230206      0      1 0.9294666 -0.3528736      0      1
##           [,323]    [,324]
## [1,] -0.26753035 0.6593778
## [2,] -0.06739767 0.7896170
## [3,]  0.01470670 0.8532682
## [4,] -0.19521844 0.9235054
## [5,] -0.16929659 1.0520026
## [6,] -0.25297763 0.9737236

## End(No test)

[Package mirt version 1.31.9 Index]