Function to calculate the empirical (marginal) reliability

Given secondary latent trait estimates and their associated standard errors returned from fscores, compute the empirical reliability.

Usage

empirical_rxx(Theta_SE, T_as_X = FALSE)

Arguments

Theta_SE

a matrix of latent trait estimates returned from fscores with the options full.scores = TRUE and full.scores.SE = TRUE

T_as_X

logical; should the observed variance be equal to var(X) = var(T) + E(E^2) or var(X) = var(T) when computing empirical reliability estimates? Default (FALSE) uses the former

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi:10.18637/jss.v048.i06

See also

Author

Phil Chalmers rphilip.chalmers@gmail.com

Examples

# \donttest{

dat <- expand.table(deAyala)
itemstats(dat)
#> $overall
#>      N mean_total.score sd_total.score ave.r  sd.r alpha SEM.alpha
#>  19601            2.912          1.434 0.233 0.074 0.608     0.898
#> 
#> $itemstats
#>            N  mean    sd total.r total.r_if_rm alpha_if_rm
#> Item.1 19601 0.887 0.316   0.447         0.246       0.605
#> Item.2 19601 0.644 0.479   0.688         0.439       0.510
#> Item.3 19601 0.566 0.496   0.680         0.416       0.523
#> Item.4 19601 0.427 0.495   0.673         0.405       0.529
#> Item.5 19601 0.387 0.487   0.602         0.312       0.581
#> 
#> $proportions
#>            0     1
#> Item.1 0.113 0.887
#> Item.2 0.356 0.644
#> Item.3 0.434 0.566
#> Item.4 0.573 0.427
#> Item.5 0.613 0.387
#> 
mod <- mirt(dat)
#> Iteration: 1, Log-Lik: -56256.669, Max-Change: 0.33820Iteration: 2, Log-Lik: -55598.944, Max-Change: 0.25622Iteration: 3, Log-Lik: -55344.183, Max-Change: 0.17901Iteration: 4, Log-Lik: -55252.392, Max-Change: 0.12170Iteration: 5, Log-Lik: -55219.175, Max-Change: 0.08221Iteration: 6, Log-Lik: -55206.845, Max-Change: 0.05582Iteration: 7, Log-Lik: -55202.096, Max-Change: 0.03866Iteration: 8, Log-Lik: -55200.153, Max-Change: 0.02623Iteration: 9, Log-Lik: -55199.330, Max-Change: 0.01879Iteration: 10, Log-Lik: -55198.648, Max-Change: 0.00688Iteration: 11, Log-Lik: -55198.588, Max-Change: 0.00523Iteration: 12, Log-Lik: -55198.554, Max-Change: 0.00418Iteration: 13, Log-Lik: -55198.501, Max-Change: 0.00094Iteration: 14, Log-Lik: -55198.498, Max-Change: 0.00025Iteration: 15, Log-Lik: -55198.497, Max-Change: 0.00024Iteration: 16, Log-Lik: -55198.497, Max-Change: 0.00021Iteration: 17, Log-Lik: -55198.497, Max-Change: 0.00096Iteration: 18, Log-Lik: -55198.496, Max-Change: 0.00057Iteration: 19, Log-Lik: -55198.496, Max-Change: 0.00015Iteration: 20, Log-Lik: -55198.496, Max-Change: 0.00007

theta_se <- fscores(mod, full.scores.SE = TRUE)
empirical_rxx(theta_se)
#>        F1 
#> 0.6200703 

theta_se <- fscores(mod, full.scores.SE = TRUE, method = 'ML')
empirical_rxx(theta_se)
#>        F1 
#> 0.5636644 
empirical_rxx(theta_se, T_as_X = TRUE)
#>        F1 
#> 0.2258948 

# }