expand.table {mirt}	R Documentation

Expand summary table of patterns and frequencies

Description

The expand.table function expands a summary table of unique response patterns to a full sized data-set. The response frequencies must be on the rightmost column of the input data.

Usage

expand.table(tabdata)

Arguments

tabdata
An object of class data.frame, matrix, or table with the unique response patterns and the number of frequencies in the rightmost column.

Value

Returns a numeric matrix with all the response patterns.

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

Examples

## No test: 
data(LSAT7)
head(LSAT7)

##   Item.1 Item.2 Item.3 Item.4 Item.5 freq
## 1      0      0      0      0      0   12
## 2      0      0      0      0      1   19
## 3      0      0      0      1      0    1
## 4      0      0      0      1      1    7
## 5      0      0      1      0      0    3
## 6      0      0      1      0      1   19

LSAT7full <- expand.table(LSAT7)
head(LSAT7full)

##      Item.1 Item.2 Item.3 Item.4 Item.5
## [1,]      0      0      0      0      0
## [2,]      0      0      0      0      0
## [3,]      0      0      0      0      0
## [4,]      0      0      0      0      0
## [5,]      0      0      0      0      0
## [6,]      0      0      0      0      0

## End(No test)

[Package mirt version 1.14 Index]