expand.table {mirt}	R Documentation

Expand summary table of patterns and frequencies

Description

The expand.table function expands a summary table of unique response patterns to a full sized data-set. The response frequencies must be on the rightmost column of the input data.

Usage

expand.table(tabdata, sample = FALSE)

Arguments

tabdata
An object of class data.frame or matrix with the unique response patterns and the number of frequencies in the rightmost column

sample
logical; randomly switch the rows in the expanded table? This does not change the expanded data, only the row locations

Value

Returns a numeric matrix with all the response patterns.

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi: 10.18637/jss.v048.i06

Examples

## No test: 
data(LSAT7)
head(LSAT7)

##   Item.1 Item.2 Item.3 Item.4 Item.5 freq
## 1      0      0      0      0      0   12
## 2      0      0      0      0      1   19
## 3      0      0      0      1      0    1
## 4      0      0      0      1      1    7
## 5      0      0      1      0      0    3
## 6      0      0      1      0      1   19

LSAT7full <- expand.table(LSAT7)
head(LSAT7full)

##   Item.1 Item.2 Item.3 Item.4 Item.5
## 1      0      0      0      0      0
## 2      0      0      0      0      0
## 3      0      0      0      0      0
## 4      0      0      0      0      0
## 5      0      0      0      0      0
## 6      0      0      0      0      0

LSAT7full <- expand.table(LSAT7, sample = TRUE)
head(LSAT7full)

##   Item.1 Item.2 Item.3 Item.4 Item.5
## 1      1      1      0      0      1
## 2      0      1      0      1      1
## 3      1      1      1      1      1
## 4      1      1      0      0      1
## 5      1      1      0      1      1
## 6      1      1      0      1      1

## End(No test)

[Package mirt version 1.25 Index]