Function to calculate expected value of item

Given an internal mirt object extracted from an estimated model compute the expected value for an item given the ability parameter(s).

Usage

expected.item(x, Theta, min = 0, include.var = FALSE)

Arguments

x

an extracted internal mirt object containing item information (see extract.item)

Theta

a vector (unidimensional) or matrix (multidimensional) of latent trait values

min

a constant value added to the expected values indicating the lowest theoretical category. Default is 0

include.var

logical; include the model-implied variance of the expected scores as well? When TRUE will return a list containing the expected values (E) and variances (VAR)

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi:10.18637/jss.v048.i06

See also

Author

Phil Chalmers rphilip.chalmers@gmail.com

Examples

mod <- mirt(Science, 1)
#> Error in make.mixed.design(item.formula = item.formula, item.design = item.design,     data = data): object 'itemdesign' not found
extr.2 <- extract.item(mod, 2)
#> Error: object 'mod' not found
Theta <- matrix(seq(-6,6, length.out=200))
expected <- expected.item(extr.2, Theta, min(Science[,1])) #min() of first item
#> Warning: restarting interrupted promise evaluation
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'ProbTrace': object 'extr.2' not found
head(data.frame(expected, Theta=Theta))
#> Error: object 'expected' not found

expected.item(extr.2, Theta, min(Science[,1]), include.var=TRUE)
#> Warning: restarting interrupted promise evaluation
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'ProbTrace': object 'extr.2' not found