expected.item {mirt}	R Documentation

Function to calculate expected value of item

Description

Given an internal mirt object extracted from an estimated model compute the expected value for an item given the ability parameter(s).

Usage

expected.item(x, Theta, min = 0)

Arguments

x
an extracted internal mirt object containing item information

Theta
a vector (unidimensional) or matrix (multidimensional) of latent trait values

min
a constant value added to the expected values indicating the lowest theoretical category. Default is 0

See Also

extract.item, , expected.test

Examples

## No test: 
mod <- mirt(Science, 1)
extr.2 <- extract.item(mod, 2)
Theta <- matrix(seq(-6,6, length.out=200))
expected <- expected.item(extr.2, Theta, min(Science[,1])) #min() of first item
head(data.frame(expected, Theta=Theta))

##   expected     Theta
## 1 1.013398 -6.000000
## 2 1.014415 -5.939698
## 3 1.015507 -5.879397
## 4 1.016681 -5.819095
## 5 1.017943 -5.758794
## 6 1.019299 -5.698492

## End(No test)

[Package mirt version 1.9 Index]