expected.test {mirt}	R Documentation

Function to calculate expected test score

Description

Given an estimated model compute the expected test score. Returns the expected values in the same form as the data used to estimate the model.

Usage

expected.test(x, Theta, group = NULL, mins = TRUE)

Arguments

x
an estimated mirt object

Theta
a matrix of latent trait values

group
a number signifying which group the item should be extracted from (applies to 'MultipleGroupClass' objects only)

mins
logical; include the minimum value constants in the dataset. If FALSE, the expected values for each item are determined from the scoring 0:(ncat-1)

See Also

expected.item

Examples

## No test: 
dat <- expand.table(deAyala)
model <- mirt.model('F = 1-5
                    CONSTRAIN = (1-5, a1)')
mod <- mirt(dat, model)

Theta <- matrix(seq(-6,6,.01))
tscore <- expected.test(mod, Theta)
tail(cbind(Theta, tscore))

##                tscore
## [1196,] 5.95 4.999138
## [1197,] 5.96 4.999150
## [1198,] 5.97 4.999162
## [1199,] 5.98 4.999174
## [1200,] 5.99 4.999186
## [1201,] 6.00 4.999198

## End(No test)

[Package mirt version 1.9 Index]