extract.item {mirt}	R Documentation

Extract an item object from mirt objects

Description

Extract the internal mirt objects from any estimated model.

Usage

extract.item(x, item, group = NULL, drop.zeros = FALSE)

Arguments

x
mirt model of class 'SingleGroupClass' or 'MultipleGroupClass'

item
a number or character signifying which item to extract

group
a number signifying which group the item should be extracted from (applies to 'MultipleGroupClass' only)

drop.zeros
logical; drop slope values that are numerically close to zero to reduce dimensionality? Useful in objects returned from bfactor or other confirmatory models that contain several zero slopes

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi: 10.18637/jss.v048.i06

See Also

extract.group, extract.mirt

Examples

## No test: 
mod <- mirt(Science, 1)
extr.1 <- extract.item(mod, 1)
## End(No test)

[Package mirt version 1.27.1 Index]