Extract various elements from estimated model objects

A generic function to extract the internal objects from estimated models.

Usage

extract.mirt(x, what, item = 1)

Arguments

x

mirt model of class 'SingleGroupClass', 'MultipleGroupClass', 'MixedClass' or 'DiscreteGroupClass'

what

a string indicating what to extract

item

if necessary, which item to extract information from. Defaults to 1 if not specified

Details

Objects which can be extracted from mirt objects include:

logLik

observed log-likelihood

logPrior

log term contributed by prior parameter distributions

G2

goodness of fit statistic

df

degrees of freedom

p

p-value for G2 statistic

RMSEA

root mean-square error of approximation based on G2

CFI

CFI fit statistic

TLI

TLI fit statistic

AIC

AIC

BIC

BIC

SABIC

sample size adjusted BIC

HQ

HQ

LLhistory

EM log-likelihood history

tabdata

a tabular version of the raw response data input. Frequencies are stored in freq

freq

frequencies associated with tabdata

K

an integer vector indicating the number of unique elements for each item

mins

an integer vector indicating the lowest category found in the input data

model

input model syntax

method

estimation method used

itemtype

a vector of item types for each respective item (e.g., 'graded', '2PL', etc)

itemnames

a vector of item names from the input data

factorNames

a vector of factor names from the model definition

rowID

an integer vector indicating all valid row numbers used in the model estimation (when all cases are used this will be 1:nrow(data))

data

raw input data of item responses

covdata

raw input data of data used as covariates

tabdatalong

similar to tabdata, however the responses have been transformed into dummy coded variables

fulldatalong

analogous to tabdatafull, but for the raw input data instead of the tabulated frequencies

EMhistory

if saved, extract the EM iteration history

exp_resp

expected probability of the unique response patterns

survey.weights

if supplied, the vector of survey weights used during estimation (NULL if missing)

converged

a logical value indicating whether the model terminated within the convergence criteria

iterations

number of iterations it took to reach the convergence criteria

nest

number of freely estimated parameters

parvec

vector containing uniquely estimated parameters

vcov

parameter covariance matrix (associated with parvec)

condnum

the condition number of the Hessian (if computed). Otherwise NA

constrain

a list of item parameter constraints to indicate which item parameters were equal during estimation

Prior

prior density distribution for the latent traits

thetaPosterior

posterior distribution for latent traits when using EM algorithm

key

if supplied, the data scoring key

nfact

number of latent traits/factors

nitems

number of items

ngroups

number of groups

groupNames

character vector of unique group names

group

a character vector indicating the group membership

invariance

a character vector indicating invariance input from multipleGroup

secondordertest

a logical indicating whether the model passed the second-order test based on the Hessian matrix. Indicates whether model is a potential local maximum solution

SEMconv

logical; check whether the supplemented EM information matrix converged. Will be NA if not applicable

time

estimation time, broken into different sections

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi:10.18637/jss.v048.i06

See also

Author

Phil Chalmers rphilip.chalmers@gmail.com

Examples

# \donttest{
mod <- mirt(Science, 1)
#> Iteration: 1, Log-Lik: -1629.361, Max-Change: 0.50660Iteration: 2, Log-Lik: -1617.374, Max-Change: 0.25442Iteration: 3, Log-Lik: -1612.894, Max-Change: 0.16991Iteration: 4, Log-Lik: -1610.306, Max-Change: 0.10461Iteration: 5, Log-Lik: -1609.814, Max-Change: 0.09162Iteration: 6, Log-Lik: -1609.534, Max-Change: 0.07363Iteration: 7, Log-Lik: -1609.030, Max-Change: 0.03677Iteration: 8, Log-Lik: -1608.988, Max-Change: 0.03200Iteration: 9, Log-Lik: -1608.958, Max-Change: 0.02754Iteration: 10, Log-Lik: -1608.878, Max-Change: 0.01443Iteration: 11, Log-Lik: -1608.875, Max-Change: 0.00847Iteration: 12, Log-Lik: -1608.873, Max-Change: 0.00515Iteration: 13, Log-Lik: -1608.872, Max-Change: 0.00550Iteration: 14, Log-Lik: -1608.872, Max-Change: 0.00318Iteration: 15, Log-Lik: -1608.871, Max-Change: 0.00462Iteration: 16, Log-Lik: -1608.871, Max-Change: 0.00277Iteration: 17, Log-Lik: -1608.870, Max-Change: 0.00145Iteration: 18, Log-Lik: -1608.870, Max-Change: 0.00175Iteration: 19, Log-Lik: -1608.870, Max-Change: 0.00126Iteration: 20, Log-Lik: -1608.870, Max-Change: 0.00025Iteration: 21, Log-Lik: -1608.870, Max-Change: 0.00285Iteration: 22, Log-Lik: -1608.870, Max-Change: 0.00108Iteration: 23, Log-Lik: -1608.870, Max-Change: 0.00022Iteration: 24, Log-Lik: -1608.870, Max-Change: 0.00059Iteration: 25, Log-Lik: -1608.870, Max-Change: 0.00014Iteration: 26, Log-Lik: -1608.870, Max-Change: 0.00068Iteration: 27, Log-Lik: -1608.870, Max-Change: 0.00065Iteration: 28, Log-Lik: -1608.870, Max-Change: 0.00019Iteration: 29, Log-Lik: -1608.870, Max-Change: 0.00061Iteration: 30, Log-Lik: -1608.870, Max-Change: 0.00012Iteration: 31, Log-Lik: -1608.870, Max-Change: 0.00012Iteration: 32, Log-Lik: -1608.870, Max-Change: 0.00058Iteration: 33, Log-Lik: -1608.870, Max-Change: 0.00055Iteration: 34, Log-Lik: -1608.870, Max-Change: 0.00015Iteration: 35, Log-Lik: -1608.870, Max-Change: 0.00052Iteration: 36, Log-Lik: -1608.870, Max-Change: 0.00010

extract.mirt(mod, 'logLik')
#> [1] -1608.87
extract.mirt(mod, 'K')  # unique categories for each item
#> [1] 4 4 4 4

#multiple group model
grp <- rep(c('G1', 'G2'), each = nrow(Science)/2)
mod2 <- multipleGroup(Science, 1, grp)
#> Iteration: 1, Log-Lik: -1629.361, Max-Change: 1.10657Iteration: 2, Log-Lik: -1609.594, Max-Change: 0.44262Iteration: 3, Log-Lik: -1603.139, Max-Change: 0.33923Iteration: 4, Log-Lik: -1600.524, Max-Change: 0.23152Iteration: 5, Log-Lik: -1599.314, Max-Change: 0.17794Iteration: 6, Log-Lik: -1598.777, Max-Change: 0.13751Iteration: 7, Log-Lik: -1598.257, Max-Change: 0.03517Iteration: 8, Log-Lik: -1598.189, Max-Change: 0.03698Iteration: 9, Log-Lik: -1598.136, Max-Change: 0.03742Iteration: 10, Log-Lik: -1597.913, Max-Change: 0.04062Iteration: 11, Log-Lik: -1597.889, Max-Change: 0.03424Iteration: 12, Log-Lik: -1597.869, Max-Change: 0.03294Iteration: 13, Log-Lik: -1597.778, Max-Change: 0.02383Iteration: 14, Log-Lik: -1597.768, Max-Change: 0.02590Iteration: 15, Log-Lik: -1597.760, Max-Change: 0.02474Iteration: 16, Log-Lik: -1597.720, Max-Change: 0.02137Iteration: 17, Log-Lik: -1597.715, Max-Change: 0.01882Iteration: 18, Log-Lik: -1597.712, Max-Change: 0.01744Iteration: 19, Log-Lik: -1597.693, Max-Change: 0.01737Iteration: 20, Log-Lik: -1597.691, Max-Change: 0.01382Iteration: 21, Log-Lik: -1597.689, Max-Change: 0.01264Iteration: 22, Log-Lik: -1597.680, Max-Change: 0.01199Iteration: 23, Log-Lik: -1597.678, Max-Change: 0.01097Iteration: 24, Log-Lik: -1597.677, Max-Change: 0.01019Iteration: 25, Log-Lik: -1597.672, Max-Change: 0.00935Iteration: 26, Log-Lik: -1597.671, Max-Change: 0.00864Iteration: 27, Log-Lik: -1597.670, Max-Change: 0.00752Iteration: 28, Log-Lik: -1597.667, Max-Change: 0.00912Iteration: 29, Log-Lik: -1597.667, Max-Change: 0.00661Iteration: 30, Log-Lik: -1597.666, Max-Change: 0.00605Iteration: 31, Log-Lik: -1597.665, Max-Change: 0.00931Iteration: 32, Log-Lik: -1597.664, Max-Change: 0.00637Iteration: 33, Log-Lik: -1597.663, Max-Change: 0.00500Iteration: 34, Log-Lik: -1597.663, Max-Change: 0.00587Iteration: 35, Log-Lik: -1597.662, Max-Change: 0.00520Iteration: 36, Log-Lik: -1597.662, Max-Change: 0.00395Iteration: 37, Log-Lik: -1597.662, Max-Change: 0.00206Iteration: 38, Log-Lik: -1597.662, Max-Change: 0.00483Iteration: 39, Log-Lik: -1597.661, Max-Change: 0.00110Iteration: 40, Log-Lik: -1597.661, Max-Change: 0.00436Iteration: 41, Log-Lik: -1597.661, Max-Change: 0.00200Iteration: 42, Log-Lik: -1597.661, Max-Change: 0.00362Iteration: 43, Log-Lik: -1597.661, Max-Change: 0.00585Iteration: 44, Log-Lik: -1597.661, Max-Change: 0.00115Iteration: 45, Log-Lik: -1597.661, Max-Change: 0.00069Iteration: 46, Log-Lik: -1597.661, Max-Change: 0.00062Iteration: 47, Log-Lik: -1597.661, Max-Change: 0.00067Iteration: 48, Log-Lik: -1597.661, Max-Change: 0.00063Iteration: 49, Log-Lik: -1597.661, Max-Change: 0.00068Iteration: 50, Log-Lik: -1597.661, Max-Change: 0.00076Iteration: 51, Log-Lik: -1597.661, Max-Change: 0.00110Iteration: 52, Log-Lik: -1597.661, Max-Change: 0.00064Iteration: 53, Log-Lik: -1597.661, Max-Change: 0.00092Iteration: 54, Log-Lik: -1597.661, Max-Change: 0.00025Iteration: 55, Log-Lik: -1597.661, Max-Change: 0.00112Iteration: 56, Log-Lik: -1597.661, Max-Change: 0.00758Iteration: 57, Log-Lik: -1597.660, Max-Change: 0.00125Iteration: 58, Log-Lik: -1597.660, Max-Change: 0.00113Iteration: 59, Log-Lik: -1597.660, Max-Change: 0.00051Iteration: 60, Log-Lik: -1597.660, Max-Change: 0.00057Iteration: 61, Log-Lik: -1597.660, Max-Change: 0.00057Iteration: 62, Log-Lik: -1597.660, Max-Change: 0.00070Iteration: 63, Log-Lik: -1597.660, Max-Change: 0.00092Iteration: 64, Log-Lik: -1597.660, Max-Change: 0.00060Iteration: 65, Log-Lik: -1597.660, Max-Change: 0.00067Iteration: 66, Log-Lik: -1597.660, Max-Change: 0.00083Iteration: 67, Log-Lik: -1597.660, Max-Change: 0.00060Iteration: 68, Log-Lik: -1597.660, Max-Change: 0.00066Iteration: 69, Log-Lik: -1597.660, Max-Change: 0.00090Iteration: 70, Log-Lik: -1597.660, Max-Change: 0.00061Iteration: 71, Log-Lik: -1597.660, Max-Change: 0.00075Iteration: 72, Log-Lik: -1597.660, Max-Change: 0.00107Iteration: 73, Log-Lik: -1597.660, Max-Change: 0.00060Iteration: 74, Log-Lik: -1597.660, Max-Change: 0.00086Iteration: 75, Log-Lik: -1597.660, Max-Change: 0.00024Iteration: 76, Log-Lik: -1597.660, Max-Change: 0.00106Iteration: 77, Log-Lik: -1597.660, Max-Change: 0.00030Iteration: 78, Log-Lik: -1597.660, Max-Change: 0.00079Iteration: 79, Log-Lik: -1597.660, Max-Change: 0.00334Iteration: 80, Log-Lik: -1597.660, Max-Change: 0.00179Iteration: 81, Log-Lik: -1597.660, Max-Change: 0.00081Iteration: 82, Log-Lik: -1597.660, Max-Change: 0.00280Iteration: 83, Log-Lik: -1597.660, Max-Change: 0.00173Iteration: 84, Log-Lik: -1597.660, Max-Change: 0.00079Iteration: 85, Log-Lik: -1597.660, Max-Change: 0.00022Iteration: 86, Log-Lik: -1597.660, Max-Change: 0.00059Iteration: 87, Log-Lik: -1597.660, Max-Change: 0.00079Iteration: 88, Log-Lik: -1597.660, Max-Change: 0.00054Iteration: 89, Log-Lik: -1597.660, Max-Change: 0.00067Iteration: 90, Log-Lik: -1597.660, Max-Change: 0.00092Iteration: 91, Log-Lik: -1597.660, Max-Change: 0.00057Iteration: 92, Log-Lik: -1597.660, Max-Change: 0.00076Iteration: 93, Log-Lik: -1597.660, Max-Change: 0.00107Iteration: 94, Log-Lik: -1597.660, Max-Change: 0.00059Iteration: 95, Log-Lik: -1597.660, Max-Change: 0.00084Iteration: 96, Log-Lik: -1597.660, Max-Change: 0.00023Iteration: 97, Log-Lik: -1597.660, Max-Change: 0.00021Iteration: 98, Log-Lik: -1597.660, Max-Change: 0.00054Iteration: 99, Log-Lik: -1597.660, Max-Change: 0.00077Iteration: 100, Log-Lik: -1597.660, Max-Change: 0.00055Iteration: 101, Log-Lik: -1597.660, Max-Change: 0.00078Iteration: 102, Log-Lik: -1597.660, Max-Change: 0.00021Iteration: 103, Log-Lik: -1597.660, Max-Change: 0.00096Iteration: 104, Log-Lik: -1597.660, Max-Change: 0.00027Iteration: 105, Log-Lik: -1597.660, Max-Change: 0.00071Iteration: 106, Log-Lik: -1597.660, Max-Change: 0.00295Iteration: 107, Log-Lik: -1597.660, Max-Change: 0.00150Iteration: 108, Log-Lik: -1597.660, Max-Change: 0.00054Iteration: 109, Log-Lik: -1597.660, Max-Change: 0.00435Iteration: 110, Log-Lik: -1597.659, Max-Change: 0.00083Iteration: 111, Log-Lik: -1597.659, Max-Change: 0.00072Iteration: 112, Log-Lik: -1597.659, Max-Change: 0.00047Iteration: 113, Log-Lik: -1597.659, Max-Change: 0.00053Iteration: 114, Log-Lik: -1597.659, Max-Change: 0.00072Iteration: 115, Log-Lik: -1597.659, Max-Change: 0.00053Iteration: 116, Log-Lik: -1597.659, Max-Change: 0.00073Iteration: 117, Log-Lik: -1597.659, Max-Change: 0.00021Iteration: 118, Log-Lik: -1597.659, Max-Change: 0.00092Iteration: 119, Log-Lik: -1597.659, Max-Change: 0.00026Iteration: 120, Log-Lik: -1597.659, Max-Change: 0.00066Iteration: 121, Log-Lik: -1597.659, Max-Change: 0.00307Iteration: 122, Log-Lik: -1597.659, Max-Change: 0.00111Iteration: 123, Log-Lik: -1597.659, Max-Change: 0.00044Iteration: 124, Log-Lik: -1597.659, Max-Change: 0.00416Iteration: 125, Log-Lik: -1597.659, Max-Change: 0.00091Iteration: 126, Log-Lik: -1597.659, Max-Change: 0.00216Iteration: 127, Log-Lik: -1597.659, Max-Change: 0.00062Iteration: 128, Log-Lik: -1597.659, Max-Change: 0.00065Iteration: 129, Log-Lik: -1597.659, Max-Change: 0.00017Iteration: 130, Log-Lik: -1597.659, Max-Change: 0.00078Iteration: 131, Log-Lik: -1597.659, Max-Change: 0.00022Iteration: 132, Log-Lik: -1597.659, Max-Change: 0.00057Iteration: 133, Log-Lik: -1597.659, Max-Change: 0.00029Iteration: 134, Log-Lik: -1597.659, Max-Change: 0.00076Iteration: 135, Log-Lik: -1597.659, Max-Change: 0.00021Iteration: 136, Log-Lik: -1597.659, Max-Change: 0.00019Iteration: 137, Log-Lik: -1597.659, Max-Change: 0.00048Iteration: 138, Log-Lik: -1597.659, Max-Change: 0.00069Iteration: 139, Log-Lik: -1597.659, Max-Change: 0.00045Iteration: 140, Log-Lik: -1597.659, Max-Change: 0.00065Iteration: 141, Log-Lik: -1597.659, Max-Change: 0.00018Iteration: 142, Log-Lik: -1597.659, Max-Change: 0.00079Iteration: 143, Log-Lik: -1597.659, Max-Change: 0.00022Iteration: 144, Log-Lik: -1597.659, Max-Change: 0.00059Iteration: 145, Log-Lik: -1597.659, Max-Change: 0.00029Iteration: 146, Log-Lik: -1597.659, Max-Change: 0.00074Iteration: 147, Log-Lik: -1597.659, Max-Change: 0.00021Iteration: 148, Log-Lik: -1597.659, Max-Change: 0.00019Iteration: 149, Log-Lik: -1597.659, Max-Change: 0.00049Iteration: 150, Log-Lik: -1597.659, Max-Change: 0.00067Iteration: 151, Log-Lik: -1597.659, Max-Change: 0.00046Iteration: 152, Log-Lik: -1597.659, Max-Change: 0.00064Iteration: 153, Log-Lik: -1597.659, Max-Change: 0.00018Iteration: 154, Log-Lik: -1597.659, Max-Change: 0.00081Iteration: 155, Log-Lik: -1597.659, Max-Change: 0.00022Iteration: 156, Log-Lik: -1597.659, Max-Change: 0.00058Iteration: 157, Log-Lik: -1597.659, Max-Change: 0.00029Iteration: 158, Log-Lik: -1597.659, Max-Change: 0.00076Iteration: 159, Log-Lik: -1597.659, Max-Change: 0.00021Iteration: 160, Log-Lik: -1597.659, Max-Change: 0.00019Iteration: 161, Log-Lik: -1597.659, Max-Change: 0.00048Iteration: 162, Log-Lik: -1597.659, Max-Change: 0.00069Iteration: 163, Log-Lik: -1597.659, Max-Change: 0.00045Iteration: 164, Log-Lik: -1597.659, Max-Change: 0.00064Iteration: 165, Log-Lik: -1597.659, Max-Change: 0.00018Iteration: 166, Log-Lik: -1597.659, Max-Change: 0.00016Iteration: 167, Log-Lik: -1597.659, Max-Change: 0.00041Iteration: 168, Log-Lik: -1597.659, Max-Change: 0.00058Iteration: 169, Log-Lik: -1597.659, Max-Change: 0.00043Iteration: 170, Log-Lik: -1597.659, Max-Change: 0.00061Iteration: 171, Log-Lik: -1597.659, Max-Change: 0.00017Iteration: 172, Log-Lik: -1597.659, Max-Change: 0.00076Iteration: 173, Log-Lik: -1597.659, Max-Change: 0.00021Iteration: 174, Log-Lik: -1597.659, Max-Change: 0.00056Iteration: 175, Log-Lik: -1597.659, Max-Change: 0.00027Iteration: 176, Log-Lik: -1597.659, Max-Change: 0.00071Iteration: 177, Log-Lik: -1597.659, Max-Change: 0.00020Iteration: 178, Log-Lik: -1597.659, Max-Change: 0.00018Iteration: 179, Log-Lik: -1597.659, Max-Change: 0.00047Iteration: 180, Log-Lik: -1597.659, Max-Change: 0.00064Iteration: 181, Log-Lik: -1597.659, Max-Change: 0.00044Iteration: 182, Log-Lik: -1597.659, Max-Change: 0.00061Iteration: 183, Log-Lik: -1597.659, Max-Change: 0.00017Iteration: 184, Log-Lik: -1597.659, Max-Change: 0.00077Iteration: 185, Log-Lik: -1597.659, Max-Change: 0.00021Iteration: 186, Log-Lik: -1597.659, Max-Change: 0.00055Iteration: 187, Log-Lik: -1597.659, Max-Change: 0.00028Iteration: 188, Log-Lik: -1597.659, Max-Change: 0.00072Iteration: 189, Log-Lik: -1597.659, Max-Change: 0.00020Iteration: 190, Log-Lik: -1597.659, Max-Change: 0.00018Iteration: 191, Log-Lik: -1597.659, Max-Change: 0.00046Iteration: 192, Log-Lik: -1597.659, Max-Change: 0.00066Iteration: 193, Log-Lik: -1597.659, Max-Change: 0.00042Iteration: 194, Log-Lik: -1597.659, Max-Change: 0.00061Iteration: 195, Log-Lik: -1597.659, Max-Change: 0.00017Iteration: 196, Log-Lik: -1597.659, Max-Change: 0.00015Iteration: 197, Log-Lik: -1597.659, Max-Change: 0.00039Iteration: 198, Log-Lik: -1597.659, Max-Change: 0.00055Iteration: 199, Log-Lik: -1597.659, Max-Change: 0.00040Iteration: 200, Log-Lik: -1597.659, Max-Change: 0.00058Iteration: 201, Log-Lik: -1597.659, Max-Change: 0.00016Iteration: 202, Log-Lik: -1597.659, Max-Change: 0.00071Iteration: 203, Log-Lik: -1597.659, Max-Change: 0.00020Iteration: 204, Log-Lik: -1597.659, Max-Change: 0.00053Iteration: 205, Log-Lik: -1597.659, Max-Change: 0.00026Iteration: 206, Log-Lik: -1597.659, Max-Change: 0.00067Iteration: 207, Log-Lik: -1597.659, Max-Change: 0.00019Iteration: 208, Log-Lik: -1597.659, Max-Change: 0.00017Iteration: 209, Log-Lik: -1597.659, Max-Change: 0.00044Iteration: 210, Log-Lik: -1597.659, Max-Change: 0.00061Iteration: 211, Log-Lik: -1597.659, Max-Change: 0.00041Iteration: 212, Log-Lik: -1597.659, Max-Change: 0.00057Iteration: 213, Log-Lik: -1597.659, Max-Change: 0.00016Iteration: 214, Log-Lik: -1597.659, Max-Change: 0.00072Iteration: 215, Log-Lik: -1597.659, Max-Change: 0.00020Iteration: 216, Log-Lik: -1597.659, Max-Change: 0.00052Iteration: 217, Log-Lik: -1597.659, Max-Change: 0.00026Iteration: 218, Log-Lik: -1597.659, Max-Change: 0.00068Iteration: 219, Log-Lik: -1597.659, Max-Change: 0.00019Iteration: 220, Log-Lik: -1597.659, Max-Change: 0.00017Iteration: 221, Log-Lik: -1597.659, Max-Change: 0.00044Iteration: 222, Log-Lik: -1597.659, Max-Change: 0.00062Iteration: 223, Log-Lik: -1597.659, Max-Change: 0.00040Iteration: 224, Log-Lik: -1597.659, Max-Change: 0.00057Iteration: 225, Log-Lik: -1597.659, Max-Change: 0.00016Iteration: 226, Log-Lik: -1597.659, Max-Change: 0.00014Iteration: 227, Log-Lik: -1597.659, Max-Change: 0.00036Iteration: 228, Log-Lik: -1597.659, Max-Change: 0.00052Iteration: 229, Log-Lik: -1597.659, Max-Change: 0.00038Iteration: 230, Log-Lik: -1597.659, Max-Change: 0.00054Iteration: 231, Log-Lik: -1597.659, Max-Change: 0.00015Iteration: 232, Log-Lik: -1597.659, Max-Change: 0.00067Iteration: 233, Log-Lik: -1597.659, Max-Change: 0.00019Iteration: 234, Log-Lik: -1597.659, Max-Change: 0.00049Iteration: 235, Log-Lik: -1597.659, Max-Change: 0.00024Iteration: 236, Log-Lik: -1597.658, Max-Change: 0.00063Iteration: 237, Log-Lik: -1597.658, Max-Change: 0.00018Iteration: 238, Log-Lik: -1597.658, Max-Change: 0.00016Iteration: 239, Log-Lik: -1597.658, Max-Change: 0.00041Iteration: 240, Log-Lik: -1597.658, Max-Change: 0.00057Iteration: 241, Log-Lik: -1597.658, Max-Change: 0.00039Iteration: 242, Log-Lik: -1597.658, Max-Change: 0.00054Iteration: 243, Log-Lik: -1597.658, Max-Change: 0.00015Iteration: 244, Log-Lik: -1597.658, Max-Change: 0.00068Iteration: 245, Log-Lik: -1597.658, Max-Change: 0.00019Iteration: 246, Log-Lik: -1597.658, Max-Change: 0.00049Iteration: 247, Log-Lik: -1597.658, Max-Change: 0.00025Iteration: 248, Log-Lik: -1597.658, Max-Change: 0.00064Iteration: 249, Log-Lik: -1597.658, Max-Change: 0.00018Iteration: 250, Log-Lik: -1597.658, Max-Change: 0.00016Iteration: 251, Log-Lik: -1597.658, Max-Change: 0.00041Iteration: 252, Log-Lik: -1597.658, Max-Change: 0.00058Iteration: 253, Log-Lik: -1597.658, Max-Change: 0.00038Iteration: 254, Log-Lik: -1597.658, Max-Change: 0.00054Iteration: 255, Log-Lik: -1597.658, Max-Change: 0.00015Iteration: 256, Log-Lik: -1597.658, Max-Change: 0.00013Iteration: 257, Log-Lik: -1597.658, Max-Change: 0.00034Iteration: 258, Log-Lik: -1597.658, Max-Change: 0.00049Iteration: 259, Log-Lik: -1597.658, Max-Change: 0.00036Iteration: 260, Log-Lik: -1597.658, Max-Change: 0.00051Iteration: 261, Log-Lik: -1597.658, Max-Change: 0.00014Iteration: 262, Log-Lik: -1597.658, Max-Change: 0.00063Iteration: 263, Log-Lik: -1597.658, Max-Change: 0.00018Iteration: 264, Log-Lik: -1597.658, Max-Change: 0.00046Iteration: 265, Log-Lik: -1597.658, Max-Change: 0.00023Iteration: 266, Log-Lik: -1597.658, Max-Change: 0.00059Iteration: 267, Log-Lik: -1597.658, Max-Change: 0.00017Iteration: 268, Log-Lik: -1597.658, Max-Change: 0.00015Iteration: 269, Log-Lik: -1597.658, Max-Change: 0.00039Iteration: 270, Log-Lik: -1597.658, Max-Change: 0.00054Iteration: 271, Log-Lik: -1597.658, Max-Change: 0.00037Iteration: 272, Log-Lik: -1597.658, Max-Change: 0.00050Iteration: 273, Log-Lik: -1597.658, Max-Change: 0.00014Iteration: 274, Log-Lik: -1597.658, Max-Change: 0.00064Iteration: 275, Log-Lik: -1597.658, Max-Change: 0.00018Iteration: 276, Log-Lik: -1597.658, Max-Change: 0.00046Iteration: 277, Log-Lik: -1597.658, Max-Change: 0.00023Iteration: 278, Log-Lik: -1597.658, Max-Change: 0.00060Iteration: 279, Log-Lik: -1597.658, Max-Change: 0.00017Iteration: 280, Log-Lik: -1597.658, Max-Change: 0.00015Iteration: 281, Log-Lik: -1597.658, Max-Change: 0.00039Iteration: 282, Log-Lik: -1597.658, Max-Change: 0.00055Iteration: 283, Log-Lik: -1597.658, Max-Change: 0.00035Iteration: 284, Log-Lik: -1597.658, Max-Change: 0.00050Iteration: 285, Log-Lik: -1597.658, Max-Change: 0.00014Iteration: 286, Log-Lik: -1597.658, Max-Change: 0.00012Iteration: 287, Log-Lik: -1597.658, Max-Change: 0.00032Iteration: 288, Log-Lik: -1597.658, Max-Change: 0.00046Iteration: 289, Log-Lik: -1597.658, Max-Change: 0.00034Iteration: 290, Log-Lik: -1597.658, Max-Change: 0.00048Iteration: 291, Log-Lik: -1597.658, Max-Change: 0.00013Iteration: 292, Log-Lik: -1597.658, Max-Change: 0.00059Iteration: 293, Log-Lik: -1597.658, Max-Change: 0.00017Iteration: 294, Log-Lik: -1597.658, Max-Change: 0.00044Iteration: 295, Log-Lik: -1597.658, Max-Change: 0.00021Iteration: 296, Log-Lik: -1597.658, Max-Change: 0.00056Iteration: 297, Log-Lik: -1597.658, Max-Change: 0.00016Iteration: 298, Log-Lik: -1597.658, Max-Change: 0.00014Iteration: 299, Log-Lik: -1597.658, Max-Change: 0.00037Iteration: 300, Log-Lik: -1597.658, Max-Change: 0.00051Iteration: 301, Log-Lik: -1597.658, Max-Change: 0.00034Iteration: 302, Log-Lik: -1597.658, Max-Change: 0.00047Iteration: 303, Log-Lik: -1597.658, Max-Change: 0.00013Iteration: 304, Log-Lik: -1597.658, Max-Change: 0.00060Iteration: 305, Log-Lik: -1597.658, Max-Change: 0.00017Iteration: 306, Log-Lik: -1597.658, Max-Change: 0.00043Iteration: 307, Log-Lik: -1597.658, Max-Change: 0.00022Iteration: 308, Log-Lik: -1597.658, Max-Change: 0.00057Iteration: 309, Log-Lik: -1597.658, Max-Change: 0.00016Iteration: 310, Log-Lik: -1597.658, Max-Change: 0.00014Iteration: 311, Log-Lik: -1597.658, Max-Change: 0.00036Iteration: 312, Log-Lik: -1597.658, Max-Change: 0.00052Iteration: 313, Log-Lik: -1597.658, Max-Change: 0.00033Iteration: 314, Log-Lik: -1597.658, Max-Change: 0.00047Iteration: 315, Log-Lik: -1597.658, Max-Change: 0.00013Iteration: 316, Log-Lik: -1597.658, Max-Change: 0.00012Iteration: 317, Log-Lik: -1597.658, Max-Change: 0.00030Iteration: 318, Log-Lik: -1597.658, Max-Change: 0.00043Iteration: 319, Log-Lik: -1597.658, Max-Change: 0.00032Iteration: 320, Log-Lik: -1597.658, Max-Change: 0.00045Iteration: 321, Log-Lik: -1597.658, Max-Change: 0.00012Iteration: 322, Log-Lik: -1597.658, Max-Change: 0.00056Iteration: 323, Log-Lik: -1597.658, Max-Change: 0.00016Iteration: 324, Log-Lik: -1597.658, Max-Change: 0.00041Iteration: 325, Log-Lik: -1597.658, Max-Change: 0.00020Iteration: 326, Log-Lik: -1597.658, Max-Change: 0.00052Iteration: 327, Log-Lik: -1597.658, Max-Change: 0.00015Iteration: 328, Log-Lik: -1597.658, Max-Change: 0.00013Iteration: 329, Log-Lik: -1597.658, Max-Change: 0.00035Iteration: 330, Log-Lik: -1597.658, Max-Change: 0.00048Iteration: 331, Log-Lik: -1597.658, Max-Change: 0.00032Iteration: 332, Log-Lik: -1597.658, Max-Change: 0.00045Iteration: 333, Log-Lik: -1597.658, Max-Change: 0.00013Iteration: 334, Log-Lik: -1597.658, Max-Change: 0.00057Iteration: 335, Log-Lik: -1597.658, Max-Change: 0.00016Iteration: 336, Log-Lik: -1597.658, Max-Change: 0.00041Iteration: 337, Log-Lik: -1597.658, Max-Change: 0.00021Iteration: 338, Log-Lik: -1597.658, Max-Change: 0.00054Iteration: 339, Log-Lik: -1597.658, Max-Change: 0.00015Iteration: 340, Log-Lik: -1597.658, Max-Change: 0.00013Iteration: 341, Log-Lik: -1597.658, Max-Change: 0.00034Iteration: 342, Log-Lik: -1597.658, Max-Change: 0.00049Iteration: 343, Log-Lik: -1597.658, Max-Change: 0.00031Iteration: 344, Log-Lik: -1597.658, Max-Change: 0.00045Iteration: 345, Log-Lik: -1597.658, Max-Change: 0.00012Iteration: 346, Log-Lik: -1597.658, Max-Change: 0.00011Iteration: 347, Log-Lik: -1597.658, Max-Change: 0.00029Iteration: 348, Log-Lik: -1597.658, Max-Change: 0.00041Iteration: 349, Log-Lik: -1597.658, Max-Change: 0.00030Iteration: 350, Log-Lik: -1597.658, Max-Change: 0.00043Iteration: 351, Log-Lik: -1597.658, Max-Change: 0.00012Iteration: 352, Log-Lik: -1597.658, Max-Change: 0.00052Iteration: 353, Log-Lik: -1597.658, Max-Change: 0.00015Iteration: 354, Log-Lik: -1597.658, Max-Change: 0.00039Iteration: 355, Log-Lik: -1597.658, Max-Change: 0.00019Iteration: 356, Log-Lik: -1597.658, Max-Change: 0.00049Iteration: 357, Log-Lik: -1597.658, Max-Change: 0.00014Iteration: 358, Log-Lik: -1597.658, Max-Change: 0.00012Iteration: 359, Log-Lik: -1597.658, Max-Change: 0.00033Iteration: 360, Log-Lik: -1597.658, Max-Change: 0.00045Iteration: 361, Log-Lik: -1597.658, Max-Change: 0.00031Iteration: 362, Log-Lik: -1597.658, Max-Change: 0.00042Iteration: 363, Log-Lik: -1597.658, Max-Change: 0.00012Iteration: 364, Log-Lik: -1597.658, Max-Change: 0.00053Iteration: 365, Log-Lik: -1597.658, Max-Change: 0.00015Iteration: 366, Log-Lik: -1597.658, Max-Change: 0.00038Iteration: 367, Log-Lik: -1597.658, Max-Change: 0.00019Iteration: 368, Log-Lik: -1597.658, Max-Change: 0.00051Iteration: 369, Log-Lik: -1597.658, Max-Change: 0.00014Iteration: 370, Log-Lik: -1597.658, Max-Change: 0.00013Iteration: 371, Log-Lik: -1597.658, Max-Change: 0.00032Iteration: 372, Log-Lik: -1597.658, Max-Change: 0.00046Iteration: 373, Log-Lik: -1597.658, Max-Change: 0.00030Iteration: 374, Log-Lik: -1597.658, Max-Change: 0.00042Iteration: 375, Log-Lik: -1597.658, Max-Change: 0.00012Iteration: 376, Log-Lik: -1597.658, Max-Change: 0.00010Iteration: 377, Log-Lik: -1597.658, Max-Change: 0.00027Iteration: 378, Log-Lik: -1597.658, Max-Change: 0.00038Iteration: 379, Log-Lik: -1597.658, Max-Change: 0.00028Iteration: 380, Log-Lik: -1597.658, Max-Change: 0.00040Iteration: 381, Log-Lik: -1597.658, Max-Change: 0.00011Iteration: 382, Log-Lik: -1597.658, Max-Change: 0.00049Iteration: 383, Log-Lik: -1597.658, Max-Change: 0.00014Iteration: 384, Log-Lik: -1597.658, Max-Change: 0.00037Iteration: 385, Log-Lik: -1597.658, Max-Change: 0.00018Iteration: 386, Log-Lik: -1597.658, Max-Change: 0.00046Iteration: 387, Log-Lik: -1597.658, Max-Change: 0.00013Iteration: 388, Log-Lik: -1597.658, Max-Change: 0.00012Iteration: 389, Log-Lik: -1597.658, Max-Change: 0.00031Iteration: 390, Log-Lik: -1597.658, Max-Change: 0.00042Iteration: 391, Log-Lik: -1597.658, Max-Change: 0.00029Iteration: 392, Log-Lik: -1597.658, Max-Change: 0.00040Iteration: 393, Log-Lik: -1597.658, Max-Change: 0.00011Iteration: 394, Log-Lik: -1597.658, Max-Change: 0.00050Iteration: 395, Log-Lik: -1597.658, Max-Change: 0.00014Iteration: 396, Log-Lik: -1597.658, Max-Change: 0.00036Iteration: 397, Log-Lik: -1597.658, Max-Change: 0.00018Iteration: 398, Log-Lik: -1597.658, Max-Change: 0.00048Iteration: 399, Log-Lik: -1597.658, Max-Change: 0.00013Iteration: 400, Log-Lik: -1597.658, Max-Change: 0.00012Iteration: 401, Log-Lik: -1597.658, Max-Change: 0.00031Iteration: 402, Log-Lik: -1597.658, Max-Change: 0.00044Iteration: 403, Log-Lik: -1597.658, Max-Change: 0.00028Iteration: 404, Log-Lik: -1597.658, Max-Change: 0.00040Iteration: 405, Log-Lik: -1597.658, Max-Change: 0.00011Iteration: 406, Log-Lik: -1597.658, Max-Change: 0.00010

grp1 <- extract.group(mod2, 1) #extract single group model
extract.mirt(mod2, 'parvec')
#>  [1]  0.8316800  4.8913029  2.5597789 -1.3592138  0.8289135  2.4772334
#>  [7]  0.7025447 -2.1135580  2.8809926  5.4820292  2.2126281 -2.5451325
#> [13]  0.6920437  3.0133748  0.8608953 -1.5516666  1.1955161  4.8900152
#> [19]  2.6929286 -1.5545256  1.7621713  3.6416308  1.1724290 -2.5499919
#> [25]  2.4324588  6.4188352  2.6336468 -1.8297283  1.4875546  3.8081350
#> [31]  1.1473714 -1.8343958
extract.mirt(grp1, 'parvec')
#>  [1]  0.8316800  4.8913029  2.5597789 -1.3592138  0.8289135  2.4772334
#>  [7]  0.7025447 -2.1135580  2.8809926  5.4820292  2.2126281 -2.5451325
#> [13]  0.6920437  3.0133748  0.8608953 -1.5516666

# }