Compute latent regression fixed effect expected values

Create expected values for fixed effects parameters in latent regression models.

fixef(x)

Arguments

x
an estimated model object from the mixedmirt or mirt function

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi: 10.18637/jss.v048.i06

Chalmers, R. P. (2015). Extended Mixed-Effects Item Response Models with the MH-RM Algorithm. Journal of Educational Measurement, 52, 200-222. doi: 10.1111/jedm.12072

See also

Author

Phil Chalmers rphilip.chalmers@gmail.com

Examples

# \donttest{

#simulate data
set.seed(1234)
N <- 1000

# covariates
X1 <- rnorm(N); X2 <- rnorm(N)
covdata <- data.frame(X1, X2)
Theta <- matrix(0.5 * X1 + -1 * X2 + rnorm(N, sd = 0.5))

#items and response data
a <- matrix(1, 20); d <- matrix(rnorm(20))
dat <- simdata(a, d, 1000, itemtype = '2PL', Theta=Theta)

#conditional model using X1 and X2 as predictors of Theta
mod1 <- mirt(dat, 1, 'Rasch', covdata=covdata, formula = ~ X1 + X2)

#latent regression fixed effects (i.e., expected values)
fe <- fixef(mod1)
head(fe)

#>           F1
#> 1  0.6129074
#> 2 -0.1661799
#> 3  2.1652793
#> 4 -1.8932638
#> 5 -0.5021655
#> 6  2.2405972

# with mixedmirt()
mod1b <- mixedmirt(dat, covdata, 1, lr.fixed = ~ X1 + X2, fixed = ~ 0 + items)
fe2 <- fixef(mod1b)
head(fe2)

#>           F1
#> 1  0.6165760
#> 2 -0.1671332
#> 3  2.1737979
#> 4 -1.8995517
#> 5 -0.5047064
#> 6  2.2499237

# }