Compute latent regression fixed effect expected values

Create expected values for fixed effects parameters in latent regression models.

Usage

fixef(x)

Arguments

x

an estimated model object from the mixedmirt or mirt function

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi:10.18637/jss.v048.i06

Chalmers, R. P. (2015). Extended Mixed-Effects Item Response Models with the MH-RM Algorithm. Journal of Educational Measurement, 52, 200-222. doi:10.1111/jedm.12072

See also

Author

Phil Chalmers rphilip.chalmers@gmail.com

Examples

# \donttest{

#simulate data
set.seed(1234)
N <- 1000

# covariates
X1 <- rnorm(N); X2 <- rnorm(N)
covdata <- data.frame(X1, X2)
Theta <- matrix(0.5 * X1 + -1 * X2 + rnorm(N, sd = 0.5))

#items and response data
a <- matrix(1, 20); d <- matrix(rnorm(20))
dat <- simdata(a, d, 1000, itemtype = '2PL', Theta=Theta)

#conditional model using X1 and X2 as predictors of Theta
mod1 <- mirt(dat, 1, 'Rasch', covdata=covdata, formula = ~ X1 + X2)
#> Iteration: 1, Log-Lik: -11123.874, Max-Change: 0.77434Iteration: 2, Log-Lik: -10724.065, Max-Change: 0.53447Iteration: 3, Log-Lik: -10569.146, Max-Change: 0.20072Iteration: 4, Log-Lik: -10525.011, Max-Change: 0.08901Iteration: 5, Log-Lik: -10508.664, Max-Change: 0.04954Iteration: 6, Log-Lik: -10501.108, Max-Change: 0.03132Iteration: 7, Log-Lik: -10497.152, Max-Change: 0.02140Iteration: 8, Log-Lik: -10494.904, Max-Change: 0.01541Iteration: 9, Log-Lik: -10493.551, Max-Change: 0.01150Iteration: 10, Log-Lik: -10492.672, Max-Change: 0.00901Iteration: 11, Log-Lik: -10492.127, Max-Change: 0.00697Iteration: 12, Log-Lik: -10491.775, Max-Change: 0.00549Iteration: 13, Log-Lik: -10491.541, Max-Change: 0.00449Iteration: 14, Log-Lik: -10491.376, Max-Change: 0.00359Iteration: 15, Log-Lik: -10491.270, Max-Change: 0.00290Iteration: 16, Log-Lik: -10491.197, Max-Change: 0.00240Iteration: 17, Log-Lik: -10491.146, Max-Change: 0.00196Iteration: 18, Log-Lik: -10491.112, Max-Change: 0.00161Iteration: 19, Log-Lik: -10491.088, Max-Change: 0.00137Iteration: 20, Log-Lik: -10491.070, Max-Change: 0.00112Iteration: 21, Log-Lik: -10491.058, Max-Change: 0.00093Iteration: 22, Log-Lik: -10491.050, Max-Change: 0.00078Iteration: 23, Log-Lik: -10491.044, Max-Change: 0.00065Iteration: 24, Log-Lik: -10491.040, Max-Change: 0.00054Iteration: 25, Log-Lik: -10491.037, Max-Change: 0.00046Iteration: 26, Log-Lik: -10491.035, Max-Change: 0.00039Iteration: 27, Log-Lik: -10491.034, Max-Change: 0.00032Iteration: 28, Log-Lik: -10491.033, Max-Change: 0.00027Iteration: 29, Log-Lik: -10491.032, Max-Change: 0.00023Iteration: 30, Log-Lik: -10491.032, Max-Change: 0.00019Iteration: 31, Log-Lik: -10491.031, Max-Change: 0.00026Iteration: 32, Log-Lik: -10491.031, Max-Change: 0.00014Iteration: 33, Log-Lik: -10491.031, Max-Change: 0.00012Iteration: 34, Log-Lik: -10491.031, Max-Change: 0.00010

#latent regression fixed effects (i.e., expected values)
fe <- fixef(mod1)
head(fe)
#>              F1
#> [1,]  0.6128940
#> [2,] -0.1661763
#> [3,]  2.1652373
#> [4,] -1.8932285
#> [5,] -0.5021551
#> [6,]  2.2405531

# with mixedmirt()
mod1b <- mixedmirt(dat, covdata, 1, lr.fixed = ~ X1 + X2, fixed = ~ 0 + items)
#> Stage 1 = 1, CDLL = -13402.5, AR(0.81) = [0.48], Max-Change = 0.2000Stage 1 = 2, CDLL = -12841.6, AR(0.81) = [0.52], Max-Change = 0.2000Stage 1 = 3, CDLL = -12458.4, AR(0.81) = [0.51], Max-Change = 0.2000Stage 1 = 4, CDLL = -12077.8, AR(0.81) = [0.48], Max-Change = 0.1965Stage 1 = 5, CDLL = -11809.3, AR(0.81) = [0.48], Max-Change = 0.1716Stage 1 = 6, CDLL = -11634.0, AR(0.81) = [0.48], Max-Change = 0.1474Stage 1 = 7, CDLL = -11477.1, AR(0.81) = [0.47], Max-Change = 0.1256Stage 1 = 8, CDLL = -11425.1, AR(0.81) = [0.52], Max-Change = 0.1060Stage 1 = 9, CDLL = -11328.3, AR(0.81) = [0.50], Max-Change = 0.0884Stage 1 = 10, CDLL = -11231.0, AR(0.81) = [0.48], Max-Change = 0.0838Stage 1 = 11, CDLL = -11190.5, AR(0.81) = [0.49], Max-Change = 0.0746Stage 1 = 12, CDLL = -11177.4, AR(0.81) = [0.45], Max-Change = 0.0642Stage 1 = 13, CDLL = -11137.4, AR(0.81) = [0.47], Max-Change = 0.0512Stage 1 = 14, CDLL = -11116.8, AR(0.81) = [0.46], Max-Change = 0.0449Stage 1 = 15, CDLL = -11074.5, AR(0.81) = [0.45], Max-Change = 0.0425Stage 1 = 16, CDLL = -11056.4, AR(0.81) = [0.47], Max-Change = 0.0340Stage 1 = 17, CDLL = -11044.3, AR(0.81) = [0.44], Max-Change = 0.0309Stage 1 = 18, CDLL = -11054.9, AR(0.81) = [0.48], Max-Change = 0.0248Stage 1 = 19, CDLL = -11061.5, AR(0.81) = [0.45], Max-Change = 0.0231Stage 1 = 20, CDLL = -11028.0, AR(0.81) = [0.46], Max-Change = 0.0157Stage 1 = 21, CDLL = -11038.7, AR(0.81) = [0.45], Max-Change = 0.0173Stage 1 = 22, CDLL = -10993.1, AR(0.81) = [0.46], Max-Change = 0.0177Stage 1 = 23, CDLL = -11012.2, AR(0.81) = [0.47], Max-Change = 0.0151Stage 1 = 24, CDLL = -10994.9, AR(0.81) = [0.44], Max-Change = 0.0149Stage 1 = 25, CDLL = -10944.5, AR(0.81) = [0.45], Max-Change = 0.0139Stage 1 = 26, CDLL = -10949.3, AR(0.81) = [0.46], Max-Change = 0.0103Stage 1 = 27, CDLL = -10983.0, AR(0.81) = [0.47], Max-Change = 0.0071Stage 1 = 28, CDLL = -10958.1, AR(0.81) = [0.45], Max-Change = 0.0092Stage 1 = 29, CDLL = -10935.4, AR(0.81) = [0.44], Max-Change = 0.0086Stage 1 = 30, CDLL = -10946.0, AR(0.81) = [0.47], Max-Change = 0.0073Stage 1 = 31, CDLL = -10960.1, AR(0.81) = [0.44], Max-Change = 0.0079Stage 1 = 32, CDLL = -10932.1, AR(0.81) = [0.45], Max-Change = 0.0073Stage 1 = 33, CDLL = -10913.2, AR(0.81) = [0.44], Max-Change = 0.0063Stage 1 = 34, CDLL = -10918.6, AR(0.81) = [0.44], Max-Change = 0.0070Stage 1 = 35, CDLL = -10922.4, AR(0.81) = [0.44], Max-Change = 0.0069Stage 1 = 36, CDLL = -10925.7, AR(0.81) = [0.44], Max-Change = 0.0032Stage 1 = 37, CDLL = -10960.0, AR(0.81) = [0.46], Max-Change = 0.0047Stage 1 = 38, CDLL = -10925.8, AR(0.81) = [0.44], Max-Change = 0.0042Stage 1 = 39, CDLL = -10941.5, AR(0.81) = [0.43], Max-Change = 0.0033Stage 1 = 40, CDLL = -10898.4, AR(0.81) = [0.44], Max-Change = 0.0047Stage 1 = 41, CDLL = -10941.7, AR(0.81) = [0.44], Max-Change = 0.0034Stage 1 = 42, CDLL = -10902.9, AR(0.81) = [0.43], Max-Change = 0.0043Stage 1 = 43, CDLL = -10913.7, AR(0.81) = [0.43], Max-Change = 0.0048Stage 1 = 44, CDLL = -10906.0, AR(0.81) = [0.45], Max-Change = 0.0044Stage 1 = 45, CDLL = -10888.3, AR(0.81) = [0.40], Max-Change = 0.0080Stage 1 = 46, CDLL = -10922.2, AR(0.81) = [0.44], Max-Change = 0.0022Stage 1 = 47, CDLL = -10944.3, AR(0.81) = [0.47], Max-Change = 0.0017Stage 1 = 48, CDLL = -10909.3, AR(0.81) = [0.42], Max-Change = 0.0045Stage 1 = 49, CDLL = -10886.0, AR(0.81) = [0.44], Max-Change = 0.0037Stage 1 = 50, CDLL = -10910.5, AR(0.81) = [0.43], Max-Change = 0.0041Stage 1 = 51, CDLL = -10912.9, AR(0.81) = [0.44], Max-Change = 0.0038Stage 1 = 52, CDLL = -10921.7, AR(0.81) = [0.43], Max-Change = 0.0040Stage 1 = 53, CDLL = -10933.5, AR(0.81) = [0.42], Max-Change = 0.0035Stage 1 = 54, CDLL = -10893.6, AR(0.81) = [0.45], Max-Change = 0.0043Stage 1 = 55, CDLL = -10908.2, AR(0.81) = [0.45], Max-Change = 0.0025Stage 1 = 56, CDLL = -10900.9, AR(0.81) = [0.44], Max-Change = 0.0057Stage 1 = 57, CDLL = -10918.8, AR(0.81) = [0.41], Max-Change = 0.0039Stage 1 = 58, CDLL = -10910.8, AR(0.81) = [0.44], Max-Change = 0.0038Stage 1 = 59, CDLL = -10937.0, AR(0.81) = [0.43], Max-Change = 0.0032Stage 1 = 60, CDLL = -10919.9, AR(0.81) = [0.45], Max-Change = 0.0017Stage 1 = 61, CDLL = -10922.3, AR(0.81) = [0.45], Max-Change = 0.0055Stage 1 = 62, CDLL = -10871.1, AR(0.81) = [0.41], Max-Change = 0.0042Stage 1 = 63, CDLL = -10886.9, AR(0.81) = [0.43], Max-Change = 0.0036Stage 1 = 64, CDLL = -10881.8, AR(0.81) = [0.45], Max-Change = 0.0026Stage 1 = 65, CDLL = -10924.8, AR(0.81) = [0.43], Max-Change = 0.0024Stage 1 = 66, CDLL = -10874.9, AR(0.81) = [0.43], Max-Change = 0.0054Stage 1 = 67, CDLL = -10914.3, AR(0.81) = [0.43], Max-Change = 0.0044Stage 1 = 68, CDLL = -10916.5, AR(0.81) = [0.45], Max-Change = 0.0109Stage 1 = 69, CDLL = -10864.7, AR(0.81) = [0.44], Max-Change = 0.0031Stage 1 = 70, CDLL = -10879.3, AR(0.81) = [0.43], Max-Change = 0.0054Stage 1 = 71, CDLL = -10888.2, AR(0.81) = [0.43], Max-Change = 0.0019Stage 1 = 72, CDLL = -10910.4, AR(0.81) = [0.42], Max-Change = 0.0011Stage 1 = 73, CDLL = -10911.6, AR(0.81) = [0.44], Max-Change = 0.0019Stage 1 = 74, CDLL = -10876.7, AR(0.81) = [0.44], Max-Change = 0.0055Stage 1 = 75, CDLL = -10907.4, AR(0.81) = [0.42], Max-Change = 0.0013Stage 1 = 76, CDLL = -10861.1, AR(0.81) = [0.42], Max-Change = 0.0055Stage 1 = 77, CDLL = -10899.3, AR(0.81) = [0.41], Max-Change = 0.0033Stage 1 = 78, CDLL = -10895.6, AR(0.81) = [0.43], Max-Change = 0.0044Stage 1 = 79, CDLL = -10933.6, AR(0.81) = [0.46], Max-Change = 0.0044Stage 1 = 80, CDLL = -10908.2, AR(0.81) = [0.41], Max-Change = 0.0031Stage 1 = 81, CDLL = -10899.0, AR(0.81) = [0.42], Max-Change = 0.0014Stage 1 = 82, CDLL = -10900.8, AR(0.81) = [0.43], Max-Change = 0.0081Stage 1 = 83, CDLL = -10875.1, AR(0.81) = [0.42], Max-Change = 0.0021Stage 1 = 84, CDLL = -10862.2, AR(0.81) = [0.42], Max-Change = 0.0045Stage 1 = 85, CDLL = -10880.9, AR(0.81) = [0.42], Max-Change = 0.0032Stage 1 = 86, CDLL = -10860.7, AR(0.81) = [0.42], Max-Change = 0.0042Stage 1 = 87, CDLL = -10880.0, AR(0.81) = [0.45], Max-Change = 0.0034Stage 1 = 88, CDLL = -10891.3, AR(0.81) = [0.42], Max-Change = 0.0071Stage 1 = 89, CDLL = -10881.9, AR(0.81) = [0.47], Max-Change = 0.0059Stage 1 = 90, CDLL = -10874.1, AR(0.81) = [0.44], Max-Change = 0.0019Stage 1 = 91, CDLL = -10878.4, AR(0.81) = [0.44], Max-Change = 0.0014Stage 1 = 92, CDLL = -10849.4, AR(0.81) = [0.43], Max-Change = 0.0050Stage 1 = 93, CDLL = -10857.4, AR(0.81) = [0.42], Max-Change = 0.0038Stage 1 = 94, CDLL = -10875.6, AR(0.81) = [0.42], Max-Change = 0.0012Stage 1 = 95, CDLL = -10871.2, AR(0.81) = [0.44], Max-Change = 0.0046Stage 1 = 96, CDLL = -10872.4, AR(0.81) = [0.42], Max-Change = 0.0034Stage 1 = 97, CDLL = -10910.2, AR(0.81) = [0.41], Max-Change = 0.0024Stage 1 = 98, CDLL = -10870.4, AR(0.81) = [0.41], Max-Change = 0.0047Stage 1 = 99, CDLL = -10879.9, AR(0.81) = [0.41], Max-Change = 0.0046Stage 1 = 100, CDLL = -10900.4, AR(0.81) = [0.41], Max-Change = 0.0034Stage 1 = 101, CDLL = -10845.4, AR(0.81) = [0.42], Max-Change = 0.0041Stage 1 = 102, CDLL = -10906.2, AR(0.81) = [0.44], Max-Change = 0.0044Stage 1 = 103, CDLL = -10878.5, AR(0.81) = [0.43], Max-Change = 0.0024Stage 1 = 104, CDLL = -10846.2, AR(0.81) = [0.41], Max-Change = 0.0054Stage 1 = 105, CDLL = -10827.9, AR(0.81) = [0.41], Max-Change = 0.0059Stage 1 = 106, CDLL = -10853.4, AR(0.81) = [0.41], Max-Change = 0.0032Stage 1 = 107, CDLL = -10863.9, AR(0.81) = [0.45], Max-Change = 0.0045Stage 1 = 108, CDLL = -10847.1, AR(0.81) = [0.43], Max-Change = 0.0028Stage 1 = 109, CDLL = -10866.1, AR(0.81) = [0.44], Max-Change = 0.0024Stage 1 = 110, CDLL = -10858.8, AR(0.81) = [0.39], Max-Change = 0.0044Stage 1 = 111, CDLL = -10849.4, AR(0.81) = [0.40], Max-Change = 0.0050Stage 1 = 112, CDLL = -10878.9, AR(0.81) = [0.43], Max-Change = 0.0011Stage 1 = 113, CDLL = -10850.8, AR(0.81) = [0.44], Max-Change = 0.0037Stage 1 = 114, CDLL = -10876.2, AR(0.81) = [0.43], Max-Change = 0.0037Stage 1 = 115, CDLL = -10838.0, AR(0.81) = [0.40], Max-Change = 0.0026Stage 1 = 116, CDLL = -10842.8, AR(0.81) = [0.43], Max-Change = 0.0029Stage 1 = 117, CDLL = -10871.7, AR(0.81) = [0.44], Max-Change = 0.0040Stage 1 = 118, CDLL = -10885.7, AR(0.81) = [0.41], Max-Change = 0.0037Stage 1 = 119, CDLL = -10899.8, AR(0.81) = [0.42], Max-Change = 0.0056Stage 1 = 120, CDLL = -10867.2, AR(0.81) = [0.42], Max-Change = 0.0046Stage 1 = 121, CDLL = -10897.5, AR(0.81) = [0.43], Max-Change = 0.0020Stage 1 = 122, CDLL = -10879.8, AR(0.81) = [0.45], Max-Change = 0.0016Stage 1 = 123, CDLL = -10857.0, AR(0.81) = [0.40], Max-Change = 0.0041Stage 1 = 124, CDLL = -10834.5, AR(0.81) = [0.40], Max-Change = 0.0064Stage 1 = 125, CDLL = -10864.6, AR(0.81) = [0.44], Max-Change = 0.0059Stage 1 = 126, CDLL = -10850.4, AR(0.81) = [0.41], Max-Change = 0.0042Stage 1 = 127, CDLL = -10888.8, AR(0.81) = [0.43], Max-Change = 0.0026Stage 1 = 128, CDLL = -10829.7, AR(0.81) = [0.41], Max-Change = 0.0036Stage 1 = 129, CDLL = -10826.6, AR(0.81) = [0.43], Max-Change = 0.0028Stage 1 = 130, CDLL = -10828.4, AR(0.81) = [0.42], Max-Change = 0.0044Stage 1 = 131, CDLL = -10835.0, AR(0.81) = [0.43], Max-Change = 0.0035Stage 1 = 132, CDLL = -10829.3, AR(0.81) = [0.40], Max-Change = 0.0026Stage 1 = 133, CDLL = -10826.7, AR(0.81) = [0.44], Max-Change = 0.0034Stage 1 = 134, CDLL = -10835.8, AR(0.81) = [0.40], Max-Change = 0.0037Stage 1 = 135, CDLL = -10840.4, AR(0.81) = [0.44], Max-Change = 0.0045Stage 1 = 136, CDLL = -10833.1, AR(0.81) = [0.40], Max-Change = 0.0049Stage 1 = 137, CDLL = -10850.0, AR(0.81) = [0.39], Max-Change = 0.0034Stage 1 = 138, CDLL = -10845.5, AR(0.81) = [0.41], Max-Change = 0.0028Stage 1 = 139, CDLL = -10824.1, AR(0.81) = [0.41], Max-Change = 0.0026Stage 1 = 140, CDLL = -10862.6, AR(0.81) = [0.40], Max-Change = 0.0077Stage 1 = 141, CDLL = -10857.9, AR(0.81) = [0.43], Max-Change = 0.0028Stage 1 = 142, CDLL = -10861.4, AR(0.81) = [0.44], Max-Change = 0.0037Stage 1 = 143, CDLL = -10831.9, AR(0.81) = [0.42], Max-Change = 0.0035Stage 1 = 144, CDLL = -10827.8, AR(0.81) = [0.42], Max-Change = 0.0021Stage 1 = 145, CDLL = -10857.0, AR(0.81) = [0.39], Max-Change = 0.0026Stage 1 = 146, CDLL = -10875.8, AR(0.81) = [0.40], Max-Change = 0.0023Stage 1 = 147, CDLL = -10884.5, AR(0.81) = [0.41], Max-Change = 0.0033Stage 1 = 148, CDLL = -10878.8, AR(0.81) = [0.42], Max-Change = 0.0049Stage 1 = 149, CDLL = -10862.7, AR(0.81) = [0.42], Max-Change = 0.0058Stage 1 = 150, CDLL = -10894.5, AR(0.93) = [0.40], Max-Change = 0.0024Stage 2 = 1, CDLL = -10872.1, AR(0.93) = [0.39], Max-Change = 0.0040Stage 2 = 2, CDLL = -10861.9, AR(0.93) = [0.42], Max-Change = 0.0035Stage 2 = 3, CDLL = -10870.3, AR(0.93) = [0.39], Max-Change = 0.0030Stage 2 = 4, CDLL = -10813.0, AR(0.93) = [0.41], Max-Change = 0.0035Stage 2 = 5, CDLL = -10854.2, AR(0.93) = [0.39], Max-Change = 0.0052Stage 2 = 6, CDLL = -10860.4, AR(0.93) = [0.37], Max-Change = 0.0016Stage 2 = 7, CDLL = -10836.9, AR(0.93) = [0.39], Max-Change = 0.0026Stage 2 = 8, CDLL = -10873.0, AR(0.93) = [0.40], Max-Change = 0.0044Stage 2 = 9, CDLL = -10864.2, AR(0.93) = [0.38], Max-Change = 0.0045Stage 2 = 10, CDLL = -10877.9, AR(0.93) = [0.41], Max-Change = 0.0029Stage 2 = 11, CDLL = -10828.1, AR(0.93) = [0.40], Max-Change = 0.0046Stage 2 = 12, CDLL = -10817.5, AR(0.93) = [0.38], Max-Change = 0.0050Stage 2 = 13, CDLL = -10852.7, AR(0.93) = [0.40], Max-Change = 0.0030Stage 2 = 14, CDLL = -10847.0, AR(0.93) = [0.38], Max-Change = 0.0010Stage 2 = 15, CDLL = -10873.1, AR(0.93) = [0.40], Max-Change = 0.0043Stage 2 = 16, CDLL = -10874.4, AR(0.93) = [0.41], Max-Change = 0.0055Stage 2 = 17, CDLL = -10906.8, AR(0.93) = [0.39], Max-Change = 0.0051Stage 2 = 18, CDLL = -10880.3, AR(0.93) = [0.41], Max-Change = 0.0041Stage 2 = 19, CDLL = -10825.7, AR(0.93) = [0.39], Max-Change = 0.0045Stage 2 = 20, CDLL = -10879.6, AR(0.93) = [0.39], Max-Change = 0.0042Stage 2 = 21, CDLL = -10890.7, AR(0.93) = [0.42], Max-Change = 0.0023Stage 2 = 22, CDLL = -10883.9, AR(0.93) = [0.38], Max-Change = 0.0017Stage 2 = 23, CDLL = -10933.3, AR(0.93) = [0.40], Max-Change = 0.0029Stage 2 = 24, CDLL = -10865.1, AR(0.93) = [0.38], Max-Change = 0.0010Stage 2 = 25, CDLL = -10866.9, AR(0.93) = [0.42], Max-Change = 0.0035Stage 2 = 26, CDLL = -10887.3, AR(0.93) = [0.40], Max-Change = 0.0013Stage 2 = 27, CDLL = -10874.7, AR(0.93) = [0.41], Max-Change = 0.0021Stage 2 = 28, CDLL = -10862.0, AR(0.93) = [0.40], Max-Change = 0.0036Stage 2 = 29, CDLL = -10888.5, AR(0.93) = [0.39], Max-Change = 0.0040Stage 2 = 30, CDLL = -10845.2, AR(0.93) = [0.42], Max-Change = 0.0047Stage 2 = 31, CDLL = -10846.9, AR(0.93) = [0.39], Max-Change = 0.0043Stage 2 = 32, CDLL = -10826.9, AR(0.93) = [0.41], Max-Change = 0.0035Stage 2 = 33, CDLL = -10877.7, AR(0.93) = [0.39], Max-Change = 0.0026Stage 2 = 34, CDLL = -10844.3, AR(0.93) = [0.38], Max-Change = 0.0032Stage 2 = 35, CDLL = -10865.1, AR(0.93) = [0.39], Max-Change = 0.0013Stage 2 = 36, CDLL = -10881.9, AR(0.93) = [0.38], Max-Change = 0.0048Stage 2 = 37, CDLL = -10852.5, AR(0.93) = [0.40], Max-Change = 0.0029Stage 2 = 38, CDLL = -10864.6, AR(0.93) = [0.41], Max-Change = 0.0019Stage 2 = 39, CDLL = -10894.5, AR(0.93) = [0.42], Max-Change = 0.0034Stage 2 = 40, CDLL = -10853.0, AR(0.93) = [0.40], Max-Change = 0.0019Stage 2 = 41, CDLL = -10844.2, AR(0.93) = [0.41], Max-Change = 0.0031Stage 2 = 42, CDLL = -10866.6, AR(0.93) = [0.41], Max-Change = 0.0015Stage 2 = 43, CDLL = -10878.1, AR(0.93) = [0.40], Max-Change = 0.0041Stage 2 = 44, CDLL = -10911.5, AR(0.93) = [0.39], Max-Change = 0.0040Stage 2 = 45, CDLL = -10863.0, AR(0.93) = [0.42], Max-Change = 0.0041Stage 2 = 46, CDLL = -10872.1, AR(0.93) = [0.40], Max-Change = 0.0038Stage 2 = 47, CDLL = -10900.3, AR(0.93) = [0.40], Max-Change = 0.0033Stage 2 = 48, CDLL = -10867.7, AR(0.93) = [0.39], Max-Change = 0.0021Stage 2 = 49, CDLL = -10905.5, AR(0.93) = [0.41], Max-Change = 0.0025Stage 2 = 50, CDLL = -10857.4, AR(0.93) = [0.40], Max-Change = 0.0032Stage 2 = 51, CDLL = -10912.8, AR(0.93) = [0.41], Max-Change = 0.0031Stage 2 = 52, CDLL = -10873.0, AR(0.93) = [0.42], Max-Change = 0.0047Stage 2 = 53, CDLL = -10891.6, AR(0.93) = [0.39], Max-Change = 0.0022Stage 2 = 54, CDLL = -10890.5, AR(0.93) = [0.40], Max-Change = 0.0030Stage 2 = 55, CDLL = -10893.5, AR(0.93) = [0.40], Max-Change = 0.0035Stage 2 = 56, CDLL = -10894.0, AR(0.93) = [0.38], Max-Change = 0.0071Stage 2 = 57, CDLL = -10942.4, AR(0.93) = [0.39], Max-Change = 0.0041Stage 2 = 58, CDLL = -10927.1, AR(0.93) = [0.43], Max-Change = 0.0028Stage 2 = 59, CDLL = -10911.1, AR(0.93) = [0.40], Max-Change = 0.0055Stage 2 = 60, CDLL = -10892.3, AR(0.93) = [0.41], Max-Change = 0.0067Stage 2 = 61, CDLL = -10906.7, AR(0.93) = [0.45], Max-Change = 0.0018Stage 2 = 62, CDLL = -10874.8, AR(0.93) = [0.41], Max-Change = 0.0019Stage 2 = 63, CDLL = -10886.6, AR(0.93) = [0.41], Max-Change = 0.0058Stage 2 = 64, CDLL = -10904.5, AR(0.93) = [0.41], Max-Change = 0.0030Stage 2 = 65, CDLL = -10829.2, AR(0.93) = [0.43], Max-Change = 0.0064Stage 2 = 66, CDLL = -10888.3, AR(0.93) = [0.40], Max-Change = 0.0038Stage 2 = 67, CDLL = -10840.9, AR(0.93) = [0.40], Max-Change = 0.0062Stage 2 = 68, CDLL = -10894.0, AR(0.93) = [0.41], Max-Change = 0.0029Stage 2 = 69, CDLL = -10883.3, AR(0.93) = [0.43], Max-Change = 0.0028Stage 2 = 70, CDLL = -10859.7, AR(0.93) = [0.42], Max-Change = 0.0040Stage 2 = 71, CDLL = -10852.0, AR(0.93) = [0.39], Max-Change = 0.0042Stage 2 = 72, CDLL = -10853.9, AR(0.93) = [0.42], Max-Change = 0.0031Stage 2 = 73, CDLL = -10897.8, AR(0.93) = [0.41], Max-Change = 0.0026Stage 2 = 74, CDLL = -10913.4, AR(0.93) = [0.43], Max-Change = 0.0022Stage 2 = 75, CDLL = -10860.5, AR(0.93) = [0.41], Max-Change = 0.0069Stage 2 = 76, CDLL = -10890.3, AR(0.93) = [0.36], Max-Change = 0.0058Stage 2 = 77, CDLL = -10843.8, AR(0.93) = [0.42], Max-Change = 0.0021Stage 2 = 78, CDLL = -10869.2, AR(0.93) = [0.41], Max-Change = 0.0035Stage 2 = 79, CDLL = -10858.1, AR(0.93) = [0.40], Max-Change = 0.0068Stage 2 = 80, CDLL = -10881.9, AR(0.93) = [0.41], Max-Change = 0.0051Stage 2 = 81, CDLL = -10855.6, AR(0.93) = [0.39], Max-Change = 0.0015Stage 2 = 82, CDLL = -10891.1, AR(0.93) = [0.40], Max-Change = 0.0036Stage 2 = 83, CDLL = -10860.2, AR(0.93) = [0.42], Max-Change = 0.0043Stage 2 = 84, CDLL = -10834.0, AR(0.93) = [0.40], Max-Change = 0.0034Stage 2 = 85, CDLL = -10849.3, AR(0.93) = [0.41], Max-Change = 0.0013Stage 2 = 86, CDLL = -10865.6, AR(0.93) = [0.42], Max-Change = 0.0022Stage 2 = 87, CDLL = -10877.7, AR(0.93) = [0.37], Max-Change = 0.0038Stage 2 = 88, CDLL = -10846.5, AR(0.93) = [0.40], Max-Change = 0.0036Stage 2 = 89, CDLL = -10855.1, AR(0.93) = [0.39], Max-Change = 0.0039Stage 2 = 90, CDLL = -10851.5, AR(0.93) = [0.37], Max-Change = 0.0021Stage 2 = 91, CDLL = -10833.6, AR(0.93) = [0.40], Max-Change = 0.0030Stage 2 = 92, CDLL = -10855.9, AR(0.93) = [0.40], Max-Change = 0.0031Stage 2 = 93, CDLL = -10863.3, AR(0.93) = [0.43], Max-Change = 0.0058Stage 2 = 94, CDLL = -10858.8, AR(0.93) = [0.39], Max-Change = 0.0060Stage 2 = 95, CDLL = -10889.0, AR(0.93) = [0.39], Max-Change = 0.0024Stage 2 = 96, CDLL = -10840.8, AR(0.93) = [0.40], Max-Change = 0.0018Stage 2 = 97, CDLL = -10877.8, AR(0.93) = [0.41], Max-Change = 0.0020Stage 2 = 98, CDLL = -10868.9, AR(0.93) = [0.39], Max-Change = 0.0038Stage 2 = 99, CDLL = -10869.5, AR(0.93) = [0.40], Max-Change = 0.0064Stage 2 = 100, CDLL = -10861.5, AR(0.93) = [0.41], Max-Change = 0.0051Stage 3 = 1, CDLL = -10895.9, AR(0.93) = [0.40], gam = 0.0000, Max-Change = 0.0000Stage 3 = 2, CDLL = -10875.6, AR(0.93) = [0.41], gam = 0.1778, Max-Change = 0.0053Stage 3 = 3, CDLL = -10850.5, AR(0.93) = [0.39], gam = 0.1057, Max-Change = 0.0033Stage 3 = 4, CDLL = -10845.9, AR(0.93) = [0.43], gam = 0.0780, Max-Change = 0.0010Stage 3 = 5, CDLL = -10847.6, AR(0.93) = [0.41], gam = 0.0629, Max-Change = 0.0008Stage 3 = 6, CDLL = -10894.7, AR(0.93) = [0.43], gam = 0.0532, Max-Change = 0.0017Stage 3 = 7, CDLL = -10894.2, AR(0.93) = [0.40], gam = 0.0464, Max-Change = 0.0011Stage 3 = 8, CDLL = -10845.0, AR(0.93) = [0.42], gam = 0.0413, Max-Change = 0.0007Stage 3 = 9, CDLL = -10863.4, AR(0.93) = [0.42], gam = 0.0374, Max-Change = 0.0009Stage 3 = 10, CDLL = -10856.9, AR(0.93) = [0.43], gam = 0.0342, Max-Change = 0.0006
#> 
#> Calculating information matrix...
#> 
#> Calculating log-likelihood...
fe2 <- fixef(mod1b)
head(fe2)
#>              F1
#> [1,]  0.6165760
#> [2,] -0.1671332
#> [3,]  2.1737979
#> [4,] -1.8995517
#> [5,] -0.5047064
#> [6,]  2.2499237

# }