fixef {mirt}	R Documentation

Compute latent regression fixed effect expected values

Description

Create expected values for fixed effects parameters in latent regression models.

Usage

fixef(x)

Arguments

x
an estimated model object from the mixedmirt or mirt function

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

See Also

mirt, mixedmirt

Examples

## No test: 

#simulate data
set.seed(1234)
N <- 1000

# covariates
X1 <- rnorm(N); X2 <- rnorm(N)
covdata <- data.frame(X1, X2)
Theta <- matrix(0.5 * X1 + -1 * X2 + rnorm(N, sd = 0.5))

#items and response data
a <- matrix(1, 20); d <- matrix(rnorm(20))
dat <- simdata(a, d, 1000, itemtype = 'dich', Theta=Theta)

#conditional model using X1 and X2 as predictors of Theta
mod1 <- mirt(dat, 1, 'Rasch', covdata=covdata, formula = ~ X1 + X2)

#latent regression fixed effects (i.e., expected values)
fe <- fixef(mod1)
head(fe)

##           F1
## 1  0.5922383
## 2 -0.1609665
## 3  2.1341843
## 4 -1.8770290
## 5 -0.4896192
## 6  2.2035914

# with mixedmirt()
mod1b <- mixedmirt(dat, covdata, 1, lr.fixed = ~ X1 + X2, fixed = ~ 0 + items)
fe2 <- fixef(mod1b)
head(fe2)

##           F1
## 1  0.5885127
## 2 -0.1599947
## 3  2.1251367
## 4 -1.8701929
## 5 -0.4869973
## 6  2.1937551

## End(No test)

[Package mirt version 1.11 Index]