fscores {mirt}	R Documentation

Compute factor score estimates (a.k.a, ability estimates, latent trait estimates, etc)

Description

Computes MAP, EAP, ML (Embretson & Reise, 2000), EAP for sum-scores (Thissen et al., 1995), or WLE (Warm, 1989) factor scores with a multivariate normal prior distribution using equally spaced quadrature. EAP scores for models with more than three factors are generally not recommended since the integration grid becomes very large, resulting in slower estimation and less precision if the quadpts are too low. Therefore, MAP scores should be used instead of EAP scores for higher dimensional models. Multiple imputation variants are possible for each estimator if a parameter information matrix was computed, which are useful if the sample size/number of items were small. As well, if the model contained latent regression predictors this information will be used in computing MAP and EAP estimates (for these models, full.scores=TRUE by default). Finaly, plausible value imputation is also available, and will also account for latent regression predictor effects.

Usage

fscores(object, rotate = NULL, full.scores = FALSE, method = "EAP",
  quadpts = NULL, response.pattern = NULL, plausible.draws = 0,
  returnER = FALSE, return.acov = FALSE, mean = NULL, cov = NULL,
  verbose = TRUE, scores.only = TRUE, full.scores.SE = FALSE,
  theta_lim = c(-6, 6), MI = 0, QMC = FALSE, custom_den = NULL,
  custom_theta = NULL, ...)

Arguments

object
a computed model object of class SingleGroupClass, MultipleGroupClass, or DiscreteClass

rotate
rotation declaration to be used when estimating the factor scores. If "" then the object@rotate default value is used (only applicable to SingleGroupClass objects which have non-unique slopes)

full.scores
if FALSE (default) then a summary table with factor scores for each unique pattern is displayed. Otherwise the original data matrix is returned with the computed factor scores appended to the rightmost column

method
type of factor score estimation method. Can be expected a-posteriori ("EAP"), Bayes modal ("MAP"), weighted likelihood estimation ("WLE"), maximum likelihood ("ML"), or expected a-posteriori for sum scores ("EAPsum"). Can also be "plausible" for a single plausible value imputation for each case, and this is equivalent to setting plausible.draws = 1

quadpts
number of quadratures to use per dimension. If not specified, a suitable one will be created which decreases as the number of dimensions increases (and therefore for estimates such as EAP, will be less accurate). This is determined from the switch statement quadpts <- switch(as.character(nfact), '1'=61, '2'=31, '3'=15, '4'=9, '5'=7, 3)

response.pattern
an optional argument used to calculate the factor scores and standard errors for a given response vector or matrix/data.frame

plausible.draws
number of plausible values to draw for future researchers to perform secondary analyses of the latent trait scores. Typically used in conjunction with latent regression predictors (see mirt for details), but can also be generated when no predictor variables were modeled. If plausible.draws is greater than 0 a list of plausible values will be returned

returnER
logical; return empirical reliability (also known as marginal reliability) estimates as a numeric values?

return.acov
logical; return a list containing covariance matricies instead of factors scores? impute = TRUE not supported with this option

mean
a vector for custom latent variable means. If NULL, the default for 'group' values from the computed mirt object will be used

cov
a custom matrix of the latent variable covariance matrix. If NULL, the default for 'group' values from the computed mirt object will be used

verbose
logical; print verbose output messages?

scores.only
logical; return only the factor scores (only applicable when full.scores = TRUE)

full.scores.SE
logical; when full.scores == TRUE, also return the standard errors associated with each respondent? Default is FALSE

theta_lim
lower and upper range to evaluate latent trait integral for each dimension

MI
a number indicating how many multiple imputation draws to perform. Default is 0, indicating that no MI draws will be performed

QMC
logical; use quasi-Monte Carlo integration? If quadpts is omitted the default number of nodes is 2000

custom_den
a function used to define the integration density (if required). The NULL default assumes that the multivariate normal distribution with the 'GroupPars' hyper-parameters are used. At the minimum must be of the form:

function(Theta, ...)

where Theta is a matrix of latent trait values (will be a grid of values if method == 'EAPsum' or method == 'EAP', otherwise Theta will have only 1 row). Additional arguments may included and are caught through the fscores(...) input. The function must return a numeric vector of density weights (one for each row in Theta)

custom_theta
a matrix of custom integration nodes to use instead of the default, where each column correspends to the respective dimension in the model

...
additional arguments to be passed to nlm

Details

The function will return either a table with the computed scores and standard errors, the original data matrix with scores appended to the rightmost column, or the scores only. By default the latent means and covariances are determined from the estimated object, though these can be overwritten. Iterative estimation methods can be estimated in parallel to decrease estimation times if a mirtCluster object is available.

If the input object is a discrete latent class object estimated from mdirt then the returned results will be with respect to the posterior classification for each individual. The method inputs for 'DiscreteClass' objects may only be 'EAP', for posterior classification of each response pattern, or 'EAPsum' for posterior classification based on the raw sum-score.

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

References

Embretson, S. E. & Reise, S. P. (2000). Item Response Theory for Psychologists. Erlbaum.

Thissen, D., Pommerich, M., Billeaud, K., & Williams, V. S. L. (1995). Item Response Theory for Scores on Tests Including Polytomous Items with Ordered Responses. Applied Psychological Measurement, 19, 39-49.

Warm, T. A. (1989). Weighted likelihood estimation of ability in item response theory. Psychometrika, 54, 427-450.

See Also

averageMI

Examples

## No test: 

mod <- mirt(Science, 1)
tabscores <- fscores(mod)

## 
## Method:  EAP
## 
## Empirical Reliability:
##     F1 
## 0.6667

head(tabscores)

##      Comfort Work Future Benefit      F1  SE_F1
## [1,]       1    1      1       1 -2.7491 0.6293
## [2,]       1    3      2       1 -1.4196 0.5771
## [3,]       1    4      2       3 -0.7145 0.6198
## [4,]       1    4      3       1 -0.4465 0.6508
## [5,]       2    1      1       1 -2.5438 0.5909
## [6,]       2    1      2       4 -1.2479 0.5839

fullscores <- fscores(mod, full.scores = TRUE)
fullscores_with_SE <- fscores(mod, full.scores = TRUE, full.scores.SE=TRUE)
head(fullscores)

##               F1
## [1,]  0.40132315
## [2,]  0.05198892
## [3,] -0.89089337
## [4,] -0.89089337
## [5,]  0.76579641
## [6,]  0.66916315

head(fullscores_with_SE)

##               F1     SE_F1
## [1,]  0.40132315 0.5977583
## [2,]  0.05198892 0.5548735
## [3,] -0.89089337 0.5420682
## [4,] -0.89089337 0.5420682
## [5,]  0.76579641 0.6384553
## [6,]  0.66916315 0.5760913

#chage method argument to use MAP estimates
fullscores <- fscores(mod, full.scores = TRUE, method='MAP')
head(fullscores)

##               F1
## [1,]  0.42120536
## [2,]  0.05863428
## [3,] -0.91959392
## [4,] -0.91959392
## [5,]  0.79046251
## [6,]  0.68266911

#calculate MAP for a given response vector
fscores(mod, method='MAP', response.pattern = c(1,2,3,4))

##      Comfort Work Future Benefit         F1    SE_F1
## [1,]       1    2      3       4 -0.3702802 0.605315

#or matrix
fscores(mod, method='MAP', response.pattern = rbind(c(1,2,3,4), c(2,2,1,3)))

##      Comfort Work Future Benefit         F1     SE_F1
## [1,]       1    2      3       4 -0.3702802 0.6053150
## [2,]       2    2      1       3 -1.6942324 0.5772351

#use custom latent variable properties (diffuse prior for MAP is very close to ML)
fscores(mod, method='MAP', cov = matrix(1000))

## 
## Method:  MAP
## 
## Empirical Reliability:
##     F1 
## 0.4207

##       Comfort Work Future Benefit      F1   SE_F1
##  [1,]       1    1      1       1 -9.3411  9.6360
##  [2,]       1    3      2       1 -2.1031  0.6890
##  [3,]       1    4      2       3 -1.1509  0.7047
##  [4,]       1    4      3       1 -0.7801  0.8069
##  [5,]       2    1      1       1 -4.3947  1.1799
##  [6,]       2    1      2       4 -1.8624  0.6677
##  [7,]       2    2      1       1 -3.2476  0.8123
##  [8,]       2    2      2       2 -1.9110  0.5874
##  [9,]       2    2      2       3 -1.6054  0.6083
## [10,]       2    2      3       1 -1.4668  0.7134
## [11,]       2    2      3       2 -1.1676  0.6345
## [12,]       2    2      3       3 -0.7962  0.6388
## [13,]       2    2      4       3  0.2414  0.7882
## [14,]       2    3      1       3 -2.1427  0.7014
## [15,]       2    3      2       2 -1.6083  0.6257
## [16,]       2    3      2       3 -1.2540  0.6300
## [17,]       2    3      3       2 -0.7552  0.6557
## [18,]       2    3      3       3 -0.3314  0.6980
## [19,]       2    3      3       4  0.0830  0.7612
## [20,]       2    3      4       1  0.1907  0.8772
## [21,]       2    3      4       3  0.7632  0.6803
## [22,]       2    4      2       1 -1.8216  0.6940
## [23,]       2    4      4       3  1.3074  0.7534
## [24,]       2    4      4       4  2.0946  1.0395
## [25,]       3    1      1       1 -3.5383  1.0018
## [26,]       3    1      1       3 -2.5099  0.6970
## [27,]       3    1      2       2 -1.9252  0.6146
## [28,]       3    1      2       3 -1.5868  0.6432
## [29,]       3    1      3       2 -1.0869  0.6739
## [30,]       3    1      3       3 -0.6577  0.6953
## [31,]       3    1      3       4 -0.2808  0.8071
## [32,]       3    1      4       3  0.5432  0.7334
## [33,]       3    1      4       4  1.0380  0.7356
## [34,]       3    2      1       2 -2.3483  0.6245
## [35,]       3    2      1       4 -1.9137  0.7090
## [36,]       3    2      2       1 -1.8609  0.6157
## [37,]       3    2      2       2 -1.5766  0.5938
## [38,]       3    2      2       3 -1.2550  0.6021
## [39,]       3    2      3       1 -1.0245  0.6625
## [40,]       3    2      3       2 -0.7974  0.6289
## [41,]       3    2      3       3 -0.4088  0.6675
## [42,]       3    2      3       4 -0.0375  0.7436
## [43,]       3    2      4       1  0.0163  0.8998
## [44,]       3    2      4       2  0.2066  0.7882
## [45,]       3    2      4       3  0.6436  0.6838
## [46,]       3    2      4       4  1.1018  0.7182
## [47,]       3    3      1       3 -1.6304  0.7797
## [48,]       3    3      2       1 -1.5174  0.6589
## [49,]       3    3      2       2 -1.2392  0.6167
## [50,]       3    3      2       3 -0.8828  0.6298
## [51,]       3    3      2       4 -0.6216  0.7076
## [52,]       3    3      3       1 -0.5488  0.7105
## [53,]       3    3      3       2 -0.3472  0.6873
## [54,]       3    3      3       3  0.0863  0.6868
## [55,]       3    3      3       4  0.4899  0.6760
## [56,]       3    3      4       2  0.7329  0.6809
## [57,]       3    3      4       3  1.0480  0.6498
## [58,]       3    3      4       4  1.5247  0.7456
## [59,]       3    4      1       3 -1.3833  0.9419
## [60,]       3    4      2       3 -0.6092  0.7252
## [61,]       3    4      3       2  0.0993  0.7674
## [62,]       3    4      3       3  0.5376  0.6772
## [63,]       3    4      3       4  0.9605  0.6743
## [64,]       3    4      4       1  1.2182  0.7583
## [65,]       3    4      4       2  1.2721  0.7490
## [66,]       3    4      4       3  1.5994  0.7663
## [67,]       3    4      4       4  2.4226  1.0404
## [68,]       4    1      1       4 -2.1062  0.7718
## [69,]       4    1      2       2 -1.6378  0.6539
## [70,]       4    1      2       4 -1.0146  0.7235
## [71,]       4    1      3       4  0.3608  0.7776
## [72,]       4    2      2       1 -1.5677  0.6510
## [73,]       4    2      2       3 -0.9356  0.6273
## [74,]       4    2      3       3  0.0535  0.7107
## [75,]       4    2      3       4  0.4913  0.7093
## [76,]       4    2      4       2  0.7627  0.7226
## [77,]       4    2      4       4  1.7596  0.9307
## [78,]       4    3      2       3 -0.4860  0.7034
## [79,]       4    3      3       2  0.1390  0.7180
## [80,]       4    3      3       3  0.5323  0.6521
## [81,]       4    3      3       4  0.9283  0.6552
## [82,]       4    3      4       2  1.2157  0.7192
## [83,]       4    3      4       3  1.5265  0.7343
## [84,]       4    3      4       4  2.2584  0.9680
## [85,]       4    4      3       2  0.6392  0.7067
## [86,]       4    4      3       3  0.9804  0.6611
## [87,]       4    4      3       4  1.4633  0.7485
## [88,]       4    4      4       2  2.0278  1.0103
## [89,]       4    4      4       3  2.3900  1.0153
## [90,]       4    4      4       4  7.1128 10.6164

fscores(mod, method='ML')

## 
## Method:  ML
## 
## Empirical Reliability:
##    F1 
## 0.716

##       Comfort Work Future Benefit      F1  SE_F1
##  [1,]       1    1      1       1    -Inf     NA
##  [2,]       1    3      2       1 -2.1041 0.6891
##  [3,]       1    4      2       3 -1.1514 0.7049
##  [4,]       1    4      3       1 -0.7806 0.8071
##  [5,]       2    1      1       1 -4.4009 1.1827
##  [6,]       2    1      2       4 -1.8632 0.6678
##  [7,]       2    2      1       1 -3.2498 0.8132
##  [8,]       2    2      2       2 -1.9117 0.5875
##  [9,]       2    2      2       3 -1.6060 0.6084
## [10,]       2    2      3       1 -1.4676 0.7137
## [11,]       2    2      3       2 -1.1680 0.6347
## [12,]       2    2      3       3 -0.7965 0.6389
## [13,]       2    2      4       3  0.2416 0.7884
## [14,]       2    3      1       3 -2.1438 0.7015
## [15,]       2    3      2       2 -1.6090 0.6258
## [16,]       2    3      2       3 -1.2545 0.6302
## [17,]       2    3      3       2 -0.7556 0.6558
## [18,]       2    3      3       3 -0.3316 0.6981
## [19,]       2    3      3       4  0.0831 0.7614
## [20,]       2    3      4       1  0.1908 0.8775
## [21,]       2    3      4       3  0.7635 0.6805
## [22,]       2    4      2       1 -1.8224 0.6941
## [23,]       2    4      4       3  1.3082 0.7538
## [24,]       2    4      4       4  2.0969 1.0410
## [25,]       3    1      1       1 -3.5419 1.0039
## [26,]       3    1      1       3 -2.5111 0.6974
## [27,]       3    1      2       2 -1.9259 0.6147
## [28,]       3    1      2       3 -1.5874 0.6434
## [29,]       3    1      3       2 -1.0874 0.6740
## [30,]       3    1      3       3 -0.6581 0.6955
## [31,]       3    1      3       4 -0.2810 0.8073
## [32,]       3    1      4       3  0.5435 0.7336
## [33,]       3    1      4       4  1.0386 0.7358
## [34,]       3    2      1       2 -2.3492 0.6247
## [35,]       3    2      1       4 -1.9147 0.7090
## [36,]       3    2      2       1 -1.8616 0.6158
## [37,]       3    2      2       2 -1.5772 0.5939
## [38,]       3    2      2       3 -1.2555 0.6022
## [39,]       3    2      3       1 -1.0250 0.6626
## [40,]       3    2      3       2 -0.7977 0.6290
## [41,]       3    2      3       3 -0.4090 0.6677
## [42,]       3    2      3       4 -0.0375 0.7439
## [43,]       3    2      4       1  0.0164 0.9002
## [44,]       3    2      4       2  0.2067 0.7884
## [45,]       3    2      4       3  0.6439 0.6839
## [46,]       3    2      4       4  1.1024 0.7184
## [47,]       3    3      1       3 -1.6314 0.7798
## [48,]       3    3      2       1 -1.5181 0.6591
## [49,]       3    3      2       2 -1.2396 0.6168
## [50,]       3    3      2       3 -0.8832 0.6299
## [51,]       3    3      2       4 -0.6219 0.7078
## [52,]       3    3      3       1 -0.5490 0.7107
## [53,]       3    3      3       2 -0.3474 0.6875
## [54,]       3    3      3       3  0.0864 0.6869
## [55,]       3    3      3       4  0.4901 0.6761
## [56,]       3    3      4       2  0.7332 0.6810
## [57,]       3    3      4       3  1.0484 0.6499
## [58,]       3    3      4       4  1.5255 0.7460
## [59,]       3    4      1       3 -1.3845 0.9419
## [60,]       3    4      2       3 -0.6095 0.7253
## [61,]       3    4      3       2  0.0993 0.7677
## [62,]       3    4      3       3  0.5378 0.6773
## [63,]       3    4      3       4  0.9609 0.6745
## [64,]       3    4      4       1  1.2189 0.7586
## [65,]       3    4      4       2  1.2728 0.7493
## [66,]       3    4      4       3  1.6003 0.7668
## [67,]       3    4      4       4  2.4252 1.0420
## [68,]       4    1      1       4 -2.1074 0.7719
## [69,]       4    1      2       2 -1.6385 0.6540
## [70,]       4    1      2       4 -1.0151 0.7236
## [71,]       4    1      3       4  0.3610 0.7778
## [72,]       4    2      2       1 -1.5683 0.6512
## [73,]       4    2      2       3 -0.9359 0.6274
## [74,]       4    2      3       3  0.0535 0.7108
## [75,]       4    2      3       4  0.4915 0.7095
## [76,]       4    2      4       2  0.7631 0.7228
## [77,]       4    2      4       4  1.7611 0.9318
## [78,]       4    3      2       3 -0.4862 0.7035
## [79,]       4    3      3       2  0.1391 0.7181
## [80,]       4    3      3       3  0.5325 0.6522
## [81,]       4    3      3       4  0.9287 0.6554
## [82,]       4    3      4       2  1.2163 0.7195
## [83,]       4    3      4       3  1.5273 0.7347
## [84,]       4    3      4       4  2.2605 0.9692
## [85,]       4    4      3       2  0.6396 0.7069
## [86,]       4    4      3       3  0.9809 0.6613
## [87,]       4    4      3       4  1.4642 0.7489
## [88,]       4    4      4       2  2.0299 1.0117
## [89,]       4    4      4       3  2.3925 1.0167
## [90,]       4    4      4       4     Inf     NA

#WLE estimation, run in parallel using available cores
mirtCluster()

## mirtCluster() has already been defined

fscores(mod, method='WLE')

## 
## Method:  WLE
## 
## Empirical Reliability:
##     F1 
## 0.7513

##       Comfort Work Future Benefit      F1  SE_F1
##  [1,]       1    1      1       1 -5.6992 1.5787
##  [2,]       1    3      2       1 -2.1183 0.6330
##  [3,]       1    4      2       3 -1.1387 0.6554
##  [4,]       1    4      3       1 -0.8483 0.6995
##  [5,]       2    1      1       1 -4.0119 1.1430
##  [6,]       2    1      2       4 -1.8954 0.6696
##  [7,]       2    2      1       1 -3.0119 0.7173
##  [8,]       2    2      2       2 -1.9364 0.5815
##  [9,]       2    2      2       3 -1.6085 0.6503
## [10,]       2    2      3       1 -1.4470 0.7519
## [11,]       2    2      3       2 -1.1557 0.6000
## [12,]       2    2      3       3 -0.8405 0.5818
## [13,]       2    2      4       3  0.3564 0.7914
## [14,]       2    3      1       3 -2.1517 0.6355
## [15,]       2    3      2       2 -1.6121 0.6717
## [16,]       2    3      2       3 -1.2353 0.6113
## [17,]       2    3      3       2 -0.8084 0.5958
## [18,]       2    3      3       3 -0.4278 0.7154
## [19,]       2    3      3       4  0.1533 0.9073
## [20,]       2    3      4       1  0.3267 0.8910
## [21,]       2    3      4       3  0.7866 0.5966
## [22,]       2    4      2       1 -1.8560 0.7085
## [23,]       2    4      4       3  1.1906 0.6436
## [24,]       2    4      4       4  1.7589 0.8992
## [25,]       3    1      1       1 -3.1948 0.8598
## [26,]       3    1      1       3 -2.4394 0.6035
## [27,]       3    1      2       2 -1.9525 0.6040
## [28,]       3    1      2       3 -1.5873 0.6929
## [29,]       3    1      3       2 -1.0844 0.6214
## [30,]       3    1      3       3 -0.7329 0.6310
## [31,]       3    1      3       4 -0.4064 0.8474
## [32,]       3    1      4       3  0.6191 0.6503
## [33,]       3    1      4       4  0.9906 0.6262
## [34,]       3    2      1       2 -2.3201 0.5607
## [35,]       3    2      1       4 -1.9499 0.6894
## [36,]       3    2      2       1 -1.8889 0.6192
## [37,]       3    2      2       2 -1.5757 0.6318
## [38,]       3    2      2       3 -1.2377 0.5864
## [39,]       3    2      3       1 -1.0318 0.6064
## [40,]       3    2      3       2 -0.8404 0.5741
## [41,]       3    2      3       3 -0.4995 0.6579
## [42,]       3    2      3       4 -0.0390 0.9532
## [43,]       3    2      4       1  0.0777 1.2483
## [44,]       3    2      4       2  0.3182 0.8166
## [45,]       3    2      4       3  0.6932 0.6073
## [46,]       3    2      4       4  1.0400 0.6150
## [47,]       3    3      1       3 -1.6424 0.8701
## [48,]       3    3      2       1 -1.5070 0.7037
## [49,]       3    3      2       2 -1.2220 0.5968
## [50,]       3    3      2       3 -0.9120 0.5740
## [51,]       3    3      2       4 -0.7040 0.6399
## [52,]       3    3      3       1 -0.6420 0.6579
## [53,]       3    3      3       2 -0.4415 0.6980
## [54,]       3    3      3       3  0.1403 0.7953
## [55,]       3    3      3       4  0.5643 0.6213
## [56,]       3    3      4       2  0.7632 0.5986
## [57,]       3    3      4       3  1.0064 0.5712
## [58,]       3    3      4       4  1.3700 0.6561
## [59,]       3    4      1       3 -1.3351 0.9727
## [60,]       3    4      2       3 -0.6967 0.6530
## [61,]       3    4      3       2  0.1780 0.8994
## [62,]       3    4      3       3  0.6053 0.6142
## [63,]       3    4      3       4  0.9371 0.5877
## [64,]       3    4      4       1  1.1211 0.6433
## [65,]       3    4      4       2  1.1643 0.6388
## [66,]       3    4      4       3  1.4272 0.6768
## [67,]       3    4      4       4  2.0826 0.9256
## [68,]       4    1      1       4 -2.1239 0.6920
## [69,]       4    1      2       2 -1.6469 0.7059
## [70,]       4    1      2       4 -1.0248 0.6506
## [71,]       4    1      3       4  0.4715 0.7254
## [72,]       4    2      2       1 -1.5651 0.7013
## [73,]       4    2      2       3 -0.9560 0.5735
## [74,]       4    2      3       3  0.0994 0.8544
## [75,]       4    2      3       4  0.5720 0.6463
## [76,]       4    2      4       2  0.7885 0.6249
## [77,]       4    2      4       4  1.5067 0.7762
## [78,]       4    3      2       3 -0.5825 0.6604
## [79,]       4    3      3       2  0.2173 0.8035
## [80,]       4    3      3       3  0.5968 0.5979
## [81,]       4    3      3       4  0.9133 0.5750
## [82,]       4    3      4       2  1.1267 0.6181
## [83,]       4    3      4       3  1.3754 0.6486
## [84,]       4    3      4       4  1.9588 0.8741
## [85,]       4    4      3       2  0.6924 0.6230
## [86,]       4    4      3       3  0.9532 0.5787
## [87,]       4    4      3       4  1.3172 0.6531
## [88,]       4    4      4       2  1.7119 0.8687
## [89,]       4    4      4       3  2.0647 0.9075
## [90,]       4    4      4       4  3.3086 1.3473

#multiple imputation using 30 draws for EAP scores. Requires information matrix
mod <- mirt(Science, 1, SE=TRUE)
fscores(mod, MI = 30)

## 
## Method:  EAP
## 
## Empirical Reliability:
##     F1 
## 0.6646

##       Comfort Work Future Benefit      F1  SE_F1
##  [1,]       1    1      1       1 -2.7247 0.6379
##  [2,]       1    3      2       1 -1.4250 0.5809
##  [3,]       1    4      2       3 -0.7459 0.6232
##  [4,]       1    4      3       1 -0.4635 0.6621
##  [5,]       2    1      1       1 -2.5276 0.5961
##  [6,]       2    1      2       4 -1.2145 0.5835
##  [7,]       2    2      1       1 -2.1987 0.5795
##  [8,]       2    2      2       2 -1.3923 0.5196
##  [9,]       2    2      2       3 -1.1488 0.5345
## [10,]       2    2      3       1 -0.9651 0.5975
## [11,]       2    2      3       2 -0.7894 0.5659
## [12,]       2    2      3       3 -0.5057 0.5668
## [13,]       2    2      4       3  0.1775 0.6505
## [14,]       2    3      1       3 -1.3522 0.6586
## [15,]       2    3      2       2 -1.1540 0.5431
## [16,]       2    3      2       3 -0.8844 0.5622
## [17,]       2    3      3       2 -0.4962 0.5767
## [18,]       2    3      3       3 -0.2060 0.5719
## [19,]       2    3      3       4  0.0517 0.5946
## [20,]       2    3      4       1  0.0771 0.7083
## [21,]       2    3      4       3  0.5145 0.6150
## [22,]       2    4      2       1 -1.2067 0.5973
## [23,]       2    4      4       3  0.8542 0.6286
## [24,]       2    4      4       4  1.2319 0.6483
## [25,]       3    1      1       1 -2.2137 0.6017
## [26,]       3    1      1       3 -1.6212 0.6364
## [27,]       3    1      2       2 -1.3492 0.5381
## [28,]       3    1      2       3 -1.0831 0.5577
## [29,]       3    1      3       2 -0.6918 0.5854
## [30,]       3    1      3       3 -0.3866 0.5835
## [31,]       3    1      3       4 -0.1303 0.6110
## [32,]       3    1      4       3  0.3515 0.6433
## [33,]       3    1      4       4  0.6987 0.6442
## [34,]       3    2      1       2 -1.6363 0.5875
## [35,]       3    2      1       4 -1.1360 0.6885
## [36,]       3    2      2       1 -1.3264 0.5380
## [37,]       3    2      2       2 -1.1476 0.5230
## [38,]       3    2      2       3 -0.8922 0.5423
## [39,]       3    2      3       1 -0.6695 0.5834
## [40,]       3    2      3       2 -0.5213 0.5573
## [41,]       3    2      3       3 -0.2426 0.5565
## [42,]       3    2      3       4  0.0032 0.5815
## [43,]       3    2      4       1  0.0207 0.6887
## [44,]       3    2      4       2  0.1479 0.6447
## [45,]       3    2      4       3  0.4541 0.6068
## [46,]       3    2      4       4  0.7780 0.6157
## [47,]       3    3      1       3 -0.9826 0.6800
## [48,]       3    3      2       1 -1.0634 0.5652
## [49,]       3    3      2       2 -0.8912 0.5500
## [50,]       3    3      2       3 -0.6110 0.5727
## [51,]       3    3      2       4 -0.3782 0.6261
## [52,]       3    3      3       1 -0.3534 0.5889
## [53,]       3    3      3       2 -0.2272 0.5637
## [54,]       3    3      3       3  0.0481 0.5549
## [55,]       3    3      3       4  0.3099 0.5726
## [56,]       3    3      4       2  0.4850 0.6121
## [57,]       3    3      4       3  0.7492 0.5779
## [58,]       3    3      4       4  1.0664 0.5911
## [59,]       3    4      1       3 -0.7246 0.7510
## [60,]       3    4      2       3 -0.3821 0.6298
## [61,]       3    4      3       2  0.0253 0.5936
## [62,]       3    4      3       3  0.3126 0.5773
## [63,]       3    4      3       4  0.6081 0.5949
## [64,]       3    4      4       1  0.7564 0.6516
## [65,]       3    4      4       2  0.8235 0.6273
## [66,]       3    4      4       3  1.0759 0.5973
## [67,]       3    4      4       4  1.4378 0.6257
## [68,]       4    1      1       4 -1.0503 0.7894
## [69,]       4    1      2       2 -1.1088 0.5661
## [70,]       4    1      2       4 -0.5756 0.6584
## [71,]       4    1      3       4  0.2163 0.6181
## [72,]       4    2      2       1 -1.0847 0.5651
## [73,]       4    2      2       3 -0.6270 0.5783
## [74,]       4    2      3       3  0.0476 0.5661
## [75,]       4    2      3       4  0.3226 0.5905
## [76,]       4    2      4       2  0.5099 0.6327
## [77,]       4    2      4       4  1.1441 0.6263
## [78,]       4    3      2       3 -0.3102 0.6129
## [79,]       4    3      3       2  0.0696 0.5711
## [80,]       4    3      3       3  0.3389 0.5574
## [81,]       4    3      3       4  0.6203 0.5762
## [82,]       4    3      4       2  0.8278 0.6038
## [83,]       4    3      4       3  1.0686 0.5798
## [84,]       4    3      4       4  1.4167 0.6099
## [85,]       4    4      3       2  0.3549 0.5979
## [86,]       4    4      3       3  0.6258 0.5785
## [87,]       4    4      3       4  0.9407 0.6032
## [88,]       4    4      4       2  1.1989 0.6327
## [89,]       4    4      4       3  1.4286 0.6131
## [90,]       4    4      4       4  1.8416 0.6595

# plausible values for future work
pv <- fscores(mod, plausible.draws = 5)
lapply(pv, function(x) c(mean=mean(x), var=var(x), min=min(x), max=max(x)))

## [[1]]
##        mean         var         min         max 
##  0.01092245  1.00252941 -2.70895978  3.56168446 
## 
## [[2]]
##        mean         var         min         max 
##  0.02274287  0.98033503 -3.10107185  2.62536094 
## 
## [[3]]
##        mean         var         min         max 
## -0.03116357  0.93957061 -2.76893891  2.93205496 
## 
## [[4]]
##        mean         var         min         max 
## -0.01109465  0.94347942 -2.62062737  2.54214683 
## 
## [[5]]
##        mean         var         min         max 
## -0.03759606  1.01966220 -3.58598333  2.94641157

## define a custom_den function. EAP with a uniform prior between -3 and 3
fun <- function(Theta, ...) as.numeric(dunif(Theta, min = -3, max = 3))
head(fscores(mod, custom_den = fun))

## 
## Method:  EAP
## 
## Empirical Reliability:
##     F1 
## 0.7383

##      Comfort Work Future Benefit      F1  SE_F1
## [1,]       1    1      1       1 -2.7661 0.2822
## [2,]       1    3      2       1 -2.0028 0.6061
## [3,]       1    4      2       3 -1.1022 0.7532
## [4,]       1    4      3       1 -0.7839 0.8612
## [5,]       2    1      1       1 -2.7203 0.3114
## [6,]       2    1      2       4 -1.7939 0.6390

# custom MAP prior: standard truncated normal between 5 and -2
library(msm)
# need the :: scope for parallel to see the function (not require if no mirtCluster() defined)
fun <- function(Theta, ...) msm::dtnorm(Theta, mean = 0, sd = 1, lower = -2, upper = 5)
head(fscores(mod, custom_den = fun, method = 'MAP'))

## 
## Method:  MAP
## 
## Empirical Reliability:
##     F1 
## 0.6638

##      Comfort Work Future Benefit      F1  SE_F1
## [1,]       1    1      1       1 -2.0000 0.5954
## [2,]       1    3      2       1 -1.4232 0.5679
## [3,]       1    4      2       3 -0.7631 0.5918
## [4,]       1    4      3       1 -0.4618 0.6528
## [5,]       2    1      1       1 -2.0000 0.5705
## [6,]       2    1      2       4 -1.2729 0.5679

## End(No test)

[Package mirt version 1.9 Index]