Compute factor score estimates (a.k.a, ability estimates, latent trait estimates, etc)

Computes MAP, EAP, ML (Embretson & Reise, 2000), EAP for sum-scores (Thissen et al., 1995), or WLE (Warm, 1989) factor scores with a multivariate normal prior distribution using equally spaced quadrature. EAP scores for models with more than three factors are generally not recommended since the integration grid becomes very large, resulting in slower estimation and less precision if the quadpts are too low. Therefore, MAP scores should be used instead of EAP scores for higher dimensional models. Multiple imputation variants are possible for each estimator if a parameter information matrix was computed, which are useful if the sample size/number of items were small. As well, if the model contained latent regression predictors this information will be used in computing MAP and EAP estimates (for these models, full.scores=TRUE will always be used). Finally, plausible value imputation is also available, and will also account for latent regression predictor effects.

fscores(
  object,
  method = "EAP",
  full.scores = TRUE,
  rotate = "oblimin",
  Target = NULL,
  response.pattern = NULL,
  append_response.pattern = TRUE,
  na.rm = FALSE,
  plausible.draws = 0,
  plausible.type = "normal",
  quadpts = NULL,
  returnER = FALSE,
  return.acov = FALSE,
  mean = NULL,
  cov = NULL,
  verbose = TRUE,
  full.scores.SE = FALSE,
  theta_lim = c(-6, 6),
  MI = 0,
  use_dentype_estimate = FALSE,
  QMC = FALSE,
  custom_den = NULL,
  custom_theta = NULL,
  min_expected = 1,
  max_theta = 20,
  start = NULL,
  ...
)

Arguments

object
a computed model object of class SingleGroupClass, MultipleGroupClass, or DiscreteClass

method
type of factor score estimation method. Can be:

"EAP" for the expected a-posteriori (default)

"MAP" for the maximum a-posteriori (i.e, Bayes modal)

"ML" for maximum likelihood

"WLE" for weighted likelihood estimation

"EAPsum" for the expected a-posteriori for each sum score

"plausible" for a single plausible value imputation for each case. This is equivalent to setting plausible.draws = 1

"classify" for the posteriori classification probabilities (only applicable when the input model was of class MixtureClass)

full.scores
if FALSE then a summary table with factor scores for each unique pattern is displayed. Otherwise, a matrix of factor scores for each response pattern in the data is returned (default)

rotate
prior rotation to be used when estimating the factor scores. See summary-method for details. If the object is not an exploratory model then this argument is ignored

Target
target rotation; see summary-method for details

response.pattern
an optional argument used to calculate the factor scores and standard errors for a given response vector or matrix/data.frame

append_response.pattern
logical; should the inputs from response.pattern also be appended to the factor score output?

na.rm
logical; remove rows with any missing values? This is generally not required due to the nature of computing factors scores, however for the "EAPsum" method this may be necessary to ensure that the sum-scores correspond to the same composite score

plausible.draws
number of plausible values to draw for future researchers to perform secondary analyses of the latent trait scores. Typically used in conjunction with latent regression predictors (see mirt for details), but can also be generated when no predictor variables were modeled. If plausible.draws is greater than 0 a list of plausible values will be returned

plausible.type
type of plausible values to obtain. Can be either 'normal' (default) to use a normal approximation based on the ACOV matrix, or 'MH' to obtain Metropolis-Hastings samples from the posterior (silently passes object to mirt, therefore arguments like technical can be supplied to increase the number of burn-in draws and discarded samples)

quadpts
number of quadratures to use per dimension. If not specified, a suitable one will be created which decreases as the number of dimensions increases (and therefore for estimates such as EAP, will be less accurate). This is determined from the switch statement quadpts <- switch(as.character(nfact), '1'=61, '2'=31, '3'=15, '4'=9, '5'=7, 3)

returnER
logical; return empirical reliability (also known as marginal reliability) estimates as a numeric values?

return.acov
logical; return a list containing covariance matrices instead of factors scores? impute = TRUE not supported with this option

mean
a vector for custom latent variable means. If NULL, the default for 'group' values from the computed mirt object will be used

cov
a custom matrix of the latent variable covariance matrix. If NULL, the default for 'group' values from the computed mirt object will be used

verbose
logical; print verbose output messages?

full.scores.SE
logical; when full.scores == TRUE, also return the standard errors associated with each respondent? Default is FALSE

theta_lim
lower and upper range to evaluate latent trait integral for each dimension. If omitted, a range will be generated automatically based on the number of dimensions

MI
a number indicating how many multiple imputation draws to perform. Default is 0, indicating that no MI draws will be performed

use_dentype_estimate
logical; if the density of the latent trait was estimated in the model (e.g., via Davidian curves or empirical histograms), should this information be used to compute the latent trait estimates? Only applicable for EAP-based estimates (EAP, EAPsum, and plausible)

QMC
logical; use quasi-Monte Carlo integration? If quadpts is omitted the default number of nodes is 5000

custom_den
a function used to define the integration density (if required). The NULL default assumes that the multivariate normal distribution with the 'GroupPars' hyper-parameters are used. At the minimum must be of the form:

function(Theta, ...)

where Theta is a matrix of latent trait values (will be a grid of values if method == 'EAPsum' or method == 'EAP', otherwise Theta will have only 1 row). Additional arguments may included and are caught through the fscores(...) input. The function must return a numeric vector of density weights (one for each row in Theta)

custom_theta
a matrix of custom integration nodes to use instead of the default, where each column corresponds to the respective dimension in the model

min_expected
when computing goodness of fit tests when method = 'EAPsum', this value is used to collapse across the conditioned total scores until the expected values are greater than this value. Note that this only affect the goodness of fit tests and not the returned EAP for sum scores table

max_theta
the maximum/minimum value any given factor score estimate will achieve using any modal estimator method (e.g., MAP, WLE, ML)

start
a matrix of starting values to use for iterative estimation methods. Default will start at a vector of 0's for each response pattern, or will start at the EAP estimates (unidimensional models only). Must be in the form that matches full.scores = FALSE (mostly used in the mirtCAT package)

...
additional arguments to be passed to nlm

Details

The function will return either a table with the computed scores and standard errors, the original data matrix with scores appended to the rightmost column, or the scores only. By default the latent means and covariances are determined from the estimated object, though these can be overwritten. Iterative estimation methods can be estimated in parallel to decrease estimation times if a mirtCluster object is available.

If the input object is a discrete latent class object estimated from mdirt then the returned results will be with respect to the posterior classification for each individual. The method inputs for 'DiscreteClass' objects may only be 'EAP', for posterior classification of each response pattern, or 'EAPsum' for posterior classification based on the raw sum-score. For more information on these algorithms refer to the mirtCAT package and the associated JSS paper (Chalmers, 2016).

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi: 10.18637/jss.v048.i06

Chalmers, R. P. (2016). Generating Adaptive and Non-Adaptive Test Interfaces for Multidimensional Item Response Theory Applications. Journal of Statistical Software, 71(5), 1-39. doi: 10.18637/jss.v071.i05

Embretson, S. E. & Reise, S. P. (2000). Item Response Theory for Psychologists. Erlbaum.

Thissen, D., Pommerich, M., Billeaud, K., & Williams, V. S. L. (1995). Item Response Theory for Scores on Tests Including Polytomous Items with Ordered Responses. Applied Psychological Measurement, 19, 39-49.

Warm, T. A. (1989). Weighted likelihood estimation of ability in item response theory. Psychometrika, 54, 427-450.

See also

Author

Phil Chalmers rphilip.chalmers@gmail.com

Examples

mod <- mirt(Science, 1)
tabscores <- fscores(mod, full.scores = FALSE)

#> 
#> Method:  EAP
#> 
#> Empirical Reliability:
#> 
#>     F1 
#> 0.6666 

head(tabscores)

#>      Comfort Work Future Benefit         F1     SE_F1
#> [1,]       1    1      1       1 -2.7492687 0.6293569
#> [2,]       1    3      2       1 -1.4198318 0.5772364
#> [3,]       1    4      2       3 -0.7141976 0.6200139
#> [4,]       1    4      3       1 -0.4469265 0.6509531
#> [5,]       2    1      1       1 -2.5437808 0.5909117
#> [6,]       2    1      2       4 -1.2478570 0.5840105

# \donttest{
fullscores <- fscores(mod)
fullscores_with_SE <- fscores(mod, full.scores.SE=TRUE)
head(fullscores)

#>              F1
#> [1,]  0.4015613
#> [2,]  0.0520324
#> [3,] -0.8906436
#> [4,] -0.8906436
#> [5,]  0.7653806
#> [6,]  0.6695350

head(fullscores_with_SE)

#>              F1     SE_F1
#> [1,]  0.4015613 0.5978747
#> [2,]  0.0520324 0.5549554
#> [3,] -0.8906436 0.5421855
#> [4,] -0.8906436 0.5421855
#> [5,]  0.7653806 0.6385998
#> [6,]  0.6695350 0.5761860

#change method argument to use MAP estimates
fullscores <- fscores(mod, method='MAP')
head(fullscores)

#>               F1
#> [1,]  0.42140793
#> [2,]  0.05866659
#> [3,] -0.91936052
#> [4,] -0.91936052
#> [5,]  0.79013883
#> [6,]  0.68302194

#calculate MAP for a given response vector
fscores(mod, method='MAP', response.pattern = c(1,2,3,4))

#>      Comfort Work Future Benefit         F1     SE_F1
#> [1,]       1    2      3       4 -0.3704194 0.6054169

#or matrix
fscores(mod, method='MAP', response.pattern = rbind(c(1,2,3,4), c(2,2,1,3)))

#>      Comfort Work Future Benefit         F1     SE_F1
#> [1,]       1    2      3       4 -0.3704194 0.6054169
#> [2,]       2    2      1       3 -1.6937559 0.5773972

# return only the scores and their SEs
fscores(mod, method='MAP', response.pattern = c(1,2,3,4),
  append_response.pattern=FALSE)

#>              F1     SE_F1
#> [1,] -0.3704194 0.6054169

#use custom latent variable properties (diffuse prior for MAP is very close to ML)
fscores(mod, method='MAP', cov = matrix(1000), full.scores = FALSE)

#> 
#> Method:  MAP
#> 
#> Empirical Reliability:
#> 
#>     F1 
#> 0.4207 

#>       Comfort Work Future Benefit          F1      SE_F1
#>  [1,]       1    1      1       1 -9.33951652  9.6355561
#>  [2,]       1    3      2       1 -2.10399820  0.6892497
#>  [3,]       1    4      2       3 -1.15079772  0.7050200
#>  [4,]       1    4      3       1 -0.78096400  0.8072694
#>  [5,]       2    1      1       1 -4.39410045  1.1794941
#>  [6,]       2    1      2       4 -1.86282369  0.6679187
#>  [7,]       2    2      1       1 -3.24765372  0.8121324
#>  [8,]       2    2      2       2 -1.91138883  0.5875056
#>  [9,]       2    2      2       3 -1.60556797  0.6084689
#> [10,]       2    2      3       1 -1.46798831  0.7136838
#> [11,]       2    2      3       2 -1.16829291  0.6346994
#> [12,]       2    2      3       3 -0.79657516  0.6390082
#> [13,]       2    2      4       3  0.24025764  0.7886564
#> [14,]       2    3      1       3 -2.14238696  0.7017130
#> [15,]       2    3      2       2 -1.60857831  0.6258231
#> [16,]       2    3      2       3 -1.25398200  0.6302088
#> [17,]       2    3      3       2 -0.75571789  0.6558530
#> [18,]       2    3      3       3 -0.33161734  0.6981086
#> [19,]       2    3      3       4  0.08303458  0.7613508
#> [20,]       2    3      4       1  0.18902647  0.8780412
#> [21,]       2    3      4       3  0.76269417  0.6806576
#> [22,]       2    4      2       1 -1.82212924  0.6941785
#> [23,]       2    4      4       3  1.30730680  0.7534851
#> [24,]       2    4      4       4  2.09457768  1.0392353
#> [25,]       3    1      1       1 -3.53814277  1.0014201
#> [26,]       3    1      1       3 -2.50975652  0.6971827
#> [27,]       3    1      2       2 -1.92551599  0.6147422
#> [28,]       3    1      2       3 -1.58688138  0.6433970
#> [29,]       3    1      3       2 -1.08753431  0.6741187
#> [30,]       3    1      3       3 -0.65801199  0.6955365
#> [31,]       3    1      3       4 -0.28080385  0.8073110
#> [32,]       3    1      4       3  0.54246942  0.7338744
#> [33,]       3    1      4       4  1.03787099  0.7358268
#> [34,]       3    2      1       2 -2.34808003  0.6246704
#> [35,]       3    2      1       4 -1.91303987  0.7094593
#> [36,]       3    2      2       1 -1.86127492  0.6158454
#> [37,]       3    2      2       2 -1.57674751  0.5939631
#> [38,]       3    2      2       3 -1.25489834  0.6022571
#> [39,]       3    2      3       1 -1.02517060  0.6627051
#> [40,]       3    2      3       2 -0.79779682  0.6290335
#> [41,]       3    2      3       3 -0.40891429  0.6676663
#> [42,]       3    2      3       4 -0.03738214  0.7437682
#> [43,]       3    2      4       1  0.01467552  0.9004291
#> [44,]       3    2      4       2  0.20538260  0.7886293
#> [45,]       3    2      4       3  0.64303231  0.6840798
#> [46,]       3    2      4       4  1.10168772  0.7183789
#> [47,]       3    3      1       3 -1.62932695  0.7802080
#> [48,]       3    3      2       1 -1.51761901  0.6591228
#> [49,]       3    3      2       2 -1.23910216  0.6168897
#> [50,]       3    3      2       3 -0.88245744  0.6300829
#> [51,]       3    3      2       4 -0.62079890  0.7080556
#> [52,]       3    3      3       1 -0.54910678  0.7107281
#> [53,]       3    3      3       2 -0.34739935  0.6874561
#> [54,]       3    3      3       3  0.08637281  0.6868696
#> [55,]       3    3      3       4  0.49018549  0.6761531
#> [56,]       3    3      4       2  0.73237004  0.6812077
#> [57,]       3    3      4       3  1.04777614  0.6499124
#> [58,]       3    3      4       4  1.52472344  0.7455910
#> [59,]       3    4      1       3 -1.38148971  0.9427079
#> [60,]       3    4      2       3 -0.60836892  0.7256019
#> [61,]       3    4      3       2  0.09927069  0.7675852
#> [62,]       3    4      3       3  0.53785991  0.6774054
#> [63,]       3    4      3       4  0.96115397  0.6745446
#> [64,]       3    4      4       1  1.21807122  0.7584567
#> [65,]       3    4      4       2  1.27195710  0.7490778
#> [66,]       3    4      4       3  1.59934323  0.7662279
#> [67,]       3    4      4       4  2.42236673  1.0401223
#> [68,]       4    1      1       4 -2.10558080  0.7724116
#> [69,]       4    1      2       2 -1.63796061  0.6540725
#> [70,]       4    1      2       4 -1.01407917  0.7239107
#> [71,]       4    1      3       4  0.36120593  0.7777451
#> [72,]       4    2      2       1 -1.56782447  0.6512224
#> [73,]       4    2      2       3 -0.93515695  0.6275737
#> [74,]       4    2      3       3  0.05367208  0.7107570
#> [75,]       4    2      3       4  0.49170860  0.7095398
#> [76,]       4    2      4       2  0.76227140  0.7229547
#> [77,]       4    2      4       4  1.75983975  0.9306370
#> [78,]       4    3      2       3 -0.48507197  0.7037542
#> [79,]       4    3      3       2  0.13911259  0.7180795
#> [80,]       4    3      3       3  0.53266734  0.6522718
#> [81,]       4    3      3       4  0.92904321  0.6554712
#> [82,]       4    3      4       2  1.21563093  0.7193802
#> [83,]       4    3      4       3  1.52654526  0.7343333
#> [84,]       4    3      4       4  2.25842250  0.9678449
#> [85,]       4    4      3       2  0.63968593  0.7069825
#> [86,]       4    4      3       3  0.98114245  0.6613476
#> [87,]       4    4      3       4  1.46456801  0.7487625
#> [88,]       4    4      4       2  2.02777164  1.0101184
#> [89,]       4    4      4       3  2.38986103  1.0150510
#> [90,]       4    4      4       4  7.11188116 10.6159261

fscores(mod, method='ML', full.scores = FALSE)

#> 
#> Method:  ML
#> 
#> Empirical Reliability:
#> 
#>    F1 
#> 0.716 

#>       Comfort Work Future Benefit          F1     SE_F1
#>  [1,]       1    1      1       1        -Inf        NA
#>  [2,]       1    3      2       1 -2.10499828 0.6894269
#>  [3,]       1    4      2       3 -1.15136437 0.7052004
#>  [4,]       1    4      3       1 -0.78147320 0.8074821
#>  [5,]       2    1      1       1 -4.40023233 1.1822486
#>  [6,]       2    1      2       4 -1.86365477 0.6680213
#>  [7,]       2    2      1       1 -3.24979940 0.8130422
#>  [8,]       2    2      2       2 -1.91204974 0.5875929
#>  [9,]       2    2      2       3 -1.60616322 0.6085831
#> [10,]       2    2      3       1 -1.46873658 0.7139920
#> [11,]       2    2      3       2 -1.16876387 0.6348582
#> [12,]       2    2      3       3 -0.79690047 0.6391173
#> [13,]       2    2      4       3  0.24040721 0.7888702
#> [14,]       2    3      1       3 -2.14344342 0.7018391
#> [15,]       2    3      2       2 -1.60920919 0.6259497
#> [16,]       2    3      2       3 -1.25448025 0.6303475
#> [17,]       2    3      3       2 -0.75604300 0.6559689
#> [18,]       2    3      3       3 -0.33177894 0.6982616
#> [19,]       2    3      3       4  0.08308250 0.7615680
#> [20,]       2    3      4       1  0.18917228 0.8783298
#> [21,]       2    3      4       3  0.76304766 0.6807989
#> [22,]       2    4      2       1 -1.82300735 0.6942902
#> [23,]       2    4      4       3  1.30804997 0.7538642
#> [24,]       2    4      4       4  2.09684241 1.0407200
#> [25,]       3    1      1       1 -3.54170079 1.0035497
#> [26,]       3    1      1       3 -2.51097738 0.6975388
#> [27,]       3    1      2       2 -1.92624519 0.6148335
#> [28,]       3    1      2       3 -1.58753955 0.6435255
#> [29,]       3    1      3       2 -1.08802880 0.6742920
#> [30,]       3    1      3       3 -0.65833030 0.6956683
#> [31,]       3    1      3       4 -0.28098699 0.8075531
#> [32,]       3    1      4       3  0.54276314 0.7340180
#> [33,]       3    1      4       4  1.03843251 0.7360830
#> [34,]       3    2      1       2 -2.34899690 0.6248345
#> [35,]       3    2      1       4 -1.91400269 0.7094700
#> [36,]       3    2      2       1 -1.86198198 0.6159398
#> [37,]       3    2      2       2 -1.57730475 0.5940658
#> [38,]       3    2      2       3 -1.25535368 0.6023707
#> [39,]       3    2      3       1 -1.02562104 0.6628596
#> [40,]       3    2      3       2 -0.79811247 0.6291362
#> [41,]       3    2      3       3 -0.40909616 0.6677936
#> [42,]       3    2      3       4 -0.03740134 0.7439739
#> [43,]       3    2      4       1  0.01469413 0.9007890
#> [44,]       3    2      4       2  0.20551043 0.7888495
#> [45,]       3    2      4       3  0.64333323 0.6842145
#> [46,]       3    2      4       4  1.10225664 0.7186515
#> [47,]       3    3      1       3 -1.63031901 0.7802420
#> [48,]       3    3      2       1 -1.51827905 0.6592758
#> [49,]       3    3      2       2 -1.23957389 0.6170141
#> [50,]       3    3      2       3 -0.88280650 0.6301834
#> [51,]       3    3      2       4 -0.62111017 0.7081726
#> [52,]       3    3      3       1 -0.54938465 0.7108727
#> [53,]       3    3      3       2 -0.34756340 0.6876002
#> [54,]       3    3      3       3  0.08641366 0.6870296
#> [55,]       3    3      3       4  0.49040925 0.6762795
#> [56,]       3    3      4       2  0.73271002 0.6813470
#> [57,]       3    3      4       3  1.04821971 0.6500852
#> [58,]       3    3      4       4  1.52557212 0.7460043
#> [59,]       3    4      1       3 -1.38271827 0.9426967
#> [60,]       3    4      2       3 -0.60868923 0.7257241
#> [61,]       3    4      3       2  0.09932876 0.7678067
#> [62,]       3    4      3       3  0.53810625 0.6775295
#> [63,]       3    4      3       4  0.96159153 0.6747114
#> [64,]       3    4      4       1  1.21877250 0.7588065
#> [65,]       3    4      4       2  1.27267168 0.7494432
#> [66,]       3    4      4       3  1.60028348 0.7666937
#> [67,]       3    4      4       4  2.42499221 1.0416690
#> [68,]       4    1      1       4 -2.10683745 0.7724672
#> [69,]       4    1      2       2 -1.63866162 0.6541976
#> [70,]       4    1      2       4 -1.01461152 0.7240600
#> [71,]       4    1      3       4  0.36142674 0.7779367
#> [72,]       4    2      2       1 -1.56848943 0.6513580
#> [73,]       4    2      2       3 -0.93552371 0.6276734
#> [74,]       4    2      3       3  0.05369990 0.7109361
#> [75,]       4    2      3       4  0.49195637 0.7096903
#> [76,]       4    2      4       2  0.76267000 0.7231379
#> [77,]       4    2      4       4  1.76136607 0.9317187
#> [78,]       4    3      2       3 -0.48531220 0.7038789
#> [79,]       4    3      3       2  0.13918630 0.7182604
#> [80,]       4    3      3       3  0.53289398 0.6523889
#> [81,]       4    3      3       4  0.92944258 0.6556238
#> [82,]       4    3      4       2  1.21626054 0.7196884
#> [83,]       4    3      4       3  1.52736964 0.7347313
#> [84,]       4    3      4       4  2.26054125 0.9690341
#> [85,]       4    4      3       2  0.64000595 0.7071286
#> [86,]       4    4      3       3  0.98157196 0.6615132
#> [87,]       4    4      3       4  1.46539075 0.7491620
#> [88,]       4    4      4       2  2.02984488 1.0114985
#> [89,]       4    4      4       3  2.39232715 1.0164811
#> [90,]       4    4      4       4         Inf        NA

# EAPsum table of values based on total scores
fscores(mod, method = 'EAPsum', full.scores = FALSE)

#>       df       X2       p.X2    rxx_F1
#> stats 10 16.31202 0.09104204 0.6241528
#> 

#>    Sum.Scores          F1     SE_F1 observed   expected
#> 4           4 -2.74926868 0.6293569        2  0.1241449
#> 5           5 -2.43096708 0.6167669        1  0.7902670
#> 6           6 -2.08095804 0.6101826        2  2.7598787
#> 7           7 -1.71797996 0.6019326        1  7.2515944
#> 8           8 -1.36354348 0.5979810       11 15.8925644
#> 9           9 -1.01186764 0.6038596       32 29.6740255
#> 10         10 -0.64926712 0.6101301       58 48.3632492
#> 11         11 -0.28688279 0.6048405       70 68.4845962
#> 12         12  0.08240446 0.5996636       91 80.5530280
#> 13         13  0.48709473 0.6133534       56 65.6680929
#> 14         14  0.93406958 0.6170256       36 42.6372130
#> 15         15  1.38421709 0.6218261       20 22.3313442
#> 16         16  1.85351394 0.6544099       12  7.4700016

#WLE estimation, run in parallel using available cores
mirtCluster()

#> mirtCluster() has already been defined

head(fscores(mod, method='WLE', full.scores = FALSE))

#> 
#> Method:  WLE
#> 
#> Empirical Reliability:
#> 
#>     F1 
#> 0.7513 

#>      Comfort Work Future Benefit         F1     SE_F1
#> [1,]       1    1      1       1 -5.6980733 1.5782656
#> [2,]       1    3      2       1 -2.1191038 0.6332802
#> [3,]       1    4      2       3 -1.1387624 0.6557002
#> [4,]       1    4      3       1 -0.8489387 0.7000115
#> [5,]       2    1      1       1 -4.0112458 1.1423935
#> [6,]       2    1      2       4 -1.8957020 0.6698434

#multiple imputation using 30 draws for EAP scores. Requires information matrix
mod <- mirt(Science, 1, SE=TRUE)
fs <- fscores(mod, MI = 30)
head(fs)

#>               F1
#> [1,]  0.40797422
#> [2,]  0.03677095
#> [3,] -0.89622654
#> [4,] -0.89622654
#> [5,]  0.74916372
#> [6,]  0.66769600

# plausible values for future work
pv <- fscores(mod, plausible.draws = 5)
lapply(pv, function(x) c(mean=mean(x), var=var(x), min=min(x), max=max(x)))

#> [[1]]
#>        mean         var         min         max 
#>  0.01879158  1.02560651 -3.18083996  3.04179713 
#> 
#> [[2]]
#>       mean        var        min        max 
#> -0.0298637  1.0311903 -3.7932765  2.8635437 
#> 
#> [[3]]
#>        mean         var         min         max 
#>  0.00374964  1.01285130 -3.05922459  3.29830312 
#> 
#> [[4]]
#>       mean        var        min        max 
#>  0.0120314  1.0011140 -3.8553957  2.7213822 
#> 
#> [[5]]
#>        mean         var         min         max 
#> -0.03109173  0.94487520 -3.11421746  2.58403499 
#> 

## define a custom_den function. EAP with a uniform prior between -3 and 3
fun <- function(Theta, ...) as.numeric(dunif(Theta, min = -3, max = 3))
head(fscores(mod, custom_den = fun))

#>               F1
#> [1,]  0.62811597
#> [2,]  0.07362511
#> [3,] -1.23497086
#> [4,] -1.23497086
#> [5,]  1.25827841
#> [6,]  1.00860231

# custom MAP prior: standard truncated normal between 5 and -2
library(msm)

#> Error in library(msm): there is no package called ‘msm’

# need the :: scope for parallel to see the function (not require if no mirtCluster() defined)
fun <- function(Theta, ...) msm::dtnorm(Theta, mean = 0, sd = 1, lower = -2, upper = 5)
head(fscores(mod, custom_den = fun, method = 'MAP', full.scores = FALSE))

#> Error in loadNamespace(name): there is no package called ‘msm’

# }