imputeMissing {mirt}	R Documentation

Imputing plausible data for missing values

Description

Given an estimated model from any of mirt's model fitting functions and an estimate of the latent trait, impute plausible missing data values. Returns the original data in a data.frame without any NA values. If a list of Theta values is supplied then a list of complete datasets is returned instead.

Usage

imputeMissing(x, Theta, ...)

Arguments

x
an estimated model x from the mirt package

Theta
a matrix containing the estimates of the latent trait scores (e.g., via fscores)

...
additional arguments to pass

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

Examples

## No test: 
dat <- expand.table(LSAT7)
(original <- mirt(dat, 1))

## 
## Call:
## mirt(data = dat, model = 1)
## 
## Full-information item factor analysis with 1 factor(s).
## Converged within 1e-04 tolerance after 28 EM iterations.
## mirt version: 1.15 
## M-step optimizer: BFGS 
## EM acceleration: Ramsay
## Number of rectangular quadrature: 61
## 
## Log-likelihood = -2658.805
## AIC = 5337.61; AICc = 5337.833
## BIC = 5386.688; SABIC = 5354.927
## G2 (21) = 31.7, p = 0.0628
## RMSEA = 0.023, CFI = 0.939, TLI = 0.924

NAperson <- sample(1:nrow(dat), 20, replace = TRUE)
NAitem <- sample(1:ncol(dat), 20, replace = TRUE)
for(i in 1:20)
    dat[NAperson[i], NAitem[i]] <- NA
(mod <- mirt(dat, 1))

## 
## Call:
## mirt(data = dat, model = 1)
## 
## Full-information item factor analysis with 1 factor(s).
## Converged within 1e-04 tolerance after 31 EM iterations.
## mirt version: 1.15 
## M-step optimizer: BFGS 
## EM acceleration: Ramsay
## Number of rectangular quadrature: 61
## 
## Log-likelihood = -2646.876
## AIC = 5313.752; AICc = 5313.974
## BIC = 5362.83; SABIC = 5331.069

scores <- fscores(mod, method = 'MAP')

#re-estimate imputed dataset (good to do this multiple times and average over)
fulldata <- imputeMissing(mod, scores)
(fullmod <- mirt(fulldata, 1))

## 
## Call:
## mirt(data = fulldata, model = 1)
## 
## Full-information item factor analysis with 1 factor(s).
## Converged within 1e-04 tolerance after 28 EM iterations.
## mirt version: 1.15 
## M-step optimizer: BFGS 
## EM acceleration: Ramsay
## Number of rectangular quadrature: 61
## 
## Log-likelihood = -2656.843
## AIC = 5333.686; AICc = 5333.908
## BIC = 5382.763; SABIC = 5351.003
## G2 (21) = 29.83, p = 0.0956
## RMSEA = 0.021, CFI = 0.948, TLI = 0.936

#with multipleGroup
set.seed(1)
group <- sample(c('group1', 'group2'), 1000, TRUE)
mod2 <- multipleGroup(dat, 1, group, TOL=1e-2)
fs <- fscores(mod2)
fulldata2 <- imputeMissing(mod2, fs)

## End(No test)

[Package mirt version 1.15 Index]