itemstats {mirt}	R Documentation

Generic item summary statistics

Description

Function to compute generic item summary statistics that do not require prior fitting of IRT models. Contains information about coefficient alpha (and alpha if an item is deleted), mean/SD and frequency of total scores, reduced item-total correlations, average/sd of the correlation between items, response frequencies, and conditional mean/sd information given the unweighted sum scores.

Usage

itemstats(
  data,
  group = NULL,
  use_ts = TRUE,
  proportions = TRUE,
  ts.tables = FALSE
)

Arguments

data
An object of class data.frame or matrix with the response patterns

group
optional grouping variable to condition on when computing summary information

use_ts
logical; include information that is conditional on a meaningful total score?

proportions
logical; include response proportion information for each item?

ts.tables
logical; include mean/sd summary information pertaining to the unweighted total score?

Value

Returns a list containing the summary statistics

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi: 10.18637/jss.v048.i06

See Also

empirical_plot

Examples

# dichotomous data example
LSAT7full <- expand.table(LSAT7)
head(LSAT7full)

##   Item.1 Item.2 Item.3 Item.4 Item.5
## 1      0      0      0      0      0
## 2      0      0      0      0      0
## 3      0      0      0      0      0
## 4      0      0      0      0      0
## 5      0      0      0      0      0
## 6      0      0      0      0      0

itemstats(LSAT7full)

## $overall
##     N mean_total.score sd_total.score ave.r  sd.r alpha
##  1000            3.707          1.199 0.143 0.052 0.453
## 
## $itemstats
##           N  mean    sd total.r total.r_if_rm alpha_if_rm
## Item.1 1000 0.828 0.378   0.530         0.246       0.396
## Item.2 1000 0.658 0.475   0.600         0.247       0.394
## Item.3 1000 0.772 0.420   0.611         0.313       0.345
## Item.4 1000 0.606 0.489   0.592         0.223       0.415
## Item.5 1000 0.843 0.364   0.461         0.175       0.438
## 
## $proportions
##            0     1
## Item.1 0.172 0.828
## Item.2 0.342 0.658
## Item.3 0.228 0.772
## Item.4 0.394 0.606
## Item.5 0.157 0.843

# behaviour with missing data
LSAT7full[1:5,1] <- NA
itemstats(LSAT7full)

## $overall
##  N.complete    N mean_total.score sd_total.score ave.r  sd.r alpha
##         995 1000            3.726          1.172 0.137 0.052 0.426
## 
## $itemstats
##           N  mean    sd total.r total.r_if_rm alpha_if_rm
## Item.1  995 0.832 0.374   0.515         0.222       0.396
## Item.2 1000 0.658 0.475   0.595         0.232       0.364
## Item.3 1000 0.772 0.420   0.602         0.295       0.316
## Item.4 1000 0.606 0.489   0.589         0.209       0.384
## Item.5 1000 0.843 0.364   0.442         0.149       0.418
## 
## $proportions
##            0     1    NA
## Item.1 0.167 0.828 0.005
## Item.2 0.342 0.658    NA
## Item.3 0.228 0.772    NA
## Item.4 0.394 0.606    NA
## Item.5 0.157 0.843    NA

# data with no meaningful total score
head(SAT12)

##   Item.1 Item.2 Item.3 Item.4 Item.5 Item.6 Item.7 Item.8 Item.9 Item.10
## 1      1      4      5      2      3      1      2      1      3       1
## 2      3      4      2     NA      3      3      2     NA      3       1
## 3      1      4      5      4      3      2      2      3      3       2
## 4      2      4      4      2      3      3      2      4      3       2
## 5      2      4      5      2      3      2      2      1      1       2
## 6      1      4      3      1      3      2      2      3      3       1
##   Item.11 Item.12 Item.13 Item.14 Item.15 Item.16 Item.17 Item.18 Item.19
## 1       2       4       2       1       5       3       4       4       1
## 2       2      NA       2       1       5       2       4       1       1
## 3       2       1       3       1       5       5       4       1       3
## 4       2       4       2       1       5       2       4       1       3
## 5       2       4       2       1       5       4       4       5       1
## 6       2       3       2       1       5       5       4       4       1
##   Item.20 Item.21 Item.22 Item.23 Item.24 Item.25 Item.26 Item.27 Item.28
## 1       4       3       3       4       1       3       5       1       3
## 2       4       3       3      NA       1      NA       4       1       4
## 3       4       3       3       1       1       3       4       1       3
## 4       4       3       1       5       2       5       4       1       3
## 5       4       3       3       3       1       1       5       1       3
## 6       4       3       3       4       1       1       4       1       4
##   Item.29 Item.30 Item.31 Item.32
## 1       1       5       4       5
## 2       5      NA       4      NA
## 3       4       4       4       1
## 4       4       2       4       2
## 5       1       2       4       1
## 6       2       3       4       3

itemstats(SAT12, use_ts=FALSE)

## $overall
##   N.complete   N
## 1        572 600
## 
## $itemstats
##           N  mean    sd
## Item.1  599 2.487 1.168
## Item.2  599 3.377 1.344
## Item.3  592 3.147 1.439
## Item.4  595 2.718 1.289
## Item.5  599 2.860 0.887
## Item.6  600 2.358 1.135
## Item.7  599 2.412 0.880
## Item.8  598 2.908 1.340
## Item.9  600 2.907 0.567
## Item.10 598 2.301 1.456
## Item.11 600 2.017 0.199
## Item.12 595 3.605 1.119
## Item.13 600 2.317 0.956
## Item.14 598 1.778 1.389
## Item.15 599 4.529 1.078
## Item.16 599 3.361 1.120
## Item.17 600 3.968 0.343
## Item.18 597 2.995 1.476
## Item.19 600 1.900 1.053
## Item.20 599 3.863 0.453
## Item.21 599 2.928 0.514
## Item.22 600 2.985 0.442
## Item.23 597 2.729 1.392
## Item.24 599 1.491 1.003
## Item.25 595 2.696 1.299
## Item.26 599 3.917 1.255
## Item.27 598 1.217 0.660
## Item.28 597 3.238 0.877
## Item.29 595 2.237 1.202
## Item.30 594 3.660 1.499
## Item.31 599 3.781 0.883
## Item.32 593 2.965 1.193
## 
## $proportions
##             1     2     3     4     5    NA
## Item.1  0.283 0.203 0.267 0.232 0.013 0.002
## Item.2  0.212 0.022 0.070 0.568 0.127 0.002
## Item.3  0.165 0.183 0.260 0.098 0.280 0.013
## Item.4  0.165 0.378 0.148 0.172 0.128 0.008
## Item.5  0.093 0.143 0.620 0.093 0.048 0.002
## Item.6  0.160 0.582 0.107 0.043 0.108    NA
## Item.7  0.025 0.760 0.007 0.190 0.017 0.002
## Item.8  0.202 0.205 0.207 0.250 0.133 0.003
## Item.9  0.065 0.010 0.885 0.033 0.007    NA
## Item.10 0.422 0.215 0.165 0.028 0.167 0.003
## Item.11 0.003 0.983 0.008 0.003 0.002    NA
## Item.12 0.072 0.082 0.218 0.415 0.205 0.008
## Item.13 0.110 0.662 0.070 0.118 0.040    NA
## Item.14 0.723 0.027 0.108 0.022 0.117 0.003
## Item.15 0.035 0.062 0.060 0.025 0.817 0.002
## Item.16 0.070 0.105 0.413 0.215 0.195 0.002
## Item.17 0.008 0.005 0.010 0.963 0.013    NA
## Item.18 0.303 0.033 0.165 0.352 0.142 0.005
## Item.19 0.548 0.053 0.358 0.030 0.010    NA
## Item.20 0.012 0.002 0.105 0.873 0.007 0.002
## Item.21 0.050 0.008 0.915 0.013 0.012 0.002
## Item.22 0.028 0.005 0.935 0.017 0.015    NA
## Item.23 0.290 0.177 0.128 0.313 0.087 0.005
## Item.24 0.728 0.162 0.042 0.022 0.045 0.002
## Item.25 0.240 0.170 0.375 0.065 0.142 0.008
## Item.26 0.020 0.227 0.030 0.262 0.460 0.002
## Item.27 0.862 0.093 0.012 0.020 0.010 0.003
## Item.28 0.082 0.010 0.530 0.337 0.037 0.005
## Item.29 0.340 0.295 0.205 0.085 0.067 0.008
## Item.30 0.150 0.110 0.107 0.183 0.440 0.010
## Item.31 0.075 0.020 0.012 0.833 0.058 0.002
## Item.32 0.125 0.183 0.443 0.075 0.162 0.012

# extra total scores tables
dat <- key2binary(SAT12,
                   key = c(1,4,5,2,3,1,2,1,3,1,2,4,2,1,
                           5,3,4,4,1,4,3,3,4,1,3,5,1,3,1,5,4,5))
itemstats(dat, ts.tables=TRUE)

## $overall
##  N.complete   N mean_total.score sd_total.score ave.r  sd.r alpha
##         572 600           18.367           4.93 0.107 0.075 0.787
## 
## $itemstats
##           N  mean    sd total.r total.r_if_rm alpha_if_rm
## Item.1  599 0.284 0.451   0.398         0.317       0.780
## Item.2  599 0.569 0.496   0.521         0.442       0.774
## Item.3  592 0.284 0.451   0.454         0.377       0.777
## Item.4  595 0.382 0.486   0.339         0.248       0.783
## Item.5  599 0.621 0.486   0.418         0.331       0.779
## Item.6  600 0.160 0.367   0.415         0.351       0.779
## Item.7  599 0.761 0.427   0.359         0.281       0.782
## Item.8  598 0.202 0.402   0.305         0.227       0.784
## Item.9  600 0.885 0.319   0.166         0.103       0.788
## Item.10 598 0.423 0.494   0.460         0.375       0.777
## Item.11 600 0.983 0.128   0.163         0.139       0.787
## Item.12 595 0.418 0.494   0.150         0.051       0.795
## Item.13 600 0.662 0.474   0.418         0.334       0.779
## Item.14 598 0.726 0.447   0.399         0.319       0.780
## Item.15 599 0.818 0.386   0.376         0.307       0.781
## Item.16 599 0.414 0.493   0.356         0.264       0.783
## Item.17 600 0.963 0.188   0.208         0.173       0.786
## Item.18 597 0.353 0.478   0.576         0.505       0.773
## Item.19 600 0.548 0.498   0.387         0.296       0.781
## Item.20 599 0.875 0.331   0.354         0.295       0.781
## Item.21 599 0.917 0.277   0.181         0.127       0.787
## Item.22 600 0.935 0.247   0.263         0.217       0.784
## Item.23 597 0.315 0.465   0.332         0.244       0.784
## Item.24 599 0.730 0.445   0.424         0.346       0.779
## Item.25 595 0.378 0.485   0.387         0.298       0.781
## Item.26 599 0.461 0.499   0.549         0.472       0.772
## Item.27 598 0.865 0.342   0.407         0.348       0.779
## Item.28 597 0.533 0.499   0.465         0.380       0.777
## Item.29 595 0.343 0.475   0.420         0.336       0.780
## Item.30 594 0.444 0.497   0.229         0.131       0.790
## Item.31 599 0.835 0.372   0.453         0.391       0.778
## Item.32 593 0.164 0.370   0.094         0.019       0.794
## 
## $proportions
##             0     1    NA
## Item.1  0.715 0.283 0.002
## Item.2  0.430 0.568 0.002
## Item.3  0.707 0.280 0.013
## Item.4  0.613 0.378 0.008
## Item.5  0.378 0.620 0.002
## Item.6  0.840 0.160    NA
## Item.7  0.238 0.760 0.002
## Item.8  0.795 0.202 0.003
## Item.9  0.115 0.885    NA
## Item.10 0.575 0.422 0.003
## Item.11 0.017 0.983    NA
## Item.12 0.577 0.415 0.008
## Item.13 0.338 0.662    NA
## Item.14 0.273 0.723 0.003
## Item.15 0.182 0.817 0.002
## Item.16 0.585 0.413 0.002
## Item.17 0.037 0.963    NA
## Item.18 0.643 0.352 0.005
## Item.19 0.452 0.548    NA
## Item.20 0.125 0.873 0.002
## Item.21 0.083 0.915 0.002
## Item.22 0.065 0.935    NA
## Item.23 0.682 0.313 0.005
## Item.24 0.270 0.728 0.002
## Item.25 0.617 0.375 0.008
## Item.26 0.538 0.460 0.002
## Item.27 0.135 0.862 0.003
## Item.28 0.465 0.530 0.005
## Item.29 0.652 0.340 0.008
## Item.30 0.550 0.440 0.010
## Item.31 0.165 0.833 0.002
## Item.32 0.827 0.162 0.012
## 
## $total.score_frequency
##      5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27 28 29 30
## Freq 1 1 1 5 7 10 11 16 33 51 44 39 44 49 43 43 20 34 30 17 15 18 19  7  7  3
##      31 32
## Freq  1  3
## 
## $total.score_means
##                0        1
## Item.1  17.11794 21.44848
## Item.2  15.35000 20.54819
## Item.3  16.94853 21.89634
## Item.4  17.06197 20.50230
## Item.5  15.72685 19.96910
## Item.6  17.46555 23.01075
## Item.7  15.19853 19.35550
## Item.8  17.59292 21.28333
## Item.9  16.07692 18.66075
## Item.10 16.42727 21.01240
## Item.11 12.00000 18.46892
## Item.12 17.73795 19.23750
## Item.13 15.46073 19.82415
## Item.14 15.14194 19.56595
## Item.15 14.32323 19.21353
## Item.16 16.87009 20.42324
## Item.17 12.84211 18.55696
## Item.18 16.26287 22.19212
## Item.19 16.27413 20.09904
## Item.20 13.58209 19.00198
## Item.21 15.34783 18.63118
## Item.22 13.36111 18.70336
## Item.23 17.26904 20.79775
## Item.24 14.90850 19.63007
## Item.25 16.87288 20.79358
## Item.26 15.82895 21.24627
## Item.27 13.20000 19.14688
## Item.28 15.85769 20.45833
## Item.29 16.86096 21.21212
## Item.30 17.34395 19.61240
## Item.31 13.24176 19.33680
## Item.32 18.16109 19.41489
## 
## $total.score_sds
##                0        1
## Item.1  4.300401 5.041784
## Item.2  3.600674 4.604027
## Item.3  4.144775 4.968143
## Item.4  4.256811 5.213164
## Item.5  4.077667 4.712782
## Item.6  4.286102 5.420361
## Item.7  4.111087 4.747872
## Item.8  4.402453 5.685280
## Item.9  4.583625 4.900119
## Item.10 4.107505 4.730480
## Item.11 3.535534 4.884467
## Item.12 4.716815 5.093873
## Item.13 4.070080 4.676148
## Item.14 3.816299 4.759716
## Item.15 3.872531 4.705621
## Item.16 4.145612 5.182836
## Item.17 4.349498 4.841408
## Item.18 3.794125 4.437413
## Item.19 4.148193 4.858884
## Item.20 3.614572 4.730641
## Item.21 4.761967 4.860423
## Item.22 3.448832 4.833650
## Item.23 4.402024 5.173116
## Item.24 3.773801 4.695427
## Item.25 4.224332 5.037057
## Item.26 3.586326 4.661479
## Item.27 3.579219 4.628601
## Item.28 4.131868 4.555996
## Item.29 4.087767 5.135272
## Item.30 4.640031 4.994643
## Item.31 3.081633 4.605576
## Item.32 4.744322 5.697405

# grouping information
group <- gl(2, 300, labels=c('G1', 'G2'))
itemstats(dat, group=group)

## $G1
## $G1$overall
##  N.complete   N mean_total.score sd_total.score ave.r sd.r alpha
##         287 300           18.167          4.924 0.105 0.09 0.784
## 
## $G1$itemstats
##           N  mean    sd total.r total.r_if_rm alpha_if_rm
## Item.1  300 0.290 0.455   0.427         0.346       0.776
## Item.2  300 0.573 0.495   0.516         0.437       0.771
## Item.3  297 0.293 0.456   0.451         0.372       0.774
## Item.4  298 0.356 0.480   0.344         0.254       0.781
## Item.5  300 0.630 0.484   0.469         0.387       0.774
## Item.6  300 0.130 0.337   0.407         0.347       0.777
## Item.7  299 0.729 0.445   0.371         0.289       0.780
## Item.8  299 0.214 0.411   0.351         0.274       0.779
## Item.9  300 0.897 0.305   0.171         0.111       0.785
## Item.10 299 0.405 0.492   0.423         0.336       0.776
## Item.11 300 0.993 0.082   0.062         0.050       0.785
## Item.12 297 0.418 0.494   0.116         0.016       0.793
## Item.13 300 0.647 0.479   0.427         0.343       0.776
## Item.14 300 0.727 0.446   0.402         0.322       0.777
## Item.15 300 0.793 0.406   0.380         0.308       0.778
## Item.16 299 0.378 0.486   0.329         0.236       0.782
## Item.17 300 0.957 0.204   0.198         0.159       0.783
## Item.18 299 0.361 0.481   0.561         0.489       0.769
## Item.19 300 0.537 0.499   0.418         0.329       0.777
## Item.20 299 0.866 0.341   0.332         0.270       0.779
## Item.21 300 0.910 0.287   0.192         0.135       0.784
## Item.22 300 0.927 0.261   0.224         0.174       0.783
## Item.23 298 0.262 0.440   0.330         0.247       0.781
## Item.24 299 0.709 0.455   0.433         0.354       0.776
## Item.25 297 0.374 0.485   0.410         0.324       0.777
## Item.26 299 0.475 0.500   0.525         0.445       0.771
## Item.27 298 0.862 0.345   0.457         0.400       0.775
## Item.28 298 0.540 0.499   0.428         0.340       0.776
## Item.29 298 0.346 0.476   0.399         0.313       0.778
## Item.30 296 0.446 0.498   0.223         0.123       0.789
## Item.31 299 0.813 0.391   0.493         0.430       0.773
## Item.32 295 0.183 0.387   0.020        -0.060       0.797
## 
## $G1$proportions
##             0     1    NA
## Item.1  0.710 0.290    NA
## Item.2  0.427 0.573    NA
## Item.3  0.700 0.290 0.010
## Item.4  0.640 0.353 0.007
## Item.5  0.370 0.630    NA
## Item.6  0.870 0.130    NA
## Item.7  0.270 0.727 0.003
## Item.8  0.783 0.213 0.003
## Item.9  0.103 0.897    NA
## Item.10 0.593 0.403 0.003
## Item.11 0.007 0.993    NA
## Item.12 0.577 0.413 0.010
## Item.13 0.353 0.647    NA
## Item.14 0.273 0.727    NA
## Item.15 0.207 0.793    NA
## Item.16 0.620 0.377 0.003
## Item.17 0.043 0.957    NA
## Item.18 0.637 0.360 0.003
## Item.19 0.463 0.537    NA
## Item.20 0.133 0.863 0.003
## Item.21 0.090 0.910    NA
## Item.22 0.073 0.927    NA
## Item.23 0.733 0.260 0.007
## Item.24 0.290 0.707 0.003
## Item.25 0.620 0.370 0.010
## Item.26 0.523 0.473 0.003
## Item.27 0.137 0.857 0.007
## Item.28 0.457 0.537 0.007
## Item.29 0.650 0.343 0.007
## Item.30 0.547 0.440 0.013
## Item.31 0.187 0.810 0.003
## Item.32 0.803 0.180 0.017
## 
## 
## $G2
## $G2$overall
##  N.complete   N mean_total.score sd_total.score ave.r sd.r alpha
##         285 300           18.568          4.937  0.11 0.08  0.79
## 
## $G2$itemstats
##           N  mean    sd total.r total.r_if_rm alpha_if_rm
## Item.1  299 0.278 0.449   0.372         0.289       0.784
## Item.2  299 0.565 0.497   0.527         0.449       0.776
## Item.3  295 0.275 0.447   0.461         0.385       0.780
## Item.4  297 0.407 0.492   0.332         0.239       0.786
## Item.5  299 0.612 0.488   0.370         0.280       0.785
## Item.6  300 0.190 0.393   0.422         0.353       0.782
## Item.7  300 0.793 0.406   0.344         0.268       0.785
## Item.8  299 0.191 0.393   0.260         0.183       0.788
## Item.9  300 0.873 0.333   0.166         0.100       0.791
## Item.10 299 0.441 0.497   0.495         0.413       0.778
## Item.11 300 0.973 0.161   0.231         0.199       0.788
## Item.12 298 0.419 0.494   0.184         0.085       0.798
## Item.13 300 0.677 0.469   0.406         0.323       0.783
## Item.14 298 0.725 0.447   0.397         0.317       0.783
## Item.15 299 0.843 0.365   0.369         0.304       0.785
## Item.16 300 0.450 0.498   0.380         0.289       0.784
## Item.17 300 0.970 0.171   0.221         0.191       0.788
## Item.18 298 0.346 0.476   0.594         0.526       0.777
## Item.19 300 0.560 0.497   0.354         0.261       0.786
## Item.20 300 0.883 0.322   0.376         0.320       0.784
## Item.21 299 0.923 0.267   0.169         0.117       0.790
## Item.22 300 0.943 0.232   0.306         0.263       0.786
## Item.23 299 0.368 0.483   0.330         0.239       0.787
## Item.24 300 0.750 0.434   0.413         0.337       0.782
## Item.25 298 0.383 0.487   0.362         0.271       0.785
## Item.26 300 0.447 0.498   0.576         0.503       0.773
## Item.27 300 0.867 0.341   0.358         0.296       0.784
## Item.28 299 0.525 0.500   0.503         0.421       0.778
## Item.29 297 0.340 0.475   0.443         0.360       0.782
## Item.30 298 0.443 0.498   0.237         0.138       0.792
## Item.31 300 0.857 0.351   0.406         0.346       0.782
## Item.32 298 0.144 0.352   0.185         0.116       0.790
## 
## $G2$proportions
##             0     1    NA
## Item.1  0.720 0.277 0.003
## Item.2  0.433 0.563 0.003
## Item.3  0.713 0.270 0.017
## Item.4  0.587 0.403 0.010
## Item.5  0.387 0.610 0.003
## Item.6  0.810 0.190    NA
## Item.7  0.207 0.793    NA
## Item.8  0.807 0.190 0.003
## Item.9  0.127 0.873    NA
## Item.10 0.557 0.440 0.003
## Item.11 0.027 0.973    NA
## Item.12 0.577 0.417 0.007
## Item.13 0.323 0.677    NA
## Item.14 0.273 0.720 0.007
## Item.15 0.157 0.840 0.003
## Item.16 0.550 0.450    NA
## Item.17 0.030 0.970    NA
## Item.18 0.650 0.343 0.007
## Item.19 0.440 0.560    NA
## Item.20 0.117 0.883    NA
## Item.21 0.077 0.920 0.003
## Item.22 0.057 0.943    NA
## Item.23 0.630 0.367 0.003
## Item.24 0.250 0.750    NA
## Item.25 0.613 0.380 0.007
## Item.26 0.553 0.447    NA
## Item.27 0.133 0.867    NA
## Item.28 0.473 0.523 0.003
## Item.29 0.653 0.337 0.010
## Item.30 0.553 0.440 0.007
## Item.31 0.143 0.857    NA
## Item.32 0.850 0.143 0.007

#####
# polytomous data example
itemstats(Science)

## $overall
##    N mean_total.score sd_total.score ave.r  sd.r alpha
##  392           11.668          2.003 0.275 0.098 0.598
## 
## $itemstats
##           N  mean    sd total.r total.r_if_rm alpha_if_rm
## Comfort 392 3.120 0.588   0.596         0.352       0.552
## Work    392 2.722 0.807   0.666         0.332       0.567
## Future  392 2.990 0.757   0.748         0.488       0.437
## Benefit 392 2.837 0.802   0.684         0.363       0.541
## 
## $proportions
##             1     2     3     4
## Comfort 0.013 0.082 0.679 0.227
## Work    0.084 0.250 0.526 0.140
## Future  0.036 0.184 0.536 0.245
## Benefit 0.054 0.255 0.492 0.199

# polytomous data with missing
newScience <- Science
newScience[1:5,1] <- NA
itemstats(newScience)

## $overall
##  N.complete   N mean_total.score sd_total.score ave.r sd.r alpha
##         387 392           11.672          2.011 0.276  0.1 0.605
## 
## $itemstats
##           N  mean    sd total.r total.r_if_rm alpha_if_rm
## Comfort 387 3.119 0.590   0.596         0.352       0.552
## Work    392 2.722 0.807   0.666         0.335       0.576
## Future  392 2.990 0.757   0.749         0.491       0.449
## Benefit 392 2.837 0.802   0.695         0.382       0.538
## 
## $proportions
##             1     2     3     4    NA
## Comfort 0.013 0.082 0.668 0.224 0.013
## Work    0.084 0.250 0.526 0.140    NA
## Future  0.036 0.184 0.536 0.245    NA
## Benefit 0.054 0.255 0.492 0.199    NA

# unequal categories
newScience[,1] <- ifelse(Science[,1] == 1, NA, Science[,1])
itemstats(newScience)

## $overall
##  N.complete   N mean_total.score sd_total.score ave.r  sd.r alpha
##         387 392           11.731          1.917  0.26 0.092 0.572
## 
## $itemstats
##           N  mean    sd total.r total.r_if_rm alpha_if_rm
## Comfort 387 3.147 0.540   0.556         0.314       0.552
## Work    392 2.722 0.807   0.680         0.339       0.517
## Future  392 2.990 0.757   0.738         0.460       0.409
## Benefit 392 2.837 0.802   0.668         0.328       0.525
## 
## $proportions
##             1     2     3     4    NA
## Comfort    NA 0.082 0.679 0.227 0.013
## Work    0.084 0.250 0.526 0.140    NA
## Future  0.036 0.184 0.536 0.245    NA
## Benefit 0.054 0.255 0.492 0.199    NA

merged <- data.frame(LSAT7full[1:392,], Science)
itemstats(merged)

## $overall
##  N.complete   N mean_total.score sd_total.score ave.r  sd.r alpha
##         387 392           14.331          2.231 0.037 0.167 0.379
## 
## $itemstats
##           N  mean    sd total.r total.r_if_rm alpha_if_rm
## Item.1  387 0.568 0.496   0.193        -0.030       0.417
## Item.2  392 0.232 0.423   0.033        -0.156       0.443
## Item.3  392 0.605 0.490   0.216        -0.003       0.405
## Item.4  392 0.467 0.500   0.261         0.038       0.392
## Item.5  392 0.760 0.428   0.178        -0.011       0.402
## Comfort 392 3.120 0.588   0.527         0.295       0.286
## Work    392 2.722 0.807   0.620         0.314       0.251
## Future  392 2.990 0.757   0.680         0.421       0.185
## Benefit 392 2.837 0.802   0.608         0.299       0.261
## 
## $proportions
##             0     1     2     3     4    NA
## Item.1  0.426 0.561    NA    NA    NA 0.013
## Item.2  0.768 0.232    NA    NA    NA    NA
## Item.3  0.395 0.605    NA    NA    NA    NA
## Item.4  0.533 0.467    NA    NA    NA    NA
## Item.5  0.240 0.760    NA    NA    NA    NA
## Comfort    NA 0.013 0.082 0.679 0.227    NA
## Work       NA 0.084 0.250 0.526 0.140    NA
## Future     NA 0.036 0.184 0.536 0.245    NA
## Benefit    NA 0.054 0.255 0.492 0.199    NA

[Package mirt version 1.36.2 Index]