key2binary {mirt}	R Documentation

Convert response patterns to binary data

Description

The key2binary function will convert response pattern data to a dichotomous format, given a response key.

Usage

key2binary(fulldata, key)

Arguments

fulldata
an object of class data.frame, matrix, or table with the response patterns

key
a vector consisting of the 'correct' response to the items. Each value corresponds to each column in fulldata

Value

Returns a numeric matrix with all the response patterns in dichotomous format.

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

Examples

## No test: 
data(SAT12)
head(SAT12)

##   Item.1 Item.2 Item.3 Item.4 Item.5 Item.6 Item.7 Item.8 Item.9 Item.10
## 1      1      4      5      2      3      1      2      1      3       1
## 2      3      4      2      8      3      3      2      8      3       1
## 3      1      4      5      4      3      2      2      3      3       2
## 4      2      4      4      2      3      3      2      4      3       2
## 5      2      4      5      2      3      2      2      1      1       2
## 6      1      4      3      1      3      2      2      3      3       1
##   Item.11 Item.12 Item.13 Item.14 Item.15 Item.16 Item.17 Item.18 Item.19
## 1       2       4       2       1       5       3       4       4       1
## 2       2       8       2       1       5       2       4       1       1
## 3       2       1       3       1       5       5       4       1       3
## 4       2       4       2       1       5       2       4       1       3
## 5       2       4       2       1       5       4       4       5       1
## 6       2       3       2       1       5       5       4       4       1
##   Item.20 Item.21 Item.22 Item.23 Item.24 Item.25 Item.26 Item.27 Item.28
## 1       4       3       3       4       1       3       5       1       3
## 2       4       3       3       8       1       8       4       1       4
## 3       4       3       3       1       1       3       4       1       3
## 4       4       3       1       5       2       5       4       1       3
## 5       4       3       3       3       1       1       5       1       3
## 6       4       3       3       4       1       1       4       1       4
##   Item.29 Item.30 Item.31 Item.32
## 1       1       5       4       5
## 2       5       8       4       8
## 3       4       4       4       1
## 4       4       2       4       2
## 5       1       2       4       1
## 6       2       3       4       3

key <- c(1,4,5,2,3,1,2,1,3,1,2,4,2,1,5,3,4,4,1,4,3,3,4,1,3,5,1,3,1,5,4,5)

dicho.SAT12 <- key2binary(SAT12,key)

## End(No test)

[Package mirt version 1.9 Index]