lagrange {mirt}	R Documentation

Lagrange test for freeing parameters

Description

Lagrange (i.e., score) test to test whether parameters should be freed from a more constrained baseline model.

Usage

lagrange(mod, parnum, SE.type = "crossprod", type = "central", ...)

Arguments

mod
an estimated model

parnum
a vector, or list of vectors, containing one or more parameter locations/sets of locations to be tested. See objects returned from mod2values for the locations

SE.type
type of information matrix estimator to use. See mirt for further details

type
type of numerical algorithm passed to numerical_deriv to obtain the gradient terms

...
additional arguments to pass to mirt

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

See Also

wald

Examples

## No test: 
dat <- expand.table(LSAT7)
mod <- mirt(dat, 1, 'Rasch')
(values <- mod2values(mod))

##    group   item     class   name parnum     value lbound ubound   est
## 1    all Item.1      dich     a1      1 1.0000000   -Inf    Inf FALSE
## 2    all Item.1      dich      d      2 1.8678174   -Inf    Inf  TRUE
## 3    all Item.1      dich      g      3 0.0000000  0e+00      1 FALSE
## 4    all Item.1      dich      u      4 1.0000000  0e+00      1 FALSE
## 5    all Item.2      dich     a1      5 1.0000000   -Inf    Inf FALSE
## 6    all Item.2      dich      d      6 0.7907818   -Inf    Inf  TRUE
## 7    all Item.2      dich      g      7 0.0000000  0e+00      1 FALSE
## 8    all Item.2      dich      u      8 1.0000000  0e+00      1 FALSE
## 9    all Item.3      dich     a1      9 1.0000000   -Inf    Inf FALSE
## 10   all Item.3      dich      d     10 1.4606248   -Inf    Inf  TRUE
## 11   all Item.3      dich      g     11 0.0000000  0e+00      1 FALSE
## 12   all Item.3      dich      u     12 1.0000000  0e+00      1 FALSE
## 13   all Item.4      dich     a1     13 1.0000000   -Inf    Inf FALSE
## 14   all Item.4      dich      d     14 0.5213304   -Inf    Inf  TRUE
## 15   all Item.4      dich      g     15 0.0000000  0e+00      1 FALSE
## 16   all Item.4      dich      u     16 1.0000000  0e+00      1 FALSE
## 17   all Item.5      dich     a1     17 1.0000000   -Inf    Inf FALSE
## 18   all Item.5      dich      d     18 1.9924961   -Inf    Inf  TRUE
## 19   all Item.5      dich      g     19 0.0000000  0e+00      1 FALSE
## 20   all Item.5      dich      u     20 1.0000000  0e+00      1 FALSE
## 21   all  GROUP GroupPars MEAN_1     21 0.0000000   -Inf    Inf FALSE
## 22   all  GROUP GroupPars COV_11     22 1.0217669  1e-04    Inf  TRUE
##    prior.type prior_1 prior_2
## 1        none     NaN     NaN
## 2        none     NaN     NaN
## 3        none     NaN     NaN
## 4        none     NaN     NaN
## 5        none     NaN     NaN
## 6        none     NaN     NaN
## 7        none     NaN     NaN
## 8        none     NaN     NaN
## 9        none     NaN     NaN
## 10       none     NaN     NaN
## 11       none     NaN     NaN
## 12       none     NaN     NaN
## 13       none     NaN     NaN
## 14       none     NaN     NaN
## 15       none     NaN     NaN
## 16       none     NaN     NaN
## 17       none     NaN     NaN
## 18       none     NaN     NaN
## 19       none     NaN     NaN
## 20       none     NaN     NaN
## 21       none     NaN     NaN
## 22       none     NaN     NaN

#test all fixed slopes individually
parnum <- values$parnum[values$name == 'a1']
lagrange(mod, parnum)

##            X2 df          p
## 1  0.41507728  1 0.51940367
## 5  0.05290187  1 0.81808860
## 9  6.57950332  1 0.01031597
## 13 1.53306126  1 0.21565359
## 17 2.35942853  1 0.12452780

# compare to LR test for first two slopes
mod2 <- mirt(dat, 'F = 1-5
                   FREE = (1, a1)', 'Rasch')
coef(mod2, simplify=TRUE)$items

##           a1     d g u
## Item.1 1.155 1.938 0 1
## Item.2 1.000 0.785 0 1
## Item.3 1.000 1.450 0 1
## Item.4 1.000 0.518 0 1
## Item.5 1.000 1.979 0 1

anova(mod, mod2)

## 
## Model 1: mirt(data = dat, model = 1, itemtype = "Rasch")
## Model 2: mirt(data = dat, model = "F = 1-5\n                   FREE = (1, a1)", 
##     itemtype = "Rasch")

##        AIC     AICc    SABIC      BIC    logLik    X2  df      p
## 1 5341.802 5341.886 5352.192 5371.248 -2664.901   NaN NaN    NaN
## 2 5343.264 5343.377 5355.386 5377.619 -2664.632 0.537   1 0.4637

mod2 <- mirt(dat, 'F = 1-5
                   FREE = (2, a1)', 'Rasch')
coef(mod2, simplify=TRUE)$items

##           a1     d g u
## Item.1 1.000 1.874 0 1
## Item.2 0.948 0.781 0 1
## Item.3 1.000 1.466 0 1
## Item.4 1.000 0.523 0 1
## Item.5 1.000 1.999 0 1

anova(mod, mod2)

## 
## Model 1: mirt(data = dat, model = 1, itemtype = "Rasch")
## Model 2: mirt(data = dat, model = "F = 1-5\n                   FREE = (2, a1)", 
##     itemtype = "Rasch")

##        AIC     AICc    SABIC      BIC    logLik    X2  df      p
## 1 5341.802 5341.886 5352.192 5371.248 -2664.901   NaN NaN    NaN
## 2 5343.721 5343.833 5355.842 5378.075 -2664.860 0.081   1 0.7759

mod2 <- mirt(dat, 'F = 1-5
                   FREE = (3, a1)', 'Rasch')
coef(mod2, simplify=TRUE)$items

##           a1     d g u
## Item.1 1.000 1.811 0 1
## Item.2 1.000 0.765 0 1
## Item.3 1.831 1.771 0 1
## Item.4 1.000 0.505 0 1
## Item.5 1.000 1.933 0 1

anova(mod, mod2)

## 
## Model 1: mirt(data = dat, model = 1, itemtype = "Rasch")
## Model 2: mirt(data = dat, model = "F = 1-5\n                   FREE = (3, a1)", 
##     itemtype = "Rasch")

##        AIC     AICc    SABIC      BIC    logLik    X2  df      p
## 1 5341.802 5341.886 5352.192 5371.248 -2664.901   NaN NaN    NaN
## 2 5335.133 5335.246 5347.255 5369.487 -2660.566 8.669   1 0.0032

# test slopes first two slopes and last three slopes jointly
lagrange(mod, list(parnum[1:2], parnum[3:5]))

##                 X2 df           p
## 1.5      0.5016988  2 0.778139545
## 9.13.17 13.7707674  3 0.003234379

# test all 5 slopes and first + last jointly
lagrange(mod, list(parnum[1:5], parnum[c(1, 5)]))

##                    X2 df          p
## 1.5.9.13.17 14.394180  5 0.01329021
## 1.17         3.151478  2 0.20685460

## End(No test)

[Package mirt version 1.20 Index]