Convert ordered Likert-scale responses (character or factors) to integers

Given a matrix or data.frame object consisting of Likert responses return an object of the same dimensions with integer values.

likert2int(x, levels = NULL)

Arguments

x
a matrix of character values or data.frame of character/factor vectors

levels
a named character vector indicating which integer values should be assigned to which elements. If omitted, the order of the elements will be determined after converting each column in x to a factor variable

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi: 10.18637/jss.v048.i06

See also

Author

Phil Chalmers rphilip.chalmers@gmail.com

Examples

# \donttest{

#simulate data

dat1 <- matrix(sample(c('Disagree', 'Strongly Disagree', 'Agree',
                        'Neutral', 'Strongly Agree'), 1000*5, replace=TRUE),
               nrow=1000, ncol=5)
dat1[2,2] <- dat1[3,3] <- dat1[1,3] <- NA # NAs added for flavour
dat2 <- matrix(sample(c('D', 'SD', 'A', 'N', 'SA'), 1000*5, replace=TRUE),
               nrow=1000, ncol=5)
dat <- cbind(dat1, dat2)

# separately
intdat1 <- likert2int(dat1)
head(dat1)

#>      [,1]                [,2]                [,3]               
#> [1,] "Neutral"           "Agree"             NA                 
#> [2,] "Disagree"          NA                  "Agree"            
#> [3,] "Disagree"          "Strongly Agree"    NA                 
#> [4,] "Strongly Agree"    "Neutral"           "Strongly Disagree"
#> [5,] "Disagree"          "Strongly Disagree" "Strongly Disagree"
#> [6,] "Strongly Disagree" "Neutral"           "Strongly Agree"   
#>      [,4]                [,5]            
#> [1,] "Strongly Agree"    "Agree"         
#> [2,] "Disagree"          "Neutral"       
#> [3,] "Strongly Disagree" "Agree"         
#> [4,] "Strongly Disagree" "Agree"         
#> [5,] "Agree"             "Strongly Agree"
#> [6,] "Neutral"           "Agree"         

head(intdat1)

#>   V1 V2 V3 V4 V5
#> 1  3  1 NA  4  1
#> 2  2 NA  1  2  3
#> 3  2  4 NA  5  1
#> 4  4  3  5  5  1
#> 5  2  5  5  1  4
#> 6  5  3  4  3  1

# more useful with explicit levels
lvl1 <- c('Strongly Disagree'=1, 'Disagree'=2, 'Neutral'=3, 'Agree'=4,
          'Strongly Agree'=5)
intdat1 <- likert2int(dat1, levels = lvl1)
head(dat1)

#>      [,1]                [,2]                [,3]               
#> [1,] "Neutral"           "Agree"             NA                 
#> [2,] "Disagree"          NA                  "Agree"            
#> [3,] "Disagree"          "Strongly Agree"    NA                 
#> [4,] "Strongly Agree"    "Neutral"           "Strongly Disagree"
#> [5,] "Disagree"          "Strongly Disagree" "Strongly Disagree"
#> [6,] "Strongly Disagree" "Neutral"           "Strongly Agree"   
#>      [,4]                [,5]            
#> [1,] "Strongly Agree"    "Agree"         
#> [2,] "Disagree"          "Neutral"       
#> [3,] "Strongly Disagree" "Agree"         
#> [4,] "Strongly Disagree" "Agree"         
#> [5,] "Agree"             "Strongly Agree"
#> [6,] "Neutral"           "Agree"         

head(intdat1)

#>   V1 V2 V3 V4 V5
#> 1  3  4 NA  5  4
#> 2  2 NA  4  2  3
#> 3  2  5 NA  1  4
#> 4  5  3  1  1  4
#> 5  2  1  1  4  5
#> 6  1  3  5  3  4

# second data
lvl2 <- c('SD'=1, 'D'=2, 'N'=3, 'A'=4, 'SA'=5)
intdat2 <- likert2int(dat2, levels = lvl2)
head(dat2)

#>      [,1] [,2] [,3] [,4] [,5]
#> [1,] "SD" "SD" "SD" "SA" "SA"
#> [2,] "N"  "D"  "A"  "SD" "SD"
#> [3,] "SA" "N"  "A"  "N"  "A" 
#> [4,] "D"  "N"  "A"  "A"  "SD"
#> [5,] "SA" "N"  "D"  "SD" "SA"
#> [6,] "A"  "SD" "D"  "A"  "SD"

head(intdat2)

#>   V1 V2 V3 V4 V5
#> 1  1  1  1  5  5
#> 2  3  2  4  1  1
#> 3  5  3  4  3  4
#> 4  2  3  4  4  1
#> 5  5  3  2  1  5
#> 6  4  1  2  4  1

# full dataset (using both mapping schemes)
intdat <- likert2int(dat, levels = c(lvl1, lvl2))
head(dat)

#>      [,1]                [,2]                [,3]               
#> [1,] "Neutral"           "Agree"             NA                 
#> [2,] "Disagree"          NA                  "Agree"            
#> [3,] "Disagree"          "Strongly Agree"    NA                 
#> [4,] "Strongly Agree"    "Neutral"           "Strongly Disagree"
#> [5,] "Disagree"          "Strongly Disagree" "Strongly Disagree"
#> [6,] "Strongly Disagree" "Neutral"           "Strongly Agree"   
#>      [,4]                [,5]             [,6] [,7] [,8] [,9] [,10]
#> [1,] "Strongly Agree"    "Agree"          "SD" "SD" "SD" "SA" "SA" 
#> [2,] "Disagree"          "Neutral"        "N"  "D"  "A"  "SD" "SD" 
#> [3,] "Strongly Disagree" "Agree"          "SA" "N"  "A"  "N"  "A"  
#> [4,] "Strongly Disagree" "Agree"          "D"  "N"  "A"  "A"  "SD" 
#> [5,] "Agree"             "Strongly Agree" "SA" "N"  "D"  "SD" "SA" 
#> [6,] "Neutral"           "Agree"          "A"  "SD" "D"  "A"  "SD" 

head(intdat)

#>   V1 V2 V3 V4 V5 V6 V7 V8 V9 V10
#> 1  3  4 NA  5  4  1  1  1  5   5
#> 2  2 NA  4  2  3  3  2  4  1   1
#> 3  2  5 NA  1  4  5  3  4  3   4
#> 4  5  3  1  1  4  2  3  4  4   1
#> 5  2  1  1  4  5  5  3  2  1   5
#> 6  1  3  5  3  4  4  1  2  4   1

#####
# data.frame as input with ordered factors

dat1 <- data.frame(dat1)
dat2 <- data.frame(dat2)
dat.old <- cbind(dat1, dat2)
colnames(dat.old) <- paste0('Item_', 1:10)
str(dat.old) # factors are leveled alphabetically by default

#> 'data.frame':	1000 obs. of  10 variables:
#>  $ Item_1 : Factor w/ 5 levels "Agree","Disagree",..: 3 2 2 4 2 5 1 1 1 4 ...
#>  $ Item_2 : Factor w/ 5 levels "Agree","Disagree",..: 1 NA 4 3 5 3 1 3 1 2 ...
#>  $ Item_3 : Factor w/ 5 levels "Agree","Disagree",..: NA 1 NA 5 5 4 5 1 1 2 ...
#>  $ Item_4 : Factor w/ 5 levels "Agree","Disagree",..: 4 2 5 5 1 3 5 5 5 1 ...
#>  $ Item_5 : Factor w/ 5 levels "Agree","Disagree",..: 1 3 1 1 4 1 4 1 2 5 ...
#>  $ Item_6 : Factor w/ 5 levels "A","D","N","SA",..: 5 3 4 2 4 1 4 5 3 4 ...
#>  $ Item_7 : Factor w/ 5 levels "A","D","N","SA",..: 5 2 3 3 3 5 5 4 5 3 ...
#>  $ Item_8 : Factor w/ 5 levels "A","D","N","SA",..: 5 1 1 1 2 2 4 3 4 2 ...
#>  $ Item_9 : Factor w/ 5 levels "A","D","N","SA",..: 4 5 3 1 5 1 2 3 5 5 ...
#>  $ Item_10: Factor w/ 5 levels "A","D","N","SA",..: 4 5 1 5 4 5 3 1 5 4 ...

# create explicit ordering in factor variables
for(i in 1:ncol(dat1))
   levels(dat1[[i]]) <- c('Strongly Disagree', 'Disagree', 'Neutral', 'Agree',
                          'Strongly Agree')

for(i in 1:ncol(dat2))
   levels(dat2[[i]]) <- c('SD', 'D', 'N', 'A', 'SA')

dat <- cbind(dat1, dat2)
colnames(dat) <- colnames(dat.old)
str(dat) # note ordering

#> 'data.frame':	1000 obs. of  10 variables:
#>  $ Item_1 : Factor w/ 5 levels "Strongly Disagree",..: 3 2 2 4 2 5 1 1 1 4 ...
#>  $ Item_2 : Factor w/ 5 levels "Strongly Disagree",..: 1 NA 4 3 5 3 1 3 1 2 ...
#>  $ Item_3 : Factor w/ 5 levels "Strongly Disagree",..: NA 1 NA 5 5 4 5 1 1 2 ...
#>  $ Item_4 : Factor w/ 5 levels "Strongly Disagree",..: 4 2 5 5 1 3 5 5 5 1 ...
#>  $ Item_5 : Factor w/ 5 levels "Strongly Disagree",..: 1 3 1 1 4 1 4 1 2 5 ...
#>  $ Item_6 : Factor w/ 5 levels "SD","D","N","A",..: 5 3 4 2 4 1 4 5 3 4 ...
#>  $ Item_7 : Factor w/ 5 levels "SD","D","N","A",..: 5 2 3 3 3 5 5 4 5 3 ...
#>  $ Item_8 : Factor w/ 5 levels "SD","D","N","A",..: 5 1 1 1 2 2 4 3 4 2 ...
#>  $ Item_9 : Factor w/ 5 levels "SD","D","N","A",..: 4 5 3 1 5 1 2 3 5 5 ...
#>  $ Item_10: Factor w/ 5 levels "SD","D","N","A",..: 4 5 1 5 4 5 3 1 5 4 ...

intdat <- likert2int(dat)
head(dat)

#>           Item_1            Item_2            Item_3            Item_4
#> 1        Neutral Strongly Disagree              <NA>             Agree
#> 2       Disagree              <NA> Strongly Disagree          Disagree
#> 3       Disagree             Agree              <NA>    Strongly Agree
#> 4          Agree           Neutral    Strongly Agree    Strongly Agree
#> 5       Disagree    Strongly Agree    Strongly Agree Strongly Disagree
#> 6 Strongly Agree           Neutral             Agree           Neutral
#>              Item_5 Item_6 Item_7 Item_8 Item_9 Item_10
#> 1 Strongly Disagree     SA     SA     SA      A       A
#> 2           Neutral      N      D     SD     SA      SA
#> 3 Strongly Disagree      A      N     SD      N      SD
#> 4 Strongly Disagree      D      N     SD     SD      SA
#> 5             Agree      A      N      D     SA       A
#> 6 Strongly Disagree     SD     SA      D     SD      SA

head(intdat)

#>   Item_1 Item_2 Item_3 Item_4 Item_5 Item_6 Item_7 Item_8 Item_9 Item_10
#> 1      3      1     NA      4      1      5      5      5      4       4
#> 2      2     NA      1      2      3      3      2      1      5       5
#> 3      2      4     NA      5      1      4      3      1      3       1
#> 4      4      3      5      5      1      2      3      1      1       5
#> 5      2      5      5      1      4      4      3      2      5       4
#> 6      5      3      4      3      1      1      5      2      1       5

# }