likert2int {mirt}	R Documentation

Convert ordered Likert-scale responses (character or factors) to integers

Description

Given a matrix or data.frame object consisting of Likert responses return an object of the same dimensions with integer values.

Usage

likert2int(x, levels = NULL)

Arguments

x
a matrix of character values or data.frame of character/factor vectors

levels
a named character vector indicating which integer values should be assigned to which elements. If omitted, the order of the elements will be determined after converting each column in x to a factor variable

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi: 10.18637/jss.v048.i06

See Also

key2binary, poly2dich

Examples

## No test: 

#simulate data

dat1 <- matrix(sample(c('Disagree', 'Strongly Disagree', 'Agree',
                        'Neutral', 'Strongly Agree'), 1000*5, replace=TRUE),
               nrow=1000, ncol=5)
dat1[2,2] <- dat1[3,3] <- dat1[1,3] <- NA # NAs added for flavour
dat2 <- matrix(sample(c('D', 'SD', 'A', 'N', 'SA'), 1000*5, replace=TRUE),
               nrow=1000, ncol=5)
dat <- cbind(dat1, dat2)

# separately
intdat1 <- likert2int(dat1)
head(dat1)

##      [,1]                [,2]                [,3]               
## [1,] "Agree"             "Strongly Agree"    NA                 
## [2,] "Neutral"           NA                  "Neutral"          
## [3,] "Strongly Agree"    "Disagree"          NA                 
## [4,] "Disagree"          "Strongly Disagree" "Agree"            
## [5,] "Strongly Disagree" "Agree"             "Strongly Disagree"
## [6,] "Strongly Disagree" "Strongly Disagree" "Neutral"          
##      [,4]             [,5]               
## [1,] "Disagree"       "Agree"            
## [2,] "Strongly Agree" "Strongly Disagree"
## [3,] "Strongly Agree" "Neutral"          
## [4,] "Disagree"       "Strongly Agree"   
## [5,] "Disagree"       "Strongly Agree"   
## [6,] "Agree"          "Disagree"

head(intdat1)

##   V1 V2 V3 V4 V5
## 1 NA NA NA NA NA
## 2 NA NA NA NA NA
## 3 NA NA NA NA NA
## 4 NA NA NA NA NA
## 5 NA NA NA NA NA
## 6 NA NA NA NA NA

# more useful with explicit levels
lvl1 <- c('Strongly Disagree'=1, 'Disagree'=2, 'Neutral'=3, 'Agree'=4,
          'Strongly Agree'=5)
intdat1 <- likert2int(dat1, levels = lvl1)
head(dat1)

##      [,1]                [,2]                [,3]               
## [1,] "Agree"             "Strongly Agree"    NA                 
## [2,] "Neutral"           NA                  "Neutral"          
## [3,] "Strongly Agree"    "Disagree"          NA                 
## [4,] "Disagree"          "Strongly Disagree" "Agree"            
## [5,] "Strongly Disagree" "Agree"             "Strongly Disagree"
## [6,] "Strongly Disagree" "Strongly Disagree" "Neutral"          
##      [,4]             [,5]               
## [1,] "Disagree"       "Agree"            
## [2,] "Strongly Agree" "Strongly Disagree"
## [3,] "Strongly Agree" "Neutral"          
## [4,] "Disagree"       "Strongly Agree"   
## [5,] "Disagree"       "Strongly Agree"   
## [6,] "Agree"          "Disagree"

head(intdat1)

##   V1 V2 V3 V4 V5
## 1  4  5 NA  2  4
## 2  3 NA  3  5  1
## 3  5  2 NA  5  3
## 4  2  1  4  2  5
## 5  1  4  1  2  5
## 6  1  1  3  4  2

# second data
lvl2 <- c('SD'=1, 'D'=2, 'N'=3, 'A'=4, 'SA'=5)
intdat2 <- likert2int(dat2, levels = lvl2)
head(dat2)

##      [,1] [,2] [,3] [,4] [,5]
## [1,] "N"  "SD" "SA" "N"  "N" 
## [2,] "SA" "SD" "SD" "SD" "SD"
## [3,] "N"  "N"  "A"  "D"  "A" 
## [4,] "SA" "D"  "D"  "N"  "SD"
## [5,] "N"  "SA" "A"  "A"  "SA"
## [6,] "SD" "N"  "SD" "N"  "A"

head(intdat2)

##   V1 V2 V3 V4 V5
## 1  3  1  5  3  3
## 2  5  1  1  1  1
## 3  3  3  4  2  4
## 4  5  2  2  3  1
## 5  3  5  4  4  5
## 6  1  3  1  3  4

# full dataset (using both mapping schemes)
intdat <- likert2int(dat, levels = c(lvl1, lvl2))
head(dat)

##      [,1]                [,2]                [,3]               
## [1,] "Agree"             "Strongly Agree"    NA                 
## [2,] "Neutral"           NA                  "Neutral"          
## [3,] "Strongly Agree"    "Disagree"          NA                 
## [4,] "Disagree"          "Strongly Disagree" "Agree"            
## [5,] "Strongly Disagree" "Agree"             "Strongly Disagree"
## [6,] "Strongly Disagree" "Strongly Disagree" "Neutral"          
##      [,4]             [,5]                [,6] [,7] [,8] [,9] [,10]
## [1,] "Disagree"       "Agree"             "N"  "SD" "SA" "N"  "N"  
## [2,] "Strongly Agree" "Strongly Disagree" "SA" "SD" "SD" "SD" "SD" 
## [3,] "Strongly Agree" "Neutral"           "N"  "N"  "A"  "D"  "A"  
## [4,] "Disagree"       "Strongly Agree"    "SA" "D"  "D"  "N"  "SD" 
## [5,] "Disagree"       "Strongly Agree"    "N"  "SA" "A"  "A"  "SA" 
## [6,] "Agree"          "Disagree"          "SD" "N"  "SD" "N"  "A"

head(intdat)

##   V1 V2 V3 V4 V5 V6 V7 V8 V9 V10
## 1  4  5 NA  2  4  3  1  5  3   3
## 2  3 NA  3  5  1  5  1  1  1   1
## 3  5  2 NA  5  3  3  3  4  2   4
## 4  2  1  4  2  5  5  2  2  3   1
## 5  1  4  1  2  5  3  5  4  4   5
## 6  1  1  3  4  2  1  3  1  3   4

#####
# data.frame as input with ordered factors

dat1 <- data.frame(dat1)
dat2 <- data.frame(dat2)
dat.old <- cbind(dat1, dat2)
colnames(dat.old) <- paste0('Item_', 1:10)
str(dat.old) # factors are leveled alphabetically by default

## 'data.frame':    1000 obs. of  10 variables:
##  $ Item_1 : chr  "Agree" "Neutral" "Strongly Agree" "Disagree" ...
##  $ Item_2 : chr  "Strongly Agree" NA "Disagree" "Strongly Disagree" ...
##  $ Item_3 : chr  NA "Neutral" NA "Agree" ...
##  $ Item_4 : chr  "Disagree" "Strongly Agree" "Strongly Agree" "Disagree" ...
##  $ Item_5 : chr  "Agree" "Strongly Disagree" "Neutral" "Strongly Agree" ...
##  $ Item_6 : chr  "N" "SA" "N" "SA" ...
##  $ Item_7 : chr  "SD" "SD" "N" "D" ...
##  $ Item_8 : chr  "SA" "SD" "A" "D" ...
##  $ Item_9 : chr  "N" "SD" "D" "N" ...
##  $ Item_10: chr  "N" "SD" "A" "SD" ...

# create explicit ordering in factor variables
for(i in 1:ncol(dat1))
   levels(dat1[[i]]) <- c('Strongly Disagree', 'Disagree', 'Neutral', 'Agree',
                          'Strongly Agree')

for(i in 1:ncol(dat2))
   levels(dat2[[i]]) <- c('SD', 'D', 'N', 'A', 'SA')

dat <- cbind(dat1, dat2)
colnames(dat) <- colnames(dat.old)
str(dat) # note ordering

## 'data.frame':    1000 obs. of  10 variables:
##  $ Item_1 : chr  "Agree" "Neutral" "Strongly Agree" "Disagree" ...
##   ..- attr(*, "levels")= chr [1:5] "Strongly Disagree" "Disagree" "Neutral" "Agree" ...
##  $ Item_2 : chr  "Strongly Agree" NA "Disagree" "Strongly Disagree" ...
##   ..- attr(*, "levels")= chr [1:5] "Strongly Disagree" "Disagree" "Neutral" "Agree" ...
##  $ Item_3 : chr  NA "Neutral" NA "Agree" ...
##   ..- attr(*, "levels")= chr [1:5] "Strongly Disagree" "Disagree" "Neutral" "Agree" ...
##  $ Item_4 : chr  "Disagree" "Strongly Agree" "Strongly Agree" "Disagree" ...
##   ..- attr(*, "levels")= chr [1:5] "Strongly Disagree" "Disagree" "Neutral" "Agree" ...
##  $ Item_5 : chr  "Agree" "Strongly Disagree" "Neutral" "Strongly Agree" ...
##   ..- attr(*, "levels")= chr [1:5] "Strongly Disagree" "Disagree" "Neutral" "Agree" ...
##  $ Item_6 : chr  "N" "SA" "N" "SA" ...
##   ..- attr(*, "levels")= chr [1:5] "SD" "D" "N" "A" ...
##  $ Item_7 : chr  "SD" "SD" "N" "D" ...
##   ..- attr(*, "levels")= chr [1:5] "SD" "D" "N" "A" ...
##  $ Item_8 : chr  "SA" "SD" "A" "D" ...
##   ..- attr(*, "levels")= chr [1:5] "SD" "D" "N" "A" ...
##  $ Item_9 : chr  "N" "SD" "D" "N" ...
##   ..- attr(*, "levels")= chr [1:5] "SD" "D" "N" "A" ...
##  $ Item_10: chr  "N" "SD" "A" "SD" ...
##   ..- attr(*, "levels")= chr [1:5] "SD" "D" "N" "A" ...

intdat <- likert2int(dat)
head(dat)

##              Item_1            Item_2            Item_3         Item_4
## 1             Agree    Strongly Agree              <NA>       Disagree
## 2           Neutral              <NA>           Neutral Strongly Agree
## 3    Strongly Agree          Disagree              <NA> Strongly Agree
## 4          Disagree Strongly Disagree             Agree       Disagree
## 5 Strongly Disagree             Agree Strongly Disagree       Disagree
## 6 Strongly Disagree Strongly Disagree           Neutral          Agree
##              Item_5 Item_6 Item_7 Item_8 Item_9 Item_10
## 1             Agree      N     SD     SA      N       N
## 2 Strongly Disagree     SA     SD     SD     SD      SD
## 3           Neutral      N      N      A      D       A
## 4    Strongly Agree     SA      D      D      N      SD
## 5    Strongly Agree      N     SA      A      A      SA
## 6          Disagree     SD      N     SD      N       A

head(intdat)

##   Item_1 Item_2 Item_3 Item_4 Item_5 Item_6 Item_7 Item_8 Item_9 Item_10
## 1      4      5     NA      2      4      3      1      5      3       3
## 2      3     NA      3      5      1      5      1      1      1       1
## 3      5      2     NA      5      3      3      3      4      2       4
## 4      2      1      4      2      5      5      2      2      3       1
## 5      1      4      1      2      5      3      5      4      4       5
## 6      1      1      3      4      2      1      3      1      3       4

## End(No test)

[Package mirt version 1.36.2 Index]