likert2int {mirt}	R Documentation

Convert ordered Likert-scale responses (character or factors) to integers

Description

Given a matrix or data.frame object consisting of Likert responses return an object of the same dimensions with integer values.

Usage

likert2int(x, levels = NULL)

Arguments

x
a matrix of character values or data.frame of character/factor vectors

levels
a named character vector indicating which integer values should be assigned to which elements. If omitted, the order of the elements will be determined after converting each column in x to a factor variable

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi:10.18637/jss.v048.i06

See Also

key2binary, poly2dich

Examples

## No test: 

# simulate data

dat1 <- matrix(sample(c('Disagree', 'Strongly Disagree', 'Agree',
                        'Neutral', 'Strongly Agree'), 1000*5, replace=TRUE),
               nrow=1000, ncol=5)
dat1[2,2] <- dat1[3,3] <- dat1[1,3] <- NA # NAs added for flavour
dat2 <- matrix(sample(c('D', 'SD', 'A', 'N', 'SA'), 1000*5, replace=TRUE),
               nrow=1000, ncol=5)
dat <- cbind(dat1, dat2)

# separately
intdat1 <- likert2int(dat1)
head(dat1)

##      [,1]             [,2]             [,3]       [,4]               
## [1,] "Agree"          "Neutral"        NA         "Strongly Disagree"
## [2,] "Strongly Agree" NA               "Neutral"  "Strongly Agree"   
## [3,] "Agree"          "Strongly Agree" NA         "Strongly Disagree"
## [4,] "Strongly Agree" "Agree"          "Disagree" "Strongly Agree"   
## [5,] "Strongly Agree" "Disagree"       "Neutral"  "Neutral"          
## [6,] "Neutral"        "Agree"          "Agree"    "Neutral"          
##      [,5]      
## [1,] "Agree"   
## [2,] "Disagree"
## [3,] "Agree"   
## [4,] "Agree"   
## [5,] "Agree"   
## [6,] "Agree"

head(intdat1)

##   V1 V2 V3 V4 V5
## 1 NA NA NA NA NA
## 2 NA NA NA NA NA
## 3 NA NA NA NA NA
## 4 NA NA NA NA NA
## 5 NA NA NA NA NA
## 6 NA NA NA NA NA

# more useful with explicit levels
lvl1 <- c('Strongly Disagree'=1, 'Disagree'=2, 'Neutral'=3, 'Agree'=4,
          'Strongly Agree'=5)
intdat1 <- likert2int(dat1, levels = lvl1)
head(dat1)

##      [,1]             [,2]             [,3]       [,4]               
## [1,] "Agree"          "Neutral"        NA         "Strongly Disagree"
## [2,] "Strongly Agree" NA               "Neutral"  "Strongly Agree"   
## [3,] "Agree"          "Strongly Agree" NA         "Strongly Disagree"
## [4,] "Strongly Agree" "Agree"          "Disagree" "Strongly Agree"   
## [5,] "Strongly Agree" "Disagree"       "Neutral"  "Neutral"          
## [6,] "Neutral"        "Agree"          "Agree"    "Neutral"          
##      [,5]      
## [1,] "Agree"   
## [2,] "Disagree"
## [3,] "Agree"   
## [4,] "Agree"   
## [5,] "Agree"   
## [6,] "Agree"

head(intdat1)

##   V1 V2 V3 V4 V5
## 1  4  3 NA  1  4
## 2  5 NA  3  5  2
## 3  4  5 NA  1  4
## 4  5  4  2  5  4
## 5  5  2  3  3  4
## 6  3  4  4  3  4

# second data
lvl2 <- c('SD'=1, 'D'=2, 'N'=3, 'A'=4, 'SA'=5)
intdat2 <- likert2int(dat2, levels = lvl2)
head(dat2)

##      [,1] [,2] [,3] [,4] [,5]
## [1,] "SD" "SD" "A"  "A"  "N" 
## [2,] "SD" "D"  "N"  "SA" "A" 
## [3,] "D"  "A"  "N"  "SD" "N" 
## [4,] "D"  "SD" "A"  "SA" "SA"
## [5,] "N"  "D"  "SA" "SA" "D" 
## [6,] "A"  "D"  "SD" "SA" "SA"

head(intdat2)

##   V1 V2 V3 V4 V5
## 1  1  1  4  4  3
## 2  1  2  3  5  4
## 3  2  4  3  1  3
## 4  2  1  4  5  5
## 5  3  2  5  5  2
## 6  4  2  1  5  5

# full dataset (using both mapping schemes)
intdat <- likert2int(dat, levels = c(lvl1, lvl2))
head(dat)

##      [,1]             [,2]             [,3]       [,4]               
## [1,] "Agree"          "Neutral"        NA         "Strongly Disagree"
## [2,] "Strongly Agree" NA               "Neutral"  "Strongly Agree"   
## [3,] "Agree"          "Strongly Agree" NA         "Strongly Disagree"
## [4,] "Strongly Agree" "Agree"          "Disagree" "Strongly Agree"   
## [5,] "Strongly Agree" "Disagree"       "Neutral"  "Neutral"          
## [6,] "Neutral"        "Agree"          "Agree"    "Neutral"          
##      [,5]       [,6] [,7] [,8] [,9] [,10]
## [1,] "Agree"    "SD" "SD" "A"  "A"  "N"  
## [2,] "Disagree" "SD" "D"  "N"  "SA" "A"  
## [3,] "Agree"    "D"  "A"  "N"  "SD" "N"  
## [4,] "Agree"    "D"  "SD" "A"  "SA" "SA" 
## [5,] "Agree"    "N"  "D"  "SA" "SA" "D"  
## [6,] "Agree"    "A"  "D"  "SD" "SA" "SA"

head(intdat)

##   V1 V2 V3 V4 V5 V6 V7 V8 V9 V10
## 1  4  3 NA  1  4  1  1  4  4   3
## 2  5 NA  3  5  2  1  2  3  5   4
## 3  4  5 NA  1  4  2  4  3  1   3
## 4  5  4  2  5  4  2  1  4  5   5
## 5  5  2  3  3  4  3  2  5  5   2
## 6  3  4  4  3  4  4  2  1  5   5

#####
# data.frame as input with ordered factors

dat1 <- data.frame(dat1)
dat2 <- data.frame(dat2)
dat.old <- cbind(dat1, dat2)
colnames(dat.old) <- paste0('Item_', 1:10)
str(dat.old) # factors are leveled alphabetically by default

## 'data.frame':	1000 obs. of  10 variables:
##  $ Item_1 : chr  "Agree" "Strongly Agree" "Agree" "Strongly Agree" ...
##  $ Item_2 : chr  "Neutral" NA "Strongly Agree" "Agree" ...
##  $ Item_3 : chr  NA "Neutral" NA "Disagree" ...
##  $ Item_4 : chr  "Strongly Disagree" "Strongly Agree" "Strongly Disagree" "Strongly Agree" ...
##  $ Item_5 : chr  "Agree" "Disagree" "Agree" "Agree" ...
##  $ Item_6 : chr  "SD" "SD" "D" "D" ...
##  $ Item_7 : chr  "SD" "D" "A" "SD" ...
##  $ Item_8 : chr  "A" "N" "N" "A" ...
##  $ Item_9 : chr  "A" "SA" "SD" "SA" ...
##  $ Item_10: chr  "N" "A" "N" "SA" ...

# create explicit ordering in factor variables
for(i in 1:ncol(dat1))
   levels(dat1[[i]]) <- c('Strongly Disagree', 'Disagree', 'Neutral', 'Agree',
                          'Strongly Agree')

for(i in 1:ncol(dat2))
   levels(dat2[[i]]) <- c('SD', 'D', 'N', 'A', 'SA')

dat <- cbind(dat1, dat2)
colnames(dat) <- colnames(dat.old)
str(dat) # note ordering

## 'data.frame':	1000 obs. of  10 variables:
##  $ Item_1 : chr  "Agree" "Strongly Agree" "Agree" "Strongly Agree" ...
##   ..- attr(*, "levels")= chr [1:5] "Strongly Disagree" "Disagree" "Neutral" "Agree" ...
##  $ Item_2 : chr  "Neutral" NA "Strongly Agree" "Agree" ...
##   ..- attr(*, "levels")= chr [1:5] "Strongly Disagree" "Disagree" "Neutral" "Agree" ...
##  $ Item_3 : chr  NA "Neutral" NA "Disagree" ...
##   ..- attr(*, "levels")= chr [1:5] "Strongly Disagree" "Disagree" "Neutral" "Agree" ...
##  $ Item_4 : chr  "Strongly Disagree" "Strongly Agree" "Strongly Disagree" "Strongly Agree" ...
##   ..- attr(*, "levels")= chr [1:5] "Strongly Disagree" "Disagree" "Neutral" "Agree" ...
##  $ Item_5 : chr  "Agree" "Disagree" "Agree" "Agree" ...
##   ..- attr(*, "levels")= chr [1:5] "Strongly Disagree" "Disagree" "Neutral" "Agree" ...
##  $ Item_6 : chr  "SD" "SD" "D" "D" ...
##   ..- attr(*, "levels")= chr [1:5] "SD" "D" "N" "A" ...
##  $ Item_7 : chr  "SD" "D" "A" "SD" ...
##   ..- attr(*, "levels")= chr [1:5] "SD" "D" "N" "A" ...
##  $ Item_8 : chr  "A" "N" "N" "A" ...
##   ..- attr(*, "levels")= chr [1:5] "SD" "D" "N" "A" ...
##  $ Item_9 : chr  "A" "SA" "SD" "SA" ...
##   ..- attr(*, "levels")= chr [1:5] "SD" "D" "N" "A" ...
##  $ Item_10: chr  "N" "A" "N" "SA" ...
##   ..- attr(*, "levels")= chr [1:5] "SD" "D" "N" "A" ...

intdat <- likert2int(dat)
head(dat)

##           Item_1         Item_2   Item_3            Item_4   Item_5 Item_6
## 1          Agree        Neutral     <NA> Strongly Disagree    Agree     SD
## 2 Strongly Agree           <NA>  Neutral    Strongly Agree Disagree     SD
## 3          Agree Strongly Agree     <NA> Strongly Disagree    Agree      D
## 4 Strongly Agree          Agree Disagree    Strongly Agree    Agree      D
## 5 Strongly Agree       Disagree  Neutral           Neutral    Agree      N
## 6        Neutral          Agree    Agree           Neutral    Agree      A
##   Item_7 Item_8 Item_9 Item_10
## 1     SD      A      A       N
## 2      D      N     SA       A
## 3      A      N     SD       N
## 4     SD      A     SA      SA
## 5      D     SA     SA       D
## 6      D     SD     SA      SA

head(intdat)

##   Item_1 Item_2 Item_3 Item_4 Item_5 Item_6 Item_7 Item_8 Item_9 Item_10
## 1      4      3     NA      1      4      1      1      4      4       3
## 2      5     NA      3      5      2      1      2      3      5       4
## 3      4      5     NA      1      4      2      4      3      1       3
## 4      5      4      2      5      4      2      1      4      5       5
## 5      5      2      3      3      4      3      2      5      5       2
## 6      3      4      4      3      4      4      2      1      5       5

## End(No test)

[Package mirt version 1.43 Index]