marginal_rxx {mirt}	R Documentation

Function to calculate the marginal reliability

Description

Given an estimated model and a prior density function, compute the marginal reliability. This is only available for unidimensional tests.

Usage

marginal_rxx(mod, density = dnorm, ...)

Arguments

mod
an object of class 'SingleGroupClass'

density
a density function to use for integration. Default assumes the latent traits are from a normal (Gaussian) distribution

...
additional arguments passed to the density function

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi: 10.18637/jss.v048.i06

See Also

empirical_rxx, extract.group, testinfo

Examples

## No test: 

dat <- expand.table(deAyala)
mod <- mirt(dat, 1)

# marginal estimate
marginal_rxx(mod)

## [1] 0.6092894

# empirical estimate (assuming the same prior)
fscores(mod, returnER = TRUE)

##        F1 
## 0.6200703

# empirical rxx the alternative way, given theta scores and SEs
fs <- fscores(mod, full.scores.SE=TRUE)
head(fs)

##             F1     SE_F1
## [1,] -1.580133 0.6699425
## [2,] -1.580133 0.6699425
## [3,] -1.580133 0.6699425
## [4,] -1.580133 0.6699425
## [5,] -1.580133 0.6699425
## [6,] -1.580133 0.6699425

empirical_rxx(fs)

##        F1 
## 0.6200703

## End(No test)

[Package mirt version 1.25 Index]