mirtCluster {mirt}	R Documentation

Define a parallel cluster object to be used in internal functions

Description

This function defines a object that is placed in a relevant internal environment defined in mirt. Internal functions such as calcLogLik, fscores, etc, will utilize this object automatically to capitalize on parallel processing architecture. The object defined is a call from parallel::makeCluster(). Note that if you are defining other parallel objects (for simulation desings, for example) it is not recommended to define a mirtCluster.

Usage

mirtCluster(ncores, remove = FALSE)

Arguments

ncores
number of cores to be used in the returned object which is passed to parallel::makeCluster(). If no input is given the maximum number of available cores will be used

remove
logical; remove previously defined mirtCluster()?

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

Examples

## No test: 

#make 4 cores available for parallel computing
mirtCluster(4)

## mirtCluster() has already been defined

#' #stop and remove cores
mirtCluster(remove = TRUE)

#use all available cores
mirtCluster()

## End(No test)

[Package mirt version 1.9 Index]