numerical_deriv {mirt}	R Documentation

Compute numerical derivatives

Description

Compute numerical derivatives using forward/backword difference, central difference, or Richardson extropolation.

Usage

numerical_deriv(par, f, ..., delta = 1e-05, gradient = TRUE,
  type = "forward")

Arguments

par
a vector of parameters

f
the objective function being evaluated

...
additional arguments to be passed to f and the numDeriv package when the Richardson type is used

delta
the term used to perturb the f function. Default is 1e-5

gradient
logical; compute the gradient terms? If FALSE then the Hessian is computed instead

type
type of difference to compute. Can be either 'forward' for the forward difference, 'central' for the central difference, or 'Richardson' for the Richardson extropolation. Backword difference is acheived by supplying a negative delta value

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

Examples

## No test: 
f <- function(x) 3*x[1]^3 - 4*x[2]^2
par <- c(3,8)

# grad = 9 * x^2 , -8 * y
(actual <- c(9 * par[1]^2, -8 * par[2]))

## [1]  81 -64

numerical_deriv(par, f, type = 'forward')

## [1]  81.00027 -64.00004

numerical_deriv(par, f, type = 'central')

## [1]  81 -64

numerical_deriv(par, f, type = 'Richardson')

## [1]  81 -64

# hessian = h11 -> 18 * x, h22 -> -8, h12 -> h21 -> 0
(actual <- matrix(c(18 * par[1], 0, 0, -8), 2, 2))

##      [,1] [,2]
## [1,]   54    0
## [2,]    0   -8

numerical_deriv(par, f, type = 'forward', gradient = FALSE)

##          [,1]      [,2]
## [1,] 54.00011  0.000000
## [2,]  0.00000 -7.999574

numerical_deriv(par, f, type = 'central', gradient = FALSE)

##          [,1]      [,2]
## [1,] 53.99997  0.000000
## [2,]  0.00000 -8.000356

numerical_deriv(par, f, type = 'Richardson', gradient = FALSE)

##              [,1]          [,2]
## [1,] 5.400000e+01  5.918015e-12
## [2,] 5.918015e-12 -8.000000e+00

## End(No test)

[Package mirt version 1.21 Index]