personfit {mirt}	R Documentation

Person fit statistics

Description

personfit calculates the Zh values from Drasgow, Levine and Williams (1985) for unidimensional and multidimensional models. For Rasch models infit and outfit statistics are also produced. The returned object is a data.frame consisting either of the tabulated data or full data with the statistics appended to the rightmost columns.

Usage

personfit(x, method = "EAP", Theta = NULL, stats.only = TRUE, ...)

Arguments

x
a computed model object of class SingleGroupClass or MultipleGroupClass

method
type of factor score estimation method. See fscores for more detail

Theta
a matrix of factor scores used for statistics that require empirical estimates. If supplied, arguments typically passed to fscores() will be ignored and these values will be used instead

stats.only
logical; return only the person fit statistics without their associated response pattern?

...
additional arguments to be passed to fscores()

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi: 10.18637/jss.v048.i06

Drasgow, F., Levine, M. V., & Williams, E. A. (1985). Appropriateness measurement with polychotomous item response models and standardized indices. British Journal of Mathematical and Statistical Psychology, 38, 67-86.

Reise, S. P. (1990). A comparison of item- and person-fit methods of assessing model-data fit in IRT. Applied Psychological Measurement, 14, 127-137.

Wright B. D. & Masters, G. N. (1982). Rating scale analysis. MESA Press.

See Also

itemfit

Examples

## No test: 

#make some data
set.seed(1)
a <- matrix(rlnorm(20),ncol=1)
d <- matrix(rnorm(20),ncol=1)
items <- rep('2PL', 20)
data <- simdata(a,d, 2000, items)

x <- mirt(data, 1)
fit <- personfit(x)
head(fit)

##           Zh
## 1  0.9931930
## 2  0.1018533
## 3 -0.5419248
## 4 -0.2966258
## 5  0.4062379
## 6  1.4271776

#using precomputed Theta
Theta <- fscores(x, method = 'MAP', full.scores = TRUE)
head(personfit(x, Theta=Theta))

##           Zh
## 1  0.9942162
## 2  0.1051419
## 3 -0.4409453
## 4 -0.2828724
## 5  0.6001267
## 6  1.4404214

#multiple group Rasch model example
set.seed(12345)
a <- matrix(rep(1, 15), ncol=1)
d <- matrix(rnorm(15,0,.7),ncol=1)
itemtype <- rep('dich', nrow(a))
N <- 1000
dataset1 <- simdata(a, d, N, itemtype)
dataset2 <- simdata(a, d, N, itemtype, sigma = matrix(1.5))
dat <- rbind(dataset1, dataset2)
group <- c(rep('D1', N), rep('D2', N))
models <- 'F1 = 1-15'
mod_Rasch <- multipleGroup(dat, models, itemtype = 'Rasch', group = group)
coef(mod_Rasch, simplify=TRUE)

## $D1
## $items
##         a1      d g u
## Item_1   1  0.392 0 1
## Item_2   1  0.421 0 1
## Item_3   1 -0.035 0 1
## Item_4   1 -0.220 0 1
## Item_5   1  0.392 0 1
## Item_6   1 -1.198 0 1
## Item_7   1  0.472 0 1
## Item_8   1 -0.112 0 1
## Item_9   1 -0.161 0 1
## Item_10  1 -0.651 0 1
## Item_11  1 -0.044 0 1
## Item_12  1  1.333 0 1
## Item_13  1  0.229 0 1
## Item_14  1  0.392 0 1
## Item_15  1 -0.508 0 1
## 
## $means
## F1 
##  0 
## 
## $cov
##       F1
## F1 1.029
## 
## 
## $D2
## $items
##         a1      d g u
## Item_1   1  0.471 0 1
## Item_2   1  0.413 0 1
## Item_3   1 -0.123 0 1
## Item_4   1 -0.399 0 1
## Item_5   1  0.413 0 1
## Item_6   1 -1.067 0 1
## Item_7   1  0.413 0 1
## Item_8   1 -0.247 0 1
## Item_9   1 -0.174 0 1
## Item_10  1 -0.640 0 1
## Item_11  1 -0.009 0 1
## Item_12  1  1.123 0 1
## Item_13  1  0.261 0 1
## Item_14  1  0.360 0 1
## Item_15  1 -0.537 0 1
## 
## $means
## F1 
##  0 
## 
## $cov
##       F1
## F1 1.464

pf <- personfit(mod_Rasch, method='MAP')
head(pf)

##      outfit    z.outfit     infit     z.infit         Zh
## 1 1.1366820  0.88135170 1.0779259  0.62077020 -0.6799647
## 2 0.9338326 -0.07995345 0.9711313 -0.01293448  0.1247230
## 3 0.7476346 -0.68481894 0.8478365 -0.48758078  0.6198994
## 4 1.0974680  0.42845303 1.1108062  0.54327414 -0.4527551
## 5 0.9951124  0.05173539 1.0384492  0.27241674 -0.1407463
## 6 0.6611781 -1.27330009 0.7182209 -1.29180661  1.2367474

## End(No test)

[Package mirt version 1.27 Index]