poly2dich {mirt}	R Documentation

Change polytomous items to dichotomous item format

Description

Tranforms a matrix of items into a new matrix where the select polytomous items have been converted into comperable dichotomous items with the same information.

Usage

poly2dich(data, which.items = 1:ncol(data))

Arguments

data
an object of class data.frame or matrix

which.items
a vector indicating which items should be transformed into the dichotomous form. Default uses all input items

Value

Returns an integer matrix

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

Examples

## No test: 
data(Science)

head(Science)

##   Comfort Work Future Benefit
## 1       4    4      3       2
## 2       3    3      3       3
## 3       3    2      2       3
## 4       3    2      2       3
## 5       3    4      4       1
## 6       4    4      3       3

newScience <- poly2dich(Science)
head(newScience)

##      Comfort.1 Comfort.2 Comfort.3 Comfort.4 Work.1 Work.2 Work.3 Work.4
## [1,]         0         0         0         1      0      0      0      1
## [2,]         0         0         1         0      0      0      1      0
## [3,]         0         0         1         0      0      1      0      0
## [4,]         0         0         1         0      0      1      0      0
## [5,]         0         0         1         0      0      0      0      1
## [6,]         0         0         0         1      0      0      0      1
##      Future.1 Future.2 Future.3 Future.4 Benefit.1 Benefit.2 Benefit.3
## [1,]        0        0        1        0         0         1         0
## [2,]        0        0        1        0         0         0         1
## [3,]        0        1        0        0         0         0         1
## [4,]        0        1        0        0         0         0         1
## [5,]        0        0        0        1         1         0         0
## [6,]        0        0        1        0         0         0         1
##      Benefit.4
## [1,]         0
## [2,]         0
## [3,]         0
## [4,]         0
## [5,]         0
## [6,]         0

newScience2 <- poly2dich(Science, which.items = 2)
head(newScience2)

##   Comfort Work.1 Work.2 Work.3 Work.4 Future Benefit
## 1       4      0      0      0      1      3       2
## 2       3      0      0      1      0      3       3
## 3       3      0      1      0      0      2       3
## 4       3      0      1      0      0      2       3
## 5       3      0      0      0      1      4       1
## 6       4      0      0      0      1      3       3

## End(No test)

[Package mirt version 1.22 Index]