probtrace {mirt}	R Documentation

Function to calculate probability trace lines

Description

Given an internal mirt object extracted from an estimated model compute the probability trace lines for all categories.

Usage

probtrace(x, Theta)

Arguments

x
an extracted internal mirt object containing item information

Theta
a vector (unidimensional) or matrix (unidimensional/multidimensional) of latent trait values

See Also

extract.item

Examples

## No test: 
mod <- mirt(Science, 1)
extr.2 <- extract.item(mod, 2)
Theta <- matrix(seq(-4,4, by = .1))
traceline <- probtrace(extr.2, Theta)

head(data.frame(traceline, Theta=Theta))

##         P.1       P.2        P.3          P.4 Theta
## 1 0.8786203 0.1034360 0.01717506 0.0007686851  -4.0
## 2 0.8649284 0.1148350 0.01936773 0.0008688481  -3.9
## 3 0.8499558 0.1272286 0.02183352 0.0009820499  -3.8
## 4 0.8336428 0.1406426 0.02460462 0.0011099843  -3.7
## 5 0.8159402 0.1550887 0.02771652 0.0012545642  -3.6
## 6 0.7968130 0.1705608 0.03120820 0.0014179495  -3.5

## End(No test)

[Package mirt version 1.9 Index]