probtrace {mirt}	R Documentation

Function to calculate probability trace lines

Description

Given an internal mirt object extracted from an estimated model compute the probability trace lines for all categories.

Usage

probtrace(x, Theta)

Arguments

x
an extracted internal mirt object containing item information (see extract.item)

Theta
a vector (unidimensional) or matrix (unidimensional/multidimensional) of latent trait values

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

See Also

extract.item

Examples

## No test: 
mod <- mirt(Science, 1)
extr.2 <- extract.item(mod, 2)
Theta <- matrix(seq(-4,4, by = .1))
traceline <- probtrace(extr.2, Theta)

head(data.frame(traceline, Theta=Theta))

##         P.1       P.2        P.3          P.4 Theta
## 1 0.8786646 0.1033965 0.01717048 0.0007684150  -4.0
## 2 0.8649759 0.1147928 0.01936274 0.0008685506  -3.9
## 3 0.8500065 0.1271837 0.02182811 0.0009817226  -3.8
## 4 0.8336966 0.1405950 0.02459876 0.0011096245  -3.7
## 5 0.8159972 0.1550385 0.02771018 0.0012541689  -3.6
## 6 0.7968730 0.1705082 0.03120137 0.0014175155  -3.5

## End(No test)

[Package mirt version 1.21 Index]