randef {mirt}	R Documentation

Compute posterior estimates of random effect

Description

Stochastically compute random effects for MixedClass objects with Metropolis-Hastings samplers and averaging over the draws. Returns a list of the estimated effects.

Usage

randef(x, ndraws = 1000, thin = 10, return.draws = FALSE)

Arguments

x
an estimated model object from the mixedmirt function

ndraws
total number of draws to perform. Default is 1000

thin
amount of thinning to apply. Default is to use every 10th draw

return.draws
logical; return a list containing the thinned draws of the posterior?

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29.

Chalmers, R. P. (2015). Extended Mixed-Effects Item Response Models with the MH-RM Algorithm. Journal of Educational Measurement, 52, 200-222. doi: 10.1111/jedm.12072 doi: 10.18637/jss.v048.i06

Examples

## No test: 
#make an arbitrary groups
covdat <- data.frame(group = rep(paste0('group', 1:49), each=nrow(Science)/49))

#partial credit model
mod <- mixedmirt(Science, covdat, model=1, random = ~ 1|group)
summary(mod)

## 
## Call:
## mixedmirt(data = Science, covdata = covdat, model = 1, random = ~1 | 
##     group)
## 
## 
## --------------
## RANDOM EFFECT COVARIANCE(S):
## Correlations on upper diagonal
## 
## $Theta
##      F1
## F1 1.01
## 
## $group
##           COV_group
## COV_group    0.0398

effects <- randef(mod, ndraws = 2000, thin = 20)
head(effects$Theta)

##               F1
## [1,]  0.48761231
## [2,] -0.03643771
## [3,] -0.65546901
## [4,] -0.71429552
## [5,]  0.03588258
## [6,]  1.03734597

head(effects$group)

##               group
## group1 -0.005567870
## group2 -0.003206529
## group3  0.028617636
## group4 -0.017325862
## group5 -0.018154265
## group6  0.027484231

## End(No test)

[Package mirt version 1.25 Index]