Compute posterior estimates of random effect

Stochastically compute random effects for MixedClass objects with Metropolis-Hastings samplers and averaging over the draws to obtain expected a posteriori predictions. Returns a list of the estimated effects.

Usage

randef(x, ndraws = 1000, thin = 10, return.draws = FALSE)

Arguments

x

an estimated model object from the mixedmirt function

ndraws

total number of draws to perform. Default is 1000

thin

amount of thinning to apply. Default is to use every 10th draw

return.draws

logical; return a list containing the thinned draws of the posterior?

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29.

Chalmers, R. P. (2015). Extended Mixed-Effects Item Response Models with the MH-RM Algorithm. Journal of Educational Measurement, 52, 200-222. doi:10.1111/jedm.12072 doi:10.18637/jss.v048.i06

Author

Phil Chalmers rphilip.chalmers@gmail.com

Examples

# \donttest{
# make an arbitrary groups
covdat <- data.frame(group = rep(paste0('group', 1:49), each=nrow(Science)/49))

# partial credit model
mod <- mixedmirt(Science, covdat, model=1, random = ~ 1|group)
#> Stage 1 = 1, CDLL = -2033.5, AR(1.64) = [0.43], Max-Change = 0.2000Stage 1 = 2, CDLL = -2049.6, AR(1.64) = [0.56], Max-Change = 0.2000Stage 1 = 3, CDLL = -2028.4, AR(1.64) = [0.50], Max-Change = 0.2000Stage 1 = 4, CDLL = -2026.6, AR(1.64) = [0.51], Max-Change = 0.1915Stage 1 = 5, CDLL = -2043.7, AR(1.64) = [0.53], Max-Change = 0.1573Stage 1 = 6, CDLL = -2022.4, AR(1.64) = [0.49], Max-Change = 0.1007Stage 1 = 7, CDLL = -1980.3, AR(1.64) = [0.53], Max-Change = 0.0881Stage 1 = 8, CDLL = -2001.3, AR(1.64) = [0.52], Max-Change = 0.0682Stage 1 = 9, CDLL = -2001.9, AR(1.64) = [0.47], Max-Change = 0.0656Stage 1 = 10, CDLL = -1992.8, AR(1.64) = [0.54], Max-Change = 0.0356Stage 1 = 11, CDLL = -2012.9, AR(1.64) = [0.45], Max-Change = 0.0255Stage 1 = 12, CDLL = -2024.6, AR(1.64) = [0.52], Max-Change = 0.0751Stage 1 = 13, CDLL = -2007.2, AR(1.64) = [0.53], Max-Change = 0.0268Stage 1 = 14, CDLL = -1995.1, AR(1.64) = [0.55], Max-Change = 0.0304Stage 1 = 15, CDLL = -1997.2, AR(1.64) = [0.48], Max-Change = 0.0229Stage 1 = 16, CDLL = -1993.9, AR(1.64) = [0.52], Max-Change = 0.0200Stage 1 = 17, CDLL = -1998.8, AR(1.64) = [0.50], Max-Change = 0.0445Stage 1 = 18, CDLL = -1993.3, AR(1.64) = [0.50], Max-Change = 0.0308Stage 1 = 19, CDLL = -1993.4, AR(1.64) = [0.46], Max-Change = 0.0444Stage 1 = 20, CDLL = -1962.8, AR(1.64) = [0.48], Max-Change = 0.0135Stage 1 = 21, CDLL = -1984.4, AR(1.64) = [0.49], Max-Change = 0.0410Stage 1 = 22, CDLL = -2012.5, AR(1.64) = [0.49], Max-Change = 0.0253Stage 1 = 23, CDLL = -2008.3, AR(1.64) = [0.54], Max-Change = 0.0290Stage 1 = 24, CDLL = -2000.6, AR(1.64) = [0.51], Max-Change = 0.0157Stage 1 = 25, CDLL = -2002.8, AR(1.64) = [0.54], Max-Change = 0.0171Stage 1 = 26, CDLL = -1986.9, AR(1.64) = [0.51], Max-Change = 0.0140Stage 1 = 27, CDLL = -1981.5, AR(1.64) = [0.52], Max-Change = 0.0117Stage 1 = 28, CDLL = -1981.1, AR(1.64) = [0.50], Max-Change = 0.0350Stage 1 = 29, CDLL = -1989.0, AR(1.64) = [0.45], Max-Change = 0.0598Stage 1 = 30, CDLL = -2005.0, AR(1.64) = [0.48], Max-Change = 0.0187Stage 1 = 31, CDLL = -1967.3, AR(1.64) = [0.50], Max-Change = 0.0132Stage 1 = 32, CDLL = -1991.7, AR(1.64) = [0.49], Max-Change = 0.0261Stage 1 = 33, CDLL = -1994.8, AR(1.64) = [0.48], Max-Change = 0.0297Stage 1 = 34, CDLL = -1981.2, AR(1.64) = [0.52], Max-Change = 0.0074Stage 1 = 35, CDLL = -1973.8, AR(1.64) = [0.48], Max-Change = 0.0052Stage 1 = 36, CDLL = -2001.8, AR(1.64) = [0.49], Max-Change = 0.0281Stage 1 = 37, CDLL = -1994.1, AR(1.64) = [0.47], Max-Change = 0.0186Stage 1 = 38, CDLL = -1998.4, AR(1.64) = [0.49], Max-Change = 0.0302Stage 1 = 39, CDLL = -2004.0, AR(1.64) = [0.52], Max-Change = 0.0417Stage 1 = 40, CDLL = -1973.4, AR(1.64) = [0.52], Max-Change = 0.0215Stage 1 = 41, CDLL = -1990.5, AR(1.64) = [0.50], Max-Change = 0.0501Stage 1 = 42, CDLL = -2014.0, AR(1.64) = [0.53], Max-Change = 0.0178Stage 1 = 43, CDLL = -1994.0, AR(1.64) = [0.48], Max-Change = 0.0388Stage 1 = 44, CDLL = -1970.3, AR(1.64) = [0.48], Max-Change = 0.0248Stage 1 = 45, CDLL = -1997.2, AR(1.64) = [0.51], Max-Change = 0.0528Stage 1 = 46, CDLL = -1964.9, AR(1.64) = [0.49], Max-Change = 0.0227Stage 1 = 47, CDLL = -1977.2, AR(1.64) = [0.54], Max-Change = 0.0224Stage 1 = 48, CDLL = -1978.0, AR(1.64) = [0.49], Max-Change = 0.0226Stage 1 = 49, CDLL = -1990.3, AR(1.64) = [0.50], Max-Change = 0.0182Stage 1 = 50, CDLL = -1962.7, AR(1.64) = [0.46], Max-Change = 0.0037Stage 1 = 51, CDLL = -1991.2, AR(1.64) = [0.46], Max-Change = 0.0091Stage 1 = 52, CDLL = -2003.4, AR(1.64) = [0.48], Max-Change = 0.0254Stage 1 = 53, CDLL = -1973.9, AR(1.64) = [0.48], Max-Change = 0.0213Stage 1 = 54, CDLL = -1989.3, AR(1.64) = [0.51], Max-Change = 0.0231Stage 1 = 55, CDLL = -2009.0, AR(1.64) = [0.52], Max-Change = 0.0188Stage 1 = 56, CDLL = -1988.4, AR(1.64) = [0.49], Max-Change = 0.0261Stage 1 = 57, CDLL = -2001.1, AR(1.64) = [0.48], Max-Change = 0.0165Stage 1 = 58, CDLL = -1994.4, AR(1.64) = [0.50], Max-Change = 0.0136Stage 1 = 59, CDLL = -1981.7, AR(1.64) = [0.48], Max-Change = 0.0445Stage 1 = 60, CDLL = -1979.0, AR(1.64) = [0.49], Max-Change = 0.0255Stage 1 = 61, CDLL = -1995.1, AR(1.64) = [0.49], Max-Change = 0.0269Stage 1 = 62, CDLL = -1988.2, AR(1.64) = [0.44], Max-Change = 0.0198Stage 1 = 63, CDLL = -1981.6, AR(1.64) = [0.48], Max-Change = 0.0410Stage 1 = 64, CDLL = -2001.1, AR(1.64) = [0.56], Max-Change = 0.0160Stage 1 = 65, CDLL = -1997.7, AR(1.64) = [0.46], Max-Change = 0.0145Stage 1 = 66, CDLL = -1978.7, AR(1.64) = [0.49], Max-Change = 0.0096Stage 1 = 67, CDLL = -1979.2, AR(1.64) = [0.47], Max-Change = 0.0160Stage 1 = 68, CDLL = -2002.8, AR(1.64) = [0.51], Max-Change = 0.0495Stage 1 = 69, CDLL = -1982.5, AR(1.64) = [0.51], Max-Change = 0.0382Stage 1 = 70, CDLL = -1969.5, AR(1.64) = [0.52], Max-Change = 0.0194Stage 1 = 71, CDLL = -1995.1, AR(1.64) = [0.51], Max-Change = 0.0273Stage 1 = 72, CDLL = -1998.4, AR(1.64) = [0.46], Max-Change = 0.0555Stage 1 = 73, CDLL = -1981.3, AR(1.64) = [0.48], Max-Change = 0.0118Stage 1 = 74, CDLL = -2013.7, AR(1.64) = [0.49], Max-Change = 0.0238Stage 1 = 75, CDLL = -1976.3, AR(1.64) = [0.48], Max-Change = 0.0052Stage 1 = 76, CDLL = -1988.1, AR(1.64) = [0.49], Max-Change = 0.0371Stage 1 = 77, CDLL = -1996.7, AR(1.64) = [0.50], Max-Change = 0.0250Stage 1 = 78, CDLL = -1996.1, AR(1.64) = [0.48], Max-Change = 0.0414Stage 1 = 79, CDLL = -1978.6, AR(1.64) = [0.51], Max-Change = 0.0157Stage 1 = 80, CDLL = -1972.1, AR(1.64) = [0.51], Max-Change = 0.0294Stage 1 = 81, CDLL = -2001.6, AR(1.64) = [0.44], Max-Change = 0.0270Stage 1 = 82, CDLL = -1997.7, AR(1.64) = [0.48], Max-Change = 0.0228Stage 1 = 83, CDLL = -1986.8, AR(1.64) = [0.47], Max-Change = 0.0059Stage 1 = 84, CDLL = -1971.8, AR(1.64) = [0.52], Max-Change = 0.0323Stage 1 = 85, CDLL = -1979.4, AR(1.64) = [0.51], Max-Change = 0.0277Stage 1 = 86, CDLL = -1992.7, AR(1.64) = [0.51], Max-Change = 0.0092Stage 1 = 87, CDLL = -1962.2, AR(1.64) = [0.46], Max-Change = 0.0094Stage 1 = 88, CDLL = -1979.9, AR(1.64) = [0.48], Max-Change = 0.0214Stage 1 = 89, CDLL = -1974.6, AR(1.64) = [0.51], Max-Change = 0.0300Stage 1 = 90, CDLL = -1981.3, AR(1.64) = [0.52], Max-Change = 0.0302Stage 1 = 91, CDLL = -1973.2, AR(1.64) = [0.48], Max-Change = 0.0401Stage 1 = 92, CDLL = -1989.6, AR(1.64) = [0.51], Max-Change = 0.0091Stage 1 = 93, CDLL = -1999.2, AR(1.64) = [0.49], Max-Change = 0.0820Stage 1 = 94, CDLL = -1977.5, AR(1.64) = [0.47], Max-Change = 0.0138Stage 1 = 95, CDLL = -1971.5, AR(1.64) = [0.52], Max-Change = 0.0163Stage 1 = 96, CDLL = -1997.7, AR(1.64) = [0.49], Max-Change = 0.0619Stage 1 = 97, CDLL = -1966.0, AR(1.64) = [0.46], Max-Change = 0.0539Stage 1 = 98, CDLL = -1973.6, AR(1.64) = [0.43], Max-Change = 0.0059Stage 1 = 99, CDLL = -1990.4, AR(1.64) = [0.47], Max-Change = 0.0250Stage 1 = 100, CDLL = -1989.7, AR(1.53) = [0.50], Max-Change = 0.0126Stage 1 = 101, CDLL = -1981.4, AR(1.53; 4.08) = [0.49; 0.49], Max-Change = 0.0472Stage 1 = 102, CDLL = -1966.5, AR(1.53; 4.08) = [0.47; 0.43], Max-Change = 0.0278Stage 1 = 103, CDLL = -1982.0, AR(1.53; 4.08) = [0.54; 0.41], Max-Change = 0.0999Stage 1 = 104, CDLL = -1973.0, AR(1.53; 4.08) = [0.49; 0.31], Max-Change = 0.0248Stage 1 = 105, CDLL = -1997.9, AR(1.53; 4.08) = [0.46; 0.35], Max-Change = 0.0933Stage 1 = 106, CDLL = -1967.3, AR(1.53; 4.08) = [0.45; 0.47], Max-Change = 0.2000Stage 1 = 107, CDLL = -1969.5, AR(1.53; 4.08) = [0.48; 0.27], Max-Change = 0.0784Stage 1 = 108, CDLL = -1971.4, AR(1.53; 4.08) = [0.47; 0.20], Max-Change = 0.0183Stage 1 = 109, CDLL = -2009.8, AR(1.53; 4.08) = [0.49; 0.29], Max-Change = 0.0312Stage 1 = 110, CDLL = -2008.5, AR(1.53; 4.08) = [0.48; 0.29], Max-Change = 0.0170Stage 1 = 111, CDLL = -1981.4, AR(1.53; 4.08) = [0.51; 0.24], Max-Change = 0.0667Stage 1 = 112, CDLL = -1982.2, AR(1.53; 4.08) = [0.52; 0.24], Max-Change = 0.0366Stage 1 = 113, CDLL = -1984.9, AR(1.53; 4.08) = [0.54; 0.33], Max-Change = 0.1640Stage 1 = 114, CDLL = -1972.0, AR(1.53; 4.08) = [0.50; 0.24], Max-Change = 0.0398Stage 1 = 115, CDLL = -1989.5, AR(1.53; 4.08) = [0.51; 0.22], Max-Change = 0.0143Stage 1 = 116, CDLL = -1992.0, AR(1.53; 4.08) = [0.50; 0.22], Max-Change = 0.0223Stage 1 = 117, CDLL = -2001.4, AR(1.53; 4.08) = [0.49; 0.24], Max-Change = 0.0374Stage 1 = 118, CDLL = -1990.8, AR(1.53; 4.08) = [0.46; 0.20], Max-Change = 0.0143Stage 1 = 119, CDLL = -2011.9, AR(1.53; 4.08) = [0.49; 0.14], Max-Change = 0.0153Stage 1 = 120, CDLL = -1980.4, AR(1.53; 4.08) = [0.51; 0.18], Max-Change = 0.0348Stage 1 = 121, CDLL = -1992.9, AR(1.53; 4.08) = [0.51; 0.16], Max-Change = 0.0118Stage 1 = 122, CDLL = -1988.2, AR(1.53; 4.08) = [0.44; 0.27], Max-Change = 0.0261Stage 1 = 123, CDLL = -2013.9, AR(1.53; 4.08) = [0.51; 0.10], Max-Change = 0.1477Stage 1 = 124, CDLL = -1994.9, AR(1.53; 4.08) = [0.53; 0.20], Max-Change = 0.0443Stage 1 = 125, CDLL = -2018.6, AR(1.53; 4.08) = [0.49; 0.18], Max-Change = 0.0091Stage 1 = 126, CDLL = -2000.9, AR(1.53; 4.08) = [0.47; 0.10], Max-Change = 0.0144Stage 1 = 127, CDLL = -1990.4, AR(1.53; 4.08) = [0.53; 0.22], Max-Change = 0.0168Stage 1 = 128, CDLL = -1974.1, AR(1.53; 4.08) = [0.53; 0.24], Max-Change = 0.0444Stage 1 = 129, CDLL = -2016.0, AR(1.53; 4.08) = [0.49; 0.14], Max-Change = 0.0190Stage 1 = 130, CDLL = -1999.5, AR(1.53; 4.08) = [0.49; 0.06], Max-Change = 0.0157Stage 1 = 131, CDLL = -2011.7, AR(1.53; 4.08) = [0.48; 0.12], Max-Change = 0.0231Stage 1 = 132, CDLL = -2012.0, AR(1.53; 4.08) = [0.55; 0.08], Max-Change = 0.0213Stage 1 = 133, CDLL = -1990.6, AR(1.53; 4.08) = [0.52; 0.14], Max-Change = 0.0231Stage 1 = 134, CDLL = -2002.5, AR(1.53; 4.08) = [0.47; 0.12], Max-Change = 0.0278Stage 1 = 135, CDLL = -1990.7, AR(1.53; 4.08) = [0.53; 0.16], Max-Change = 0.0082Stage 1 = 136, CDLL = -2001.1, AR(1.53; 4.08) = [0.49; 0.20], Max-Change = 0.0440Stage 1 = 137, CDLL = -2033.8, AR(1.53; 4.08) = [0.53; 0.12], Max-Change = 0.0320Stage 1 = 138, CDLL = -2006.4, AR(1.53; 4.08) = [0.50; 0.06], Max-Change = 0.0238Stage 1 = 139, CDLL = -1991.4, AR(1.53; 4.08) = [0.49; 0.22], Max-Change = 0.0408Stage 1 = 140, CDLL = -2010.7, AR(1.53; 4.08) = [0.51; 0.10], Max-Change = 0.0082Stage 1 = 141, CDLL = -1974.1, AR(1.53; 4.08) = [0.50; 0.10], Max-Change = 0.0251Stage 1 = 142, CDLL = -1986.8, AR(1.53; 4.08) = [0.49; 0.12], Max-Change = 0.0223Stage 1 = 143, CDLL = -2010.3, AR(1.53; 4.08) = [0.50; 0.18], Max-Change = 0.2000Stage 1 = 144, CDLL = -1989.8, AR(1.53; 4.08) = [0.49; 0.02], Max-Change = 0.0111Stage 1 = 145, CDLL = -1981.1, AR(1.53; 4.08) = [0.53; 0.10], Max-Change = 0.0245Stage 1 = 146, CDLL = -2003.6, AR(1.53; 4.08) = [0.51; 0.04], Max-Change = 0.0052Stage 1 = 147, CDLL = -1989.0, AR(1.53; 4.08) = [0.49; 0.14], Max-Change = 0.0256Stage 1 = 148, CDLL = -1971.4, AR(1.53; 4.08) = [0.50; 0.18], Max-Change = 0.0124Stage 1 = 149, CDLL = -2020.7, AR(1.53; 4.08) = [0.49; 0.06], Max-Change = 0.0316Stage 1 = 150, CDLL = -1995.1, AR(1.53; 4.08) = [0.48; 0.12], Max-Change = 0.0363Stage 1 = 151, CDLL = -1996.9, AR(1.53; 4.08) = [0.48; 0.16], Max-Change = 0.0302Stage 1 = 152, CDLL = -1996.6, AR(1.53; 4.08) = [0.49; 0.08], Max-Change = 0.0088Stage 1 = 153, CDLL = -1999.6, AR(1.53; 4.08) = [0.47; 0.04], Max-Change = 0.0356Stage 1 = 154, CDLL = -1976.6, AR(1.53; 4.08) = [0.55; 0.06], Max-Change = 0.0298Stage 1 = 155, CDLL = -2001.8, AR(1.53; 4.08) = [0.49; 0.08], Max-Change = 0.0329Stage 1 = 156, CDLL = -2005.3, AR(1.53; 4.08) = [0.51; 0.08], Max-Change = 0.0517Stage 1 = 157, CDLL = -2005.6, AR(1.53; 4.08) = [0.52; 0.06], Max-Change = 0.0219Stage 1 = 158, CDLL = -2000.8, AR(1.53; 4.08) = [0.54; 0.14], Max-Change = 0.0085Stage 1 = 159, CDLL = -1990.1, AR(1.53; 4.08) = [0.51; 0.10], Max-Change = 0.0108Stage 1 = 160, CDLL = -1985.4, AR(1.53; 4.08) = [0.53; 0.06], Max-Change = 0.0283Stage 1 = 161, CDLL = -1991.1, AR(1.53; 4.08) = [0.48; 0.04], Max-Change = 0.0234Stage 1 = 162, CDLL = -1999.2, AR(1.53; 4.08) = [0.51; 0.18], Max-Change = 0.0235Stage 1 = 163, CDLL = -1992.0, AR(1.53; 4.08) = [0.55; 0.04], Max-Change = 0.0381Stage 1 = 164, CDLL = -1996.4, AR(1.53; 4.08) = [0.49; 0.10], Max-Change = 0.0311Stage 1 = 165, CDLL = -1971.4, AR(1.53; 4.08) = [0.50; 0.14], Max-Change = 0.0693Stage 1 = 166, CDLL = -1980.4, AR(1.53; 4.08) = [0.50; 0.04], Max-Change = 0.0397Stage 1 = 167, CDLL = -1974.1, AR(1.53; 4.08) = [0.53; 0.06], Max-Change = 0.0596Stage 1 = 168, CDLL = -1954.9, AR(1.53; 4.08) = [0.51; 0.08], Max-Change = 0.0471Stage 1 = 169, CDLL = -1987.3, AR(1.53; 4.08) = [0.47; 0.02], Max-Change = 0.0343Stage 1 = 170, CDLL = -1970.4, AR(1.53; 4.08) = [0.53; 0.00], Max-Change = 0.0371Stage 1 = 171, CDLL = -1984.1, AR(1.53; 4.08) = [0.52; 0.06], Max-Change = 0.0371Stage 1 = 172, CDLL = -1995.1, AR(1.53; 4.08) = [0.51; 0.04], Max-Change = 0.0160Stage 1 = 173, CDLL = -1996.4, AR(1.53; 4.08) = [0.47; 0.04], Max-Change = 0.0327Stage 1 = 174, CDLL = -1989.0, AR(1.53; 4.08) = [0.53; 0.12], Max-Change = 0.0152Stage 1 = 175, CDLL = -1981.9, AR(1.53; 4.08) = [0.48; 0.10], Max-Change = 0.0242Stage 1 = 176, CDLL = -1991.0, AR(1.53; 4.08) = [0.52; 0.04], Max-Change = 0.0370Stage 1 = 177, CDLL = -1988.0, AR(1.53; 4.08) = [0.50; 0.08], Max-Change = 0.0250Stage 1 = 178, CDLL = -1982.1, AR(1.53; 4.08) = [0.52; 0.06], Max-Change = 0.0365Stage 1 = 179, CDLL = -1985.2, AR(1.53; 4.08) = [0.47; 0.02], Max-Change = 0.0145Stage 1 = 180, CDLL = -1976.5, AR(1.53; 4.08) = [0.51; 0.06], Max-Change = 0.0242Stage 1 = 181, CDLL = -2011.0, AR(1.53; 4.08) = [0.48; 0.06], Max-Change = 0.0182Stage 1 = 182, CDLL = -1986.6, AR(1.53; 4.08) = [0.52; 0.06], Max-Change = 0.0203Stage 1 = 183, CDLL = -1998.2, AR(1.53; 4.08) = [0.48; 0.12], Max-Change = 0.0343Stage 1 = 184, CDLL = -1978.5, AR(1.53; 4.08) = [0.52; 0.06], Max-Change = 0.0251Stage 1 = 185, CDLL = -1988.9, AR(1.53; 4.08) = [0.54; 0.06], Max-Change = 0.0231Stage 1 = 186, CDLL = -1989.9, AR(1.53; 4.08) = [0.47; 0.02], Max-Change = 0.0061Stage 1 = 187, CDLL = -2004.9, AR(1.53; 4.08) = [0.52; 0.04], Max-Change = 0.0145Stage 1 = 188, CDLL = -2000.9, AR(1.53; 4.08) = [0.50; 0.06], Max-Change = 0.0328Stage 1 = 189, CDLL = -1988.7, AR(1.53; 4.08) = [0.48; 0.10], Max-Change = 0.0212Stage 1 = 190, CDLL = -1984.3, AR(1.53; 4.08) = [0.53; 0.04], Max-Change = 0.0181Stage 1 = 191, CDLL = -2005.1, AR(1.53; 4.08) = [0.47; 0.00], Max-Change = 0.0260Stage 1 = 192, CDLL = -1972.9, AR(1.53; 4.08) = [0.52; 0.12], Max-Change = 0.0101Stage 1 = 193, CDLL = -1997.1, AR(1.53; 4.08) = [0.49; 0.04], Max-Change = 0.0139Stage 1 = 194, CDLL = -1997.9, AR(1.53; 4.08) = [0.48; 0.12], Max-Change = 0.0107Stage 1 = 195, CDLL = -1978.6, AR(1.53; 4.08) = [0.54; 0.10], Max-Change = 0.0053Stage 1 = 196, CDLL = -1971.3, AR(1.53; 4.08) = [0.48; 0.10], Max-Change = 0.0552Stage 1 = 197, CDLL = -2002.7, AR(1.53; 4.08) = [0.48; 0.04], Max-Change = 0.0186Stage 1 = 198, CDLL = -1973.0, AR(1.53; 4.08) = [0.52; 0.04], Max-Change = 0.0305Stage 1 = 199, CDLL = -1997.9, AR(1.53; 4.08) = [0.48; 0.02], Max-Change = 0.0174Stage 1 = 200, CDLL = -1964.2, AR(2.82; 0.00) = [0.39; 0.80], Max-Change = 0.0232Stage 2 = 1, CDLL = -1998.7, AR(2.82; 0.00) = [0.39; 0.76], Max-Change = 0.0176Stage 2 = 2, CDLL = -1987.4, AR(2.82; 0.00) = [0.43; 0.63], Max-Change = 0.0055Stage 2 = 3, CDLL = -1966.7, AR(2.82; 0.00) = [0.43; 0.65], Max-Change = 0.0257Stage 2 = 4, CDLL = -1966.2, AR(2.82; 0.00) = [0.40; 0.67], Max-Change = 0.0202Stage 2 = 5, CDLL = -1971.0, AR(2.82; 0.00) = [0.43; 0.67], Max-Change = 0.0191Stage 2 = 6, CDLL = -1972.1, AR(2.82; 0.00) = [0.41; 0.82], Max-Change = 0.0085Stage 2 = 7, CDLL = -1999.1, AR(2.82; 0.00) = [0.39; 0.63], Max-Change = 0.0236Stage 2 = 8, CDLL = -1976.5, AR(2.82; 0.00) = [0.41; 0.80], Max-Change = 0.0117Stage 2 = 9, CDLL = -1998.3, AR(2.82; 0.00) = [0.38; 0.65], Max-Change = 0.0123Stage 2 = 10, CDLL = -1970.9, AR(2.82; 0.00) = [0.38; 0.71], Max-Change = 0.0412Stage 2 = 11, CDLL = -1977.9, AR(2.82; 0.00) = [0.41; 0.82], Max-Change = 0.0324Stage 2 = 12, CDLL = -1975.0, AR(2.82; 0.00) = [0.40; 0.73], Max-Change = 0.0294Stage 2 = 13, CDLL = -2006.0, AR(2.82; 0.00) = [0.44; 0.65], Max-Change = 0.0184Stage 2 = 14, CDLL = -1980.9, AR(2.82; 0.00) = [0.44; 0.71], Max-Change = 0.0159Stage 2 = 15, CDLL = -1996.4, AR(2.82; 0.00) = [0.46; 0.71], Max-Change = 0.0174Stage 2 = 16, CDLL = -1975.3, AR(2.82; 0.00) = [0.41; 0.73], Max-Change = 0.0132Stage 2 = 17, CDLL = -1969.6, AR(2.82; 0.00) = [0.42; 0.80], Max-Change = 0.0290Stage 2 = 18, CDLL = -1988.2, AR(2.82; 0.00) = [0.39; 0.69], Max-Change = 0.0222Stage 2 = 19, CDLL = -1994.8, AR(2.82; 0.00) = [0.38; 0.82], Max-Change = 0.0331Stage 2 = 20, CDLL = -1987.2, AR(2.82; 0.00) = [0.39; 0.84], Max-Change = 0.0446Stage 2 = 21, CDLL = -2017.2, AR(2.82; 0.00) = [0.42; 0.71], Max-Change = 0.0547Stage 2 = 22, CDLL = -1981.2, AR(2.82; 0.00) = [0.41; 0.76], Max-Change = 0.0381Stage 2 = 23, CDLL = -2003.7, AR(2.82; 0.00) = [0.39; 0.71], Max-Change = 0.0320Stage 2 = 24, CDLL = -1969.6, AR(2.82; 0.00) = [0.39; 0.73], Max-Change = 0.0527Stage 2 = 25, CDLL = -1955.5, AR(2.82; 0.00) = [0.36; 0.71], Max-Change = 0.0478Stage 2 = 26, CDLL = -1974.2, AR(2.82; 0.00) = [0.41; 0.76], Max-Change = 0.0351Stage 2 = 27, CDLL = -1971.7, AR(2.82; 0.00) = [0.42; 0.73], Max-Change = 0.0380Stage 2 = 28, CDLL = -1994.5, AR(2.82; 0.00) = [0.40; 0.69], Max-Change = 0.0211Stage 2 = 29, CDLL = -1991.1, AR(2.82; 0.00) = [0.42; 0.73], Max-Change = 0.0213Stage 2 = 30, CDLL = -1958.1, AR(2.82; 0.00) = [0.42; 0.80], Max-Change = 0.0547Stage 2 = 31, CDLL = -1986.5, AR(2.82; 0.00) = [0.41; 0.71], Max-Change = 0.0155Stage 2 = 32, CDLL = -1997.2, AR(2.82; 0.00) = [0.42; 0.67], Max-Change = 0.0092Stage 2 = 33, CDLL = -2030.1, AR(2.82; 0.00) = [0.39; 0.69], Max-Change = 0.0178Stage 2 = 34, CDLL = -2022.2, AR(2.82; 0.00) = [0.38; 0.61], Max-Change = 0.0393Stage 2 = 35, CDLL = -1990.6, AR(2.82; 0.00) = [0.41; 0.65], Max-Change = 0.0074Stage 2 = 36, CDLL = -1993.2, AR(2.82; 0.00) = [0.36; 0.76], Max-Change = 0.0178Stage 2 = 37, CDLL = -1978.5, AR(2.82; 0.00) = [0.40; 0.80], Max-Change = 0.0199Stage 2 = 38, CDLL = -1980.2, AR(2.82; 0.00) = [0.38; 0.63], Max-Change = 0.0740Stage 2 = 39, CDLL = -1971.3, AR(2.82; 0.00) = [0.41; 0.67], Max-Change = 0.0120Stage 2 = 40, CDLL = -1982.5, AR(2.82; 0.00) = [0.39; 0.67], Max-Change = 0.0230Stage 2 = 41, CDLL = -1992.0, AR(2.82; 0.00) = [0.40; 0.71], Max-Change = 0.0285Stage 2 = 42, CDLL = -2009.1, AR(2.82; 0.00) = [0.46; 0.69], Max-Change = 0.0158Stage 2 = 43, CDLL = -2007.6, AR(2.82; 0.00) = [0.42; 0.67], Max-Change = 0.0081Stage 2 = 44, CDLL = -1987.9, AR(2.82; 0.00) = [0.45; 0.67], Max-Change = 0.0257Stage 2 = 45, CDLL = -1999.9, AR(2.82; 0.00) = [0.39; 0.65], Max-Change = 0.0076Stage 2 = 46, CDLL = -1995.1, AR(2.82; 0.00) = [0.42; 0.73], Max-Change = 0.0203Stage 2 = 47, CDLL = -1967.6, AR(2.82; 0.00) = [0.39; 0.76], Max-Change = 0.0119Stage 2 = 48, CDLL = -1997.1, AR(2.82; 0.00) = [0.38; 0.59], Max-Change = 0.0317Stage 2 = 49, CDLL = -1974.2, AR(2.82; 0.00) = [0.40; 0.67], Max-Change = 0.0153Stage 2 = 50, CDLL = -1973.6, AR(2.82; 0.00) = [0.43; 0.67], Max-Change = 0.0327Stage 2 = 51, CDLL = -1975.7, AR(2.82; 0.00) = [0.39; 0.69], Max-Change = 0.0362Stage 2 = 52, CDLL = -1983.8, AR(2.82; 0.00) = [0.39; 0.61], Max-Change = 0.0118Stage 2 = 53, CDLL = -1960.9, AR(2.82; 0.00) = [0.38; 0.80], Max-Change = 0.0187Stage 2 = 54, CDLL = -1949.8, AR(2.82; 0.00) = [0.39; 0.73], Max-Change = 0.0571Stage 2 = 55, CDLL = -1977.7, AR(2.82; 0.00) = [0.40; 0.80], Max-Change = 0.0288Stage 2 = 56, CDLL = -1963.4, AR(2.82; 0.00) = [0.41; 0.78], Max-Change = 0.0133Stage 2 = 57, CDLL = -1982.5, AR(2.82; 0.00) = [0.37; 0.73], Max-Change = 0.0365Stage 2 = 58, CDLL = -2004.1, AR(2.82; 0.00) = [0.40; 0.67], Max-Change = 0.0157Stage 2 = 59, CDLL = -1965.3, AR(2.82; 0.00) = [0.37; 0.71], Max-Change = 0.0135Stage 2 = 60, CDLL = -2019.5, AR(2.82; 0.00) = [0.40; 0.67], Max-Change = 0.0211Stage 2 = 61, CDLL = -1999.1, AR(2.82; 0.00) = [0.41; 0.80], Max-Change = 0.0368Stage 2 = 62, CDLL = -1983.1, AR(2.82; 0.00) = [0.42; 0.80], Max-Change = 0.0104Stage 2 = 63, CDLL = -1981.7, AR(2.82; 0.00) = [0.41; 0.65], Max-Change = 0.0265Stage 2 = 64, CDLL = -1983.4, AR(2.82; 0.00) = [0.38; 0.69], Max-Change = 0.0138Stage 2 = 65, CDLL = -1992.3, AR(2.82; 0.00) = [0.42; 0.63], Max-Change = 0.0203Stage 2 = 66, CDLL = -1969.3, AR(2.82; 0.00) = [0.42; 0.71], Max-Change = 0.0366Stage 2 = 67, CDLL = -1989.0, AR(2.82; 0.00) = [0.39; 0.65], Max-Change = 0.0227Stage 2 = 68, CDLL = -1966.9, AR(2.82; 0.00) = [0.40; 0.63], Max-Change = 0.0356Stage 2 = 69, CDLL = -1989.1, AR(2.82; 0.00) = [0.35; 0.61], Max-Change = 0.0329Stage 2 = 70, CDLL = -1968.1, AR(2.82; 0.00) = [0.41; 0.69], Max-Change = 0.0566Stage 2 = 71, CDLL = -2002.8, AR(2.82; 0.00) = [0.40; 0.69], Max-Change = 0.0124Stage 2 = 72, CDLL = -1961.4, AR(2.82; 0.00) = [0.35; 0.67], Max-Change = 0.0127Stage 2 = 73, CDLL = -1985.2, AR(2.82; 0.00) = [0.41; 0.57], Max-Change = 0.0163Stage 2 = 74, CDLL = -2015.9, AR(2.82; 0.00) = [0.40; 0.57], Max-Change = 0.0133Stage 2 = 75, CDLL = -1985.4, AR(2.82; 0.00) = [0.33; 0.71], Max-Change = 0.0307Stage 2 = 76, CDLL = -1965.5, AR(2.82; 0.00) = [0.42; 0.76], Max-Change = 0.0247Stage 2 = 77, CDLL = -1968.8, AR(2.82; 0.00) = [0.43; 0.76], Max-Change = 0.0069Stage 2 = 78, CDLL = -1985.5, AR(2.82; 0.00) = [0.40; 0.69], Max-Change = 0.0221Stage 2 = 79, CDLL = -1984.8, AR(2.82; 0.00) = [0.40; 0.80], Max-Change = 0.0158Stage 2 = 80, CDLL = -1983.8, AR(2.82; 0.00) = [0.42; 0.67], Max-Change = 0.0242Stage 2 = 81, CDLL = -1963.3, AR(2.82; 0.00) = [0.39; 0.84], Max-Change = 0.0052Stage 2 = 82, CDLL = -1979.9, AR(2.82; 0.00) = [0.45; 0.76], Max-Change = 0.0228Stage 2 = 83, CDLL = -1975.8, AR(2.82; 0.00) = [0.38; 0.76], Max-Change = 0.0170Stage 2 = 84, CDLL = -1975.7, AR(2.82; 0.00) = [0.43; 0.86], Max-Change = 0.0107Stage 2 = 85, CDLL = -1967.1, AR(2.82; 0.00) = [0.37; 0.86], Max-Change = 0.0293Stage 2 = 86, CDLL = -1985.1, AR(2.82; 0.00) = [0.41; 0.73], Max-Change = 0.0277Stage 2 = 87, CDLL = -1958.6, AR(2.82; 0.00) = [0.38; 0.80], Max-Change = 0.0145Stage 2 = 88, CDLL = -1979.4, AR(2.82; 0.00) = [0.39; 0.78], Max-Change = 0.0471Stage 2 = 89, CDLL = -1984.6, AR(2.82; 0.00) = [0.42; 0.67], Max-Change = 0.0053Stage 2 = 90, CDLL = -1998.5, AR(2.82; 0.00) = [0.42; 0.65], Max-Change = 0.0236Stage 2 = 91, CDLL = -1974.2, AR(2.82; 0.00) = [0.39; 0.78], Max-Change = 0.0284Stage 2 = 92, CDLL = -1969.4, AR(2.82; 0.00) = [0.40; 0.86], Max-Change = 0.0147Stage 2 = 93, CDLL = -1966.0, AR(2.82; 0.00) = [0.43; 0.78], Max-Change = 0.0189Stage 2 = 94, CDLL = -1982.4, AR(2.82; 0.00) = [0.37; 0.71], Max-Change = 0.0147Stage 2 = 95, CDLL = -1966.0, AR(2.82; 0.00) = [0.40; 0.73], Max-Change = 0.0253Stage 2 = 96, CDLL = -1974.5, AR(2.82; 0.00) = [0.44; 0.67], Max-Change = 0.0429Stage 2 = 97, CDLL = -1994.2, AR(2.82; 0.00) = [0.40; 0.71], Max-Change = 0.0081Stage 2 = 98, CDLL = -1965.6, AR(2.82; 0.00) = [0.40; 0.80], Max-Change = 0.0276Stage 2 = 99, CDLL = -1972.5, AR(2.82; 0.00) = [0.41; 0.80], Max-Change = 0.0364Stage 2 = 100, CDLL = -1980.7, AR(2.82; 0.00) = [0.38; 0.69], Max-Change = 0.0184Stage 3 = 1, CDLL = -1965.8, AR(2.82; 0.00) = [0.40; 0.82], gam = 0.0000, Max-Change = 0.0000Stage 3 = 2, CDLL = -2005.8, AR(2.82; 0.00) = [0.38; 0.59], gam = 0.1778, Max-Change = 0.0217Stage 3 = 3, CDLL = -1966.7, AR(2.82; 0.00) = [0.40; 0.73], gam = 0.1057, Max-Change = 0.0110Stage 3 = 4, CDLL = -1972.9, AR(2.82; 0.00) = [0.41; 0.76], gam = 0.0780, Max-Change = 0.0088Stage 3 = 5, CDLL = -1990.0, AR(2.82; 0.00) = [0.40; 0.67], gam = 0.0629, Max-Change = 0.0031Stage 3 = 6, CDLL = -1974.0, AR(2.82; 0.00) = [0.40; 0.71], gam = 0.0532, Max-Change = 0.0044Stage 3 = 7, CDLL = -1989.0, AR(2.82; 0.00) = [0.42; 0.73], gam = 0.0464, Max-Change = 0.0037Stage 3 = 8, CDLL = -2001.7, AR(2.82; 0.00) = [0.43; 0.80], gam = 0.0413, Max-Change = 0.0045Stage 3 = 9, CDLL = -1962.5, AR(2.82; 0.00) = [0.40; 0.73], gam = 0.0374, Max-Change = 0.0044Stage 3 = 10, CDLL = -1961.2, AR(2.82; 0.00) = [0.42; 0.82], gam = 0.0342, Max-Change = 0.0029Stage 3 = 11, CDLL = -1973.4, AR(2.82; 0.00) = [0.39; 0.69], gam = 0.0316, Max-Change = 0.0062Stage 3 = 12, CDLL = -1988.5, AR(2.82; 0.00) = [0.39; 0.55], gam = 0.0294, Max-Change = 0.0045Stage 3 = 13, CDLL = -1999.6, AR(2.82; 0.00) = [0.41; 0.65], gam = 0.0276, Max-Change = 0.0019Stage 3 = 14, CDLL = -1994.8, AR(2.82; 0.00) = [0.41; 0.67], gam = 0.0260, Max-Change = 0.0010Stage 3 = 15, CDLL = -1998.9, AR(2.82; 0.00) = [0.39; 0.65], gam = 0.0246, Max-Change = 0.0030Stage 3 = 16, CDLL = -1980.1, AR(2.82; 0.00) = [0.44; 0.69], gam = 0.0233, Max-Change = 0.0011Stage 3 = 17, CDLL = -2004.4, AR(2.82; 0.00) = [0.40; 0.63], gam = 0.0222, Max-Change = 0.0026Stage 3 = 18, CDLL = -1987.6, AR(2.82; 0.00) = [0.44; 0.73], gam = 0.0212, Max-Change = 0.0026Stage 3 = 19, CDLL = -1978.5, AR(2.82; 0.00) = [0.45; 0.61], gam = 0.0203, Max-Change = 0.0017Stage 3 = 20, CDLL = -1979.8, AR(2.82; 0.00) = [0.44; 0.78], gam = 0.0195, Max-Change = 0.0023Stage 3 = 21, CDLL = -1984.1, AR(2.82; 0.00) = [0.39; 0.73], gam = 0.0188, Max-Change = 0.0007Stage 3 = 22, CDLL = -1997.7, AR(2.82; 0.00) = [0.40; 0.55], gam = 0.0181, Max-Change = 0.0036Stage 3 = 23, CDLL = -1982.3, AR(2.82; 0.00) = [0.41; 0.69], gam = 0.0175, Max-Change = 0.0007Stage 3 = 24, CDLL = -1995.8, AR(2.82; 0.00) = [0.42; 0.67], gam = 0.0169, Max-Change = 0.0025Stage 3 = 25, CDLL = -1981.6, AR(2.82; 0.00) = [0.41; 0.71], gam = 0.0164, Max-Change = 0.0033Stage 3 = 26, CDLL = -1978.9, AR(2.82; 0.00) = [0.44; 0.80], gam = 0.0159, Max-Change = 0.0020Stage 3 = 27, CDLL = -1977.6, AR(2.82; 0.00) = [0.41; 0.71], gam = 0.0154, Max-Change = 0.0011Stage 3 = 28, CDLL = -1978.4, AR(2.82; 0.00) = [0.40; 0.73], gam = 0.0150, Max-Change = 0.0030Stage 3 = 29, CDLL = -1969.3, AR(2.82; 0.00) = [0.40; 0.78], gam = 0.0146, Max-Change = 0.0009Stage 3 = 30, CDLL = -1973.4, AR(2.82; 0.00) = [0.45; 0.76], gam = 0.0142, Max-Change = 0.0022Stage 3 = 31, CDLL = -1995.8, AR(2.82; 0.00) = [0.39; 0.59], gam = 0.0139, Max-Change = 0.0029Stage 3 = 32, CDLL = -1983.6, AR(2.82; 0.00) = [0.43; 0.73], gam = 0.0135, Max-Change = 0.0014Stage 3 = 33, CDLL = -1979.1, AR(2.82; 0.00) = [0.38; 0.73], gam = 0.0132, Max-Change = 0.0010Stage 3 = 34, CDLL = -1983.0, AR(2.82; 0.00) = [0.40; 0.65], gam = 0.0129, Max-Change = 0.0010Stage 3 = 35, CDLL = -1985.6, AR(2.82; 0.00) = [0.37; 0.69], gam = 0.0126, Max-Change = 0.0017Stage 3 = 36, CDLL = -1983.4, AR(2.82; 0.00) = [0.37; 0.69], gam = 0.0124, Max-Change = 0.0010Stage 3 = 37, CDLL = -1976.5, AR(2.82; 0.00) = [0.37; 0.67], gam = 0.0121, Max-Change = 0.0014Stage 3 = 38, CDLL = -1996.1, AR(2.82; 0.00) = [0.42; 0.57], gam = 0.0119, Max-Change = 0.0025Stage 3 = 39, CDLL = -1978.1, AR(2.82; 0.00) = [0.42; 0.67], gam = 0.0116, Max-Change = 0.0012Stage 3 = 40, CDLL = -2001.4, AR(2.82; 0.00) = [0.43; 0.65], gam = 0.0114, Max-Change = 0.0010Stage 3 = 41, CDLL = -1977.5, AR(2.82; 0.00) = [0.47; 0.76], gam = 0.0112, Max-Change = 0.0003Stage 3 = 42, CDLL = -2003.2, AR(2.82; 0.00) = [0.42; 0.63], gam = 0.0110, Max-Change = 0.0009
#> 
#> Calculating information matrix...
#> 
#> Calculating log-likelihood...
summary(mod)
#> 
#> Call:
#> mixedmirt(data = Science, covdata = covdat, model = 1, random = ~1 | 
#>     group)
#> 
#> 
#> --------------
#> RANDOM EFFECT COVARIANCE(S):
#> Correlations on upper diagonal
#> 
#> $Theta
#>      F1
#> F1 0.95
#> 
#> $group
#>           COV_group
#> COV_group    0.0185
#> 

effects <- randef(mod, ndraws = 2000, thin = 20)
head(effects$Theta)
#>              F1
#> [1,]  0.5782723
#> [2,]  0.1919645
#> [3,] -0.6486663
#> [4,] -0.7797778
#> [5,]  0.1012223
#> [6,]  0.9040769
head(effects$group)
#>               group
#> group1 -0.001573162
#> group2 -0.046624079
#> group3  0.014948268
#> group4 -0.019287051
#> group5 -0.015659220
#> group6  0.021948734

# }