residuals-method {mirt}	R Documentation

Compute model residuals

Description

Return model implied residuals for linear dependencies between items or at the person level.

Usage

## S4 method for signature 'SingleGroupClass'
residuals(object, type = "LD", digits = 3,
  df.p = FALSE, full.scores = FALSE, printvalue = NULL, tables = FALSE,
  verbose = TRUE, Theta = NULL, suppress = 1, ...)

Arguments

object
an object of class SingleGroupClass or MultipleGroupClass. Bifactor models are automatically detected and utilized for better accuracy

type
type of residuals to be displayed. Can be either 'LD' or 'LDG2' for a local dependence matrix based on the X2 or G2 statistics (Chen & Thissen, 1997), 'Q3' for the statistic proposed by Yen (1984), or 'exp' for the expected values for the frequencies of every response pattern

digits
number of significant digits to be rounded

df.p
logical; print the degrees of freedom and p-values?

full.scores
logical; compute relevant statistics for each subject in the original data?

printvalue
a numeric value to be specified when using the res='exp' option. Only prints patterns that have standardized residuals greater than abs(printvalue). The default (NULL) prints all response patterns

tables
logical; for LD type, return the observed, expected, and standardized residual tables for each item combination?

verbose
logical; allow information to be printed to the console?

Theta
a matrix of factor scores used for statistics that require empirical estimates (i.e., Q3). If supplied, arguments typically passed to fscores() will be ignored and these values will be used instead

suppress
a numeric value indiciating which parameter local dependency combinations to flag as being too high. Absolute values for the standardized estimates greater than this value will be returned, while all values less than this value will be set to NA

...
additional arguments to be passed to fscores()

References

Chen, W. H. & Thissen, D. (1997). Local dependence indices for item pairs using item response theory. Journal of Educational and Behavioral Statistics, 22, 265-289.

Yen, W. (1984). Effects of local item dependence on the fit and equating performance of the three parameter logistic model. Applied Psychological Measurement, 8, 125-145.

Examples

## No test: 

x <- mirt(Science, 1)
residuals(x)

## LD matrix (lower triangle) and standardized values:

##         Comfort    Work  Future Benefit
## Comfort      NA  -0.147  -0.101   0.152
## Work    -25.505      NA   0.088  -0.141
## Future  -12.003   9.206      NA  -0.122
## Benefit  27.332 -23.230 -17.465      NA

residuals(x, tables = TRUE)

## $Comfort_Work
## $Comfort_Work$Obs
##    
##       1   2   3   4
##   1   2   0   1   2
##   2   2  11  15   4
##   3  24  71 148  23
##   4   5  16  42  26
## 
## $Comfort_Work$Exp
##           [,1]      [,2]       [,3]       [,4]
## [1,]  1.045907  1.788133   1.923417  0.2630185
## [2,]  5.720902 11.386664  13.626073  1.9725198
## [3,] 23.320217 69.352985 140.026966 31.7182764
## [4,]  3.487657 14.112488  50.763038 21.4917381
## 
## $Comfort_Work$std_res
##    
##              1          2          3          4
##   1  0.9329195 -1.3372108 -0.6658267  3.3868976
##   2 -1.5556646 -0.1145871  0.3722016  1.4435968
##   3  0.1407680  0.1977721  0.6737795 -1.5480174
##   4  0.8098107  0.5024443 -1.2299315  0.9724636
## 
## 
## $Comfort_Future
## $Comfort_Future$Obs
##    
##       1   2   3   4
##   1   2   2   1   0
##   2   3  11  11   7
##   3   8  51 155  52
##   4   1   8  43  37
## 
## $Comfort_Future$Exp
##           [,1]      [,2]       [,3]      [,4]
## [1,] 0.7681441  1.813184   2.091831  0.347316
## [2,] 3.6750813 11.241541  15.083865  2.705672
## [3,] 9.1965787 52.386802 148.576424 54.258639
## [4,] 0.7523699  6.828174  42.538685 39.735692
## 
## $Comfort_Future$std_res
##    
##               1           2           3           4
##   1  1.40552502  0.13873696 -0.75490461 -0.58933526
##   2 -0.35214587 -0.07204074 -1.05151395  2.61070422
##   3 -0.39457370 -0.19160356  0.52698942 -0.30662841
##   4  0.28548797  0.44844679  0.07073026 -0.43398713
## 
## 
## $Comfort_Benefit
## $Comfort_Benefit$Obs
##    
##       1   2   3   4
##   1   4   0   1   0
##   2   5  10  13   4
##   3  11  81 133  41
##   4   1   9  46  33
## 
## $Comfort_Benefit$Exp
##            [,1]      [,2]       [,3]       [,4]
## [1,]  0.6709255  1.952142   1.960338  0.4370702
## [2,]  3.6233594 12.167645  13.682855  3.2322998
## [3,] 14.8467641 71.278452 131.341200 46.9520278
## [4,]  2.3182281 14.557981  44.803183 28.1755290
## 
## $Comfort_Benefit$std_res
##    
##              1          2          3          4
##   1  4.0643056 -1.3971907 -0.6858964 -0.6611129
##   2  0.7232108 -0.6214194 -0.1846036  0.4270078
##   3 -0.9983427  1.1514797  0.1447416 -0.8686359
##   4 -0.8657898 -1.4566878  0.1788024  0.9088949
## 
## 
## $Work_Future
## $Work_Future$Obs
##    
##       1   2   3   4
##   1   7  10  14   2
##   2   3  28  57  10
##   3   3  31 122  50
##   4   1   3  17  34
## 
## $Work_Future$Exp
##           [,1]      [,2]      [,3]      [,4]
## [1,] 4.6011602 12.483223  14.34091  2.149386
## [2,] 5.8024232 27.792074  52.26704 10.778733
## [3,] 3.6883121 28.933731 118.01358 55.703873
## [4,] 0.3002785  3.060673  23.66927 28.415327
## 
## $Work_Future$std_res
##    
##               1           2           3           4
##   1  1.11832378 -0.70283344 -0.09002332 -0.10189508
##   2 -1.16339989  0.03944103  0.65466460 -0.23719446
##   3 -0.35840293  0.38413562  0.36695879 -0.76423542
##   4  1.27691840 -0.03468052 -1.37083792  1.04766250
## 
## 
## $Work_Benefit
## $Work_Benefit$Obs
##    
##       1   2   3   4
##   1   4   8  12   9
##   2   6  34  47  11
##   3   8  52 111  35
##   4   3   6  23  23
## 
## $Work_Benefit$Exp
##          [,1]      [,2]      [,3]      [,4]
## [1,] 4.232417 13.168136  13.31561  2.858521
## [2,] 7.661162 32.195853  44.96520 11.818055
## [3,] 8.418174 46.929789 106.54865 44.442884
## [4,] 1.147524  7.662442  26.95812 19.677467
## 
## $Work_Benefit$std_res
##    
##              1          2          3          4
##   1 -0.1129726 -1.4242026 -0.3605341  3.6324723
##   2 -0.6001567  0.3179597  0.3034474 -0.2379631
##   3 -0.1441278  0.7401193  0.4312390 -1.4164574
##   4  1.7293047 -0.6005689 -0.7623319  0.7490049
## 
## 
## $Future_Benefit
## $Future_Benefit$Obs
##    
##       1   2   3   4
##   1   5   1   6   2
##   2   5  32  30   5
##   3   8  53 118  31
##   4   3  14  39  40
## 
## $Future_Benefit$Exp
##          [,1]      [,2]       [,3]       [,4]
## [1,] 2.960982  6.706231   4.142357  0.5826046
## [2,] 7.761205 29.144047  29.887350  5.4770998
## [3,] 9.227525 52.775488 109.946606 36.3411869
## [4,] 1.509566 11.330454  47.811264 36.3960355
## 
## $Future_Benefit$std_res
##    
##               1           2           3           4
##   1  1.18495856 -2.20348490  0.91272216  1.85696641
##   2 -0.99113779  0.52902494  0.02060568 -0.20386089
##   3 -0.40409882  0.03090459  0.76804746 -0.88600919
##   4  1.21307256  0.79307414 -1.27430418  0.59738383

residuals(x, type = 'exp')

##    Comfort Work Future Benefit freq    exp    res
## 86       1    1      1       1    2  0.124  5.325
## 54       1    3      2       1    1  0.067  3.605
## 16       1    4      2       3    1  0.019  7.045
## 85       1    4      3       1    1  0.006 12.640
## 90       2    1      1       1    1  0.460  0.796
## 57       2    1      2       4    1  0.095  2.930
## 29       2    2      1       1    1  0.351  1.095
## 72       2    2      2       2    4  2.147  1.264
## 53       2    2      2       3    2  1.616  0.302
## 74       2    2      3       1    1  0.377  1.016
## 89       2    2      3       2    1  1.715 -0.546
## 73       2    2      3       3    1  2.228 -0.823
## 24       2    2      4       3    1  0.251  1.497
## 14       2    3      1       3    1  0.213  1.708
## 23       2    3      2       2    2  1.545  0.366
## 65       2    3      2       3    1  1.489 -0.401
## 62       2    3      3       2    3  2.248  0.502
## 70       2    3      3       3    3  3.910 -0.460
## 46       2    3      3       4    2  1.036  0.947
## 48       2    3      4       1    1  0.041  4.764
## 30       2    3      4       3    2  0.866  1.218
## 32       2    4      2       1    1  0.029  5.689
## 79       2    4      4       3    2  0.259  3.417
## 49       2    4      4       4    1  0.184  1.901
## 83       3    1      1       1    1  0.644  0.443
## 36       3    1      1       3    2  0.638  1.704
## 13       3    1      2       2    2  3.923 -0.971
## 10       3    1      2       3    4  3.078  0.526
## 18       3    1      3       2    5  3.565  0.760
## 15       3    1      3       3    5  5.095 -0.042
## 69       3    1      3       4    3  1.160  1.709
## 87       3    1      4       3    1  0.764  0.270
## 42       3    1      4       4    1  0.320  1.203
## 37       3    2      1       2    1  1.782 -0.586
## 64       3    2      1       4    1  0.154  2.157
## 88       3    2      2       1    1  2.250 -0.834
## 38       3    2      2       2   10  8.473  0.524
## 3        3    2      2       3    7  8.070 -0.377
## 40       3    2      3       1    1  2.221 -0.819
## 12       3    2      3       2   16 11.742  1.243
## 17       3    2      3       3   22 19.586  0.546
## 6        3    2      3       4    5  4.951  0.022
## 82       3    2      4       1    1  0.193  1.835
## 50       3    2      4       2    1  1.293 -0.257
## 75       3    2      4       3    3  3.778 -0.400
## 60       3    2      4       4    2  1.717  0.216
## 66       3    3      1       3    2  0.971  1.045
## 59       3    3      2       1    1  1.764 -0.575
## 19       3    3      2       2   13  7.852  1.837
## 26       3    3      2       3   10  9.779  0.071
## 76       3    3      2       4    2  1.974  0.018
## 39       3    3      3       1    5  3.364  0.892
## 11       3    3      3       2   23 20.446  0.565
## 2        3    3      3       3   52 45.241  1.005
## 21       3    3      3       4    8 14.758 -1.759
## 7        3    3      4       2    7  4.389  1.246
## 9        3    3      4       3   13 16.981 -0.966
## 31       3    3      4       4   12 10.326  0.521
## 58       3    4      1       3    1  0.090  3.032
## 34       3    4      2       3    1  1.097 -0.093
## 55       3    4      3       2    2  3.043 -0.598
## 28       3    4      3       3    4  8.561 -1.559
## 51       3    4      3       4    4  3.635  0.191
## 4        3    4      4       1    1  0.160  2.099
## 78       3    4      4       2    1  1.287 -0.253
## 56       3    4      4       3    6  6.467 -0.184
## 63       3    4      4       4    3  5.660 -1.118
## 47       4    1      1       4    1  0.006 12.588
## 80       4    1      2       2    1  0.325  1.183
## 35       4    1      2       4    2  0.057  8.157
## 61       4    1      3       4    1  0.300  1.278
## 25       4    2      2       1    1  0.193  1.835
## 68       4    2      2       3    3  1.017  1.966
## 22       4    2      3       3    8  4.463  1.674
## 20       4    2      3       4    2  1.448  0.459
## 45       4    2      4       2    1  0.433  0.862
## 71       4    2      4       4    1  1.080 -0.077
## 84       4    3      2       3    1  1.665 -0.515
## 81       4    3      3       2    2  4.874 -1.302
## 8        4    3      3       3   21 14.000  1.871
## 33       4    3      3       4    3  5.963 -1.213
## 27       4    3      4       2    2  2.097 -0.067
## 67       4    3      4       3    5 10.420 -1.679
## 44       4    3      4       4    8  8.831 -0.280
## 1        4    4      3       2    1  0.966  0.035
## 5        4    4      3       3    2  3.569 -0.830
## 52       4    4      3       4    3  2.061  0.654
## 77       4    4      4       2    2  0.906  1.149
## 43       4    4      4       3    6  5.780  0.092
## 41       4    4      4       4   12  7.469  1.658

residuals(x, suppress = .15)

## LD matrix (lower triangle) and standardized values:

##         Comfort Work Future Benefit
## Comfort      NA   NA     NA   0.152
## Work         NA   NA     NA      NA
## Future       NA   NA     NA      NA
## Benefit      NA   NA     NA      NA

# with and without supplied factor scores
Theta <- fscores(x, full.scores=TRUE, scores.only=TRUE)
residuals(x, type = 'Q3', Theta=Theta)

## Q3 matrix:

##         Comfort   Work Future Benefit
## Comfort   1.000 -0.150 -0.097   0.112
## Work     -0.150  1.000 -0.127  -0.250
## Future   -0.097 -0.127  1.000  -0.430
## Benefit   0.112 -0.250 -0.430   1.000

residuals(x, type = 'Q3', method = 'ML')

## Q3 matrix:

##         Comfort   Work Future Benefit
## Comfort   1.000 -0.277 -0.348   0.033
## Work     -0.277  1.000 -0.391  -0.289
## Future   -0.348 -0.391  1.000  -0.532
## Benefit   0.033 -0.289 -0.532   1.000

## End(No test)

[Package mirt version 1.9 Index]