residuals-method {mirt}	R Documentation

Compute model residuals

Description

Return model implied residuals for linear dependencies between items or at the person level.

Usage

## S4 method for signature 'SingleGroupClass'
residuals(object, type = "LD", digits = 3,
  df.p = FALSE, full.scores = FALSE, printvalue = NULL, tables = FALSE,
  verbose = TRUE, Theta = NULL, suppress = 1, theta_lim = c(-6, 6),
  quadpts = NULL, ...)

Arguments

object
an object of class SingleGroupClass or MultipleGroupClass. Bifactor models are automatically detected and utilized for better accuracy

type
type of residuals to be displayed. Can be either 'LD' or 'LDG2' for a local dependence matrix based on the X2 or G2 statistics (Chen & Thissen, 1997), 'Q3' for the statistic proposed by Yen (1984), or 'exp' for the expected values for the frequencies of every response pattern

digits
number of significant digits to be rounded

df.p
logical; print the degrees of freedom and p-values?

full.scores
logical; compute relevant statistics for each subject in the original data?

printvalue
a numeric value to be specified when using the res='exp' option. Only prints patterns that have standardized residuals greater than abs(printvalue). The default (NULL) prints all response patterns

tables
logical; for LD type, return the observed, expected, and standardized residual tables for each item combination?

verbose
logical; allow information to be printed to the console?

Theta
a matrix of factor scores used for statistics that require empirical estimates (i.e., Q3). If supplied, arguments typically passed to fscores() will be ignored and these values will be used instead

suppress
a numeric value indicating which parameter local dependency combinations to flag as being too high. Absolute values for the standardized estimates greater than this value will be returned, while all values less than this value will be set to NA

theta_lim
range for the integration grid

quadpts
number of quadrature nodes to use. The default is extracted from model (if available) or generated automatically if not available

...
additional arguments to be passed to fscores()

References

Chen, W. H. & Thissen, D. (1997). Local dependence indices for item pairs using item response theory. Journal of Educational and Behavioral Statistics, 22, 265-289.

Yen, W. (1984). Effects of local item dependence on the fit and equating performance of the three parameter logistic model. Applied Psychological Measurement, 8, 125-145.

Examples

## No test: 

x <- mirt(Science, 1)
residuals(x)

## LD matrix (lower triangle) and standardized values:

##         Comfort    Work  Future Benefit
## Comfort      NA  -0.147  -0.101   0.152
## Work    -25.512      NA   0.088  -0.141
## Future  -12.002   9.208      NA  -0.122
## Benefit  27.321 -23.235 -17.461      NA

residuals(x, tables = TRUE)

## $Comfort_Work
## $Comfort_Work$Obs
##    
##       1   2   3   4
##   1   2   0   1   2
##   2   2  11  15   4
##   3  24  71 148  23
##   4   5  16  42  26
## 
## $Comfort_Work$Exp
##           [,1]      [,2]       [,3]      [,4]
## [1,]  1.046123  1.787854   1.922925  0.262908
## [2,]  5.722535 11.386923  13.625623  1.972172
## [3,] 23.322258 69.348472 140.029013 31.718143
## [4,]  3.487224 14.109434  50.763258 21.495135
## 
## $Comfort_Work$std_res
##    
##              1          2          3          4
##   1  0.9326115 -1.3371066 -0.6655567  3.3878247
##   2 -1.5561252 -0.1146625  0.3723298  1.4439717
##   3  0.1403392  0.1983204  0.6736015 -1.5479971
##   4  0.8100928  0.5033119 -1.2299596  0.9716541
## 
## 
## $Comfort_Future
## $Comfort_Future$Obs
##    
##       1   2   3   4
##   1   2   2   1   0
##   2   3  11  11   7
##   3   8  51 155  52
##   4   1   8  43  37
## 
## $Comfort_Future$Exp
##           [,1]      [,2]       [,3]       [,4]
## [1,] 0.7680477  1.812936   2.091572  0.3472546
## [2,] 3.6749826 11.241583  15.084907  2.7057791
## [3,] 9.1976778 52.384929 148.579175 54.2561060
## [4,] 0.7527106  6.828614  42.541147 39.7325783
## 
## $Comfort_Future$std_res
##    
##               1           2           3           4
##   1  1.40572316  0.13893088 -0.75477217 -0.58928310
##   2 -0.35209910 -0.07205318 -1.05174604  2.61058722
##   3 -0.39491253 -0.19134814  0.52675890 -0.30629167
##   4  0.28503062  0.44826372  0.07035076 -0.43351011
## 
## 
## $Comfort_Benefit
## $Comfort_Benefit$Obs
##    
##       1   2   3   4
##   1   4   0   1   0
##   2   5  10  13   4
##   3  11  81 133  41
##   4   1   9  46  33
## 
## $Comfort_Benefit$Exp
##            [,1]     [,2]       [,3]       [,4]
## [1,]  0.6710457  1.95217   1.959779  0.4368153
## [2,]  3.6241191 12.16983  13.682082  3.2312240
## [3,] 14.8454892 71.28246 131.343241 46.9466978
## [4,]  2.3173267 14.55596  44.803655 28.1781067
## 
## $Comfort_Benefit$std_res
##    
##              1          2          3          4
##   1  4.0637950 -1.3972006 -0.6855953 -0.6609200
##   2  0.7227359 -0.6219892 -0.1844000  0.4276774
##   3 -0.9980547  1.1509727  0.1445624 -0.8679073
##   4 -0.8653661 -1.4562597  0.1787310  0.9083677
## 
## 
## $Work_Future
## $Work_Future$Obs
##    
##       1   2   3   4
##   1   7  10  14   2
##   2   3  28  57  10
##   3   3  31 122  50
##   4   1   3  17  34
## 
## $Work_Future$Exp
##           [,1]      [,2]      [,3]      [,4]
## [1,] 4.6004355 12.483087  14.34426  2.150355
## [2,] 5.8020446 27.786418  52.26417 10.780046
## [3,] 3.6903551 28.936581 118.01439 55.699488
## [4,] 0.3005835  3.061977  23.67397 28.411829
## 
## $Work_Future$std_res
##    
##               1           2           3           4
##   1  1.11874979 -0.70279865 -0.09089724 -0.10253275
##   2 -1.16328067  0.04051800  0.65507895 -0.23757996
##   3 -0.35936723  0.38358699  0.36688237 -0.76367799
##   4  1.27571400 -0.03541829 -1.37166696  1.04838334
## 
## 
## $Work_Benefit
## $Work_Benefit$Obs
##    
##       1   2   3   4
##   1   4   8  12   9
##   2   6  34  47  11
##   3   8  52 111  35
##   4   3   6  23  23
## 
## $Work_Benefit$Exp
##          [,1]      [,2]      [,3]      [,4]
## [1,] 4.233251 13.170723  13.31610  2.858063
## [2,] 7.660483 32.196149  44.96112 11.814929
## [3,] 8.417022 46.931424 106.55195 44.440425
## [4,] 1.147224  7.662122  26.95958 19.679427
## 
## $Work_Benefit$std_res
##    
##              1          2          3          4
##   1 -0.1133671 -1.4247756 -0.3606627  3.6330344
##   2 -0.5999380  0.3179060  0.3040694 -0.2370851
##   3 -0.1437407  0.7398678  0.4309125 -1.4161278
##   4  1.7298113 -0.6004660 -0.7625934  0.7485258
## 
## 
## $Future_Benefit
## $Future_Benefit$Obs
##    
##       1   2   3   4
##   1   5   1   6   2
##   2   5  32  30   5
##   3   8  53 118  31
##   4   3  14  39  40
## 
## $Future_Benefit$Exp
##          [,1]      [,2]       [,3]       [,4]
## [1,] 2.960508  6.706374   4.143629  0.5829072
## [2,] 7.760422 29.142832  29.887302  5.4775063
## [3,] 9.227453 52.779344 109.948223 36.3417805
## [4,] 1.509597 11.331867  47.809604 36.3906499
## 
## $Future_Benefit$std_res
##    
##               1           2           3           4
##   1  1.18532911 -2.20351681  0.91195681  1.85608815
##   2 -0.99090687  0.52926095  0.02061456 -0.20402700
##   3 -0.40407692  0.03037266  0.76788760 -0.88610042
##   4  1.21303423  0.79260501 -1.27408623  0.59832081

residuals(x, type = 'exp')

##    Comfort Work Future Benefit freq    exp    res
## 86       1    1      1       1    2  0.124  5.324
## 54       1    3      2       1    1  0.067  3.606
## 16       1    4      2       3    1  0.019  7.046
## 85       1    4      3       1    1  0.006 12.642
## 90       2    1      1       1    1  0.460  0.796
## 57       2    1      2       4    1  0.095  2.930
## 29       2    2      1       1    1  0.351  1.095
## 72       2    2      2       2    4  2.147  1.264
## 53       2    2      2       3    2  1.616  0.302
## 74       2    2      3       1    1  0.377  1.015
## 89       2    2      3       2    1  1.716 -0.547
## 73       2    2      3       3    1  2.228 -0.823
## 24       2    2      4       3    1  0.251  1.497
## 14       2    3      1       3    1  0.213  1.707
## 23       2    3      2       2    2  1.545  0.366
## 65       2    3      2       3    1  1.489 -0.401
## 62       2    3      3       2    3  2.248  0.501
## 70       2    3      3       3    3  3.910 -0.460
## 46       2    3      3       4    2  1.035  0.948
## 48       2    3      4       1    1  0.041  4.763
## 30       2    3      4       3    2  0.866  1.218
## 32       2    4      2       1    1  0.029  5.690
## 79       2    4      4       3    2  0.259  3.417
## 49       2    4      4       4    1  0.184  1.902
## 83       3    1      1       1    1  0.644  0.444
## 36       3    1      1       3    2  0.638  1.704
## 13       3    1      2       2    2  3.923 -0.971
## 10       3    1      2       3    4  3.078  0.526
## 18       3    1      3       2    5  3.568  0.758
## 15       3    1      3       3    5  5.097 -0.043
## 69       3    1      3       4    3  1.160  1.709
## 87       3    1      4       3    1  0.764  0.269
## 42       3    1      4       4    1  0.320  1.203
## 37       3    2      1       2    1  1.781 -0.585
## 64       3    2      1       4    1  0.154  2.157
## 88       3    2      2       1    1  2.249 -0.833
## 38       3    2      2       2   10  8.471  0.525
## 3        3    2      2       3    7  8.067 -0.376
## 40       3    2      3       1    1  2.221 -0.819
## 12       3    2      3       2   16 11.743  1.242
## 17       3    2      3       3   22 19.583  0.546
## 6        3    2      3       4    5  4.949  0.023
## 82       3    2      4       1    1  0.193  1.835
## 50       3    2      4       2    1  1.293 -0.258
## 75       3    2      4       3    3  3.777 -0.400
## 60       3    2      4       4    2  1.716  0.217
## 66       3    3      1       3    2  0.971  1.044
## 59       3    3      2       1    1  1.764 -0.575
## 19       3    3      2       2   13  7.851  1.837
## 26       3    3      2       3   10  9.780  0.070
## 76       3    3      2       4    2  1.975  0.018
## 39       3    3      3       1    5  3.363  0.893
## 11       3    3      3       2   23 20.448  0.564
## 2        3    3      3       3   52 45.244  1.004
## 21       3    3      3       4    8 14.757 -1.759
## 7        3    3      4       2    7  4.390  1.246
## 9        3    3      4       3   13 16.979 -0.966
## 31       3    3      4       4   12 10.321  0.522
## 58       3    4      1       3    1  0.090  3.030
## 34       3    4      2       3    1  1.098 -0.093
## 55       3    4      3       2    2  3.043 -0.598
## 28       3    4      3       3    4  8.563 -1.559
## 51       3    4      3       4    4  3.637  0.191
## 4        3    4      4       1    1  0.160  2.100
## 78       3    4      4       2    1  1.287 -0.253
## 56       3    4      4       3    6  6.466 -0.183
## 63       3    4      4       4    3  5.659 -1.118
## 47       4    1      1       4    1  0.006 12.587
## 80       4    1      2       2    1  0.325  1.183
## 35       4    1      2       4    2  0.057  8.158
## 61       4    1      3       4    1  0.300  1.278
## 25       4    2      2       1    1  0.193  1.836
## 68       4    2      2       3    3  1.017  1.967
## 22       4    2      3       3    8  4.462  1.675
## 20       4    2      3       4    2  1.447  0.459
## 45       4    2      4       2    1  0.433  0.862
## 71       4    2      4       4    1  1.080 -0.077
## 84       4    3      2       3    1  1.665 -0.516
## 81       4    3      3       2    2  4.874 -1.302
## 8        4    3      3       3   21 14.003  1.870
## 33       4    3      3       4    3  5.965 -1.214
## 27       4    3      4       2    2  2.097 -0.067
## 67       4    3      4       3    5 10.419 -1.679
## 44       4    3      4       4    8  8.828 -0.279
## 1        4    4      3       2    1  0.966  0.034
## 5        4    4      3       3    2  3.571 -0.831
## 52       4    4      3       4    3  2.063  0.652
## 77       4    4      4       2    2  0.906  1.149
## 43       4    4      4       3    6  5.780  0.091
## 41       4    4      4       4   12  7.470  1.657

residuals(x, suppress = .15)

## LD matrix (lower triangle) and standardized values:

##         Comfort Work Future Benefit
## Comfort      NA   NA     NA   0.152
## Work         NA   NA     NA      NA
## Future       NA   NA     NA      NA
## Benefit      NA   NA     NA      NA

# with and without supplied factor scores
Theta <- fscores(x, full.scores=TRUE)
residuals(x, type = 'Q3', Theta=Theta)

## Q3 matrix:

##         Comfort   Work Future Benefit
## Comfort   1.000 -0.150 -0.097   0.112
## Work     -0.150  1.000 -0.127  -0.251
## Future   -0.097 -0.127  1.000  -0.430
## Benefit   0.112 -0.251 -0.430   1.000

residuals(x, type = 'Q3', method = 'ML')

## Q3 matrix:

##         Comfort   Work Future Benefit
## Comfort   1.000 -0.277 -0.348   0.033
## Work     -0.277  1.000 -0.391  -0.289
## Future   -0.348 -0.391  1.000  -0.532
## Benefit   0.033 -0.289 -0.532   1.000

## End(No test)

[Package mirt version 1.10 Index]