summary-method {mirt}	R Documentation

Summary of model object

Description

Transforms coefficients into a standardized factor loading's metric. For MixedClass objects, the fixed and random coefficients are printed.

Usage

## S4 method for signature 'SingleGroupClass'
summary(object, rotate = NULL, Target = NULL,
  suppress = 0, digits = 3, printCI = FALSE, verbose = TRUE, ...)

Arguments

object
an object of class SingleGroupClass, MultipleGroupClass, or MixedClass

rotate
a string indicating which rotation to use for exploratory models, primarily from the GPArotation package (see documentation therein).

Rotations currently supported are: 'promax', 'oblimin', 'varimax', 'quartimin', 'targetT', 'targetQ', 'pstT', 'pstQ', 'oblimax', 'entropy', 'quartimax', 'simplimax', 'bentlerT', 'bentlerQ', 'tandemI', 'tandemII', 'geominT', 'geominQ', 'cfT', 'cfQ', 'infomaxT', 'infomaxQ', 'mccammon', 'bifactorT', 'bifactorQ'

Target
a dummy variable matrix indicting a target rotation pattern

suppress
a numeric value indicating which (possibly rotated) factor loadings should be suppressed. Typical values are around .3 in most statistical software. Default is 0 for no suppression

digits
number of significant digits to be rounded

printCI
print a confidence interval for standardized loadings (e.g., printCI = .95 gives a 95% confidence interval)

verbose
logical; allow information to be printed to the console?

...
additional arguments to be passed

See Also

coef-method

Examples

## No test: 
x <- mirt(Science, 2)
summary(x)

## 
## Rotation:  oblimin 
## 
## Rotated factor loadings: 
## 
##             F1     F2    h2
## Comfort  0.598  0.035 0.380
## Work    -0.056  0.800 0.597
## Future   0.333  0.512 0.546
## Benefit  0.728 -0.026 0.511
## 
## Rotated SS loadings:  1.001 0.904 
## 
## Factor correlations: 
## 
##       F1    F2
## F1 1.000 0.506
## F2 0.506 1.000

summary(x, rotate = 'varimax')

## 
## Rotation:  varimax 
## 
## Rotated factor loadings: 
## 
##            F1     F2    h2
## Comfort 0.217 -0.577 0.380
## Work    0.763 -0.121 0.597
## Future  0.602 -0.429 0.546
## Benefit 0.197 -0.687 0.511
## 
## Rotated SS loadings:  1.031 1.004 
## 
## Factor correlations: 
## 
##    F1 F2
## F1  1  0
## F2  0  1

#print confidence interval (requires computed information matrix)
x2 <- mirt(Science, 1, SE=TRUE)
summary(x2, printCI=.95)

##            F1    h2
## Comfort 0.522 0.272
## Work    0.584 0.342
## Future  0.803 0.645
## Benefit 0.541 0.293
## 
## SS loadings:  1.552 
## Proportion Var:  0.388 
## 
## Factor correlations: 
## 
##    F1
## F1  1

## End(No test)

[Package mirt version 1.9 Index]