Summary of model object

Transforms coefficients into a standardized factor loading's metric. For MixedClass objects, the fixed and random coefficients are printed. Note that while the output to the console is rounded to three digits, the returned list of objects is not. For simulations, use output <- summary(mod, verbose = FALSE) to suppress the console messages.

Usage

# S4 method for class 'SingleGroupClass'
summary(
  object,
  SE = TRUE,
  rotate = "oblimin",
  Target = NULL,
  suppress = 0,
  suppress.cor = 0,
  verbose = TRUE,
  ...
)

Arguments

object

an object of class SingleGroupClass, MultipleGroupClass, or MixedClass

SE

logical; include the standard errors for the standardized loadings? Requires the initial model to have included and estimated of the asymptotic covariance matrix (via, for instance, mirt(..., SE = TRUE)). If TRUE SEs are computed using the delta method

rotate

a string indicating which rotation to use for exploratory models, primarily from the GPArotation package (see documentation therein).

Rotations currently supported are: 'promax', 'oblimin', 'varimax', 'quartimin', 'targetT', 'targetQ', 'pstT', 'pstQ', 'oblimax', 'entropy', 'quartimax', 'simplimax', 'bentlerT', 'bentlerQ', 'tandemI', 'tandemII', 'geominT', 'geominQ', 'cfT', 'cfQ', 'infomaxT', 'infomaxQ', 'mccammon', 'bifactorT', 'bifactorQ'.

For models that are not exploratory this input will automatically be set to 'none'

Target

a dummy variable matrix indicting a target rotation pattern. This is required for rotations such as 'targetT', 'targetQ', 'pstT', and 'pstQ'

suppress

a numeric value indicating which (possibly rotated) factor loadings should be suppressed. Typical values are around .3 in most statistical software. Default is 0 for no suppression

suppress.cor

same as suppress, but for the correlation matrix output

verbose

logical; allow information to be printed to the console?

...

additional arguments to be passed

References

Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29. doi:10.18637/jss.v048.i06

See also

Examples

# \donttest{
x <- mirt(Science, 2)
#> Iteration: 1, Log-Lik: -1634.562, Max-Change: 0.49365Iteration: 2, Log-Lik: -1608.496, Max-Change: 0.17069Iteration: 3, Log-Lik: -1604.784, Max-Change: 0.08523Iteration: 4, Log-Lik: -1604.195, Max-Change: 0.05345Iteration: 5, Log-Lik: -1603.777, Max-Change: 0.03491Iteration: 6, Log-Lik: -1603.581, Max-Change: 0.02948Iteration: 7, Log-Lik: -1603.263, Max-Change: 0.02614Iteration: 8, Log-Lik: -1603.177, Max-Change: 0.02281Iteration: 9, Log-Lik: -1603.099, Max-Change: 0.02311Iteration: 10, Log-Lik: -1602.717, Max-Change: 0.01961Iteration: 11, Log-Lik: -1602.673, Max-Change: 0.01556Iteration: 12, Log-Lik: -1602.635, Max-Change: 0.01694Iteration: 13, Log-Lik: -1602.443, Max-Change: 0.01534Iteration: 14, Log-Lik: -1602.421, Max-Change: 0.01311Iteration: 15, Log-Lik: -1602.401, Max-Change: 0.01384Iteration: 16, Log-Lik: -1602.303, Max-Change: 0.00987Iteration: 17, Log-Lik: -1602.288, Max-Change: 0.01006Iteration: 18, Log-Lik: -1602.275, Max-Change: 0.00939Iteration: 19, Log-Lik: -1602.206, Max-Change: 0.00972Iteration: 20, Log-Lik: -1602.198, Max-Change: 0.01016Iteration: 21, Log-Lik: -1602.189, Max-Change: 0.00670Iteration: 22, Log-Lik: -1602.165, Max-Change: 0.00679Iteration: 23, Log-Lik: -1602.158, Max-Change: 0.00694Iteration: 24, Log-Lik: -1602.152, Max-Change: 0.00555Iteration: 25, Log-Lik: -1602.116, Max-Change: 0.00689Iteration: 26, Log-Lik: -1602.112, Max-Change: 0.00465Iteration: 27, Log-Lik: -1602.107, Max-Change: 0.00478Iteration: 28, Log-Lik: -1602.082, Max-Change: 0.00389Iteration: 29, Log-Lik: -1602.078, Max-Change: 0.00745Iteration: 30, Log-Lik: -1602.074, Max-Change: 0.00380Iteration: 31, Log-Lik: -1602.069, Max-Change: 0.00373Iteration: 32, Log-Lik: -1602.066, Max-Change: 0.00792Iteration: 33, Log-Lik: -1602.063, Max-Change: 0.00397Iteration: 34, Log-Lik: -1602.060, Max-Change: 0.00364Iteration: 35, Log-Lik: -1602.057, Max-Change: 0.00808Iteration: 36, Log-Lik: -1602.055, Max-Change: 0.00332Iteration: 37, Log-Lik: -1602.052, Max-Change: 0.00613Iteration: 38, Log-Lik: -1602.049, Max-Change: 0.00319Iteration: 39, Log-Lik: -1602.047, Max-Change: 0.00596Iteration: 40, Log-Lik: -1602.043, Max-Change: 0.00303Iteration: 41, Log-Lik: -1602.040, Max-Change: 0.00302Iteration: 42, Log-Lik: -1602.038, Max-Change: 0.00611Iteration: 43, Log-Lik: -1602.034, Max-Change: 0.00392Iteration: 44, Log-Lik: -1602.033, Max-Change: 0.00329Iteration: 45, Log-Lik: -1602.030, Max-Change: 0.00406Iteration: 46, Log-Lik: -1602.026, Max-Change: 0.00303Iteration: 47, Log-Lik: -1602.025, Max-Change: 0.00261Iteration: 48, Log-Lik: -1602.023, Max-Change: 0.00457Iteration: 49, Log-Lik: -1602.021, Max-Change: 0.00265Iteration: 50, Log-Lik: -1602.020, Max-Change: 0.00290Iteration: 51, Log-Lik: -1602.018, Max-Change: 0.00268Iteration: 52, Log-Lik: -1602.010, Max-Change: 0.00310Iteration: 53, Log-Lik: -1602.008, Max-Change: 0.00361Iteration: 54, Log-Lik: -1602.007, Max-Change: 0.00238Iteration: 55, Log-Lik: -1602.006, Max-Change: 0.00327Iteration: 56, Log-Lik: -1602.005, Max-Change: 0.00269Iteration: 57, Log-Lik: -1602.004, Max-Change: 0.00324Iteration: 58, Log-Lik: -1602.002, Max-Change: 0.00218Iteration: 59, Log-Lik: -1602.001, Max-Change: 0.00198Iteration: 60, Log-Lik: -1602.000, Max-Change: 0.00346Iteration: 61, Log-Lik: -1601.999, Max-Change: 0.00226Iteration: 62, Log-Lik: -1601.998, Max-Change: 0.00247Iteration: 63, Log-Lik: -1601.997, Max-Change: 0.00264Iteration: 64, Log-Lik: -1601.996, Max-Change: 0.00380Iteration: 65, Log-Lik: -1601.995, Max-Change: 0.00236Iteration: 66, Log-Lik: -1601.994, Max-Change: 0.00244Iteration: 67, Log-Lik: -1601.992, Max-Change: 0.00164Iteration: 68, Log-Lik: -1601.992, Max-Change: 0.00214Iteration: 69, Log-Lik: -1601.991, Max-Change: 0.00193Iteration: 70, Log-Lik: -1601.990, Max-Change: 0.00267Iteration: 71, Log-Lik: -1601.989, Max-Change: 0.00227Iteration: 72, Log-Lik: -1601.989, Max-Change: 0.00232Iteration: 73, Log-Lik: -1601.988, Max-Change: 0.00172Iteration: 74, Log-Lik: -1601.987, Max-Change: 0.00149Iteration: 75, Log-Lik: -1601.987, Max-Change: 0.00245Iteration: 76, Log-Lik: -1601.986, Max-Change: 0.00149Iteration: 77, Log-Lik: -1601.985, Max-Change: 0.00215Iteration: 78, Log-Lik: -1601.985, Max-Change: 0.00206Iteration: 79, Log-Lik: -1601.984, Max-Change: 0.00223Iteration: 80, Log-Lik: -1601.983, Max-Change: 0.00190Iteration: 81, Log-Lik: -1601.983, Max-Change: 0.00181Iteration: 82, Log-Lik: -1601.982, Max-Change: 0.00125Iteration: 83, Log-Lik: -1601.981, Max-Change: 0.00136Iteration: 84, Log-Lik: -1601.981, Max-Change: 0.00207Iteration: 85, Log-Lik: -1601.981, Max-Change: 0.00124Iteration: 86, Log-Lik: -1601.980, Max-Change: 0.00187Iteration: 87, Log-Lik: -1601.980, Max-Change: 0.00123Iteration: 88, Log-Lik: -1601.980, Max-Change: 0.00193Iteration: 89, Log-Lik: -1601.980, Max-Change: 0.00123Iteration: 90, Log-Lik: -1601.979, Max-Change: 0.00198Iteration: 91, Log-Lik: -1601.979, Max-Change: 0.00118Iteration: 92, Log-Lik: -1601.979, Max-Change: 0.00121Iteration: 93, Log-Lik: -1601.979, Max-Change: 0.00196Iteration: 94, Log-Lik: -1601.978, Max-Change: 0.00112Iteration: 95, Log-Lik: -1601.978, Max-Change: 0.00117Iteration: 96, Log-Lik: -1601.978, Max-Change: 0.00193Iteration: 97, Log-Lik: -1601.978, Max-Change: 0.00106Iteration: 98, Log-Lik: -1601.977, Max-Change: 0.00112Iteration: 99, Log-Lik: -1601.977, Max-Change: 0.00187Iteration: 100, Log-Lik: -1601.977, Max-Change: 0.00103Iteration: 101, Log-Lik: -1601.977, Max-Change: 0.00109Iteration: 102, Log-Lik: -1601.977, Max-Change: 0.00127Iteration: 103, Log-Lik: -1601.976, Max-Change: 0.00118Iteration: 104, Log-Lik: -1601.976, Max-Change: 0.00115Iteration: 105, Log-Lik: -1601.976, Max-Change: 0.00115Iteration: 106, Log-Lik: -1601.975, Max-Change: 0.00108Iteration: 107, Log-Lik: -1601.975, Max-Change: 0.00108Iteration: 108, Log-Lik: -1601.975, Max-Change: 0.00107Iteration: 109, Log-Lik: -1601.974, Max-Change: 0.00103Iteration: 110, Log-Lik: -1601.974, Max-Change: 0.00102Iteration: 111, Log-Lik: -1601.974, Max-Change: 0.00101Iteration: 112, Log-Lik: -1601.974, Max-Change: 0.00083Iteration: 113, Log-Lik: -1601.974, Max-Change: 0.00088Iteration: 114, Log-Lik: -1601.974, Max-Change: 0.00090Iteration: 115, Log-Lik: -1601.973, Max-Change: 0.00108Iteration: 116, Log-Lik: -1601.973, Max-Change: 0.00037Iteration: 117, Log-Lik: -1601.973, Max-Change: 0.00138Iteration: 118, Log-Lik: -1601.973, Max-Change: 0.00320Iteration: 119, Log-Lik: -1601.973, Max-Change: 0.00085Iteration: 120, Log-Lik: -1601.973, Max-Change: 0.00039Iteration: 121, Log-Lik: -1601.973, Max-Change: 0.00034Iteration: 122, Log-Lik: -1601.973, Max-Change: 0.00026Iteration: 123, Log-Lik: -1601.973, Max-Change: 0.00109Iteration: 124, Log-Lik: -1601.973, Max-Change: 0.00128Iteration: 125, Log-Lik: -1601.973, Max-Change: 0.00113Iteration: 126, Log-Lik: -1601.973, Max-Change: 0.00094Iteration: 127, Log-Lik: -1601.973, Max-Change: 0.00080Iteration: 128, Log-Lik: -1601.972, Max-Change: 0.00028Iteration: 129, Log-Lik: -1601.972, Max-Change: 0.00105Iteration: 130, Log-Lik: -1601.972, Max-Change: 0.00291Iteration: 131, Log-Lik: -1601.972, Max-Change: 0.00082Iteration: 132, Log-Lik: -1601.972, Max-Change: 0.00034Iteration: 133, Log-Lik: -1601.972, Max-Change: 0.00029Iteration: 134, Log-Lik: -1601.972, Max-Change: 0.00112Iteration: 135, Log-Lik: -1601.972, Max-Change: 0.00244Iteration: 136, Log-Lik: -1601.972, Max-Change: 0.00039Iteration: 137, Log-Lik: -1601.972, Max-Change: 0.00027Iteration: 138, Log-Lik: -1601.972, Max-Change: 0.00022Iteration: 139, Log-Lik: -1601.972, Max-Change: 0.00112Iteration: 140, Log-Lik: -1601.972, Max-Change: 0.00024Iteration: 141, Log-Lik: -1601.972, Max-Change: 0.00098Iteration: 142, Log-Lik: -1601.972, Max-Change: 0.00261Iteration: 143, Log-Lik: -1601.972, Max-Change: 0.00031Iteration: 144, Log-Lik: -1601.972, Max-Change: 0.00023Iteration: 145, Log-Lik: -1601.972, Max-Change: 0.00208Iteration: 146, Log-Lik: -1601.972, Max-Change: 0.00074Iteration: 147, Log-Lik: -1601.972, Max-Change: 0.00031Iteration: 148, Log-Lik: -1601.972, Max-Change: 0.00027Iteration: 149, Log-Lik: -1601.971, Max-Change: 0.00104Iteration: 150, Log-Lik: -1601.971, Max-Change: 0.00114Iteration: 151, Log-Lik: -1601.971, Max-Change: 0.00026Iteration: 152, Log-Lik: -1601.971, Max-Change: 0.00099Iteration: 153, Log-Lik: -1601.971, Max-Change: 0.00235Iteration: 154, Log-Lik: -1601.971, Max-Change: 0.00028Iteration: 155, Log-Lik: -1601.971, Max-Change: 0.00022Iteration: 156, Log-Lik: -1601.971, Max-Change: 0.00019Iteration: 157, Log-Lik: -1601.971, Max-Change: 0.00104Iteration: 158, Log-Lik: -1601.971, Max-Change: 0.00023Iteration: 159, Log-Lik: -1601.971, Max-Change: 0.00088Iteration: 160, Log-Lik: -1601.971, Max-Change: 0.00085Iteration: 161, Log-Lik: -1601.971, Max-Change: 0.00030Iteration: 162, Log-Lik: -1601.971, Max-Change: 0.00019Iteration: 163, Log-Lik: -1601.971, Max-Change: 0.00075Iteration: 164, Log-Lik: -1601.971, Max-Change: 0.00106Iteration: 165, Log-Lik: -1601.971, Max-Change: 0.00093Iteration: 166, Log-Lik: -1601.971, Max-Change: 0.00081Iteration: 167, Log-Lik: -1601.971, Max-Change: 0.00068Iteration: 168, Log-Lik: -1601.971, Max-Change: 0.00022Iteration: 169, Log-Lik: -1601.971, Max-Change: 0.00020Iteration: 170, Log-Lik: -1601.971, Max-Change: 0.00080Iteration: 171, Log-Lik: -1601.971, Max-Change: 0.00104Iteration: 172, Log-Lik: -1601.971, Max-Change: 0.00026Iteration: 173, Log-Lik: -1601.971, Max-Change: 0.00019Iteration: 174, Log-Lik: -1601.971, Max-Change: 0.00078Iteration: 175, Log-Lik: -1601.971, Max-Change: 0.00103Iteration: 176, Log-Lik: -1601.971, Max-Change: 0.00071Iteration: 177, Log-Lik: -1601.971, Max-Change: 0.00014Iteration: 178, Log-Lik: -1601.971, Max-Change: 0.00066Iteration: 179, Log-Lik: -1601.971, Max-Change: 0.00021Iteration: 180, Log-Lik: -1601.971, Max-Change: 0.00065Iteration: 181, Log-Lik: -1601.971, Max-Change: 0.00147Iteration: 182, Log-Lik: -1601.970, Max-Change: 0.00020Iteration: 183, Log-Lik: -1601.970, Max-Change: 0.00017Iteration: 184, Log-Lik: -1601.970, Max-Change: 0.00062Iteration: 185, Log-Lik: -1601.970, Max-Change: 0.00058Iteration: 186, Log-Lik: -1601.970, Max-Change: 0.00064Iteration: 187, Log-Lik: -1601.970, Max-Change: 0.00105Iteration: 188, Log-Lik: -1601.970, Max-Change: 0.00019Iteration: 189, Log-Lik: -1601.970, Max-Change: 0.00077Iteration: 190, Log-Lik: -1601.970, Max-Change: 0.00071Iteration: 191, Log-Lik: -1601.970, Max-Change: 0.00021Iteration: 192, Log-Lik: -1601.970, Max-Change: 0.00077Iteration: 193, Log-Lik: -1601.970, Max-Change: 0.00081Iteration: 194, Log-Lik: -1601.970, Max-Change: 0.00028Iteration: 195, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 196, Log-Lik: -1601.970, Max-Change: 0.00069Iteration: 197, Log-Lik: -1601.970, Max-Change: 0.00063Iteration: 198, Log-Lik: -1601.970, Max-Change: 0.00020Iteration: 199, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 200, Log-Lik: -1601.970, Max-Change: 0.00068Iteration: 201, Log-Lik: -1601.970, Max-Change: 0.00064Iteration: 202, Log-Lik: -1601.970, Max-Change: 0.00023Iteration: 203, Log-Lik: -1601.970, Max-Change: 0.00016Iteration: 204, Log-Lik: -1601.970, Max-Change: 0.00062Iteration: 205, Log-Lik: -1601.970, Max-Change: 0.00064Iteration: 206, Log-Lik: -1601.970, Max-Change: 0.00021Iteration: 207, Log-Lik: -1601.970, Max-Change: 0.00072Iteration: 208, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 209, Log-Lik: -1601.970, Max-Change: 0.00061Iteration: 210, Log-Lik: -1601.970, Max-Change: 0.00020Iteration: 211, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 212, Log-Lik: -1601.970, Max-Change: 0.00067Iteration: 213, Log-Lik: -1601.970, Max-Change: 0.00065Iteration: 214, Log-Lik: -1601.970, Max-Change: 0.00021Iteration: 215, Log-Lik: -1601.970, Max-Change: 0.00014Iteration: 216, Log-Lik: -1601.970, Max-Change: 0.00058Iteration: 217, Log-Lik: -1601.970, Max-Change: 0.00060Iteration: 218, Log-Lik: -1601.970, Max-Change: 0.00020Iteration: 219, Log-Lik: -1601.970, Max-Change: 0.00068Iteration: 220, Log-Lik: -1601.970, Max-Change: 0.00019Iteration: 221, Log-Lik: -1601.970, Max-Change: 0.00059Iteration: 222, Log-Lik: -1601.970, Max-Change: 0.00019Iteration: 223, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 224, Log-Lik: -1601.970, Max-Change: 0.00064Iteration: 225, Log-Lik: -1601.970, Max-Change: 0.00013Iteration: 226, Log-Lik: -1601.970, Max-Change: 0.00059Iteration: 227, Log-Lik: -1601.970, Max-Change: 0.00020Iteration: 228, Log-Lik: -1601.970, Max-Change: 0.00013Iteration: 229, Log-Lik: -1601.970, Max-Change: 0.00054Iteration: 230, Log-Lik: -1601.970, Max-Change: 0.00052Iteration: 231, Log-Lik: -1601.970, Max-Change: 0.00015Iteration: 232, Log-Lik: -1601.970, Max-Change: 0.00014Iteration: 233, Log-Lik: -1601.970, Max-Change: 0.00055Iteration: 234, Log-Lik: -1601.970, Max-Change: 0.00052Iteration: 235, Log-Lik: -1601.970, Max-Change: 0.00019Iteration: 236, Log-Lik: -1601.970, Max-Change: 0.00013Iteration: 237, Log-Lik: -1601.970, Max-Change: 0.00053Iteration: 238, Log-Lik: -1601.970, Max-Change: 0.00054Iteration: 239, Log-Lik: -1601.970, Max-Change: 0.00017Iteration: 240, Log-Lik: -1601.970, Max-Change: 0.00062Iteration: 241, Log-Lik: -1601.970, Max-Change: 0.00017Iteration: 242, Log-Lik: -1601.970, Max-Change: 0.00053Iteration: 243, Log-Lik: -1601.970, Max-Change: 0.00017Iteration: 244, Log-Lik: -1601.970, Max-Change: 0.00016Iteration: 245, Log-Lik: -1601.970, Max-Change: 0.00058Iteration: 246, Log-Lik: -1601.970, Max-Change: 0.00011Iteration: 247, Log-Lik: -1601.970, Max-Change: 0.00054Iteration: 248, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 249, Log-Lik: -1601.970, Max-Change: 0.00062Iteration: 250, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 251, Log-Lik: -1601.970, Max-Change: 0.00054Iteration: 252, Log-Lik: -1601.970, Max-Change: 0.00017Iteration: 253, Log-Lik: -1601.970, Max-Change: 0.00016Iteration: 254, Log-Lik: -1601.970, Max-Change: 0.00058Iteration: 255, Log-Lik: -1601.970, Max-Change: 0.00012Iteration: 256, Log-Lik: -1601.970, Max-Change: 0.00055Iteration: 257, Log-Lik: -1601.970, Max-Change: 0.00018Iteration: 258, Log-Lik: -1601.970, Max-Change: 0.00012Iteration: 259, Log-Lik: -1601.970, Max-Change: 0.00050Iteration: 260, Log-Lik: -1601.969, Max-Change: 0.00048Iteration: 261, Log-Lik: -1601.969, Max-Change: 0.00014Iteration: 262, Log-Lik: -1601.969, Max-Change: 0.00013Iteration: 263, Log-Lik: -1601.969, Max-Change: 0.00051Iteration: 264, Log-Lik: -1601.969, Max-Change: 0.00049Iteration: 265, Log-Lik: -1601.969, Max-Change: 0.00017Iteration: 266, Log-Lik: -1601.969, Max-Change: 0.00059Iteration: 267, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 268, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 269, Log-Lik: -1601.969, Max-Change: 0.00047Iteration: 270, Log-Lik: -1601.969, Max-Change: 0.00013Iteration: 271, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 272, Log-Lik: -1601.969, Max-Change: 0.00049Iteration: 273, Log-Lik: -1601.969, Max-Change: 0.00047Iteration: 274, Log-Lik: -1601.969, Max-Change: 0.00016Iteration: 275, Log-Lik: -1601.969, Max-Change: 0.00058Iteration: 276, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 277, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 278, Log-Lik: -1601.969, Max-Change: 0.00046Iteration: 279, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 280, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 281, Log-Lik: -1601.969, Max-Change: 0.00048Iteration: 282, Log-Lik: -1601.969, Max-Change: 0.00047Iteration: 283, Log-Lik: -1601.969, Max-Change: 0.00016Iteration: 284, Log-Lik: -1601.969, Max-Change: 0.00056Iteration: 285, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 286, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 287, Log-Lik: -1601.969, Max-Change: 0.00045Iteration: 288, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 289, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 290, Log-Lik: -1601.969, Max-Change: 0.00047Iteration: 291, Log-Lik: -1601.969, Max-Change: 0.00046Iteration: 292, Log-Lik: -1601.969, Max-Change: 0.00016Iteration: 293, Log-Lik: -1601.969, Max-Change: 0.00055Iteration: 294, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 295, Log-Lik: -1601.969, Max-Change: 0.00010Iteration: 296, Log-Lik: -1601.969, Max-Change: 0.00044Iteration: 297, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 298, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 299, Log-Lik: -1601.969, Max-Change: 0.00046Iteration: 300, Log-Lik: -1601.969, Max-Change: 0.00045Iteration: 301, Log-Lik: -1601.969, Max-Change: 0.00015Iteration: 302, Log-Lik: -1601.969, Max-Change: 0.00054Iteration: 303, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 304, Log-Lik: -1601.969, Max-Change: 0.00010Iteration: 305, Log-Lik: -1601.969, Max-Change: 0.00043Iteration: 306, Log-Lik: -1601.969, Max-Change: 0.00012Iteration: 307, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 308, Log-Lik: -1601.969, Max-Change: 0.00045Iteration: 309, Log-Lik: -1601.969, Max-Change: 0.00044Iteration: 310, Log-Lik: -1601.969, Max-Change: 0.00015Iteration: 311, Log-Lik: -1601.969, Max-Change: 0.00053Iteration: 312, Log-Lik: -1601.969, Max-Change: 0.00011Iteration: 313, Log-Lik: -1601.969, Max-Change: 0.00010
summary(x)
#> 
#> Rotation:  oblimin 
#> 
#> Rotated factor loadings: 
#> 
#>              F1      F2    h2
#> Comfort  0.6016  0.0312 0.382
#> Work    -0.0573  0.7971 0.592
#> Future   0.3302  0.5153 0.548
#> Benefit  0.7231 -0.0239 0.506
#> 
#> Rotated SS loadings:  0.997 0.902 
#> 
#> Factor correlations: 
#> 
#>       F1 F2
#> F1 1.000   
#> F2 0.511  1
summary(x, rotate = 'varimax')
#> 
#> Rotation:  varimax 
#> 
#> Rotated factor loadings: 
#> 
#>            F1    F2    h2
#> Comfort 0.579 0.216 0.382
#> Work    0.121 0.760 0.592
#> Future  0.428 0.605 0.548
#> Benefit 0.683 0.200 0.506
#> 
#> Rotated SS loadings:  0.999 1.03 
#> 
#> Factor correlations: 
#> 
#>    F1 F2
#> F1  1   
#> F2  0  1

# }