wald {mirt}	R Documentation

Wald statistics for mirt models

Description

Compute a Wald test given an L vector or matrix of numeric contrasts. Requires that the model information matrix be computed (including SE = TRUE when using the EM method). Use wald(model) to observe how the information matrix columns are named, especially if the estimated model contains constrained parameters (e.g., 1PL).

Usage

wald(object, L, C = 0)

Arguments

object
estimated object from mirt, bfactor, multipleGroup, mixedmirt, or mdirt

L
a coefficient matrix with dimensions nconstrasts x npars. Omitting this value will return the column names of the information matrix used to identify the (potentially constrained) parameters

C
a constant vector of population parameters to be compared along side L, where length(C) == ncol(L). By default a vector of 0's is constructed

Author(s)

Phil Chalmers rphilip.chalmers@gmail.com

Examples

## No test: 
#View parnumber index
data(LSAT7)
data <- expand.table(LSAT7)
mod <- mirt(data, 1, SE = TRUE)
coef(mod)

## $Item.1
##            a1     d  g  u
## par     0.988 1.856  0  1
## CI_2.5  0.642 1.600 NA NA
## CI_97.5 1.334 2.112 NA NA
## 
## $Item.2
##            a1     d  g  u
## par     1.081 0.808  0  1
## CI_2.5  0.749 0.629 NA NA
## CI_97.5 1.413 0.987 NA NA
## 
## $Item.3
##            a1     d  g  u
## par     1.706 1.804  0  1
## CI_2.5  1.077 1.404 NA NA
## CI_97.5 2.335 2.205 NA NA
## 
## $Item.4
##            a1     d  g  u
## par     0.765 0.486  0  1
## CI_2.5  0.500 0.339 NA NA
## CI_97.5 1.030 0.633 NA NA
## 
## $Item.5
##            a1     d  g  u
## par     0.736 1.855  0  1
## CI_2.5  0.435 1.630 NA NA
## CI_97.5 1.036 2.079 NA NA
## 
## $GroupPars
##         MEAN_1 COV_11
## par          0      1
## CI_2.5      NA     NA
## CI_97.5     NA     NA

# see how the information matrix relates to estimated parameters, and how it lines up
#   with the parameter index
(infonames <- wald(mod))

##  a1.1   d.2  a1.5   d.6  a1.9  d.10 a1.13  d.14 a1.17  d.18 
## 0.988 1.856 1.081 0.808 1.706 1.804 0.765 0.486 0.736 1.855

index <- mod2values(mod)
index[index$est, ]

##    group   item class name parnum     value lbound ubound  est prior.type
## 1    all Item.1  dich   a1      1 0.9879254   -Inf    Inf TRUE       none
## 2    all Item.1  dich    d      2 1.8560605   -Inf    Inf TRUE       none
## 5    all Item.2  dich   a1      5 1.0808847   -Inf    Inf TRUE       none
## 6    all Item.2  dich    d      6 0.8079786   -Inf    Inf TRUE       none
## 9    all Item.3  dich   a1      9 1.7058006   -Inf    Inf TRUE       none
## 10   all Item.3  dich    d     10 1.8042187   -Inf    Inf TRUE       none
## 13   all Item.4  dich   a1     13 0.7651853   -Inf    Inf TRUE       none
## 14   all Item.4  dich    d     14 0.4859966   -Inf    Inf TRUE       none
## 17   all Item.5  dich   a1     17 0.7357980   -Inf    Inf TRUE       none
## 18   all Item.5  dich    d     18 1.8545127   -Inf    Inf TRUE       none
##    prior_1 prior_2
## 1      NaN     NaN
## 2      NaN     NaN
## 5      NaN     NaN
## 6      NaN     NaN
## 9      NaN     NaN
## 10     NaN     NaN
## 13     NaN     NaN
## 14     NaN     NaN
## 17     NaN     NaN
## 18     NaN     NaN

#second item slope equal to 0?
L <- matrix(0, 1, 10)
L[1,3] <- 1
wald(mod, L)

##          W df            p
## 1 40.72587  1 1.751531e-10

#simultaneously test equal factor slopes for item 1 and 2, and 4 and 5
L <- matrix(0, 2, 10)
L[1,1] <- L[2, 7] <- 1
L[1,3] <- L[2, 9] <- -1
L

##      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
## [1,]    1    0   -1    0    0    0    0    0    0     0
## [2,]    0    0    0    0    0    0    1    0   -1     0

wald(mod, L)

##           W df         p
## 1 0.1796746  2 0.9140799

#logLiklihood tests (requires estimating a new model)
cmodel <- 'theta = 1-5
           CONSTRAIN = (1,2, a1), (4,5, a1)'
mod2 <- mirt(data, cmodel)
#or, eqivalently
#mod2 <- mirt(data, 1, constrain = list(c(1,5), c(13,17)))
anova(mod2, mod)

## 
## Model 1: mirt(data = data, model = cmodel)
## Model 2: mirt(data = data, model = 1, SE = TRUE)

##        AIC     AICc    SABIC      BIC    logLik    X2  df      p
## 1 5333.763 5333.908 5347.616 5373.025 -2658.881   NaN NaN    NaN
## 2 5337.610 5337.833 5354.927 5386.688 -2658.805 0.152   2 0.9268

## End(No test)

[Package mirt version 1.21 Index]